% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{abb_mat}
\alias{abb_mat}
\title{Print abbreviated matrix}
\usage{
abb_mat(exprObj, nrows, ncols)
}
\description{
print abbreviated matrix exprObj. Works for Matrix pkg denseMatrix,
matrix, data.frame and classes that inherit them.
}
\keyword{internal}
