% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{activeSpatUnit-generic}
\alias{activeSpatUnit-generic}
\alias{activeSpatUnit}
\alias{activeSpatUnit<-}
\alias{activeSpatUnit,giotto-method}
\alias{activeSpatUnit<-,giotto,character-method}
\title{Active spatial unit}
\usage{
activeSpatUnit(gobject, ...)

\S4method{activeSpatUnit}{giotto}(gobject)

\S4method{activeSpatUnit}{giotto,character}(gobject) <- value
}
\arguments{
\item{gobject}{giotto object}
}
\description{
Retrieve or set the active spatial unit. This value will be the
default spatial unit that the giotto object uses.
}
