% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_section.R
\name{adapt_aspect_ratio}
\alias{adapt_aspect_ratio}
\title{adapt_aspect_ratio}
\usage{
adapt_aspect_ratio(
  current_ratio,
  cell_locations,
  sdimx = NULL,
  sdimy = NULL,
  sdimz = NULL,
  mesh_obj = NULL
)
}
\description{
adapt the aspact ratio after inserting cross section mesh grid lines
}
