% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{add_img_array_alpha}
\alias{add_img_array_alpha}
\title{Add alpha channel to image array}
\usage{
add_img_array_alpha(x, alpha)
}
\arguments{
\item{x}{image array to use}

\item{alpha}{global alpha value to use. Numeric. Scales from 0 to 1, with 0
being fully transparent and 1 being fully visible}
}
\value{
image array with 4th channel for transparency
}
\description{
Add alpha channel to image array
}
\details{
Add 4th alpha channel to 3 channel RGB image arrays
}
\keyword{internal}
