% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction.R
\name{addCellIntMetadata}
\alias{addCellIntMetadata}
\title{addCellIntMetadata}
\usage{
addCellIntMetadata(
  gobject,
  spatial_network = "spatial_network",
  cluster_column,
  cell_interaction,
  name = "select_int",
  return_gobject = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{spatial_network}{name of spatial network to use}

\item{cluster_column}{column of cell types}

\item{cell_interaction}{cell-cell interaction to use}

\item{name}{name for the new metadata column}

\item{return_gobject}{return an updated giotto object}
}
\value{
Giotto object
}
\description{
Creates an additional metadata column with information about interacting and non-interacting cell types of the
selected cell-cell interaction.
}
\details{
This function will create an additional metadata column which selects interacting cell types for a specific cell-cell
interaction. For example, if you want to color interacting astrocytes and oligodendrocytes it will create a new metadata column with
the values "select_astrocytes", "select_oligodendrocytes", "other_astrocytes", "other_oligodendroyctes" and "other". Where "other" is all
other cell types found within the selected cell type column.
}
