% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{addCellMetadata}
\alias{addCellMetadata}
\title{addCellMetadata}
\usage{
addCellMetadata(gobject, new_metadata, by_column = F, column_cell_ID = NULL)
}
\arguments{
\item{gobject}{giotto object}

\item{new_metadata}{new metadata to use}

\item{by_column}{merge metadata based on cell_ID column in pDataDT}

\item{column_cell_ID}{column name of new metadata to use if by_column = TRUE}
}
\value{
giotto object
}
\description{
adds cell metadata to the giotto object
}
\details{
Description of how to add cell metadata ...
}
\examples{
    addCellMetadata(gobject)
}
