% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visuals.R
\name{addCellsFromPolygon}
\alias{addCellsFromPolygon}
\title{Add corresponding polygon IDs to cell metadata}
\usage{
addCellsFromPolygon(gobject, cellsFromPolygon, feat_type = "rna")
}
\arguments{
\item{gobject}{A Giotto object}

\item{cellsFromPolygon}{A `SpatVector` with cell IDs located inside each polygon}

\item{feat_type}{feature name where metadata will be added}
}
\value{
A Giotto object with a modified cell_metadata slot that includes the
polygon name where each cell is located or NA if the cell is not located
within a polygon area
}
\description{
Add corresponding polygon IDs to cell metadata
}
\examples{

## Plot interactive polygons
my_polygon_coords <- plotInteractivePolygons(my_spatPlot)

## Add polygon coordinates to Giotto object
my_giotto_polygons <- createGiottoPolygonsFromDfr(my_polygon_coords)
my_giotto_object <- addGiottoPolygons(gobject = my_giotto_object,
                                      gpolygons = list(my_giotto_polygons))

## Get cells located within polygons area
my_polygon_cells <- getCellsFromPolygon(my_giotto_object)

my_giotto_object <- addCellsFromPolygon(my_giotto_object, my_polygon_cells)
}
