% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{addFeatMetadata}
\alias{addFeatMetadata}
\title{Add feature metadata}
\usage{
addFeatMetadata(
  gobject,
  feat_type = NULL,
  spat_unit = NULL,
  new_metadata,
  by_column = F,
  column_feat_ID = NULL,
  vector_name = NULL
)
}
\arguments{
\item{gobject}{giotto object}

\item{feat_type}{feature type}

\item{spat_unit}{spatial unit}

\item{new_metadata}{new metadata to use)}

\item{by_column}{merge metadata based on \emph{feat_ID} column in \code{\link{fDataDT}}}

\item{column_feat_ID}{column name of new metadata to use if by_column = TRUE}

\item{vector_name}{(optional) custom name if you provide a single vector}
}
\value{
giotto object
}
\description{
Adds feature metadata to the giotto object
}
\details{
You can add additional feature metadata in two manners: \cr
1. Provide a data.table or data.frame with feature annotations in the same order as the \emph{feat_ID} column in fDataDT(gobject) \cr
2. Provide a data.table or data.frame with feature annotations and specify which column contains the feature IDs,
these feature IDs need to match with the \emph{feat_ID} column in fDataDT(gobject)
}
