% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{addFeatStatistics}
\alias{addFeatStatistics}
\title{Add feature statistics}
\usage{
addFeatStatistics(
  gobject,
  feat_type = NULL,
  spat_unit = NULL,
  expression_values = c("normalized", "scaled", "custom"),
  detection_threshold = 0,
  return_gobject = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{feat_type}{feature type}

\item{spat_unit}{spatial unit}

\item{expression_values}{expression values to use}

\item{detection_threshold}{detection threshold to consider a gene detected}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}
}
\value{
giotto object if return_gobject = TRUE
}
\description{
Adds feature statistics to the giotto object
}
\details{
This function will add the following statistics to feature metadata:
\itemize{
  \item{nr_cells: }{Denotes in how many cells the feature is detected}
  \item{per_cells: }{Denotes in what percentage of cells the feature is detected}
  \item{total_expr: }{Shows the total sum of feature expression in all cells}
  \item{mean_expr: }{Average feature expression in all cells}
  \item{mean_expr_det: }{Average feature expression in cells with detectable levels of the gene}
}
}
