% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{addGeneMetadata}
\alias{addGeneMetadata}
\title{addGeneMetadata}
\usage{
addGeneMetadata(gobject, new_metadata, by_column = F, column_gene_ID = NULL)
}
\arguments{
\item{gobject}{giotto object}

\item{new_metadata}{new metadata to use}

\item{by_column}{merge metadata based on gene_ID column in fDataDT}

\item{column_cell_ID}{column name of new metadata to use if by_column = TRUE}
}
\value{
giotto object
}
\description{
adds gene metadata to the giotto object
}
\details{
Description of how to add gene metadata ...
}
\examples{
    addGeneMetadata(gobject)
}
