% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{addGiottoLargeImage}
\alias{addGiottoLargeImage}
\title{addGiottoLargeImage}
\usage{
addGiottoLargeImage(
  gobject = NULL,
  largeImages = NULL,
  spat_loc_name = NULL,
  scale_factor = NULL,
  negative_y = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{largeImages}{list of giottoLargeImage objects}

\item{spat_loc_name}{provide spatial location slot in Giotto to align images. (optional)}

\item{scale_factor}{provide scale of image pixel dimensions relative to spatial coordinates.}

\item{negative_y}{map image to negative y spatial values if TRUE during automatic alignment. Meaning that origin is in upper left instead of lower left.}

\item{verbose}{be verbose}
}
\value{
an updated Giotto object with access to the list of images
}
\description{
Adds giotto image objects to your giotto object
}
