% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_structures.R
\name{addGiottoPoints3D}
\alias{addGiottoPoints3D}
\title{Add sub cellular 3D coordinates to Giotto object}
\usage{
addGiottoPoints3D(gobject, coords, feat_type = "rna")
}
\arguments{
\item{gobject}{A Giotto object.}

\item{coords}{A \link{data.frame} or `spatVector` with at least xyz coordinates and feature ids.}

\item{feat_type}{a character. The feat_type must previously exist in the Giotto object. Default = "rna".}
}
\value{
A Giotto object with a `spatVector` object in the feat_info slot
}
\description{
Add sub cellular 3D coordinates to Giotto object
}
