% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_hmrf.R
\name{addHMRF_V2}
\alias{addHMRF_V2}
\title{addHMRF_V2}
\usage{
addHMRF_V2(gobject, HMRFoutput, name = "hmrf")
}
\arguments{
\item{gobject}{giotto object}

\item{HMRFoutput}{result object from HMRF model}

\item{name}{name of HMRF models}
}
\value{
giotto object
}
\description{
function to add HMRF Domain Type to cell meta data
}
\details{
This function appends HMRF domain clusters to corresponding cell meta data
for all the beta values, with the given HMRF model names. For example, if
name = ‘hmrf1’ and name of result in HMRFoutput is ‘k=8 b=0.00’, the
appended cell meta data column will be named with ‘hmrf1 k=8 b=0.00’
}
\examples{
g <- GiottoData::loadGiottoMini("visium")
g <- binSpect(g, return_gobject = TRUE)
HMRF_init_obj <- initHMRF_V2(gobject = g, cl.method = "km")
HMRFoutput <- doHMRF_V2(HMRF_init_obj = HMRF_init_obj, betas = c(0, 5, 2))

addHMRF_V2(gobject = g, HMRFoutput = HMRFoutput)
}
