% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{addImage}
\alias{addImage}
\title{addImage}
\usage{
addImage(gobject, images)
}
\arguments{
\item{gobject}{giotto object}

\item{images}{list of giotto image objects, see \code{\link{createGiottoImage}}}
}
\value{
an updated Giotto object with access to the list of images
}
\description{
Adds giotto image objects to your giotto object
}
\examples{
    addImage(mg_object)
}
