% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_structures.R
\name{addSpatialCentroidLocationsLayer}
\alias{addSpatialCentroidLocationsLayer}
\title{addSpatialCentroidLocationsLayer}
\usage{
addSpatialCentroidLocationsLayer(
  gobject,
  poly_info = "cell",
  feat_type = NULL,
  provenance = poly_info,
  spat_loc_name = "raw",
  return_gobject = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{poly_info}{polygon information}

\item{feat_type}{feature type}

\item{provenance}{(optional) provenance to assign to generated spatLocsObj. If
not provided, provenance will default to \code{poly_info}}

\item{spat_loc_name}{name to give to the created spatial locations}

\item{return_gobject}{return giotto object (default: TRUE)}
}
\value{
If \code{return_gobject = TRUE} the giotto object containing the calculated
  polygon centroids will be returned. If \code{return_gobject = FALSE} only the
  generated polygon centroids will be returned as spatLocsObj.
}
\description{
Calculates the centroid locations for the polygons within one selected layer
}
\concept{centroid}
