% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_visuals.R
\name{aes_string2}
\alias{aes_string2}
\title{aes_string2}
\usage{
aes_string2(...)
}
\arguments{
\item{\dots}{aes_string parameters}
}
\description{
makes sure aes_string can also be used with names that start with numeric values
}
\keyword{internal}
