% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto.R
\docType{class}
\name{aggregateExprObj-class}
\alias{aggregateExprObj-class}
\title{S4 aggregateExprObj}
\description{
Framework to store aggregated expression information
}
\section{Slots}{

\describe{
\item{\code{name}}{name of aggregateExprObj}

\item{\code{sparse}}{(boolean) if matrix is sparse}

\item{\code{expressionMat}}{matrix of expression information}

\item{\code{provenance}}{origin data of aggregated expression information (if applicable)}

\item{\code{spat_unit}}{spatial unit of aggregated expression (e.g. 'cell')}

\item{\code{feat_type}}{feature type of aggregated expression (e.g. 'rna', 'protein')}

\item{\code{misc}}{misc}
}}

