% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{aggregateStacks}
\alias{aggregateStacks}
\title{aggregateStacks}
\usage{
aggregateStacks(
  gobject,
  spat_units,
  feat_type,
  values,
  summarize_expression = "sum",
  summarize_locations = "mean",
  new_spat_unit = "aggregate"
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_units}{spatial units to aggregate}

\item{feat_type}{feature type}

\item{values}{values to use}

\item{summarize_expression}{method to summarize expression information}

\item{summarize_locations}{method to summarize spatial location information}

\item{new_spat_unit}{new name for aggregated spatial unit}
}
\value{
giotto object
}
\description{
aggregate expression matrices from different z-stacks
}
\details{
Combines both \code{\link{aggregateStacksExpression}} and \code{\link{aggregateStacksLocations}}
}
\seealso{
Other aggregate stacks: 
\code{\link{aggregateStacksExpression}()},
\code{\link{aggregateStacksLocations}()},
\code{\link{aggregateStacksPolygonOverlaps}()},
\code{\link{aggregateStacksPolygons}()}
}
\concept{aggregate stacks}
