% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{aggregateStacksExpression}
\alias{aggregateStacksExpression}
\title{aggregateStacksExpression}
\usage{
aggregateStacksExpression(
  gobject,
  spat_units,
  feat_type,
  values = "raw",
  summarize = "sum",
  new_spat_unit = "aggregate",
  verbose = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_units}{spatial units to aggregate}

\item{feat_type}{feature type}

\item{values}{values to use}

\item{summarize}{method to summarize expression information}

\item{new_spat_unit}{new name for aggregated spatial unit}

\item{verbose}{verbosity}
}
\value{
giotto object
}
\description{
aggregate expression matrices from different z-stacks
}
\seealso{
Other aggregate stacks: 
\code{\link{aggregateStacksLocations}()},
\code{\link{aggregateStacksPolygonOverlaps}()},
\code{\link{aggregateStacksPolygons}()},
\code{\link{aggregateStacks}()}
}
\concept{aggregate stacks}
