% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{aggregateStacksLocations}
\alias{aggregateStacksLocations}
\title{aggregateStacksLocations}
\usage{
aggregateStacksLocations(
  gobject,
  spat_units,
  values = "raw",
  summarize = "mean",
  new_spat_unit = "aggregate"
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_units}{spatial units to aggregate}

\item{values}{values to use}

\item{summarize}{method to summarize spatial location information}

\item{new_spat_unit}{new name for aggregated spatial unit}
}
\value{
giotto object
}
\description{
aggregate expression matrices from different z-stacks
}
\seealso{
Other aggregate stacks: 
\code{\link{aggregateStacksExpression}()},
\code{\link{aggregateStacksPolygonOverlaps}()},
\code{\link{aggregateStacksPolygons}()},
\code{\link{aggregateStacks}()}
}
\concept{aggregate stacks}
