% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{aggregateStacksPolygonOverlaps}
\alias{aggregateStacksPolygonOverlaps}
\title{aggregateStacksPolygonOverlaps}
\usage{
aggregateStacksPolygonOverlaps(
  gobject,
  spat_units,
  feat_type,
  new_spat_unit = "aggregate"
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_units}{spatial units to aggregate}

\item{feat_type}{feature type used for overlap calculations}

\item{new_spat_unit}{new name for aggregated spatial unit}
}
\value{
giotto object
}
\description{
aggregate polygons overlap information from different z-stacks
}
\seealso{
Other aggregate stacks: 
\code{\link{aggregateStacksExpression}()},
\code{\link{aggregateStacksLocations}()},
\code{\link{aggregateStacksPolygons}()},
\code{\link{aggregateStacks}()}
}
\concept{aggregate stacks}
