% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction.R
\name{allCellCellcommunicationsScores}
\alias{allCellCellcommunicationsScores}
\title{allCellCellcommunicationsScores}
\usage{
allCellCellcommunicationsScores(gobject,
  spatial_network_name = "spatial_network",
  cluster_column = "cell_types", random_iter = 100, gene_set_1,
  gene_set_2, log2FC_addendum = 0.1, min_observations = 2,
  verbose = c("a little", "a lot", "none"))
}
\arguments{
\item{gobject}{giotto object to use}

\item{spatial_network_name}{spatial network to use for identifying interacting cells}

\item{cluster_column}{cluster column with cell type information}

\item{random_iter}{number of iterations}

\item{gene_set_1}{first specific gene set from gene pairs}

\item{gene_set_2}{second specific gene set from gene pairs}

\item{log2FC_addendum}{addendum to add when calculating log2FC}

\item{min_observations}{minimum number of interactions needed to be considered}

\item{verbose}{verbose}
}
\value{
Cell-Cell communication scores for gene pairs based on spatial interaction
}
\description{
All Cell-Cell communication scores based on spatial expression of interacting cells
}
\details{
Details will follow.
}
\examples{
    allCellCellcommunicationsScores(gobject)
}
