% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_genes.R
\name{analyzeData}
\alias{analyzeData}
\alias{analyzeData,allMatrix,covGroupsParam-method}
\alias{analyzeData,allMatrix,covLoessParam-method}
\alias{analyzeData,allMatrix,varParam-method}
\title{Data Analysis via Parameter Dispatch}
\usage{
\S4method{analyzeData}{allMatrix,covGroupsParam}(x, param, ...)

\S4method{analyzeData}{allMatrix,covLoessParam}(x, param, ...)

\S4method{analyzeData}{allMatrix,varParam}(x, param, ...)
}
\arguments{
\item{x}{data to analyze}

\item{param}{an [analyzeParam-class] inheriting object defining the analysis
operation and its settings}

\item{\dots}{additional params passed to specific methods}
}
\value{
a `data.table` of computed values
}
\description{
Compute statistics or scores from matrix-type data. `analyzeData()` is a
generic that dispatches on both `x` (the data) and `param` (the analysis
operation). Methods return a `data.table` of computed values; any downstream
thresholding or selection is a separate step.
}
