% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interoperability.R
\name{anndataToGiotto}
\alias{anndataToGiotto}
\title{Convert anndata to Giotto}
\usage{
anndataToGiotto(
  anndata_path = NULL,
  n_key_added = NULL,
  spatial_key_added = NULL,
  spat_unit = NULL,
  feat_type = NULL,
  python_path = NULL
)
}
\arguments{
\item{anndata_path}{path to the .h5ad file}

\item{n_key_added}{equivalent of "key_added" argument from scanpy.pp.neighbors().
Cannot be "spatial". This becomes the name of the nearest network in the gobject.}

\item{spatial_key_added}{Default = NULL}

\item{spat_unit}{desired spatial unit for conversion, default NULL}

\item{feat_type}{desired feature type for conversion, default NULL}

\item{python_path}{path to python executable within a conda/miniconda environment}
}
\value{
Giotto object
}
\description{
Converts a spatial anndata (e.g. scanpy) .h5ad file into a Giotto object
}
\details{
Function in beta. Converts a .h5ad file into a Giotto object.
   The returned Giotto Object will take default insructions with the
   exception of the python path, which may be customized.
   See \code{\link{changeGiottoInstructions}} to modify instructions after creation.
}
