% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_structures.R
\name{annotateSpatialGrid}
\alias{annotateSpatialGrid}
\title{annotateSpatialGrid}
\usage{
annotateSpatialGrid(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  spat_loc_name = "raw",
  spatial_grid_name = "spatial_grid",
  cluster_columns = NULL
)
}
\arguments{
\item{gobject}{Giotto object}

\item{spat_unit}{spatial unit}

\item{feat_type}{feature type}

\item{spat_loc_name}{name of spatial locations}

\item{spatial_grid_name}{name of spatial grid, see \code{\link{showGiottoSpatGrids}}}

\item{cluster_columns}{names of cell metadata, see \code{\link{pDataDT}}}
}
\value{
annotated spatial grid data.table
}
\description{
annotate spatial grid with cell ID and cell metadata (optional)
}
