% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{as.data.frame.giottoTree}
\alias{as.data.frame.giottoTree}
\alias{as.data.table.giottoTree}
\title{Splits of a cluster tree as a table}
\usage{
\method{as.data.frame}{giottoTree}(x, ...)

\method{as.data.table}{giottoTree}(x, ...)
}
\arguments{
\item{x}{a `giottoTree`}

\item{...}{ignored}
}
\value{
a `data.frame`, or a `data.table` from `as.data.table()`
}
\description{
One row per split (internal node) of a tree from
[calculateClusterTree()], root first:
* `nodeID` --- the `tree$merge` row. The same node numbering
  [findClusterTreeMarkers()], [writeClusterTreeQuery()] and
  [annotateClusterTree()] (`labels$nodes`) use.
* `node_h` --- the height of the join
* `left`, `right` --- the clusters on each side, as character vectors
}
\examples{
g <- GiottoData::loadGiottoMini("visium")
tree <- calculateClusterTree(g, cluster_column = "leiden_clus")
data.table::as.data.table(tree)
}
