% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction.R
\name{average_gene_gene_expression_in_groups}
\alias{average_gene_gene_expression_in_groups}
\title{average_gene_gene_expression_in_groups}
\usage{
average_gene_gene_expression_in_groups(
  gobject,
  cluster_column = "cell_types",
  gene_set_1,
  gene_set_2
)
}
\arguments{
\item{gobject}{giotto object to use}

\item{cluster_column}{cluster column with cell type information}

\item{gene_set_1}{first specific gene set from gene pairs}

\item{gene_set_2}{second specific gene set from gene pairs}
}
\value{
data.table with average expression scores for each cluster
}
\description{
calculate average expression per cluster
}
\keyword{internal}
