% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_genes.R
\name{binGetSpatialGenes}
\alias{binGetSpatialGenes}
\title{binGetSpatialGenes}
\usage{
binGetSpatialGenes(
  gobject,
  bin_method = c("kmeans", "rank"),
  expression_values = c("normalized", "scaled", "custom"),
  spatial_network_name = "spatial_network",
  nstart = 3,
  iter_max = 10,
  percentage_rank = 10,
  do_fisher_test = F,
  community_expectation = 5,
  verbose = F
)
}
\arguments{
\item{gobject}{giotto object}

\item{bin_method}{method to binarize gene expression}

\item{expression_values}{expression values to use}

\item{spatial_network_name}{name of spatial network to use (default = 'spatial_network')}

\item{nstart}{kmeans: nstart parameter}

\item{iter_max}{kmeans: iter.max parameter}

\item{do_fisher_test}{perform fisher test}

\item{community_expectation}{cell degree expectation in spatial communities}

\item{verbose}{be verbose}

\item{rank_percentage}{percentage of top cells for binarization}
}
\value{
giotto object spatial genes appended to fDataDT
}
\description{
compute genes that are spatially clustered
}
\details{
Description of how we compute spatial genes.
}
\examples{
    binGetSpatialGenes(gobject)
}
