% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_genes.R
\name{binSpectMultiMatrix}
\alias{binSpectMultiMatrix}
\title{binSpectMultiMatrix}
\usage{
binSpectMultiMatrix(
  expression_matrix,
  spatial_networks,
  bin_method = c("kmeans", "rank"),
  subset_feats = NULL,
  kmeans_algo = c("kmeans", "kmeans_arma", "kmeans_arma_subset"),
  nstart = 3,
  iter_max = 10,
  extreme_nr = 50,
  sample_nr = 50,
  percentage_rank = c(10, 30),
  do_fisher_test = TRUE,
  adjust_method = "fdr",
  calc_hub = FALSE,
  hub_min_int = 3,
  get_av_expr = TRUE,
  get_high_expr = TRUE,
  implementation = c("data.table", "simple", "matrix"),
  group_size = "automatic",
  do_parallel = TRUE,
  cores = NA,
  verbose = T,
  knn_params = NULL,
  set.seed = NULL,
  summarize = c("adj.p.value", "p.value")
)
}
\arguments{
\item{expression_matrix}{expression matrix}

\item{spatial_networks}{list of spatial networks in data.table format}

\item{bin_method}{method to binarize gene expression}

\item{subset_feats}{only select a subset of features to test}

\item{kmeans_algo}{kmeans algorithm to use (kmeans, kmeans_arma, kmeans_arma_subset)}

\item{nstart}{kmeans: nstart parameter}

\item{iter_max}{kmeans: iter.max parameter}

\item{extreme_nr}{number of top and bottom cells (see details)}

\item{sample_nr}{total number of cells to sample (see details)}

\item{percentage_rank}{vector of percentages of top cells for binarization}

\item{do_fisher_test}{perform fisher test}

\item{adjust_method}{p-value adjusted method to use (see \code{\link[stats]{p.adjust}})}

\item{calc_hub}{calculate the number of hub cells}

\item{hub_min_int}{minimum number of cell-cell interactions for a hub cell}

\item{get_av_expr}{calculate the average expression per gene of the high expressing cells}

\item{get_high_expr}{calculate the number of high expressing cells  per gene}

\item{implementation}{enrichment implementation (data.table, simple, matrix)}

\item{group_size}{number of genes to process together with data.table implementation (default = automatic)}

\item{do_parallel}{run calculations in parallel with mclapply}

\item{cores}{number of cores to use if do_parallel = TRUE}

\item{verbose}{be verbose}

\item{knn_params}{list of parameters to create spatial kNN network}

\item{set.seed}{set a seed before kmeans binarization}

\item{summarize}{summarize the p-values or adjusted p-values}
}
\value{
data.table with results
}
\description{
binSpect for a single spatial network and a provided expression matrix
}
