% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{box_chars}
\alias{box_chars}
\alias{is_latex_output}
\alias{is_utf8_output}
\title{Box characters}
\usage{
box_chars()

is_latex_output()

is_utf8_output()
}
\description{
Helper function to print unicode characters using escape codes.
}
\details{
Much inspiration taken from \pkg{fs} \href{https://rdrr.io/cran/fs/src/R/tree.R}{tree.R}
}
\section{Functions}{
\itemize{
\item \code{is_latex_output()}: Determine if print is latex output

\item \code{is_utf8_output()}: Determine if system is using UTF-8 encoding

}}
\keyword{internal}
