% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONTraC_wrapper.R
\name{cal_cell_niche_cluster_bin}
\alias{cal_cell_niche_cluster_bin}
\title{cal_cell_niche_cluster_bin}
\usage{
cal_cell_niche_cluster_bin(
  gobject,
  spat_unit = "cell",
  feat_type = "niche cluster"
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit (e.g. "cell")}

\item{feat_type}{feature type (e.g. "rna", "dna", "protein")}
}
\value{
gobject with binarized cell-level niche cluster assignment
}
\description{
calculate binarized cell-level niche cluster assignment
}
