% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_genes.R
\name{calc_cov_group_HVF}
\alias{calc_cov_group_HVF}
\title{calc_cov_group_HVF}
\usage{
calc_cov_group_HVF(
  feat_in_cells_detected,
  nr_expression_groups = 20,
  zscore_threshold = 1,
  show_plot = NA,
  return_plot = NA,
  save_plot = NA
)
}
\description{
calc_cov_group_HVF
}
\keyword{internal}
