% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_genes.R
\name{calc_cov_group_HVG}
\alias{calc_cov_group_HVG}
\title{calc_cov_group_HVG}
\usage{
calc_cov_group_HVG(
  feat_in_cells_detected,
  nr_expression_groups = 20,
  zscore_threshold = 1
)
}
\description{
calc_cov_group_HVG
}
\keyword{internal}
