% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_genes.R
\name{calc_cov_loess_HVF}
\alias{calc_cov_loess_HVF}
\title{calc_cov_loess_HVF}
\usage{
calc_cov_loess_HVF(
  feat_in_cells_detected,
  difference_in_cov = 0.1,
  show_plot = NA,
  return_plot = NA,
  save_plot = NA
)
}
\description{
calc_cov_loess_HVF
}
\keyword{internal}
