% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_genes.R
\name{calc_cov_loess_HVG}
\alias{calc_cov_loess_HVG}
\title{calc_cov_loess_HVG}
\usage{
calc_cov_loess_HVG(feat_in_cells_detected, difference_in_cov = 0.1)
}
\description{
calc_cov_loess_HVG
}
\keyword{internal}
