% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_genes.R
\name{calc_spatial_enrichment_DT}
\alias{calc_spatial_enrichment_DT}
\title{calc_spatial_enrichment_DT}
\usage{
calc_spatial_enrichment_DT(
  bin_matrix,
  spatial_network,
  calc_hub = F,
  hub_min_int = 3,
  group_size = "automatic",
  do_fisher_test = TRUE,
  adjust_method = "fdr",
  cores = NA
)
}
\description{
calculate spatial enrichment using the data.table implementation
}
\keyword{internal}
