% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_genes.R
\name{calc_spatial_enrichment_matrix}
\alias{calc_spatial_enrichment_matrix}
\title{calc_spatial_enrichment_matrix}
\usage{
calc_spatial_enrichment_matrix(
  spatial_network,
  bin_matrix,
  adjust_method = "fdr",
  do_fisher_test = TRUE,
  do_parallel = TRUE,
  cores = NA,
  calc_hub = FALSE,
  hub_min_int = 3,
  verbose = TRUE
)
}
\description{
calculate spatial enrichment using a matrix approach
}
\keyword{internal}
