% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_genes.R
\name{calc_spatial_enrichment_minimum}
\alias{calc_spatial_enrichment_minimum}
\title{calc_spatial_enrichment_minimum}
\usage{
calc_spatial_enrichment_minimum(
  spatial_network,
  bin_matrix,
  adjust_method = "fdr",
  do_fisher_test = TRUE
)
}
\description{
calculate spatial enrichment using a simple and efficient for loop
}
\keyword{internal}
