% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_genes.R
\name{calc_var_HVF}
\alias{calc_var_HVF}
\title{calc_var_HVF}
\usage{
calc_var_HVF(
  scaled_matrix,
  var_threshold = 1.5,
  var_number = NULL,
  show_plot = NA,
  return_plot = NA,
  save_plot = NA
)
}
\description{
calc_var_HVF
}
\keyword{internal}
