% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_genes.R
\name{calculate_binarized_spatial_genes}
\alias{calculate_binarized_spatial_genes}
\title{calculate_binarized_spatial_genes}
\usage{
calculate_binarized_spatial_genes(gobject,
  expression_values = c("normalized", "scaled", "custom"),
  spatial_network_name = "spatial_network", bin_method = c("kmeans",
  "rank"), percentage_rank = 10)
}
\description{
compute spatial genes on binarized gene values
}
