% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_structures.R
\name{calculate_centroids_polygons}
\alias{calculate_centroids_polygons}
\title{Calculate polygon centroids}
\usage{
calculate_centroids_polygons(
  gpolygon,
  name = "centroids",
  append_gpolygon = TRUE
)
}
\description{
calculates centroids from selected polygons
}
\keyword{internal}
