% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_structures.R
\name{calculate_distance_and_weight}
\alias{calculate_distance_and_weight}
\title{calculate_distance_and_weight}
\usage{
calculate_distance_and_weight(
  networkDT,
  sdimx = "sdimx",
  sdimy = "sdimy",
  sdimz = "sdimz",
  d2_or_d3 = c(2, 3)
)
}
\description{
calculate_distance_and_weight
}
