% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_genes.R
\name{calculate_spatial_genes}
\alias{calculate_spatial_genes}
\title{calculateSpatialGenes}
\usage{
calculate_spatial_genes(gobject, expression_values = c("normalized",
  "scaled", "custom"), spatial_network_name = "spatial_network",
  method = c("gini"), simulations = 0, detection_threshold = 0)
}
\description{
compute spatial genes on gene expression products between neighbors
}
