% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_genes.R
\name{calculate_spatial_genes_python}
\alias{calculate_spatial_genes_python}
\title{calculate_spatial_genes_python}
\usage{
calculate_spatial_genes_python(
  gobject,
  expression_values = c("normalized", "scaled", "custom"),
  metric = "euclidean",
  subset_genes = NULL,
  rbp_p = 0.95,
  examine_top = 0.3,
  python_path = NULL
)
}
\arguments{
\item{gobject}{giotto object}

\item{expression_values}{expression values to use}

\item{metric}{distance metric to use}

\item{subset_genes}{only run on this subset of genes}

\item{rbp_p}{fractional binarization threshold}

\item{examine_top}{top fraction to evaluate with silhouette}

\item{python_path}{specify specific path to python if required}
}
\value{
data.table with spatial scores
}
\description{
Calculate spatial genes using distance matrix.
}
\details{
Description of how we compute spatial pattern genes.
}
\examples{
    calculate_spatial_genes_python(gobject)
}
