% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_genes.R
\name{calculateHVG}
\alias{calculateHVG}
\title{calculateHVG}
\usage{
calculateHVG(
  gobject,
  expression_values = c("normalized", "scaled", "custom"),
  method = c("cov_groups", "cov_loess", "gini_loess"),
  reverse_log_scale = T,
  logbase = 2,
  expression_threshold = 0,
  nr_expression_groups = 20,
  zscore_threshold = 1.5,
  HVGname = "hvg",
  difference_in_variance = 1,
  show_plot = T,
  return_gobject = T
)
}
\arguments{
\item{gobject}{giotto object}

\item{expression_values}{expression values to use}

\item{method}{method to calculate highly variable genes}

\item{reverse_log_scale}{reverse log-scale of expression values}

\item{logbase}{if reverse_log_scale is TRUE, which log base was used?}

\item{expression_threshold}{expression threshold to consider a gene detected}

\item{nr_expression_groups}{number of expression groups for cov_groups}

\item{zscore_threshold}{zscore to select hvg for cov_groups}

\item{HVGname}{name for highly variable genes in cell metadata}

\item{difference_in_variance}{minimum difference in variance required}

\item{show_plot}{show plots}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}
}
\value{
giotto object highly variable genes appended to gene metadata (fDataDT)
}
\description{
compute highly variable genes
}
\details{
Description of how we compute highly variable genes.
}
\examples{
    calculateHVG(gobject)
}
