% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{calculateMetaTable}
\alias{calculateMetaTable}
\title{calculateMetaTable}
\usage{
calculateMetaTable(
  gobject,
  feat_type = NULL,
  spat_unit = NULL,
  expression_values = c("normalized", "scaled", "custom"),
  metadata_cols = NULL,
  selected_feats = NULL,
  selected_genes = NULL
)
}
\arguments{
\item{gobject}{giotto object}

\item{feat_type}{feature type}

\item{spat_unit}{spatial unit}

\item{expression_values}{expression values to use}

\item{metadata_cols}{annotation columns found in pDataDT(gobject)}

\item{selected_genes}{subset of genes to use}
}
\value{
data.table with average expression values for each gene per (combined) annotation
}
\description{
calculates the average gene expression for one or more (combined) annotation columns.
}
