% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{calculateMetaTableCells}
\alias{calculateMetaTableCells}
\title{calculateMetaTableCells}
\usage{
calculateMetaTableCells(
  gobject,
  value_cols = NULL,
  metadata_cols = NULL,
  spat_enr_names = NULL
)
}
\arguments{
\item{gobject}{giotto object}

\item{value_cols}{metadata or enrichment value columns to use}

\item{metadata_cols}{annotation columns found in pDataDT(gobject)}

\item{spat_enr_names}{which spatial enrichment results to include}
}
\value{
data.table with average metadata values per (combined) annotation
}
\description{
calculates the average metadata values for one or more (combined) annotation columns.
}
