% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_structures.R
\name{calculateOverlapOLD}
\alias{calculateOverlapOLD}
\title{calculateOverlapOLD}
\usage{
calculateOverlapOLD(
  gobject,
  poly_info = "cell",
  feat_info = "rna",
  x_step = 200,
  y_step = 200,
  return_gobject = TRUE,
  name_overlap = NULL
)
}
\arguments{
\item{gobject}{giotto object}

\item{poly_info}{polygon information}

\item{feat_info}{feature information}

\item{x_step}{x-direction step to travel over the polygon landscape}

\item{y_step}{y-direction step to travel over the polygon landscape}

\item{return_gobject}{return giotto object (default: TRUE)}

\item{name_overlap}{name for the overlap results (default to feat_info parameter)}
}
\value{
giotto object or spatVector with overlapping information
}
\description{
calculate overlap between cellular structures (polygons) and features (points)
}
\keyword{calculateOverlapOLD}
