% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_structures.R
\name{calculateOverlapPolygonImages}
\alias{calculateOverlapPolygonImages}
\title{calculateOverlapPolygonImages}
\usage{
calculateOverlapPolygonImages(
  gobject,
  name_overlap = "images",
  spatial_info = "cell",
  poly_ID_names = NULL,
  image_names = NULL,
  return_gobject = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{name_overlap}{name for the overlap results (default to feat_info parameter)}

\item{spatial_info}{polygon information}

\item{poly_ID_names}{(optional) list of poly_IDs to use}

\item{image_names}{names of the images with raw data}

\item{return_gobject}{return giotto object (default: TRUE)}

\item{verbose}{be verbose}
}
\value{
giotto object or data.table with overlapping information
}
\description{
calculate overlap between cellular structures (polygons) and images (intensities)
}
\concept{overlap}
