% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_structures.R
\name{calculateOverlapRaster}
\alias{calculateOverlapRaster}
\title{calculateOverlapRaster}
\usage{
calculateOverlapRaster(
  gobject,
  name_overlap,
  spatial_info,
  poly_ID_names = NULL,
  feat_info,
  return_gobject,
  verbose = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{name_overlap}{name for the overlap results (default to feat_info parameter)}

\item{feat_info}{feature information}

\item{return_gobject}{return giotto object (default: TRUE)}

\item{verbose}{be verbose}

\item{poly_info}{polygon information}
}
\value{
giotto object or spatVector with overlapping information
}
\description{
calculate overlap between cellular structures (polygons) and features (points)
}
\details{
Serial overlapping function.
}
\keyword{overlap}
