% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_structures.R
\name{calculateOverlapSerial}
\alias{calculateOverlapSerial}
\title{calculateOverlapSerial}
\usage{
calculateOverlapSerial(
  gobject,
  name_overlap = NULL,
  spatial_info = "cell",
  feat_info = "rna",
  poly_ID_names = "all",
  polygon_group_size = 500,
  return_gobject = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{gobject}{giotto object}

\item{name_overlap}{name for the overlap results (default to feat_info parameter)}

\item{feat_info}{feature information}

\item{return_gobject}{return giotto object (default: TRUE)}

\item{verbose}{be verbose}

\item{poly_info}{polygon information}
}
\value{
giotto object or spatVector with overlapping information
}
\description{
calculate overlap between cellular structures (polygons) and features (points)
}
\details{
Serial overlapping function.
}
\keyword{overlap}
