% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{calculateSpatCellMetadataProportions}
\alias{calculateSpatCellMetadataProportions}
\title{calculateSpatCellMetadataProportions}
\usage{
calculateSpatCellMetadataProportions(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  spat_network = NULL,
  metadata_column = NULL,
  name = "proportion",
  return_gobject = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit}

\item{feat_type}{feature type}

\item{spat_network}{spatial network}

\item{metadata_column}{metadata column to use}

\item{name}{descriptve name for the calculated proportions}

\item{return_gobject}{return giotto object}

\item{metadata_cols}{annotation columns found in \code{pDataDT(gobject)}}
}
\value{
giotto object (default) or enrichment object if return_gobject = FALSE
}
\description{
calculates a proportion table for a cell metadata column (e.g. cluster labels)
for all the spatial neighbors of a source cell. In other words it calculates the
niche composition for a given annotation for each cell.
}
