% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spdep.R
\name{callSpdep}
\alias{callSpdep}
\title{Call the spdep function with required parameters}
\usage{
callSpdep(method, ...)
}
\arguments{
\item{method}{Specify method name to call from spdep with its required
parameters.}

\item{...}{Additional parameters for the function. See spdep documentation 
for relevant parameters.}
}
\value{
Computed statistics from the specified method.
}
\description{
Call the spdep function with required parameters
}
\seealso{
\pkg{\link{spdep}}
}
