% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction_spot.R
\name{cell_proximity_spots_external}
\alias{cell_proximity_spots_external}
\title{cell_proximity_spots_external}
\usage{
cell_proximity_spots_external(pairs, dwls_values)
}
\arguments{
\item{pairs}{data.table of paired spots. Format: cell_ID1, cell_ID2, N}

\item{dwls_values}{data.table of cell type enrichment in each spot and multiply
by cell number in each spot}
}
\value{
List of cell proximity observed value in data.table format. Columns:
unified_int, internal value, type_int.
}
\description{
Compute cell-cell interactions observed value for interacted spots
}
