% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction_spot.R
\name{cell_proximity_spots_internal}
\alias{cell_proximity_spots_internal}
\title{cell_proximity_spots_internal}
\usage{
cell_proximity_spots_internal(cell_IDs, dwls_values)
}
\arguments{
\item{cell_IDs}{cell_IDs}

\item{dwls_values}{data.table of cell type enrichment in each spot and multiply
by cell number in each spot}
}
\value{
List of cell proximity observed value in data.table format. Columns:
unified_int, internal value, type_int.
}
\description{
Compute cell-cell interactions observed value inner each spot
}
\keyword{internal}
