% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction.R
\name{cell_proximity_tests}
\alias{cell_proximity_tests}
\alias{.do_ttest}
\alias{.do_limmatest}
\alias{.do_wilctest}
\alias{.do_permuttest}
\title{Interaction changed features test methods}
\usage{
.do_ttest(
  expr_values,
  select_ind,
  other_ind,
  adjust_method,
  mean_method,
  offset = 0.1
)

.do_limmatest(expr_values, select_ind, other_ind, mean_method, offset = 0.1)

.do_wilctest(
  expr_values,
  select_ind,
  other_ind,
  adjust_method,
  mean_method,
  offset = 0.1
)

.do_permuttest(
  expr_values,
  select_ind,
  other_ind,
  n_perm = 1000,
  adjust_method = "fdr",
  mean_method,
  offset = 0.1,
  set_seed = TRUE,
  seed_number = 1234
)
}
\value{
cell proximity values
}
\description{
Perform specified test on subsets of a matrix
}
\section{Functions}{
\itemize{
\item \code{.do_ttest()}: t.test

\item \code{.do_limmatest()}: limma t.test

\item \code{.do_wilctest()}: wilcoxon

\item \code{.do_permuttest()}: random permutation

}}
\keyword{internal}
