% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction_visuals.R
\name{cellProximityBarplot}
\alias{cellProximityBarplot}
\title{cellProximityBarplot}
\usage{
cellProximityBarplot(CPscore, min_orig_ints = 5, min_sim_ints = 5,
  p_val = 0.05)
}
\arguments{
\item{CPscore}{CPscore, output from cellProximityEnrichment()}

\item{min_orig_ints}{filter on minimum original cell-cell interactions}

\item{min_sim_ints}{filter on minimum simulated cell-cell interactions}

\item{p_val}{p-value}
}
\value{
ggplot barplot
}
\description{
Create barplot from cell-cell proximity scores
}
\details{
This function creates a barplot that shows the  spatial proximity
 enrichment or depletion of cell type pairs.
}
\examples{
    cellProximityBarplot(CPscore)
}
