% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction_spot.R
\name{cellProximityEnrichmentEachSpot}
\alias{cellProximityEnrichmentEachSpot}
\title{cellProximityEnrichmentEachSpot}
\usage{
cellProximityEnrichmentEachSpot(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  spatial_network_name = "spatial_network",
  cluster_column = "cell_ID"
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit (e.g. 'cell')}

\item{feat_type}{feature type (e.g. 'rna')}

\item{spatial_network_name}{name of spatial network to use}

\item{cluster_column}{name of column to use for clusters}
}
\value{
matrix that rownames are cell-cell interaction pairs and colnames 
are cell_IDs
}
\description{
Compute cell-cell interaction enrichment for each spot with its
interacted spots (observed)
}
\examples{
g <- GiottoData::loadGiottoMini("visium")
x <- findMarkers_one_vs_all(g, 
cluster_column = "leiden_clus", min_feats = 20)
sign_gene <- x$feats

sign_matrix <- matrix(rnorm(length(sign_gene)*8, mean  = 10), 
nrow = length(sign_gene))
rownames(sign_matrix) <- sign_gene
colnames(sign_matrix) <- paste0("celltype_",unique(x$cluster))

g <- runDWLSDeconv(gobject = g, sign_matrix = sign_matrix)

cellProximityEnrichmentEachSpot(gobject = g)
}
