% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction_visuals.R
\name{cellProximityHeatmap}
\alias{cellProximityHeatmap}
\title{cellProximityBarplot}
\usage{
cellProximityHeatmap(CPscore, scale = T, order_cell_types = T)
}
\arguments{
\item{CPscore}{CPscore, output from cellProximityEnrichment()}

\item{scale}{scale cell-cell proximity interaction scores}

\item{order_cell_types}{order}
}
\value{
ggplot heatmap
}
\description{
Create heatmap from cell-cell proximity scores
}
\details{
Give more details ...
}
\examples{
    cellProximityHeatmap(CPscore)
}
