% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction_visuals.R
\name{cellProximityNetwork}
\alias{cellProximityNetwork}
\title{cellProximityNetwork}
\usage{
cellProximityNetwork(CPscore, remove_self_edges = FALSE,
  color_depletion = "blue", color_enrichment = "red",
  rescale_edge_weights = TRUE, edge_weight_range_depletion = c(0.1, 1),
  edge_weight_range_enrichment = c(1, 5), layout = "Fruchterman",
  show_plot = F, return_plot = TRUE, save_plot = FALSE,
  save_param = list(...))
}
\arguments{
\item{CPscore}{CPscore, output from cellProximityEnrichment()}

\item{remove_self_edges}{remove enrichment/depletion edges with itself}

\item{color_depletion}{color for depleted cell-cell interactions}

\item{color_enrichment}{color for enriched cell-cell interactions}

\item{rescale_edge_weights}{rescale edge weights (boolean)}

\item{edge_weight_range_depletion}{numerical vector of length 2 to rescale depleted edge weights}

\item{edge_weight_range_enrichment}{numerical vector of length 2 to rescale enriched edge weights}

\item{layout}{layout algorithm to use to draw nodes and edges}

\item{show_plot}{show plots}

\item{return_plot}{return ggplot object}

\item{save_plot}{directly save the plot [boolean]}

\item{save_param}{list of saving parameters from all_plots_save_function()}
}
\value{
igraph plot
}
\description{
Create network from cell-cell proximity scores
}
\details{
This function creates a network that shows the  spatial proximity
 enrichment or depletion of cell type pairs.
}
\examples{
    cellProximityNetwork(CPscore)
}
