% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction_visuals.R
\name{cellProximitySpatPlot2D}
\alias{cellProximitySpatPlot2D}
\title{cellProximitySpatPlot2D}
\usage{
cellProximitySpatPlot2D(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  spat_loc_name = NULL,
  interaction_name = NULL,
  cluster_column = NULL,
  sdimx = "sdimx",
  sdimy = "sdimy",
  cell_color = NULL,
  cell_color_code = NULL,
  color_as_factor = T,
  show_other_cells = F,
  show_network = F,
  show_other_network = F,
  network_color = NULL,
  spatial_network_name = "Delaunay_network",
  show_grid = F,
  grid_color = NULL,
  spatial_grid_name = "spatial_grid",
  coord_fix_ratio = 1,
  show_legend = T,
  point_size_select = 2,
  point_select_border_col = "black",
  point_select_border_stroke = 0.05,
  point_size_other = 1,
  point_alpha_other = 0.3,
  point_other_border_col = "lightgrey",
  point_other_border_stroke = 0.01,
  show_plot = NULL,
  return_plot = NULL,
  save_plot = NULL,
  save_param = list(),
  default_save_name = "cellProximitySpatPlot2D"
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit (e.g. "cell")}

\item{feat_type}{feature type (e.g. "rna", "dna", "protein")}

\item{spat_loc_name}{spatial locations to use}

\item{interaction_name}{cell-cell interaction name}

\item{cluster_column}{cluster column with cell clusters}

\item{sdimx}{x-axis dimension name (default = 'sdimx')}

\item{sdimy}{y-axis dimension name (default = 'sdimy')}

\item{cell_color}{character. what to color cells by (e.g. metadata col or
spatial enrichment col)}

\item{cell_color_code}{character. discrete colors to use. palette to use or
named vector of colors}

\item{color_as_factor}{logical. convert color column to factor. discrete colors
are used when this is TRUE. continuous colors when FALSE.}

\item{show_other_cells}{decide if show cells not in network}

\item{show_network}{show spatial network of selected cells}

\item{show_other_network}{show spatial network of not selected cells}

\item{network_color}{color of spatial network}

\item{spatial_network_name}{name of spatial network to use}

\item{show_grid}{show spatial grid}

\item{grid_color}{color of spatial grid}

\item{spatial_grid_name}{name of spatial grid to use}

\item{coord_fix_ratio}{fix ratio between x and y-axis}

\item{show_legend}{logical. show legend}

\item{point_size_select}{size of selected points}

\item{point_select_border_col}{border color of selected points}

\item{point_select_border_stroke}{stroke size of selected points}

\item{point_size_other}{size of other points}

\item{point_alpha_other}{opacity of other points}

\item{point_other_border_col}{border color of other points}

\item{point_other_border_stroke}{stroke size of other points}

\item{show_plot}{logical. show plot}

\item{return_plot}{logical. return ggplot object}

\item{save_plot}{logical. save the plot}

\item{save_param}{list of saving parameters, see \code{\link{showSaveParameters}}}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}
}
\value{
ggplot
}
\description{
Visualize 2D cell-cell interactions according to spatial coordinates in ggplot mode
}
\details{
Description of parameters.
}
