% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction_visuals.R
\name{cellProximityVisPlot}
\alias{cellProximityVisPlot}
\title{cellProximityVisPlot}
\usage{
cellProximityVisPlot(gobject, interaction_name = NULL,
  cluster_column = NULL, sdimx = NULL, sdimy = NULL, sdimz = NULL,
  cell_color = NULL, cell_color_code = NULL, color_as_factor = T,
  show_network = F, network_color = NULL,
  spatial_network_name = "spatial_network", show_grid = F,
  grid_color = NULL, spatial_grid_name = "spatial_grid",
  coord_fix_ratio = 0.6, show.legend = T, point_size_select = 2,
  point_select.border.col = "black", point_select.border.stroke = 0.05,
  point_size_other = 1, point_other.border.col = "lightgrey",
  point_other.border.stroke = 0.01, ...)
}
\arguments{
\item{gobject}{giotto object}

\item{interaction_name}{cell-cell interaction name}

\item{cluster_column}{cluster column with cell clusters}

\item{sdimx}{x-axis dimension name (default = 'sdimx')}

\item{sdimy}{y-axis dimension name (default = 'sdimy')}

\item{sdimz}{z-axis dimension name (default = 'sdimz')}

\item{cell_color}{color for cells (see details)}

\item{cell_color_code}{named vector with colors}

\item{color_as_factor}{convert color column to factor}

\item{show_network}{show underlying spatial network}

\item{network_color}{color of spatial network}

\item{spatial_network_name}{name of spatial network to use}

\item{show_grid}{show spatial grid}

\item{grid_color}{color of spatial grid}

\item{spatial_grid_name}{name of spatial grid to use}

\item{coord_fix_ratio}{fix ratio between x and y-axis}

\item{show.legend}{show legend}

\item{point_size_select}{size of selected points}

\item{point_select.border.col}{border color of selected points}

\item{point_select.border.stroke}{stroke size of selected points}

\item{point_size_other}{size of other points}

\item{point_other.border.col}{border color of other points}

\item{point_other.border.stroke}{stroke size of other points}
}
\value{
ggplot
}
\description{
Visualize cell-cell interactions according to spatial coordinates
}
\details{
Description of parameters.
}
\examples{
    cellProximityVisPlot(gobject)
}
