% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{cells-generic}
\alias{cells-generic}
\title{Cell and feature names}
\usage{
cells(x, spat_unit, ...)

features(x, feat_type, ...)

## Default S4 method for signatures:
'giotto', 'exprObj', 'spatLocsObj', 'cellMetaObj', 'spatialNetworkObj' 'dimObj'
}
\arguments{
\item{x}{an object}

\item{spat_unit}{(optional) specify which spatial unit}

\item{...}{additional parameters to pass}

\item{feat_type}{(optional) specify which feature type}
}
\description{
Get the cells and feature names of a giotto object or subobject
}
