% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{centroids-generic}
\alias{centroids-generic}
\alias{centroids}
\alias{centroids,giottoPolygon-method}
\title{centroids-generic}
\usage{
\S4method{centroids}{giottoPolygon}(x)
}
\arguments{
\item{x}{object}
}
\description{
Access centroids information from polygon objects
}
\details{
For giottoPolygon, if centroids already exist, pulls from
\code{spatVectorCentroids} slot. Otherwise, generates from
\code{spatVector} slot de novo
}
