% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto.R
\name{changeGiottoInstructions}
\alias{changeGiottoInstructions}
\title{Change giotto instruction(s) associated with giotto object}
\usage{
changeGiottoInstructions(
  gobject,
  params = NULL,
  new_values = NULL,
  return_gobject = TRUE,
  init_gobject = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{params}{parameter(s) to change}

\item{new_values}{new value(s) for parameter(s)}

\item{return_gobject}{(boolean, default = TRUE) return giotto object}

\item{init_gobject}{(boolean, default = TRUE) initialize gobject if returning}
}
\value{
giotto object with one or more changed instructions
}
\description{
Function to change one or more instructions from giotto object.
If more than one item is supplied to \code{params} and \code{new_values}, use
a vector of values. Does not call \code{initialize} on the giotto object
}
