% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{changeImageBg}
\alias{changeImageBg}
\title{changeImageBg}
\usage{
changeImageBg(
  mg_object,
  bg_color,
  perc_range = 10,
  new_color = "#FFFFFF",
  new_name = NULL
)
}
\arguments{
\item{mg_object}{magick image or giotto image object}

\item{bg_color}{estimated current background color}

\item{perc_range}{range around estimated background color to include (percentage)}

\item{new_color}{new background color}

\item{new_name}{change name of Giotto image}
}
\value{
magick image or giotto image object with updated background color
}
\description{
Function to change the background color of a magick image plot to another color
}
