% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{check_cell_meta_obj}
\alias{check_cell_meta_obj}
\title{Check cell metadata object}
\usage{
check_cell_meta_obj(object)
}
\arguments{
\item{object}{S4 cellMetaObj to check}
}
\description{
Function to check S4 cellMetaObj
}
\keyword{internal}
