% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{check_expr_obj}
\alias{check_expr_obj}
\title{Check exprObj}
\usage{
check_expr_obj(object)
}
\arguments{
\item{object}{S4 exprObj to check}
}
\description{
Check function for S4 exprObj
}
\keyword{internal}
