% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{check_feat_meta_obj}
\alias{check_feat_meta_obj}
\title{Check feature metadata object}
\usage{
check_feat_meta_obj(object)
}
\arguments{
\item{object}{S4 featMetaObj to check}
}
\description{
Function to check S4 featMetaObj
}
\keyword{internal}
