% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_help.R
\name{check_github_suite_ver}
\alias{check_github_suite_ver}
\title{Check for updates to Giotto Suite}
\usage{
check_github_suite_ver()
}
\description{
Checks the Giotto Suite github repository and compares the version
number to the currently installed.
}
\keyword{internal}
