% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto.R
\name{check_prov_match}
\alias{check_prov_match}
\title{Check provenance info matches across list of S4 subobjects}
\usage{
check_prov_match(s4_list)
}
\description{
Check provenance info matches across list of S4 subobjects
}
\keyword{internal}
