% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interoperability.R
\name{check_py_for_scanpy}
\alias{check_py_for_scanpy}
\title{Check Scanpy Installation}
\usage{
check_py_for_scanpy()
}
\description{
checks current python environment for scanpy 1.9.0
}
\keyword{internal}
