% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{check_spat_grid_obj}
\alias{check_spat_grid_obj}
\title{Check spatialGridObj}
\usage{
check_spat_grid_obj(object)
}
\arguments{
\item{object}{S4 spatialGridObj to check}
}
\description{
Check function for S4 spatialGridObj
}
\keyword{internal}
