% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{check_spat_locs_obj}
\alias{check_spat_locs_obj}
\title{Check spatLocsObj}
\usage{
check_spat_locs_obj(object)
}
\arguments{
\item{object}{S4 spatLocsObj to check}
}
\description{
Check function for S4 spatLocsObj
}
\keyword{internal}
