% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{check_spat_net_obj}
\alias{check_spat_net_obj}
\title{Check spatialNetworkObj}
\usage{
check_spat_net_obj(object)
}
\arguments{
\item{object}{S4 spatialNetworkObj to check}
}
\description{
Check function for S4 spatialNetworkObj
}
\keyword{internal}
