% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto.R
\name{checkAggExprObj}
\alias{checkAggExprObj}
\title{Check aggregateExprObj}
\usage{
checkAggExprObj(object)
}
\arguments{
\item{object}{S4 aggregateExprObj to check}
}
\description{
Check function for S4 aggregateExprObj
}
\keyword{internal}
