% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_hmrf.R
\name{checkAndFixSpatialGenes}
\alias{checkAndFixSpatialGenes}
\title{checkAndFixSpatialGenes}
\usage{
checkAndFixSpatialGenes(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  use_spatial_genes,
  use_score = FALSE
)
}
\arguments{
\item{gobject}{Giotto object}

\item{spat_unit}{spatial unit}

\item{feat_type}{feature type}

\item{use_spatial_genes}{test name of spatial gene set to check}

\item{use_score}{logical variable to select silhouetteRank score}
}
\value{
character
}
\description{
function to check the selected test name for spatial gene set
in Giotto object
}
\details{
This function checks the user specified test name of spatial gene set in
Giotto object.
SilhouetteRank works only with score, and SilhouetteRankTest works only
with pval. Use parameter use_score to specify.
}
\keyword{internal}
