% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_structures.R
\name{checkSpatGridObj}
\alias{checkSpatGridObj}
\title{Check spatialGridObj}
\usage{
checkSpatGridObj(object)
}
\arguments{
\item{object}{S4 spatialGridObj to check}
}
\description{
Check function for S4 spatialGridObj
}
\keyword{internal}
