% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_structures.R
\name{checkSpatNetObj}
\alias{checkSpatNetObj}
\title{Check spatialNetworkObj}
\usage{
checkSpatNetObj(object)
}
\arguments{
\item{object}{S4 spatialNetworkObj to check}
}
\description{
Check function for S4 spatialNetworkObj
}
\keyword{internal}
