% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_hmrf.R
\name{chooseAvailableSpatialGenes}
\alias{chooseAvailableSpatialGenes}
\title{chooseAvailableSpatialGenes}
\usage{
chooseAvailableSpatialGenes(gobject)
}
\arguments{
\item{gobject}{TBD}
}
\description{
chooseAvailableSpatialGenes
}
