% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_structures.R
\name{circleVertices}
\alias{circleVertices}
\title{Generate circle polygon vertices}
\usage{
circleVertices(radius, npoints = 25)
}
\arguments{
\item{radius}{radius of circle to be drawn}

\item{npoints}{number of vertices to generate}
}
\value{
a data.table of circle vertices
}
\description{
Generates vertex coordinates for a circle around (0,0) with the
given radius. Modified from \pkg{packcircles}.
}
\seealso{
polyStamp rectVertices hexVertices
}
