% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_enrichment.R
\name{cluster_enrich_analysis}
\alias{cluster_enrich_analysis}
\title{cluster_enrich_analysis}
\usage{
cluster_enrich_analysis(exp_matrix, cluster_info, enrich_sig_matrix)
}
\description{
Rui to fill in
}
\keyword{internal}
