% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clusterData}
\alias{clusterData}
\alias{clusterData,allMatrix,BlusterParam-method}
\alias{clusterData,allMatrix,NNClusParam-method}
\alias{clusterData,exprObj,BlusterParam-method}
\alias{clusterData,dimObj,BlusterParam-method}
\alias{clusterData,spatEnrObj,BlusterParam-method}
\alias{clusterData,nnNetObj,NNClusParam-method}
\alias{clusterData,igraph,LouvainCommunityClusParam-method}
\alias{clusterData,igraph,LouvainMultinetClusParam-method}
\alias{clusterData,igraph,LeidenPythonClusParam-method}
\alias{clusterData,igraph,LeidenIgraphClusParam-method}
\alias{clusterData,gAny,BlusterParam-method}
\alias{clusterData,gAny,NNClusParam-method}
\title{Data clustering}
\usage{
\S4method{clusterData}{allMatrix,BlusterParam}(
  x,
  param,
  cell_subset = NULL,
  feat_subset = NULL,
  cells_are_rows = TRUE,
  set_seed = TRUE,
  seed_number = 1234,
  output = c("data.table", "factor", "full"),
  ...
)

\S4method{clusterData}{allMatrix,NNClusParam}(x, param, ...)

\S4method{clusterData}{exprObj,BlusterParam}(
  x,
  param,
  cell_subset = NULL,
  feat_subset = NULL,
  set_seed = TRUE,
  seed_number = 1234,
  output = c("data.table", "factor", "full"),
  ...
)

\S4method{clusterData}{dimObj,BlusterParam}(
  x,
  param,
  cell_subset = NULL,
  feat_subset = 1:10,
  set_seed = TRUE,
  seed_number = 1234,
  output = c("data.table", "factor", "full"),
  ...
)

\S4method{clusterData}{spatEnrObj,BlusterParam}(
  x,
  param,
  cell_subset = NULL,
  feat_subset = NULL,
  set_seed = TRUE,
  seed_number = 1234,
  output = c("data.table", "factor", "full"),
  ...
)

\S4method{clusterData}{nnNetObj,NNClusParam}(
  x,
  param,
  set_seed = TRUE,
  seed_number = 1234,
  output = c("data.table", "factor"),
  ...
)

\S4method{clusterData}{igraph,LouvainCommunityClusParam}(
  x,
  param,
  set_seed = TRUE,
  seed_number = 1234,
  output = c("data.table", "factor")
)

\S4method{clusterData}{igraph,LouvainMultinetClusParam}(
  x,
  param,
  set_seed = TRUE,
  seed_number = 1234,
  output = c("data.table", "factor")
)

\S4method{clusterData}{igraph,LeidenPythonClusParam}(
  x,
  param,
  set_seed = TRUE,
  seed_number = 1234,
  output = c("data.table", "factor")
)

\S4method{clusterData}{igraph,LeidenIgraphClusParam}(
  x,
  param,
  set_seed = TRUE,
  seed_number = 1234,
  output = c("data.table", "factor"),
  ...
)

\S4method{clusterData}{gAny,BlusterParam}(
  x,
  param,
  what = c("expression", "dimension_reduction", "spatial_enrichment"),
  name = paste(class(param), "clusters", sep = "_"),
  dim_reduction_to_use = "pca",
  dim_reduction_name = "pca",
  expression_values = "normalized",
  spat_enr_name = NULL,
  spat_unit = NULL,
  feat_type = NULL,
  output = c("gobject", "data.table", "factor", "full"),
  .n = 1L,
  ...
)

\S4method{clusterData}{gAny,NNClusParam}(
  x,
  param,
  name = paste(class(param), "clusters", sep = "_"),
  nn_network_to_use = "sNN",
  network_name = "sNN.pca",
  spat_unit = NULL,
  feat_type = NULL,
  output = c("gobject", "data.table", "factor"),
  .n = 1L,
  ...
)
}
\arguments{
\item{x}{data object to cluster}

\item{param}{S4 param class inheriting from \link[bluster:BlusterParam-class]{bluster::BlusterParam} or
\link[=NNClusParam-class]{NNClusParam}}

\item{cell_subset}{numeric, logical, factor, or character vector of cells to
use. (matrix-like clustering only)}

\item{feat_subset}{numeric, logical, factor, or character vector of features
or dimensions to use. (matrix-like clustering only)}

\item{cells_are_rows}{logical. Transposes the input matrix when \code{FALSE}.
Mainly here to highlight that clustering is performed on row-wise cells.}

\item{set_seed}{logical, default = \code{TRUE}. Whether to set a seed for
reproducibility.}

\item{seed_number}{numeric, default = 1234. If \code{set_seed = TRUE}, seed number
to use.}

\item{output}{character. Type of output.
\itemize{
\item \code{"factor"} returns a numeric factor
of clustering results, named by cell_ID.
\item \code{"data.table"} returns a two column
(or one if no cell_IDs are present) with columns \code{"cell_ID"} and \code{"cluster"}.
\item \code{"full"} returns the full output. It is equivalent to setting \code{full = TRUE}
for \code{\link[bluster:clusterRows]{bluster::clusterRows()}}.
\item \code{"gobject"} can only be selected when \code{x} is a \code{giotto} object. The
clustering results are appended to the cell metadata column designated by
\code{name} param and the \code{giotto} object is returned.
}}

\item{\dots}{additional params to pass to underlying algorithms, if any.}

\item{what}{character. One of \code{"expression"}, \code{"dimension_reduction"}, or
\code{"spatial_enrichment"} indicating what data in \code{giotto} object to cluster.
(matrix-like clustering only)}

\item{name}{character. When \code{output = "gobject"}, the name to assign the
column of clustering results in cell metadata.}

\item{dim_reduction_to_use}{character. Type of dimension reduction to use.}

\item{dim_reduction_name}{character. Name of dimension reduction results to
use.}

\item{expression_values}{character. Name of expression values to use}

\item{spat_enr_name}{character. Name of spatial enrichment values to use.}

\item{spat_unit}{character. Spatial unit to use. NULL will autodetect based
on active spatial unit.}

\item{feat_type}{character. Feature type to use. NULL will autodetect based
on active feature type.}

\item{.n}{integer. Used in recording object history
(see \code{\link[GiottoClass:update_giotto_params]{update_giotto_params()}}). Number of stack frames back to record call
of interest. Set to a large negative number (e.g. -1000) to avoid recording
altogether (useful when there is another record call elsewhere in the stack
you want to use.)}

\item{nn_network_to_use}{character. Type of nearest neighbor network to use.
Currently one of \code{"sNN"} or \code{"kNN"}.}

\item{network_name}{character. Name of nearest neighbor network to use.}
}
\description{
\pkg{Giotto} provides several options for clustering using the
\code{\link[GiottoClass:clusterData]{GiottoClass::clusterData()}} generic. Clustering for matrix-like
data hooks into the \pkg{bluster} framework. Network community clustering
type methods instead use Giotto-specific implementations that work on top
of Giotto's calculated nearest neighbor networks.
}
\examples{
x <- clusterParam("kmeans", centers = 2)
x@centers

m <- matrix(runif(9), nrow = 3)
clusterData(m, x)
# add ids
rownames(m) <- paste("id", seq_len(3), sep = "_")
clusterData(m, x)
}
\seealso{
\code{\link[=clusterParam]{clusterParam()}} for a convenient factory function to generate
clustering param classes.
}
