% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_genes.R
\name{clusterSpatialCorFeats}
\alias{clusterSpatialCorFeats}
\title{clusterSpatialCorFeats}
\usage{
clusterSpatialCorFeats(
  spatCorObject,
  name = "spat_clus",
  hclust_method = "ward.D",
  k = 10,
  return_obj = TRUE
)
}
\arguments{
\item{spatCorObject}{spatial correlation object}

\item{name}{name for spatial clustering results}

\item{hclust_method}{method for hierarchical clustering}

\item{k}{number of clusters to extract}

\item{return_obj}{return spatial correlation object (spatCorObject)}
}
\value{
spatCorObject or cluster results
}
\description{
Cluster based on spatially correlated features
}
\examples{
g <- GiottoData::loadGiottoMini("visium")

clusterSpatialCorFeats(spatCorObject = detectSpatialCorFeats(
    g,
    method = "network"
))
}
