% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_genes.R
\name{clusterSpatialCorGenes}
\alias{clusterSpatialCorGenes}
\title{clusterSpatialCorGenes}
\usage{
clusterSpatialCorGenes(
  spatCorObject,
  name = "spat_clus",
  hclust_method = "ward.D",
  k = 10,
  return_obj = TRUE
)
}
\arguments{
\item{spatCorObject}{spatial correlation object}

\item{name}{name for spatial clustering results}

\item{hclust_method}{method for hierarchical clustering}

\item{k}{number of clusters to extract}

\item{return_obj}{return spatial correlation object (spatCorObject)}
}
\value{
spatCorObject or cluster results
}
\description{
Cluster based on spatially correlated genes
}
