% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{colMeans_giotto}
\alias{colMeans_giotto}
\title{colMeans_giotto}
\usage{
colMeans_giotto(mymatrix)
}
\arguments{
\item{mymatrix}{matrix object}
}
\value{
numeric vector
}
\description{
colMeans function that works with multiple matrix representations
}
