% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{color_tag}
\alias{color_tag}
\title{Colorize print text}
\usage{
color_tag()
}
\description{
Colorize print text
}
\details{
supported colors checking is modified from \pkg{cli}
\href{https://github.com/r-lib/cli/blob/HEAD/R/num-ansi-colors.R}{aab-num-ansi-colors.R}
}
\keyword{internal}
