% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{colSums_giotto}
\alias{colSums_giotto}
\title{colSums_giotto}
\usage{
colSums_giotto(mymatrix)
}
\arguments{
\item{mymatrix}{matrix object}
}
\value{
numeric vector
}
\description{
colSums function that works with multiple matrix representations
}
