% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction.R
\name{combCCcom}
\alias{combCCcom}
\title{Combine cell cell communication tables}
\usage{
combCCcom(
  spatialCC,
  exprCC,
  min_lig_nr = 3,
  min_rec_nr = 3,
  min_padj_value = 1,
  min_log2fc = 0,
  min_av_diff = 0,
  detailed = FALSE
)
}
\arguments{
\item{spatialCC}{spatial cell-cell communication scores}

\item{exprCC}{expression cell-cell communication scores}

\item{min_lig_nr}{minimum number of ligand cells}

\item{min_rec_nr}{minimum number of receptor cells}

\item{min_padj_value}{minimum adjusted p-value}

\item{min_log2fc}{minimum log2 fold-change}

\item{min_av_diff}{minimum average expression difference}

\item{detailed}{detailed option used with \code{\link{spatCellCellcom}}
(default = FALSE)}
}
\value{
combined data.table with spatial and expression communication data
}
\description{
Combine spatial and expression based cell-cell communication
data.tables
}
\examples{
g <- GiottoData::loadGiottoMini("visium")

exprCC <- exprCellCellcom(g,
    cluster_column = "leiden_clus",
    feat_set_1 = "Gm19935",
    feat_set_2 = "9630013A20Rik"
)

spatialCC <- spatCellCellcom(
    gobject = g,
    cluster_column = "leiden_clus",
    feat_set_1 = "Gm19935",
    feat_set_2 = "9630013A20Rik",
    verbose = "a lot",
    random_iter = 10
)

combCC <- combCCcom(spatialCC = spatialCC, exprCC = exprCC)
force(combCC)
}
