% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction.R
\name{combine_ints_f}
\alias{combine_ints_f}
\title{combine_ints_f}
\usage{
combine_ints_f(
  cell_int,
  all_ints,
  unif_gene_scores,
  specific_genes_1 = NULL,
  specific_genes_2 = NULL,
  min_cells = 5,
  min_fdr = 0.05,
  min_spat_diff = 0.2,
  min_log2_fc = 0.5
)
}
\arguments{
\item{cell_int}{selected cell interaction}

\item{all_ints}{all interactions}

\item{unif_gene_scores}{unif_gene_scores results}

\item{specific_genes_1}{specific source genes (see details)}

\item{specific_genes_2}{specific target genes (see details)}

\item{min_cells}{min number of cells threshold}

\item{min_spat_diff}{spatial difference threshold}

\item{min_log2_fc}{log2 fold-change threshold}

\item{min_pval}{p-value threshold}
}
\value{
Gene to gene scores in data.table format
}
\description{
function to combine gene enrichment interactions
}
