% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{combine_stack_spatVectors}
\alias{combine_stack_spatVectors}
\title{combine_stack_spatVectors}
\usage{
combine_stack_spatVectors(gobject, spat_units)
}
\description{
combines/aggregates polygons with the same cell ID from different z-stacks
}
\keyword{internal}
