% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_structures.R
\name{combineCellData}
\alias{combineCellData}
\title{combineCellData}
\usage{
combineCellData(
  gobject,
  feat_type = "rna",
  include_spat_locs = TRUE,
  spat_loc_name = "cell",
  include_poly_info = TRUE,
  poly_info = "cell"
)
}
\arguments{
\item{gobject}{giotto object}

\item{feat_type}{feature type}

\item{include_spat_locs}{include information about spatial locations}

\item{spat_loc_name}{spatial location name}

\item{include_poly_info}{include information about polygon}

\item{poly_info}{polygon information name}
}
\value{
data.table with combined spatial information
}
\description{
combine cell data information
}
\keyword{cell}
\keyword{combine}
\keyword{metadata}
