% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_structures.R
\name{combineFeatureData}
\alias{combineFeatureData}
\title{combineFeatureData}
\usage{
combineFeatureData(
  gobject,
  feat_type = NULL,
  spat_unit = NULL,
  sel_feats = NULL
)
}
\arguments{
\item{gobject}{giotto object}

\item{feat_type}{feature type}

\item{sel_feats}{selected features (default: NULL or no selection)}
}
\value{
data.table with combined spatial feature information
}
\description{
combine feature data information
}
\keyword{combine}
\keyword{feature}
\keyword{metadata}
