% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction.R
\name{combineICF}
\alias{combineICF}
\title{combineICF}
\usage{
combineICF(
  icfObject,
  selected_ints = NULL,
  selected_feats = NULL,
  specific_feats_1 = NULL,
  specific_feats_2 = NULL,
  min_cells = 5,
  min_int_cells = 3,
  min_fdr = 0.05,
  min_spat_diff = 0,
  min_log2_fc = 0.5,
  do_parallel = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{icfObject}{ICF (interaction changed feat) score object}

\item{selected_ints}{subset of selected cell-cell interactions (optional)}

\item{selected_feats}{subset of selected Feats (optional)}

\item{specific_feats_1}{specific Featset combo
(need to position match specific_genes_2)}

\item{specific_feats_2}{specific Featset combo
(need to position match specific_genes_1)}

\item{min_cells}{minimum number of target cell type}

\item{min_int_cells}{minimum number of interacting cell type}

\item{min_fdr}{minimum adjusted p-value}

\item{min_spat_diff}{minimum absolute spatial expression difference}

\item{min_log2_fc}{minimum absolute log2 fold-change}

\item{do_parallel}{run calculations in parallel with mclapply}

\item{verbose}{verbose}
}
\value{
icfObject that contains the filtered differential feats scores
}
\description{
Combine ICF scores in a pairwise manner.
}
\examples{
g <- GiottoData::loadGiottoMini("visium")
g_icf <- findInteractionChangedFeats(g, cluster_column = "leiden_clus",
selected_feats = c("Gna12", "Ccnd2", "Btbd17"), nr_permutations = 10)

combineICF(g_icf)
}
