% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{combineSpatialCellFeatureInfo}
\alias{combineSpatialCellFeatureInfo}
\title{combineSpatialCellFeatureInfo}
\usage{
combineSpatialCellFeatureInfo(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  selected_features = NULL
)
}
\arguments{
\item{gobject}{Giotto object}

\item{spat_unit}{spatial unit}

\item{feat_type}{feature type(s)}

\item{selected_features}{select set of features}
}
\value{
list of data.table(s)
}
\description{
Combine spatial cell information (e.g. polygon)
and spatial feature information (e.g. transcript locations)
}
\details{
The returned data.table has the following columns: \cr
\itemize{
  \item{sdimx: spatial feature location on the x-axis}
  \item{sdimy: spatial feature location on the y-axis}
  \item{feat_ID: unique feature ID}
  \item{cell_ID: unique cell ID}
  \item{used: how often was the feature used/assigned to a cell}
  \item{feat: selected feature(s)}
}
}
