% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{combineSpatialCellMetadataInfo}
\alias{combineSpatialCellMetadataInfo}
\title{combineSpatialCellMetadataInfo}
\usage{
combineSpatialCellMetadataInfo(gobject, feat_type = NULL)
}
\arguments{
\item{gobject}{Giotto object}

\item{feat_type}{feature type(s)}
}
\value{
list of data.table(s)
}
\description{
Combine cell metadata with spatial cell information (e.g. polygon)
}
\details{
The returned data.table has the following columns: \cr
\itemize{
  \item{sdimx: spatial feature location on the x-axis}
  \item{sdimy: spatial feature location on the y-axis}
  \item{cell_ID: unique cell ID}
  \item{feat: selected feature(s)}
  \item{other columns that are part of the cell metadata}
}
}
