% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactivity.R
\name{comparePolygonExpression}
\alias{comparePolygonExpression}
\title{Compare gene expression between polygon areas}
\usage{
comparePolygonExpression(
  gobject,
  polygon_name = "selections",
  spat_unit = "cell",
  feat_type = "rna",
  selected_feats = "top_genes",
  expression_values = "normalized",
  method = "scran",
  ...
)
}
\arguments{
\item{gobject}{A Giotto object}

\item{polygon_name}{name of polygon selections}

\item{spat_unit}{spatial unit (e.g. "cell")}

\item{feat_type}{feature type (e.g. "rna", "dna", "protein")}

\item{selected_feats}{vector of selected features to plot}

\item{expression_values}{gene expression values to use 
("normalized", "scaled", "custom")}

\item{method}{method to use to detect differentially expressed feats 
("scran", "gini", "mast")}

\item{\dots}{Arguments passed to \link[ComplexHeatmap]{Heatmap}}
}
\value{
A ComplexHeatmap::Heatmap object
}
\description{
Compare gene expression between polygon areas
}
