% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_structures.R
\name{compatible_spatial_network}
\alias{compatible_spatial_network}
\title{Compatible spatial network}
\usage{
compatible_spatial_network(spatial_network, expression_matrix)
}
\description{
Function to evaluate if a spatial network is compatible
with a provided expression matrix
}
\keyword{internal}
