% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{convert_mgImage_to_array_DT}
\alias{convert_mgImage_to_array_DT}
\title{convert_mgImage_to_array_DT}
\usage{
convert_mgImage_to_array_DT(mg_object)
}
\arguments{
\item{mg_object}{magick image or Giotto image object}
}
\value{
data.table with image pixel information
}
\description{
converts a magick image object to a data.table
}
\keyword{internal}
