% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_structures.R
\name{convert_to_reduced_spatial_network}
\alias{convert_to_reduced_spatial_network}
\title{convert_to_reduced_spatial_network}
\usage{
convert_to_reduced_spatial_network(full_spatial_network_DT)
}
\arguments{
\item{full_spatial_network_DT}{full spatial network in data.table format}
}
\description{
convert to a reduced spatial network
}
\keyword{internal}
