% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_help.R
\name{convertEnsemblToGeneSymbol}
\alias{convertEnsemblToGeneSymbol}
\title{convertEnsemblToGeneSymbol}
\usage{
convertEnsemblToGeneSymbol(matrix, species = c("mouse", "human"))
}
\arguments{
\item{matrix}{an expression matrix with ensembl gene IDs as rownames}

\item{species}{species to use for gene symbol conversion}
}
\value{
expression matrix with gene symbols as rownames
}
\description{
This function convert ensembl gene IDs from a matrix to official gene symbols
}
\details{
This function requires that the biomaRt library is installed
}
