% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{convertGiottoLargeImageToMG}
\alias{convertGiottoLargeImageToMG}
\title{convertGiottoLargeImageToMG}
\usage{
convertGiottoLargeImageToMG(
  gobject = NULL,
  largeImage_name = NULL,
  giottoLargeImage = NULL,
  mg_name = NULL,
  spat_unit = NULL,
  spat_loc_name = NULL,
  crop_extent = NULL,
  xmax_crop = NULL,
  xmin_crop = NULL,
  ymax_crop = NULL,
  ymin_crop = NULL,
  resample_size = 5e+05,
  max_intensity = NULL,
  return_gobject = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{gobject}{gobject containing giottoLargeImage}

\item{largeImage_name}{name of giottoLargeImage}

\item{mg_name}{name to assign converted magick image based giottoImage. Defaults to name of giottoLargeImage}

\item{spat_unit}{spatial unit}

\item{spat_loc_name}{gobject spatial location name to map giottoImage to (optional)}

\item{crop_extent}{extent object to focus on specific region of image}

\item{xmax_crop}{assign crop boundary}

\item{xmin_crop}{assign crop boundary}

\item{ymax_crop}{assign crop boundary}

\item{ymin_crop}{assign crop boundary}

\item{resample_size}{maximum number of pixels to use when resampling}

\item{max_intensity}{value to treat as maximum intensity in color scale}

\item{return_gobject}{return as giotto object}

\item{verbose}{be verbose}
}
\value{
a giotto object or an updated giotto image if return_gobject = F
}
\description{
convert a giottoLargeImage by downsampling into a normal magick based giottoImage
}
