% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_enrichment.R
\name{convertSignListToMatrix}
\alias{convertSignListToMatrix}
\title{convertSignListToMatrix}
\usage{
convertSignListToMatrix(sign_names, sign_list)
}
\arguments{
\item{sign_names}{vector with names for each provided gene signature}

\item{sign_list}{list of gene signatures}
}
\value{
matrix
}
\description{
Function to convert list of signature genes (e.g. for cell types) into
a binary matrix format that can be used with the PAGE enrichment option.
}
\examples{
    convertSignListToMatrix()
}
