% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{coordDataDT-class}
\alias{coordDataDT-class}
\title{Basic class for classes with coordinate information}
\description{
coordDataDT is the specific flavor that deals with objects where the coordinate
information is stored within data.table objects and should work similarly to
data.table when interacting with some basic generic operators for data
retreival and setting.
}
