% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{copy-generic}
\alias{copy-generic}
\alias{copy}
\alias{copy,spatialLocationsObj-method}
\alias{copy,data.table-method}
\title{Copy an entire object}
\usage{
copy(x)

\S4method{copy}{spatialLocationsObj}(x)

\S4method{copy}{data.table}(x)
}
\arguments{
\item{x}{a Giotto S4 class subobject}
}
\description{
S4 generic for Giotto's S4 subobjects to return with full copies of
certain subobjects that usually return referenced information.
}
\section{Functions}{
\itemize{
\item \code{copy(spatialLocationsObj)}: Copy \emph{data.table}-based spatial locations object.

\item \code{copy(data.table)}: See \code{\link[data.table]{copy}}

}}
\seealso{
\code{\link[data.table]{copy}} \code{\link[terra]{deepcopy}}
}
