% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction.R
\name{create_cell_type_random_cell_IDs}
\alias{create_cell_type_random_cell_IDs}
\title{create_cell_type_random_cell_IDs}
\usage{
create_cell_type_random_cell_IDs(gobject, cluster_column = "cell_types",
  needed_cell_types)
}
\arguments{
\item{gobject}{giotto object to use}

\item{cluster_column}{cluster column with cell type information}

\item{needed_cell_types}{vector of cell type names for which a random id will be found}
}
\value{
list of randomly sampled cell ids with same cell type composition
}
\description{
creates randomized cell ids within a selection of cell types
}
\details{
Details will follow.
}
\examples{
    create_cell_type_random_cell_IDs(gobject)
}
