% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{create_cluster_matrix}
\alias{create_cluster_matrix}
\title{create_cluster_matrix}
\usage{
create_cluster_matrix(
  gobject,
  expression_values = c("normalized", "scaled", "custom"),
  cluster_column,
  gene_subset = NULL
)
}
\description{
creates aggregated matrix for a given clustering column
}
\keyword{internal}
