% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_section.R
\name{create_crossSection_object}
\alias{create_crossSection_object}
\title{create_crossSection_object}
\usage{
create_crossSection_object(
  name = NULL,
  method = NULL,
  thickness_unit = NULL,
  slice_thickness = NULL,
  plane_equation = NULL,
  mesh_grid_n = NULL,
  mesh_obj = NULL,
  cell_subset = NULL,
  cell_subset_spatial_locations = NULL,
  cell_subset_projection_locations = NULL,
  cell_subset_projection_PCA = NULL,
  cell_subset_projection_coords = NULL
)
}
\description{
create a crossSection object
}
