% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_structures.R
\name{create_delaunayNetwork2D_v2}
\alias{create_delaunayNetwork2D_v2}
\title{create_delaunayNetwork2D_v2}
\usage{
create_delaunayNetwork2D_v2(
  gobject,
  method = c("delaunayn_geometry", "RTriangle", "deldir"),
  sdimx = "sdimx",
  sdimy = "sdimy",
  name = "delaunay_network",
  maximum_distance = "auto",
  minimum_k = 0,
  options = "Pp",
  Y = TRUE,
  j = TRUE,
  S = 0,
  verbose = T,
  return_gobject = TRUE,
  ...
)
}
\description{
Create a spatial Delaunay network.
}
\examples{
    create_delaunayNetwork2D_v2(gobject)
}
