% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_structures.R
\name{create_delaunayNetwork3D}
\alias{create_delaunayNetwork3D}
\title{create_delaunayNetwork3D}
\usage{
create_delaunayNetwork3D(
  gobject,
  method = "delaunayn_geometry",
  dimensions = c("sdimx", "sdimy", "sdimz"),
  name = "delaunay_network_3D",
  maximum_distance = "auto",
  return_gobject = TRUE,
  ...
)
}
\description{
create delaunay network for 3D data
}
