% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_structures.R
\name{create_delaunayNetwork_geometry}
\alias{create_delaunayNetwork_geometry}
\title{create_delaunayNetwork_geometry}
\usage{
create_delaunayNetwork_geometry(spatial_locations)
}
\description{
create delaunay network based on package geometry
}
