% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_structures.R
\name{create_delaunayNetwork_geometry_3D_v2}
\alias{create_delaunayNetwork_geometry_3D_v2}
\title{create_delaunayNetwork_geometry_3D_v2}
\usage{
create_delaunayNetwork_geometry_3D_v2(
  spatial_locations,
  sdimx = "sdimx",
  sdimy = "sdimy",
  sdimz = "sdimz",
  ...
)
}
\description{
Create a spatial Delaunay network.
}
\examples{
    create_delaunayNetwork_geometry_3D_v2(gobject)
}
