% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_structures.R
\name{create_delaunayNetwork_RTriangle}
\alias{create_delaunayNetwork_RTriangle}
\title{create_delaunayNetwork_RTriangle}
\usage{
create_delaunayNetwork_RTriangle(
  spatial_locations,
  sdimx = "sdimx",
  sdimy = "sdimy",
  Y = TRUE,
  j = TRUE,
  S = 0,
  ...
)
}
\description{
Create a spatial Delaunay network with RTriangle
}
\keyword{internal}
