% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_structures.R
\name{create_delaunayNetwork_RTriangle_v2}
\alias{create_delaunayNetwork_RTriangle_v2}
\title{create_delaunayNetwork_RTriangle_v2}
\usage{
create_delaunayNetwork_RTriangle_v2(
  spatial_locations,
  sdimx = "sdimx",
  sdimy = "sdimy",
  Y = TRUE,
  j = TRUE,
  S = 0,
  ...
)
}
\description{
Create a spatial Delaunay network.
}
\examples{
    create_delaunayNetwork_RTriangle_v2(gobject)
}
