% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{create_dim_obj}
\alias{create_dim_obj}
\title{Create S4 dimObj}
\usage{
create_dim_obj(
  name = "test",
  reduction = "cells",
  reduction_method = NA_character_,
  coordinates = NULL,
  spat_unit = "cell",
  feat_type = "rna",
  provenance = NULL,
  misc = NULL,
  my_rownames = NULL
)
}
\arguments{
\item{name}{name of dimObj}

\item{reduction}{reduction on columns (e.g. cells) or rows (e.g. features)}

\item{reduction_method}{method used to generate dimension reduction}

\item{coordinates}{embedding coordinates}

\item{spat_unit}{spatial unit of aggregated expression (e.g. 'cell')}

\item{feat_type}{feature type of aggregated expression (e.g. 'rna', 'protein')}

\item{provenance}{origin data of aggregated expression information (if applicable)}

\item{misc}{misc}

\item{my_rownames}{(optional) if needed, set coordinates rowname values here}
}
\description{
Create an S4 dimObj
}
\keyword{internal}
