% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.R
\name{create_dimObject}
\alias{create_dimObject}
\title{create_dimObject}
\usage{
create_dimObject(
  name = "test",
  spat_unit = "cell",
  feat_type = "rna",
  reduction_method = NULL,
  coordinates = NULL,
  misc = NULL,
  my_rownames = NULL
)
}
\arguments{
\item{name}{arbitrary name for object}

\item{spat_unit}{spatial unit}

\item{feat_type}{feature type}

\item{reduction_method}{method used to reduce dimensions}

\item{coordinates}{accepts the coordinates after dimension reduction}

\item{misc}{any additional information will be added to this slot}

\item{my_rownames}{rownames}
}
\value{
number of distinct colors
}
\description{
Creates an object that stores a dimension reduction output
}
\keyword{internal}
