% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{create_expr_obj}
\alias{create_expr_obj}
\title{Create S4 exprObj}
\usage{
create_expr_obj(
  name = "test",
  exprMat = NULL,
  spat_unit = "cell",
  feat_type = "rna",
  provenance = NULL,
  misc = NULL
)
}
\arguments{
\item{name}{name of exprObj}

\item{exprMat}{matrix of expression information}

\item{spat_unit}{spatial unit of expression (e.g. 'cell')}

\item{feat_type}{feature type of expression (e.g. 'rna', 'protein')}

\item{provenance}{origin data of expression information (if applicable)}

\item{misc}{misc}
}
\description{
Create an S4 exprObj
}
\keyword{internal}
