% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{create_feat_meta_obj}
\alias{create_feat_meta_obj}
\title{Create S4 featMetaObj}
\usage{
create_feat_meta_obj(
  metaDT = NULL,
  col_desc = NA_character_,
  spat_unit = "cell",
  feat_type = "rna",
  provenance = NULL
)
}
\arguments{
\item{col_desc}{(optional) character vector describing columns of the metadata}

\item{spat_unit}{spatial unit of aggregated expression (e.g. 'cell')}

\item{feat_type}{feature type of aggregated expression (e.g. 'rna', 'protein')}

\item{provenance}{origin data of aggregated expression information (if applicable)}

\item{metadata}{metadata info}

\item{misc}{misc}
}
\description{
Create an S4 featMetaObj
}
\keyword{internal}
