% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{create_featureNetwork_object}
\alias{create_featureNetwork_object}
\title{Create feature network object}
\usage{
create_featureNetwork_object(
  name = "feat_network",
  network_datatable = NULL,
  network_lookup_id = NULL,
  full = NULL
)
}
\arguments{
\item{name}{name to assign the created feature network object}

\item{network_datatable}{network data.table object}

\item{network_lookup_id}{network lookup id}

\item{full}{fully connected status}
}
\description{
Create feature network object
}
\keyword{internal}
