% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{create_giotto_points_object}
\alias{create_giotto_points_object}
\title{Create giotto points object}
\usage{
create_giotto_points_object(
  feat_type = "rna",
  spatVector = NULL,
  networks = NULL
)
}
\arguments{
\item{feat_type}{feature type}

\item{spatVector}{terra spatVector object containing point data}

\item{networks}{feature network object}
}
\description{
Create giotto points object
}
\keyword{internal}
