% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_structures.R
\name{create_giotto_polygon_object}
\alias{create_giotto_polygon_object}
\title{Create a giotto polygon object}
\usage{
create_giotto_polygon_object(
  name = "cell",
  spatVector = NULL,
  spatVectorCentroids = NULL,
  overlaps = NULL
)
}
\description{
Create a giotto polygon object
}
\keyword{internal}
