% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_structures.R
\name{create_KNNnetwork_dbscan}
\alias{create_KNNnetwork_dbscan}
\title{create_KNNnetwork_dbscan}
\usage{
create_KNNnetwork_dbscan(
  spatial_locations,
  sdimx = "sdimx",
  sdimy = "sdimy",
  sdimz = "sdimz",
  k = 4,
  ...
)
}
\description{
Create a spatial knn network with dbscan
}
\keyword{internal}
