% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_section.R
\name{create_mesh_grid_lines}
\alias{create_mesh_grid_lines}
\title{create_mesh_grid_lines}
\usage{
create_mesh_grid_lines(
  cell_subset_projection_locations,
  extend_ratio,
  mesh_grid_n
)
}
\arguments{
\item{cell_subset_projection_locations}{cell_subset_projection_locations}

\item{extend_ratio}{extend_ratio}

\item{mesh_grid_n}{mesh_grid_n}
}
\description{
create mesh grid lines for cross section
}
\keyword{internal}
