% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{create_nn_net_obj}
\alias{create_nn_net_obj}
\title{Create S4 nnNetObj}
\usage{
create_nn_net_obj(
  name = "test",
  nn_type = NA_character_,
  igraph = NULL,
  spat_unit = "cell",
  feat_type = "rna",
  provenance = NULL,
  misc = NULL
)
}
\arguments{
\item{name}{name of nnNetObj}

\item{nn_type}{type of nearest neighbor network}

\item{igraph}{igraph object containing nearest neighbor information}

\item{misc}{misc}
}
\description{
Create an S4 nnNetObj
}
\section{Slots}{

\describe{
\item{\code{spat_unit}}{spatial unit of data}

\item{\code{feat_type}}{feature type of data}

\item{\code{provenance}}{origin of aggregated information (if applicable)}
}}

\keyword{internal}
