% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{create_pb}
\alias{create_pb}
\title{Create a txt progressbar}
\usage{
create_pb(max, min = 0, width = min(35, getOption("width")), char = "=")
}
\description{
Create a txt progressbar
}
\keyword{internal}
