% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.R
\name{create_screeplot}
\alias{create_screeplot}
\title{create_screeplot}
\usage{
create_screeplot(eigs, ncp = 20, ylim = c(0, 20))
}
\arguments{
\item{eigs}{numeric. Vector of pca eigenvalues}

\item{ncp}{numeric. max number of principal components to plot}

\item{ylim}{numeric. y-axis limits on scree plot}
}
\value{
ggplot
}
\description{
create screeplot with ggplot
}
\examples{
\dontrun{
dr <- GiottoData::loadSubObjectMini("dimObj")
scree <- create_screeplot(methods::slot(dr, "misc")$eigenvalues)
scree
}
}
