% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_structures.R
\name{create_segm_polygons}
\alias{create_segm_polygons}
\title{Create segmentation polygons}
\usage{
create_segm_polygons(
  maskfile,
  name = "cell",
  poly_IDs = NULL,
  flip_vertical = TRUE,
  shift_vertical_step = TRUE,
  flip_horizontal = TRUE,
  shift_horizontal_step = TRUE,
  remove_background_polygon = FALSE
)
}
\value{
giotto polygon
}
\description{
creates giotto polygons from segmentation mask data
}
\keyword{internal}
