% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{create_spat_enr_obj}
\alias{create_spat_enr_obj}
\title{Create S4 spatEnrObj}
\usage{
create_spat_enr_obj(
  name = "test",
  method = NA_character_,
  enrichDT = NULL,
  spat_unit = "cell",
  feat_type = "rna",
  provenance = NULL,
  misc = NULL
)
}
\arguments{
\item{name}{name of S4 spatEnrObj}

\item{method}{method used to generate spatial enrichment information}

\item{enrichDT}{spatial enrichment results, provided as a data.table}

\item{spat_unit}{spatial unit of aggregated expression (e.g. 'cell')}

\item{feat_type}{feature type of aggregated expression (e.g. 'rna', 'protein')}

\item{provenance}{origin data of aggregated expression information (if applicable)}

\item{misc}{misc additional information about he spatial enrichment or how it
was generated}
}
\description{
Create S4 spatEnrObj
}
\keyword{internal}
