% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{create_spat_grid_obj}
\alias{create_spat_grid_obj}
\title{Create S4 spatialGridObj}
\usage{
create_spat_grid_obj(
  name = "test",
  method = NA_character_,
  parameters = NULL,
  gridDT = NULL,
  spat_unit = "cell",
  feat_type = "rna",
  provenance = NULL,
  misc = NULL
)
}
\arguments{
\item{name}{name of spatialGridObj}

\item{method}{method used to generate spatial grid}

\item{parameters}{additional method-specific parameters used during spatial grid generation}

\item{gridDT}{data.table holding the spatial grid information}

\item{spat_unit}{spatial unit}

\item{feat_type}{feature type}

\item{provenance}{origin of aggregated information (if applicable)}

\item{misc}{misc}
}
\description{
Create S4 spatialGridObj
}
\keyword{internal}
