% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{create_spat_locs_obj}
\alias{create_spat_locs_obj}
\alias{createSpatLocsObj}
\title{Create S4 spatLocsObj}
\usage{
createSpatLocsObj(
  coordinates,
  name = "test",
  spat_unit = "cell",
  provenance = NULL,
  misc = NULL,
  verbose = TRUE
)
}
\arguments{
\item{coordinates}{spatial coordinates}

\item{name}{name of spatLocsObj}

\item{spat_unit}{spatial unit of aggregated expression (e.g. 'cell')}

\item{provenance}{origin data of aggregated expression information (if applicable)}

\item{misc}{misc}
}
\description{
Create an S4 spatLocsObj
}
