% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{create_spat_net_obj}
\alias{create_spat_net_obj}
\title{Create S4 spatialNetworkObj}
\usage{
create_spat_net_obj(
  name = "test",
  method = NA_character_,
  parameters = NULL,
  outputObj = NULL,
  networkDT = NULL,
  networkDT_before_filter = NULL,
  cellShapeObj = NULL,
  crossSectionObjects = NULL,
  spat_unit = "cell",
  provenance = NULL,
  misc = NULL
)
}
\arguments{
\item{name}{name of spatialNetworkObj}

\item{method}{method used to generate spatial network}

\item{parameters}{additional method-specific parameters used during spatial network generation}

\item{outputObj}{network geometry object}

\item{networkDT}{data.table of network connections, distances, and weightings}

\item{networkDT_before_filter}{unfiltered data.table  of network connections, distances, and weightings}

\item{cellShapeObj}{network cell shape information}

\item{crossSectionObjects}{crossSectionObjects (see \code{\link{create_crossSection_object}})}

\item{spat_unit}{spatial unit tag}

\item{provenance}{origin of aggregated information (if applicable)}

\item{misc}{misc}
}
\description{
Create S4 spatialNetworkObj
}
\keyword{internal}
