% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_structures.R
\name{create_spatial_grid}
\alias{create_spatial_grid}
\title{create_spatial_grid}
\usage{
create_spatial_grid(spatial_coord, sdimx_coord = NULL, sdimx_start,
  sdimx_end, sdimx_stepsize, sdimy_coord = NULL, sdimy_start, sdimy_end,
  sdimy_stepsize, sdimz_coord = NULL, sdimz_start, sdimz_end,
  sdimz_stepsize)
}
\description{
helper function to create spatial grid
}
