% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_structures.R
\name{create_spatialGrid_default_3D}
\alias{create_spatialGrid_default_3D}
\title{create_spatialGrid_default_3D}
\usage{
create_spatialGrid_default_3D(
  gobject,
  sdimx_stepsize = NULL,
  sdimy_stepsize = NULL,
  sdimz_stepsize = NULL,
  minimum_padding = 1
)
}
\description{
create a 3D spatial grid
}
\keyword{internal}
