% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_structures.R
\name{create_spatialGridObject}
\alias{create_spatialGridObject}
\title{create_spatialGridObject}
\usage{
create_spatialGridObject(
  name = NULL,
  method = NULL,
  parameters = NULL,
  gridDT = NULL,
  outputObj = NULL,
  misc = NULL
)
}
\description{
create a spatial grid object
}
\keyword{internal}
