% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{create_spatlocs_obj}
\alias{create_spatlocs_obj}
\title{Create S4 spatLocsObj}
\usage{
create_spatlocs_obj(
  name = "test",
  coordinates = NULL,
  spat_unit = "cell",
  provenance = NULL,
  misc = NULL
)
}
\arguments{
\item{name}{name of spatLocsObj}

\item{coordinates}{spatial coordinates}

\item{spat_unit}{spatial unit of aggregated expression (e.g. 'cell')}

\item{provenance}{origin data of aggregated expression information (if applicable)}

\item{misc}{misc}
}
\description{
Create an S4 spatLocsObj
}
\keyword{internal}
