% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_structures.R
\name{create_spatvector_object_from_dfr}
\alias{create_spatvector_object_from_dfr}
\title{Create terra spatvector object from a data.frame}
\usage{
create_spatvector_object_from_dfr(x)
}
\arguments{
\item{x}{data.frame object}
}
\description{
create terra spatvector from a data.frame
}
\keyword{internal}
