% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{create_terra_spatRaster}
\alias{create_terra_spatRaster}
\title{Load image as a terra spatRaster object}
\usage{
create_terra_spatRaster(image_path)
}
\arguments{
\item{image_path}{existing full filepath to image to be loaded as a terra spatRaster}
}
\description{
Load image as a terra spatRaster object
}
\keyword{internal}
