% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_bento.R
\name{createBentoAdata}
\alias{createBentoAdata}
\title{Create bento adata object from gobject}
\usage{
createBentoAdata(gobject = NULL, env_to_use = NULL)
}
\arguments{
\item{gobject}{Giotto object}

\item{env_to_use}{Python environment within which bento is installed.
If it is not already installed, the user
will be prompted to install `bento-tools`
DEFAULT: "giotto_env"}
}
\value{
bento_adata bento adata object
}
\description{
Create bento adata object from gobject
}
