% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{createDimObj}
\alias{createDimObj}
\title{Create S4 dimObj}
\usage{
createDimObj(
  coordinates,
  name = "test",
  spat_unit = "cell",
  feat_type = "rna",
  method = NULL,
  reduction = "cells",
  provenance = NULL,
  misc = NULL,
  my_rownames = NULL
)
}
\arguments{
\item{coordinates}{embedding coordinates}

\item{name}{name of dimObj}

\item{spat_unit}{spatial unit of aggregated expression (e.g. 'cell')}

\item{feat_type}{feature type of aggregated expression (e.g. 'rna', 'protein')}

\item{reduction}{reduction on columns (e.g. cells) or rows (e.g. features)}

\item{provenance}{origin data of aggregated expression information (if applicable)}

\item{misc}{misc}

\item{my_rownames}{(optional) if needed, set coordinates rowname values here}

\item{reduction_method}{method used to generate dimension reduction}
}
\description{
Create an S4 dimObj
}
