% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createEpiObject.R
\name{createEpiObject}
\alias{createEpiObject}
\title{Create epigenetic object}
\usage{
createEpiObject(
  fragmentsPath,
  genome = c("hg19", "hg38", "mm9", "mm10"),
  expression = NULL,
  expression_feat = "atac",
  spatial_locs = NULL,
  minFrags = 0,
  maxFrags = 1e+07,
  minTSS = 0,
  addTileMat = TRUE,
  addGeneScoreMat = TRUE,
  offsetPlus = 0,
  offsetMinus = 0,
  TileMatParams = list(tileSize = 5000),
  sampleName = "epiSample",
  outputDirectory = "epiSample",
  copyArrows = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{fragmentsPath}{A character vector containing the paths to the input files to use to generate the ArrowFiles. These files can be in one of the following formats: (i) scATAC tabix files, (ii) fragment files, or (iii) bam files.}

\item{genome}{A string indicating the default genome to be used for all ArchR functions. Currently supported values include "hg19","hg38","mm9", and "mm10". This value is stored as a global environment variable, not part of the ArchRProject. This can be overwritten on a per-function basis using the given function's geneAnnotationand genomeAnnotation parameter. For something other than one of the currently supported, see createGeneAnnnotation() and createGenomeAnnnotation()}

\item{expression}{expression information}

\item{expression_feat}{available features (e.g. atac, rna, ...)}

\item{spatial_locs}{data.table or data.frame with coordinates for cell centroids}

\item{minFrags}{The minimum number of mapped ATAC-seq fragments required per cell to pass filtering for use in downstream analyses. Cells containing greater than or equal to minFrags total fragments wll be retained.}

\item{maxFrags}{The maximum number of mapped ATAC-seq fragments required per cell to pass filtering for use in downstream analyses. Cells containing greater than or equal to maxFrags total fragments wll be retained.}

\item{minTSS}{The minimum numeric transcription start site (TSS) enrichment score required for a cell to pass filtering for use in downstream analyses. Cells with a TSS enrichment score greater than or equal to minTSS will be retained. TSS enrichment score is a measurement of signal-to-background in ATAC-seq.}

\item{addTileMat}{A boolean value indicating whether to add a "Tile Matrix" to each ArrowFile. A Tile Matrix is a counts matrix that, instead of using peaks, uses a fixed-width sliding window of bins across the whole genome. This matrix can be used in many downstream ArchR operations.}

\item{addGeneScoreMat}{A boolean value indicating whether to add a Gene-Score Matrix to each ArrowFile. A Gene-Score Matrix uses ATAC-seq signal proximal to the TSS to estimate gene activity.}

\item{offsetPlus}{The numeric offset to apply to a "+" stranded Tn5 insertion to account for the precise Tn5 binding site. This parameter only applies to bam file input and it is assumed that fragment files have already been offset which is the standard from 10x output. See Buenrostro et al. Nature Methods 2013.}

\item{offsetMinus}{The numeric offset to apply to a "-" stranded Tn5 insertion to account for the precise Tn5 binding site. This parameter only applies to bam file input and it is assumed that fragment files have already been offset which is the standard from 10x output. See Buenrostro et al. Nature Methods 2013.}

\item{TileMatParams}{A list of parameters to pass to the ArchR::addTileMatrix() function. See ArchR::addTileMatrix() for options.}

\item{sampleName}{sample name. Default = 'epiSample'}

\item{outputDirectory}{A name for the relative path of the outputDirectory for ArchR results. Relative to the current working directory.}

\item{copyArrows}{A boolean value indicating whether ArrowFiles should be copied into outputDirectory.}

\item{verbose}{Default = TRUE}

\item{...}{Additional params passed to createGiottoObject()}
}
\value{
A Giotto object with at least an atac or epigenetic modality
}
\description{
Create epigenetic object
}
