% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{createExprObj}
\alias{createExprObj}
\title{Create S4 exprObj}
\usage{
createExprObj(
  expression_data,
  name = "test",
  spat_unit = "cell",
  feat_type = "rna",
  provenance = NULL,
  expression_matrix_class = c("Matrix", "HDF5Matrix", "dbMatrix"),
  sparse = TRUE,
  cores = determine_cores(),
  remote_dir = ":temp:",
  db_extension = ".duckdb",
  misc = NULL,
  ...
)
}
\arguments{
\item{expression_data}{expression data}

\item{name}{name of exprObj}

\item{spat_unit}{spatial unit of expression (e.g. 'cell')}

\item{feat_type}{feature type of expression (e.g. 'rna', 'protein')}

\item{provenance}{origin data of expression information (if applicable)}

\item{cores}{number of cores to use}

\item{remote_dir}{(optional) directory of remote backend. Default is tempdir()}

\item{db_extension}{(optional) database file extension}

\item{misc}{misc}

\item{...}{additional params to pass to matrix reading functions}

\item{matrix_type}{type of matrix to use}
}
\description{
Create an S4 exprObj
}
