% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto.R
\name{createGiottoCosMxObject_aggregate}
\alias{createGiottoCosMxObject_aggregate}
\title{Load and create a CosMx Giotto object from aggregate info}
\usage{
createGiottoCosMxObject_aggregate(
  dir_items,
  cores,
  verbose = TRUE,
  instructions = NULL
)
}
\arguments{
\item{cores}{how many cores or threads to use to read data if paths are provided}

\item{verbose}{be verbose when building Giotto object}

\item{instructions}{list of instructions or output result from \code{\link{createGiottoInstructions}}}
}
\description{
Load and create a CosMx Giotto object from aggregate info
}
\keyword{internal}
