% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto.R
\name{createGiottoCosMxObject_subcellular}
\alias{createGiottoCosMxObject_subcellular}
\title{Load and create a CosMx Giotto object from subcellular info}
\usage{
createGiottoCosMxObject_subcellular(
  dir_items,
  FOVs = NULL,
  cores,
  verbose = TRUE,
  instructions = NULL
)
}
\arguments{
\item{FOVs}{field of views to load (only affects subcellular data and images)}

\item{cores}{how many cores or threads to use to read data if paths are provided}

\item{verbose}{be verbose when building Giotto object}

\item{instructions}{list of instructions or output result from \code{\link{createGiottoInstructions}}}
}
\description{
Load and create a CosMx Giotto object from subcellular info
}
\keyword{internal}
