% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto.R
\name{createGiottoInstructions}
\alias{createGiottoInstructions}
\title{createGiottoInstructions}
\usage{
createGiottoInstructions(
  python_path = NULL,
  show_plot = NULL,
  return_plot = NULL,
  save_plot = NULL,
  save_dir = NULL,
  plot_format = NULL,
  dpi = NULL,
  units = NULL,
  height = NULL,
  width = NULL,
  is_docker = FALSE
)
}
\arguments{
\item{python_path}{The path to your custom python binary (only at the beginning of the script).}

\item{show_plot}{Show the plot to the console, default = TRUE.}

\item{return_plot}{Return the plot as a \code{\link{ggplot2}} object (e.g. to further modify or save the plot), default = TRUE.}

\item{save_plot}{Automatically save the plot to your designated save_dir, default = FALSE.}

\item{save_dir}{Path of the directory to save plot to.}

\item{plot_format}{The output format of plots. Can be "png" (the default), "tiff" or "pdf".}

\item{dpi}{Plot resolution in dpi for raster images.}

\item{units}{The units in which height and width are given. Can be px(pixels), in (inches, the default), cm or mm.}

\item{height}{Plot height in `units`.}

\item{width}{Plot width in `units`.}

\item{is_docker}{Logical. Whether using docker implementation of Giotto or not (default = FALSE).}
}
\value{
A \code{list} with Giotto instructions.
}
\description{
Function to set global instructions for giotto functions.
}
\details{
Instructions can be set at the beginning of each workflow with \code{\link{createGiottoInstructions}}.

The provided instructions for a giotto object can be viewed with \code{\link{showGiottoInstructions}}.

One or more parameters can be changed with \code{\link{changeGiottoInstructions}}.

All instructions can be replaced with \code{\link{replaceGiottoInstructions}}.

Specific saving instructions can also be directly provided for each plotting function. See \code{\link{showSaveParameters}} for more details.
}
\examples{
instructions <- createGiottoInstruction(python_path = "/path/to/python",
                                        save_plot = TRUE,
                                        save_dir = "/path/to/working_dir")

giotto_object <- createGiottoObject(raw_exprs = "/path/to/expression_matrix",
                                    spatial_locs = "/path/to/spatial_locations",
                                    instructions = instructions)

}
\seealso{
More online information can be found here \url{https://rubd.github.io/Giotto_site/articles/instructions_and_plotting.html}
}
