% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{createGiottoLargeImage}
\alias{createGiottoLargeImage}
\title{createGiottoLargeImage}
\usage{
createGiottoLargeImage(
  raster_object,
  name = "image",
  negative_y = TRUE,
  extent = NULL,
  use_rast_ext = FALSE,
  image_transformations = NULL,
  flip_vertical = FALSE,
  flip_horizontal = FALSE,
  xmax_bound = NULL,
  xmin_bound = NULL,
  ymax_bound = NULL,
  ymin_bound = NULL,
  scale_factor = 1,
  verbose = TRUE
)
}
\arguments{
\item{raster_object}{terra SpatRaster image object}

\item{name}{name for the image}

\item{negative_y}{Map image to negative y spatial values if TRUE. Meaning that origin is in upper left instead of lower left.}

\item{extent}{SpatExtent object to assign spatial extent. Takes priority unless use_rast_ext is TRUE.}

\item{use_rast_ext}{Use extent from input raster object}

\item{image_transformations}{vector of sequential image transformations - under construction}

\item{flip_vertical}{flip raster in a vertical manner}

\item{flip_horizontal}{flip raster in a horizontal manner}

\item{xmax_bound, xmin_bound, ymax_bound, ymin_bound}{assign min and max x and y
values for image spatial placement}

\item{scale_factor}{scaling of image dimensions relative to spatial coordinates}

\item{verbose}{be verbose}
}
\value{
a giottoLargeImage object
}
\description{
Creates a large giotto image that can be added to a Giotto subcellular object. Generates deep copy of SpatRaster
}
