% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{createGiottoLargeImageList}
\alias{createGiottoLargeImageList}
\title{createGiottoLargeImageList}
\usage{
createGiottoLargeImageList(
  raster_objects,
  names = "image",
  negative_y = TRUE,
  extent = NULL,
  use_rast_ext = FALSE,
  image_transformations = NULL,
  xmax_bound = NULL,
  xmin_bound = NULL,
  ymax_bound = NULL,
  ymin_bound = NULL,
  scale_factor = 1,
  verbose = TRUE
)
}
\arguments{
\item{raster_objects}{vector of image paths or terra SpatRaster image objects}

\item{names}{vector of names for the images}

\item{negative_y}{Map image to negative y spatial values if TRUE. Meaning that origin is in upper left instead of lower left.}

\item{extent}{SpatExtent object to assign spatial extent. Takes priority unless use_rast_ext is TRUE.}

\item{image_transformations}{vector of sequential image transformations - under construction}

\item{scale_factor}{scaling of image dimensions relative to spatial coordinates}

\item{verbose}{be verbose}

\item{use_rast_extent}{Use extent from input raster object}

\item{xmax_adj}{adjustment of the maximum x-value to align the image}

\item{xmin_adj}{adjustment of the minimum x-value to align the image}

\item{ymax_adj}{adjustment of the maximum y-value to align the image}

\item{ymin_adj}{adjustment of the minimum y-value to align the image}
}
\value{
a list with giottoLargeImage objects
}
\description{
Creates a list of large giotto images that can be added to a Giotto object. Generates deep copy of SpatRaster
}
\details{
See \code{\link{createGiottoLargeImage}}
}
