% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto.R
\name{createGiottoObject}
\alias{createGiottoObject}
\title{create Giotto object}
\usage{
createGiottoObject(raw_exprs, spatial_locs = NULL, norm_expr = NULL,
  norm_scaled_expr = NULL, custom_expr = NULL, cell_metadata = NULL,
  gene_metadata = NULL, spatial_network = NULL,
  spatial_network_name = NULL, spatial_grid = NULL,
  spatial_grid_name = NULL, dimension_reduction = NULL,
  nn_network = NULL, offset_file = NULL, instructions = NULL)
}
\arguments{
\item{raw_exprs}{matrix with raw expression counts [required]}

\item{spatial_locs}{data.table with coordinates for cell centroids [required]}

\item{norm_expr}{normalized expression values}

\item{norm_scaled_expr}{scaled expression values}

\item{custom_expr}{custom expression values}

\item{cell_metadata}{cell metadata}

\item{gene_metadata}{gene metadata}

\item{spatial_network}{list of spatial network(s)}

\item{spatial_network_name}{list of spatial network name(s)}

\item{spatial_grid}{list of spatial grid(s)}

\item{spatial_grid_name}{list of spatial grid name(s)}

\item{dimension_reduction}{list of dimension reduction(s)}

\item{nn_network}{list of nearest neighbor network(s)}

\item{offset_file}{file used to stitch fields together (optional)}
}
\value{
giotto object
}
\description{
Function to create a giotto object
}
\examples{
    createGiottoObject(raw_exprs, spatial_locs)
}
\keyword{giotto}
