% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{createGiottoObjectfromArchR}
\alias{createGiottoObjectfromArchR}
\title{Create a Giotto object from an ArchR project}
\usage{
createGiottoObjectfromArchR(
  archRproj,
  expression = NULL,
  expression_feat = "atac",
  spatial_locs = NULL,
  sampleNames = "sample1",
  ...
)
}
\arguments{
\item{archRproj}{}

\item{expression}{expression information}

\item{expression_feat}{Giotto object available features (e.g. atac, rna, ...)}

\item{spatial_locs}{data.table or data.frame with coordinates for cell centroids}

\item{sampleNames}{A character vector containing the ArchR project sample name}

\item{...}{additional arguments passed to `createGiottoObject`}
}
\value{
A Giotto object with at least an atac or epigenetic modality
}
\description{
Create a Giotto object from an ArchR project
}
