% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_structures.R
\name{createGiottoPoints}
\alias{createGiottoPoints}
\title{Create giotto points object}
\usage{
createGiottoPoints(x, feat_type = "rna")
}
\arguments{
\item{x}{spatVector or data.frame-like object with points coordinate information (x, y, feat ID)}

\item{feat_type}{feature type}
}
\value{
giottoPoints
}
\description{
Creates Giotto point object from a structured dataframe-like object
}
\concept{polygon}
