% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_structures.R
\name{createGiottoPolygonsFromDfr}
\alias{createGiottoPolygonsFromDfr}
\title{createGiottoPolygonsFromDfr}
\usage{
createGiottoPolygonsFromDfr(segmdfr, name = "cell", calc_centroids = FALSE)
}
\arguments{
\item{segmdfr}{data.frame-like object with polygon coordinate information (x, y, ID)}

\item{name}{name for polygons}

\item{calc_centroids}{calculate centroids for polygons}
}
\value{

}
\description{
Creates Giotto polygon object from a structured dataframe-like object
}
\keyword{polygon}
