% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_general.R
\name{createMerscopeLargeImage}
\alias{createMerscopeLargeImage}
\title{Create Vizgen MERSCOPE largeImage}
\usage{
createMerscopeLargeImage(image_file, transforms_file, name = "image")
}
\arguments{
\item{image_file}{character. Path to one or more MERSCOPE images to load}

\item{transforms_file}{character. Path to MERSCOPE transforms file. Usually
in the same folder as the images and named
'micron_to_mosaic_pixel_transform.csv'}

\item{name}{character. name to assign the image. Multiple should be provided
if image_file is a list.}
}
\value{
giottoLargeImage
}
\description{
Read MERSCOPE stitched images as giottoLargeImage. Images will also be
transformed to match the spatial coordinate reference system of the paired
points and polygon data.
}
