% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{createNearestNetObj}
\alias{createNearestNetObj}
\title{Create S4 nnNetObj}
\usage{
createNearestNetObj(
  name = "test",
  network,
  nn_type = NULL,
  spat_unit = "cell",
  feat_type = "rna",
  provenance = NULL,
  misc = NULL
)
}
\arguments{
\item{name}{name of nnNetObj}

\item{network}{igraph object or data.frame containing nearest neighbor
information (see details)}

\item{nn_type}{type of nearest neighbor network}

\item{misc}{misc}
}
\description{
Create an S4 nnNetObj
}
\details{
igraph and dataframe-like inputs must include certain information.
For igraph, it must have, at minimum vertex 'name' attributes and 'distance'
edge attribute.
dataframe-like inputs must have 'from', 'to', and 'distance' columns
}
\section{Slots}{

\describe{
\item{\code{spat_unit}}{spatial unit of data}

\item{\code{feat_type}}{feature type of data}

\item{\code{provenance}}{origin of aggregated information (if applicable)}
}}

