% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NN_network.R
\name{createNearestNetwork}
\alias{createNearestNetwork}
\title{createNearestNetwork}
\usage{
createNearestNetwork(
  gobject,
  expression_values = c("normalized", "scaled", "custom"),
  type = c("sNN", "kNN"),
  dim_reduction_to_use = "pca",
  dim_reduction_name = "pca",
  dimensions_to_use = 1:10,
  genes_to_use = NULL,
  name = "sNN.pca",
  return_gobject = TRUE,
  k = 30,
  minimum_shared = 5,
  top_shared = 3,
  verbose = T,
  ...
)
}
\arguments{
\item{gobject}{giotto object}

\item{expression_values}{expression values to use}

\item{type}{kNN or sNN}

\item{dim_reduction_to_use}{dimension reduction method to use}

\item{dim_reduction_name}{name of dimension reduction set to use}

\item{dimensions_to_use}{number of dimensions to use as input}

\item{genes_to_use}{if dim_reduction_to_use = NULL, which genes to use}

\item{name}{arbitrary name for NN network}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}

\item{k}{number of k neighbors to use}

\item{minimum_shared}{minimum shared neighbors}

\item{top_shared}{keep at ...}

\item{verbose}{be verbose}

\item{...}{additional parameters}
}
\value{
giotto object with updated NN network
}
\description{
create a nearest neighbour network based on previously computed \cr
dimension reductions
}
\details{
Description of nearest neighbor network creation and filter steps.
}
\examples{
    createNearestNetwork(gobject)
}
