% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registration.R
\name{createRegZStackGobject}
\alias{createRegZStackGobject}
\title{createRegZStackGobject}
\usage{
createRegZStackGobject(
  expression_list,
  spatlocs_list,
  xvals,
  yvals,
  scalefactor = 1,
  transformXML,
  z_list,
  instructions = NULL
)
}
\arguments{
\item{expression_list}{list of expression values to use}

\item{spatlocs_list}{list of spatlocs to use}

\item{xvals}{x value spatial input}

\item{yvals}{y value spatial input}

\item{scalefactor}{scale spatial coords to pixel coords}

\item{transformXML}{transformation files to use}

\item{z_list}{z values to use in z stacking. In order of expression_list and spatloc_list.}

\item{instructions}{instructions for Giotto processing}
}
\description{
create registered z stacked 3D gobject. Images not supported
}
