% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_deconvolution.R
\name{createSpatialDeconvolution}
\alias{createSpatialDeconvolution}
\title{createSpatialDeconvolution}
\usage{
createSpatialDeconvolution(gobject, sign_matrix, n_cell = 50)
}
\arguments{
\item{sign_matrix}{sig matrix for deconvolution}

\item{method}{list of genes (signature)}
}
\value{
matrix
}
\description{
Function to deconvolution based on single cell expression data
}
\examples{
createSpatialDeconvolution()
}
