% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_structures.R
\name{createSpatialDefaultGrid}
\alias{createSpatialDefaultGrid}
\title{createSpatialDefaultGrid}
\usage{
createSpatialDefaultGrid(
  gobject,
  sdimx_stepsize = NULL,
  sdimy_stepsize = NULL,
  sdimz_stepsize = NULL,
  minimum_padding = 1,
  name = NULL,
  return_gobject = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{sdimx_stepsize}{stepsize along the x-axis}

\item{sdimy_stepsize}{stepsize along the y-axis}

\item{sdimz_stepsize}{stepsize along the z-axis}

\item{minimum_padding}{minimum padding on the edges}

\item{name}{name for spatial grid (default = 'spatial_grid')}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}
}
\value{
giotto object with updated spatial grid slot
}
\description{
Create a spatial grid using the default method
}
\details{
Creates a spatial grid with defined x, y (and z) dimensions.
The dimension units are based on the provided spatial location units.
}
