% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_structures.R
\name{createSpatialFeaturesKNNnetwork}
\alias{createSpatialFeaturesKNNnetwork}
\title{createSpatialFeaturesKNNnetwork}
\usage{
createSpatialFeaturesKNNnetwork(
  gobject,
  method = c("dbscan"),
  feat_type = NULL,
  name = "knn_feats_network",
  k = 4,
  maximum_distance = NULL,
  minimum_k = 0,
  add_feat_ids = FALSE,
  verbose = TRUE,
  return_gobject = TRUE,
  toplevel_params = 2,
  ...
)
}
\arguments{
\item{gobject}{giotto object}

\item{method}{kNN algorithm method}

\item{feat_type}{feature type to build feature network}

\item{name}{name of network}

\item{k}{number of neighbors}

\item{maximum_distance}{maximum distance bewteen features}

\item{minimum_k}{minimum number of neighbors to find}

\item{add_feat_ids}{add feature id names (default = FALSE, increases object size)}

\item{verbose}{be verbose}

\item{return_gobject}{return giotto object (default: TRUE)}
}
\value{

}
\description{
Calculates the centroid locations for the giotto polygons
}
\keyword{Features}
