% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_structures.R
\name{createSpatialFeaturesKNNnetwork_dbscan}
\alias{createSpatialFeaturesKNNnetwork_dbscan}
\title{Create kNN spatial feature network using dbscan}
\usage{
createSpatialFeaturesKNNnetwork_dbscan(
  gobject,
  feat_type = NULL,
  name = "knn_feats_network",
  k = 4,
  maximum_distance = NULL,
  minimum_k = 0,
  add_feat_ids = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{gobject}{giotto object}

\item{feat_type}{feature type}

\item{name}{name to assign generated feature network}

\item{k}{number of neighbors for kNN to find}

\item{maximum_distance}{network maximum distance allowed}

\item{minimum_k}{minimum neighbors allowed}

\item{add_feat_ids}{whether to add feature information [boolean]}

\item{verbose}{be verbose}

\item{...}{additional parameters to pass}
}
\description{
to create a feature kNN spatial network using dbscan
}
\keyword{internal}
