% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{createSpatialGenomicsObject}
\alias{createSpatialGenomicsObject}
\title{Create Spatial Genomics Giotto Object}
\usage{
createSpatialGenomicsObject(sg_dir = NULL, instructions = NULL)
}
\arguments{
\item{sg_dir}{full path to the exported Spatial Genomics directory}

\item{instructions}{new instructions (e.g. result from createGiottoInstructions)}
}
\description{
Given the path to a Spatial Genomics data directory, creates a 
Giotto object.
}
