% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_structures.R
\name{createSpatialGrid}
\alias{createSpatialGrid}
\title{createSpatialGrid}
\usage{
createSpatialGrid(
  gobject,
  name = NULL,
  method = c("default"),
  sdimx_stepsize = NULL,
  sdimy_stepsize = NULL,
  sdimz_stepsize = NULL,
  minimum_padding = 1,
  return_gobject = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{name}{name for spatial grid}

\item{method}{method to create a spatial grid}

\item{sdimx_stepsize}{stepsize along the x-axis}

\item{sdimy_stepsize}{stepsize along the y-axis}

\item{sdimz_stepsize}{stepsize along the z-axis}

\item{minimum_padding}{minimum padding on the edges}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}
}
\value{
giotto object with updated spatial grid slot
}
\description{
Create a spatial grid using the default method
}
\details{
Creates a spatial grid with defined x, y (and z) dimensions.
The dimension units are based on the provided spatial location units.
\itemize{
  \item{default method: }{\code{\link{createSpatialDefaultGrid}}}
}
}
