% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_structures.R
\name{createSpatialKNNnetwork_v2}
\alias{createSpatialKNNnetwork_v2}
\title{createSpatialKNNnetwork_v2}
\usage{
createSpatialKNNnetwork_v2(
  gobject,
  method = "dbscan",
  dimensions = "all",
  name = "knn_network",
  k = 4,
  maximum_distance = NULL,
  minimum_k = 0,
  verbose = F,
  return_gobject = TRUE,
  ...
)
}
\arguments{
\item{gobject}{giotto object}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}
}
\value{
giotto object with updated spatial network slot
}
\description{
Create a spatial knn network.
}
\examples{
    createSpatialKNNnetwork_v2(gobject)
}
