% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_structures.R
\name{createSpatialNetworkOLD}
\alias{createSpatialNetworkOLD}
\title{createSpatialNetworkOLD}
\usage{
createSpatialNetworkOLD(
  gobject,
  k = 4,
  dimensions = "all",
  maximum_distance = NULL,
  minimum_k = 0,
  name = "spatial_network",
  verbose = F,
  return_gobject = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{k}{number of nearest neighbors based on physical distance}

\item{dimensions}{which spatial dimensions to use (default = all)}

\item{maximum_distance}{distance cuttof for nearest neighbors to consider}

\item{minimum_k}{minimum nearest neigbhours if maximum_distance != NULL}

\item{name}{name for spatial network (default = 'spatial_network')}

\item{verbose}{verbose}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}
}
\value{
giotto object with updated spatial network slot
}
\description{
Create a spatial network based on cell centroid physical distances.
}
\details{
Creates a spatial network connecting single-cells based on their physical distance to each other.
Number of neighbors can be determined by k, maximum distance from each cell with or without
setting a minimum k for each cell.

\strong{dimensions: } default = 'all' which takes all possible dimensions.
Alternatively you can provide a character vector that specififies the spatial dimensions to use, e.g. c("sdimx', "sdimy")
or a numerical vector, e.g. 2:3

\strong{maximum_distance: } to create a network based on maximum distance only, you also need to set k to a very high value, e.g. k = 100
}
\examples{
    createSpatialNetworkOLD(gobject)
}
