% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{createSpatNetObj}
\alias{createSpatNetObj}
\title{Create S4 spatialNetworkObj}
\usage{
createSpatNetObj(
  network,
  name = "test",
  networkDT_before_filter = NULL,
  method = NULL,
  spat_unit = "cell",
  provenance = NULL,
  parameters = NULL,
  outputObj = NULL,
  cellShapeObj = NULL,
  crossSectionObjects = NULL,
  misc = NULL
)
}
\arguments{
\item{network}{network data with connections, distances, and weightings}

\item{name}{name of spatialNetworkObj}

\item{networkDT_before_filter}{(optional) unfiltered data.table  of network connections, distances, and weightings}

\item{method}{method used to generate spatial network}

\item{spat_unit}{spatial unit tag}

\item{provenance}{(optional) origin of aggregated information (if applicable)}

\item{parameters}{(optional) additional method-specific parameters used during spatial network generation}

\item{outputObj}{(optional) network geometry object}

\item{cellShapeObj}{(optional) network cell shape information}

\item{crossSectionObjects}{(optional) crossSectionObjects (see \code{\link{create_crossSection_object}})}

\item{misc}{misc}
}
\description{
Create S4 spatialNetworkObj
}
