% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{cropGiottoLargeImage}
\alias{cropGiottoLargeImage}
\title{cropGiottoLargeImage}
\usage{
cropGiottoLargeImage(
  gobject = NULL,
  largeImage_name = NULL,
  giottoLargeImage = NULL,
  crop_name = "image",
  crop_extent = NULL,
  xmax_crop = NULL,
  xmin_crop = NULL,
  ymax_crop = NULL,
  ymin_crop = NULL
)
}
\arguments{
\item{gobject}{gobject holding the giottoLargeImage}

\item{largeImage_name}{name of giottoLargeImage within gobject}

\item{giottoLargeImage}{a giottoLargeImage}

\item{crop_name}{arbitrary name for cropped giottoLargeImage}

\item{crop_extent}{terra extent object used to crop the giottoLargeImage}

\item{xmax_crop}{crop xmax bound}

\item{xmin_crop}{crop xmin bound}

\item{ymax_crop}{crop ymax bound}

\item{ymin_crop}{crop ymin bound}
}
\value{
a giottoLargeImage
}
\description{
crop a giottoLargeImage based on crop_extent argument or given values
}
