% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_section.R
\name{crossSectionFeatPlot3D}
\alias{crossSectionFeatPlot3D}
\title{crossSectionFeatPlot3D}
\usage{
crossSectionFeatPlot3D(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  crossSection_obj = NULL,
  name = NULL,
  spatial_network_name = "Delaunay_network",
  show_other_cells = TRUE,
  other_cell_color = alpha("lightgrey", 0),
  default_save_name = "crossSectionGenePlot3D",
  ...
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit}

\item{feat_type}{feature type}

\item{crossSection_obj}{cross section object as alternative input. default = NULL.}

\item{name}{name of virtual cross section to use}

\item{spatial_network_name}{name of spatial network to use}

\item{show_other_cells}{logical. Default = TRUE}

\item{other_cell_color}{color of cells outside the cross section.
default = transparent.}

\item{default_save_name}{default save name for saving, don't change, change
save_name in save_param}

\item{...}{parameters for spatGenePlot3D}
}
\value{
ggplot
}
\description{
Visualize cells and feature expression in a virtual cross
section according to spatial coordinates
}
\details{
Description of parameters.
}
