% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_section.R
\name{crossSectionGenePlot}
\alias{crossSectionGenePlot}
\title{crossSectionGenePlot}
\usage{
crossSectionGenePlot(
  gobject = NULL,
  spat_loc_name = "raw",
  crossSection_obj = NULL,
  name = NULL,
  spatial_network_name = "Delaunay_network",
  default_save_name = "crossSectionGenePlot",
  ...
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_loc_name}{name of spatial locations}

\item{crossSection_obj}{crossSection object}

\item{name}{name of virtual cross section to use}

\item{spatial_network_name}{name of spatial network to use}

\item{default_save_name}{default save name for saving, don't change,
change save_name in save_param}

\item{...}{parameters for spatFeatPlot2D}
}
\value{
ggplot
}
\description{
Visualize cells and gene expression in a virtual cross section
according to spatial coordinates
}
\details{
Description of parameters.
}
\seealso{
\link[GiottoVisuals:spatFeatPlot3D]{GiottoVisuals::spatGenePlot3D} and \link[GiottoVisuals:spatFeatPlot2D]{GiottoVisuals::spatFeatPlot2D}
}
