% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_section.R
\name{crossSectionGenePlot3D}
\alias{crossSectionGenePlot3D}
\title{crossSectionGenePlot3D}
\usage{
crossSectionGenePlot3D(
  gobject,
  crossSection_obj = NULL,
  name = NULL,
  spatial_network_name = "Delaunay_network",
  other_cell_color = alpha("lightgrey", 0),
  default_save_name = "crossSectionGenePlot3D",
  ...
)
}
\arguments{
\item{gobject}{giotto object}

\item{crossSection_obj}{cross section object as alternative input. default = NULL.}

\item{name}{name of virtual cross section to use}

\item{spatial_network_name}{name of spatial network to use}

\item{other_cell_color}{color of cells outside the cross section. default = transparent.}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}

\item{...}{parameters for spatGenePlot3D}
}
\value{
ggplot
}
\description{
Visualize cells and gene expression in a virtual cross section according to spatial coordinates
}
\details{
Description of parameters.
}
