% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{data_access}
\alias{data_access}
\title{Giotto object data accessors}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit (e.g. "cell")}

\item{feat_type}{feature type (e.g. "rna", "dna", "protein")}

\item{return_uniques}{return unique nesting names (ignores if final object exists/is correct class)}

\item{output}{what format in which to get information (e.g. "data.table")}
}
\description{
Access or examine slots within the giotto object
}
\keyword{internal}
