% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd.R
\name{data_access_params}
\alias{data_access_params}
\title{data_access_params}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit (e.g. "cell")}

\item{feat_type}{feature type (e.g. "rna", "dna", "protein")}

\item{return_uniques}{return unique nesting names (ignores if final object exists/is correct class)}

\item{output}{what format in which to get information (e.g. "data.table")}

\item{set_defaults}{set default spat_unit and feat_type. Change to FALSE only when
expression and spat_info are not expected to exist.}

\item{copy_obj}{whether to deep copy/duplicate when getting the object (default = TRUE)}

\item{initialize}{(default = FALSE) whether to initialize the gobject before
returning}
}
\description{
data_access_params
}
\keyword{internal}
