% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{dbPath-generic}
\alias{dbPath-generic}
\alias{dbPath}
\alias{dbPath<-}
\alias{dbPath,giotto-method}
\alias{dbPath<-,giotto,character-method}
\title{Database backend path}
\usage{
dbPath(gobject, ...)

\S4method{dbPath}{giotto}(gobject)

\S4method{dbPath}{giotto,character}(gobject, extension = ".duckdb", ...) <- value
}
\arguments{
\item{gobject}{giotto object}
}
\description{
Retrieve or set the database backend path in the Giotto instructions.
Immediately updates the hash ID contained within any existing GiottoDB backendInfo
object in the gobject dbInfo slot.
When dbData objects access that backendInfo object for reconnection, they will
also update their hash ID values
}
