% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{degrees}
\alias{degrees}
\alias{radians}
\title{Radian/degree conversions}
\usage{
radians(deg)

degrees(rad)
}
\arguments{
\item{deg}{degrees}

\item{rad}{radians}
}
\description{
Convert radians to degrees and vice versa
}
\section{Functions}{
\itemize{
\item \code{radians()}: Degrees to radians

\item \code{degrees()}: Radians to degrees

}}
\keyword{internal}
