% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{density_giottoLargeImage}
\alias{density_giottoLargeImage}
\alias{dist_giottoLargeImage}
\title{Plot smoothed curve of giotto largeImage intensity values}
\usage{
dist_giottoLargeImage(
  gobject = NULL,
  image_name = NULL,
  giottoLargeImage = NULL,
  method = "dens"
)
}
\arguments{
\item{gobject}{giotto object}

\item{giottoLargeImage}{giotto large image object}

\item{method}{method of plotting image distribution}

\item{image_type}{image object type (only supports largeImage and is set as
default)}
}
\description{
Plot smoothed curve of giotto largeImage intensity values
}
\keyword{internal}
