% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto.R
\name{depth}
\alias{depth}
\title{Find depth of subnesting}
\usage{
depth(this, method = c("max", "min"), sig = "data.frame")
}
\arguments{
\item{this}{object to evaluate}

\item{method}{max (default) or min nesting to detect}

\item{sig}{signature or class to check for. Default is 'data.frame'}
}
\description{
Recursively determines how many max or min layers of subnesting
there is, with the end object (defined by param sig or a list of length 0)
being layer 0
}
\details{
https://stackoverflow.com/questions/13432863/determine-level-of-nesting-in-r
}
\keyword{internal}
