% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_genes.R
\name{detectSpatialCorFeats}
\alias{detectSpatialCorFeats}
\alias{detectSpatialCorFeatsMatrix}
\title{Detect spatially correlated features}
\usage{
detectSpatialCorFeats(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  spat_loc_name = "raw",
  method = c("grid", "network"),
  expression_values = c("normalized", "scaled", "custom"),
  subset_feats = NULL,
  spatial_network_name = "Delaunay_network",
  network_smoothing = NULL,
  spatial_grid_name = "spatial_grid",
  min_cells_per_grid = 4,
  cor_method = c("pearson", "kendall", "spearman")
)

detectSpatialCorFeatsMatrix(
  expression_matrix,
  method = c("grid", "network"),
  spatial_network,
  spatial_grid,
  spatial_locs,
  subset_feats = NULL,
  network_smoothing = NULL,
  min_cells_per_grid = 4,
  cor_method = c("pearson", "kendall", "spearman")
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit}

\item{feat_type}{feature type}

\item{spat_loc_name}{name for spatial locations}

\item{method}{method to use for spatial averaging}

\item{expression_values}{gene expression values to use}

\item{subset_feats}{subset of features to use}

\item{spatial_network_name}{name of spatial network to use}

\item{network_smoothing}{smoothing factor between 0 and 1
(has automatic default, see details)}

\item{spatial_grid_name}{name of spatial grid to use}

\item{min_cells_per_grid}{minimum number of cells to consider a grid}

\item{cor_method}{correlation method}

\item{expression_matrix}{provided expression matrix}

\item{spatial_network}{provided spatial network}

\item{spatial_grid}{provided spatial grid}

\item{spatial_locs}{provided spatial locations}
}
\value{
returns a spatial correlation object: \code{spatCorObject}
}
\description{
Detect features that are spatially correlated. Functions for
starting from either a gobject (\code{detectSpatialCorFeats()}) or individual
pieces of data (\code{detectSpatialCorFeatsMatrix()}) are provided.
}
\details{
For \code{method = network}, it expects a fully connected spatial network.
You can make sure to create a
fully connected network by setting minimal_k > 0 in the
\code{\link{createSpatialNetwork}} function.
\enumerate{
\item \strong{grid-averaging:} average gene expression values within a predefined
spatial grid
\item \strong{network-averaging:} smoothens the gene expression matrix by
averaging the expression within one cell by using the neighbours within
the predefined spatial network. \eqn{b} is a smoothening factor passed by
\code{network_smoothing} param that defaults to \eqn{1 - 1/k}, where \eqn{k}
is the median number of k-neighbors in the selected spatial network.
Setting \eqn{b = 0} means no smoothing and \eqn{b = 1} means no
contribution from its own expression.
}

The \code{spatCorObject} can be further explored with \code{showSpatialCorFeats()}
}
\examples{
g <- GiottoData::loadGiottoMini("visium")

# Perform with data in a gobject
detectSpatialCorFeats(g, method = "network")

# This analysis can also be performed with data outside of the gobject
detectSpatialCorFeatsMatrix(
    expression_matrix = getExpression(
        g,
        output = "matrix"
    ),
    method = "network",
    spatial_network = getSpatialNetwork(g, output = "networkDT")
)

}
\seealso{
\code{\link{showSpatialCorFeats}}
}
