% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_genes.R
\name{detectSpatialCorFeatsMatrix}
\alias{detectSpatialCorFeatsMatrix}
\title{detectSpatialCorFeatsMatrix}
\usage{
detectSpatialCorFeatsMatrix(
  expression_matrix,
  method = c("grid", "network"),
  spatial_network,
  spatial_grid,
  spatial_locs,
  subset_feats = NULL,
  network_smoothing = NULL,
  min_cells_per_grid = 4,
  cor_method = c("pearson", "kendall", "spearman")
)
}
\arguments{
\item{expression_matrix}{provided expression matrix}

\item{method}{method to use for spatial averaging}

\item{spatial_network}{provided spatial network}

\item{spatial_grid}{provided spatial grid}

\item{spatial_locs}{provided spatial locations}

\item{subset_feats}{subset of features to use}

\item{network_smoothing}{smoothing factor beteen 0 and 1 
(default: automatic)}

\item{min_cells_per_grid}{minimum number of cells to consider a grid}

\item{cor_method}{correlation method}
}
\value{
returns a spatial correlation object: "spatCorObject"
}
\description{
Detect genes that are spatially correlated
}
\details{
For method = network, it expects a fully connected spatial network. 
You can make sure to create a
fully connected network by setting minimal_k > 0 in the 
\code{\link{createSpatialNetwork}} function.
\itemize{
 \item{1. grid-averaging: }{average gene expression values within a predefined spatial grid}
 \item{2. network-averaging: }{smoothens the gene expression matrix by averaging the expression within one cell
 by using the neighbours within the predefined spatial network. b is a smoothening factor
 that defaults to 1 - 1/k, where k is the median number of  k-neighbors in the
 selected spatial network. Setting b = 0 means no smoothing and b = 1 means no contribution
 from its own expression.}
}
The spatCorObject can be further explored with showSpatialCorGenes()
}
\seealso{
\code{\link{showSpatialCorFeats}}
}
