% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_genes.R
\name{detectSpatialCorGenes}
\alias{detectSpatialCorGenes}
\title{detectSpatialCorGenes}
\usage{
detectSpatialCorGenes(
  gobject,
  method = c("grid", "network"),
  expression_values = c("normalized", "scaled", "custom"),
  subset_genes = NULL,
  spatial_network_name = "Delaunay_network",
  network_smoothing = NULL,
  spatial_grid_name = "spatial_grid",
  min_cells_per_grid = 4,
  cor_method = c("pearson", "kendall", "spearman")
)
}
\arguments{
\item{gobject}{giotto object}

\item{method}{method to use for spatial averaging}

\item{expression_values}{gene expression values to use}

\item{subset_genes}{subset of genes to use}

\item{spatial_network_name}{name of spatial network to use}

\item{network_smoothing}{smoothing factor beteen 0 and 1 (default: automatic)}

\item{spatial_grid_name}{name of spatial grid to use}

\item{min_cells_per_grid}{minimum number of cells to consider a grid}

\item{cor_method}{correlation method}
}
\value{
returns a spatial correlation object: "spatCorObject"
}
\description{
Detect genes that are spatially correlated
}
\details{
For method = network, it expects a fully connected spatial network. You can make sure to create a
fully connected network by setting minimal_k > 0 in the \code{\link{createSpatialNetwork}} function.
\itemize{
 \item{1. grid-averaging: }{average gene expression values within a predefined spatial grid}
 \item{2. network-averaging: }{smoothens the gene expression matrix by averaging the expression within one cell
 by using the neighbours within the predefined spatial network. b is a smoothening factor
 that defaults to 1 - 1/k, where k is the median number of  k-neighbors in the
 selected spatial network. Setting b = 0 means no smoothing and b = 1 means no contribution
 from its own expression.}
}
The spatCorObject can be further explored with showSpatialCorGenes()
}
\seealso{
\code{\link{showSpatialCorGenes}}
}
