% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_genes.R
\name{detectSpatialPatterns}
\alias{detectSpatialPatterns}
\title{detectSpatialPatterns}
\usage{
detectSpatialPatterns(
  gobject,
  expression_values = c("normalized", "scaled", "custom"),
  spatial_grid_name = "spatial_grid",
  min_cells_per_grid = 4,
  scale_unit = F,
  ncp = 100,
  show_plot = T,
  PC_zscore = 1.5
)
}
\arguments{
\item{gobject}{giotto object}

\item{expression_values}{expression values to use}

\item{spatial_grid_name}{name of spatial grid to use (default = 'spatial_grid')}

\item{min_cells_per_grid}{minimum number of cells in a grid to be considered}

\item{scale_unit}{scale features}

\item{ncp}{number of principal components to calculate}

\item{show_plot}{show plots}

\item{PC_zscore}{minimum z-score of variance explained by a PC}
}
\value{
spatial pattern object 'spatPatObj'
}
\description{
Identify spatial patterns through PCA on average expression in a spatial grid.
}
\details{
Steps to identify spatial patterns:
\itemize{
  \item{1. average gene expression for cells within a grid, see createSpatialGrid}
  \item{2. perform PCA on the average grid expression profiles}
  \item{3. convert variance of principlal components (PCs) to z-scores and select PCs based on a z-score threshold}
}
}
