% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_help.R
\name{determine_cores}
\alias{determine_cores}
\title{determine_cores}
\usage{
determine_cores(cores, min_cores = 1, max_cores = 10)
}
\value{
numeric
}
\description{
guesses how many cores to use
}
\keyword{internal}
