% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_influence.r
\name{determine_switch_string_unequal}
\alias{determine_switch_string_unequal}
\title{determine_switch_string_unequal}
\usage{
determine_switch_string_unequal(num_orig = NULL, num_new = NULL)
}
\arguments{
\item{num_orig}{sorted vector of cluster numbers in the original metadata}

\item{num_new}{sorted vector of cluster numbers in the new, resized metadata}
}
\value{
switch_str, a vector of corresponding cluster numbers in strings
}
\description{
determine_switch_string_unequal
}
\details{
determines how to create a string in the format c("x_1-y_1", "x_2-y_2"..."x_n, y_m")
Where:
   x_n is a cluster number from the original spatial unit
   y_m is a cluster number from the resized spatial unit
   n is the number of clusters in the original spatial unit
   m is the number of clusters in the new spatial unit

Essentially determines iteration order for create_switch_string_unequal()
}
\keyword{internal}
