% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visuals.R
\name{dimGenePlot}
\alias{dimGenePlot}
\title{dimGenePlot}
\usage{
dimGenePlot(gobject, ...)
}
\arguments{
\item{gobject}{giotto object}

\item{...}{parameters for cowplot::save_plot()}

\item{expression_values}{gene expression values to use}

\item{genes}{genes to show}

\item{dim_reduction_to_use}{dimension reduction to use}

\item{dim_reduction_name}{dimension reduction name}

\item{dim1_to_use}{dimension to use on x-axis}

\item{dim2_to_use}{dimension to use on y-axis}

\item{show_NN_network}{show underlying NN network}

\item{nn_network_to_use}{type of NN network to use (kNN vs sNN)}

\item{network_name}{name of NN network to use, if show_NN_network = TRUE}

\item{edge_alpha}{column to use for alpha of the edges}

\item{scale_alpha_with_expression}{scale expression with ggplot alpha parameter}

\item{point_size}{size of point (cell)}

\item{point_border_col}{color of border around points}

\item{point_border_stroke}{stroke size of border around points}

\item{midpoint}{size of point (cell)}

\item{cow_n_col}{cowplot param: how many columns}

\item{cow_rel_h}{cowplot param: relative height}

\item{cow_rel_w}{cowplot param: relative width}

\item{cow_align}{cowplot param: how to align}

\item{show_legend}{show legend}

\item{show_plot}{show plots}

\item{return_plot}{return ggplot object}

\item{save_plot}{directly save the plot [boolean]}

\item{save_param}{list of saving parameters from all_plots_save_function()}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}
}
\value{
ggplot
}
\description{
Visualize cells and gene expression according to dimension reduction coordinates
}
\details{
Description of parameters.
}
\examples{
    dimGenePlot(gobject)
}
\seealso{
\code{\link{dimGenePlot3D}}
}
