% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{dimObj-class}
\alias{dimObj-class}
\title{S4 dimObj Class}
\description{
Framework to store dimension reduction information
}
\section{Slots}{

\describe{
\item{\code{name}}{name of dimObject}

\item{\code{feat_type}}{feature type of data}

\item{\code{spat_unit}}{spatial unit of data}

\item{\code{provenance}}{origin of aggregated information (if applicable)}

\item{\code{reduction}}{whether reduction was performed on 'feats' or 'cells'}

\item{\code{reduction_method}}{method used to generate dimension reduction}

\item{\code{coordinates}}{embedding coordinates}

\item{\code{misc}}{method-specific additional outputs}
}}

