% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{dims-generic}
\alias{dims-generic}
\alias{nrow,giottoPoints-method}
\alias{nrow,giottoPolygon-method}
\alias{nrow,coordDataDT-method}
\alias{nrow,exprData-method}
\alias{nrow,metaData-method}
\alias{ncol,exprData-method}
\alias{ncol,metaData-method}
\alias{dim,exprData-method}
\alias{dim,metaData-method}
\title{Dimensions of giotto objects}
\usage{
\S4method{nrow}{giottoPoints}(x)

\S4method{nrow}{giottoPolygon}(x)

\S4method{nrow}{coordDataDT}(x)

\S4method{nrow}{exprData}(x)

\S4method{nrow}{metaData}(x)

\S4method{ncol}{exprData}(x)

\S4method{ncol}{metaData}(x)

\S4method{dim}{exprData}(x)

\S4method{dim}{metaData}(x)
}
\arguments{
\item{x}{object to check dimensions of}
}
\description{
Find the dimensions of an object
}
\section{Functions}{
\itemize{
\item \code{nrow,giottoPoints-method}: Find rows of giottoPoints object

\item \code{nrow,giottoPolygon-method}: Find rows of giottoPolygon object

\item \code{nrow,coordDataDT-method}: Find rows of giotto S4s with data.table based \code{coordinates} slots

\item \code{nrow,exprData-method}: Find rows of giotto S4s with data.table based \code{coordinates} slots

\item \code{nrow,metaData-method}: Find rows of giotto S4s with data.table based \code{coordinates} slots

\item \code{ncol,exprData-method}: Find cols of giotto S4s with Matrix based \code{exprMat} slots

\item \code{ncol,metaData-method}: Find cols of giotto S4s with data.table based \code{metaDT} slots

\item \code{dim,exprData-method}: Find dimensions of giotto S4s with Matrix based \code{exprMat} slots

\item \code{dim,metaData-method}: Find dimensions of giotto S4s with data.table based \code{metaDT} slots
}}

