% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction_visuals.R
\name{direction_test_CPG}
\alias{direction_test_CPG}
\alias{direction_test}
\title{direction_test_CPG}
\usage{
direction_test(x, min_pval = 0.05)
}
\description{
shows direction of change
}
\examples{
    direction_test_CPG()
}
