% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{distGiottoImage}
\alias{distGiottoImage}
\title{Plot distribution of image intensity values}
\usage{
distGiottoImage(
  gobject = NULL,
  image_type = "largeImage",
  image_name = NULL,
  giottoLargeImage = NULL,
  method = c("dens", "hist")
)
}
\arguments{
\item{gobject}{giotto object}

\item{image_type}{image object type (only supports largeImage and is set as
default)}

\item{image_name}{name of image object to use}

\item{giottoLargeImage}{giotto large image object}

\item{method}{plot type to show image intensity distribution}
}
\description{
Plot distribution of intensity values using either a density plot
  or a histogram. Useful for finding image artefact outliers and determining
  reasonable scaling cutoffs.
}
\details{
Plot is generated from a downsampling of the original image
}
\section{Plotting method \code{'dens'}}{

  Density plot of intensity values for image objects. \strong{N} total values
  examined. \strong{Bandwidth} refers to the curve smoothing value applied.
}

\section{Plotting method \code{'hist'}}{

  Histogram of intensity values for image objects.
}

