% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction.R
\name{do_cell_proximity_test}
\alias{do_cell_proximity_test}
\title{do_cell_proximity_test}
\usage{
do_cell_proximity_test(
  expr_values,
  select_ind,
  other_ind,
  diff_test = c("permutation", "limma", "t.test", "wilcox"),
  mean_method = c("arithmic", "geometric"),
  offset = 0.1,
  n_perm = 100,
  adjust_method = c("bonferroni", "BH", "holm", "hochberg", "hommel", "BY", "fdr",
    "none"),
  cores = 2,
  set_seed = TRUE,
  seed_number = 1234
)
}
\description{
Performs a selected differential test on subsets of a matrix
}
\keyword{internal}
