% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction_spot.R
\name{do_cell_proximity_test_spot}
\alias{do_cell_proximity_test_spot}
\title{do_cell_proximity_test_spot}
\usage{
do_cell_proximity_test_spot(
  sel_int,
  other_ints,
  select_ind,
  other_ind,
  proximityMat,
  expr_residual,
  diff_test,
  n_perm = 100,
  adjust_method = "fdr",
  cores = 2,
  set_seed = TRUE,
  seed_number = 1234
)
}
\description{
Performs a selected differential test on subsets of a matrix for spots
}
\keyword{internal}
