% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_structures.R
\name{do_gpoly}
\alias{do_gpoly}
\title{do_gpoly}
\usage{
do_gpoly(x, what, args = NULL)
}
\arguments{
\item{x}{giottoPolygon}

\item{what}{a call to do}

\item{args}{a \code{list} of additional args}
}
\description{
Perform function on all spatVector-based slots of giottoPolygon
}
\keyword{internal}
