% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction.R
\name{do_limmatest}
\alias{do_limmatest}
\title{do_limmatest}
\usage{
do_limmatest(expr_values, select_ind, other_ind, mean_method, offset = 0.1)
}
\description{
Performs limma t.test on subsets of a matrix
}
\keyword{internal}
