% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction_spot.R
\name{do_multi_permuttest_random_spot}
\alias{do_multi_permuttest_random_spot}
\title{do_multi_permuttest_random_spot}
\usage{
do_multi_permuttest_random_spot(
  sel_int,
  other_ints,
  select_ind,
  other_ind,
  proximityMat,
  expr_residual,
  n = 100,
  cores = 2,
  set_seed = TRUE,
  seed_number = 1234
)
}
\description{
calculate multiple random values for spots
}
\keyword{internal}
