% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_enrichment.R
\name{do_page_permutation}
\alias{do_page_permutation}
\title{do_page_permutation}
\usage{
do_page_permutation(gobject, sig_gene, ntimes)
}
\description{
creates permutation for the PAGEEnrich test
}
\keyword{internal}
