% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction.R
\name{do_permuttest_original}
\alias{do_permuttest_original}
\title{do_permuttest_original}
\usage{
do_permuttest_original(
  expr_values,
  select_ind,
  other_ind,
  name = "orig",
  mean_method,
  offset = 0.1
)
}
\description{
calculate original values
}
\keyword{internal}
