% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction_spot.R
\name{do_permuttest_original_spot}
\alias{do_permuttest_original_spot}
\title{do_permuttest_original_spot}
\usage{
do_permuttest_original_spot(
  sel_int,
  other_ints,
  select_ind,
  other_ind,
  name = "orig",
  proximityMat,
  expr_residual
)
}
\description{
calculate original values for spots
}
\keyword{internal}
