% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction.R
\name{do_permuttest_random}
\alias{do_permuttest_random}
\alias{do_permuttest}
\title{do_permuttest_random}
\usage{
do_permuttest_random(
  expr_values,
  select_ind,
  other_ind,
  name = "perm_1",
  mean_method,
  offset = 0.1
)

do_permuttest(
  expr_values,
  select_ind,
  other_ind,
  n_perm = 1000,
  adjust_method = "fdr",
  mean_method,
  offset = 0.1,
  cores = 2
)
}
\description{
calculate random values

Performs permutation test on subsets of a matrix
}
\examples{
    do_permuttest_random()
    do_permuttest_random()
}
