% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction_spot.R
\name{do_permuttest_random_spot}
\alias{do_permuttest_random_spot}
\title{do_permuttest_random_spot}
\usage{
do_permuttest_random_spot(
  sel_int,
  other_ints,
  select_ind,
  other_ind,
  name = "perm_1",
  proximityMat,
  expr_residual,
  set_seed = TRUE,
  seed_number = 1234
)
}
\description{
calculate random values for spots
}
\keyword{internal}
