% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_enrichment.R
\name{do_rank_permutation}
\alias{do_rank_permutation}
\title{do_rank_permutation}
\usage{
do_rank_permutation(sc_gene, n)
}
\description{
creates permutation for the rankEnrich test
}
\keyword{internal}
