% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_genes.R
\name{do_spatial_grid_averaging}
\alias{do_spatial_grid_averaging}
\title{do_spatial_grid_averaging}
\usage{
do_spatial_grid_averaging(
  expression_matrix,
  spatial_grid,
  spatial_locs,
  subset_feats = NULL,
  min_cells_per_grid = 4
)
}
\value{
matrix with smoothened gene expression values based on spatial grid
}
\description{
smooth gene expression over a defined spatial grid
}
\keyword{internal}
