% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction.R
\name{do_ttest}
\alias{do_ttest}
\alias{do_wilctest}
\title{do_ttest}
\usage{
do_ttest(expr_values, select_ind, other_ind, adjust_method)

do_wilctest(expr_values, select_ind, other_ind, adjust_method)
}
\description{
Performs t.test on subsets of a matrix

Performs wilcoxon on subsets of a matrix
}
\examples{
    do_ttest()
    do_ttest()
}
