% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction.R
\name{do_wilctest}
\alias{do_wilctest}
\title{do_wilctest}
\usage{
do_wilctest(
  expr_values,
  select_ind,
  other_ind,
  adjust_method,
  mean_method,
  offset = 0.1
)
}
\description{
Performs wilcoxon on subsets of a matrix
}
\keyword{internal}
