% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_segmentation.R
\name{doCellSegmentation}
\alias{doCellSegmentation}
\title{doCellSegmentation}
\usage{
doCellSegmentation(
  raster_img,
  folder_path,
  reduce_resolution = 4,
  overlapping_pixels = 50
)
}
\arguments{
\item{raster_img}{raster image; Dapi image of cells}

\item{folder_path}{character; where to save the file}

\item{reduce_resolution}{numeric; the original Dapi image from Vizgen works
better in the Mesmer algorithm if its resolution is reduced 4 times.}

\item{overlapping_pixels}{numeric; the number of pixels to overlap when
calculating the rolling window}
}
\description{
segment cells in Dapi image
}
\details{
This function is a wrapper for the Mesmer algorithm implemented in python.
It segments the cells in tissue by applying a rolling window on the complete
image. It saves all the files in the specified folder with the coordinates
of the tile: sx (start x), ex (end x), sy, and ey.
}
