% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{doClusterProjection}
\alias{doClusterProjection}
\title{Projection of cluster labels}
\usage{
doClusterProjection(
  target_gobject,
  target_cluster_label_name = "knn_labels",
  spat_unit = NULL,
  feat_type = NULL,
  source_gobject,
  source_cluster_labels = NULL,
  reduction = "cells",
  reduction_method = "pca",
  reduction_name = "pca",
  dimensions_to_use = 1:10,
  knn_k = 10,
  prob = FALSE,
  algorithm = c("kd_tree", "cover_tree", "brute"),
  return_gobject = TRUE
)
}
\arguments{
\item{target_gobject}{target giotto object}

\item{target_cluster_label_name}{name for predicted clusters}

\item{spat_unit}{spatial unit}

\item{feat_type}{feature type}

\item{source_gobject}{source giotto object with annotation data}

\item{source_cluster_labels}{annotation/labels to use to train KNN classifier}

\item{reduction}{reduction on cells or features (default = cells)}

\item{reduction_method}{shared reduction method (default = pca space)}

\item{reduction_name}{name of shared reduction space (default name = 'pca')}

\item{dimensions_to_use}{dimensions to use in shared reduction space
(default = 1:10)}

\item{knn_k}{number of k-neighbors to train a KNN classifier}

\item{prob}{output probabilities together with label predictions}

\item{algorithm}{nearest neighbor search algorithm}

\item{return_gobject}{return giotto object}
}
\value{
giotto object (default) or data.table with cell metadata
}
\description{
Use a fast KNN classifier to predict labels from a smaller
giotto object
}
\details{
Function to train a KNN with \code{\link[FNN]{knn}}. The training
data is obtained from the source giotto object (source_gobject) using
existing annotations within the cell metadata. Cells without
annotation/labels from the target giotto object (target_gobject) will
receive predicted labels (and optional probabilities with prob = TRUE).

**IMPORTANT** This projection assumes that you're using the same dimension
reduction space (e.g. PCA) and number of dimensions (e.g. first 10 PCs) to
train the KNN classifier as you used to create the initial
annotations/labels in the source Giotto object.

Altogether this is a convenience function that allow you to work with very
big data as you can predict cell labels on a smaller & subsetted Giotto
object and then project the cell labels to the remaining cells in the target
Giotto object.
}
