% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{doGiottoClustree}
\alias{doGiottoClustree}
\title{doGiottoClustree}
\usage{
doGiottoClustree(
  gobject,
  res_vector = NULL,
  res_seq = NULL,
  return_gobject = FALSE,
  show_plot = NULL,
  save_plot = NULL,
  return_plot = NULL,
  save_param = list(),
  default_save_name = "clustree",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{gobject}{giotto object}

\item{res_vector}{vector of different resolutions to test}

\item{res_seq}{list of float numbers indicating start, end, and step size for resolution testing, i.e. (0.1, 0.6, 0.1)}

\item{return_gobject}{default FALSE. See details for more info.}

\item{show_plot}{by default, pulls from provided gobject instructions}

\item{save_plot}{by default, pulls from provided gobject instructions}

\item{return_plot}{by default, pulls from provided gobject instructions}

\item{save_param}{list of saving parameters from \code{\link[GiottoVisuals:plot_save]{GiottoVisuals::all_plots_save_function()}}}

\item{default_save_name}{name of saved plot, default "clustree"}

\item{verbose}{be verbose}

\item{...}{
  Arguments passed on to \code{\link[clustree:clustree]{clustree::clustree}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
a plot object (default), OR a giotto object (if specified)
}
\description{
cluster cells using leiden methodology to visualize different resolutions
}
\details{
This function tests different resolutions for Leiden clustering and provides a visualization
of cluster sizing as resolution varies.

By default, the tested leiden clusters are NOT saved to the Giotto object, and a plot is returned.

If return_gobject is set to TRUE, and a giotto object with \emph{all} tested leiden cluster information
will be returned.
}
\seealso{
\code{\link{doLeidenCluster}}
}
