% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{doHclust}
\alias{doHclust}
\title{doHclust}
\usage{
doHclust(
  gobject,
  expression_values = c("normalized", "scaled", "custom"),
  genes_to_use = NULL,
  dim_reduction_to_use = c("cells", "pca", "umap", "tsne"),
  dim_reduction_name = "pca",
  dimensions_to_use = 1:10,
  distance_method = c("pearson", "spearman", "original", "euclidean", "maximum",
    "manhattan", "canberra", "binary", "minkowski"),
  agglomeration_method = c("ward.D2", "ward.D", "single", "complete", "average",
    "mcquitty", "median", "centroid"),
  k = 10,
  h = NULL,
  name = "hclust",
  return_gobject = TRUE,
  set_seed = T,
  seed_number = 1234
)
}
\arguments{
\item{gobject}{giotto object}

\item{expression_values}{expression values to use}

\item{genes_to_use}{subset of genes to use}

\item{dim_reduction_to_use}{dimension reduction to use}

\item{dim_reduction_name}{dimensions reduction name}

\item{dimensions_to_use}{dimensions to use}

\item{distance_method}{distance method}

\item{agglomeration_method}{agglomeration method for hclust}

\item{k}{number of final clusters}

\item{h}{cut hierarchical tree at height = h}

\item{name}{name for hierarchical clustering}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}

\item{set_seed}{set seed}

\item{seed_number}{number for seed}
}
\value{
giotto object with new clusters appended to cell metadata
}
\description{
cluster cells using hierarchical clustering algorithm
}
\details{
Description on how to use Kmeans clustering method.
}
\examples{

data(mini_giotto_single_cell)

mini_giotto_single_cell = doHclust(mini_giotto_single_cell, k = 4, name = 'hier_clus')
plotUMAP_2D(mini_giotto_single_cell, cell_color = 'hier_clus', point_size = 3)

}
\seealso{
\code{\link[stats]{hclust}}
}
