% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_hmrf.R
\name{doHMRFextend}
\alias{doHMRFextend}
\title{doHMRFextend}
\usage{
doHMRFextend(gobject, expression_values = c("normalized", "scaled",
  "custom"), spatial_network_name = "spatial_network",
  spatial_genes = NULL, name = "test", k = 10, betas = c(0, 2, 50),
  tolerance = 1e-10, zscore = c("rowcol", "colrow", "None"),
  numinit = 100, python_path = NULL, reader_path = NULL,
  get_result_path = NULL, output_folder = NULL)
}
\arguments{
\item{gobject}{giotto object}

\item{expression_values}{expression values to use}

\item{spatial_network_name}{name of spatial network to use for HMRF}

\item{spatial_genes}{spatial genes to use for HMRF}

\item{name}{name of HMRF run}

\item{k}{number of HMRF domains}

\item{betas}{betas to test for}

\item{tolerance}{tolerance}

\item{zscore}{zscore}

\item{numinit}{number of initializations}

\item{python_path}{python path to use}

\item{reader_path}{path to reader script}

\item{get_result_path}{path to get results script}

\item{output_folder}{output folder to save results}
}
\value{
Creates a directory with results that can be viewed with viewHMRFresults
}
\description{
Run HMRF
}
\details{
Description of HMRF parameters ...
}
\examples{
    doHMRFextend(gobject)
}
