% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{doKmeans}
\alias{doKmeans}
\title{doKmeans}
\usage{
doKmeans(
  gobject,
  feat_type = NULL,
  spat_unit = NULL,
  expression_values = c("normalized", "scaled", "custom"),
  feats_to_use = NULL,
  dim_reduction_to_use = c("cells", "pca", "umap", "tsne"),
  dim_reduction_name = "pca",
  dimensions_to_use = 1:10,
  distance_method = c("original", "pearson", "spearman", "euclidean", "maximum",
    "manhattan", "canberra", "binary", "minkowski"),
  centers = 10,
  iter_max = 100,
  nstart = 1000,
  algorithm = "Hartigan-Wong",
  name = "kmeans",
  return_gobject = TRUE,
  set_seed = TRUE,
  seed_number = 1234
)
}
\arguments{
\item{gobject}{giotto object}

\item{feat_type}{feature type (e.g. "cell")}

\item{spat_unit}{spatial unit (e.g. "rna", "dna", "protein")}

\item{expression_values}{expression values from list_expression()
(e.g. "normalized", "scaled", "custom")}

\item{feats_to_use}{(optional) subset of features to use}

\item{dim_reduction_to_use}{dimension reduction from list_dim_reductions()
(e.g. "pca", "umap", "tsne")}

\item{dim_reduction_name}{dimensions reduction name, default to "pca"}

\item{dimensions_to_use}{dimensions to use, default = 1:10}

\item{distance_method}{distance method (e.g. "original", "pearson",
"spearman", "euclidean", "maximum", "manhattan", "canberra", "binary",
"minkowski")}

\item{centers}{number of final clusters, default = 10}

\item{iter_max}{kmeans maximum iterations, default = 100}

\item{nstart}{kmeans nstart, default = 1000}

\item{algorithm}{kmeans algorithm, default to "Hartigan-Wong"}

\item{name}{name for kmeans clustering, default to "kmeans"}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}

\item{set_seed}{set seed (default = TRUE)}

\item{seed_number}{number for seed}
}
\value{
if return_gobject = TRUE: giotto object with new clusters appended 
to cell metadata
}
\description{
cluster cells using kmeans algorithm
}
\details{
The default settings will use dimension reduction results as input.
Set dim_reduction_to_use = NULL if you want to directly use expression 
values as input.
By providing a feature vector to feats_to_use you can subset the expression 
matrix.
}
\examples{
g <- GiottoData::loadGiottoMini("visium")

doKmeans(g)
}
\seealso{
\code{\link[stats]{kmeans}}
}
