% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{doLeidenCluster}
\alias{doLeidenCluster}
\alias{doLeidenClusterIgraph}
\title{doLeidenCluster}
\usage{
doLeidenClusterIgraph(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  name = "leiden_clus",
  nn_network_to_use = "sNN",
  network_name = "sNN.pca",
  objective_function = c("modularity", "CPM"),
  weights = NULL,
  resolution = 1,
  beta = 0.01,
  initial_membership = NULL,
  n_iterations = 20,
  return_gobject = TRUE,
  set_seed = TRUE,
  seed_number = 1234,
  verbose = NULL,
  ...
)

doLeidenCluster(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  name = "leiden_clus",
  nn_network_to_use = "sNN",
  network_name = "sNN.pca",
  objective_function = c("modularity", "CPM"),
  weights = NULL,
  resolution = 1,
  beta = 0.01,
  initial_membership = NULL,
  n_iterations = 20,
  return_gobject = TRUE,
  set_seed = TRUE,
  seed_number = 1234,
  verbose = NULL,
  ...
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit (e.g. "cell")}

\item{feat_type}{feature type (e.g. "rna", "dna", "protein")}

\item{name}{name for cluster, default to "leiden_clus"}

\item{nn_network_to_use}{type of NN network to use (kNN vs sNN), default to
"sNN"}

\item{network_name}{name of NN network to use, default to "sNN.pca"}

\item{objective_function}{objective function for the leiden algo}

\item{weights}{weights of edges}

\item{resolution}{resolution, default = 1}

\item{beta}{leiden randomness}

\item{initial_membership}{initial membership of cells for the partition}

\item{n_iterations}{number of interations to run the Leiden algorithm.}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}

\item{set_seed}{set seed}

\item{seed_number}{number for seed}

\item{verbose}{verbosity of function}

\item{...}{
  Arguments passed on to \code{\link[igraph:cluster_leiden]{igraph::cluster_leiden}}
  \describe{
    \item{\code{resolution_parameter}}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}} Use \code{resolution} instead.}
    \item{\code{vertex_weights}}{the vertex weights used in the Leiden algorithm.
If this is not provided, it will be automatically determined on the basis
of the \code{objective_function}. Please see the details of this function
how to interpret the vertex weights.}
  }}
}
\value{
giotto object with new clusters appended to cell metadata
}
\description{
cluster cells using a NN-network and the Leiden community
detection algorithm as implemented in igraph
}
\details{
This function is a wrapper for the Leiden algorithm implemented in igraph,
which can detect communities in graphs of millions of nodes (cells),
as long as they can fit in memory. See \code{\link[igraph]{cluster_leiden}}
for more information.

Set \emph{weights = NULL} to use the vertices weights associated with the
igraph network.
Set \emph{weights = NA} if you don't want to use vertices weights
}
\examples{
g <- GiottoData::loadGiottoMini("visium")

doLeidenCluster(g)
}
