% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{doLeidenSubCluster}
\alias{doLeidenSubCluster}
\title{doLeidenSubCluster}
\usage{
doLeidenSubCluster(gobject, cluster_column = NULL,
  hvg_param = list(reverse_log_scale = T, difference_in_variance = 1,
  expression_values = "normalized"), hvg_min_perc_cells = 5,
  hvg_mean_expr_det = 1, pca_param = list(expression_values = "custom",
  scale.unit = T), nn_param = list(dimensions_to_use = 1:20),
  k_neighbors = 10, resolution = 0.5, n_iterations = 500,
  python_path = "/Users/rubendries/Bin/anaconda3/envs/py36/bin/python",
  nn_network_to_use = "sNN", network_name = "sNN.pca",
  name = "sub_pleiden_clus", return_gobject = TRUE, verbose = T, ...)
}
\arguments{
\item{gobject}{giotto object}

\item{cluster_column}{cluster column to subcluster}

\item{hvg_param}{parameters for calculateHVG}

\item{hvg_min_perc_cells}{threshold for detection in min percentage of cells}

\item{hvg_mean_expr_det}{threshold for mean expression level in cells with detection}

\item{pca_param}{parameters for runPCA}

\item{nn_param}{parameters for parameters for runPCA}

\item{k_neighbors}{number of k for sNN}

\item{resolution}{resolution for Leiden clustering}

\item{n_iterations}{number of iterations for Leiden clustering}

\item{python_path}{python path to use for Leiden clustering}

\item{nn_network_to_use}{NN network to use}

\item{network_name}{NN network name}

\item{name}{name of clustering}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}

\item{verbose}{be verbose}

\item{...}{additional parameters}
}
\value{
giotto object appended with new cluster
}
\description{
cluster cells iteratively
}
\details{
Description of iterative clustering.
}
\examples{
    doLeidenSubCluster(gobject)
}
