% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{doLouvainCluster}
\alias{doLouvainCluster}
\title{doLouvainCluster}
\usage{
doLouvainCluster(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  version = c("community", "multinet"),
  name = "louvain_clus",
  nn_network_to_use = "sNN",
  network_name = "sNN.pca",
  python_path = deprecated(),
  resolution = 1,
  weight_col = NULL,
  gamma = 1,
  omega = 1,
  louv_random = FALSE,
  return_gobject = TRUE,
  set_seed = FALSE,
  seed_number = 1234,
  ...
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit (e.g. "cell")}

\item{feat_type}{feature type (e.g. "rna", "dna", "protein")}

\item{version}{implemented version of Louvain clustering to use}

\item{name}{name for cluster, default to "louvain_clus"}

\item{nn_network_to_use}{type of NN network to use (kNN vs sNN), default to
"sNN"}

\item{network_name}{name of NN network to use, default to "sNN.pca"}

\item{python_path}{deprecated.}

\item{resolution}{[community] resolution, default = 1}

\item{weight_col}{weight column name}

\item{gamma}{[multinet] Resolution parameter for modularity in the
generalized louvain method, default = 1}

\item{omega}{[multinet] Inter-layer weight parameter in the generalized
louvain method, default = 1}

\item{louv_random}{[community] Will randomize the node evaluation order and
the community evaluation order to get different partitions at each call
(default = FALSE)}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}

\item{set_seed}{set seed (default = FALSE)}

\item{seed_number}{number for seed}

\item{...}{arguments passed to \code{\link{.doLouvainCluster_community}}}
}
\value{
giotto object with new clusters appended to cell metadata
}
\description{
cluster cells using a NN-network and the Louvain algorithm.
}
\details{
Louvain clustering using the community or multinet implementation
of the louvain clustering algorithm.
}
\examples{
g <- GiottoData::loadGiottoMini("visium")

doLouvainCluster(g)
}
\seealso{
\code{\link{.doLouvainCluster_community}} and
\code{\link{.doLouvainCluster_multinet}}
}
