% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{doLouvainCluster_community}
\alias{doLouvainCluster_community}
\title{doLouvainCluster_community}
\usage{
doLouvainCluster_community(gobject, name = "louvain_clus",
  nn_network_to_use = "sNN", network_name = "sNN.pca",
  python_path = NULL, resolution = 1, weight_col = NULL,
  louv_random = F, return_gobject = TRUE, set_seed = F,
  seed_number = 1234, ...)
}
\arguments{
\item{gobject}{giotto object}

\item{name}{name for cluster}

\item{nn_network_to_use}{type of NN network to use (kNN vs sNN)}

\item{network_name}{name of NN network to use}

\item{python_path}{specify specific path to python if required}

\item{resolution}{resolution}

\item{weight_col}{weight column}

\item{louv_random}{random}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}

\item{set_seed}{set seed}

\item{seed_number}{number for seed}

\item{...}{additional parameters}
}
\value{
giotto object appended with new cluster
}
\description{
cluster cells using a NN-network and the Louvain algorithm from the community module in Python
}
\details{
Description of Leiden clustering method.
}
\examples{
    doLouvainCluster_community(gobject)
}
