% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{doLouvainCluster_multinet}
\alias{doLouvainCluster_multinet}
\title{doLouvainCluster_multinet}
\usage{
doLouvainCluster_multinet(
  gobject,
  name = "louvain_clus",
  nn_network_to_use = "sNN",
  network_name = "sNN.pca",
  gamma = 1,
  omega = 1,
  return_gobject = TRUE,
  set_seed = F,
  seed_number = 1234
)
}
\arguments{
\item{gobject}{giotto object}

\item{name}{name for cluster}

\item{nn_network_to_use}{type of NN network to use (kNN vs sNN)}

\item{network_name}{name of NN network to use}

\item{gamma}{Resolution parameter for modularity in the generalized louvain method.}

\item{omega}{Inter-layer weight parameter in the generalized louvain method.}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}

\item{set_seed}{set seed}

\item{seed_number}{number for seed}
}
\value{
giotto object with new clusters appended to cell metadata
}
\description{
cluster cells using a NN-network and the Louvain algorithm from the multinet package in R.
}
\details{
See \code{\link[multinet]{glouvain_ml}} from the multinet package in R for
more information.
}
\keyword{internal}
