% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{doLouvainRCluster}
\alias{doLouvainRCluster}
\title{doLouvainRCluster}
\usage{
doLouvainRCluster(gobject, nn_network_to_use = "sNN",
  network_name = "sNN.pca", gamma = 1, omega = 1,
  name = "rlouvain_clus", return_gobject = TRUE, set_seed = F,
  seed_number = 1234, ...)
}
\arguments{
\item{gobject}{giotto object}

\item{nn_network_to_use}{type of NN network to use (kNN vs sNN)}

\item{network_name}{name of NN network to use}

\item{gamma}{gamma parameter}

\item{omega}{omega parameter}

\item{name}{name for cluster}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}

\item{set_seed}{set seed}

\item{seed_number}{number for seed}

\item{...}{additional parameters}
}
\value{
giotto object appended with new cluster
}
\description{
cluster cells using a NN-network and the R Louvain community detection algorithm
}
\details{
Description of R Louvain clustering method.
}
\examples{
    doLouvainRCluster(gobject)
}
