% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{doLouvainSubCluster_community}
\alias{doLouvainSubCluster_community}
\title{doLouvainSubCluster_community}
\usage{
doLouvainSubCluster_community(gobject, name = "sub_louvain_comm_clus",
  cluster_column = NULL, selected_clusters = NULL,
  hvg_param = list(reverse_log_scale = T, difference_in_variance = 1,
  expression_values = "normalized"), hvg_min_perc_cells = 5,
  hvg_mean_expr_det = 1, pca_param = list(expression_values = "custom",
  scale.unit = T), nn_param = list(dimensions_to_use = 1:20),
  k_neighbors = 10, resolution = 0.5, python_path = NULL,
  nn_network_to_use = "sNN", network_name = "sNN.pca",
  return_gobject = TRUE, verbose = T, ...)
}
\arguments{
\item{gobject}{giotto object}

\item{name}{name for new clustering result}

\item{cluster_column}{cluster column to subcluster}

\item{selected_clusters}{only do subclustering on these clusters}

\item{hvg_param}{parameters for calculateHVG}

\item{hvg_min_perc_cells}{threshold for detection in min percentage of cells}

\item{hvg_mean_expr_det}{threshold for mean expression level in cells with detection}

\item{pca_param}{parameters for runPCA}

\item{nn_param}{parameters for parameters for createNearestNetwork}

\item{k_neighbors}{number of k for createNearestNetwork}

\item{resolution}{resolution}

\item{python_path}{specify specific path to python if required}

\item{nn_network_to_use}{type of NN network to use (kNN vs sNN)}

\item{network_name}{name of NN network to use}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}

\item{verbose}{verbose}

\item{...}{additional parameters}
}
\value{
giotto object appended with new cluster
}
\description{
subcluster cells using a NN-network and the Louvain community detection algorithm
}
\details{
Description of Leiden clustering method.
}
\examples{
    doLouvainSubCluster_community(gobject)
}
