% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{doLouvainSubCluster_multinet}
\alias{doLouvainSubCluster_multinet}
\title{doLouvainSubCluster_multinet}
\usage{
doLouvainSubCluster_multinet(gobject, cluster_column = NULL,
  selected_clusters = NULL, hvg_param = list(reverse_log_scale = T,
  difference_in_variance = 1, expression_values = "normalized"),
  hvg_min_perc_cells = 5, hvg_mean_expr_det = 1,
  pca_param = list(expression_values = "custom", scale.unit = T),
  nn_param = list(dimensions_to_use = 1:20), k_neighbors = 10,
  gamma = 1, omega = 1, python_path = NULL,
  nn_network_to_use = "sNN", network_name = "sNN.pca",
  name = "sub_louvain_mult_clus", return_gobject = TRUE, verbose = T,
  ...)
}
\arguments{
\item{gobject}{giotto object}

\item{cluster_column}{cluster column to subcluster}

\item{selected_clusters}{only do subclustering on these clusters}

\item{hvg_param}{parameters for calculateHVG}

\item{hvg_min_perc_cells}{threshold for detection in min percentage of cells}

\item{hvg_mean_expr_det}{threshold for mean expression level in cells with detection}

\item{pca_param}{parameters for runPCA}

\item{nn_param}{parameters for parameters for createNearestNetwork}

\item{k_neighbors}{number of k for createNearestNetwork}

\item{gamma}{gamma}

\item{omega}{omega}

\item{python_path}{specify specific path to python if required}

\item{nn_network_to_use}{type of NN network to use (kNN vs sNN)}

\item{network_name}{name of NN network to use}

\item{name}{name for new clustering result}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}

\item{verbose}{verbose}

\item{...}{additional parameters}
}
\value{
giotto object appended with new cluster
}
\description{
subcluster cells using a NN-network and the Louvain multinet detection algorithm
}
\details{
Description of Louvain clustering method.
}
\examples{
    doLouvainSubCluster_multinet(gobject)
}
