% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{doRandomWalkCluster}
\alias{doRandomWalkCluster}
\title{doRandomWalkCluster}
\usage{
doRandomWalkCluster(gobject, nn_network_to_use = "sNN",
  network_name = "sNN.pca", walk_steps = 4, walk_clusters = 10,
  walk_weights = NA, name = "random_walk_clus",
  return_gobject = TRUE, set_seed = F, seed_number = 1234, ...)
}
\arguments{
\item{gobject}{giotto object}

\item{nn_network_to_use}{type of NN network to use (kNN vs sNN)}

\item{network_name}{name of NN network to use}

\item{walk_steps}{number of random steps}

\item{walk_clusters}{number of clusters}

\item{walk_weights}{weight column}

\item{name}{name for cluster}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}

\item{set_seed}{set seed}

\item{seed_number}{number for seed}

\item{...}{additional parameters}
}
\value{
giotto object appended with new cluster
}
\description{
cluster cells using a NN-network and the R Louvain community detection algorithm
}
\details{
Description of igraph Random Walk clustering method.
}
\examples{
    doRandomWalkCluster(gobject)
}
