% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{doSNNCluster}
\alias{doSNNCluster}
\title{doSNNCluster}
\usage{
doSNNCluster(gobject, nn_network_to_use = "kNN",
  network_name = "kNN.pca", k = 20, eps = 4, minPts = 16,
  borderPoints = TRUE, name = "sNN_clus", return_gobject = TRUE,
  set_seed = F, seed_number = 1234, ...)
}
\arguments{
\item{gobject}{giotto object}

\item{nn_network_to_use}{type of NN network to use (kNN vs sNN)}

\item{network_name}{name of NN network to use}

\item{k}{number of clusters}

\item{eps}{eps parameter dbscan}

\item{minPts}{minimum Points}

\item{borderPoints}{include points at border}

\item{name}{name for cluster}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}

\item{set_seed}{set seed}

\item{seed_number}{number for seed}

\item{...}{additional parameters}
}
\value{
giotto object appended with new cluster
}
\description{
cluster cells using a NN-network and the SNNclust community detection algorithm
}
\details{
Description of SNNclust clustering method.
}
\examples{
    doSNNCluster(gobject)
}
