% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction.R
\name{.average_feat_feat_expression_in_groups}
\alias{.average_feat_feat_expression_in_groups}
\title{average_feat_feat_expression_in_groups}
\usage{
.average_feat_feat_expression_in_groups(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  cluster_column = "cell_types",
  feat_set_1,
  feat_set_2
)
}
\arguments{
\item{gobject}{giotto object to use}

\item{spat_unit}{spatial unit}

\item{feat_type}{feature type}

\item{cluster_column}{cluster column with cell type information}

\item{feat_set_1}{first specific feat set from feat pairs}

\item{feat_set_2}{second specific feat set from feat pairs}
}
\value{
data.table with average expression scores for each cluster
}
\description{
calculate average expression per cluster
}
\keyword{internal}
