% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction_spot.R
\name{.cal_diff_per_interaction}
\alias{.cal_diff_per_interaction}
\title{Calculate difference per interaction}
\usage{
.cal_diff_per_interaction(
  sel_int,
  other_ints,
  select_ind,
  other_ind,
  proximityMat,
  expr_residual
)
}
\value{
data.table
}
\description{
calculate correlation between expression residual and
cell proximity score of selected cell for spots
}
\keyword{internal}
