% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction_spot.R
\name{.cal_expr_residual}
\alias{.cal_expr_residual}
\title{Calculate feature expression residual}
\usage{
.cal_expr_residual(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  expression_values = c("normalized", "scaled", "custom"),
  ave_celltype_exp
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit (e.g. 'cell')}

\item{feat_type}{feature type (e.g. 'rna')}

\item{expression_values}{expression values to use 
(e.g. 'normalized', 'scaled', 'custom')}

\item{ave_celltype_exp}{average expression matrix in cell types}
}
\value{
matrix
}
\description{
Calculate feature expression residual 
(observed_exp - DWLS_predicted)
}
\keyword{internal}
