% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction.R
\name{.combineInteractionChangedFeatures_per_interaction}
\alias{.combineInteractionChangedFeatures_per_interaction}
\title{Combine ICF scores per interaction}
\usage{
.combineInteractionChangedFeatures_per_interaction(
  icfObject,
  sel_int,
  selected_feats = NULL,
  specific_feats_1 = NULL,
  specific_feats_2 = NULL,
  min_cells = 5,
  min_int_cells = 3,
  min_fdr = 0.05,
  min_spat_diff = 0,
  min_log2_fc = 0.5
)
}
\description{
Combine ICF scores per interaction
}
\keyword{internal}
