% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{.createGiottoCosMxObject_subcellular}
\alias{.createGiottoCosMxObject_subcellular}
\title{Load and create a CosMx Giotto object from subcellular info}
\usage{
.createGiottoCosMxObject_subcellular(
  dir_items,
  FOVs = NULL,
  remove_background_polygon = TRUE,
  background_algo = c("range"),
  remove_unvalid_polygons = TRUE,
  cores,
  verbose = TRUE,
  instructions = NULL
)
}
\arguments{
\item{FOVs}{field of views to load (only affects subcellular data and images)}

\item{remove_background_polygon}{try to remove background polygon 
(default: FALSE)}

\item{background_algo}{algorithm to remove background polygon}

\item{remove_unvalid_polygons}{remove unvalid polygons (default: TRUE)}

\item{cores}{how many cores or threads to use to read data if paths are
provided}

\item{verbose}{be verbose when building Giotto object}

\item{instructions}{list of instructions or output result
from \code{\link[GiottoClass]{createGiottoInstructions}}}
}
\value{
giotto object
}
\description{
Load and create a CosMx Giotto object from subcellular info
}
\keyword{internal}
