% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{.createGiottoXeniumObject_subcellular}
\alias{.createGiottoXeniumObject_subcellular}
\title{Create a Xenium Giotto object from subcellular info}
\usage{
.createGiottoXeniumObject_subcellular(
  data_list,
  key_list = NULL,
  qv_threshold = 20,
  instructions = NULL,
  cores = NA,
  verbose = TRUE
)
}
\arguments{
\item{data_list}{list of data loaded by \code{\link{.load_xenium_folder}}}

\item{key_list}{regex-based search keys for feature IDs to allow separation
into separate giottoPoints objects by feat_type}

\item{qv_threshold}{Minimum Phred-scaled quality score cutoff to be included
as a subcellular transcript detection (default = 20)}

\item{instructions}{list of instructions or output result
from \code{\link[GiottoClass]{createGiottoInstructions}}}

\item{cores}{how many cores or threads to use to read data if paths are
provided}

\item{verbose}{be verbose when building Giotto object}
}
\value{
giotto object
}
\description{
Subcellular workflow for createGiottoXeniumObject
}
\seealso{
createGiottoXeniumObject .createGiottoXeniumObject_aggregate
}
\keyword{internal}
