% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_influence.R
\name{.determine_switch_string_equal}
\alias{.determine_switch_string_equal}
\title{Determine switch string equal}
\usage{
.determine_switch_string_equal(
  cell_meta = NULL,
  cell_meta_new = NULL,
  clus_name = NULL
)
}
\arguments{
\item{cell_meta}{cell_metadata from the original spatial_unit}

\item{cell_meta_new}{cell_metadata from the resized spatial unit}

\item{clus_name}{name of the cluster type, likely "kmeans"}
}
\value{
switch_str, a vector of corresponding cluster numbers in strings
}
\description{
Determine switch string equal
}
\details{
creates a string in the format c("x_1-y_1", "x_2-y_2"..."x_n, y_n")
Where:
   x_n is a cluster number from the original spatial unit
   y_m is a cluster number from the resized spatial unit
   n is the number of clusters

Clusters are determined to be corresponding based on % overlap in cell_IDs 
in each cluster.
}
\keyword{internal}
