% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{.doLouvainCluster_community}
\alias{.doLouvainCluster_community}
\title{doLouvainCluster community}
\usage{
.doLouvainCluster_community(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  name = "louvain_clus",
  nn_network_to_use = "sNN",
  network_name = "sNN.pca",
  python_path = NULL,
  resolution = 1,
  weight_col = NULL,
  louv_random = FALSE,
  return_gobject = TRUE,
  set_seed = FALSE,
  seed_number = 1234,
  ...
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit (e.g. "cell")}

\item{feat_type}{feature type (e.g. "rna", "dna", "protein")}

\item{name}{name for cluster, default to "louvain_clus"}

\item{nn_network_to_use}{type of NN network to use (kNN vs sNN), default to 
"sNN"}

\item{network_name}{name of NN network to use, default to "sNN.pca"}

\item{python_path}{specify specific path to python if required}

\item{resolution}{resolution, default = 1}

\item{weight_col}{weight column to use for edges}

\item{louv_random}{Will randomize the node evaluation order and the 
community evaluation order to get different partitions at each call 
(default = FALSE)}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}

\item{set_seed}{set seed (default = FALSE)}

\item{seed_number}{number for seed}

\item{\dots}{additional params to pass}
}
\value{
giotto object with new clusters appended to cell metadata
}
\description{
cluster cells using a NN-network and the Louvain algorithm 
from the community module in Python
}
\details{
This function is a wrapper for the Louvain algorithm implemented in 
Python, which can detect communities in graphs of nodes (cells).
See the 
\url{https://python-louvain.readthedocs.io/en/latest/index.html}{readthedocs}
page for more information.

Set \emph{weight_col = NULL} to give equal weight (=1) to each edge.
}
\keyword{internal}
