% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{.doLouvainCluster_multinet}
\alias{.doLouvainCluster_multinet}
\title{doLouvainCluster multinet}
\usage{
.doLouvainCluster_multinet(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  name = "louvain_clus",
  nn_network_to_use = "sNN",
  network_name = "sNN.pca",
  gamma = 1,
  omega = 1,
  return_gobject = TRUE,
  set_seed = FALSE,
  seed_number = 1234
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit (e.g. "cell")}

\item{feat_type}{feature type (e.g. "rna", "dna", "protein")}

\item{name}{name for cluster, default to "louvain_clus"}

\item{nn_network_to_use}{type of NN network to use (kNN vs sNN), default to "sNN"}

\item{network_name}{name of NN network to use, default to "sNN.pca"}

\item{gamma}{Resolution parameter for modularity in the generalized louvain method. default  = 1}

\item{omega}{Inter-layer weight parameter in the generalized louvain method. default = 1}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}

\item{set_seed}{set seed (default = FALSE)}

\item{seed_number}{number for seed}
}
\value{
giotto object with new clusters appended to cell metadata
}
\description{
cluster cells using a NN-network and the Louvain algorithm from the multinet package in R.
}
\details{
See \code{\link[multinet]{glouvain_ml}} from the multinet package in R for
more information.
}
\keyword{internal}
