% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registration.R
\name{.estimate_transform_from_matched_descriptor}
\alias{.estimate_transform_from_matched_descriptor}
\title{Estimate affine transformation from matched descriptor}
\usage{
.estimate_transform_from_matched_descriptor(
  keypoints1,
  keypoints2,
  match,
  estimate_fun,
  ...,
  pkg_ptr
)
}
\arguments{
\item{keypoints1}{keypoints extracted from target image via .sift_detect}

\item{match}{a 2 col matrix of x to y index matched descriptors via .match_descriptor_single}
}
\value{
a list of model and inliners
}
\description{
Estimate affine transformation from matched descriptor
}
