% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_help.R
\name{.h5_read_bare}
\alias{.h5_read_bare}
\title{Read dataset from opened HDF5 with C functions}
\usage{
.h5_read_bare(file, name = "", dapl)
}
\arguments{
\item{file}{opened HDF5 file}

\item{name}{dataset name within}

\item{dapl}{HDF5 property list (H5Pcreate('H5P_DATASET_ACCESS'))}
}
\value{
HDF5 contents
}
\description{
Read dataset from opened HDF5 with C functions
}
\keyword{internal}
