% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_clusters.R
\name{.igraph_vertex_membership}
\alias{.igraph_vertex_membership}
\title{igraph vertex membership}
\usage{
.igraph_vertex_membership(g, clus_name, all_ids = NULL, missing_id_name)
}
\arguments{
\item{g}{igraph}

\item{clus_name}{character. name to assign column of clustering info}

\item{all_ids}{(optional) character vector with all ids}

\item{missing_id_name}{character and name for vertices that were missing from g}
}
\value{
data.table
}
\description{
Get which weakly connected set of vertices each vertex is part of
}
\keyword{internal}
