% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{.lib_norm_giotto}
\alias{.lib_norm_giotto}
\title{Normalize expression matrix for library size}
\usage{
.lib_norm_giotto(mymatrix, scalefactor)
}
\arguments{
\item{mymatrix}{matrix object}

\item{scalefactor}{scalefactor}
}
\value{
matrix
}
\description{
Normalize expression matrix for library size
}
\keyword{internal}
