% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{.log_norm_giotto}
\alias{.log_norm_giotto}
\title{Log normalize expression matrix}
\usage{
.log_norm_giotto(mymatrix, base, offset)
}
\description{
Log normalize expression matrix
}
\keyword{internal}
