% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registration.R
\name{.match_descriptor}
\alias{.match_descriptor}
\title{Match image descriptors}
\usage{
.match_descriptor(
  descriptor_list,
  target_idx = 1L,
  cross_check = TRUE,
  max_ratio = 0.8,
  ...,
  pkg_ptr
)
}
\arguments{
\item{descriptor_list}{list of descriptor matrices}

\item{target_idx}{which item in the list is the target image. Default is 1}

\item{cross_check}{whether to check that only the best match is returned}

\item{max_ratio}{Maximum ratio of distances between first and second closest
descriptor in the second set of descriptors. This threshold is useful to
filter ambiguous matches between the two descriptor sets. The choice of this
value depends on the statistics of the chosen descriptor, e.g., for SIFT
descriptors a value of 0.8 is usually chosen, see D.G. Lowe, "Distinctive
Image Features from Scale-Invariant Keypoints", International Journal of
Computer Vision, 2004.}

\item{...}{additional params to pass to `skimage.feature.match_descriptors()`}
}
\value{
list
}
\description{
Brute force matching of descriptors using \pkg{scikit-image}. Find matching
image descriptors between moving images and a target image.
}
