% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{.mean_expr_det_test}
\alias{.mean_expr_det_test}
\title{Mean expression detected test}
\usage{
.mean_expr_det_test(mymatrix, detection_threshold = 1)
}
\arguments{
\item{mymatrix}{matrix of expression info}

\item{detection_threshold}{detection threshold. Defaults to 1 count.}
}
\value{
numeric
}
\description{
Mean expression detected test
}
\keyword{internal}
