% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registration.R
\name{.plot_matched_descriptors}
\alias{.plot_matched_descriptors}
\title{plot matched descriptors}
\usage{
.plot_matched_descriptors(x, y, keypoints1, keypoints2, match, pkg_ptr)
}
\arguments{
\item{x}{source image from .sift_preprocess}

\item{y}{target image from .sift_preprocess}

\item{keypoints1}{keypoints extracted from target image via .sift_detect}

\item{match}{a 2 col matrix of x to y index matched descriptors via 
.match_descriptor_single}
}
\value{
None
}
\description{
A wrapper function for the plot_matches for the SIFT feature extractor and 
descriptor pipeline
}
