% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_cosmx.R
\name{.read_cosmx_folder}
\alias{.read_cosmx_folder}
\title{Read a structured CosMx folder}
\usage{
.read_cosmx_folder(cosmx_dir, verbose = TRUE)
}
\arguments{
\item{cosmx_dir}{full path to the exported cosmx directory}

\item{verbose}{be verbose when building Giotto object}
}
\value{
path_list a list of cosmx files discovered and their filepaths. NULL
values denote missing items
}
\description{
Read a structured CosMx folder
}
\seealso{
createGiottoCosMxObject load_cosmx_folder
}
\keyword{internal}
