% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{.read_xenium_folder}
\alias{.read_xenium_folder}
\title{Read a structured xenium folder}
\usage{
.read_xenium_folder(
  xenium_dir,
  data_to_use = "subcellular",
  bounds_to_load = c("cell"),
  load_format = "csv",
  h5_expression = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{xenium_dir}{full path to the exported xenium directory}

\item{data_to_use}{which type(s) of expression data to build the gobject with
(e.g. default: \strong{'subcellular'}, 'aggregate', or 'all')}

\item{bounds_to_load}{vector of boundary information to load (e.g. \code{'cell'}
or \code{'nucleus'} by themselves or \code{c('cell', 'nucleus')} to load both
at the same time.)}

\item{load_format}{files formats from which to load the data. Either `csv` or
`parquet` currently supported.}

\item{h5_expression}{(boolean) whether to load cell_feature_matrix from .h5 file.
Default is \code{TRUE}}

\item{verbose}{be verbose when building Giotto object}
}
\value{
path_list a list of xenium files discovered and their filepaths. NULL
values denote missing items
}
\description{
Read a structured xenium folder
}
\keyword{internal}
