% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{.rna_pears_resid_normalization}
\alias{.rna_pears_resid_normalization}
\title{RNA pearson residuals normalization}
\usage{
.rna_pears_resid_normalization(
  gobject,
  raw_expr,
  feat_type,
  spat_unit,
  theta = 100,
  name = "scaled",
  verbose = TRUE
)
}
\value{
giotto object
}
\description{
function for RNA normalization according to Lause/Kobak et al
paper
Adapted from https://gist.github.com/hypercompetent/51a3c428745e1c06d826d76c3671797c#file-pearson_residuals-r
}
\keyword{internal}
