% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{.rna_standard_normalization}
\alias{.rna_standard_normalization}
\title{RNA standard normalization}
\usage{
.rna_standard_normalization(
  gobject,
  raw_expr,
  feat_type,
  spat_unit,
  library_size_norm = TRUE,
  scalefactor = 6000,
  log_norm = TRUE,
  log_offset = 1,
  logbase = 2,
  scale_feats = TRUE,
  scale_cells = TRUE,
  scale_order = c("first_feats", "first_cells"),
  verbose = TRUE
)
}
\value{
giotto object
}
\description{
standard function for RNA normalization
}
\keyword{internal}
