% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.R
\name{.run_pca_biocsingular}
\alias{.run_pca_biocsingular}
\title{Run PCA - BiocSingular}
\usage{
.run_pca_biocsingular(
  x,
  ncp = 100,
  center = TRUE,
  scale = TRUE,
  rev = FALSE,
  set_seed = TRUE,
  seed_number = 1234,
  BSPARAM = c("irlba", "exact", "random"),
  BPPARAM = BiocParallel::SerialParam(),
  ...
)
}
\arguments{
\item{x}{matrix or object that can be converted to matrix}

\item{ncp}{number of principal components to calculate}

\item{center}{center the matrix before pca}

\item{scale}{scale features}

\item{rev}{reverse PCA}

\item{set_seed}{use of seed}

\item{seed_number}{seed number to use}

\item{BSPARAM}{method to use}

\item{BPPARAM}{BiocParallelParam object}
}
\value{
list of eigenvalues, loadings and pca coordinates
}
\description{
Performs PCA based on the biocSingular package
}
\keyword{internal}
