% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.R
\name{.run_pca_biocsingular_irlba_projection}
\alias{.run_pca_biocsingular_irlba_projection}
\title{Run PCA - BiocSingular irlba projection}
\usage{
.run_pca_biocsingular_irlba_projection(
  x,
  ncp = 100,
  center = TRUE,
  scale = TRUE,
  rev = FALSE,
  set_seed = TRUE,
  seed_number = 1234,
  BPPARAM = BiocParallel::SerialParam(),
  random_subset = 500,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{matrix or object that can be converted to matrix}

\item{ncp}{number of principal components to calculate}

\item{center}{center the matrix before pca}

\item{scale}{scale features}

\item{rev}{reverse PCA}

\item{set_seed}{use of seed}

\item{seed_number}{seed number to use}

\item{BPPARAM}{BiocParallelParam object}

\item{random_subset}{random subset to perform PCA on}

\item{verbose}{verbosity level}
}
\value{
list of eigenvalues, loadings and pca coordinates
}
\description{
Performs PCA based on the biocSingular package on a
subset of the matrix. It uses the obtained loadings to predicted coordinates
for the remaining matrix.
}
\keyword{internal}
