% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_enrichment.R
\name{.run_spat_autocor_local}
\alias{.run_spat_autocor_local}
\title{.run_spat_autocor_local}
\usage{
.run_spat_autocor_local(
  use_values,
  feats,
  weight_matrix,
  method,
  test_method,
  IDs
)
}
\value{
data.table
}
\description{
.run_spat_autocor_local
}
\keyword{internal}
