% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registration.R
\name{.sift_detect}
\alias{.sift_detect}
\title{Run SIFT feature detector and descriptor extractor}
\usage{
.sift_detect(x, ..., pkg_ptr)
}
\arguments{
\item{x}{input matrix or preprocessed image to extract feature and 
descriptor from}

\item{...}{additional params to pass to `skimage.feature.SIFT()`}
}
\value{
list of keypoints and descriptors
}
\description{
Perform feature detector and descriptor extractor on a matrix object or 
preprocessed image object
}
