% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_general.R
\name{.visium_spot_poly}
\alias{.visium_spot_poly}
\title{Create Polygons for Visium Data}
\usage{
.visium_spot_poly(spatlocs = NULL, json_scalefactors)
}
\arguments{
\item{spatlocs}{spatial locations data.table or \code{spatLocsObj} containing
centroid locations of visium spots}

\item{json_scalefactors}{list of scalefactors from
.visium_read_scalefactors()}
}
\value{
giottoPolygon object
}
\description{
Create Polygons for Visium Data
}
\details{
Creates circular polygons for spatial representation of
Visium spots.
}
\keyword{internal}
