% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registration.R
\name{.warp_transformed_image}
\alias{.warp_transformed_image}
\title{Warp transformed images from estimated transformation}
\usage{
.warp_transformed_image(x, y, model, outpath = NULL, pkg_ptr)
}
\arguments{
\item{x}{source image from .sift_preprocess}

\item{y}{target image from .sift_preprocess}

\item{model}{estimated transformation object from .estimate_transform_from_matched_descriptor}
}
\value{
None, it will write to a output path
}
\description{
Warp transformed images from estimated transformation
}
