% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_structures.R
\name{dt_to_spatVector_points}
\alias{dt_to_spatVector_points}
\title{Convert point data data.table to spatVector}
\usage{
dt_to_spatVector_points(dt, include_values = TRUE, specific_values = NULL)
}
\arguments{
\item{dt}{data.table}

\item{include_values}{include additional values from data.table as attributes paired with created terra spatVector [boolean]}

\item{specific_values}{specific values to include as attributes if include_values == TRUE}
}
\description{
data.table to spatVector for points
}
\keyword{internal}
