% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_structures.R
\name{dt_to_spatVector_polygon}
\alias{dt_to_spatVector_polygon}
\title{Convert data.table to polygon spatVector}
\usage{
dt_to_spatVector_polygon(dt, include_values = TRUE, specific_values = NULL)
}
\description{
convert data.table to spatVector for polygons
}
\keyword{internal}
