% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_enrichment.R
\name{enrich_deconvolution}
\alias{enrich_deconvolution}
\title{enrich_deconvolution}
\usage{
enrich_deconvolution(expr, log_expr, cluster_info, ct_exp, cutoff)
}
\description{
Rui to fill in
}
\keyword{internal}
