% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_genes.R
\name{enrichSpatialCorGroups}
\alias{enrichSpatialCorGroups}
\title{enrichSpatialCorGroups}
\usage{
enrichSpatialCorGroups(
  gobject,
  spatCorObject,
  expression_values = c("normalized", "scaled", "custom"),
  use_clus_name = NULL,
  select_clusters = NULL,
  name = "spatclus_enr",
  convert_enrich_to_cluster = FALSE,
  enrich_to_cluster_name = "enrich_cluster",
  return_gobject = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{spatCorObject}{show the accompanying scatter plot}

\item{expression_values}{gene expression values to use}

\item{use_clus_name}{name of clusters to visualize (from clusterSpatialCorGenes())}

\item{select_clusters}{subset of clusters to use}

\item{name}{name of the spatial enrichment results}

\item{convert_enrich_to_cluster}{create clusters based on enrichment scores}

\item{enrich_to_cluster_name}{name for clusters based on enrichment scores}

\item{return_gobject}{return giotto object}
}
\value{
giotto object
}
\description{
Create enrichment scores based on the metagene expression of the spatially correlated gene groups.
}
\details{
This function calculates the metagene expression for each identified cluster
of spatially correlated genes (use_clus_name) and provides a new enrichment score
(name) that can be visualized. Additionally, these spatial gene correlation enrichment
scores can be converted into clusters by selecting the highest z-score per cell after
z-scoring first columns and then rows.
}
\examples{
    enrichSpatialCorGroups(gobject)
}
