% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registration.R
\name{estimateAutomatedImageRegistrationWithSIFT}
\alias{estimateAutomatedImageRegistrationWithSIFT}
\title{Estimate Automated ImageRegistration With SIFT}
\usage{
estimateAutomatedImageRegistrationWithSIFT(
  x,
  y,
  plot_match = F,
  max_ratio = 0.6,
  estimate_fun = "affine",
  save_warp = NULL,
  verbose = T
)
}
\arguments{
\item{x}{required. Source matrix input, could be generated from preprocessImageToMatrix}

\item{y}{required. Source matrix input, could be generated from preprocessImageToMatrix}

\item{plot_match}{whether or not to plot the matching descriptors.Default False}

\item{max_ratio}{max_ratio parameter for matching descriptors, default 0.6}

\item{estimate_fun}{default Affine. The transformation model to use estimation}

\item{save_warp}{default NULL, if not NULL, please provide an output image path to save the warpped image.}
}
\value{
a list of the estimated transformation object
example estimation <- estimateAutomatedImageRegistrationWithSIFT(x = image_mtx1,y = image_mtx2)
}
\description{
Automatically estimate a transform with SIFT feature detection, descriptor match and returns a transformation object to use
}
