% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_section.R
\name{estimateCellCellDistance}
\alias{estimateCellCellDistance}
\title{estimateCellCellDistance}
\usage{
estimateCellCellDistance(
  gobject,
  spatial_network_name = "Delaunay_network",
  method = c("mean", "median")
)
}
\arguments{
\item{gobject}{gobject}

\item{spatial_network_name}{spatial_network_name}

\item{method}{method}
}
\description{
estimate average distance between neighboring cells
}
\keyword{internal}
