% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{estimateImageBg}
\alias{estimateImageBg}
\title{estimateImageBg}
\usage{
estimateImageBg(mg_object, top_color_range = 1:50)
}
\arguments{
\item{mg_object}{magick image or Giotto image object}

\item{top_color_range}{top possible background colors to return}
}
\value{
vector of pixel color frequencies and an associated barplot
}
\description{
helps to estimate which color is the background color of your plot
}
