% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto.R
\name{evaluate_expr_matrix}
\alias{evaluate_expr_matrix}
\title{evaluate_expr_matrix}
\usage{
evaluate_expr_matrix(inputmatrix, sparse = TRUE)
}
\arguments{
\item{inputmatrix}{inputmatrix to evaluate}
}
\value{
sparse matrix
}
\description{
Evaluate expression matrices.
}
\details{
The inputmatrix can be a matrix, sparse matrix, data.frame, data.table or path to any of these.
}
\examples{
    evaluate_expr_matrix()
}
