% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto.R
\name{evaluate_feat_info}
\alias{evaluate_feat_info}
\title{Evaluate feature info}
\usage{
evaluate_feat_info(spatial_feat_info, feat_type, cores = 1, feat_ID)
}
\arguments{
\item{spatial_feat_info}{spatial feature information to evaluate}

\item{cores}{how many cores to use}

\item{feat_ID}{feature IDs to check with}
}
\value{
data.table
}
\description{
Evaluate spatial feature information input
}
\keyword{internal}
