% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_genes.R
\name{evaluate_provided_spatial_locations}
\alias{evaluate_provided_spatial_locations}
\title{Evaluate provided spatial locations}
\usage{
evaluate_provided_spatial_locations(spatial_locs)
}
\description{
Evaluate provided spatial locations
}
\keyword{internal}
