% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto.R
\name{evaluate_spatial_locations}
\alias{evaluate_spatial_locations}
\title{evaluate_spatial_locations}
\usage{
evaluate_spatial_locations(spatial_locs, cores = 1, dummy_n = 2)
}
\arguments{
\item{spatial_locs}{spatial locations to evaluate}

\item{cores}{how many cores to use}
}
\value{
data.table
}
\description{
Evaluate spatial location input
}
\keyword{internal}
