% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_structures.R
\name{evaluate_spatial_network}
\alias{evaluate_spatial_network}
\title{Evaluate spatial network}
\usage{
evaluate_spatial_network(spatial_network)
}
\description{
function to evaluate a spatial network
}
\keyword{internal}
