% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_in_situ_visuals.R
\name{expand_feature_info}
\alias{expand_feature_info}
\title{expand_feature_info}
\usage{
expand_feature_info(
  spatial_feat_info,
  expand_counts = FALSE,
  count_info_column = "count",
  jitter = c(0, 0),
  verbose = TRUE
)
}
\value{
data.table
}
\description{
low level function to expand feature coordinates
}
\keyword{internal}
