% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto.R
\name{explainGiottoInstructions}
\alias{explainGiottoInstructions}
\title{explainGiottoInstructions}
\usage{
explainGiottoInstructions()
}
\value{
Instructions in txt format
}
\description{
Simple function to explain how to use \code{\link{createGiottoInstructions}}
}
