% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction.R
\name{exprCellCellcom}
\alias{exprCellCellcom}
\title{exprCellCellcom}
\usage{
exprCellCellcom(
  gobject,
  feat_type = NULL,
  spat_unit = NULL,
  cluster_column = "cell_types",
  random_iter = 1000,
  feat_set_1,
  feat_set_2,
  log2FC_addendum = 0.1,
  detailed = FALSE,
  adjust_method = c("fdr", "bonferroni", "BH", "holm", "hochberg", "hommel", "BY",
    "none"),
  adjust_target = c("feats", "cells"),
  set_seed = TRUE,
  seed_number = 1234,
  verbose = TRUE
)
}
\arguments{
\item{gobject}{giotto object to use}

\item{feat_type}{feature type}

\item{spat_unit}{spatial unit}

\item{cluster_column}{cluster column with cell type information}

\item{random_iter}{number of iterations}

\item{feat_set_1}{first specific feature set from feature pairs}

\item{feat_set_2}{second specific feature set from feature pairs}

\item{log2FC_addendum}{addendum to add when calculating log2FC}

\item{detailed}{provide more detailed information (random variance and z-score)}

\item{adjust_method}{which method to adjust p-values}

\item{adjust_target}{adjust multiple hypotheses at the cell or feature level}

\item{set_seed}{set seed for random simulations (default = TRUE)}

\item{seed_number}{seed number}

\item{verbose}{verbose}
}
\value{
Cell-Cell communication scores for feature pairs based on expression only
}
\description{
Cell-Cell communication scores based on expression only
}
\details{
Statistical framework to identify if pairs of features (such as ligand-receptor combinations)
are expressed at higher levels than expected based on a reshuffled null distribution of feature expression values,
without considering the spatial position of cells.
More details will follow soon.
}
