% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{exprObj-class}
\alias{exprObj-class}
\title{S4 exprObj}
\description{
Framework to store aggregated expression information
}
\section{Slots}{

\describe{
\item{\code{name}}{name of exprObj}

\item{\code{exprMat}}{matrix of expression information}

\item{\code{spat_unit}}{spatial unit of expression (e.g. 'cell')}

\item{\code{feat_type}}{feature type of expression (e.g. 'rna', 'protein')}

\item{\code{provenance}}{origin data of expression information (if applicable)}

\item{\code{misc}}{misc}
}}

