% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction.R
\name{exprOnlyCellCellcommunicationScores}
\alias{exprOnlyCellCellcommunicationScores}
\title{exprOnlyCellCellcommunicationScores}
\usage{
exprOnlyCellCellcommunicationScores(gobject,
  cluster_column = "cell_types", random_iter = 100, gene_set_1,
  gene_set_2, log2FC_addendum = 0.1, verbose = T)
}
\arguments{
\item{gobject}{giotto object to use}

\item{cluster_column}{cluster column with cell type information}

\item{random_iter}{number of iterations}

\item{gene_set_1}{first specific gene set from gene pairs}

\item{gene_set_2}{second specific gene set from gene pairs}

\item{log2FC_addendum}{addendum to add when calculating log2FC}

\item{verbose}{verbose}
}
\value{
Cell-Cell communication scores for gene pairs based on expression only
}
\description{
Cell-Cell communication scores based on expression only
}
\details{
Details will follow.
}
\examples{
    exprOnlyCellCellcommunicationScores(gobject)
}
