% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_manipulation.R
\name{ext-generic}
\alias{ext-generic}
\alias{ext,spatLocsObj-method}
\alias{ext,giottoPolygon-method}
\alias{ext,giottoPoints-method}
\alias{ext,spatialNetworkObj-method}
\title{Get a SpatExtent}
\usage{
\S4method{ext}{spatLocsObj}(x, ...)

\S4method{ext}{giottoPolygon}(x, ...)

\S4method{ext}{giottoPoints}(x, ...)

\S4method{ext}{spatialNetworkObj}(x, ...)
}
\arguments{
\item{x}{spatial object}

\item{...}{additional params to pass}
}
\description{
Get a SpatExtent of an object. This is the spatial minmax x and y
that the object is mapped to.
}
\section{Functions}{
\itemize{
\item \code{ext(spatLocsObj)}: Get extent of spatLocsObj

\item \code{ext(giottoPolygon)}: Get extent of giottoPolygon

\item \code{ext(giottoPoints)}: Get extent of giottoPoints

\item \code{ext(spatialNetworkObj)}: Get extent of spatialNetworkObj

}}
