% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_section.R
\name{extend_vector}
\alias{extend_vector}
\title{extend_vector}
\usage{
extend_vector(x, extend_ratio)
}
\arguments{
\item{x}{x}

\item{extend_ratio}{extend_ratio}
}
\description{
extend the range of a vector by a given ratio
}
\keyword{internal}
