% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_help.R
\name{extended_gini_fun}
\alias{extended_gini_fun}
\title{extended_gini_fun}
\usage{
extended_gini_fun(x, weights = rep(1, length = length(x)), minimum_length = 16)
}
\value{
gini coefficient
}
\description{
calculate gini coefficient on a minimum length vector
}
\keyword{internal}
