% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{extract-generic}
\alias{extract-generic}
\alias{$,coordDataDT-method}
\alias{$<-,coordDataDT-method}
\alias{[,coordDataDT,missing,ANY,missing-method}
\alias{[,coordDataDT,ANY,ANY,missing-method}
\alias{[,coordDataDT,missing,missing,missing-method}
\alias{[,metaData,missing,ANY,missing-method}
\alias{[,metaData,ANY,ANY,missing-method}
\alias{[,metaData,missing,missing,missing-method}
\alias{[,dimObj,ANY,ANY,missing-method}
\title{Extract or replace parts of an object}
\usage{
\S4method{$}{coordDataDT}(x, name)

\S4method{$}{coordDataDT}(x, name) <- value

\S4method{[}{coordDataDT,missing,ANY,missing}(x, i, j)

\S4method{[}{coordDataDT,missing,ANY,missing}(x, i, j)

\S4method{[}{coordDataDT,ANY,ANY,missing}(x, i, j)

\S4method{[}{coordDataDT,missing,missing,missing}(x, i, j)

\S4method{[}{metaData,missing,ANY,missing}(x, i, j)

\S4method{[}{metaData,missing,ANY,missing}(x, i, j)

\S4method{[}{metaData,ANY,ANY,missing}(x, i, j)

\S4method{[}{metaData,missing,missing,missing}(x, i, j)

\S4method{[}{dimObj,ANY,ANY,missing}(x, i, j)
}
\arguments{
\item{x}{Giotto S4 object to extract columns from}

\item{name}{A literal character string (possibly backtick quoted).}

\item{value}{value(s) to set
This is normally matched to the colnames of the data.table object within the S4.}

\item{i, j}{indices specifying elements to extract or replace. Indices are
numeric or character vectors or empty}
}
\description{
Operators Giotto S4 internal data.tables to extract
or replace parts.
}
\section{\code{`$`} methods}{



  Select by colname from giotto S4 data.table coordinates slot.
  Works via data.table methods
}

\section{\code{`$<-`} methods}{



  Set values by colname into giotto S4 data.table coordinates slot.
  Works via data.table methods
}

\section{\code{`[`} methods}{



  Select rows (i) and cols (j) from giotto S4 coordinates slot


  Return \code{coordinates} slot data.table from giotto S4


  Select rows (i) and cols (j) from giotto S4 metaDT slot


  Return \code{coordinates} slot data.table from giotto S4
}

\section{\code{`[<-`} methods}{

  TODO
}

