% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_structures.R
\name{extract_points_list}
\alias{extract_points_list}
\title{Extract list of giotto points objects}
\usage{
extract_points_list(pointslist)
}
\arguments{
\item{pointslist}{list of inputs from which to create giotto points objects}
}
\description{
to extract list of giotto points
}
\keyword{internal}
