% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_structures.R
\name{extract_polygon_list}
\alias{extract_polygon_list}
\title{Extract list of polygons}
\usage{
extract_polygon_list(
  polygonlist,
  polygon_mask_list_params,
  polygon_dfr_list_params
)
}
\description{
to extract list of polygons
}
\keyword{internal}
