% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NN_network.R
\name{extractNearestNetwork}
\alias{extractNearestNetwork}
\title{extractNearestNetwork}
\usage{
extractNearestNetwork(
  gobject,
  nn_network_to_use = "sNN",
  network_name = "sNN.pca"
)
}
\arguments{
\item{gobject}{giotto object}

\item{nn_network_to_use}{kNN or sNN}

\item{network_name}{name of NN network to be used}
}
\value{
igraph object
}
\description{
Extracts a NN-network from a Giotto object as an igraph object
}
\examples{
    extractNearestNetwork(gobject)
}
