% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R, R/accessors.R
\name{fDataDT}
\alias{fDataDT}
\alias{get_FeatMetadata}
\title{fDataDT}
\usage{
fDataDT(gobject, spat_unit = NULL, feat_type = NULL)

get_FeatMetadata(gobject, spat_unit = NULL, feat_type = NULL)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit}

\item{feat_type}{feature type}
}
\value{
data.table with gene metadata
}
\description{
show gene metadata
}
