% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{featData-class}
\alias{featData-class}
\title{Basic class for classes with feature information}
\description{
Features in Giotto are a blanket term for any features that are detected, covering
modalities such as, but not limited to rna, protein, ATAC, and even QC probes.
Classes that inherit from this class will contain a feat_type slot that describes
which feature type the data is. Within slots that deal with classes that contain
featData, there is a nesting structure that usually first nests by spatial unit
and then by feature type
}
