% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction_spot.R
\name{featExpDWLS}
\alias{featExpDWLS}
\title{featExpDWLS}
\usage{
featExpDWLS(gobject, spat_unit = NULL, feat_type = NULL, ave_celltype_exp)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit (e.g. 'cell')}

\item{feat_type}{feature type (e.g. 'rna')}

\item{ave_celltype_exp}{data.table of feature expression in each cell type}
}
\value{
matrix
}
\description{
Compute predicted feature expression value by spatialDWSL
results and average feature expression for cell type
}
