% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{featType-generic}
\alias{featType-generic}
\alias{featType}
\alias{featType,featData-method}
\alias{featType<-,featData-method}
\title{Feature type information}
\usage{
featType(x)

\S4method{featType}{featData}(x)

\S4method{featType}{featData}(x) <- value
}
\arguments{
\item{x}{a Giotto S4 class subobject with feature type}

\item{value}{value to set as feature type}
}
\description{
access and set feat_type slot of S4 subobject
}
\section{Functions}{
\itemize{
\item \code{featType(featData)}: Get feature type information

\item \code{featType(featData) <- value}: Set feature type information

}}
