% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_structures.R
\name{filter_network}
\alias{filter_network}
\title{filter_network}
\usage{
filter_network(networkDT = NULL, maximum_distance = NULL, minimum_k = NULL)
}
\arguments{
\item{networkDT}{spatial network in data.table format}

\item{maximum_distance}{maximum distance between cell centroids}

\item{minimum_k}{minimum number of neighbors}
}
\description{
function to filter a spatial network
}
\keyword{internal}
