% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction.R
\name{filterCPG}
\alias{filterCPG}
\title{filterCPG}
\usage{
filterCPG(
  cpgObject,
  min_cells = 5,
  min_int_cells = 3,
  min_fdr = 0.05,
  min_spat_diff = 0.2,
  min_log2_fc = 0.5,
  direction = c("both", "up", "down")
)
}
\arguments{
\item{cpgObject}{cell proximity gene score object}

\item{min_cells}{minimum number of target cell type}

\item{min_int_cells}{minimum number of interacting cell type}

\item{min_fdr}{minimum adjusted p-value}

\item{min_spat_diff}{minimum absolute spatial expression difference}

\item{min_log2_fc}{minimum absolute log2 fold-change}

\item{direction}{differential expression directions to keep}
}
\value{
cpgObject that contains the filtered differential gene scores
}
\description{
Filter cell proximity gene scores.
}
\examples{
    filterCPG(gobject)
}
