% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction_visuals.R
\name{filterCPGscores}
\alias{filterCPGscores}
\title{filterCPGscores}
\usage{
filterCPGscores(
  CPGscore,
  min_cells = 5,
  min_fdr = 0.05,
  min_spat_diff = 0.2,
  min_log2_fc = 0.5,
  keep_int_duplicates = TRUE,
  direction = c("both", "up", "down")
)
}
\arguments{
\item{min_cells}{min number of cells threshold}

\item{min_fdr}{false_discovery threshold}

\item{min_spat_diff}{spatial difference threshold}

\item{min_log2_fc}{min log2 fold-change}

\item{keep_int_duplicates}{keep both cell_A-cell_B and cell_B-cell_A}

\item{direction}{expression changes to keep}

\item{method}{visualization method}
}
\value{
Gene to gene scores in data.table format
}
\description{
visualize Cell Proximity Gene enrichment scores
}
\details{
This function filters the output from getCellProximityGeneScores based on
false-discovery rate, minimum absolute difference, minimum log fold-change and
direction of change.
}
\examples{
    filterCPGscores(CPGscore)
}
