% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction_spot.R
\name{filterICGSpot}
\alias{filterICGSpot}
\title{filterICGSpot}
\usage{
filterICGSpot(
  cpgObject,
  min_cells = 4,
  min_cells_expr_resi = 0.05,
  min_int_cells = 4,
  min_int_cells_expr_resi = 0.05,
  min_fdr = 0.5,
  min_pcc_diff = 0.05,
  min_zscore = 0.05,
  zscores_column = c("cell_type", "genes"),
  direction = c("both", "up", "down")
)
}
\arguments{
\item{cpgObject}{ICG (interaction changed gene) score object}

\item{min_cells}{minimum number of source cell type}

\item{min_cells_expr_resi}{minimum expression residual level for source cell type}

\item{min_int_cells}{minimum number of interacting neighbor cell type}

\item{min_int_cells_expr_resi}{minimum expression residual level for interacting neighbor cell type}

\item{min_fdr}{minimum adjusted p-value}

\item{min_pcc_diff}{minimum absolute pcc difference difference}

\item{min_zscore}{minimum z-score change}

\item{zscores_column}{calculate z-scores over cell types or genes}

\item{direction}{differential expression directions to keep}
}
\value{
cpgObject that contains the filtered differential gene scores
}
\description{
Filter Interaction Changed Gene scores for spots.
}
