% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction.R
\name{filterInteractionChangedFeats}
\alias{filterInteractionChangedFeats}
\title{filterInteractionChangedFeats}
\usage{
filterInteractionChangedFeats(
  icfObject,
  min_cells = 4,
  min_cells_expr = 1,
  min_int_cells = 4,
  min_int_cells_expr = 1,
  min_fdr = 0.1,
  min_spat_diff = 0.2,
  min_log2_fc = 0.2,
  min_zscore = 2,
  zscores_column = c("cell_type", "feats"),
  direction = c("both", "up", "down")
)
}
\arguments{
\item{icfObject}{ICF (interaction changed feature) score object}

\item{min_cells}{minimum number of source cell type}

\item{min_cells_expr}{minimum expression level for source cell type}

\item{min_int_cells}{minimum number of interacting neighbor cell type}

\item{min_int_cells_expr}{minimum expression level for interacting neighbor 
cell type}

\item{min_fdr}{minimum adjusted p-value}

\item{min_spat_diff}{minimum absolute spatial expression difference}

\item{min_log2_fc}{minimum log2 fold-change}

\item{min_zscore}{minimum z-score change}

\item{zscores_column}{calculate z-scores over cell types or genes}

\item{direction}{differential expression directions to keep}
}
\value{
icfObject that contains the filtered differential feature scores
}
\description{
Filter Interaction Changed Feature scores.
}
