% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_hmrf.R
\name{filterSpatialGenes}
\alias{filterSpatialGenes}
\title{filterSpatialGenes}
\usage{
filterSpatialGenes(
  gobject,
  spatial_genes,
  max = 2500,
  name = c("binSpect", "silhouetteRank", "silhouetteRankTest"),
  method = c("none", "elbow")
)
}
\arguments{
\item{gobject}{TBD}

\item{spatial_genes}{TBD}

\item{max}{TBD}

\item{name}{TBD}

\item{method}{TBD}
}
\description{
filterSpatialGenes
}
