% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_enrichment.R
\name{find_dampening_constant}
\alias{find_dampening_constant}
\title{find_dampening_constant}
\usage{
find_dampening_constant(S, B, goldStandard)
}
\description{
find a dampening constant for the weights using cross-validation
}
\keyword{internal}
