% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{find_terra_writeRaster_dataType}
\alias{find_terra_writeRaster_dataType}
\title{find_terra_writeRaster_dataType}
\usage{
find_terra_writeRaster_dataType(
  giottoLargeImage = NULL,
  quick_INTS_maxval = NULL,
  max_intensity = NULL,
  min_intensity = NULL,
  is_int = NULL,
  signed = NULL,
  bitDepth = NULL,
  verbose = TRUE
)
}
\arguments{
\item{giottoLargeImage}{giottoLargeImage object to determine max_intensity,
min_intensity, is_int settings from}

\item{max_intensity, min_intensity}{value given as image maximum/minimum intensity}

\item{is_int}{if image is integer (TRUE) or floating point (FALSE)}

\item{signed}{if image is signed (TRUE) or unsigned (TRUE)}

\item{bitDepth}{image bitDepth}

\item{verbose}{be verbose}

\item{quick_INTU_maxval}{Treat as maximum intensity to find compatible unsigned
integer settings}
}
\value{
datatype for terra writeRaster function
}
\description{
find likely compatible datatype for given image characteristics.
  Values given in arguments take priority over those found from giottoLargeImage
  metadata
}
\keyword{internal}
