% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_section.R
\name{find_x_y_ranges}
\alias{find_x_y_ranges}
\title{find_x_y_ranges}
\usage{
find_x_y_ranges(data, extend_ratio)
}
\arguments{
\item{data}{data}

\item{extend_ratio}{extend_ratio}
}
\description{
get the extended ranges of x and y
}
\keyword{internal}
