% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction.R
\name{findCellProximityGenes_per_interaction}
\alias{findCellProximityGenes_per_interaction}
\title{findCellProximityGenes_per_interaction}
\usage{
findCellProximityGenes_per_interaction(
  expr_values,
  cell_metadata,
  annot_spatnetwork,
  sel_int,
  cluster_column = NULL,
  minimum_unique_cells = 1,
  minimum_unique_int_cells = 1,
  exclude_selected_cells_from_test = T,
  diff_test = c("permutation", "limma", "t.test", "wilcox"),
  mean_method = c("arithmic", "geometric"),
  offset = 0.1,
  adjust_method = "bonferroni",
  nr_permutations = 100,
  cores = 1,
  set_seed = TRUE,
  seed_number = 1234
)
}
\description{
Identifies genes that are differentially expressed due to proximity to other cell types.
}
\keyword{internal}
