% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction.R
\name{findCPG}
\alias{findCPG}
\title{findCPG}
\usage{
findCPG(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=findICG]{findICG}}
  \describe{
    \item{\code{gobject}}{giotto object}
    \item{\code{expression_values}}{expression values to use}
    \item{\code{selected_genes}}{subset of selected genes (optional)}
    \item{\code{cluster_column}}{name of column to use for cell types}
    \item{\code{spatial_network_name}}{name of spatial network to use}
    \item{\code{minimum_unique_cells}}{minimum number of target cells required}
    \item{\code{minimum_unique_int_cells}}{minimum number of interacting cells required}
    \item{\code{diff_test}}{which differential expression test}
    \item{\code{mean_method}}{method to use to calculate the mean}
    \item{\code{offset}}{offset value to use when calculating log2 ratio}
    \item{\code{adjust_method}}{which method to adjust p-values}
    \item{\code{nr_permutations}}{number of permutations if diff_test = permutation}
    \item{\code{exclude_selected_cells_from_test}}{exclude interacting cells other cells}
    \item{\code{do_parallel}}{run calculations in parallel with mclapply}
    \item{\code{cores}}{number of cores to use if do_parallel = TRUE}
    \item{\code{set_seed}}{set a seed for reproducibility}
    \item{\code{seed_number}}{seed number}
  }}
}
\description{
Identifies cell-to-cell Interaction Changed Genes (ICG),
i.e. genes that are differentially expressed due to proximity to other cell types.
}
\seealso{
\code{\link{findICG}}
}
