% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction.R
\name{findICG}
\alias{findICG}
\title{findICG}
\usage{
findICG(
  gobject,
  expression_values = "normalized",
  selected_genes = NULL,
  cluster_column,
  spatial_network_name = "Delaunay_network",
  minimum_unique_cells = 1,
  minimum_unique_int_cells = 1,
  diff_test = c("permutation", "limma", "t.test", "wilcox"),
  mean_method = c("arithmic", "geometric"),
  offset = 0.1,
  adjust_method = c("bonferroni", "BH", "holm", "hochberg", "hommel", "BY", "fdr",
    "none"),
  nr_permutations = 100,
  exclude_selected_cells_from_test = T,
  do_parallel = TRUE,
  cores = NA,
  set_seed = TRUE,
  seed_number = 1234
)
}
\arguments{
\item{gobject}{giotto object}

\item{expression_values}{expression values to use}

\item{selected_genes}{subset of selected genes (optional)}

\item{cluster_column}{name of column to use for cell types}

\item{spatial_network_name}{name of spatial network to use}

\item{minimum_unique_cells}{minimum number of target cells required}

\item{minimum_unique_int_cells}{minimum number of interacting cells required}

\item{diff_test}{which differential expression test}

\item{mean_method}{method to use to calculate the mean}

\item{offset}{offset value to use when calculating log2 ratio}

\item{adjust_method}{which method to adjust p-values}

\item{nr_permutations}{number of permutations if diff_test = permutation}

\item{exclude_selected_cells_from_test}{exclude interacting cells other cells}

\item{do_parallel}{run calculations in parallel with mclapply}

\item{cores}{number of cores to use if do_parallel = TRUE}

\item{set_seed}{set a seed for reproducibility}

\item{seed_number}{seed number}
}
\value{
cpgObject that contains the differential gene scores
}
\description{
Identifies cell-to-cell Interaction Changed Genes (ICG),
i.e. genes that are differentially expressed due to proximity to other cell types.
}
\details{
Function to calculate if genes are differentially expressed in cell types
 when they interact (approximated by physical proximity) with other cell types.
 The results data.table in the cpgObject contains - at least - the following columns:
\itemize{
 \item{genes:}{ All or selected list of tested genes}
 \item{sel:}{ average gene expression in the interacting cells from the target cell type }
 \item{other:}{ average gene expression in the NOT-interacting cells from the target cell type }
 \item{log2fc:}{ log2 fold-change between sel and other}
 \item{diff:}{ spatial expression difference between sel and other}
 \item{p.value:}{ associated p-value}
 \item{p.adj:}{ adjusted p-value}
 \item{cell_type:}{ target cell type}
 \item{int_cell_type:}{ interacting cell type}
 \item{nr_select:}{ number of cells for selected target cell type}
 \item{int_nr_select:}{ number of cells for interacting cell type}
 \item{nr_other:}{ number of other cells of selected target cell type}
 \item{int_nr_other:}{ number of other cells for interacting cell type}
 \item{unif_int:}{ cell-cell interaction}
}
}
