% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction_spot.R
\name{findICGSpot}
\alias{findICGSpot}
\title{findICGSpot}
\usage{
findICGSpot(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  expression_values = c("normalized", "scaled", "custom"),
  ave_celltype_exp,
  selected_genes = NULL,
  spatial_network_name = "Delaunay_network",
  minimum_unique_cells = 5,
  minimum_unique_int_cells = 5,
  CCI_cell_score = 0.1,
  dwls_cutoff = 0.001,
  diff_test = "permutation",
  nr_permutations = 100,
  adjust_method = "fdr",
  do_parallel = TRUE,
  cores = 2,
  set_seed = TRUE,
  seed_number = 1234
)
}
\arguments{
\item{gobject}{giotto object}

\item{expression_values}{expression values to use}

\item{ave_celltype_exp}{averege gene expression in each cell type}

\item{selected_genes}{subset of selected genes (optional)}

\item{spatial_network_name}{name of spatial network to use}

\item{minimum_unique_cells}{minimum number of target cells required}

\item{minimum_unique_int_cells}{minimum number of interacting cells required}

\item{CCI_cell_score}{cell proximity score to filter no interacted cell}

\item{dwls_cutoff}{cell type proportion cutoff to label the cell}

\item{diff_test}{which differential expression test}

\item{nr_permutations}{number of permutations if diff_test = permutation}

\item{adjust_method}{which method to adjust p-values}

\item{do_parallel}{run calculations in parallel with mclapply}

\item{cores}{number of cores to use if do_parallel = TRUE}

\item{set_seed}{set a seed for reproducibility}

\item{seed_number}{seed number}
}
\value{
cpgObject that contains the differential gene scores
}
\description{
Identifies cell-to-cell Interaction Changed Genes (ICG) for spots,
i.e. genes expression residual that are different due to proximity to other cell types.
}
\details{
Function to calculate if genes expreesion residual are differentially expressed in cell types
 when they interact (approximated by physical proximity) with other cell types.
 Gene expression residual calculated as:
 (observed expression in spot - cell_type_proportion * average_expressed_in_cell_type)
 The results data.table in the cpgObject contains - at least - the following columns:
\itemize{
 \item{genes:}{ All or selected list of tested genes}
 \item{sel:}{ average gene expression residual in the interacting cells from the target cell type }
 \item{other:}{ average gene expression residual in the NOT-interacting cells from the target cell type }
 \item{pcc_sel:}{ correlation between cell proximity score and expression residual in the interacting cells from the target cell type}
 \item{pcc_other:}{ correlation between cell proximity score and expression residual in the NOT-interacting cells from the target cell type }
 \item{pcc_diff:}{ correlation difference between sel and other}
 \item{p.value:}{ associated p-value}
 \item{p.adj:}{ adjusted p-value}
 \item{cell_type:}{ target cell type}
 \item{int_cell_type:}{ interacting cell type}
 \item{nr_select:}{ number of cells for selected target cell type}
 \item{int_nr_select:}{ number of cells for interacting cell type}
 \item{unif_int:}{ cell-cell interaction}
}
}
