% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction.R
\name{findInteractionChangedFeats}
\alias{findInteractionChangedFeats}
\title{findInteractionChangedFeats}
\usage{
findInteractionChangedFeats(
  gobject,
  feat_type = NULL,
  spat_unit = NULL,
  expression_values = "normalized",
  selected_feats = NULL,
  cluster_column,
  spatial_network_name = "Delaunay_network",
  minimum_unique_cells = 1,
  minimum_unique_int_cells = 1,
  diff_test = c("permutation", "limma", "t.test", "wilcox"),
  mean_method = c("arithmic", "geometric"),
  offset = 0.1,
  adjust_method = c("bonferroni", "BH", "holm", "hochberg", "hommel", "BY", "fdr",
    "none"),
  nr_permutations = 1000,
  exclude_selected_cells_from_test = TRUE,
  do_parallel = TRUE,
  set_seed = TRUE,
  seed_number = 1234
)
}
\arguments{
\item{gobject}{giotto object}

\item{feat_type}{feature type}

\item{spat_unit}{spatial unit}

\item{expression_values}{expression values to use}

\item{selected_feats}{subset of selected features (optional)}

\item{cluster_column}{name of column to use for cell types}

\item{spatial_network_name}{name of spatial network to use}

\item{minimum_unique_cells}{minimum number of target cells required}

\item{minimum_unique_int_cells}{minimum number of interacting cells required}

\item{diff_test}{which differential expression test}

\item{mean_method}{method to use to calculate the mean}

\item{offset}{offset value to use when calculating log2 ratio}

\item{adjust_method}{which method to adjust p-values}

\item{nr_permutations}{number of permutations if diff_test = permutation}

\item{exclude_selected_cells_from_test}{exclude interacting cells other cells}

\item{do_parallel}{run calculations in parallel with mclapply}

\item{set_seed}{set a seed for reproducibility}

\item{seed_number}{seed number}
}
\value{
icfObject that contains the Interaction Changed differential feature
scores
}
\description{
Identifies cell-to-cell Interaction Changed Features (ICF),
i.e. features that are differentially expressed due to proximity to other
cell types.
}
\details{
Function to calculate if features are differentially expressed in
cell types when they interact (approximated by physical proximity) with
other cell types. The results data.table in the icfObject contains
- at least - the following columns:
\itemize{
 \item{features:}{ All or selected list of tested features}
 \item{sel:}{ average feature expression in the interacting cells from the target cell type }
 \item{other:}{ average feature expression in the NOT-interacting cells from the target cell type }
 \item{log2fc:}{ log2 fold-change between sel and other}
 \item{diff:}{ spatial expression difference between sel and other}
 \item{p.value:}{ associated p-value}
 \item{p.adj:}{ adjusted p-value}
 \item{cell_type:}{ target cell type}
 \item{int_cell_type:}{ interacting cell type}
 \item{nr_select:}{ number of cells for selected target cell type}
 \item{int_nr_select:}{ number of cells for interacting cell type}
 \item{nr_other:}{ number of other cells of selected target cell type}
 \item{int_nr_other:}{ number of other cells for interacting cell type}
 \item{unif_int:}{ cell-cell interaction}
}
}
\examples{
g <- GiottoData::loadGiottoMini("visium")

findInteractionChangedFeats(g, cluster_column = "leiden_clus",
selected_feats = c("Gna12", "Ccnd2", "Btbd17"), nr_permutations = 10)
}
