% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction.R
\name{findInteractionChangedGenes}
\alias{findInteractionChangedGenes}
\title{findInteractionChangedGenes}
\usage{
findInteractionChangedGenes(...)
}
\arguments{
\item{...}{params to pass to \code{findInteractionChangedFeats}}
}
\value{
interaction changed genes
}
\description{
Identifies cell-to-cell Interaction Changed Features (ICF),
i.e. genes that are differentially expressed due to interactions with other
cell types.
}
\seealso{
\code{\link{findInteractionChangedFeats}}
}
