% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{findMastMarkers_one_vs_all}
\alias{findMastMarkers_one_vs_all}
\title{findMastMarkers_one_vs_all}
\usage{
findMastMarkers_one_vs_all(gobject, expression_values = c("normalized",
  "scaled", "custom"), cluster_column, subset_clusters = NULL,
  adjust_columns = NULL, pval = 0.001, logFC = 1, min_genes = 10,
  verbose = TRUE, ...)
}
\arguments{
\item{gobject}{giotto object}

\item{expression_values}{gene expression values to use}

\item{cluster_column}{clusters to use}

\item{subset_clusters}{selection of clusters to compare}

\item{adjust_columns}{column in pDataDT to adjust for (e.g. detection rate)}

\item{pval}{filter on minimal p-value}

\item{logFC}{filter on logFC}

\item{min_genes}{minimum genes to keep per cluster, overrides pval and logFC}

\item{verbose}{be verbose}

\item{...}{additional parameters for the zlm function in MAST}
}
\value{
data.table with marker genes
}
\description{
Identify marker genes for all clusters based on the MAST package.
}
\details{
This is a minimal convenience wrapper around the MAST functions to detect differentially expressed genes.
}
\examples{
    findMastMarkers_one_vs_all(gobject)
}
