% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{findNetworkNeighbors}
\alias{findNetworkNeighbors}
\title{findNetworkNeighbors}
\usage{
findNetworkNeighbors(
  gobject,
  spatial_network_name,
  source_cell_ids = NULL,
  name = "nb_cells"
)
}
\arguments{
\item{gobject}{Giotto object}

\item{spatial_network_name}{name of spatial network}

\item{source_cell_ids}{cell ids for which you want to know the spatial neighbors}

\item{name}{name of the results}
}
\value{
data.table
}
\description{
Find the spatial neighbors for a selected group of cells within the selected spatial network.
}
\examples{

data(mini_giotto_single_cell)

# get all cells
all_cells = slot(mini_giotto_single_cell, 'cell_ID')

# find all the spatial neighbours for the first 5 cells
# within the Delaunay network
findNetworkNeighbors(mini_giotto_single_cell,
                     spatial_network_name = 'Delaunay_network',
                     source_cell_ids = all_cells[1:5])

}
