% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{findScranMarkers}
\alias{findScranMarkers}
\title{findScranMarkers}
\usage{
findScranMarkers(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  expression_values = c("normalized", "scaled", "custom"),
  cluster_column,
  subset_clusters = NULL,
  group_1 = NULL,
  group_1_name = NULL,
  group_2 = NULL,
  group_2_name = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit}

\item{feat_type}{feature type}

\item{expression_values}{gene expression values to use}

\item{cluster_column}{clusters to use}

\item{subset_clusters}{selection of clusters to compare}

\item{group_1}{group 1 cluster IDs from cluster_column for pairwise 
comparison}

\item{group_1_name}{custom name for group_1 clusters}

\item{group_2}{group 2 cluster IDs from cluster_column for pairwise 
comparison}

\item{group_2_name}{custom name for group_2 clusters}

\item{verbose}{be verbose (default = FALSE)}

\item{...}{additional parameters for the findMarkers function in scran}
}
\value{
data.table with marker genes
}
\description{
Identify marker genes for all or selected clusters based on 
scran's implementation of findMarkers.
}
\details{
This is a minimal convenience wrapper around
the \code{\link[scran]{findMarkers}} function from the scran package.

To perform differential expression between custom selected groups of cells
you need to specify the cell_ID column to parameter \emph{cluster_column}
and provide the individual cell IDs to the parameters \emph{group_1} and 
\emph{group_2}

By default group names will be created by pasting the different id names 
within each selected group.
When you have many different ids in a single group
it is recommend to provide names for both groups to \emph{group_1_name} and 
\emph{group_2_name}
}
\examples{
g <- GiottoData::loadGiottoMini("visium")

findScranMarkers(g, cluster_column = "leiden_clus")
}
