% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_structures.R
\name{fix_multipart_geoms}
\alias{fix_multipart_geoms}
\title{Split multi-part polygons}
\usage{
fix_multipart_geoms(spatVector)
}
\description{
function to split geoms (polygons) that have multiple parts
}
\keyword{internal}
