% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{flex_lapply}
\alias{flex_lapply}
\title{flex_lapply}
\usage{
flex_lapply(X, FUN, cores = NA, fun = NULL, ...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=lapply_flex]{lapply_flex}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
flex_lapply
}
\seealso{
\code{\link{lapply_flex}}
}
\keyword{internal}
