% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_manipulation.R
\name{flip-generic}
\alias{flip-generic}
\alias{flip,giottoPolygon-method}
\alias{flip,giottoPoints-method}
\alias{flip,spatLocsObj-method}
\alias{flip,spatialNetworkObj-method}
\alias{flip,giottoLargeImage-method}
\alias{flip,SpatExtent-method}
\title{Flip an object}
\usage{
\S4method{flip}{giottoPolygon}(x, direction = "vertical", x0 = 0, y0 = 0, ...)

\S4method{flip}{giottoPoints}(x, direction = "vertical", x0 = 0, y0 = 0, ...)

\S4method{flip}{spatLocsObj}(x, direction = "vertical", x0 = 0, y0 = 0, ...)

\S4method{flip}{spatialNetworkObj}(x, direction = "vertical", x0 = 0, y0 = 0, ...)

\S4method{flip}{giottoLargeImage}(x, direction = "vertical", ...)

\S4method{flip}{SpatExtent}(x, direction = "vertical", x0 = 0, y0 = 0)
}
\arguments{
\item{x}{object}

\item{direction}{character. Direction to flip. Should be either partial match to 'vertical' or 'horizontal'}

\item{x0}{x value to flip horizontally over (ignored for vertical). Pass NULL
to flip over the extent}

\item{y0}{y value to flip vertically over (ignored for horizontal). Pass NULL
to flip over the extent}

\item{...}{additional args to pass}
}
\description{
Flip an object over a designated x or y value depending on
direction param input. Note that this behavior may be different from terra's
}
\section{Functions}{
\itemize{
\item \code{flip(giottoPolygon)}: Flip a giottoPolygon object

\item \code{flip(giottoPoints)}: Flip a giottoPoints object

\item \code{flip(spatLocsObj)}: Flip a spatLocsObj

\item \code{flip(spatialNetworkObj)}: Flip a spatialNetworkObj

\item \code{flip(giottoLargeImage)}: Flip a giottoLargeImage

\item \code{flip(SpatExtent)}: Flip a SpatExtent

}}
