% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{fread_colmatch}
\alias{fread_colmatch}
\title{Fread rows based on column matches}
\usage{
fread_colmatch(file, col, sep = NULL, values_to_match, verbose = FALSE, ...)
}
\arguments{
\item{file}{path to file to load}

\item{col}{name of col to match from}

\item{sep}{grep term to match as column delimiters within the file}

\item{values_to_match}{values in \code{col} to match given as a vector}

\item{verbose}{whether to print the grep command}

\item{...}{additional parameters to pass to \code{\link[data.table]{fread}}}
}
\description{
Fread rows based on column matches
}
\keyword{internal}
