% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_genes.R
\name{FSV_show}
\alias{FSV_show}
\title{FSV_show}
\usage{
FSV_show(
  results,
  ms_results = NULL,
  size = c(4, 2, 1),
  color = c("blue", "green", "red"),
  sig_alpha = 0.5,
  unsig_alpha = 0.5
)
}
\arguments{
\item{results}{results caculated by spatial_DE}

\item{ms_results}{ms_results caculated by spatial_DE}

\item{size}{indicate different levels of qval}

\item{color}{indicate different SV features}

\item{sig_alpha}{transparency of significant genes}

\item{unsig_alpha}{transparency of unsignificant genes}
}
\value{
ggplot object
}
\description{
Visualize spatial varible genes caculated by spatial_DE
}
\details{
Description of parameters.
}
\keyword{internal}
