% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction_spot.R
\name{geneExpDWLS}
\alias{geneExpDWLS}
\title{geneExpDWLS}
\usage{
geneExpDWLS(gobject, ave_celltype_exp)
}
\arguments{
\item{gobject}{gottio object}

\item{ave_celltype_exp}{data.table of gene expression in each cell type}
}
\value{
matrix
}
\description{
Compute predicted gene expression value by spacialDWSL results and 
average gene expression for cell type
}
