% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_visuals.R
\name{general_save_function}
\alias{general_save_function}
\title{general_save_function}
\usage{
general_save_function(
  gobject,
  plot_object,
  save_dir = NULL,
  save_folder = NULL,
  save_name = NULL,
  default_save_name = "giotto_plot",
  save_format = c("png", "tiff", "pdf", "svg"),
  show_saved_plot = F,
  base_width = NULL,
  base_height = NULL,
  base_aspect_ratio = NULL,
  units = NULL,
  dpi = NULL,
  ...
)
}
\arguments{
\item{gobject}{giotto object}

\item{plot_object}{non-ggplot object to plot}

\item{save_dir}{directory to save to}

\item{save_folder}{folder in save_dir to save to}

\item{save_name}{name of plot}

\item{save_format}{format (e.g. png, tiff, pdf, ...)}

\item{show_saved_plot}{load & display the saved plot}

\item{base_width}{width}

\item{base_height}{height}

\item{base_aspect_ratio}{aspect ratio}

\item{units}{units}

\item{dpi}{Plot resolution}
}
\description{
Function to automatically save plots to directory of interest
}
\keyword{internal}
