% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_help.R
\name{get10Xmatrix}
\alias{get10Xmatrix}
\title{get10Xmatrix}
\usage{
get10Xmatrix(path_to_data)
}
\arguments{
\item{path_to_data}{path to the 10X folder}
}
\value{
expression matrix from 10X
}
\description{
This function creates an expression matrix from a 10X structured folder
}
\details{
A typical 10X folder is named raw_feature_bc_matrix or raw_feature_bc_matrix. It has 3 files:
\itemize{
  \item{barcodes}
  \item{features.tsv.gz}
  \item{matrix.mtx.gz}
}
}
\examples{
    get10Xmatrix(10Xmatrix)
}
