% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_cell_id}
\alias{get_cell_id}
\title{Get cell IDs for a given spatial unit}
\usage{
get_cell_id(gobject, spat_unit = NULL)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit (e.g. "cell")}
}
\value{
character vector of cell_IDs
}
\description{
Data for each spatial unit is expected to agree on a single set of cell_IDs
that are shared across any feature types. These cell_IDs are stored within the
giotto object's \code{cell_ID} slot. Getters and setters for this slot directly
retrieve (get) or replace (set) this slot.
}
\seealso{
set_cell_id

Other functions to set data in giotto object: 
\code{\link{get_feat_id}()},
\code{\link{set_CellMetadata}()},
\code{\link{set_FeatMetadata}()},
\code{\link{set_NearestNetwork}()},
\code{\link{set_cell_id}()},
\code{\link{set_dimReduction}()},
\code{\link{set_expression_values}()},
\code{\link{set_feat_id}()},
\code{\link{set_feature_info}()},
\code{\link{set_giottoImage}()},
\code{\link{set_polygon_info}()},
\code{\link{set_spatialGrid}()},
\code{\link{set_spatialNetwork}()},
\code{\link{set_spatial_enrichment}()},
\code{\link{set_spatial_locations}()}
}
\concept{functions to set data in giotto object}
\keyword{internal}
