% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction.R
\name{get_cell_to_cell_sorted_name_conversion}
\alias{get_cell_to_cell_sorted_name_conversion}
\title{get_cell_to_cell_sorted_name_conversion}
\usage{
get_cell_to_cell_sorted_name_conversion(all_cell_types)
}
\description{
creates unified cell-cell interaction names
}
\examples{
    get_cell_to_cell_sorted_name_conversion()
}
