% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_section.R
\name{get_cross_section_coordinates}
\alias{get_cross_section_coordinates}
\title{get_cross_section_coordinates}
\usage{
get_cross_section_coordinates(cell_subset_projection_locations)
}
\description{
get local coordinates within cross section plane
}
