% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_dimReduction}
\alias{get_dimReduction}
\title{get_dimReduction}
\usage{
get_dimReduction(
  gobject,
  spat_unit = NULL,
  reduction = c("cells", "feats"),
  reduction_method = c("pca", "umap", "tsne"),
  name = "pca",
  return_dimObj = FALSE
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit}

\item{reduction}{reduction on cells or features}

\item{reduction_method}{reduction method (e.g. pca)}

\item{name}{name of reduction results}

\item{return_dimObj}{return full dimension object result}
}
\value{
dim reduction coordinates (default) or dim reduction object
}
\description{
function to get a dimension reduction object
}
