% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_feature_metadata}
\alias{get_feature_metadata}
\title{Get feature metadata}
\usage{
get_feature_metadata(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  output = c("data.table", "featMetaObj")
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit (e.g. "cell")}

\item{feat_type}{feature type (e.g. "rna", "dna", "protein")}

\item{output}{return as either 'data.table' or 'featMetaObj'}
}
\description{
Get feature metadata from giotto object
}
\seealso{
fDataDT
}
\keyword{internal}
