% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_giottoImage}
\alias{get_giottoImage}
\title{get_giottoImage}
\usage{
get_giottoImage(
  gobject = NULL,
  image_type = c("image", "largeImage"),
  name = NULL
)
}
\arguments{
\item{gobject}{giotto object}

\item{image_type}{type of giotto image object}

\item{image_name}{name of a giotto image object \code{\link{showGiottoImageNames}}}
}
\value{
a giotto image object
}
\description{
get giotto image object from gobject
}
