% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_giottoImage_MG}
\alias{get_giottoImage_MG}
\title{Get \emph{magick}-based giotto \code{image}}
\usage{
get_giottoImage_MG(gobject = NULL, name = NULL)
}
\arguments{
\item{gobject}{giotto object}

\item{name}{name of giottoImage \code{\link{showGiottoImageNames}}}
}
\value{
a giottoImage
}
\description{
Get a giottoImage from a giotto object
}
\keyword{internal}
