% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registration.R
\name{get_img_corners}
\alias{get_img_corners}
\title{get_img_corners}
\usage{
get_img_corners(img_object)
}
\arguments{
\item{img_object}{giottoImage or magick-image to use}
}
\description{
finds four corner spatial coords of giottoImages or magick-images
}
\keyword{internal}
