% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{get_img_minmax}
\alias{get_img_minmax}
\title{get_img_minmax}
\usage{
get_img_minmax(mg_img, negative_y = TRUE)
}
\arguments{
\item{mg_img}{magick object}

\item{negative_y}{Map image to negative y spatial values if TRUE during automatic alignment. Meaning that origin is in upper left instead of lower left.}
}
\description{
get_img_minmax
}
\keyword{internal}
