% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_section.R
\name{get_meanCellDistance}
\alias{get_meanCellDistance}
\title{get_meanCellDistance}
\usage{
get_meanCellDistance(
  gobject,
  spatial_network_name = "Delaunay_network",
  plane_equation = NULL
)
}
\description{
get mean distance between neighboring cells
}
