% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_help.R
\name{get_os}
\alias{get_os}
\title{get_os}
\usage{
get_os()
}
\value{
character osx, linux or windows
}
\description{
return the type of operating system, see https://conjugateprior.org/2015/06/identifying-the-os-from-r/
}
\keyword{internal}
