% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_polygon_info}
\alias{get_polygon_info}
\title{get_polygon_info}
\usage{
get_polygon_info(gobject, polygon_name = "cell", polygon_overlap = NULL)
}
\arguments{
\item{gobject}{giotto object}

\item{polygon_name}{name of polygons}

\item{polygon_overlap}{include polygon overlap information}
}
\description{
get giotto polygon spatVector
}
