% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_section.R
\name{get_sectionThickness}
\alias{get_sectionThickness}
\title{get_sectionThickness}
\usage{
get_sectionThickness(
  gobject,
  thickness_unit = c("cell", "natural"),
  slice_thickness = 2,
  spatial_network_name = "Delaunay_network",
  plane_equation = NULL
)
}
\description{
get section thickness
}
