% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_spatial_enrichment}
\alias{get_spatial_enrichment}
\title{get_spatial_enrichment}
\usage{
get_spatial_enrichment(gobject, enrichm_name = "DWLS")
}
\arguments{
\item{gobject}{giotto object}

\item{enrichm_name}{name of spatial enrichment results}
}
\value{
data.table with fractions
}
\description{
function to get a spatial enrichment data.table
}
