% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_spatial_locations}
\alias{get_spatial_locations}
\title{get_spatial_locations}
\usage{
get_spatial_locations(gobject, spat_loc_name = "raw")
}
\arguments{
\item{gobject}{giotto object}

\item{spat_loc_name}{name of spatial locations}
}
\value{
data.table with coordinates
}
\description{
function to get a spatial location data.table
}
