% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_spatialGrid}
\alias{get_spatialGrid}
\title{get_spatialGrid}
\usage{
get_spatialGrid(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  name = NULL,
  return_grid_Obj = FALSE
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit}

\item{feat_type}{feature type}

\item{name}{name of spatial grid}

\item{return_network_Obj}{return grid object (default = FALSE)}
}
\description{
function to get spatial grid
}
