% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_spatialNetwork}
\alias{get_spatialNetwork}
\title{get_spatialNetwork}
\usage{
get_spatialNetwork(gobject, name = NULL, return_network_Obj = FALSE)
}
\arguments{
\item{gobject}{giotto object}

\item{name}{name of spatial network}

\item{return_network_Obj}{return network object (default = FALSE)}
}
\description{
function to get a spatial network
}
