% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction.R
\name{getAverageCellProximityGeneScores}
\alias{getAverageCellProximityGeneScores}
\title{getAverageCellProximityGeneScores}
\usage{
getAverageCellProximityGeneScores(gobject,
  spatial_network_name = "spatial_network",
  cluster_column = "louvain_clus.1",
  expression_values = c("normalized", "scaled", "custom"),
  do_diff_test = T, exclude_selected_cells_from_test = F,
  verbose = T)
}
\arguments{
\item{gobject}{giotto object}

\item{spatial_network_name}{name of spatial network to use}

\item{cluster_column}{name of column to use for clusters}

\item{expression_values}{expression values to use}

\item{do_diff_test}{do differential test}

\item{exclude_selected_cells_from_test}{exclude certain cells from test}

\item{verbose}{verbose}
}
\value{
Cell Proximity Gene scores (CPGscores) in data.table format
}
\description{
Compute cell-cell interaction enrichment (observed vs expected)
}
\details{
Give more details ...
}
\examples{
    getAverageCellProximityGeneScores(gobject)
}
