% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{getCellMetadata}
\alias{getCellMetadata}
\title{getCellMetadata}
\usage{
getCellMetadata(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  output = c("cellMetaObj", "data.table"),
  copy_obj = TRUE,
  set_defaults = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit (e.g. "cell")}

\item{feat_type}{feature type (e.g. "rna", "dna", "protein")}

\item{output}{return as either 'data.table' or 'cellMetaObj'}

\item{copy_obj}{whether to deep copy/duplicate when getting the object (default = TRUE)}

\item{set_defaults}{set default spat_unit and feat_type. Change to FALSE only when
expression and spat_info are not expected to exist.}
}
\description{
Get cell metadata from giotto object
}
\seealso{
pDataDT
}
