% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction.R
\name{getCellProximityGeneScores}
\alias{getCellProximityGeneScores}
\title{getCellProximityGeneScores}
\usage{
getCellProximityGeneScores(gobject,
  spatial_network_name = "spatial_network",
  cluster_column = "louvain_clus.1", selected_genes = NULL,
  expression_values = c("normalized", "scaled", "custom"),
  fold_change_addendum = 0.1, in_two_directions = TRUE,
  exclude_selected_cells_from_test = F, verbose = T)
}
\arguments{
\item{gobject}{giotto object}

\item{spatial_network_name}{name of spatial network to use}

\item{cluster_column}{name of column to use for clusters}

\item{expression_values}{expression values to use}

\item{fold_change_addendum}{constant to add when calculating log2 fold-change}

\item{in_two_directions}{shows enrichment in both directions: cell1-cell2, cell2-cell1}

\item{exclude_selected_cells_from_test}{exclude certain cells from test}

\item{verbose}{verbose}
}
\value{
Cell Proximity Gene scores (CPGscores) in data.table format
}
\description{
Compute cell-cell interaction enrichment (observed vs expected)
}
\details{
Give more details ...
}
\examples{
    getCellProximityGeneScores(gobject)
}
