% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{getClusterSimilarity}
\alias{getClusterSimilarity}
\title{getClusterSimilarity}
\usage{
getClusterSimilarity(
  gobject,
  expression_values = c("normalized", "scaled", "custom"),
  cluster_column,
  cor = c("pearson", "spearman")
)
}
\arguments{
\item{gobject}{giotto object}

\item{expression_values}{expression values to use}

\item{cluster_column}{name of column to use for clusters}

\item{cor}{correlation score to calculate distance}
}
\value{
data.table
}
\description{
Creates data.table with pairwise correlation scores between each cluster.
}
\details{
Creates data.table with pairwise correlation scores between each cluster and
the group size (# of cells) for each cluster. This information can be used together
with mergeClusters to combine very similar or small clusters into bigger clusters.
}
\examples{

data("mini_giotto_single_cell")

cluster_similarities = getClusterSimilarity(mini_giotto_single_cell,
                                            cluster_column = 'leiden_clus')

}
