% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_help.R
\name{getDistinctColors}
\alias{getDistinctColors}
\title{getDistinctColors}
\usage{
getDistinctColors(n)
}
\arguments{
\item{n}{number of colors wanted}
}
\value{
number of distinct colors
}
\description{
Returns a number of distint colors based on the RGB scale
}
