% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_help.R
\name{getGEFtxCoords}
\alias{getGEFtxCoords}
\title{getGEFtxCoords}
\usage{
getGEFtxCoords(gef_file, bin_size = "bin100")
}
\arguments{
\item{gef_file}{path to .gef file}

\item{bin_size}{bin size to select from .gef file}
}
\description{
Converts .gef file (output stereo-seq pipeline) into
transcript with coordinates
}
