% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction.R
\name{getGeneToGeneScores}
\alias{getGeneToGeneScores}
\title{getGeneToGeneScores}
\usage{
getGeneToGeneScores(CPGscore, selected_genes = NULL,
  selected_cell_interactions, specific_genes_1 = NULL,
  specific_genes_2 = NULL, verbose = FALSE)
}
\arguments{
\item{CPGscore}{CPGscore, output from getAverageCellProximityGeneScores()}

\item{selected_genes}{subset of genes to test}

\item{selected_cell_interactions}{subset of cell-cell interactions to test}

\item{specific_genes_1}{specific source genes (see details)}

\item{specific_genes_2}{specific target genes (see details)}

\item{verbose}{verbose}
}
\value{
Gene to gene scores in data.table format
}
\description{
Compute gene-gene enrichment scores.
}
\details{
Give more details ...
}
\examples{
    getGeneToGeneScores(CPGscore)
}
