% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnn.R
\name{getMultiomics}
\alias{getMultiomics}
\title{Get multiomics integration results}
\usage{
getMultiomics(
  gobject = NULL,
  spat_unit = NULL,
  feat_type = NULL,
  integration_method = "WNN",
  result_name = "theta_weighted_matrix"
)
}
\arguments{
\item{gobject}{A Giotto object}

\item{spat_unit}{spatial unit (e.g. 'cell')}

\item{feat_type}{integrated feature type (e.g. 'rna_protein')}

\item{integration_method}{multiomics integration method used. Default = 'WNN'}

\item{result_name}{Default = 'theta_weighted_matrix'}
}
\value{
A multiomics integration result (e.g. theta_weighted_matrix from WNN)
}
\description{
Get a multiomics integration result from a Giotto object
}
\seealso{
Other multiomics accessor functions: 
\code{\link{get_multiomics}()},
\code{\link{setMultiomics}()},
\code{\link{set_multiomics}()}

Other functions to get data from giotto object: 
\code{\link{getDimReduction}()},
\code{\link{getExpression}()},
\code{\link{getFeatureInfo}()},
\code{\link{getGiottoImage}()},
\code{\link{getNearestNetwork}()},
\code{\link{getPolygonInfo}()},
\code{\link{getSpatialEnrichment}()},
\code{\link{getSpatialGrid}()},
\code{\link{getSpatialLocations}()},
\code{\link{getSpatialNetwork}()},
\code{\link{get_NearestNetwork}()},
\code{\link{get_dimReduction}()},
\code{\link{get_expression_values}()},
\code{\link{get_feature_info}()},
\code{\link{get_giottoImage}()},
\code{\link{get_multiomics}()},
\code{\link{get_polygon_info}()},
\code{\link{get_spatialGrid}()},
\code{\link{get_spatialNetwork}()},
\code{\link{get_spatial_enrichment}()},
\code{\link{get_spatial_locations}()}
}
\concept{functions to get data from giotto object}
\concept{multiomics accessor functions}
