% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONTraC_wrapper.R
\name{getONTraCv1Input}
\alias{getONTraCv1Input}
\title{getONTraCv1Input}
\usage{
getONTraCv1Input(
  gobject,
  cell_type,
  output_path = getwd(),
  spat_unit = NULL,
  feat_type = NULL,
  verbose = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{cell_type}{the cell type column name in the metadata}

\item{output_path}{the path to save the output file}

\item{spat_unit}{spatial unit (e.g. "cell")}

\item{feat_type}{feature type (e.g. "rna", "dna", "protein")}

\item{verbose}{be verbose}
}
\value{
data.table with columns: Cell_ID, Sample, x, y, Cell_Type
}
\description{
generate the input data for ONTraC v1
}
\details{
This function generate the input data for ONTraC v1
}
\examples{
g <- GiottoData::loadGiottoMini("visium")

getONTraCv1Input(
  gobject = g,
  cell_type = "custom_leiden"
)
}
