% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{getSpatialEnrichment}
\alias{getSpatialEnrichment}
\title{Get spatial enrichment}
\usage{
getSpatialEnrichment(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  enrichm_name = "DWLS",
  output = c("spatEnrObj", "data.table"),
  copy_obj = TRUE,
  set_defaults = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit (e.g. "cell")}

\item{feat_type}{feature type (e.g. "rna", "dna", "protein")}

\item{enrichm_name}{name of spatial enrichment results. Default "DWLS"}

\item{output}{what format in which to get information (e.g. "data.table")}

\item{copy_obj}{whether to deep copy/duplicate when getting the object (default = TRUE)
expression and spat_info are not expected to exist.}

\item{set_defaults}{set default spat_unit and feat_type. Change to FALSE only when}
}
\value{
data.table with fractions
}
\description{
Function to get a spatial enrichment data.table
}
\seealso{
Other spatial enrichment data accessor functions: 
\code{\link{get_spatial_enrichment}()},
\code{\link{setSpatialEnrichment}()},
\code{\link{set_spatial_enrichment}()}

Other functions to get data from giotto object: 
\code{\link{getDimReduction}()},
\code{\link{getExpression}()},
\code{\link{getFeatureInfo}()},
\code{\link{getGiottoImage}()},
\code{\link{getMultiomics}()},
\code{\link{getNearestNetwork}()},
\code{\link{getPolygonInfo}()},
\code{\link{getSpatialGrid}()},
\code{\link{getSpatialLocations}()},
\code{\link{getSpatialNetwork}()},
\code{\link{get_NearestNetwork}()},
\code{\link{get_dimReduction}()},
\code{\link{get_expression_values}()},
\code{\link{get_feature_info}()},
\code{\link{get_giottoImage}()},
\code{\link{get_multiomics}()},
\code{\link{get_polygon_info}()},
\code{\link{get_spatialGrid}()},
\code{\link{get_spatialNetwork}()},
\code{\link{get_spatial_enrichment}()},
\code{\link{get_spatial_locations}()}
}
\concept{functions to get data from giotto object}
\concept{spatial enrichment data accessor functions}
