% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{getSpatialNetwork}
\alias{getSpatialNetwork}
\title{Get spatial network}
\usage{
getSpatialNetwork(
  gobject,
  spat_unit = NULL,
  name = NULL,
  output = c("spatialNetworkObj", "networkDT", "networkDT_before_filter", "outputObj"),
  set_defaults = TRUE,
  copy_obj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit (e.g. "cell")}

\item{name}{name of spatial network}

\item{output}{object type to return as. Options: 'spatialNetworkObj' (default),
'networkDT' and 'networkDT_before_filter' for data.table outputs.}

\item{set_defaults}{set default spat_unit and feat_type. Change to FALSE only when}

\item{copy_obj}{whether to copy/duplicate when getting the object (default = TRUE)}

\item{verbose}{be verbose}
}
\description{
Function to get a spatial network
}
\seealso{
Other spatial network data accessor functions: 
\code{\link{get_spatialNetwork}()},
\code{\link{setSpatialNetwork}()},
\code{\link{set_spatialNetwork}()}

Other functions to get data from giotto object: 
\code{\link{getDimReduction}()},
\code{\link{getExpression}()},
\code{\link{getFeatureInfo}()},
\code{\link{getGiottoImage}()},
\code{\link{getMultiomics}()},
\code{\link{getNearestNetwork}()},
\code{\link{getPolygonInfo}()},
\code{\link{getSpatialEnrichment}()},
\code{\link{getSpatialGrid}()},
\code{\link{getSpatialLocations}()},
\code{\link{get_NearestNetwork}()},
\code{\link{get_dimReduction}()},
\code{\link{get_expression_values}()},
\code{\link{get_feature_info}()},
\code{\link{get_giottoImage}()},
\code{\link{get_multiomics}()},
\code{\link{get_polygon_info}()},
\code{\link{get_spatialGrid}()},
\code{\link{get_spatialNetwork}()},
\code{\link{get_spatial_enrichment}()},
\code{\link{get_spatial_locations}()}
}
\concept{functions to get data from giotto object}
\concept{spatial network data accessor functions}
