% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visuals.R
\name{ggplot_save_function}
\alias{ggplot_save_function}
\title{ggplot_save_function}
\usage{
ggplot_save_function(gobject, plot_object = NULL, save_dir = NULL,
  save_folder = NULL, save_name = NULL, save_format = NULL,
  show_saved_plot = F, scale = 1, width = NA, height = NA,
  units = c("in", "cm", "mm"), dpi = 300, limitsize = TRUE)
}
\arguments{
\item{gobject}{giotto object}

\item{save_dir}{directory to save to}

\item{save_folder}{folder in save_dir to save to}

\item{save_name}{name of plot}

\item{save_format}{format (e.g. png, tiff, pdf, ...)}

\item{show_saved_plot}{load & display the saved plot}

\item{scale}{scale}

\item{width}{width}

\item{units}{units}

\item{dpi}{Plot resolution}

\item{limitsize}{When TRUE (the default), ggsave will not save images larger than 50x50 inches, to prevent the common error of specifying dimensions in pixels.}

\item{heigth}{height}
}
\description{
Function to automatically save plots to directory of interest
}
\examples{
    ggplot_save_function(gobject)
}
\seealso{
\code{\link{ggplot2::ggsave}}
}
