% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{giotto-class}
\alias{giotto-class}
\alias{giotto}
\title{S4 giotto Class}
\description{
Framework of giotto object to store and work with spatial expression data
}
\details{
[\strong{expression}] There are several ways to provide expression information:

[\strong{expression_feat}] The different features or modalities such as rna, protein, metabolites, ...
that are provided in the expression slot.
}
\section{Slots}{

\describe{
\item{\code{expression}}{expression information}

\item{\code{expression_feat}}{available features (e.g. rna, protein, ...)}

\item{\code{spatial_locs}}{spatial location coordinates for cells/spots/grids}

\item{\code{spatial_info}}{information about spatial units (Giotto spatVector)}

\item{\code{cell_metadata}}{metadata for cells}

\item{\code{feat_metadata}}{metadata for available features}

\item{\code{feat_info}}{information about features (Giotto spatVector)}

\item{\code{cell_ID}}{unique cell IDs}

\item{\code{feat_ID}}{unique feature IDs for all features or modalities}

\item{\code{spatial_network}}{spatial network in data.table/data.frame format}

\item{\code{spatial_grid}}{spatial grid in data.table/data.frame format}

\item{\code{spatial_enrichment}}{slot to save spatial enrichment-like results}

\item{\code{dimension_reduction}}{slot to save dimension reduction coordinates}

\item{\code{nn_network}}{nearest neighbor network in igraph format}

\item{\code{images}}{slot to store giotto images}

\item{\code{largeImages}}{slot to store giottoLargeImage objects}

\item{\code{parameters}}{slot to save parameters that have been used}

\item{\code{instructions}}{slot for global function instructions}

\item{\code{offset_file}}{offset file used to stitch together image fields}

\item{\code{OS_platform}}{Operating System to run Giotto analysis on}

\item{\code{join_info}}{information about joined Giotto objects}

\item{\code{multiomics}}{multiomics integration results}
}}

\concept{giotto object}
