% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_help.R
\name{giotto_alloc_dt_slots}
\alias{giotto_alloc_dt_slots}
\title{Over-allocation for giotto DT-based info}
\usage{
giotto_alloc_dt_slots(gobject)
}
\arguments{
\item{gobject}{giotto object}
}
\description{
Finds DT based objects, overallocates the data.tables, then sets
the objects back in the giotto object
}
\keyword{internal}
