% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_visuals.R
\name{giotto_point}
\alias{giotto_point}
\title{giotto_point}
\usage{
giotto_point(plot_method = c("ggplot", "scattermore", "scattermost"), ...)
}
\arguments{
\item{\dots}{geom_point parameters}
}
\description{
uses ggplot::geom_point, scattermore::geom_scattermore or scattermore::geom_scattermost
}
\keyword{internal}
