% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{giottoImage-class}
\alias{giottoImage-class}
\alias{giottoImage}
\title{S4 giottoImage Class}
\description{
Framework of giotto object to store and work with spatial expression data
}
\details{
[\strong{mg_object}] Core object is any image that can be read by the magick package

[\strong{boundaries}] Boundary adjustments can be used to manually or
automatically through a script adjust the image with the spatial data.
}
\section{Slots}{

\describe{
\item{\code{name}}{name of Giotto image}

\item{\code{mg_object}}{magick image object}

\item{\code{minmax}}{minimum and maximum of associated spatial location coordinates}

\item{\code{boundaries}}{x and y coordinate adjustments (default to 0)}

\item{\code{scale_factor}}{image scaling relative to spatial locations}

\item{\code{resolution}}{spatial location units covered per pixel}

\item{\code{file_path}}{file path to the image if given}

\item{\code{OS_platform}}{Operating System to run Giotto analysis on}
}}

\concept{giotto image object}
