% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{giottoLargeImage-class}
\alias{giottoLargeImage-class}
\alias{giottoLargeImage}
\title{S4 giottoLargeImage Class}
\description{
class to handle images too large to load in normally through magick
}
\section{Slots}{

\describe{
\item{\code{name}}{name of large Giotto image}

\item{\code{raster_object}}{terra raster object}

\item{\code{extent}}{tracks the extent of the raster object. Note that most processes should rely on the extent of the raster object instead of this.}

\item{\code{overall_extent}}{terra extent object covering the original extent of image}

\item{\code{scale_factor}}{image scaling relative to spatial locations}

\item{\code{resolution}}{spatial location units covered per pixel}

\item{\code{max_intensity}}{value to set as maximum intensity in color scaling}

\item{\code{min_intensity}}{minimum value found}

\item{\code{is_int}}{values are integers}

\item{\code{file_path}}{file path to the image if given}

\item{\code{OS_platform}}{Operating System to run Giotto analysis on}
}}

\concept{giotto object image}
