% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interoperability.R
\name{giottoMasterToSuite}
\alias{giottoMasterToSuite}
\title{Convert a master Giotto object to suite}
\usage{
giottoMasterToSuite(gobject, expression_feat = "rna")
}
\arguments{
\item{gobject}{A Giotto object created with master version}

\item{expression_feat}{available features (e.g. rna, protein, ...)}
}
\value{
A Giotto object compatible with suite version
}
\description{
Convert a master Giotto object to suite
}
