% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_structures.R
\docType{class}
\name{giottoPoints-class}
\alias{giottoPoints-class}
\alias{giottoPoints}
\title{S4 giotto points Class}
\description{
Giotto class to store and operate on points data
}
\details{
Contains vector-type feature data
}
\section{Slots}{

\describe{
\item{\code{feat_type}}{name of feature type}

\item{\code{spatVector}}{terra spatVector to store point shapes}

\item{\code{networks}}{feature networks}
}}

\keyword{class}
\keyword{giotto,}
\keyword{points,}
