% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interoperability.R
\name{giottoToAnnData}
\alias{giottoToAnnData}
\title{Convert Giotto to anndata}
\usage{
giottoToAnnData(
  gobject = NULL,
  spat_unit = NULL,
  feat_type = NULL,
  python_path = NULL,
  save_directory = NULL
)
}
\arguments{
\item{gobject}{giotto object to be converted}

\item{spat_unit}{spatial unit which will be used in conversion.}

\item{feat_type}{feature type which will be used in conversion.}

\item{python_path}{path to python executable within a conda/miniconda environment}

\item{save_directory}{directory in which the file will be saved.}
}
\value{
vector containing .h5ad file path(s)
}
\description{
Converts a Giotto object to a spatial anndata (e.g. scanpy) .h5ad file
}
\details{
Function in beta. Converts a Giotto object into .h5ad file(s).

If there are multiple spatial units and/or feature types, but only
one spatial unit and/or feature type is specified, then only the
specified spatial unit and/or feature type will be used. If NULL,
by default, all spatial units will be used in conversion.

If multiple spatial units or feature types are specified, multiple
AnnData object will be created and returned.

The save_directory will be created if it does not already exist.
The default save_directory is the working directory.
}
