% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactivity.R
\name{giottoToAnndataZarr}
\alias{giottoToAnndataZarr}
\title{Create a local anndata zarr folder}
\usage{
giottoToAnndataZarr(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  expression = "raw",
  output_path
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit (e.g. "cell")}

\item{feat_type}{feature type (e.g. "rna", "dna", "protein")}

\item{expression}{expression values to extract (e.g. "raw", "normalized", "scaled")}

\item{output_path}{path to create and save the anndata zarr folder}
}
\value{
local anndata zarr folder
}
\description{
Create a local anndata zarr folder
}
\examples{
# using the mini visium object
giotto_object <- GiottoData::loadGiottoMini("visium")

giottoToAnndataZarr(giotto_object,
    expression = "raw",
    output_path = tempdir()
)

# using the mini vizgen object
giotto_object <- GiottoData::loadGiottoMini("vizgen")

giottoToAnndataZarr(giotto_object,
    spat_unit = "aggregate",
    expression = "scaled",
    output_path = tempdir()
)
}
