% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interoperability.R
\name{giottoToSeuratV5}
\alias{giottoToSeuratV5}
\title{Convert Giotto to Seurat V5}
\usage{
giottoToSeuratV5(gobject, spat_unit = NULL, obj_use = NULL, ...)
}
\arguments{
\item{gobject}{Giotto object}

\item{spat_unit}{spatial unit (e.g. 'cell')}

\item{obj_use}{Giotto object (deprecated, use gobject)}

\item{...}{additional params to pass to \code{\link{get_spatial_locations}}}
}
\value{
Seurat object
}
\description{
Converts Giotto object into a Seurat object. This functions extracts
specific sets of data belonging to specified spatial unit.
The default values are 'cell' and 'rna' respectively.
}
