% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interoperability.R
\name{giottoToSpatialExperiment}
\alias{giottoToSpatialExperiment}
\title{Utility function to convert a Giotto object to a SpatialExperiment object.}
\usage{
giottoToSpatialExperiment(giottoObj)
}
\arguments{
\item{giottoObj}{Input Giotto object to convert to a SpatialExperiment object.}
}
\value{
A SpatialExperiment object that contains data from the input Giotto object.
}
\description{
Utility function to convert a Giotto object to a SpatialExperiment object.
}
