% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gobject_mini_vizgen}
\alias{gobject_mini_vizgen}
\title{gobject_mini_vizgen}
\format{
An object of class \code{"giotto"};
see \code{\link[Giotto]{createGiottoObject}} or
 \code{\link[Giotto]{createGiottoObjectSubcellular}}.
}
\usage{
loadGiottoMini(dataset = 'vizgen')
}
\description{
Mini Giotto object created from the Mouse Brain Vizgen data.
}
\examples{

\dontrun{

gvizg = loadGiottoMini(dataset = 'vizgen')

spatDimPlot2D(gobject = gvizg,
             spat_unit = 'z1',
             show_image = F,
             largeImage_name = 'polyT_z1',
             cell_color = 'leiden_clus',
             spat_point_size = 2,
             dim_point_size = 2)}
}
\references{
www.vizgen.com
}
\keyword{datasets}
