% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_structures.R
\name{gpoly_from_dfr_smoothed_wrapped}
\alias{gpoly_from_dfr_smoothed_wrapped}
\title{Polygon creation and smoothing for parallel}
\usage{
gpoly_from_dfr_smoothed_wrapped(
  segmdfr,
  name = "cell",
  calc_centroids = FALSE,
  smooth_polygons = FALSE,
  vertices = 20L,
  k = 3L,
  set_neg_to_zero = TRUE,
  skip_eval_dfr = FALSE,
  copy_dt = TRUE,
  verbose = TRUE
)
}
\description{
Polygon creation and smoothing for parallel
}
\keyword{internal}
