% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_structures.R
\name{hexVertices}
\alias{hexVertices}
\title{Generate regular hexagon vertices}
\usage{
hexVertices(radius, major_axis = c("v", "h"))
}
\arguments{
\item{radius}{radius of the hexagon}

\item{major_axis}{orientation of the major axis 'v' is vertical (default)
and 'h' is horizontal}
}
\value{
a data.table of regular hexagon vertices
}
\description{
Generates vertex coordinates for a regular hexagon.
}
\seealso{
polyStamp circleVertices rectVertices
}
