% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_enrichment.R
\name{hyperGeometricEnrich}
\alias{hyperGeometricEnrich}
\title{hyperGeometricEnrich}
\usage{
hyperGeometricEnrich(
  gobject,
  sign_matrix,
  expression_values = c("normalized", "scaled", "custom"),
  reverse_log_scale = TRUE,
  logbase = 2,
  top_percentage = 5,
  output_enrichment = c("original", "zscore")
)
}
\arguments{
\item{gobject}{Giotto object}

\item{sign_matrix}{Matrix of signature genes for each cell type / process}

\item{expression_values}{expression values to use}

\item{reverse_log_scale}{reverse expression values from log scale}

\item{logbase}{log base to use if reverse_log_scale = TRUE}

\item{top_percentage}{percentage of cells that will be considered to have gene expression with matrix binarization}

\item{output_enrichment}{how to return enrichment output}
}
\value{
data.table with enrichment results
}
\description{
Function to calculate gene signature enrichment scores per spatial position using a hypergeometric test.
}
\details{
The enrichment score is calculated based on the p-value from the
hypergeometric test, -log10(p-value).
}
\examples{
    hyperGeometricEnrich(gobject)
}
