% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_structures.R
\name{identify_background_range_polygons}
\alias{identify_background_range_polygons}
\title{Identify background range polygons}
\usage{
identify_background_range_polygons(spatVector)
}
\description{
function to remove background polygon based on largest range
}
\keyword{internal}
