% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_clusters.R
\name{identifyTMAcores}
\alias{identifyTMAcores}
\title{Split cluster annotations based on a spatial network}
\usage{
identifyTMAcores(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  spatial_network_name = "Delaunay_network",
  core_id_name = "core_id",
  id_fmt = "\%d",
  include_all_ids = TRUE,
  missing_id_name = "not_connected",
  min_nodes = 5,
  join_split_cores = TRUE,
  join_tolerance = 1.2,
  return_gobject = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit (e.g. "cell")}

\item{feat_type}{feature type (e.g. "rna", "dna", "protein")}

\item{spatial_network_name}{character. Name of spatial network to use}

\item{core_id_name}{metadata column name for the core information}

\item{id_fmt}{character. [sprintf] formatting to use for core ids}

\item{include_all_ids}{logical. Include all ids, including vertex ids not
found in the spatial network}

\item{missing_id_name}{character. Name for nodes that are not connected to
a core.}

\item{min_nodes}{numeric. Minimal number of nodes to not be considered
an unconnected group.}

\item{join_split_cores}{logical. Attempt to repair core IDs when a core
is split down the middle and detected as two different cores.}

\item{join_tolerance}{numeric. Max ratio allowed relative to previous max
core convex hull area when determining if a pair of cores should be joined.}

\item{return_gobject}{logical. Return giotto object}
}
\value{
cluster annotations
}
\description{
Split cluster annotations based on a spatial network
}
