% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_clusters.R
\name{igraph_remove_hetero_edges}
\alias{igraph_remove_hetero_edges}
\title{Remove hetero edges from igraph}
\usage{
igraph_remove_hetero_edges(g, clus_attr)
}
\arguments{
\item{g}{igraph object}

\item{clus_attr}{character. A categorical node attribute}
}
\value{
igraph
}
\description{
Given an igraph `g` and set of node attributes `clus_att` that encode
different spatial clusters, remove edges that connect non-similar nodes.
This can be used when data is already clustered, but these clusters should
be further broken up based on whether they are spatially touching.
}
\keyword{internal}
