% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_clusters.R
\name{igraph_vertex_membership}
\alias{igraph_vertex_membership}
\title{igraph vertex membership}
\usage{
igraph_vertex_membership(g, clus_name)
}
\arguments{
\item{g}{igraph}

\item{clus_name}{character. name to assign column of clustering info}
}
\value{
data.table
}
\description{
Get which weakly connected set of vertices each vertex is part of
}
\keyword{internal}
