% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{init_cell_metadata}
\alias{init_cell_metadata}
\title{Initialize cell metadata slot}
\usage{
init_cell_metadata(gobject, provenance = NULL)
}
\arguments{
\item{gobject}{giotto object}

\item{provenance}{provenance information (optional)
and feature type in the giotto object.}
}
\description{
Generate cellMetaObjs to hold cell metadata for each spatial unit
}
\keyword{internal}
