% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{init_feat_metadata}
\alias{init_feat_metadata}
\title{Initialize feature metadata slot}
\usage{
init_feat_metadata(gobject, provenance = NULL)
}
\arguments{
\item{gobject}{giotto object}

\item{provenance}{provenance information (optional)}
}
\description{
Generate featMetaObjs to hold all feature metadata for each spatial unit
and feature type in the giotto object.
}
\keyword{internal}
