% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_hmrf.R
\name{initHMRF_V2}
\alias{initHMRF_V2}
\title{initHMRF_V2}
\usage{
initHMRF_V2(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  expression_values = c("scaled", "normalized", "custom"),
  spatial_network_name = "Delaunay_network",
  use_spatial_genes = c("binSpect", "silhouetteRank"),
  use_score = FALSE,
  gene_list_from_top = 2500,
  filter_method = c("none", "elbow"),
  user_gene_list = NULL,
  use_pca = FALSE,
  use_pca_dim = 1:20,
  gene_samples = 500,
  gene_sampling_rate = 2,
  gene_sampling_seed = 10,
  use_metagene = FALSE,
  cluster_metagene = 50,
  top_metagene = 20,
  existing_spatial_enrichm_to_use = NULL,
  use_neighborhood_composition = FALSE,
  spatial_network_name_for_neighborhood = NULL,
  metadata_to_use = NULL,
  hmrf_seed = 100,
  cl.method = c("km", "leiden", "louvain"),
  resolution.cl = 1,
  k = 10,
  tolerance = 1e-05,
  zscore = c("none", "rowcol", "colrow"),
  nstart = 1000,
  factor_step = 1.05,
  python_path = NULL
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit}

\item{feat_type}{feature type}

\item{expression_values}{expression values to use}

\item{spatial_network_name}{name of spatial network to use for HMRF}

\item{use_spatial_genes}{which of Giotto's spatial genes to use}

\item{use_score}{use score as gene selection criterion (applies when use_spatial_genes=silhouetteRank)}

\item{gene_list_from_top}{total spatial genes before sampling}

\item{filter_method}{filter genes by top or by elbow method, prior to sampling}

\item{user_gene_list}{user-specified genes (optional)}

\item{use_pca}{if PCA is used on the spatial gene expression value for clustering}

\item{use_pca_dim}{dimensions of the PCs of the selected expression}

\item{gene_samples}{number of spatial gene subset to use for HMRF}

\item{gene_sampling_rate}{parameter (1-50) controlling proportion of gene samples from different module when sampling, 1 corresponding to equal gene samples between different modules; 50 corresponding to gene samples proportional to module size.}

\item{gene_sampling_seed}{random number seed to sample spatial genes}

\item{use_metagene}{if metagene expression is used for clustering}

\item{cluster_metagene}{number of metagenes to use}

\item{top_metagene}{= number of genes in each cluster for the metagene calculation}

\item{existing_spatial_enrichm_to_use}{name of existing spatial enrichment result to use}

\item{use_neighborhood_composition}{if neighborhood composition is used for hmrf}

\item{spatial_network_name_for_neighborhood}{spatial network used to calculate neighborhood composition}

\item{metadata_to_use}{metadata used to calculate neighborhood composition}

\item{hmrf_seed}{random number seed to generate initial mean vector of HMRF model}

\item{cl.method}{clustering method to calculate the initial mean vector, selecting from 'km', 'leiden', or 'louvain'}

\item{resolution.cl}{resolution of Leiden or Louvain clustering}

\item{k}{number of HMRF domains}

\item{tolerance}{error tolerance threshold}

\item{zscore}{type of zscore to use}

\item{nstart}{number of Kmeans initializations from which to select the best initialization}

\item{factor_step}{dampened factor step}

\item{python_path}{python_path}
}
\value{
This function returns a list containing:
\itemize{
\item \strong{y} (expression), nei (neighborhood structure)
\item \strong{numnei} (number of neighbors)
\item \strong{blocks} (graph colors)
\item \strong{damp} (dampened factor)
\item \strong{mu} (mean)
\item \strong{sigma} (covariance)
\item \strong{k}
\item \strong{genes}
\item \strong{edgelist}
\item \strong{init.cl} (initial clusters)
\item \strong{spat_unit}
\item \strong{feat_type}
}

This information is needed for the second step, \code{doHMRF()}.
}
\description{
Run initialzation for HMRF model
}
\details{
This function is the initialization step of HMRF domain clustering. This
function performs 3 major steps:\cr\cr
\strong{First}, the user specifies which of Giotto's spatial genes to run through
param \code{use_spatial_genes}. Spatial genes have been stored in the gene metadata
table. A first pass of genes will filter genes that are not significantly
spatial, as determined by \code{filter_method.} If \code{filter_method} is "none", then
the top 2500 (\code{gene_list_from_top}) genes ranked by pvalue are considered
spatial. If \code{filter_method} is "elbow", then the exact cutoff is determined'
by the elbow in the -log10 P-value vs. gene rank plot.\cr\cr
\strong{Second}, users have a few options to decrease the dimension of the spatial
genes for clustering, listed with selection priority:
\enumerate{
\item use PCA of the spatial gene expressions (selected by \code{use_pca})
\item use metagene expressions (selected by \code{use_metagene})
\item sampling to select 500 spatial genes (controlled by \code{gene_samples}).
}

\cr
\strong{Third}, once spatial genes are finalized, we are using clustering method
to initialize HMRF.
Instead of select spatial genes for domain clustering, HMRF method can also
be applied on unit neighborhood composition of any group membership
(such as cell types), specified by parameter: \code{use_neighborhood_composition},
\code{spatial_network_name_for_neighborhood} and \code{metadata_to_use}.
HMRF also allows the user to do clustering with any customized spatial
enrichment matrix (\code{existing_spatial_enrichm_to_use}).
There are 3 clustering algorithm: K-means, Leiden, and Louvain to determine
initial centroids of HMRF. The initialization is then finished.
}
