% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_section.R
\name{insertCrossSectionGenePlot3D}
\alias{insertCrossSectionGenePlot3D}
\title{insertCrossSectionGenePlot3D}
\usage{
insertCrossSectionGenePlot3D(
  gobject,
  crossSection_obj = NULL,
  name = NULL,
  spatial_network_name = "Delaunay_network",
  mesh_grid_color = "#1f77b4",
  mesh_grid_width = 3,
  mesh_grid_style = "dot",
  sdimx = "sdimx",
  sdimy = "sdimy",
  sdimz = "sdimz",
  show_other_cells = F,
  axis_scale = c("cube", "real", "custom"),
  custom_ratio = NULL,
  show_plot = NA,
  return_plot = NA,
  save_plot = NA,
  save_param = list(),
  default_save_name = "spatGenePlot3D_with_cross_section",
  ...
)
}
\arguments{
\item{gobject}{giotto object}

\item{crossSection_obj}{cross section object as alternative input. default = NULL.}

\item{name}{name of virtual cross section to use}

\item{spatial_network_name}{name of spatial network to use}

\item{mesh_grid_color}{color for the meshgrid lines}

\item{mesh_grid_width}{width for the meshgrid lines}

\item{mesh_grid_style}{style for the meshgrid lines}

\item{sdimx}{x-axis dimension name (default = 'sdimx')}

\item{sdimy}{y-axis dimension name (default = 'sdimy')}

\item{sdimz}{z-axis dimension name (default = 'sdimy')}

\item{show_other_cells}{display not selected cells}

\item{axis_scale}{axis_scale}

\item{custom_ratio}{custom_ratio}

\item{show_plot}{show plots}

\item{return_plot}{return ggplot object}

\item{save_plot}{directly save the plot [boolean]}

\item{save_param}{list of saving parameters from \code{\link{all_plots_save_function}}}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}

\item{...}{parameters for spatGenePlot3D}
}
\value{
ggplot
}
\description{
Visualize cells and gene expression in a virtual cross section according to spatial coordinates
}
\details{
Description of parameters.
}
