% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_section.R
\name{insertCrossSectionSpatPlot3D}
\alias{insertCrossSectionSpatPlot3D}
\title{insertCrossSectionSpatPlot3D}
\usage{
insertCrossSectionSpatPlot3D(
  gobject,
  crossSection_obj = NULL,
  name = NULL,
  spatial_network_name = "Delaunay_network",
  mesh_grid_color = "#1f77b4",
  mesh_grid_width = 3,
  mesh_grid_style = "dot",
  sdimx = "sdimx",
  sdimy = "sdimy",
  sdimz = "sdimz",
  show_other_cells = F,
  axis_scale = c("cube", "real", "custom"),
  custom_ratio = NULL,
  default_save_name = "spat3D_with_cross_section",
  ...
)
}
\arguments{
\item{gobject}{giotto object}

\item{crossSection_obj}{cross section object as alternative input. default = NULL.}

\item{name}{name of virtual cross section to use}

\item{spatial_network_name}{name of spatial network to use}

\item{mesh_grid_color}{color for the meshgrid lines}

\item{mesh_grid_width}{width for the meshgrid lines}

\item{mesh_grid_style}{style for the meshgrid lines}

\item{sdimx}{x-axis dimension name (default = 'sdimx')}

\item{sdimy}{y-axis dimension name (default = 'sdimy')}

\item{sdimz}{z-axis dimension name (default = 'sdimy')}

\item{show_other_cells}{display not selected cells}

\item{axis_scale}{axis_scale}

\item{custom_ratio}{custom_ratio}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}

\item{...}{parameters for spatPlot3D}
}
\value{
ggplot
}
\description{
Visualize the meshgrid lines of cross section together with cells
}
\details{
Description of parameters.
}
