% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_environment.R
\name{install_giotto_environment}
\alias{install_giotto_environment}
\title{install_giotto_environment}
\usage{
install_giotto_environment(
  force_environment = FALSE,
  packages_to_install = c("pandas", "networkx", "python-igraph", "leidenalg",
    "python-louvain", "python.app", "scikit-learn"),
  verbose = TRUE
)
}
\description{
installation options of giotto environment
}
\keyword{internal}
