% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_environment.R
\name{install_giotto_environment_specific}
\alias{install_giotto_environment_specific}
\title{install_giotto_environment_specific}
\usage{
install_giotto_environment_specific(
  packages_to_install = c("pandas", "networkx", "python-igraph", "leidenalg",
    "python-louvain", "python.app", "scikit-learn"),
  python_version = "3.10.2",
  mini_install_path = NULL,
  verbose = TRUE
)
}
\description{
installation of giotto environment
}
\keyword{internal}
