% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_environment.R
\name{install_github_link_pip}
\alias{install_github_link_pip}
\title{Install Package from GitHub Link}
\usage{
install_github_link_pip(link = NULL, env = NULL)
}
\arguments{
\item{link}{link to github repository containing a python package, 
e.g. `git+https://github.com/TencentAILabHealthcare/pysodb.git`}

\item{env}{conda environment to which `link` will be installed via pip}
}
\description{
Installs `link` to python `env`
}
\keyword{internal}
