% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_environment.R
\name{install_py_pkg_reticulate}
\alias{install_py_pkg_reticulate}
\title{Install Python Package with Reticulate}
\usage{
install_py_pkg_reticulate(package = NULL, env = NULL)
}
\arguments{
\item{package}{name of python package}

\item{env}{name of the environment into which the python 
package should be installed.}
}
\description{
Install Python Package with Reticulate
}
\details{
Installs `package` to python `env` after prompting user.
Installation is done via `py_install` from the 
`reticulate` package.
}
\keyword{internal}
