% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONTraC_wrapper.R
\name{installONTraCEnvironment}
\alias{installONTraCEnvironment}
\title{installONTraCEnvironment}
\usage{
installONTraCEnvironment(
  python_version = "3.11.9",
  ontrac_version = "latest",
  mini_install_path = NULL,
  confirm = TRUE,
  envname = "giotto_ontrac_env",
  conda = "auto",
  force_miniconda = FALSE,
  force_environment = FALSE,
  verbose = NULL
)
}
\arguments{
\item{python_version}{python version to use within the giotto conda
environment. Default is v3.11.9}

\item{ontrac_version}{ONTraC version to install. Default is "latest"}

\item{mini_install_path}{(optional) desired miniconda installation location.
Default is chosen by `reticulate::install_miniconda()`}

\item{confirm}{whether to pause for confirmation of conda environment
install location (default = TRUE)}

\item{envname}{name to assign environment. Default = "giotto_ontrac_env"}

\item{conda}{either "auto" (default) to allow reticulate to handle it, or
the full filepath to the conda executable. You can also set the option
"reticulate.conda_binary" or `Sys.setenv()` "RETICULATE_CONDA" to tell
reticulate where to look.}

\item{force_miniconda}{force reinstallation of miniconda}

\item{force_environment}{force reinstallation of the giotto environment}

\item{verbose}{be verbose}

\item{packages_to_install}{python modules (packages) to install for Giotto.}
}
\value{
installs a giotto environment using the reticulate miniconda system
}
\description{
Installs a conda environment contains ONTraC. This
includes a miniconda installation and also a set of python packages that
Giotto may often use. See details for further information
}
\details{
This function will install a local conda environment using
the miniconda system as implemented by \pkg{reticulate}. Which contains
ONTraC and a set of python packages that Giotto may often use.
}
\examples{
installONTraCEnvironment()

}
