% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{instructions-generic}
\alias{instructions-generic}
\alias{instructions}
\alias{instructions<-}
\alias{instructions,giotto,missing-method}
\alias{instructions<-,giotto,ANY,missing-method}
\alias{instructions,giotto,character-method}
\alias{instructions<-,giotto,ANY,character-method}
\title{Access giotto instructions}
\usage{
instructions(gobject, param, ...)

\S4method{instructions}{giotto,missing}(gobject)

\S4method{instructions}{giotto,ANY,missing}(gobject) <- value

\S4method{instructions}{giotto,character}(gobject, param)

\S4method{instructions}{giotto,ANY,character}(gobject, param) <- value
}
\arguments{
\item{gobject}{giotto object}

\item{param}{Specific param in instructions to access or modify}

\item{value}{value to set}
}
\description{
Retrieve or set giotto instructions. Specific instructions can
be replaced using the \code{field} param
}
