% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kriging.R
\name{interpolateFeature}
\alias{interpolateFeature}
\alias{interpolateFeature,giotto,missing-method}
\alias{interpolateFeature,spatLocsObj,data.frame-method}
\title{Spatial feature interpolation}
\usage{
\S4method{interpolateFeature}{giotto,missing}(
  x,
  spat_unit = NULL,
  feat_type = NULL,
  feats,
  spatvalues_params = list(),
  spat_loc_name = "raw",
  ext = NULL,
  buffer = 50,
  name_fmt = "\%s",
  savedir = file.path(getwd(), "interp_rasters"),
  overwrite = FALSE,
  verbose = NULL,
  ...
)

\S4method{interpolateFeature}{spatLocsObj,data.frame}(
  x,
  y,
  ext = NULL,
  buffer = 50,
  rastersize = 500,
  name_fmt = "\%s",
  savedir = file.path(getwd(), "interp_rasters"),
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{x}{object containing coordinates to use interpolation with}

\item{spat_unit}{(optional) spatial unit to use}

\item{feat_type}{(optional) feature type to use}

\item{feats}{character vector. Features to interpolate from the `giotto`
object}

\item{spatvalues_params}{list. Additional list of parameters to pass to
[spatValues()] to help with data retrieval from `giotto` object}

\item{spat_loc_name}{character. Name of spatial locations to use. Values to
be interpolated are spatially mapped to these locations by cell_ID.}

\item{ext}{`SpatExtent`. (optional) extent across which to apply the
interpolation. If not provided, will default to the extent of the spatLocsObj
expanded by the value of `buffer`. It can be helpful to set this as the
extent of any polygons that will be used in aggregation.}

\item{buffer}{numeric. (optional) default buffer to expand derived extent by
if `ext` is not provided.}

\item{name_fmt}{character. sprintf fmt to apply to `feats` when naming the
resulting interpolation `giottoLargeImage` objects. Default is no change.}

\item{savedir}{character. Output directory. Default is a new `interp_rasters`
folder in working directory.}

\item{overwrite}{logical. Whether raster outputs should be overwritten if
the same `filename` is provided.}

\item{verbose}{be verbose}

\item{...}{additional params to pass downstream methods}

\item{y}{data.frame-like. Values for interpolation. Must also have a
`cell_ID` column and that matches with `x`.}

\item{rastersize}{numeric. Length of major axis in px of interpolation
raster to create.}
}
\value{
`giotto` method returns a `giotto` object with newly made appended
feature interpolation rasters as `giottoLargeImages`\cr
}
\description{
Spatial feature interpolation
}
\details{
The data.frame method returns a `giottoLargeImage` linked to an interpolated
raster that is written to disk as GeoTIFF.
}
