% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{iterLouvainCluster_community}
\alias{iterLouvainCluster_community}
\title{iterLouvainCluster_community}
\usage{
iterLouvainCluster_community(gobject, nr_rounds = 5,
  hvg_param = list(reverse_log_scale = T, difference_in_variance = 1,
  expression_values = "normalized"), hvg_min_perc_cells = 5,
  hvg_mean_expr_det = 1, pca_param = list(expression_values = "custom",
  scale.unit = T), nn_param = list(dimensions_to_use = 1:20),
  k_neighbors = 20, resolution = 1, python_path = NULL,
  nn_network_to_use = "sNN", network_name = "sNN.pca",
  name = "iter_clus", return_gobject = TRUE, ...)
}
\arguments{
\item{gobject}{giotto object}

\item{nr_rounds}{number of iterative rounds}

\item{hvg_param}{parameters for calculateHVG}

\item{hvg_min_perc_cells}{threshold for detection in min percentage of cells}

\item{hvg_mean_expr_det}{threshold for mean expression level in cells with detection}

\item{pca_param}{parameters for runPCA}

\item{nn_param}{parameters for parameters for runPCA}

\item{k_neighbors}{k for nn-network}

\item{resolution}{resolution for Leiden clustering}

\item{python_path}{python path to use for Leiden clustering}

\item{nn_network_to_use}{NN network to use}

\item{network_name}{NN network name}

\item{name}{name of clustering}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}

\item{...}{additional parameters}
}
\value{
giotto object appended with new cluster
}
\description{
cluster cells iteratively
}
\details{
Description of iterative clustering.
}
\examples{
    iterLouvainCluster_community(gobject)
}
