% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.R
\name{jackstrawPlot}
\alias{jackstrawPlot}
\title{jackstrawPlot}
\usage{
jackstrawPlot(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  expression_values = c("normalized", "scaled", "custom"),
  reduction = c("cells", "feats"),
  feats_to_use = NULL,
  center = FALSE,
  scale_unit = FALSE,
  ncp = 20,
  ylim = c(0, 1),
  iter = 10,
  threshold = 0.01,
  random_subset = NULL,
  set_seed = TRUE,
  seed_number = 1234,
  verbose = TRUE,
  show_plot = NULL,
  return_plot = NULL,
  save_plot = NULL,
  save_param = list(),
  default_save_name = "jackstrawPlot"
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit (e.g. "cell")}

\item{feat_type}{feature type (e.g. "rna", "dna", "protein")}

\item{expression_values}{expression values to use}

\item{reduction}{cells or genes}

\item{feats_to_use}{subset of features to use for PCA}

\item{center}{center data before PCA}

\item{scale_unit}{scale features before PCA}

\item{ncp}{number of principal components to calculate}

\item{ylim}{y-axis limits on jackstraw plot}

\item{iter}{number of iterations for jackstraw}

\item{threshold}{p-value threshold to call a PC significant}

\item{random_subset}{randomized subset of matrix to use to approximate but
speed up calculation}

\item{set_seed}{logical. whether to set a seed when random_subset is used}

\item{seed_number}{seed number to use when random_subset is used}

\item{verbose}{show progress of jackstraw method}

\item{show_plot}{logical. show plot}

\item{return_plot}{logical. return ggplot object}

\item{save_plot}{logical. save the plot}

\item{save_param}{list of saving parameters, see 
\code{\link{showSaveParameters}}}

\item{default_save_name}{default save name for saving, don't change, 
change save_name in save_param}
}
\value{
ggplot object for jackstraw method
}
\description{
Identify significant principal components (PCs)
}
\details{
The Jackstraw method uses the \code{\link[jackstraw]{permutationPA}}
function. By systematically permuting genes it identifies robust, and thus
significant, PCs. This implementation makes small modifications to SVD
calculation for improved efficiency and flexibility with different matrix
types. \cr
This implementation supports both dense and sparse input matrices. \cr

\strong{steps}
\enumerate{
\item Select singular values to calculate based on matrix dims and ncp
\item Find SVD to get variance explained of each PC
\item Randomly sample across features then re-calculate randomized variance
\item Determine P-value by comparing actual vs randomized explained variance,
indicating the significance of each PC
}
}
\examples{
g <- GiottoData::loadGiottoMini("visium")

jackstrawPlot(gobject = g)
}
