% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registration.R
\name{join_expression}
\alias{join_expression}
\title{join_expression}
\usage{
join_expression(expression_list, z_list)
}
\arguments{
\item{expression_list}{list of expression values to merge}

\item{z_list}{z values to use z stacking expression values}
}
\description{
joins expression list together while appending z value to cellIDs to ensure unique colnames
}
\keyword{internal}
