% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto.R
\name{joinGiottoObjects}
\alias{joinGiottoObjects}
\title{joinGiottoObjects}
\usage{
joinGiottoObjects(
  gobject_list,
  gobject_names = NULL,
  join_method = c("shift", "z_stack"),
  z_vals = 1000,
  x_shift = NULL,
  y_shift = NULL,
  x_padding = 0,
  y_padding = 0,
  verbose = TRUE
)
}
\arguments{
\item{gobject_list}{list of giotto objects}

\item{gobject_names}{unique giotto names for each giotto object}

\item{join_method}{method to join giotto objects}

\item{z_vals}{distance(s) along z-axis if method is z-stack}

\item{x_shift}{shift along x-axis if method is shift}

\item{y_shift}{shift along y-axis if method is shift}

\item{x_padding}{padding between datasets/images if method is shift}

\item{y_padding}{padding between datasets/images if method is shift}

\item{verbose}{be verbose}
}
\value{
giotto object
}
\description{
Function to join multiple giotto objects together
}
\concept{giotto}
