% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_help.R
\name{kmeans_arma_binarize}
\alias{kmeans_arma_binarize}
\title{kmeans_arma_binarize}
\usage{
kmeans_arma_binarize(x, n_iter = 5, set.seed = NULL)
}
\description{
create binarized scores from a vector using kmeans_arma
}
\keyword{internal}
