% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_help.R
\name{kmeans_arma_subset_binarize}
\alias{kmeans_arma_subset_binarize}
\title{kmeans_arma_subset_binarize}
\usage{
kmeans_arma_subset_binarize(
  x,
  n_iter = 5,
  extreme_nr = 20,
  sample_nr = 200,
  set.seed = NULL
)
}
\description{
create binarized scores from a subsetted vector using kmeans_arma
}
\keyword{internal}
