% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_help.R
\name{kmeans_binarize}
\alias{kmeans_binarize}
\title{kmeans_binarize}
\usage{
kmeans_binarize(x, nstart = 3, iter.max = 10, set.seed = NULL)
}
\description{
create binarized scores from a vector using kmeans
}
\keyword{internal}
