% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_help.R
\name{kmeans_binarize_wrapper}
\alias{kmeans_binarize_wrapper}
\title{kmeans_binarize_wrapper}
\usage{
kmeans_binarize_wrapper(
  expr_values,
  subset_feats = NULL,
  kmeans_algo = c("kmeans", "kmeans_arma", "kmeans_arma_subset"),
  nstart = 3,
  iter_max = 10,
  extreme_nr = 50,
  sample_nr = 50,
  seed = NULL
)
}
\value{
matrix
}
\description{
wrapper for different binarization functions
}
\keyword{internal}
