% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{lapply_flex}
\alias{lapply_flex}
\title{lapply_flex}
\usage{
lapply_flex(X, FUN, cores = NA, fun = NULL, ...)
}
\arguments{
\item{X}{list to use}

\item{FUN}{function to be performed}

\item{cores}{cores to use}

\item{fun}{deprecated. Backwards compatibility for FUN}

\item{...}{other arguments to pass}
}
\description{
lapply_flex
}
\keyword{internal}
