% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\docType{class}
\name{LeidenIgraphClusParam-class}
\alias{LeidenIgraphClusParam-class}
\alias{LeidenIgraphClusParam}
\title{Leiden Clustering}
\description{
Cluster cells using a NN-network and the Leiden community detection
algorithm as implemented in igraph.
}
\details{
This function is a wrapper for the Leiden algorithm implemented in igraph,
which can detect communities in graphs of millions of nodes (cells),
as long as they can fit in memory. See \code{\link[igraph]{cluster_leiden}}
for more information.
}
\section{params}{


\tabular{ll}{
\code{resolution} \tab numeric (default = 1). Clustering resolution. \cr
\code{n_iterations} \tab numeric (default = 20). Number of iterations to run
the Leiden algorithm. \cr
\code{weights} \tab (default = \code{NULL}) weights of edges. Set \code{NULL} to use
weights associated with the igraph network. Set \code{NA} if you don't want
to use weights. \cr
\code{beta} \tab character (default = 0.01). Leiden randomness \cr
\code{objective_function} \tab character (default = \code{"modularity"}) objective
function for the leiden algorithm. One of \code{"modularity"} or \code{"CPM"} \cr
\code{initial_membership} \tab (default = \code{NULL}) initial membership of cells
for the partition.
}
}

\examples{
g <- GiottoData::loadGiottoMini("visium")
clusterData(g, clusterParam("leiden_igraph", resolution = 0.5))
}
