% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\docType{class}
\name{LeidenPythonClusParam-class}
\alias{LeidenPythonClusParam-class}
\alias{LeidenPythonClusParam}
\title{Leiden Clustering}
\description{
Cluster cells using a NN-network and the Leiden community detection algorithm
This version is implemented via the python package \pkg{leidenalg}.
}
\details{
This implementation is a wrapper for the Leiden algorithm implemented in python,
which can detect communities in graphs of millions of nodes (cells),
as long as they can fit in memory. See the
\href{https://github.com/vtraag/leidenalg}{leidenalg}
github page or the
\href{https://leidenalg.readthedocs.io/en/stable/index.html}{readthedocs}
page for more information.
}
\section{params}{


\tabular{ll}{
\code{resolution} \tab numeric (default = 1). Clustering resolution. \cr
\code{n_iterations} \tab numeric (default = 20). Number of iterations to run
the Leiden algorithm. If the number of iterations is negative, the Leiden
algorithm is run until an iteration in which there was no improvement. \cr
\code{weight_col} \tab character. (default = \code{weight}). Weight column in
network information to use for edge weights. \cr
\code{partition_type} \tab character (default = \code{"RBConfigurationVertexPartition"}).
The type of partition to use for optimization. (one of
\code{"RBConfigurationVertexPartition"} or \code{"ModularityVertexPartition"}) \cr
\code{initial_membership} \tab (default = \code{NULL}) initial membership of cells
for the partition
}
}

\section{partition types available and information}{

\itemize{
\item \strong{RBConfigurationVertexPartition:} Implements Reichardt and Bornholdt’s
Potts model with a configuration null model. This quality function is
well-defined only for positive edge weights. This quality function uses a
linear resolution parameter.
\item \strong{ModularityVertexPartition:} Implements modularity. This quality
function is well-defined only for positive edge weights. It does \emph{not}
use the resolution parameter
}

Set \code{weight_col = NULL} to give equal weight (=1) to each edge.
}

\examples{
g <- GiottoData::loadGiottoMini("visium")
clusterData(g, clusterParam("leiden_python", resolution = 0.5))
}
