% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{libNorm_giotto}
\alias{libNorm_giotto}
\title{Normalize for library size}
\usage{
libNorm_giotto(mymatrix, scalefactor)
}
\arguments{
\item{mymatrix}{matrix object}

\item{scalefactor}{scalefactor}
}
\description{
Normalize for library size
}
\keyword{internal}
