% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{list_cell_id_names}
\alias{list_cell_id_names}
\title{List cell ID names}
\usage{
list_cell_id_names(gobject)
}
\arguments{
\item{gobject}{giotto object}
}
\value{
vector with names of available sets of cell_IDs
}
\description{
lists the available cell id names. In effect, these names are the
spat_units and poly info in the gobject
}
