% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{list_dim_reductions_names}
\alias{list_dim_reductions_names}
\title{list_dim_reductions_names}
\usage{
list_dim_reductions_names(
  gobject,
  data_type = "cells",
  spat_unit,
  feat_type,
  dim_type
)
}
\arguments{
\item{gobject}{giotto object}

\item{data_type}{cells or feats dim reduction}

\item{spat_unit}{spatial unit}

\item{feat_type}{feature type}

\item{dim_type}{dimensional reduction type (method)}
}
\value{
names pf dimension reduction object
}
\description{
return the available dimension reductions object names
}
\details{
function that can be used to find which names have been used
}
