% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{list_element_exists}
\alias{list_element_exists}
\title{Test if list element exists}
\usage{
list_element_exists(x, index)
}
\arguments{
\item{x}{list}

\item{index}{element index}
}
\value{
boolean
}
\description{
Test if nth element of list exists
}
\keyword{internal}
