% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{list_expression_names}
\alias{list_expression_names}
\title{list_expression_names}
\usage{
list_expression_names(gobject, spat_unit, feat_type)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit}

\item{feat_type}{feature type}
}
\value{
vector with names of available matrices
}
\description{
lists the available matrices names for a given spatial unit and feature type
}
