% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{list_feat_id_names}
\alias{list_feat_id_names}
\title{List feat ID names}
\usage{
list_feat_id_names(gobject)
}
\arguments{
\item{gobject}{giotto object}
}
\value{
vector with names of available sets of feat_IDs
}
\description{
lists the available feat id names In effect, these names are the
feat_types and feature info in the gobject
}
