% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{list_feature_info}
\alias{list_feature_info}
\title{list_feature_info}
\usage{
list_feature_info(gobject)
}
\arguments{
\item{gobject}{giotto object}
}
\value{
names of available feature information
}
\description{
return the available giotto spatial feature information
}
