% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{list_feature_info_names}
\alias{list_feature_info_names}
\title{list_feature_info_names}
\usage{
list_feature_info_names(gobject)
}
\arguments{
\item{gobject}{giotto object}
}
\value{
vector with names of available feature information
}
\description{
return the available names for giotto feature information
}
