% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{list_image_names}
\alias{list_image_names}
\title{list_image_names}
\usage{
list_image_names(gobject)
}
\arguments{
\item{gobject}{a giotto object}

\item{return}{return named list of available images}

\item{verbose}{verbosity of function}
}
\value{
a vector of giotto image names attached to the giotto object
}
\description{
Prints the available giotto images that are attached to the Giotto object
}
