% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{list_images}
\alias{list_images}
\title{list_images}
\usage{
list_images(gobject, img_type = NULL)
}
\arguments{
\item{gobject}{giotto object}

\item{img_type}{"image" or "largeImage"}
}
\value{
data.table of giotto image names attached to the giotto object
}
\description{
Prints the available giotto images that are attached to the Giotto object
}
