% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{list_images_names}
\alias{list_images_names}
\title{list_images_names}
\usage{
list_images_names(gobject, img_type)
}
\arguments{
\item{gobject}{a giotto object}

\item{img_type}{"image" or "largeImage"}
}
\value{
vector with names of available image names
}
\description{
return the available image names for a given image type that are attached to the Giotto object
}
