% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{list_nearest_networks}
\alias{list_nearest_networks}
\title{list_nearest_networks}
\usage{
list_nearest_networks(gobject, spat_unit = NULL)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit (e.g. "cell")}
}
\value{
data.table of names and locations of available nearest networks.
}
\description{
return the available nearest networks that are attached to the Giotto object
}
