% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{list_nearest_networks_names}
\alias{list_nearest_networks_names}
\title{list_nearest_networks_names}
\usage{
list_nearest_networks_names(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  nn_type = NULL
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit (e.g. "cell")}

\item{feat_type}{feature type (e.g. "rna", "dna", "protein")}

\item{nn_type}{nearest neighbor method (e.g. "sNN", "kNN")}
}
\value{
names of nearest neighbor network object
}
\description{
return the available nearest neighbor network object names
}
\details{
function that can be used to find which names have been used
}
