% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{list_spatial_enrichments_names}
\alias{list_spatial_enrichments_names}
\title{list_spatial_enrichments_names}
\usage{
list_spatial_enrichments_names(gobject, spat_unit, feat_type)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit (e.g. "cell")}

\item{feat_type}{feature type (e.g. "rna", "dna", "protein")}
}
\value{
vector of names for available spatial enrichments
}
\description{
returns the available spatial enrichment names for a given spatial unit
}
