% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{list_spatial_grids}
\alias{list_spatial_grids}
\title{list_spatial_grids}
\usage{
list_spatial_grids(gobject)
}
\arguments{
\item{gobject}{giotto object}
}
\value{
dataframe of names and locations of available spatial grids. col order matters
}
\description{
return the available spatial grids that are attached to the Giotto object
}
