% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{list_spatial_grids_names}
\alias{list_spatial_grids_names}
\title{list_spatial_grids_names}
\usage{
list_spatial_grids_names(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  return_uniques = FALSE
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit (e.g. "cell")}

\item{feat_type}{feature type (e.g. "rna", "dna", "protein")}

\item{return_uniques}{return unique nesting names (ignores if final object exists/is correct class)}
}
\value{
vector with names of available spatial grids names
}
\description{
return the available spatial grids name for a given spatial unit that are attached to the Giotto object
}
