% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{list_spatial_info_names}
\alias{list_spatial_info_names}
\title{list_spatial_info_names}
\usage{
list_spatial_info_names(gobject)
}
\arguments{
\item{gobject}{giotto object}
}
\value{
vector with names of available spatial polygon information
}
\description{
return the available names for giotto spatial polygon information
}
