% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{list_spatial_locations_names}
\alias{list_spatial_locations_names}
\title{list_spatial_locations_names}
\usage{
list_spatial_locations_names(gobject, spat_unit = NULL)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit (e.g. "cell")}
}
\value{
vector with names of available spatial locations
}
\description{
lists the available spatial location names for a given spatial unit
}
