% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONTraC_wrapper.R
\name{load_cell_niche_cluster_prob}
\alias{load_cell_niche_cluster_prob}
\title{load_cell_niche_cluster_prob}
\usage{
load_cell_niche_cluster_prob(
  gobject,
  ontrac_results_dir = getwd(),
  spat_unit = "cell",
  feat_type = "niche cluster",
  name = "prob"
)
}
\arguments{
\item{gobject}{giotto object}

\item{ontrac_results_dir}{the directory where the ONTraC results are saved}

\item{spat_unit}{spatial unit (e.g. "cell")}

\item{feat_type}{feature type (e.g. "rna", "dna", "protein")}

\item{name}{name for the probability matrix}
}
\value{
gobject with cell-niche cluster probability matrix
}
\description{
load cell-niche cluster probability
}
\details{
This function loads the ONTraC outputed cell-niche cluster
probability as an exprObj into the giotto object.
}
