% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONTraC_wrapper.R
\name{load_cell_NT_score}
\alias{load_cell_NT_score}
\title{load_cell_NT_score}
\usage{
load_cell_NT_score(gobject, ontrac_results_dir = getwd())
}
\arguments{
\item{gobject}{giotto object}

\item{ontrac_results_dir}{the directory where the ONTraC results are saved}
}
\value{
gobject with cell-level NT score
}
\description{
load cell-level NT score
}
\details{
This function loads the ONTraC outputed cell-level NT score
}
