% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{load_cosmx_folder_aggregate}
\alias{load_cosmx_folder_aggregate}
\title{Load CosMx folder aggregate info}
\usage{
load_cosmx_folder_aggregate(dir_items, cores, verbose = TRUE)
}
\arguments{
\item{cores}{how many cores or threads to use to read data if paths are provided}

\item{verbose}{be verbose when building Giotto object}
}
\description{
Load CosMx folder aggregate info
}
\keyword{internal}
