% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{load_cosmx_folder_subcellular}
\alias{load_cosmx_folder_subcellular}
\title{Load CosMx folder subcellular info}
\usage{
load_cosmx_folder_subcellular(dir_items, FOVs = NULL, cores, verbose = TRUE)
}
\arguments{
\item{FOVs}{field of views to load (only affects subcellular data and images)}

\item{cores}{how many cores or threads to use to read data if paths are provided}

\item{verbose}{be verbose when building Giotto object}
}
\description{
loads in the feature detections information. Note that the mask
images are still required for a working subcellular object, and those are loaded
in \code{\link{createGiottoCosMxObject_subcellular}}
}
\keyword{internal}
