% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R, R/convenience_create.R
\name{load_merscope_folder}
\alias{load_merscope_folder}
\alias{load_merscope_folder_subcellular}
\alias{load_merscope_folder_aggregate}
\title{Load MERSCOPE data from folder}
\usage{
load_merscope_folder(
  dir_items,
  data_to_use,
  FOVs = NULL,
  polygon_feat_types = 0:6,
  cores = NA,
  verbose = TRUE
)

load_merscope_folder_subcellular(
  dir_items,
  data_to_use,
  polygon_feat_types = 0:6,
  cores = NA,
  verbose = TRUE,
  FOVs = NULL
)

load_merscope_folder_aggregate(
  dir_items,
  data_to_use,
  cores = NA,
  verbose = TRUE
)

load_merscope_folder(
  dir_items,
  data_to_use,
  FOVs = NULL,
  polygon_feat_types = 0:6,
  cores = NA,
  verbose = TRUE
)

load_merscope_folder_subcellular(
  dir_items,
  data_to_use,
  polygon_feat_types = 0:6,
  cores = NA,
  verbose = TRUE,
  FOVs = NULL
)

load_merscope_folder_aggregate(
  dir_items,
  data_to_use,
  cores = NA,
  verbose = TRUE
)
}
\arguments{
\item{dir_items}{list of full filepaths from \code{\link{read_merscope_folder}}}

\item{data_to_use}{which of either the 'subcellular' or 'aggregate' information
to use for object creation}

\item{FOVs}{which FOVs to use when building the subcellular object. (default is NULL)
NULL loads all FOVs (very slow)}

\item{polygon_feat_types}{which Vizgen polygon z slices are loaded (There are 0 - 6)}

\item{cores}{how many cores or threads to use to read data if paths are provided}

\item{verbose}{be verbose when building Giotto object}
}
\value{
list of loaded-in MERSCOPE data

list of loaded-in MERSCOPE data
}
\description{
Load MERSCOPE data from folder

Load MERSCOPE data from folder
}
\section{Functions}{
\itemize{
\item \code{load_merscope_folder_subcellular()}: Load items for 'subcellular' workflow

\item \code{load_merscope_folder_aggregate()}: Load items for 'aggregate' workflow

\item \code{load_merscope_folder_subcellular()}: Load items for 'subcellular' workflow

\item \code{load_merscope_folder_aggregate()}: Load items for 'aggregate' workflow

}}
\keyword{internal}
