% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONTraC_wrapper.R
\name{load_nc_connectivity}
\alias{load_nc_connectivity}
\title{load_nc_connectivity}
\usage{
load_nc_connectivity(
  gobject,
  ontrac_results_dir = getwd(),
  spat_unit = "niche cluster",
  feat_type = "connectivity",
  name = "normalized"
)
}
\arguments{
\item{gobject}{giotto object}

\item{ontrac_results_dir}{the directory where the ONTraC results are saved}

\item{spat_unit}{spatial unit (e.g. "cell")}

\item{feat_type}{feature type (e.g. "rna", "dna", "protein")}

\item{name}{name for the connectivity matrix}
}
\value{
gobject with niche cluster connectivity matrix
}
\description{
load niche cluster connectivity
}
\details{
This function loads the ONTraC outputed niche cluster connectivity
matrix as an exprObj into the giotto object.
}
