% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONTraC_wrapper.R
\name{load_niche_cluster_nt_score}
\alias{load_niche_cluster_nt_score}
\title{load_niche_cluster_nt_score}
\usage{
load_niche_cluster_nt_score(
  gobject,
  ontrac_results_dir = getwd(),
  NTScore_dir = file.path(ontrac_results_dir, "NTScore_dir"),
  spat_unit = "niche cluster",
  feat_type = "connectivity"
)
}
\arguments{
\item{gobject}{giotto object}

\item{ontrac_results_dir}{the directory where the ONTraC results are saved.
Default is getwd()}

\item{NTScore_dir}{the directory to save the NTScore results. Default is
file.path(ontrac_results_dir, "NTScore_dir")}

\item{spat_unit}{spatial unit (e.g. "cell")}

\item{feat_type}{feature type (e.g. "rna", "dna", "protein")}
}
\value{
gobject with niche cluster NT score
}
\description{
load niche cluster NT score
}
\details{
This function loads the ONTraC outputed niche cluster NT score
into the giotto object.
}
