% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_xenium.R
\name{load_xenium_folder}
\alias{load_xenium_folder}
\alias{.load_xenium_folder}
\alias{.load_xenium_folder_csv}
\alias{.load_xenium_folder_parquet}
\title{Load xenium data from folder}
\usage{
.load_xenium_folder(
  path_list,
  load_format = "csv",
  data_to_use = "subcellular",
  h5_expression = "FALSE",
  h5_gene_ids = "symbols",
  gene_column_index = 1,
  cores,
  verbose = TRUE
)

.load_xenium_folder_csv(
  path_list,
  cores,
  data_to_use = "subcellular",
  h5_expression = FALSE,
  h5_gene_ids = "symbols",
  gene_column_index = 1,
  verbose = TRUE
)

.load_xenium_folder_parquet(
  path_list,
  cores,
  data_to_use = "subcellular",
  h5_expression = FALSE,
  h5_gene_ids = "symbols",
  gene_column_index = 1,
  verbose = TRUE
)
}
\arguments{
\item{path_list}{list of full filepaths from .read_xenium_folder}

\item{load_format}{files formats from which to load the data. Either `csv` or
`parquet` currently supported.}

\item{data_to_use}{which type(s) of expression data to build the gobject with
(e.g. default: \strong{'subcellular'}, 'aggregate', or 'all')}

\item{h5_expression}{(boolean) whether to load cell_feature_matrix from .h5
file. Default is \code{TRUE}}

\item{h5_gene_ids}{use gene symbols (default) or ensembl ids for the .h5 gene
expression matrix}

\item{gene_column_index}{which column from the features or genes .tsv file
to use for row ids}

\item{cores}{how many cores or threads to use to read data if paths are
provided}

\item{verbose}{be verbose when building Giotto object}
}
\value{
list of loaded in xenium data
}
\description{
Load xenium data from folder
}
\section{Functions}{
\itemize{
\item \code{.load_xenium_folder_csv()}: Load from csv files

\item \code{.load_xenium_folder_parquet()}: Load from parquet files

}}
\keyword{internal}
