% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_help.R
\name{loadGiotto}
\alias{loadGiotto}
\title{loadGiotto}
\usage{
loadGiotto(
  path_to_folder,
  load_params = list(),
  reconnect_giottoImage = TRUE,
  python_path = NULL,
  verbose = TRUE
)
}
\arguments{
\item{path_to_folder}{path to folder where Giotto object was stored with \code{\link{saveGiotto}}}

\item{load_params}{additional parameters for loading or reading giotto object}

\item{reconnect_giottoImage}{(default = TRUE) whether to attempt reconnection of magick based image objects}

\item{python_path}{(optional) manually set your python path}

\item{verbose}{be verbose}
}
\value{
Giotto object
}
\description{
Saves a Giotto object to a specific folder structure
}
\details{
Works together with \code{\link{saveGiotto}} to save and re-load
Giotto objects.
Additional load_params need to be provided as a list and will
go to \code{\link[base]{readRDS}} or \code{\link[qs]{qread}}
You can set the python path, alternatively it will look for an existing
Giotto python environment.
}
