% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_help.R
\name{loadGiottoMini}
\alias{loadGiottoMini}
\title{loadGiottoMini}
\usage{
loadGiottoMini(
  dataset = c("visium", "seqfish", "starmap", "vizgen"),
  python_path = NULL
)
}
\arguments{
\item{dataset}{mini dataset giotto object to load}

\item{python_path}{pythan path to use}
}
\description{
This function will automatically load one of the existing mini
giotto objects. These giotto objects can be used to test Giotto functions
and run examples. If no python path is provided it will try to find and use
the Giotto python environment. Images associated with the giotto mini objects
will be reconnected if possible.

Instructions, such as for saving plots, can be changed
using the \code{\link{changeGiottoInstructions}}
}
