% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONTraC_wrapper.R
\name{loadOntraCResults}
\alias{loadOntraCResults}
\title{loadOntraCResults}
\usage{
loadOntraCResults(gobject, ontrac_results_dir = getwd())
}
\arguments{
\item{gobject}{giotto object}

\item{ontrac_results_dir}{the directory where the ONTraC results are saved}
}
\value{
gobject with ONTraC results
}
\description{
load ONTraC results
}
\details{
This function loads the ONTraC results into the giotto object.
}
