% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\docType{class}
\name{LouvainCommunityClusParam-class}
\alias{LouvainCommunityClusParam-class}
\alias{LouvainCommunityClusParam}
\title{Louvain Clustering}
\description{
Cluster cells using a NN-network and the Louvain algorithm. This utilizes the
\{community\} package from python.
}
\section{params}{


\tabular{ll}{
\code{resolution} \tab numeric (default = 1). Clustering resolution. \cr
\code{weight_col} \tab character (default = \code{NULL}). Weight column name. \cr
\code{louv_random} \tab (default = \code{FALSE}) Will randomize the node evaluation
order and the community evaluation order to get different partitions at
each call
}
}

\examples{
g <- GiottoData::loadGiottoMini("visium")
clusterData(g, clusterParam("louvain_community", resolution = 0.5))
}
