% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\docType{class}
\name{LouvainMultinetClusParam-class}
\alias{LouvainMultinetClusParam-class}
\alias{LouvainMultinetClusParam}
\title{Louvain Clustering}
\description{
Cluster cells using a NN-network and the Louvain algorithm. This utilizes the
generalized Louvain implementation from  \pkg{multinet}.
}
\section{params}{


\tabular{ll}{
\code{gamma} \tab numeric (default = 1). Resolution parameter for modularity in the
generalized louvain method, \cr
\code{omega} \tab numeric (default = 1). Inter-layer weight parameter in the
generalized louvain method
}
}

\examples{
g <- GiottoData::loadGiottoMini("visium")
clusterData(g, clusterParam("louvain_multinet"))
}
