% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_enrichment.R
\name{makeSignMatrixDWLS}
\alias{makeSignMatrixDWLS}
\title{Build a DWLS signature matrix from a Giotto object}
\usage{
makeSignMatrixDWLS(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  expression_values = c("normalized", "scaled", "custom"),
  reverse_log = TRUE,
  log_base = 2,
  sign_gene,
  cell_type_vector,
  cell_type = NULL
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit (e.g. "cell")}

\item{feat_type}{feature type (e.g. "rna", "dna", "protein")}

\item{expression_values}{character. Which expression values to use, e.g.
"normalized". A method's own default is shown in its Usage section.}

\item{reverse_log}{reverse a log-normalized expression matrix}

\item{log_base}{the logarithm base (default = 2)}

\item{sign_gene}{features to use, typically differentially expressed ones}

\item{cell_type_vector}{vector with cell types (length = ncol(matrix))}

\item{cell_type}{deprecated, use \code{cell_type_vector}}
}
\value{
matrix of mean expression, features by cell type
}
\description{
Convert expression held in a Giotto object into the
mean-expression reference \code{\link{runDWLSDeconv}} expects: signature
features by cell type, each entry the mean expression of that feature in
that type. A vector for \code{cell_type_vector} can be taken from the cell
metadata (\code{\link{pDataDT}}).
}
\examples{
g <- GiottoData::loadGiottoMini("visium")
sign_gene <- c(
    "Bcl11b", "Lmo1", "F3", "Cnih3", "Ppp1r3c", "Rims2", "Gfap",
    "Gjc3", "Chrna4", "Prkcd", "Prr18", "Grb14", "Tprn", "Clic1", "Olig2",
    "Hrh3", "Tmbim1", "Carhsp1", "Tmem88b", "Ugt8a", "Arpp19", "Lamp5",
    "Galnt6", "Hlf", "Hs3st2", "Tbr1", "Myl4", "Cygb", "Ttc9b", "Ipcef1"
)

makeSignMatrixDWLS(
    gobject = g, sign_gene = sign_gene,
    cell_type_vector = pDataDT(g)[["leiden_clus"]]
)
}
\seealso{
\code{\link{runDWLSDeconv}}

Other spatial deconvolution:
\code{\link[=makeSignMatrixDWLSfromMatrix]{makeSignMatrixDWLSfromMatrix()}},
\code{\link[=runDWLSDeconv]{runDWLSDeconv()}},
\code{\link[=runSpatialDeconv]{runSpatialDeconv()}}
}
\concept{spatial deconvolution}
