% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\docType{class}
\name{markersParam-class}
\alias{markersParam-class}
\title{Marker detection parameter}
\description{
Virtual parent for the marker detection parameters dispatched on by
\code{\link[=analyzeData]{analyzeData()}}. One subclass per detection method, mirroring the \code{method}
argument of \code{\link[=findMarkers]{findMarkers()}}.
}
\seealso{
\code{\link[=analyzeData]{analyzeData()}}, \code{\link[=scranMarkersParam]{scranMarkersParam()}}, \code{\link[=findMarkers]{findMarkers()}}
}
