% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{mat_queryDT}
\alias{mat_queryDT}
\title{Query Matrix with data.table syntax}
\usage{
mat_queryDT(i = NULL, j = NULL, by = cell_ID, mtx)
}
\arguments{
\item{i}{select by row}

\item{j}{select by column or expression to query}

\item{by}{return results by which column}

\item{mtx}{matrix to use}
}
\description{
Finds the rows of the Matrix needed to perform the query, converts
first to matrix, transposes, then to data.table after which the query is performed
just as if it had been a cell_ID (rows) by gene (cols) data.table. Only works
when j param is provided.
}
\keyword{internal}
