% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{mean_giotto}
\alias{mean_giotto}
\title{mean_giotto}
\usage{
mean_giotto(x, ...)
}
\arguments{
\item{x}{vector}

\item{\dots}{additional parameters}
}
\value{
numeric
}
\description{
mean function that works with multiple matrix representations
}
