% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{merge_spatial_locs_feat_info}
\alias{merge_spatial_locs_feat_info}
\title{merge_spatial_locs_feat_info}
\usage{
merge_spatial_locs_feat_info(spatial_info, feature_info)
}
\description{
merge spatial cell and feature location information
}
\keyword{internal}
