% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{metaData-class}
\alias{metaData-class}
\title{Basic class for classes with metadata information}
\description{
Classes that inherit from this class will contain a metadata slot that stores
information in a data.table and should work similarly to data.table when interacting
with some basic generic operators for data retrieval and setting
}
