% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mini_giotto_3D}
\alias{mini_giotto_3D}
\title{mini Giotto object for spatial single-cell 3D data}
\format{
An object of class \code{"giotto"}; see \code{\link[Giotto]{createGiottoObject}}.
}
\usage{
data(mini_giotto_3D)
}
\description{
Mini Giotto object created from the STARmap data.
}
\examples{
data(mini_giotto_3D)

\dontrun{spatPlot3D(mini_giotto_3D, cell_color = 'cell_types', point_size = 5)}
}
\references{
Wang et al. (2018) Science
(\href{https://pubmed.ncbi.nlm.nih.gov/29930089/}{PubMed})
}
\keyword{datasets}
