% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mini_giotto_multi_cell}
\alias{mini_giotto_multi_cell}
\title{mini Giotto object for spatial multi-cell resolution data}
\format{
An object of class \code{"giotto"}; see \code{\link[Giotto]{createGiottoObject}}.
}
\usage{
data(mini_giotto_multi_cell)
}
\description{
Mini Giotto object created from the Brain Visium 10X data.
}
\examples{
data(mini_giotto_multi_cell)

\dontrun{spatPlot(mini_giotto_multi_cell, cell_color = 'cell_types', point_size = 5)}
}
\references{
10 Genomics Visium technology
(\href{https://www.10xgenomics.com/spatial-transcriptomics/}{10xgenomics})
}
\keyword{datasets}
