% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mini_giotto_single_cell}
\alias{mini_giotto_single_cell}
\title{mini Giotto object for spatial single-cell resolution data}
\format{
An object of class \code{"giotto"}; see \code{\link[Giotto]{createGiottoObject}}.
}
\usage{
data(mini_giotto_single_cell)
}
\description{
Mini Giotto object created from the seqFISH+ data.
}
\examples{
data(mini_giotto_single_cell)

\dontrun{spatPlot2D(mini_giotto_single_cell,cell_color = 'cell_types', point_size = 5)}
}
\references{
Eng et al. (2019) Nature
(\href{https://www.nature.com/articles/s41586-019-1049-y}{PubMed})
}
\keyword{datasets}
