% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{miscData-class}
\alias{miscData-class}
\title{Basic class for additional miscellaneous information}
\description{
Classes (such as dimObj) that can hold information from multiple types of methods
use the misc slot to hold additional information specific to each method.
Information may be stored within as S3 structures.
}
