% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_help.R, R/utilities.R
\name{my_arowMeans}
\alias{my_arowMeans}
\title{my_arowMeans}
\usage{
my_arowMeans(x)

my_arowMeans(x)
}
\arguments{
\item{x}{data to use}
}
\description{
arithmic rowMeans that works for a single column
}
\keyword{internal}
