% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_help.R, R/utilities.R
\name{my_growMeans}
\alias{my_growMeans}
\title{my_growMeans}
\usage{
my_growMeans(x, offset = 0.1)

my_growMeans(x, offset = 0.1)
}
\arguments{
\item{x}{data to use}

\item{offset}{offset}
}
\description{
geometric rowMeans that works for a single column
}
\keyword{internal}
