% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NN_network.R
\name{nnDT_to_kNN}
\alias{nnDT_to_kNN}
\title{nnDT_to_kNN}
\usage{
nnDT_to_kNN(nnDT)
}
\arguments{
\item{nnDT}{nearest neighbor network in data.table format}
}
\value{
kNN object
}
\description{
Convert a nearest network data.table to a kNN object
}
\keyword{internal}
