% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{norm_arcsinh}
\alias{norm_arcsinh}
\title{Arcsinh Normalization}
\value{
normalized object
}
\description{
A normalization commonly used with intensity-based data (CODEX, CyCIF, IMC).
It effectively handles a wide dynamic range and zero/near-zero values while
preserving the relative differences between signals of different intensities.

\deqn{\LARGE
x'_{i,j} = \operatorname{arcsinh}\left({\frac{x_{i,j}}{c}}\right)
}

Where:
\itemize{
\item (\eqn{x_{i,j}}) is the raw intensity for feature \eqn{i} in sample \eqn{j}
\item (\eqn{x'_{i,j}}) is the normalized intensity for feature \eqn{i} in
sample \eqn{j}
\item (\eqn{c}) is a cofactor that determines the degree of transformation
}
}
\section{Note}{
The cofactor \eqn{c} prevents over-amplification of small values and allows
better differentiation of signals at different intensities.

Common values to use are:
\itemize{
\item \code{c = 5} for fluorescence imaging-based proteomics (CODEX, CyCIF)
\item \code{c = 1} or \code{5} for mass-cytometry-based imaging (IMC).
}
}

\section{params}{


\tabular{ll}{
\code{c} \tab numeric (default = 5). Expressed as \eqn{c} in the above
formula.
}
}

\seealso{
\code{\link[=process_param]{process_param()}}

Other normalization parameters: 
\code{\link{norm_default}},
\code{\link{norm_l2}},
\code{\link{norm_library}},
\code{\link{norm_log}},
\code{\link{norm_osmfish}},
\code{\link{norm_pearson}},
\code{\link{norm_quantile}},
\code{\link{norm_tfidf}}
}
\concept{normalization parameters}
