% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{norm_l2}
\alias{norm_l2}
\title{L2 Normalization}
\value{
normalized object
}
\description{
L2 normalization (also known as Euclidean normalization) scales each column
(sample) in the expression matrix to have unit Euclidean length. This
process makes samples with different sequencing depths more comparable and
improves the performance of distance-based analyses.

\deqn{\LARGE
x'_{i,j} = \frac{x_{i,j}}{\sqrt{\sum_{i} x_{i,j}^2}}
}

Where:
\itemize{
\item (\eqn{x_{i,j}}) is the expression value for feature \eqn{i} in sample \eqn{j}
\item (\eqn{x'_{i,j}}) is the L2-normalized expression value
}
}
\section{Note}{
L2 normalization can be applied to raw data, but is most commonly used after
other normalization methods such as TF-IDF or log normalization to standardize
sample-to-sample comparisons.
}

\section{params}{

None
}

\seealso{
\link{process_param}

Other normalization parameters: 
\code{\link{norm_arcsinh}},
\code{\link{norm_default}},
\code{\link{norm_library}},
\code{\link{norm_log}},
\code{\link{norm_osmfish}},
\code{\link{norm_pearson}},
\code{\link{norm_quantile}},
\code{\link{norm_tfidf}}
}
\concept{normalization parameters}
