% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{norm_library}
\alias{norm_library}
\title{Library Size Normalization}
\description{
Normalize expression matrix for total library size and then scale by
a custom scalefactor.

This method does not work well when any cells/samples
have a library size of 0, so filtering prior to this is recommended.

\deqn{\LARGE
x'_{i,j} = \frac{x_{i,j}}{\sum_{i} x_{i,j}} \times k
}
Where:
\itemize{
\item (\eqn{x_{i,j}}) is the raw count for feature \eqn{i} in sample \eqn{j}
\item (\eqn{x'_{i,j}}) is the library normalized and scaled expression value for
feature \eqn{i} in sample \eqn{j}
\item (k) is a scalefactor applied after normalization
}
}
\section{params}{


\tabular{ll}{
\code{scalefactor} \tab numeric (default = 6000). Scalefactor to use after
library size normalization. Expressed as \emph{\strong{k}} in the above equation
}
}

\seealso{
\link{process_param}

Other normalization parameters: 
\code{\link{norm_default}},
\code{\link{norm_l2}},
\code{\link{norm_log}},
\code{\link{norm_osmfish}},
\code{\link{norm_pearson}},
\code{\link{norm_quantile}},
\code{\link{norm_tfidf}}
}
\concept{normalization parameters}
