% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{norm_pearson}
\alias{norm_pearson}
\title{Lause/Kobak Pearson Residuals Normalization}
\description{
Calculate Pearson residuals with a dispersion adjustment, to identify cells
that deviate significantly from what would be expected under independence.
The normalization divides by the standard deviation of the difference, which
is adjusted by the dispersion parameter θ.

This normalization is designed for detection of highly variable features and
dimension reduction and clustering.

\deqn{\LARGE
z_{i,j} = \frac{x_{i,j} - \mu_{i,j}}{\sqrt{\mu_{i,j} + \mu_{i,j}^2 / \theta}}
}

\deqn{\LARGE
\mu_{i,j} = \frac{r_i \cdot c_j}{N}
}

Where:
\itemize{
\item (\eqn{x_{i,j}}) is the raw count for feature \eqn{i} in sample \eqn{j}
\item (\eqn{\mu_{i,j}}) is the expected value under the model
\item (\eqn{r_i}) is \eqn{\sum_j x_{i,j}}
\item (\eqn{c_j}) is \eqn{\sum_i x_{i,j}}
\item (\eqn{N}) is \eqn{\sum_{i,j} x_{i,j}}
\item (\eqn{\theta}) is a dispersion parameter
\item (\eqn{z_{i,j}}) is the Pearson residual clipped to the range
\eqn{[-\sqrt{n}, \sqrt{n}]} where \eqn{n} is the number of columns. This is
done to prevent extreme values from dominating the analysis.
}
}
\section{Note}{
Scaling is not recommended after this normalization since it is already
transforming the data to z-score-like values with a dispersion adjustment.
It is also not recommended to use this with DGE analysis.
}

\section{params}{


\tabular{ll}{
\code{theta} \tab dispersion parameter expressed as \eqn{\theta} in the above
formula
}
}

\references{
Lause, J., Berens, P. & Kobak, D. Analytic Pearson residuals for
normalization of single-cell RNA-seq UMI data. Genome Biol 22, 258 (2021).
https://doi.org/10.1186/s13059-021-02451-7
}
\seealso{
\link{process_param}

Other normalization parameters: 
\code{\link{norm_default}},
\code{\link{norm_l2}},
\code{\link{norm_library}},
\code{\link{norm_log}},
\code{\link{norm_osmfish}},
\code{\link{norm_quantile}},
\code{\link{norm_tfidf}}
}
\concept{normalization parameters}
