% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{normalizeGiotto}
\alias{normalizeGiotto}
\title{normalizeGiotto}
\usage{
normalizeGiotto(gobject, library_size_norm = T, scalefactor = 2000,
  logbase = 2, scale_genes = T, scale_cells = T)
}
\arguments{
\item{gobject}{giotto object}

\item{library_size_norm}{normalize cells by library size}

\item{scalefactor}{scale factor to use after library size normalization}

\item{logbase}{log base to use to log normalize expression values}

\item{scale_genes}{z-score genes over all cells}

\item{scale_cells}{z-score cells over all genes}
}
\value{
giotto object
}
\description{
normalize and/or scale expresion values of Giotto object
}
\details{
Description of normalization steps ...
}
\examples{
    normalizeGiotto(gobject)
}
