% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{normalizeGiotto2fast}
\alias{normalizeGiotto2fast}
\title{normalizeGiotto2}
\usage{
normalizeGiotto2fast(
  gobject,
  norm_methods = c("standard", "osmFISH"),
  library_size_norm = TRUE,
  scalefactor = 6000,
  log_norm = TRUE,
  logbase = 2,
  scale_genes = T,
  scale_cells = T,
  scale_order = c("first_genes", "first_cells"),
  verbose = F
)
}
\arguments{
\item{gobject}{giotto object}

\item{library_size_norm}{normalize cells by library size}

\item{scalefactor}{scale factor to use after library size normalization}

\item{logbase}{log base to use to log normalize expression values}

\item{scale_genes}{z-score genes over all cells}

\item{scale_cells}{z-score cells over all genes}

\item{scale_order}{order to scale genes and cells}

\item{verbose}{be verbose}
}
\value{
giotto object
}
\description{
fast normalize and/or scale expresion values of Giotto object
}
\details{
Description of normalization steps ...
}
\examples{
    normalizeGiotto2(gobject)
}
