% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{nrow-generic}
\alias{nrow-generic}
\alias{nrow}
\alias{nrow,giottoPoints-method}
\alias{nrow,giottoPolygon-method}
\title{Dimensions of giotto points and polygon objects}
\usage{
\S4method{nrow}{giottoPoints}(x)

\S4method{nrow}{giottoPolygon}(x)
}
\arguments{
\item{x}{giottoPoints or giottoPolygon object}
}
\description{
Get the number of rows (nrow)
}
\section{Functions}{
\itemize{
\item \code{nrow(giottoPoints)}: Find rows of giottoPoints object

\item \code{nrow(giottoPolygon)}: Find rows of giottoPolygon object

}}
