% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_hmrf.R
\name{numPts_below_line}
\alias{numPts_below_line}
\title{numPts_below_line}
\usage{
numPts_below_line(myVector, slope, x)
}
\arguments{
\item{myVector}{input sequence of sorted positive values from smallest to
greatest}

\item{slope}{slope to compare}

\item{x}{location point of the line to compare, integer from 1 to length of
myVector}
}
\value{
numeric
}
\description{
function to calculate the number of data points below a given
line
}
\details{
This function calculates the number of data points in a sorted sequence
below a line with given slope through a certain point on this sequence.
}
\keyword{internal}
