% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{objName-generic}
\alias{objName-generic}
\alias{objName}
\alias{objName<-}
\alias{objName,nameData-method}
\alias{objName,giottoPoints-method}
\alias{objName,giottoLargeImage-method}
\alias{objName,giottoImage-method}
\alias{objName<-,nameData-method}
\alias{objName<-,giottoImage-method}
\alias{objName<-,giottoLargeImage-method}
\alias{objName<-,giottoPoints-method}
\title{Giotto object name information}
\usage{
objName(x)

\S4method{objName}{nameData}(x)

\S4method{objName}{giottoPoints}(x)

\S4method{objName}{giottoLargeImage}(x)

\S4method{objName}{giottoImage}(x)

\S4method{objName}{nameData}(x) <- value

\S4method{objName}{giottoImage}(x) <- value

\S4method{objName}{giottoLargeImage}(x) <- value

\S4method{objName}{giottoPoints}(x) <- value
}
\arguments{
\item{x}{a Giotto S4 class subobject with name data}

\item{value}{value to set as object name}
}
\description{
access and set name slot fo S4 subobject
}
\section{Functions}{
\itemize{
\item \code{objName(nameData)}: Get name information

\item \code{objName(giottoPoints)}: Get name giottoPoints

\item \code{objName(giottoPoints) <- value}: Set name giottoPoints

}}
