% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_enrichment.R
\name{optimize_deconvolute_dwls}
\alias{optimize_deconvolute_dwls}
\title{optimize_deconvolute_dwls}
\usage{
optimize_deconvolute_dwls(exp, Signature)
}
\description{
Rui to fill in
}
\keyword{internal}
