% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_structures.R
\name{overlap_points_per_polygon}
\alias{overlap_points_per_polygon}
\title{Overlap points per polgyon}
\usage{
overlap_points_per_polygon(spatvec, pointvec, poly_ID_names, verbose = TRUE)
}
\description{
Loop to overlap each single polygon
}
\seealso{
\code{\link{overlap_points_single_polygon}}
}
\keyword{internal}
