% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_structures.R
\name{overlap_points_per_polygon_wrapped}
\alias{overlap_points_per_polygon_wrapped}
\title{Overlap points per polygon -- wrapped}
\usage{
overlap_points_per_polygon_wrapped(
  spatvec_wrapped,
  pointvec_wrapped,
  poly_ID_names
)
}
\description{
overlap wrapped polygons
}
\keyword{internal}
