% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_structures.R
\name{overlap_points_single_polygon}
\alias{overlap_points_single_polygon}
\title{Overlap points -- single polygon}
\usage{
overlap_points_single_polygon(spatvec, poly_ID_name, pointvec_dt)
}
\description{
overlap for a single polygon
}
\keyword{internal}
