% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_structures.R
\name{overlapImagesToMatrix}
\alias{overlapImagesToMatrix}
\title{overlapImagesToMatrix}
\usage{
overlapImagesToMatrix(
  gobject,
  name = "raw",
  poly_info = "cell",
  feat_info = "protein",
  name_overlap = "images",
  image_names = NULL,
  spat_locs_name = "raw",
  return_gobject = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{name}{name for the overlap count matrix}

\item{poly_info}{polygon information}

\item{feat_info}{feature information}

\item{image_names}{names of images you used}

\item{spat_locs_name}{name for spatial centroids / locations associated with matrix}

\item{return_gobject}{return giotto object (default: TRUE)}
}
\value{
giotto object or data.table with aggregated information
}
\description{
create a count matrix based on overlap results from \code{\link{calculateOverlapPolygonImages}}
}
\concept{overlap}
