% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{overlaps-generic}
\alias{overlaps-generic}
\alias{overlaps}
\alias{overlaps,giottoPolygon-method}
\title{overlaps-generic}
\usage{
overlaps(x, ...)

\S4method{overlaps}{giottoPolygon}(x, name = NULL)
}
\arguments{
\item{x}{object}

\item{name}{(optional) name of overlaps information to retrieve}
}
\description{
Access list of overlaps information from object
}
\section{Functions}{
\itemize{
\item \code{overlaps(giottoPolygon)}: Get overlaps information from giottoPolygon

}}
