% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_structures.R
\name{overlapToMatrix}
\alias{overlapToMatrix}
\title{overlapToMatrix}
\usage{
overlapToMatrix(
  gobject,
  name = "raw",
  poly_info = "cell",
  feat_info = "rna",
  return_gobject = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{name}{name for the overlap count matrix}

\item{poly_info}{polygon information}

\item{feat_info}{feature information}

\item{return_gobject}{return giotto object (default: TRUE)}
}
\value{
giotto object or count matrix
}
\description{
create a count matrix based on overlap results from \code{\link{calculateOverlapRaster}}, \code{\link{calculateOverlapSerial}}, or \code{\link{calculateOverlapParallel}}
}
\keyword{overlap}
