% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_structures.R
\name{overlapToMatrixMultiPoly}
\alias{overlapToMatrixMultiPoly}
\title{overlapToMatrixMultiPoly}
\usage{
overlapToMatrixMultiPoly(
  gobject,
  name = "raw",
  poly_info = "cell",
  feat_info = "rna",
  new_poly_info = "multi",
  return_gobject = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{name}{name for the overlap count matrix}

\item{poly_info}{vector with polygon information}

\item{feat_info}{feature information}

\item{new_poly_info}{name for new aggregated polygon information}

\item{return_gobject}{return giotto object (default: TRUE)}
}
\value{
giotto object or count matrix
}
\description{
create a count matrix based on overlap results from \code{\link{calculateOverlapRaster}}, \code{\link{calculateOverlapSerial}}, or \code{\link{calculateOverlapParallel}}
and aggregate information from multiple polygon layers (e.g. z-stacks) together
}
\keyword{overlap}
