% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_enrichment.R
\name{pagePermutation}
\alias{pagePermutation}
\title{pagePermutation}
\usage{
pagePermutation(sc_gene, gene_number, n)
}
\description{
creates permutation for the PAGEEnrich test
}
\examples{
    pagePermutation()
}
