% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registration.R
\name{parse_affine}
\alias{parse_affine}
\title{Read affine matrix for linear transforms}
\usage{
parse_affine(x)
}
\arguments{
\item{x}{object coercible to matrix with a 2x3 or 3x3 affine matrix}
}
\value{
a list of transforms information.
}
\description{
Affine transforms are linear transformations that cover scaling,
rotation, shearing, and translations. They can be represented as matrices of
2x3 or 3x3 values. This function reads the matrix and extracts the values
needed to perform them.
}
\keyword{internal}
