% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.R
\name{pca_giotto}
\alias{pca_giotto}
\title{pca_giotto}
\usage{
pca_giotto(mymatrix, center = T, scale = T, k = 50)
}
\arguments{
\item{mymatrix}{matrix or object that can be converted to matrix}

\item{center}{center data}

\item{scale}{scale features}

\item{k}{number of principal components to calculate}
}
\value{
list of eigenvalues, eigenvectors and pca coordinates
}
\description{
performs PCA based on Rfast
}
\keyword{internal}
