% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{pDataDT}
\alias{pDataDT}
\title{pDataDT}
\usage{
pDataDT(gobject, spat_unit = NULL, feat_type = NULL)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit}

\item{feat_type}{feature type}
}
\value{
data.table with cell metadata
}
\description{
show cell metadata
}
