% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visuals.R
\name{plot_auto_convert_largeImage_MG}
\alias{plot_auto_convert_largeImage_MG}
\title{plot_auto_convert_largeImage_MG}
\usage{
plot_auto_convert_largeImage_MG(
  gobject,
  giottoLargeImage = NULL,
  largeImage_name = NULL,
  feat_type = NULL,
  spat_unit = NULL,
  spat_loc_name = NULL,
  include_image_in_border = TRUE
)
}
\arguments{
\item{gobject}{gobject containing largeImage}

\item{giottoLargeImage}{giottoLargeImage}

\item{largeImage_name}{name of largeImage}

\item{feat_type}{feature type}

\item{spat_unit}{spatial unit}

\item{spat_loc_name}{name of spatial locations to plot}

\item{include_image_in_border}{expand the extent sampled to also show image in border regions not included in spatlocs}
}
\value{
a giottoImage (MG)
}
\description{
automatically convert giottoLargeImage to giottoImage for giotto plotting functions
}
\keyword{internal}
