% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visuals.R
\name{plot_auto_largeImage_resample}
\alias{plot_auto_largeImage_resample}
\title{plot_auto_largeImage_resample}
\usage{
plot_auto_largeImage_resample(
  gobject,
  giottoLargeImage = NULL,
  largeImage_name = NULL,
  spat_unit = NULL,
  spat_loc_name = NULL,
  include_image_in_border = TRUE
)
}
\arguments{
\item{gobject}{gobject containing largeImage}

\item{giottoLargeImage}{giottoLargeImage}

\item{largeImage_name}{name of largeImage}

\item{spat_unit}{spatial unit}

\item{spat_loc_name}{name of spatial locations to plot}

\item{include_image_in_border}{expand the extent sampled to also show image in border regions not included in spatlocs}
}
\value{
a giottoLargeImage cropped and resampled properly for plotting
}
\description{
resamples largeImage for plotting
}
\keyword{internal}
