% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd.R
\name{plot_cow_params}
\alias{plot_cow_params}
\title{Params documentation template: plot_cow_params}
\arguments{
\item{cow_n_col}{cowplot param: how many columns}

\item{cow_rel_h}{cowplot param: relative heights of rows (e.g. c(1,2))}

\item{cow_rel_w}{cowplot param: relative widths of columns (e.g. c(1,2))}

\item{cow_align}{cowplot param: how to align}
}
\description{
Params documentation template: plot_cow_params
}
\keyword{internal}
