% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd.R
\name{plot_dimred_params}
\alias{plot_dimred_params}
\title{Params documentation template: plot_dimred_params}
\arguments{
\item{dim_reduction_to_use}{character. dimension reduction to use}

\item{dim_reduction_name}{character. dimension reduction name}

\item{dim1_to_use}{numeric. dimension to use on x-axis}

\item{dim2_to_use}{numeric. dimension to use on y-axis}

\item{dim3_to_use}{numeric. dimension to use on z-axis}

\item{dim_point_shape}{point with border or not (border or no_border)}

\item{dim_point_size}{size of points in dim. reduction space}

\item{dim_point_alpha}{transparancy of point in dim. reduction space}

\item{dim_point_border_col}{border color of points in dim. reduction space}

\item{dim_point_border_stroke}{border stroke of points in dim. reduction space}
}
\description{
Params documentation template: plot_dimred_params
}
\keyword{internal}
