% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd.R
\name{plot_feat_params}
\alias{plot_feat_params}
\title{Params documentation template: plot_feat_params}
\arguments{
\item{feats_color_code}{code to color the provided features}

\item{feat_shape_code}{code to shape the provided feature types}
}
\description{
Params documentation template: plot_feat_params
}
\keyword{internal}
