% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_in_situ_visuals.R
\name{plot_feature_points_layer}
\alias{plot_feature_points_layer}
\title{plot_feature_points_layer}
\usage{
plot_feature_points_layer(
  ggobject,
  spatial_feat_info,
  feats,
  feats_color_code = NULL,
  feat_shape_code = NULL,
  sdimx = "x",
  sdimy = "y",
  color = "feat_ID",
  shape = "feat",
  point_size = 1.5,
  show_legend = TRUE,
  plot_method = c("ggplot", "scattermore", "scattermost")
)
}
\value{
ggplot
}
\description{
low level function to plot a points at the spatial in situ level
}
\details{
This function can plot multiple features over multiple modalities. These plots can get very big very fast.
}
\keyword{internal}
