% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{plot_giottoImage_MG}
\alias{plot_giottoImage_MG}
\title{plot_giottoImage_MG}
\usage{
plot_giottoImage_MG(gobject = NULL, image_name = NULL, giottoImage = NULL)
}
\arguments{
\item{gobject}{giotto object}

\item{image_name}{name of giotto image \code{\link{showGiottoImageNames}}}

\item{giottoImage}{giottoImage object}
}
\value{
plot
}
\description{
get and plot a giottoImage either directly or from a giotto object
}
\keyword{internal}
