% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{plot_giottoLargeImage}
\alias{plot_giottoLargeImage}
\title{plot_giottoLargeImage}
\usage{
plot_giottoLargeImage(
  gobject = NULL,
  largeImage_name = NULL,
  giottoLargeImage = NULL,
  crop_extent = NULL,
  xmax_crop = NULL,
  xmin_crop = NULL,
  ymax_crop = NULL,
  ymin_crop = NULL,
  max_intensity = NULL,
  asRGB = FALSE,
  stretch = NULL,
  maxcell = 5e+05,
  col = grDevices::grey.colors(n = 256, start = 0, end = 1, gamma = 1),
  ...
)
}
\arguments{
\item{gobject}{giotto object}

\item{largeImage_name}{name of giottoLargeImage}

\item{giottoLargeImage}{giottoLargeImage object}

\item{crop_extent}{(optional) extent object to focus on specific region of image}

\item{xmax_crop, xmin_crop, ymax_crop, ymin_crop}{(optional) crop min/max x and y bounds}

\item{max_intensity}{(optional) value to treat as maximum intensity in color scale}

\item{asRGB}{(optional) [boolean] force RGB plotting if not automatically detected}

\item{stretch}{character. Option to stretch the values to increase contrast: "lin"
linear or "hist" (histogram)}

\item{maxcell}{positive integer. Maximum number of image cells to use for the plot}

\item{col}{character. Colors for single channel images. The default is
grDevices::grey.colors(n = 256, start = 0, end = 1, gamma = 1). It can also be a
data.frame with two columns (value, color) to get a "classes" type legend or with
three columns (from, to, color) to get an "interval" type legend}

\item{...}{additional params to pass to terra::plot or terra::plotRGB depending
depending on image type}
}
\value{
plot
}
\description{
Plot a \emph{downsampled} version of giottoLargeImage. Cropping can increase plot resolution of region of interest.
}
\keyword{internal}
