% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd.R
\name{plot_image_params}
\alias{plot_image_params}
\title{Params documentation template: plot_image_params}
\arguments{
\item{show_image}{show a tissue background image}

\item{gimage}{a giotto image}

\item{image_name}{name of a giotto image or multiple images with group_by}

\item{largeImage_name}{name of a giottoLargeImage or multiple images with 
group_by}
}
\value{
plot
}
\description{
Params documentation template: plot_image_params
}
\keyword{internal}
