% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visuals.R
\name{plot_network_layer_ggplot}
\alias{plot_network_layer_ggplot}
\title{plot_network_layer_ggplot}
\usage{
plot_network_layer_ggplot(
  ggobject,
  annotated_network_DT,
  edge_alpha = NULL,
  show_legend = T
)
}
\arguments{
\item{annotated_network_DT}{annotated network data.table of selected cells}

\item{edge_alpha}{alpha of network edges}

\item{show_legend}{show legend}

\item{gobject}{giotto object}
}
\value{
ggplot
}
\description{
Visualize cells in network layer according to dimension reduction coordinates
}
\details{
Description of parameters.
}
\keyword{internal}
