% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd.R
\name{plot_output_params}
\alias{plot_output_params}
\title{Params documentation template: plot_output_params}
\arguments{
\item{show_plot}{logical. show plot}

\item{return_plot}{logical. return ggplot object}

\item{save_plot}{logical. save the plot}

\item{save_param}{list of saving parameters, see 
\code{\link{showSaveParameters}}}

\item{default_save_name}{default save name for saving, don't change, 
change save_name in save_param}
}
\value{
plot
}
\description{
Params documentation template: plot_output_params
}
\keyword{internal}
