% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visuals.R
\name{plot_point_layer_ggplot}
\alias{plot_point_layer_ggplot}
\title{plot_point_layer_ggplot}
\usage{
plot_point_layer_ggplot(ggobject, annotated_DT_selected,
  annotated_DT_other, cell_color = NULL, color_as_factor = T,
  cell_color_code = NULL, select_cell_groups = NULL,
  select_cells = NULL, show_other_cells = T,
  other_cell_color = "lightgrey", other_point_size = 0.5,
  show_cluster_center = F, show_center_label = T,
  center_point_size = 4, center_point_border_col = "black",
  center_point_border_stroke = 0.1, label_size = 4,
  label_fontface = "bold", edge_alpha = NULL, point_size = 1,
  point_border_col = "black", point_border_stroke = 0.1,
  show_legend = T)
}
\arguments{
\item{annotated_DT_selected}{annotated data.table of selected cells}

\item{annotated_DT_other}{annotated data.table of not selected cells}

\item{cell_color}{color for cells (see details)}

\item{color_as_factor}{convert color column to factor}

\item{cell_color_code}{named vector with colors}

\item{select_cell_groups}{select subset of cells/clusters based on cell_color parameter}

\item{select_cells}{select subset of cells based on cell IDs}

\item{show_other_cells}{display not selected cells}

\item{other_cell_color}{color of not selected cells}

\item{other_point_size}{size of not selected cells}

\item{show_cluster_center}{plot center of selected clusters}

\item{show_center_label}{plot label of selected clusters}

\item{center_point_size}{size of center points}

\item{label_size}{size of labels}

\item{label_fontface}{font of labels}

\item{edge_alpha}{column to use for alpha of the edges}

\item{point_size}{size of point (cell)}

\item{point_border_col}{color of border around points}

\item{point_border_stroke}{stroke size of border around points}

\item{show_legend}{show legend}

\item{gobject}{giotto object}
}
\value{
ggplot
}
\description{
Visualize cells in point layer according to dimension reduction coordinates
}
\details{
Description of parameters.
}
\examples{
    plot_point_layer_ggplot(gobject)
}
