% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visuals.R
\name{plot_spat_image_layer_ggplot}
\alias{plot_spat_image_layer_ggplot}
\title{plot_spat_image_layer_ggplot}
\usage{
plot_spat_image_layer_ggplot(
  gg_obj,
  gobject,
  gimage,
  feat_type = NULL,
  spat_unit = NULL,
  spat_loc_name = NULL,
  polygon_feat_type = NULL,
  sdimx = NULL,
  sdimy = NULL
)
}
\arguments{
\item{gg_obj}{ggplot2 object}

\item{gobject}{giotto object}

\item{gimage}{a giotto image or a list/vector of giotto images}

\item{feat_type}{feature type}

\item{spat_unit}{spatial unit}

\item{spat_loc_name}{name for spatial locations}

\item{polygon_feat_type}{name for feature type associated with polygon information}

\item{sdimx}{x-axis dimension name (default = 'sdimx')}

\item{sdimy}{y-axis dimension name (default = 'sdimy')}
}
\value{
ggplot
}
\description{
create background image in ggplot
}
\keyword{internal}
