% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visuals.R
\name{plot_spat_point_layer_ggplot}
\alias{plot_spat_point_layer_ggplot}
\title{plot_spat_point_layer_ggplot}
\usage{
plot_spat_point_layer_ggplot(ggobject, sdimx = NULL, sdimy = NULL,
  cell_locations_metadata_selected, cell_locations_metadata_other,
  cell_color = NULL, color_as_factor = T, select_cell_groups = NULL,
  select_cells = NULL, cell_color_code = NULL, show_legend = TRUE,
  point_size = 2, point_border_stroke = 0.1,
  point_border_col = "lightgrey", show_other_cells = T,
  other_cell_color = "lightgrey", other_point_size = 1)
}
\arguments{
\item{sdimx}{x-axis dimension name (default = 'sdimx')}

\item{sdimy}{y-axis dimension name (default = 'sdimy')}

\item{cell_locations_metadata_selected}{annotated location from selected cells}

\item{cell_locations_metadata_other}{annotated location from non-selected cells}

\item{cell_color}{color for cells (see details)}

\item{color_as_factor}{convert color column to factor}

\item{select_cell_groups}{select subset of cells/clusters based on cell_color parameter}

\item{select_cells}{select subset of cells based on cell IDs}

\item{cell_color_code}{named vector with colors}

\item{show_legend}{show legend}

\item{point_size}{size of point (cell)}

\item{point_border_stroke}{stroke size of border around points}

\item{point_border_col}{color of border around points}

\item{show_other_cells}{display not selected cells}

\item{other_cell_color}{color for not selected cells}

\item{other_point_size}{point size for not selected cells}

\item{gobject}{giotto object}
}
\value{
ggplot
}
\description{
creat ggplot point layer for spatial coordinates
}
\details{
Description of parameters.
}
\examples{
    plot_spat_point_layer_ggplot(gobject)
}
