% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visuals.R
\name{plot_spat_scatterpie_layer_ggplot}
\alias{plot_spat_scatterpie_layer_ggplot}
\title{plot_spat_scatterpie_layer_ggplot}
\usage{
plot_spat_scatterpie_layer_ggplot(
  ggobject,
  sdimx = "sdimx",
  sdimy = "sdimy",
  spatial_locations = NULL,
  spatial_enrichment = NULL,
  radius = 10,
  color = NA,
  cell_color_code = NULL
)
}
\arguments{
\item{sdimx}{x-axis dimension name (default = 'sdimx')}

\item{sdimy}{y-axis dimension name (default = 'sdimy')}

\item{spatial_locations}{spatial locations}

\item{spatial_enrichment}{spatial enrichment results}

\item{radius}{radius of scatterpie}

\item{color}{color of lines within pie chart}

\item{cell_color_code}{color code for the cell types}

\item{gobject}{giotto object}
}
\value{
ggplot
}
\description{
create scatterpie in ggplot
}
\keyword{internal}
