% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd.R
\name{plot_spatenr_params}
\alias{plot_spatenr_params}
\title{Params documentation template: plot_spatenr_params}
\arguments{
\item{spat_enr_names}{character. names of spatial enrichment results to include}
}
\description{
Params documentation template: plot_spatenr_params
}
\keyword{internal}
