% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd.R
\name{plot_spatnet_params}
\alias{plot_spatnet_params}
\title{Params documentation template: plot_spatnet_params}
\arguments{
\item{show_spatial_network}{show spatial network}

\item{spatial_network_name}{name of spatial network to use}

\item{spat_network_name}{name of spatial network to use}

\item{spat_network_color}{color of spatial network}

\item{spatial_network_color}{color of spatial network}

\item{spat_network_alpha}{alpha of spatial network}
}
\description{
Params documentation template: plot_spatnet_params
}
\keyword{internal}
