% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction_visuals.R
\name{plotCCcomDotplot}
\alias{plotCCcomDotplot}
\title{plotCCcomDotplot}
\usage{
plotCCcomDotplot(
  gobject,
  comScores,
  selected_LR = NULL,
  selected_cell_LR = NULL,
  show_LR_names = TRUE,
  show_cell_LR_names = TRUE,
  cluster_on = c("PI", "LR_expr", "log2fc"),
  cor_method = c("pearson", "kendall", "spearman"),
  aggl_method = c("ward.D", "ward.D2", "single", "complete", "average", "mcquitty",
    "median", "centroid"),
  dot_color_gradient = NULL,
  gradient_style = c("divergent", "sequential"),
  show_plot = NULL,
  return_plot = NULL,
  save_plot = NULL,
  save_param = list(),
  default_save_name = "plotCCcomDotplot"
)
}
\arguments{
\item{gobject}{giotto object}

\item{comScores}{communication scores from \code{\link{exprCellCellcom}} 
or \code{\link{spatCellCellcom}}}

\item{selected_LR}{selected ligand-receptor combinations}

\item{selected_cell_LR}{selected cell-cell combinations for ligand-receptor 
combinations}

\item{show_LR_names}{show ligand-receptor names}

\item{show_cell_LR_names}{show cell-cell names}

\item{cluster_on}{values to use for clustering of cell-cell and 
ligand-receptor pairs}

\item{cor_method}{correlation method used for clustering}

\item{aggl_method}{agglomeration method used by hclust}

\item{dot_color_gradient}{character. continuous colors to use. palette to
use or vector of colors to use (minimum of 2).}

\item{gradient_style}{either 'divergent' (midpoint is used in color scaling)
or 'sequential' (scaled based on data range)}

\item{show_plot}{logical. show plot}

\item{return_plot}{logical. return ggplot object}

\item{save_plot}{logical. save the plot}

\item{save_param}{list of saving parameters, see \code{\link{showSaveParameters}}}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}
}
\value{
ggplot
}
\description{
Plots dotplot for ligand-receptor communication scores in 
cell-cell interactions
}
