% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction_visuals.R
\name{plotCCcomHeatmap}
\alias{plotCCcomHeatmap}
\title{plotCCcomHeatmap}
\usage{
plotCCcomHeatmap(
  gobject,
  comScores,
  selected_LR = NULL,
  selected_cell_LR = NULL,
  show_LR_names = TRUE,
  show_cell_LR_names = TRUE,
  show = c("PI", "LR_expr", "log2fc"),
  cor_method = c("pearson", "kendall", "spearman"),
  aggl_method = c("ward.D", "ward.D2", "single", "complete", "average", "mcquitty",
    "median", "centroid"),
  gradient_color = NULL,
  gradient_style = c("divergent", "sequential"),
  show_plot = NULL,
  return_plot = NULL,
  save_plot = NULL,
  save_param = list(),
  default_save_name = "plotCCcomHeatmap"
)
}
\arguments{
\item{gobject}{giotto object}

\item{comScores}{communinication scores from \code{\link{exprCellCellcom}} 
or \code{\link{spatCellCellcom}}}

\item{selected_LR}{selected ligand-receptor combinations}

\item{selected_cell_LR}{selected cell-cell combinations for ligand-receptor 
combinations}

\item{show_LR_names}{show ligand-receptor names}

\item{show_cell_LR_names}{show cell-cell names}

\item{show}{values to show on heatmap}

\item{cor_method}{correlation method used for clustering}

\item{aggl_method}{agglomeration method used by hclust}

\item{gradient_color}{character. continuous colors to use. palette to
use or vector of colors to use (minimum of 2).}

\item{gradient_style}{either 'divergent' (midpoint is used in color scaling)
or 'sequential' (scaled based on data range)}

\item{show_plot}{logical. show plot}

\item{return_plot}{logical. return ggplot object}

\item{save_plot}{logical. save the plot}

\item{save_param}{list of saving parameters, see \code{\link{showSaveParameters}}}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}
}
\value{
ggplot
}
\description{
Plots heatmap for ligand-receptor communication scores in 
cell-cell interactions
}
\examples{
g <- GiottoData::loadGiottoMini("visium")

comScores <- exprCellCellcom(g, cluster_column = "leiden_clus", 
feat_set_1 = c("Gm19935", "2900040C04Rik", "Ccnd2"), 
feat_set_2 = c("9630013A20Rik", "Gna12", "Btbd17"))

plotCCcomHeatmap(gobject = g, comScores = comScores, show_plot = TRUE)
}
