% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction_visuals.R
\name{plotCellProximityGenes}
\alias{plotCellProximityGenes}
\title{plotCellProximityGenes}
\usage{
plotCellProximityGenes(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=plotCellProximityFeats]{plotCellProximityFeats}}
  \describe{
    \item{\code{gobject}}{giotto object}
    \item{\code{icfObject}}{ICF (interaction changed feature) score object}
    \item{\code{method}}{plotting method to use}
    \item{\code{min_cells}}{minimum number of source cell type}
    \item{\code{min_cells_expr}}{minimum expression level for source cell type}
    \item{\code{min_int_cells}}{minimum number of interacting neighbor cell type}
    \item{\code{min_int_cells_expr}}{minimum expression level for interacting neighbor cell type}
    \item{\code{min_fdr}}{minimum adjusted p-value}
    \item{\code{min_spat_diff}}{minimum absolute spatial expression difference}
    \item{\code{min_log2_fc}}{minimum log2 fold-change}
    \item{\code{min_zscore}}{minimum z-score change}
    \item{\code{zscores_column}}{calculate z-scores over cell types or featuress}
    \item{\code{direction}}{differential expression directions to keep}
    \item{\code{cell_color_code}}{vector of colors with cell types as names}
    \item{\code{show_plot}}{show plots}
    \item{\code{return_plot}}{return plotting object}
    \item{\code{save_plot}}{directly save the plot [boolean]}
    \item{\code{save_param}}{list of saving parameters from \code{\link{all_plots_save_function}}}
    \item{\code{default_save_name}}{default save name for saving, don't change, change save_name in save_param}
  }}
}
\description{
Create visualization for cell proximity feature scores
}
