% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction_visuals.R
\name{plotCellProximityGeneScores}
\alias{plotCellProximityGeneScores}
\title{plotCellProximityGeneScores}
\usage{
plotCellProximityGeneScores(CPGscores, selected_interactions = NULL,
  selected_genes = NULL, detail_plot = T, simple_plot = F,
  simple_plot_facet = c("interaction", "genes"),
  facet.scales = "fixed", facet.ncol = length(selected_genes),
  facet.nrow = length(selected_interactions), show.plot = F)
}
\arguments{
\item{CPGscores}{CPGscores, output from getCellProximityGeneScores()}

\item{selected_interactions}{interactions to show}

\item{selected_genes}{genes to show}

\item{detail_plot}{show detailed info in both interacting cell types}

\item{simple_plot}{show a simplified plot}

\item{simple_plot_facet}{facet on interactions or genes with simple plot}

\item{facet.scales}{ggplot facet scales paramter}

\item{facet.ncol}{ggplot facet ncol parameter}

\item{facet.nrow}{ggplot facet nrow parameter}

\item{show.plot}{show plot}
}
\value{
ggplot barplot
}
\description{
Create heatmap from cell-cell proximity scores
}
\details{
Give more details ...
}
\examples{
    plotCellProximityGeneScores(CPGscores)
}
