% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction_visuals.R
\name{plotCombineCCcom}
\alias{plotCombineCCcom}
\title{plotCombineCCcom}
\usage{
plotCombineCCcom(
  gobject,
  combCCcom,
  selected_LR = NULL,
  selected_cell_LR = NULL,
  detail_plot = T,
  simple_plot = F,
  simple_plot_facet = c("interaction", "genes"),
  facet_scales = "fixed",
  facet_ncol = length(selected_LR),
  facet_nrow = length(selected_cell_LR),
  colors = c("#9932CC", "#FF8C00"),
  show_plot = NA,
  return_plot = NA,
  save_plot = NA,
  save_param = list(),
  default_save_name = "plotCombineCCcom"
)
}
\arguments{
\item{gobject}{giotto object}

\item{combCCcom}{combined communcation scores, output from combCCcom()}

\item{selected_LR}{selected ligand-receptor pair}

\item{selected_cell_LR}{selected cell-cell interaction pair for ligand-receptor pair}

\item{detail_plot}{show detailed info in both interacting cell types}

\item{simple_plot}{show a simplified plot}

\item{simple_plot_facet}{facet on interactions or genes with simple plot}

\item{facet_scales}{ggplot facet scales paramter}

\item{facet_ncol}{ggplot facet ncol parameter}

\item{facet_nrow}{ggplot facet nrow parameter}

\item{colors}{vector with two colors to use}

\item{show_plot}{logical. show plot}

\item{return_plot}{logical. return ggplot object}

\item{save_plot}{logical. save the plot}

\item{save_param}{list of saving parameters, see \code{\link{showSaveParameters}}}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}
}
\value{
ggplot
}
\description{
Create visualization for combined (pairwise) cell proximity gene scores
}
