% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction_visuals.R
\name{plotCombineCellCellCommunication}
\alias{plotCombineCellCellCommunication}
\title{plotCombineCellCellCommunication}
\usage{
plotCombineCellCellCommunication(
  gobject,
  combCCcom,
  selected_LR = NULL,
  selected_cell_LR = NULL,
  detail_plot = TRUE,
  simple_plot = FALSE,
  simple_plot_facet = c("interaction", "genes"),
  facet_scales = "fixed",
  facet_ncol = length(selected_LR),
  facet_nrow = length(selected_cell_LR),
  colors = c("#9932CC", "#FF8C00"),
  show_plot = NULL,
  return_plot = NULL,
  save_plot = NULL,
  save_param = list(),
  default_save_name = "plotCombineCellCellCommunication"
)
}
\arguments{
\item{gobject}{giotto object}

\item{combCCcom}{combined communcation scores, output from combCCcom()}

\item{selected_LR}{selected ligand-receptor pair}

\item{selected_cell_LR}{selected cell-cell interaction pair for 
ligand-receptor pair}

\item{detail_plot}{show detailed info in both interacting cell types}

\item{simple_plot}{show a simplified plot}

\item{simple_plot_facet}{facet on interactions or genes with simple plot}

\item{facet_scales}{ggplot facet scales paramter}

\item{facet_ncol}{ggplot facet ncol parameter}

\item{facet_nrow}{ggplot facet nrow parameter}

\item{colors}{vector with two colors to use}

\item{show_plot}{logical. show plot}

\item{return_plot}{logical. return ggplot object}

\item{save_plot}{logical. save the plot}

\item{save_param}{list of saving parameters, see \code{\link{showSaveParameters}}}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}
}
\value{
ggplot
}
\description{
Create visualization for combined (pairwise) cell proximity 
gene scores
}
\examples{
g <- GiottoData::loadGiottoMini("visium")

comScores <- exprCellCellcom(g, cluster_column = "leiden_clus", 
feat_set_1 = c("Gm19935", "2900040C04Rik", "Ccnd2"), 
feat_set_2 = c("9630013A20Rik", "Gna12", "Btbd17"))

exprCC <- exprCellCellcom(g, cluster_column = "leiden_clus", 
feat_set_1 = "Gm19935", feat_set_2 = "9630013A20Rik")
spatialCC <- spatCellCellcom(gobject = g, cluster_column = "leiden_clus",
feat_set_1 = "Gm19935", feat_set_2 = "9630013A20Rik", verbose = "a lot")

combCCcom <- combCCcom(spatialCC = spatialCC, exprCC = exprCC)

plotCombineCellCellCommunication(gobject = g, combCCcom = combCCcom,
selected_LR = c("Gm19935-9630013A20Rik"), selected_cell_LR = c("1--1"))
}
