% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction_visuals.R
\name{plotCombineCPG}
\alias{plotCombineCPG}
\title{plotCombineCPG}
\usage{
plotCombineCPG(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=plotCombineICG]{plotCombineICG}}
  \describe{
    \item{\code{gobject}}{giotto object}
    \item{\code{combCpgObject}}{ICGscores, output from combineInteractionChangedGenes()}
    \item{\code{selected_interactions}}{interactions to show}
    \item{\code{selected_gene_to_gene}}{pairwise gene combinations to show}
    \item{\code{detail_plot}}{show detailed info in both interacting cell types}
    \item{\code{simple_plot}}{show a simplified plot}
    \item{\code{simple_plot_facet}}{facet on interactions or genes with simple plot}
    \item{\code{facet_scales}}{ggplot facet scales paramter}
    \item{\code{facet_ncol}}{ggplot facet ncol parameter}
    \item{\code{facet_nrow}}{ggplot facet nrow parameter}
    \item{\code{colors}}{vector with two colors to use}
    \item{\code{show_plot}}{show plots}
    \item{\code{return_plot}}{return plotting object}
    \item{\code{save_plot}}{directly save the plot [boolean]}
    \item{\code{save_param}}{list of saving parameters from \code{\link{all_plots_save_function}}}
    \item{\code{default_save_name}}{default save name for saving, don't change, change save_name in save_param}
  }}
}
\description{
Create visualization for combined (pairwise) ICG scores
}
\seealso{
\code{\link{plotCombineICG}}
}
