% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction_visuals.R
\name{plotCombineICF}
\alias{plotCombineICF}
\title{plotCombineICF}
\usage{
plotCombineICF(
  gobject,
  combIcfObject,
  selected_interactions = NULL,
  selected_feat_to_feat = NULL,
  detail_plot = T,
  simple_plot = F,
  simple_plot_facet = c("interaction", "feats"),
  facet_scales = "fixed",
  facet_ncol = length(selected_feat_to_feat),
  facet_nrow = length(selected_interactions),
  colors = c("#9932CC", "#FF8C00"),
  show_plot = NA,
  return_plot = NA,
  save_plot = NA,
  save_param = list(),
  default_save_name = "plotCombineICF"
)
}
\arguments{
\item{gobject}{giotto object}

\item{combIcfObject}{ICFscores, output from combineInteractionChangedFeats()}

\item{selected_interactions}{interactions to show}

\item{selected_feat_to_feat}{pairwise feature combinations to show}

\item{detail_plot}{show detailed info in both interacting cell types}

\item{simple_plot}{show a simplified plot}

\item{simple_plot_facet}{facet on interactions or feats with simple plot}

\item{facet_scales}{ggplot facet scales paramter}

\item{facet_ncol}{ggplot facet ncol parameter}

\item{facet_nrow}{ggplot facet nrow parameter}

\item{colors}{vector with two colors to use}

\item{show_plot}{show plots}

\item{return_plot}{return plotting object}

\item{save_plot}{directly save the plot [boolean]}

\item{save_param}{list of saving parameters from \code{\link{all_plots_save_function}}}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}
}
\value{
ggplot
}
\description{
Create visualization for combined (pairwise) ICF scores
}
