% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction_visuals.R
\name{plotCombineInteractionChangedFeats}
\alias{plotCombineInteractionChangedFeats}
\title{plotCombineInteractionChangedFeats}
\usage{
plotCombineInteractionChangedFeats(
  gobject,
  combIcfObject,
  selected_interactions = NULL,
  selected_feat_to_feat = NULL,
  detail_plot = TRUE,
  simple_plot = FALSE,
  simple_plot_facet = c("interaction", "feats"),
  facet_scales = "fixed",
  facet_ncol = length(selected_feat_to_feat),
  facet_nrow = length(selected_interactions),
  colors = c("#9932CC", "#FF8C00"),
  show_plot = NULL,
  return_plot = NULL,
  save_plot = NULL,
  save_param = list(),
  default_save_name = "plotCombineICF"
)
}
\arguments{
\item{gobject}{giotto object}

\item{combIcfObject}{ICFscores, output from combineInteractionChangedFeats()}

\item{selected_interactions}{interactions to show}

\item{selected_feat_to_feat}{pairwise feature combinations to show}

\item{detail_plot}{show detailed info in both interacting cell types}

\item{simple_plot}{show a simplified plot}

\item{simple_plot_facet}{facet on interactions or feats with simple plot}

\item{facet_scales}{ggplot facet scales paramter}

\item{facet_ncol}{ggplot facet ncol parameter}

\item{facet_nrow}{ggplot facet nrow parameter}

\item{colors}{vector with two colors to use}

\item{show_plot}{logical. show plot}

\item{return_plot}{logical. return ggplot object}

\item{save_plot}{logical. save the plot}

\item{save_param}{list of saving parameters, see 
\code{\link{showSaveParameters}}}

\item{default_save_name}{default save name for saving, don't change, 
change save_name in save_param}
}
\value{
ggplot
}
\description{
Create visualization for combined (pairwise) ICF scores
}
\examples{
g <- GiottoData::loadGiottoMini("visium")

g_icf <- findInteractionChangedFeats(g,
    cluster_column = "leiden_clus",
    selected_feats = c("Gna12", "Ccnd2", "Btbd17"), nr_permutations = 10
)

combIcfObject <- combineInteractionChangedFeats(g_icf)

plotCombineInteractionChangedFeats(
    gobject = g,
    combIcfObject = combIcfObject,
    selected_feat_to_feat = c("Btbd17--Ccnd2", "Btbd17--Gna12"),
    selected_interactions = "1--8"
)
}
