% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction_visuals.R
\name{plotCPF}
\alias{plotCPF}
\title{plotCPF}
\usage{
plotCPF(
  gobject,
  icfObject,
  method = c("volcano", "cell_barplot", "cell-cell", "cell_sankey", "heatmap", "dotplot"),
  min_cells = 5,
  min_cells_expr = 1,
  min_int_cells = 3,
  min_int_cells_expr = 1,
  min_fdr = 0.05,
  min_spat_diff = 0.2,
  min_log2_fc = 0.2,
  min_zscore = 2,
  zscores_column = c("cell_type", "feats"),
  direction = c("both", "up", "down"),
  cell_color_code = NULL,
  show_plot = NULL,
  return_plot = NULL,
  save_plot = NULL,
  save_param = list(),
  default_save_name = "plotCPG"
)
}
\arguments{
\item{gobject}{giotto object}

\item{icfObject}{ICF (interaction changed feature) score object}

\item{method}{plotting method to use}

\item{min_cells}{minimum number of source cell type}

\item{min_cells_expr}{minimum expression level for source cell type}

\item{min_int_cells}{minimum number of interacting neighbor cell type}

\item{min_int_cells_expr}{minimum expression level for interacting neighbor cell type}

\item{min_fdr}{minimum adjusted p-value}

\item{min_spat_diff}{minimum absolute spatial expression difference}

\item{min_log2_fc}{minimum log2 fold-change}

\item{min_zscore}{minimum z-score change}

\item{zscores_column}{calculate z-scores over cell types or features}

\item{direction}{differential expression directions to keep}

\item{cell_color_code}{character. discrete colors to use. palette to use or
named vector of colors}

\item{show_plot}{logical. show plot}

\item{return_plot}{logical. return ggplot object}

\item{save_plot}{logical. save the plot}

\item{save_param}{list of saving parameters, see \code{\link{showSaveParameters}}}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}
}
\value{
plot
}
\description{
Create visualization for cell proximity feature scores
}
