% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction_visuals.R
\name{plotCPGscores}
\alias{plotCPGscores}
\title{plotCPGscores}
\usage{
plotCPGscores(CPGscores, selected_interactions = NULL,
  selected_genes = NULL, detail_plot = T, simple_plot = F,
  simple_plot_facet = c("interaction", "genes"),
  facet_scales = "fixed", facet_ncol = length(selected_genes),
  facet_nrow = length(selected_interactions), show_plot = F)
}
\arguments{
\item{CPGscores}{CPGscores, output from getCellProximityGeneScores()}

\item{selected_interactions}{interactions to show}

\item{selected_genes}{genes to show}

\item{detail_plot}{show detailed info in both interacting cell types}

\item{simple_plot}{show a simplified plot}

\item{simple_plot_facet}{facet on interactions or genes with simple plot}

\item{facet_scales}{ggplot facet scales paramter}

\item{facet_ncol}{ggplot facet ncol parameter}

\item{facet_nrow}{ggplot facet nrow parameter}

\item{show_plot}{show plot}
}
\value{
ggplot barplot
}
\description{
Create heatmap from cell-cell proximity scores
}
\details{
Give more details ...
}
\examples{
    plotCPGscores(CPGscores)
}
