% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{plotGiottoLargeImage}
\alias{plotGiottoLargeImage}
\title{plotGiottoLargeImage}
\usage{
plotGiottoLargeImage(
  gobject = NULL,
  largeImage_name = NULL,
  giottoLargeImage = NULL,
  crop_extent = NULL,
  xmax_crop = NULL,
  xmin_crop = NULL,
  ymax_crop = NULL,
  ymin_crop = NULL,
  max_intensity = NULL
)
}
\arguments{
\item{gobject}{giotto object}

\item{largeImage_name}{name of giottoLargeImage}

\item{giottoLargeImage}{giottoLargeImage object}

\item{crop_extent}{extent object to focus on specific region of image}

\item{xmax_crop}{xmax crop boundary}

\item{xmin_crop}{xmin crop boundary}

\item{ymax_crop}{ymax crop boundary}

\item{ymin_crop}{ymin crop boundary}

\item{max_intensity}{value to treat as maximum intensity in color scale}
}
\value{
plot
}
\description{
Plot a downsampled version of giottoLargeImage. Cropping can increase plot resolution of region of interest.
}
