% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction_visuals.R
\name{plotGTGscores}
\alias{plotGTGscores}
\title{plotGTGscores}
\usage{
plotGTGscores(GTGscore, selected_interactions = NULL,
  selected_gene_to_gene = NULL, detail_plot = T, simple_plot = F,
  simple_plot_facet = c("interaction", "genes"),
  facet_scales = "fixed", facet_ncol = length(selected_gene_to_gene),
  facet_nrow = length(selected_interactions), colors = c("blue",
  "red"), show_plot = F)
}
\arguments{
\item{GTGscore}{GTGscore, output from getGeneToGeneScores()}

\item{selected_interactions}{interactions to show}

\item{detail_plot}{show detailed info in both interacting cell types}

\item{simple_plot}{show a simplified plot}

\item{simple_plot_facet}{facet on interactions or genes with simple plot}

\item{facet_scales}{ggplot facet scales paramter}

\item{facet_ncol}{ggplot facet ncol parameter}

\item{facet_nrow}{ggplot facet nrow parameter}

\item{colors}{vector with 2 colors to represent respectively all and selected cells}

\item{show_plot}{show plot}

\item{selected_genes}{genes to show}
}
\value{
ggplot barplot
}
\description{
Create heatmap from cell-cell proximity scores
}
\details{
Give more details ...
}
\examples{
    plotGTGscores(GTGscore)
}
