% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_visuals.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{plotHeatmap}
\usage{
plotHeatmap(gobject, expression_values = c("normalized", "scaled",
  "custom"), genes, cluster_column = NULL, cluster_order = c("size",
  "correlation", "custom"), cluster_custom_order = NULL,
  gene_order = c("custom", "correlation"), show_values = c("rescaled",
  "z-scaled", "original"))
}
\arguments{
\item{gobject}{giotto object}

\item{expression_values}{expression values to use}

\item{genes}{genes to plot}

\item{cluster_column}{name of column to use for clusters}

\item{cluster_order}{method to order the clusters (see details)}

\item{cluster_custom_order}{vector with custom order of clusters}

\item{gene_order}{order of genes (see details)}

\item{show_values}{values to plot on heatmap (see details)}
}
\value{
ggplot
}
\description{
Creates heatmap based on identified clusters
}
\details{
Correlation heatmap of clusters vs genes.
}
\examples{
    plotHeatmap(gobject)
}
