% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction_spot.R
\name{plotICFSpot}
\alias{plotICFSpot}
\title{plotICFSpot}
\usage{
plotICFSpot(
  gobject,
  icfObject,
  source_type,
  source_markers,
  ICF_features,
  cell_color_code = NULL,
  show_plot = NULL,
  return_plot = NULL,
  save_plot = NULL,
  save_param = list(),
  default_save_name = "plotICFSpot"
)
}
\arguments{
\item{gobject}{giotto object}

\item{icfObject}{ICF (interaction changed feature) score object}

\item{source_type}{cell type of the source cell}

\item{source_markers}{markers for the source cell type}

\item{ICF_features}{named character vector of ICF features}

\item{cell_color_code}{cell color code for the interacting cell types}

\item{show_plot}{logical. show plot}

\item{return_plot}{logical. return ggplot object}

\item{save_plot}{logical. save the plot}

\item{save_param}{list of saving parameters, see \code{\link{showSaveParameters}}}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}
}
\value{
plot
}
\description{
Create barplot to visualize interaction changed features
}
\examples{
g <- GiottoData::loadGiottoMini("visium")
icfObject <- findInteractionChangedFeats(g, cluster_column = "leiden_clus",
selected_feats = c("Gna12", "Ccnd2", "Btbd17"), nr_permutations = 10)

plotICFSpot(gobject = g, icfObject = icfObject,
source_type = "1", source_markers = "Ccnd2", 
ICF_features = c("3" = "Gna12", "1" = "Ccnd2", "8" = "Btbd17"))
}
