% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction_visuals.R
\name{plotInteractionChangedFeats}
\alias{plotInteractionChangedFeats}
\title{Plot interaction changed features}
\usage{
plotInteractionChangedFeats(
  gobject,
  icfObject,
  source_type,
  source_markers,
  ICF_feats,
  cell_color_code = NULL,
  show_plot = NA,
  return_plot = NA,
  save_plot = NA,
  save_param = list(),
  default_save_name = "plotInteractionChangedFeats"
)
}
\arguments{
\item{gobject}{giotto object}

\item{icfObject}{ICF (interaction changed feature) score object}

\item{source_type}{cell type of the source cell}

\item{source_markers}{markers for the source cell type}

\item{ICF_feats}{named character vector of ICF features}

\item{cell_color_code}{cell color code for the interacting cell types}

\item{show_plot}{show plots}

\item{return_plot}{return plotting object}

\item{save_plot}{directly save the plot [boolean]}

\item{save_param}{list of saving parameters from \code{\link{all_plots_save_function}}}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}
}
\value{
plot
}
\description{
Create barplot to visualize interaction changed features
}
