% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction_visuals.R
\name{plotInteractionChangedGenes}
\alias{plotInteractionChangedGenes}
\title{Plot interaction changed genes}
\usage{
plotInteractionChangedGenes(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=plotInteractionChangedFeats]{plotInteractionChangedFeats}}
  \describe{
    \item{\code{gobject}}{giotto object}
    \item{\code{icfObject}}{ICF (interaction changed feature) score object}
    \item{\code{source_type}}{cell type of the source cell}
    \item{\code{source_markers}}{markers for the source cell type}
    \item{\code{ICF_feats}}{named character vector of ICF features}
    \item{\code{cell_color_code}}{cell color code for the interacting cell types}
    \item{\code{show_plot}}{show plots}
    \item{\code{return_plot}}{return plotting object}
    \item{\code{save_plot}}{directly save the plot [boolean]}
    \item{\code{save_param}}{list of saving parameters from \code{\link{all_plots_save_function}}}
    \item{\code{default_save_name}}{default save name for saving, don't change, change save_name in save_param}
  }}
}
\description{
Create barplot to visualize interaction changed features
}
\seealso{
\code{\link{plotInteractionChangedFeats}}
}
