% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactivity.R
\name{plotInteractive3D}
\alias{plotInteractive3D}
\title{Plot interactive 3D spatial plot}
\usage{
plotInteractive3D(
  gobject,
  spat_unit = "cell",
  feat_type = "rna",
  cell_color = "leiden_clus",
  cell_color_code = NULL,
  point_size = 0.5,
  width = "100\%",
  height = "400px"
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit (e.g. "cell")}

\item{feat_type}{feature type (e.g. "rna", "dna", "protein")}

\item{cell_color}{character. What to color cells by
(e.g. metadata col or spatial enrichment col)}

\item{cell_color_code}{character. discrete colors to use. Palette to use or
named vector of colors}

\item{point_size}{size of point (cell)}

\item{width}{plot width}

\item{height}{plot height}
}
\value{
data.table with selected cell_IDs, spatial coordinates, and
cluster_ID.
}
\description{
Plot interactive 3D spatial plot
}
