% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_visuals.R
\name{plotly_axis_scale_2D}
\alias{plotly_axis_scale_2D}
\title{plotly_axis_scale_2D}
\usage{
plotly_axis_scale_2D(
  cell_locations,
  sdimx = NULL,
  sdimy = NULL,
  mode = c("cube", "real", "custom"),
  custom_ratio = NULL
)
}
\arguments{
\item{cell_locations}{spatial_loc in giotto object}

\item{sdimx}{x axis of cell spatial location}

\item{sdimy}{y axis of cell spatial location}

\item{mode}{axis adjustment mode}

\item{custom_ratio}{set the ratio artificially}
}
\value{
edges in spatial grid as data.table()
}
\description{
adjust the axis scale in 2D plotly plot
}
\keyword{internal}
