% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_visuals.R
\name{plotly_spatial_network}
\alias{plotly_spatial_network}
\title{plotly_spatial_network}
\usage{
plotly_spatial_network(spatial_network)
}
\arguments{
\item{gobject}{spatial_network in giotto object}
}
\value{
edges in network as data.table()
}
\description{
provide network segment to draw in 3D plot_ly()
}
\examples{
    plotly_spatial_network(gobject)
}
