% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_visuals.R
\name{plotMetaDataHeatmap}
\alias{plotMetaDataHeatmap}
\title{plotMetaDataHeatmap}
\usage{
plotMetaDataHeatmap(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  expression_values = c("normalized", "scaled", "custom"),
  metadata_cols = NULL,
  selected_feats = NULL,
  selected_genes = NULL,
  first_meta_col = NULL,
  second_meta_col = NULL,
  show_values = c("zscores", "original", "zscores_rescaled"),
  custom_cluster_order = NULL,
  clus_cor_method = "pearson",
  clus_cluster_method = "complete",
  custom_feat_order = NULL,
  custom_gene_order = NULL,
  feat_cor_method = "pearson",
  gene_cor_method = NULL,
  feat_cluster_method = "complete",
  gene_cluster_method = NULL,
  gradient_color = c("blue", "white", "red"),
  gradient_midpoint = 0,
  gradient_limits = NULL,
  x_text_size = 10,
  x_text_angle = 45,
  y_text_size = 10,
  strip_text_size = 8,
  plot_title = NULL,
  show_plot = NA,
  return_plot = NA,
  save_plot = NA,
  save_param = list(),
  default_save_name = "plotMetaDataHeatmap"
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit (e.g. "cell")}

\item{feat_type}{feature type (e.g. "rna", "dna", "protein")}

\item{expression_values}{expression values to use (e.g. "normalized", "scaled", "custom")}

\item{metadata_cols}{annotation columns found in pDataDT(gobject)}

\item{selected_feats}{subset of features to use}

\item{selected_genes}{deprecated. See \code{selected_feats} param}

\item{first_meta_col}{if more than 1 metadata column, select the x-axis factor}

\item{second_meta_col}{if more than 1 metadata column, select the facetting factor}

\item{show_values}{which values to show on heatmap (e.g. "zscores", "original", "zscores_rescaled")}

\item{custom_cluster_order}{custom cluster order (default = NULL)}

\item{clus_cor_method}{correlation method for clusters, default to "pearson"}

\item{clus_cluster_method}{hierarchical cluster method for the clusters, default to "complete"}

\item{custom_feat_order}{custom feature order (default = NULL)}

\item{custom_gene_order}{deprecated. See \code{custom_feat_order} param}

\item{feat_cor_method}{correlation method for features, default to "pearson"}

\item{gene_cor_method}{deprecated. See \code{feat_cor_method} param}

\item{feat_cluster_method}{hierarchical cluster method for the features, default to "complete"}

\item{gene_cluster_method}{deprecated. See \code{feat_cluster_method} param}

\item{gradient_color}{vector with 3 colors for numeric data}

\item{gradient_midpoint}{midpoint for color gradient}

\item{gradient_limits}{vector with lower and upper limits}

\item{x_text_size}{size of x-axis text}

\item{x_text_angle}{angle of x-axis text}

\item{y_text_size}{size of y-axis text}

\item{strip_text_size}{size of strip text}

\item{show_plot}{show plot. TRUE or FALSE}

\item{return_plot}{return ggplot object. TRUE or FALSE}

\item{save_plot}{directly save the plot. TRUE or FALSE}

\item{save_param}{list of saving parameters, see \code{\link{showSaveParameters}}}

\item{default_save_name}{default save name}
}
\value{
ggplot or data.table
}
\description{
Creates heatmap for genes within aggregated clusters.
}
\details{
Creates heatmap for the average expression of selected genes in the different annotation/cluster groups.
Calculation of cluster or gene order is done on the provided expression values, but visualization
is by default on the z-scores. Other options are the original values or z-scores rescaled per gene (-1 to 1).
}
\seealso{
\code{\link{plotMetaDataCellsHeatmap}} for numeric cell annotation instead of gene expression.
}
