% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONTraC_wrapper.R
\name{plotNicheClusterConnectivity}
\alias{plotNicheClusterConnectivity}
\title{plotNicheClusterConnectivity}
\usage{
plotNicheClusterConnectivity(
  gobject,
  spat_unit = "niche cluster",
  feat_type = "connectivity",
  values = "normalized",
  show_plot = NULL,
  return_plot = NULL,
  save_plot = NULL,
  save_param = list(),
  default_save_name = "NicheClusterConnectivity"
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{name of spatial unit niche stored cluster features}

\item{feat_type}{name of the feature type stored niche cluster connectivities}

\item{values}{name of the expression matrix stored connectivity values}

\item{show_plot}{logical. show plot}

\item{return_plot}{logical. return ggplot object}

\item{save_plot}{logical. save the plot}

\item{save_param}{list of saving parameters, see 
\code{\link{showSaveParameters}}}

\item{default_save_name}{default save name for saving, don't change, 
change save_name in save_param}
}
\description{
plot niche cluster connectivity
}
\details{
This function plots the niche cluster connectivity matrix
}
