% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visuals.R
\name{plotPCA}
\alias{plotPCA}
\title{plotPCA}
\usage{
plotPCA(gobject, ...)
}
\arguments{
\item{gobject}{giotto object}

\item{...}{other parameters that are part of visDimPlot()}
}
\value{
ggplot
}
\description{
Short wrapper for PCA visualization
}
\details{
Description of parameters.
}
\examples{
    plotPCA(gobject)
}
\seealso{
\code{\link{visDimPlot}}
}
