% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visuals.R
\name{plotPolygons}
\alias{plotPolygons}
\title{Plot stored polygons}
\usage{
plotPolygons(gobject, x, spat_unit = "cell", polygons = NULL, ...)
}
\arguments{
\item{gobject}{A Giotto object with polygon coordinates}

\item{x}{A ggplot2, spatPlot or terra::rast object}

\item{spat_unit}{spatial unit}

\item{polygons}{character. Vector of polygon names to plot. If NULL, all polygons are plotted}

\item{...}{Additional parameters passed to ggplot2::geom_polygon() or graphics::polygon}
}
\value{
A ggplot2 image
}
\description{
Plot stored polygons
}
