% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction_visuals.R
\name{plotRankSpatvsExpr}
\alias{plotRankSpatvsExpr}
\title{plotRankSpatvsExpr}
\usage{
plotRankSpatvsExpr(
  gobject,
  combCC,
  expr_rnk_column = "LR_expr_rnk",
  spat_rnk_column = "LR_spat_rnk",
  dot_color_gradient = NULL,
  midpoint = deprecated(),
  gradient_midpoint = 10,
  gradient_style = c("divergent", "sequential"),
  size_range = c(0.01, 1.5),
  xlims = NULL,
  ylims = NULL,
  selected_ranks = c(1, 10, 20),
  show_plot = NULL,
  return_plot = NULL,
  save_plot = NULL,
  save_param = list(),
  default_save_name = "plotRankSpatvsExpr"
)
}
\arguments{
\item{gobject}{giotto object}

\item{combCC}{combined communication scores from \code{\link{combCCcom}}}

\item{expr_rnk_column}{column with expression rank information to use}

\item{spat_rnk_column}{column with spatial rank information to use}

\item{dot_color_gradient}{character. continuous colors to use. palette to
use or vector of colors to use (minimum of 2).}

\item{midpoint}{deprecated}

\item{gradient_midpoint}{numeric. default = 10. midpoint of colors}

\item{gradient_style}{either 'divergent' (midpoint is used in color scaling)
or 'sequential' (scaled based on data range)}

\item{size_range}{size ranges of dotplot}

\item{xlims}{x-limits, numerical vector of 2}

\item{ylims}{y-limits, numerical vector of 2}

\item{selected_ranks}{numerical vector, will be used to print out the
percentage of top spatial ranks are recovered}

\item{show_plot}{logical. show plot}

\item{return_plot}{logical. return ggplot object}

\item{save_plot}{logical. save the plot}

\item{save_param}{list of saving parameters, see \code{\link{showSaveParameters}}}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}
}
\value{
ggplot
}
\description{
Plots dotplot to compare ligand-receptor rankings from
spatial and expression information
}
\examples{
g <- GiottoData::loadGiottoMini("visium")

exprCC <- exprCellCellcom(g,
    cluster_column = "leiden_clus",
    feat_set_1 = "Gm19935", feat_set_2 = "9630013A20Rik"
)
spatialCC <- spatCellCellcom(
    gobject = g, cluster_column = "leiden_clus",
    feat_set_1 = "Gm19935", feat_set_2 = "9630013A20Rik", verbose = "a lot",
    random_iter = 10
)

combCC <- combCCcom(spatialCC = spatialCC, exprCC = exprCC)

plotRankSpatvsExpr(gobject = g, combCC = combCC)
}
