% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction_visuals.R
\name{plotRecovery_sub}
\alias{plotRecovery_sub}
\title{plotRecovery_sub}
\usage{
plotRecovery_sub(combCC, first_col = "LR_expr_rnk", second_col = "LR_spat_rnk")
}
\arguments{
\item{combCC}{combined communinication scores from \code{\link{combCCcom}}}

\item{first_col}{first column to use}

\item{second_col}{second column to use}
}
\description{
Plots recovery plot to compare ligand-receptor rankings from spatial and expression information
}
