% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONTraC_wrapper.R
\name{plotSpatNicheClusterBin}
\alias{plotSpatNicheClusterBin}
\title{plotSpatNicheClusterBin}
\usage{
plotSpatNicheClusterBin(
  gobject,
  spat_unit = "cell",
  feat_type = "niche cluster",
  ...,
  default_save_name = "spatNicheClusterBin"
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{name of spatial unit niche stored cluster features}

\item{feat_type}{name of the feature type stored binarized niche cluster}

\item{...}{additional arguments to be passed to the spatFeatPlot2D function}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}

\item{niche_cluster_label}{name of the niche cluster label}
}
\description{
plot spatial niche cluster binarized
}
\details{
This function plots the spatial niche cluster binarized
}
