% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONTraC_wrapper.R
\name{plotSpatNicheClusterProb}
\alias{plotSpatNicheClusterProb}
\title{plotSpatNicheClusterProb}
\usage{
plotSpatNicheClusterProb(
  gobject,
  spat_unit = "cell",
  feat_type = "niche cluster",
  expression_values = "prob",
  ...,
  default_save_name = "spatNicheClusterProb"
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{name of spatial unit niche stored cluster features}

\item{feat_type}{name of the feature type stored probability matrix}

\item{expression_values}{name of the expression matrix stored probability
values}

\item{...}{additional arguments to be passed to the spatFeatPlot2D function}

\item{default_save_name}{default save name for saving, don't change, 
change save_name in save_param}
}
\description{
plot spatial niche cluster probability
}
\details{
This function plots the spatial niche cluster probability
}
