% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_structures.R
\name{plotStatDelaunayNetwork}
\alias{plotStatDelaunayNetwork}
\title{plotStatDelaunayNetwork}
\usage{
plotStatDelaunayNetwork(
  gobject,
  dimensions = c("sdimx", "sdimy"),
  name = "delaunay_network",
  maximum_distance = "auto",
  minimum_k = 0,
  Y = TRUE,
  j = TRUE,
  S = 0,
  show_plot = NA,
  return_plot = NA,
  save_plot = NA,
  save_param = list(),
  default_save_name = "plotStatDelaunayNetwork",
  ...
)

plotStatDelaunayNetwork(
  gobject,
  dimensions = c("sdimx", "sdimy"),
  name = "delaunay_network",
  maximum_distance = "auto",
  minimum_k = 0,
  Y = TRUE,
  j = TRUE,
  S = 0,
  show_plot = NA,
  return_plot = NA,
  save_plot = NA,
  save_param = list(),
  default_save_name = "plotStatDelaunayNetwork",
  ...
)
}
\arguments{
\item{gobject}{giotto object}

\item{dimensions}{which spatial dimensions to use (maximum 2 dimensions)}

\item{name}{name for spatial network (default = 'delaunay_network')}

\item{maximum_distance}{distance cuttof for Delaunay neighbors to consider}

\item{minimum_k}{minimum neigbhours if maximum_distance != NULL}

\item{Y}{(RTriangle) If TRUE prohibits the insertion of Steiner points on the mesh boundary.}

\item{j}{(RTriangle) If TRUE jettisons vertices that are not part of the final triangulation from the output.}

\item{S}{(RTriangle) Specifies the maximum number of added Steiner points.}

\item{show_plot}{show plots}

\item{return_plot}{return ggplot object}

\item{save_plot}{directly save the plot [boolean]}

\item{save_param}{list of saving parameters from \code{\link{all_plots_save_function}}}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}

\item{...}{Other parameters of the \code{\link[RTriangle]{triangulate}} function}

\item{verbose}{verbose}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}
}
\value{
giotto object with updated spatial network slot

giotto object with updated spatial network slot
}
\description{
Plots network statistics for a Delaunay network..

Create a spatial Delaunay network based on cell centroid physical distances.
}
\details{
Plots statistics for a spatial Delaunay network as explained in \code{\link[RTriangle]{triangulate}}.
This can be used to further finetune the \code{\link{createDelaunayNetwork}} function.

Creates a spatial Delaunay network as explained in \code{\link[RTriangle]{triangulate}}.
}
\examples{
    plotStatDelaunayNetwork(gobject)
    createDelaunayNetwork(gobject)
}
