% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_structures.R
\name{plotStatDelaunayNetwork_v2}
\alias{plotStatDelaunayNetwork_v2}
\title{plotStatDelaunayNetwork_v2}
\usage{
plotStatDelaunayNetwork_v2(
  gobject,
  method = c("delaunayn_geometry", "RTriangle", "deldir"),
  dimensions = "all",
  maximum_distance = "auto",
  minimum_k = 0,
  options = "Pp",
  Y = TRUE,
  j = TRUE,
  S = 0,
  show_plot = NA,
  return_plot = NA,
  save_plot = NA,
  save_param = list(),
  default_save_name = "plotStatDelaunayNetwork",
  ...
)
}
\arguments{
\item{gobject}{giotto object}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}
}
\value{
ggplot
}
\description{
Plot Delaunay network stats.
}
\examples{
    plotStatDelaunayNetwork_v2(gobject)
}
