% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visuals.R
\name{plotTSNE}
\alias{plotTSNE}
\title{plotTSNE}
\usage{
plotTSNE(gobject, ...)
}
\arguments{
\item{gobject}{giotto object}

\item{...}{other parameters that are part of visDimPlot()}
}
\value{
ggplot
}
\description{
Short wrapper for tSNE visualization
}
\details{
Description of parameters.
}
\examples{
    plotTSNE(gobject)
}
\seealso{
\code{\link{visDimPlot}}
}
