% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visuals.R
\name{plotUMAP}
\alias{plotUMAP}
\title{plotUMAP}
\usage{
plotUMAP(gobject, ...)
}
\arguments{
\item{gobject}{giotto object}

\item{...}{other parameters that are part of visDimPlot()}
}
\value{
ggplot
}
\description{
Short wrapper for UMAP visualization
}
\details{
Description of parameters.
}
\examples{
    plotUMAP(gobject)
}
\seealso{
\code{\link{visDimPlot}}
}
