% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visuals.R
\name{plotUMAP_2D}
\alias{plotUMAP_2D}
\title{plotUMAP_2D}
\usage{
plotUMAP_2D(gobject, ...)
}
\arguments{
\item{gobject}{giotto object}

\item{dim1_to_use}{dimension to use on x-axis}

\item{dim2_to_use}{dimension to use on y-axis}

\item{dim3_to_use}{dimension to use on z-axis}

\item{show_NN_network}{show underlying NN network}

\item{nn_network_to_use}{type of NN network to use (kNN vs sNN)}

\item{network_name}{name of NN network to use, if show_NN_network = TRUE}

\item{cell_color}{color for cells (see details)}

\item{color_as_factor}{convert color column to factor}

\item{cell_color_code}{named vector with colors}

\item{show_cluster_center}{plot center of selected clusters}

\item{show_center_label}{plot label of selected clusters}

\item{center_point_size}{size of center points}

\item{label_size}{size of labels}

\item{label_fontface}{font of labels}

\item{edge_alpha}{column to use for alpha of the edges}

\item{point_size}{size of point (cell)}

\item{point_border_col}{color of border around points}

\item{point_border_stroke}{stroke size of border around points}

\item{show_legend}{show legend}

\item{show_plot}{show plot}

\item{return_plot}{return ggplot object}

\item{save_plot}{directly save the plot [boolean]}

\item{save_dir}{directory to save the plot}

\item{save_folder}{(optional) folder in directory to save the plot}

\item{save_name}{name of plot}

\item{save_format}{format of plot (e.g. tiff, png, pdf, ...)}

\item{show_saved_plot}{load & display the saved plot}
}
\value{
ggplot
}
\description{
Short wrapper for UMAP visualization
}
\details{
Description of parameters.
}
\examples{
    plotUMAP_2D(gobject)
}
\seealso{
\code{\link{dimPlot2D}}
}
