% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_structures.R
\name{polygon_to_raster}
\alias{polygon_to_raster}
\title{Convert polygon to raster}
\usage{
polygon_to_raster(polygon, field = NULL)

polygon_to_raster(polygon, field = NULL)
}
\description{
function to convert terra SpatVector Polygon shape into a terra SpatRaster

convert polygon to raster
}
\keyword{internal}
