% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_structures.R
\name{polygon_to_raster_OLD}
\alias{polygon_to_raster_OLD}
\title{Convert polygon to raster}
\usage{
polygon_to_raster_OLD(polygon, field = NULL)
}
\description{
function to convert terra SpatVector Polygon shape into a terra SpatRaster
TODO: can be removed
}
\keyword{internal}
