% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction.R
\name{print.combIcfObject}
\alias{print.combIcfObject}
\title{combIcfObject print method}
\usage{
print.combIcfObject(x, ...)
}
\arguments{
\item{x}{object to print}

\item{\dots}{additional params to pass (none implemented)}
}
\description{
combIcfObject print method
}
\keyword{internal}
