% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto.R
\name{print.giotto}
\alias{print.giotto}
\title{print method for giotto class}
\usage{
print.giotto(object, ...)
}
\arguments{
\item{nr_genes}{number of genes (rows) to print}

\item{nr_cells}{number of cells (columns) to print}
}
\description{
print method for giotto class.
Prints the chosen number of genes (rows) and cells (columns) from the raw count matrix.
Also print the spatial locations for the chosen number of cells.
}
\keyword{giotto,}
\keyword{object}
