% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction.R
\name{print.icfObject}
\alias{print.icfObject}
\title{icfObject print method}
\usage{
print.icfObject(x, ...)
}
\arguments{
\item{x}{object to print}

\item{\dots}{additional params to pass (none implemented)}
}
\value{
icfObject
}
\description{
icfObject print method
}
\keyword{internal}
