% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{print_leaf}
\alias{print_leaf}
\title{Hierarchical tree printing}
\usage{
print_leaf(level_index, availableDT, inherit_last = TRUE, indent)
}
\arguments{
\item{level_index}{Which col of availability matrix to start print from}

\item{availableDT}{availability matrix given as data.table}

\item{inherit_last}{(boolean) determine behavior from previous level for last level (intended for values print)
Only TRUE behavior defined. #TODO}

\item{indent}{ident characters to print for this level (top level is '')}
}
\description{
Hierarchical tree printing
}
\details{
Much inspiration taken from https://rdrr.io/cran/fs/src/R/tree.R
}
\keyword{internal}
