% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{processGiotto}
\alias{processGiotto}
\title{processGiotto}
\usage{
processGiotto(
  gobject,
  filter_params = list(),
  norm_params = list(),
  stat_params = list(),
  adjust_params = list(),
  verbose = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{filter_params}{additional parameters to filterGiotto}

\item{norm_params}{additional parameters to normalizeGiotto}

\item{stat_params}{additional parameters to addStatistics}

\item{adjust_params}{additional parameters to adjustGiottoMatrix; set to
NULL if not required}

\item{verbose}{be verbose (default is TRUE)}
}
\value{
giotto object
}
\description{
Wrapper for the different Giotto object processing functions
}
\details{
See \code{\link{filterGiotto}}, \code{\link{normalizeGiotto}},
\code{\link{addStatistics}}, and \code{\link{adjustGiottoMatrix}}. For more
information about the different parameters in each step. If you do not
provide them it will use the default values. If no adjustment is required,
adjust_params must be set to NULL
}
