% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_section.R
\name{projection_fun}
\alias{projection_fun}
\title{projection_fun}
\usage{
projection_fun(point_to_project, plane_point, plane_norm)
}
\arguments{
\item{point_to_project}{point_to_project}

\item{plane_point}{plane_point}

\item{plane_norm}{plane_norm}
}
\description{
project a point onto a plane
}
\keyword{internal}
