% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto.R
\name{prov_match}
\alias{prov_match}
\title{Check provenance info matches across list of S4 subobjects}
\usage{
prov_match(..., verbose = FALSE)
}
\arguments{
\item{...}{list of s4 subobjects to match provenance}

\item{verbose}{print discovered provenance info}
}
\description{
Check provenance info matches across list of S4 subobjects
}
\keyword{internal}
