% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{provData-class}
\alias{provData-class}
\title{Basic class for classes with provenance information.}
\description{
This kind of information is necesssary when generating data that is aggregated
from multiple original sources of raw information. This could refer to situations
such as when producing cellxfeature expression matrices from subcellular transcript
information and polygons that are provided as multiple z layers. Provenance
is Giotto's method of mapping this aggregated information back to the original
z layers that were used in its generation.
}
