% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_environment.R
\name{py_install_prompt}
\alias{py_install_prompt}
\title{Prompt User for Python Install}
\usage{
py_install_prompt(package = NULL, env = NULL)
}
\arguments{
\item{package}{python package/github url}

\item{env}{environment into which package will be installed}
}
\description{
prompts user to install a package
}
\keyword{internal}
