% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_help.R
\name{rank_binarize_wrapper}
\alias{rank_binarize_wrapper}
\title{rank_binarize_wrapper}
\usage{
rank_binarize_wrapper(expr_values, subset_feats = NULL, percentage_rank = 30)
}
\value{
matrix
}
\description{
wrapper for rank binarization function
}
\keyword{internal}
