% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_enrichment.R
\name{rankEnrich}
\alias{rankEnrich}
\title{rankEnrich}
\usage{
rankEnrich(
  gobject,
  sign_matrix,
  expression_values = c("normalized", "scaled", "custom"),
  reverse_log_scale = TRUE,
  logbase = 2,
  output_enrichment = c("original", "zscore")
)
}
\arguments{
\item{gobject}{Giotto object}

\item{sign_matrix}{Matrix of signature genes for each cell type / process}

\item{expression_values}{expression values to use}

\item{reverse_log_scale}{reverse expression values from log scale}

\item{logbase}{log base to use if reverse_log_scale = TRUE}

\item{output_enrichment}{how to return enrichment output}
}
\value{
data.table with enrichment results
}
\description{
Function to calculate gene signature enrichment scores per spatial position using a rank based approach.
}
\details{
First a new rank is calculated as R = (R1*R2)^(1/2), where R1 is the rank of
fold-change for each gene in each spot and R2 is the rank of each marker in each cell type.
The Rank-Biased Precision is then calculated as: RBP = (1 - 0.99) * (0.99)^(R - 1)
and the final enrichment score is then calculated as the sum of top 100 RBPs.
}
\examples{
    rankEnrich(gobject)
}
