% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_enrichment.R
\name{rankPermutation}
\alias{rankPermutation}
\title{rankPermutation}
\usage{
rankPermutation(sc_gene, n)
}
\description{
creates permutation for the rankEnrich test
}
\examples{
    rankPermutation()
}
