% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_genes.R
\name{rankSpatialCorGroups}
\alias{rankSpatialCorGroups}
\title{rankSpatialCorGroups}
\usage{
rankSpatialCorGroups(spatCorObject, use_clus_name = NULL, show_plot = TRUE)
}
\arguments{
\item{use_clus_name}{name of clusters to visualize (from clusterSpatialCorGenes())}

\item{show_plot}{show the accompanying scatter plot}
}
\value{
data.table with positive (within group) and negative (outside group) scores
}
\description{
Rank spatial correlated clusters according to correlation structure
}
\examples{
    rankSpatialCorGroups(gobject)
}
