% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{rbind-generic}
\alias{rbind-generic}
\alias{rbind2,giottoPolygon,giottoPolygon-method}
\title{Combine objects by rows (Giotto-related)}
\usage{
\S4method{rbind2}{giottoPolygon,giottoPolygon}(x, y, add_list_ID = TRUE, ...)
}
\arguments{
\item{x}{item 1 to rbind}

\item{y}{item 2 to rbind}

\item{add_list_ID}{whether to generate a list_ID column when giottoPolygons
to append have different names}

\item{\dots}{additional params to pass to methods}
}
\description{
row bind two objects
}
\section{Functions}{
\itemize{
\item \code{rbind2(x = giottoPolygon, y = giottoPolygon)}: Append giottoPolygon objects

}}
