% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_structures.R
\name{rbind2_giotto_polygon_hetero}
\alias{rbind2_giotto_polygon_hetero}
\title{Append giotto polygons of different names}
\usage{
rbind2_giotto_polygon_hetero(x, y, new_name, add_list_ID = TRUE)
}
\arguments{
\item{x}{\code{giottoPolygon} 1}

\item{y}{\code{giottoPolygon} 2}

\item{add_list_ID}{whether to include the name of the origin \code{giottoPolygons}
as a new 'list_ID' attribute}

\item{poly_names}{sorted polygon names to be used in the combined \code{giottoPolygon}
object}
}
\description{
Append two giotto polygons together of different names
Performed recursively through \code{rbind2} and \code{rbind}. Generates an
additional list_ID attribute based on the original names. The object name
also becomes a combination of both previous names
}
\keyword{internal}
