% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_structures.R
\name{rbind2_giotto_polygon_homo}
\alias{rbind2_giotto_polygon_homo}
\title{Append giotto polygons of the same name}
\usage{
rbind2_giotto_polygon_homo(x, y)
}
\arguments{
\item{x}{\code{giottoPolygon} 1}

\item{y}{\code{giottoPolygon} 2}
}
\description{
Append two giotto polygons together of the same name.
Performed recursively through \code{rbind2} and \code{rbind}
}
\keyword{internal}
