% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto.R
\name{read_cell_metadata}
\alias{read_cell_metadata}
\title{Read cell metadata}
\usage{
read_cell_metadata(gobject, metadata, provenance = NULL, verbose = TRUE)
}
\arguments{
\item{gobject}{giotto object}

\item{metadata}{nested list of cell metadata information}

\item{provenance}{provenance information (optional)}

\item{verbose}{be verbose}
}
\description{
read cell metadata from list
}
\keyword{internal}
