% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{read_data_folder}
\alias{read_data_folder}
\alias{read_merscope_folder}
\title{Read a structured folder of exported data}
\usage{
read_data_folder(
  spat_method = NULL,
  data_dir = NULL,
  dir_items,
  data_to_use,
  load_format = NULL,
  require_data_DT,
  cores = NA,
  verbose = TRUE
)

read_merscope_folder(merscope_dir, data_to_use, cores = NA, verbose = TRUE)
}
\arguments{
\item{spat_method}{spatial method for which the data is being read}

\item{data_dir}{exported data directory to read from}

\item{dir_items}{named list of directory items to expect and keywords to match}

\item{data_to_use}{which type(s) of expression data to build the gobject with}

\item{require_data_DT}{data.table detailing if expected data items are required
or optional for each \code{data_to_use} workflow}

\item{cores}{cores to use}

\item{verbose}{be verbose}
}
\description{
Read the exported folder of a spatial method and
detect the presence of needed files. NULL values denote missing items.
}
\details{
Steps performed:
\itemize{
  \item{1. detection of items within \code{data_dir} by looking for keywords
  assigned through \code{dir_items}}
  \item{2. check of detected items to see if everything needed has been found.
  Dictionary of necessary vs optional items for each \code{data_to_use} workflow
  is provided through \code{require_data_DT}}
  \item{3. if multiple filepaths are found to be matching then select the first
  one. This function is only intended to find the first level subdirectories
  and files.}
}
}
\section{Functions}{
\itemize{
\item \code{read_merscope_folder()}: Read a structured MERSCOPE folder

}}
\keyword{internal}
