% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd.R
\name{read_data_params}
\alias{read_data_params}
\title{read_data_params}
\arguments{
\item{data_list}{(nested) list of input data to read}

\item{default_spat_unit}{(optional) default spat_unit to use}

\item{default_feat_type}{(optional) default feat_type to use}

\item{provenance}{(optional) provenance information}

\item{verbose}{be verbose}
}
\description{
read_data_params
}
\keyword{internal}
