% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto.R
\name{read_dimension_reduction}
\alias{read_dimension_reduction}
\title{Read dimensional reduction}
\usage{
read_dimension_reduction(gobject, dimension_reduction)
}
\description{
read dimension reduction results from list
}
\keyword{internal}
