% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto.R
\name{read_expression_data}
\alias{read_expression_data}
\title{read_expression_data}
\usage{
read_expression_data(
  expr_list = NULL,
  sparse = TRUE,
  cores = NA,
  verbose = TRUE
)
}
\arguments{
\item{expr_list}{(nested) list with expression data}

\item{sparse}{read matrix data in a sparse manner}

\item{cores}{number of cores to use}

\item{verbose}{be verbose}
}
\description{
read_expression_data
}
\details{
mylistA = list('a' = matrix(1:5), 'b' = matrix(1:5))
depth(mylistA)

mylistB = list(A = list('a' = matrix(1:5), 'b' = matrix(1:5)),
               B = list('c' = matrix(1:5),'d' = matrix(1:5)))
depth(mylistB)

mylistC = list('RNA' = list('RAW' = list('cell' = matrix(1:5), 'nucleus' = matrix(6:10)),
                            'NORM' = list('cell' = matrix(11:15),'nucleus' = matrix(20:25))),
               'PROT' = list('RAW' = list('cell' = matrix(16:20))))
depth(mylistC)

mymatD = matrix(data = 1:4)
}
\keyword{internal}
