% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto.R
\name{read_feature_metadata}
\alias{read_feature_metadata}
\title{Read feature metadata}
\usage{
read_feature_metadata(gobject, metadata)
}
\arguments{
\item{gobject}{giotto object}

\item{metadata}{nested list of feature metadata information}
}
\description{
read feature metadata from list
}
\keyword{internal}
