% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto.R
\name{read_nearest_networks}
\alias{read_nearest_networks}
\title{Read nearest neighbor networks}
\usage{
read_nearest_networks(gobject, nn_network)
}
\description{
read nearest network results from list
}
\keyword{internal}
