% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto.R
\name{read_spatial_enrichment}
\alias{read_spatial_enrichment}
\title{Read spatial enrichment}
\usage{
read_spatial_enrichment(gobject, spatial_enrichment)
}
\description{
read spatial enrichment results from list
}
\keyword{internal}
