% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto.R
\name{read_spatial_location_data}
\alias{read_spatial_location_data}
\title{Read spatial location data}
\usage{
read_spatial_location_data(
  gobject,
  spat_loc_list,
  cores = determine_cores(),
  provenance = NULL,
  verbose = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_loc_list}{list of spatial locations}

\item{cores}{how many cores to use}

\item{provenance}{provenance information (optional)}

\item{verbose}{be verbose}
}
\value{
updated giotto object
}
\description{
read spatial locations
}
\keyword{internal}
