% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto.R
\name{read_spatial_networks}
\alias{read_spatial_networks}
\title{Read spatial networks}
\usage{
read_spatial_networks(spatial_network, provenance = NULL)
}
\description{
read spatial networks from list
}
\keyword{internal}
