% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto.R
\name{readDimReducData}
\alias{readDimReducData}
\title{Read dimensional reduction data}
\usage{
readDimReducData(
  data_list,
  default_spat_unit = NULL,
  default_feat_type = NULL,
  reduction = c("cells", "feats"),
  provenance = NULL,
  verbose = TRUE
)
}
\arguments{
\item{data_list}{(nested) list of dimension reduction input data}
}
\description{
read dimension reduction results from list
}
