% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto.R
\name{readExprData}
\alias{readExprData}
\title{Read expression data}
\usage{
readExprData(
  data_list,
  sparse = TRUE,
  cores = determine_cores(),
  default_feat_type = NULL,
  verbose = TRUE,
  provenance = NULL
)
}
\arguments{
\item{sparse}{(boolean, default = TRUE) read matrix data in a sparse manner}

\item{cores}{number of cores to use}

\item{default_feat_type}{default feature type to use}

\item{verbose}{be verbose}

\item{provenance}{provenance information}

\item{expr_list}{(nested) list of expression input data}
}
\description{
Read a nested list of expression data inputs in order to
generate a list of giotto-native exprObj
}
\details{
mylistA = list('a' = matrix(1:5), 'b' = matrix(1:5))
depth(mylistA)

mylistB = list(A = list('a' = matrix(1:5), 'b' = matrix(1:5)),
               B = list('c' = matrix(1:5),'d' = matrix(1:5)))
depth(mylistB)

mylistC = list('RNA' = list('RAW' = list('cell' = matrix(1:5), 'nucleus' = matrix(6:10)),
                            'NORM' = list('cell' = matrix(11:15),'nucleus' = matrix(20:25))),
               'PROT' = list('RAW' = list('cell' = matrix(16:20))))
depth(mylistC)

mymatD = matrix(data = 1:4)
}
