% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{readExprMatrix}
\alias{readExprMatrix}
\title{Read expression matrix}
\usage{
readExprMatrix(path, cores = determine_cores(), transpose = FALSE)
}
\arguments{
\item{path}{path to the expression matrix}

\item{cores}{number of cores to use}

\item{transpose}{transpose matrix}
}
\value{
sparse matrix
}
\description{
Function to read an expression matrix into a sparse matrix.
}
\details{
The expression matrix needs to have both unique column names and row names
}
\seealso{
Other functions to read expression matrices: 
\code{\link{readExprMatrixDB}()}
}
\concept{functions to read expression matrices}
