% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{readExprMatrixDB}
\alias{readExprMatrixDB}
\title{Read a flat file in as dbMatrix}
\usage{
readExprMatrixDB(
  path,
  remote_name = "test",
  remote_path = ":temp:",
  cores = determine_cores(),
  transpose = FALSE,
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{path}{path to the expression matrix}

\item{remote_name}{name of remote table in database}

\item{remote_path}{path to database. Default is tempdir()}

\item{cores}{number of cores to use when reading}

\item{transpose}{transpose matrix}

\item{overwrite}{whether to overwrite remote table}
}
\description{
Function to read an expression matrix into a database backend.
}
\seealso{
Other functions to read expression matrices: 
\code{\link{readExprMatrix}()}
}
\concept{functions to read expression matrices}
