% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{readFeatData}
\alias{readFeatData}
\title{Read feature information}
\usage{
readFeatData(data_list, verbose = TRUE)
}
\arguments{
\item{data_list}{(nested) list of input data to read}

\item{verbose}{be verbose}
}
\description{
Function to read lists of feature information data and output
a list of generated giottoPoints objects
}
