% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{readPolygonData}
\alias{readPolygonData}
\title{Read list of polygons information}
\usage{
readPolygonData(
  data_list,
  default_name = "cell",
  input = "guess",
  polygon_mask_list_params = NULL,
  polygon_dfr_list_params = NULL,
  calc_centroids = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{default_name}{default name to assign if \code{polygonlist} is not a
list. If \code{polygonlist} is an unnamed list then \code{default_name} will
be used as part of the template for generating indexed default names.}

\item{input}{what type of input is being used. When set to 'guess', uses
'mask' if \code{polygonlist} is of type character and 'table' when
\code{polygonlist} is dataframe-like}

\item{polygon_mask_list_params}{parameters for when using 'mask' workflow}

\item{polygon_dfr_list_params}{parameters for when using 'table' workflow}

\item{calc_centroids}{whether centroids should be calculated during polygon
creation}

\item{verbose}{be verbose}
}
\description{
Function extract list of polygons when given raw input as either
mask or tabular data. Calls the respective createGiottoPolygons functions. \cr
If a \code{giottoPolygon} object is passed then no edits will be made other
than updating the \code{name} slot if the list is named.
}
