% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_help.R
\name{readPolygonFilesVizgen}
\alias{readPolygonFilesVizgen}
\alias{.h5_read_vizgen}
\title{readPolygonFilesVizgen}
\usage{
readPolygonFilesVizgen(
  gobject,
  boundaries_path,
  fovs = NULL,
  polygon_feat_types = 0:6,
  flip_x_axis = FALSE,
  flip_y_axis = FALSE,
  smooth_polygons = TRUE,
  smooth_vertices = 60,
  set_neg_to_zero = FALSE,
  return_gobject = TRUE,
  verbose = TRUE
)

.h5_read_vizgen(
  h5File,
  z_indices = 1L:7L,
  segm_to_use = "p_0",
  H5Fopen_flags = "H5F_ACC_RDWR"
)
}
\arguments{
\item{gobject}{giotto object}

\item{boundaries_path}{path to the cell_boundaries folder}

\item{fovs}{selected fovs, if NULL select all fovs within Giotto object}

\item{polygon_feat_types}{a vector containing the polygon feature types}

\item{flip_x_axis}{flip x axis of polygon coordinates (multiply by -1)}

\item{flip_y_axis}{flip y axis of polygon coordinates (multiply by -1)}

\item{smooth_polygons}{smooth polygons (default = TRUE)}

\item{smooth_vertices}{number of vertices for smoothing}

\item{set_neg_to_zero}{set negative values to zero when smoothing}

\item{return_gobject}{return giotto object}

\item{verbose}{be verbose}
}
\value{
giotto object or cell polygons list
}
\description{
Read selected polygon files for the FOVs present in the Giotto
object and add the smoothed polygons to the object
}
\section{Functions}{
\itemize{
\item \code{.h5_read_vizgen()}: (internal) Optimized .hdf5 reading for
vizgen merscope output. Returns a data.table of xyz coords and cell_id

}}
\seealso{
\code{\link{smoothGiottoPolygons}}
}
\keyword{internal}
