% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_help.R
\name{readPolygonFilesVizgenHDF5}
\alias{readPolygonFilesVizgenHDF5}
\title{readPolygonFilesVizgenHDF5}
\usage{
readPolygonFilesVizgenHDF5(
  boundaries_path,
  fovs = NULL,
  polygon_feat_types = 0:6,
  flip_x_axis = F,
  flip_y_axis = F,
  smooth_polygons = TRUE,
  smooth_vertices = 60,
  set_neg_to_zero = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{boundaries_path}{path to the cell_boundaries folder}

\item{fovs}{subset of fovs to use}

\item{polygon_feat_types}{a vector containing the polygon feature types}

\item{flip_x_axis}{flip x axis of polygon coordinates (multiply by -1)}

\item{flip_y_axis}{flip y axis of polygon coordinates (multiply by -1)}

\item{smooth_polygons}{smooth polygons (default = TRUE)}

\item{smooth_vertices}{number of vertices for smoothing}

\item{verbose}{be verbose}
}
\description{
Read and create polygons for all cells, or for only selected FOVs.
}
