% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_help.R
\name{readPolygonVizgenParquet}
\alias{readPolygonVizgenParquet}
\title{Read MERSCOPE polygons from parquet}
\usage{
readPolygonVizgenParquet(
  file,
  z_index = "all",
  calc_centroids = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{file}{parquet file to load}

\item{z_index}{either 'all' or a numeric vector of z_indices to get polygons for}

\item{calc_centroids}{calculate centroids for the polygons (default = TRUE)}

\item{verbose}{be verbose}
}
\description{
Read Vizgen exported cell boundary parquet files as giottoPolyons. The z level
can be selected.
}
