% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{readSpatEnrichData}
\alias{readSpatEnrichData}
\title{Read spatial enrichment}
\usage{
readSpatEnrichData(
  data_list,
  default_spat_unit = NULL,
  default_feat_type = NULL,
  provenance = NULL,
  verbose = TRUE
)
}
\arguments{
\item{data_list}{(nested) list of spatial enrichment input data}

\item{default_spat_unit}{(optional) default spat_unit to use}

\item{default_feat_type}{(optional) default feat_type to use}

\item{provenance}{(optional) provenance information}

\item{verbose}{be verbose}
}
\description{
read spatial enrichment results from list
}
