% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto.R
\name{readSpatLocData}
\alias{readSpatLocData}
\title{Read spatial location data}
\usage{
readSpatLocData(
  spat_loc_list,
  cores = determine_cores(),
  provenance = NULL,
  verbose = TRUE
)
}
\arguments{
\item{spat_loc_list}{(nested) list of spatial locations}

\item{cores}{how many cores to use}

\item{provenance}{provenance information (optional)}

\item{verbose}{be verbose}
}
\value{
list of spatLocsObj
}
\description{
read spatial locations/coordinates from nested list and generate
list of Giotto spatLocsObj
}
