% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{readSpatLocsData}
\alias{readSpatLocsData}
\title{Read spatial location data}
\usage{
readSpatLocsData(
  data_list,
  default_spat_unit = NULL,
  provenance = NULL,
  cores = determine_cores(),
  verbose = TRUE
)
}
\arguments{
\item{data_list}{(nested) list of spatial locations input data}

\item{default_spat_unit}{(optional) default spat_unit to use}

\item{provenance}{(optional) provenance information}

\item{cores}{how many cores to use}

\item{verbose}{be verbose}
}
\value{
list of spatLocsObj
}
\description{
read spatial locations/coordinates from nested list and generate
list of Giotto spatLocsObj
}
