% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto.R
\name{readSpatNetData}
\alias{readSpatNetData}
\title{Read spatial networks}
\usage{
readSpatNetData(
  data_list,
  default_spat_unit = NULL,
  provenance = NULL,
  verbose = TRUE
)
}
\arguments{
\item{data_list}{(nested) list of spatial network input data}
}
\description{
read spatial networks data from list
}
