% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{reconnect_giottoImage_MG}
\alias{reconnect_giottoImage_MG}
\title{reconnect_giottoImage_MG}
\usage{
reconnect_giottoImage_MG(giottoImage, image_path)
}
\arguments{
\item{giottoImage}{giottoImage to reconnect}

\item{image_path}{path to source file of giottoImage}
}
\value{
reconnected giottoImage
}
\description{
reconnect giottoImage if image pointer is broken
}
\keyword{internal}
