% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{reconnect_giottoLargeImage}
\alias{reconnect_giottoLargeImage}
\title{reconnect_giottoLargeImage}
\usage{
reconnect_giottoLargeImage(giottoLargeImage, image_path)
}
\arguments{
\item{giottoLargeImage}{giottoLargeImage to reconnect}

\item{image_path}{path to source file of giottoLargeImage}
}
\value{
reconnected giottoLargeImage
}
\description{
reconnect giottoLargeImage if image pointer is broken
}
\keyword{internal}
