% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{reconnect_image_object}
\alias{reconnect_image_object}
\title{reconnect_image_object}
\usage{
reconnect_image_object(image_object, image_type, image_path)
}
\arguments{
\item{image_object}{giotto image object}

\item{image_type}{type of giotto image object}

\item{image_path}{path to image source to reconnect image object with}
}
\value{
reconnected image_object
}
\description{
Reconnect giotto image object with dead image pointer using a filepath
  to the original image source
}
\details{
This is a simple wrapper function for image object-specific reconnection
  functions and does not include other functionality to find the specific image
  objects in the giotto object. See \link{\code{reconnectGiottoImage}}.
}
\keyword{internal}
