% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{reconnectGiottoImage}
\alias{reconnectGiottoImage}
\title{reconnectGiottoImage}
\usage{
reconnectGiottoImage(
  gobject,
  image_name = NULL,
  largeImage_name = NULL,
  image_path = NULL,
  largeImage_path = NULL
)
}
\arguments{
\item{gobject}{giotto object}

\item{image_name}{names of images to reconnect}

\item{largeImage_name}{name of large images to reconnect}

\item{image_path}{named list of paths to images to reconnect to giottoImages}

\item{largeImage_path}{named list of paths to images to reconnect to giottoLargeImages}
}
\value{
a giotto object with updated image pointer
}
\description{
reconnect dead image pointers using filepaths. Inputs can either be given as both image name and filepath args or as a named list through filepath argument alone.
}
