% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{reconnectImage}
\alias{reconnectImage}
\title{reconnectImage}
\usage{
reconnectImage(gobject, image_name = NULL, largeImage_name = NULL, path = NULL)
}
\arguments{
\item{gobject}{giotto object}

\item{image_name}{name of image to be reconnected}

\item{largeImage_name}{name of largeImage to be reconnected}
}
\value{
a giotto object with updated image pointer
}
\description{
reconnect dead image pointers using filepaths
}
