% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_structures.R
\name{rectVertices}
\alias{rectVertices}
\title{Generate rectangular polygon vertices}
\usage{
rectVertices(dims)
}
\arguments{
\item{dims}{named vector in the style of c(x = \code{numeric}, y = \code{numeric})
that defines the width (x) and height (y) of the generated rectangle polygon.}
}
\value{
a data.table of rectangle vertices
}
\description{
Generates vertex coordinates for a rectangle with dimensions given
through \code{dims} param.
}
\seealso{
polyStamp circleVertices hexVertices
}
