% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.R
\name{reduceDims}
\alias{reduceDims}
\title{Run dimension reduction method}
\usage{
reduceDims(
  gobject,
  method = c("pca", "nmf", "umap", "tsne"),
  projection = FALSE,
  toplevel = 1L,
  ...
)
}
\arguments{
\item{gobject}{giotto object}

\item{method}{character. Dimension reduction method to use}

\item{projection}{logical. Whether to run in a projection manner
(faster, but is an approximation)}

\item{toplevel}{relative stackframe the call was made at. do not use.}

\item{\dots}{additional params to pass to specific functions}
}
\value{
`giotto` object with attached dimension reduction
}
\description{
Wrapper function for Giotto dimension reduction methods for easier coding.
}
\examples{
g <- GiottoData::loadGiottoMini("vis")
x <- reduceDims(g, "tsne", spat_unit = "cell")
x <- reduceDims(x, "umap", projection = TRUE)
x <- reduceDims(x, method = "nmf")
}
