% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registration.R
\name{reg_img_minmax_finder}
\alias{reg_img_minmax_finder}
\title{reg_img_minmax_finder}
\usage{
reg_img_minmax_finder(
  gobject_list,
  image_unreg = NULL,
  largeImage_unreg = NULL,
  scale_factor,
  transform_values,
  method
)
}
\arguments{
\item{gobject_list}{list of gobjects to use}

\item{image_unreg}{name of original unregistered images}

\item{largeImage_unreg}{name of original unregistered largeImages}

\item{scale_factor}{scale factors for registered images relative to spatlocs.}

\item{transform_values}{transformation values to use}

\item{method}{method of registration}
}
\description{
finds new minmax boundaries of registration transformed images
}
\keyword{internal}
