% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registration.R
\name{register_spatial_locations}
\alias{register_spatial_locations}
\title{register_spatial_locations}
\usage{
register_spatial_locations(xvals, yvals, scalefactor = 1, transformXML)
}
\arguments{
\item{xvals}{x value spatial input}

\item{yvals}{y value spatial input}

\item{scalefactor}{scale spatial coords to pixel coords}

\item{transformXML}{transformation files to use}
}
\description{
Performs appropriate transforms to align spatial locations with registered images.
}
\keyword{internal}
