% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registration.R
\name{registerGiottoObjectList}
\alias{registerGiottoObjectList}
\title{registerGiottoObjectList}
\usage{
registerGiottoObjectList(
  gobject_list,
  spat_unit = NULL,
  method = c("fiji", "rvision"),
  image_unreg = "image",
  image_reg_name = "image",
  image_list = NULL,
  save_dir = NULL,
  spatloc_unreg = "raw",
  spatloc_reg_name = "raw",
  fiji_xml_files,
  fiji_registered_images,
  scale_factor = NULL,
  allow_rvision_autoscale = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{gobject_list}{List of gobjects to register}

\item{spat_unit}{spatial unit}

\item{method}{Method used to align gobjects. Current options are either
using FIJI register_virtual_stack_slices output or rvision}

\item{image_unreg}{Gobject image slot to use. Defaults to 'image' (optional)}

\item{image_reg_name}{Arbitrary image slot name for registered images to
occupy. Defaults to replacement of 'image' slot (optional)}

\item{image_list}{RVISION - under construction}

\item{save_dir}{RVISION - under construction}

\item{spatloc_unreg}{Unregistered spatial locations to align. Defaults to
'raw' slot (optional)}

\item{spatloc_reg_name}{Arbitrary name for registered spatial locations.
Defaults to replacement of 'raw' slot (optional)}

\item{fiji_xml_files}{Filepaths to FIJI registration XML outputs}

\item{fiji_registered_images}{Registered images output by FIJI
register_virtual_stack_slices}

\item{scale_factor}{Scaling to be applied to spatial coordinates}

\item{allow_rvision_autoscale}{Whether or not to allow rvision to
automatically scale the images when performing image registration}

\item{verbose}{Be verbose}
}
\value{
List of registered giotto objects where the registered images and
spatial locations
}
\description{
Wrapper function for registerGiottoObjectListFiji and
registerGiottoObjectListRvision
}
