% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registration.R
\name{registerGiottoObjectListFiji}
\alias{registerGiottoObjectListFiji}
\title{registerGiottoObjectListFiji}
\usage{
registerGiottoObjectListFiji(
  gobject_list,
  image = "image",
  registered_image_name = "image",
  spat_loc_values = NULL,
  spat_loc_name = "raw",
  xml_files,
  registered_images = NULL,
  scaling = 1,
  auto_comp_reg_border,
  verbose = TRUE
)
}
\arguments{
\item{gobject_list}{list of gobjects to register}

\item{image}{slot of original unregistered images}

\item{spat_loc_values}{spatial locations to use}

\item{spat_loc_name}{slot to save spatial locations to. Defaults to replacement of spat_loc_values (optional)}

\item{xml_files}{record of transformations performed in xml format exported during image registration using FIJI register_virtual_stack_slices}

\item{registered_images}{registered images output by FIJI register_virtual_stack_slices}

\item{scaling}{scaling to be applied to spatial coordinates}

\item{verbose}{be verbose}
}
\value{
list of registered giotto objects where the registered images and spatial locations
}
\description{
Function to spatially align gobject data based on FIJI image registration.
}
