% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registration.R
\name{registerGiottoObjectListRvision}
\alias{registerGiottoObjectListRvision}
\title{registerGiottoObjectListRvision}
\usage{
registerGiottoObjectListRvision(
  gobject_list = gobject_list,
  image_list = NULL,
  save_dir = NULL,
  spatloc_unreg = NULL,
  spatloc_reg_name = "raw",
  verbose = TRUE
)
}
\arguments{
\item{gobject_list}{list of gobjects to register}

\item{image_list}{Filepaths to unregistered images}

\item{save_dir}{(Optional) If given, save registered images to this directory}

\item{spatloc_unreg}{spatial locations to use}

\item{spatloc_reg_name}{name for registered spatial locations to. Defaults
to replacement of spat_unreg (optional)}

\item{verbose}{be verbose}
}
\value{
list of registered giotto objects where the registered images and
spatial locations
}
\description{
Function to spatially align gobject data based on Rvision image
registration.
}
