% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registration.R
\name{registerImagesFIJI}
\alias{registerImagesFIJI}
\title{registerImagesFIJI}
\usage{
registerImagesFIJI(
  source_img_dir,
  output_img_dir,
  transforms_save_dir,
  ref_img_name,
  init_gauss_blur = 1.6,
  steps_per_scale_octave = 3,
  min_img_size = 64,
  max_img_size = 1024,
  feat_desc_size = 8,
  feat_desc_orient_bins = 8,
  closest_next_closest_ratio = 0.92,
  max_align_err = 25,
  inlier_ratio = 0.05,
  headless = FALSE,
  batch = TRUE,
  MinMem = MaxMem,
  MaxMem = 2500,
  IncrementalGC = TRUE,
  Threads = NULL,
  fijiArgs = NULL,
  javaArgs = NULL,
  ijArgs = NULL,
  jython = FALSE,
  fijiPath = fiji(),
  DryRun = FALSE
)
}
\arguments{
\item{source_img_dir}{Folder containing images to be registered}

\item{output_img_dir}{Folder to save registered images to}

\item{transforms_save_dir}{(jython implementation only) Folder to save 
transforms to}

\item{ref_img_name}{(jython implementation only) File name of reference 
image for the registration}

\item{init_gauss_blur}{Point detector option: initial image blurring}

\item{steps_per_scale_octave}{Point detector option}

\item{min_img_size}{Point detector option}

\item{max_img_size}{Point detector option}

\item{feat_desc_size}{Feature descriptor option}

\item{feat_desc_orient_bins}{Feature descriptor option}

\item{closest_next_closest_ratio}{Feature descriptor option}

\item{max_align_err}{Geometric consensus filter option}

\item{inlier_ratio}{Geometric consensus filter option}

\item{headless}{Whether to have ImageJ/Fiji running headless #TODO}

\item{batch}{Use batch mode #TODO}

\item{MinMem, MaxMem}{Memory limits}

\item{IncrementalGC}{Whether to use incremental garbage collection}

\item{Threads}{Number of threads}

\item{fijiArgs}{Arguments for ImageJ/FIJI}

\item{javaArgs}{Arguments for Java}

\item{ijArgs}{Arguments for ImageJ}

\item{jython}{Use jython wrapper script}

\item{fijiPath}{Path to fiji executable (can be set by
\code{options(giotto.fiji="/some/path")})}

\item{DryRun}{Whether to return the command to be run rather than actually
executing it.}
}
\value{
list of registered giotto objects where the registered images and 
spatial locations
}
\description{
Wrapper function for Register Virtual Stack Slices plugin in 
FIJI
}
\details{
This function was adapted from runFijiMacro function in 
jimpipeline by jefferislab
}
