% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{removeCellAnnotation}
\alias{removeCellAnnotation}
\title{removeCellAnnotation}
\usage{
removeCellAnnotation(gobject, columns = NULL, return_gobject = TRUE)
}
\arguments{
\item{gobject}{giotto object}

\item{columns}{names of columns to remove}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}
}
\value{
giotto object
}
\description{
removes cell annotation of giotto object
}
\details{
if return_gobject = FALSE, it will return the cell metadata
}
\examples{
    removeCellAnnotation(gobject)
}
