% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{removeFeatAnnotation}
\alias{removeFeatAnnotation}
\title{removeFeatAnnotation}
\usage{
removeFeatAnnotation(
  gobject,
  feat_type = NULL,
  columns = NULL,
  return_gobject = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{feat_type}{feature type}

\item{columns}{names of columns to remove}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}
}
\value{
giotto object
}
\description{
removes feature annotation from a Giotto object for a specific feature modality
}
\details{
if return_gobject = FALSE, it will return the gene metadata
}
