% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{removeGeneAnnotation}
\alias{removeGeneAnnotation}
\title{removeGeneAnnotation}
\usage{
removeGeneAnnotation(gobject, columns = NULL, return_gobject = TRUE)
}
\arguments{
\item{gobject}{giotto object}

\item{columns}{names of columns to remove}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}
}
\value{
giotto object
}
\description{
removes gene annotation of giotto object
}
\examples{
    removeGeneAnnotation(gobject)
}
