% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_environment.R
\name{removeGiottoEnvironment}
\alias{removeGiottoEnvironment}
\title{removeGiottoEnvironment}
\usage{
removeGiottoEnvironment()
}
\description{
removeGiottoEnvironment
}
\details{
Removes a previously installed giotto environment.
See \code{\link{installGiottoEnvironment}}.
}
