% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_structures.R
\name{rescale_polygons}
\alias{rescale_polygons}
\title{rescale polygons}
\usage{
rescale_polygons(spatVector, spatVectorCentroids, fx = 0.5, fy = 0.5)
}
\description{
rescale individual polygons by a factor x and y
}
\keyword{internal}
