% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{rescaleGiottoImage}
\alias{rescaleGiottoImage}
\title{rescaleGiottoImage}
\usage{
rescaleGiottoImage(
  gimage = NULL,
  gobject = NULL,
  image_name,
  spatloc_name = "raw",
  scale_factor = NULL,
  resolution = NULL
)
}
\arguments{
\item{gimage}{giottoImage to rescale}

\item{gobject}{giottoObject containing giottoImage to rescale}

\item{image_name}{name of giottoImage in giottoObject}

\item{spatloc_name}{name of spatlocs that scaling is relative to (defaults to 'raw')}

\item{scale_factor}{scale factor to convert from spatial distance to pixel distance}

\item{resolution}{pixel distance per unit of real world distance (1/scale_factor)}
}
\description{
Change scale_factor and resolution slot values
}
