% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_structures.R
\name{rescalePolygons}
\alias{rescalePolygons}
\title{rescalePolygons}
\usage{
rescalePolygons(
  gobject,
  poly_info = "cell",
  name = "rescaled_cell",
  fx = 0.5,
  fy = 0.5,
  calculate_centroids = TRUE,
  return_gobject = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{poly_info}{polygon information name}

\item{name}{name of new polygon layer}

\item{fx}{x-scaling factor}

\item{fy}{y-scaling factor}

\item{calculate_centroids}{calculate centroids}

\item{return_gobject}{return giotto object}
}
\value{
giotto object
}
\description{
Rescale individual polygons by a x and y factor
}
\concept{polygon scaling}
