% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_section.R
\name{reshape_to_data_point}
\alias{reshape_to_data_point}
\title{reshape_to_data_point}
\usage{
reshape_to_data_point(mesh_grid_obj)
}
\arguments{
\item{mesh_grid_obj}{mesh_grid_obj}
}
\description{
reshape a mesh grid line object to data point matrix
}
\keyword{internal}
