% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_section.R
\name{reshape_to_mesh_grid_obj}
\alias{reshape_to_mesh_grid_obj}
\title{reshape_to_mesh_grid_obj}
\usage{
reshape_to_mesh_grid_obj(data_points, mesh_grid_n)
}
\arguments{
\item{data_points}{data_points}

\item{mesh_grid_n}{mesh_grid_n}
}
\description{
reshape a data point matrix to a mesh grid line object
}
\keyword{internal}
