% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_environment.R
\name{return_giotto_environment_path}
\alias{return_giotto_environment_path}
\title{return_giotto_environment_path}
\usage{
return_giotto_environment_path()
}
\value{
string path
}
\description{
returns the path to the detected giotto environment
}
\keyword{internal}
