% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_environment.R
\name{return_giotto_environment_path_executable}
\alias{return_giotto_environment_path_executable}
\title{return_giotto_environment_path_executable}
\usage{
return_giotto_environment_path_executable()
}
\value{
string path
}
\description{
returns the path to the detected giotto environment executable
}
\keyword{internal}
