% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registration.R
\name{rigid_transform_spatial_locations}
\alias{rigid_transform_spatial_locations}
\title{rigid_transform_spatial_locations}
\usage{
rigid_transform_spatial_locations(spatlocs, transform_values, method)
}
\arguments{
\item{spatlocs}{input spatial locations}

\item{transform_values}{transformation values to use}

\item{method}{which method is used for image registration}
}
\description{
Performs appropriate transforms to align spatial locations with registered images.
}
\keyword{internal}
