% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{rna_osmfish_normalization}
\alias{rna_osmfish_normalization}
\title{RNA osmfish normalization}
\usage{
rna_osmfish_normalization(
  gobject,
  raw_expr,
  feat_type,
  spat_unit,
  name = "custom",
  verbose = TRUE
)
}
\description{
function for RNA normalization according to osmFISH paper
}
\keyword{internal}
