% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_manipulation.R
\name{rotate_spatial_locations}
\alias{rotate_spatial_locations}
\title{Rotate spatial locations}
\usage{
rotate_spatial_locations(
  spatlocs,
  rotateradians = c(xy = 0, zy = 0, xz = 0),
  rcenter = c(x = 0, y = 0, z = 0)
)
}
\arguments{
\item{spatlocs}{spatial locations to use}

\item{rotateradians}{Named vector of radians for rotation along each of the 3 coordinate
axes. If only a single value is provided, it will be treated as xy rotation.}

\item{rcenter}{center of rotation given as vector xy(z) coordinates (defaults to coordinate center)}
}
\description{
Rotate given spatlocs by given radians
}
\details{
Radians are provided through \code{rotateradians} param as a named vector
with values for \code{xy} (yaw), \code{zy} (pitch), \code{xz} (roll)
}
\keyword{internal}
