% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{round_seconds}
\alias{round_seconds}
\title{Round seconds to other periods of time}
\usage{
round_seconds(sec, output = c("list", "char"))
}
\description{
Round seconds to other periods of time
}
\keyword{internal}
