% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{row-plus-colnames-generic}
\alias{row-plus-colnames-generic}
\alias{colnames}
\alias{rownames}
\alias{colnames,exprObj-method}
\alias{rownames,exprObj-method}
\title{Row and column names}
\usage{
\S4method{colnames}{exprObj}(x)

\S4method{rownames}{exprObj}(x)
}
\arguments{
\item{x}{object}
}
\value{
A character vector of row or col names
}
\description{
Retrieve or set the row or column names of an object
}
