% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{rowMeans_giotto}
\alias{rowMeans_giotto}
\title{rowMeans_giotto}
\usage{
rowMeans_giotto(mymatrix)
}
\arguments{
\item{mymatrix}{matrix object}
}
\value{
numeric vector
}
\description{
rowMeans function that works with multiple matrix representations
}
