% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{rowSums_giotto}
\alias{rowSums_giotto}
\title{rowSums_giotto}
\usage{
rowSums_giotto(mymatrix)
}
\arguments{
\item{mymatrix}{matrix object}
}
\value{
numeric vector
}
\description{
rowSums function that works with multiple matrix representations
}
