% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_enrichment.R
\name{run_spat_autocor_global}
\alias{run_spat_autocor_global}
\title{run_spat_autocor_global}
\usage{
run_spat_autocor_global(
  use_values,
  feats,
  weight_matrix,
  method,
  test_method,
  mc_nsim,
  cor_name
)
}
\description{
run_spat_autocor_global
}
\keyword{internal}
