% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_genes.R
\name{run_spatial_sim_tests_multi}
\alias{run_spatial_sim_tests_multi}
\title{run_spatial_sim_tests_multi}
\usage{
run_spatial_sim_tests_multi(
  gobject,
  pattern_name = "pattern",
  pattern_cell_ids = NULL,
  gene_name = NULL,
  spatial_probs = c(0.5, 1),
  reps = 2,
  spatial_network_name = "kNN_network",
  binSpect_km_param = list(nstart = 3, iter_max = 10, expression_values = "normalized",
    get_av_expr = FALSE, get_high_expr = FALSE),
  binSpect_rnk_param = list(percentage_rank = 30, expression_values = "normalized",
    get_av_expr = FALSE, get_high_expr = FALSE),
  spatialDE_param = list(expression_values = "raw", sig_alpha = 0.5, unsig_alpha = 0.5),
  spark_param = list(values_type = "raw", percentage = 0.1, min_count = 10, num_core =
    5),
  save_plot = F,
  save_dir = "~",
  ...
)
}
\description{
runs all spatial tests for multiple probabilities and repetitions
}
\keyword{internal}
