% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registration.R
\name{fiji}
\alias{fiji}
\title{fiji}
\usage{
fiji(fijiPath = NULL)
}
\arguments{
\item{fijiPath}{manually set filepath to Fiji executable}
}
\description{
\code{fiji} returns path to preferred Fiji executable
}
\examples{
# Path to current Fiji executable
\donttest{
fiji()
}

\dontrun{
# This function was taken and modified from jimpipeline by jefferislab #
# Set path to preferred Fiji executable (this will be remembered)
# you can also set options(giotto.fiji="/some/path")
fiji("/Applications/Fiji.app/Contents/MacOS/ImageJ-macosx")
}
}
