% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnn.R
\name{runIntegratedUMAP}
\alias{runIntegratedUMAP}
\title{Run integrated UMAP}
\usage{
runIntegratedUMAP(
  gobject,
  spat_unit = "cell",
  feat_types = c("rna", "protein"),
  integrated_feat_type = NULL,
  integration_method = "WNN",
  matrix_result_name = "theta_weighted_matrix",
  k = 20,
  spread = 5,
  min_dist = 0.01,
  force = FALSE,
  seed = 1234,
  ...
)
}
\arguments{
\item{gobject}{A giotto object}

\item{spat_unit}{spatial unit}

\item{feat_types}{feature types to integrate. Default = c("rna", "protein")}

\item{integrated_feat_type}{integrated feature type (e.g. 'rna_protein')}

\item{integration_method}{multiomics integration method used. Default = 'WNN'}

\item{matrix_result_name}{Default = 'theta_weighted_matrix'}

\item{k}{k number}

\item{spread}{UMAP param: spread}

\item{min_dist}{UMAP param: min_dist}

\item{force}{force calculation of integrated kNN. Default = FALSE}

\item{seed}{seed. Default 1234}

\item{...}{additional UMAP parameters}
}
\value{
A Giotto object with integrated UMAP
}
\description{
Run integrated UMAP
}
