% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.R
\name{runIterativeLSI}
\alias{runIterativeLSI}
\title{Run Iterative Latent Semantic Indexing (LSI)}
\usage{
runIterativeLSI(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  expression_values = "raw",
  name = NULL,
  feats_to_use = NULL,
  return_gobject = TRUE,
  iterations = 2,
  sample_cells_pre = 10000,
  sample_cells_final = NULL,
  var_features = 25000,
  dims = 30,
  scale_to = 10000,
  lsi_method = 2,
  resolution = 2,
  first_selection = "top",
  k = 10,
  seed = 1234,
  set_seed = TRUE,
  verbose = NULL
)
}
\arguments{
\item{gobject}{Giotto object}

\item{spat_unit}{Spatial unit (e.g., "cell")}

\item{feat_type}{Feature type (e.g., "spatial_atac")}

\item{expression_values}{Values to use (e.g., "raw")}

\item{name}{Name for reduction (default: feat_type + ".iterative.lsi")}

\item{feats_to_use}{Initial feature subset (NULL = all)}

\item{return_gobject}{Return Giotto object (default = TRUE)}

\item{iterations}{Number of iterations (default = 2)}

\item{sample_cells_pre}{Cells to subsample in early iterations}

\item{sample_cells_final}{Cells to subsample in final iteration (NULL = all)}

\item{var_features}{Number of variable features (default = 25000)}

\item{dims}{LSI dimensions to retain (default = seq_len(30))}

\item{scale_to}{TF-IDF scaling factor for sub_method 2 (default = 10000)}

\item{lsi_method}{TF-IDF method: 1 = Casanovich, 2 = Stuart, 3 = ArchR, 
4 = Giotto (default = 3). See [Giotto::norm_tfidf] for more information on 
how each of these are calculated.}

\item{resolution}{Clustering resolution (default = 2)}

\item{first_selection}{"top" (accessibility) or "var" (variance) 
(default = "var")}

\item{k}{Number of nearest neighbors for sNN (default = 10)}

\item{seed}{Seed for reproducibility (default = 1234)}

\item{set_seed}{Logical, whether to set a seed (default = TRUE)}

\item{verbose}{Verbosity (default = NULL, inherits from gobject)}
}
\value{
Giotto object or LSI coordinates
}
\description{
Performs Iterative LSI on a Giotto object, replicating ArchR's 
approach using Giotto functions.
}
