% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_enrichment.R
\name{runPAGEEnrich_OLD}
\alias{runPAGEEnrich_OLD}
\title{runPAGEEnrich_OLD}
\usage{
runPAGEEnrich_OLD(
  gobject,
  sign_matrix,
  expression_values = c("normalized", "scaled", "custom"),
  reverse_log_scale = TRUE,
  logbase = 2,
  output_enrichment = c("original", "zscore"),
  p_value = FALSE,
  n_times = 1000,
  name = NULL,
  return_gobject = TRUE
)
}
\arguments{
\item{gobject}{Giotto object}

\item{sign_matrix}{Matrix of signature genes for each cell type / process}

\item{expression_values}{expression values to use}

\item{reverse_log_scale}{reverse expression values from log scale}

\item{logbase}{log base to use if reverse_log_scale = TRUE}

\item{output_enrichment}{how to return enrichment output}

\item{p_value}{calculate p-values (boolean, default = FALSE)}

\item{n_times}{number of permutations to calculate for p_value}

\item{name}{to give to spatial enrichment results, default = PAGE}

\item{return_gobject}{return giotto object}
}
\value{
data.table with enrichment results
}
\description{
Function to calculate gene signature enrichment scores per spatial position using PAGE.
}
\details{
sign_matrix: a binary matrix with genes as row names and cell-types as column names.
Alternatively a list of signature genes can be provided to makeSignMatrixPAGE, which will create
the matrix for you. \cr

The enrichment Z score is calculated by using method (PAGE) from
Kim SY et al., BMC bioinformatics, 2005 as \eqn{Z = ((Sm – mu)*m^(1/2)) / delta}.
For each gene in each spot, mu is the fold change values versus the mean expression
and delta is the standard deviation. Sm is the mean fold change value of a specific marker gene set
and  m is the size of a given marker gene set.
}
\seealso{
\code{\link{makeSignMatrixPAGE}}
}
