% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_genes.R
\name{runPatternSimulation}
\alias{runPatternSimulation}
\title{runPatternSimulation}
\usage{
runPatternSimulation(
  gobject,
  pattern_name = "pattern",
  pattern_cell_ids = NULL,
  gene_names = NULL,
  spatial_probs = c(0.5, 1),
  reps = 2,
  save_plot = T,
  save_dir = "~",
  max_col = 4,
  height = 7,
  width = 7,
  ...
)
}
\arguments{
\item{gobject}{giotto object}

\item{pattern_name}{name of spatial pattern}

\item{pattern_cell_ids}{cell ids that make up the spatial pattern}

\item{gene_names}{selected genes}

\item{spatial_probs}{probabilities to test for a high expressing gene value to be part of the spatial pattern}

\item{reps}{number of random simulation repititions}

\item{save_plot}{save intermediate random simulation plots or not}

\item{save_dir}{directory to save results to}

\item{max_col}{maximum number of columns for final plots}

\item{height}{height of final plots}

\item{width}{width of final plots}

\item{\dots}{additional parameters for spatial gene detection tests}
}
\value{
data.table with results
}
\description{
Creates a known spatial pattern for selected genes one-by-one and runs the different spatial gene detection tests
}
\examples{
    runPatternSimulation(gobject)
}
