% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.R
\name{runPCAprojection}
\alias{runPCAprojection}
\title{runPCAprojection}
\usage{
runPCAprojection(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  expression_values = c("normalized", "scaled", "custom"),
  reduction = c("cells", "feats"),
  random_subset = 500,
  name = "pca.projection",
  feats_to_use = "hvf",
  return_gobject = TRUE,
  center = TRUE,
  scale_unit = TRUE,
  ncp = 100,
  method = c("irlba"),
  method_params = BiocParallel::SerialParam(),
  rev = FALSE,
  set_seed = TRUE,
  seed_number = 1234,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit}

\item{feat_type}{feature type}

\item{expression_values}{expression values to use}

\item{reduction}{cells or genes}

\item{random_subset}{random subset to perform PCA on}

\item{name}{arbitrary name for PCA run}

\item{feats_to_use}{subset of features to use for PCA}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}

\item{center}{center data first (default = TRUE)}

\item{scale_unit}{scale features before PCA (default = TRUE)}

\item{ncp}{number of principal components to calculate}

\item{method}{which implementation to use}

\item{method_params}{BiocParallelParam object}

\item{rev}{do a reverse PCA}

\item{set_seed}{use of seed}

\item{seed_number}{seed number to use}

\item{verbose}{verbosity of the function}

\item{...}{additional parameters for PCA (see details)}
}
\value{
giotto object with updated PCA dimension recuction
}
\description{
runs a Principal Component Analysis on a random subet + projection
}
\details{
See \code{\link[BiocSingular]{runPCA}} and \code{\link[FactoMineR]{PCA}} for more information about other parameters.
This PCA implementation is similar to  \code{\link{runPCA}}, except that it
performs PCA on a subset of the cells or features, and predict on the others.
This can significantly increase speed without sacrificing accuracy too much.
\itemize{
  \item feats_to_use = NULL: will use all features from the selected matrix
  \item feats_to_use = <hvg name>: can be used to select a column name of
  highly variable features, created by (see \code{\link{calculateHVF}})
  \item feats_to_use = c('geneA', 'geneB', ...): will use all manually provided features
}
}
