% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.R
\name{runUMAPprojection}
\alias{runUMAPprojection}
\title{Run UMAP dimension reduction}
\usage{
runUMAPprojection(
  gobject,
  feat_type = NULL,
  spat_unit = NULL,
  expression_values = c("normalized", "scaled", "custom"),
  reduction = c("cells", "feats"),
  dim_reduction_to_use = "pca",
  dim_reduction_name = NULL,
  dimensions_to_use = 1:10,
  random_subset = 500,
  name = NULL,
  feats_to_use = NULL,
  return_gobject = TRUE,
  n_neighbors = 40,
  n_components = 2,
  n_epochs = 400,
  min_dist = 0.01,
  n_threads = NA,
  spread = 5,
  set_seed = TRUE,
  seed_number = 1234,
  verbose = T,
  toplevel_params = 2,
  ...
)
}
\arguments{
\item{gobject}{giotto object}

\item{feat_type}{feature type}

\item{spat_unit}{spatial unit}

\item{expression_values}{expression values to use}

\item{reduction}{'cells' or 'feats'}

\item{dim_reduction_to_use}{use another dimension reduction set as input}

\item{dim_reduction_name}{name of dimension reduction set to use}

\item{dimensions_to_use}{number of dimensions to use as input}

\item{random_subset}{random subset to perform UMAP on}

\item{name}{arbitrary name for UMAP run}

\item{feats_to_use}{if dim_reduction_to_use = NULL, which features to use}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}

\item{n_neighbors}{UMAP param: number of neighbors}

\item{n_components}{UMAP param: number of components}

\item{n_epochs}{UMAP param: number of epochs}

\item{min_dist}{UMAP param: minimum distance}

\item{n_threads}{UMAP param: threads/cores to use}

\item{spread}{UMAP param: spread}

\item{set_seed}{use of seed}

\item{seed_number}{seed number to use}

\item{verbose}{verbosity of function}

\item{toplevel_params}{parameters to extract}

\item{...}{additional UMAP parameters}
}
\value{
giotto object with updated UMAP dimension reduction
}
\description{
run UMAP on subset and project on the rest
}
\details{
See \code{\link[uwot]{umap}} for more information about these and other parameters.
\itemize{
  \item Input for UMAP dimension reduction can be another dimension reduction (default = 'pca')
  \item To use gene expression as input set dim_reduction_to_use = NULL
  \item If dim_reduction_to_use = NULL, feats_to_use can be used to select a column name of
  highly variable genes (see \code{\link{calculateHVF}}) or simply provide a vector of genes
  \item multiple UMAP results can be stored by changing the \emph{name} of the analysis
}
}
