% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{S3toS4spatGridObj}
\alias{S3toS4spatGridObj}
\alias{S3toS4spatialGridObj}
\title{Spatially Binned Data}
\usage{
S3toS4spatialGridObj(object)
}
\arguments{
\item{object}{S3 spatialGridObj}
}
\description{
convert S3 spatialGridObj to S4
}
\keyword{internal}
