% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{S3toS4spatNetObj}
\alias{S3toS4spatNetObj}
\title{Spatial Networks}
\usage{
S3toS4spatNetObj(object, spat_unit = NULL)
}
\arguments{
\item{object}{S3 spatNetworkObj}

\item{spat_unit}{spatial unit metadata to append}
}
\description{
convert S3 spatialNetworkObj to S4
}
\keyword{internal}
