% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_help.R
\name{saveGiotto}
\alias{saveGiotto}
\title{saveGiotto}
\usage{
saveGiotto(
  gobject,
  foldername = "saveGiottoDir",
  dir = getwd(),
  method = c("RDS", "qs"),
  method_params = list(),
  overwrite = FALSE,
  image_filetype = "PNG",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{gobject}{Giotto object}

\item{foldername}{Folder name}

\item{dir}{Directory where to create the folder}

\item{method}{method to save main object}

\item{method_params}{additional method parameters for RDS or qs}

\item{overwrite}{Overwrite existing folders}

\item{image_filetype}{the image filetype to use, see \code{\link[terra]{writeRaster}}}

\item{verbose}{be verbose}

\item{...}{additional parameters for \code{\link[terra]{writeRaster}}}
}
\value{
Creates a directory with Giotto object information
}
\description{
Saves a Giotto object to a specific folder structure
}
\details{
Works together with \code{\link{loadGiotto}} to save and re-load
Giotto objects. Additional method_params need to be provided as a list and will
go to \code{\link[base]{saveRDS}} or \code{\link[qs]{qsave}}
}
