% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{scale_default}
\alias{scale_default}
\title{Default Giotto Scaling}
\value{
scaled object
}
\description{
2 step \link[=scale_zscore]{z-scoring} along features and samples
}
\section{params}{


\tabular{ll}{
\code{scale_feats} \tab logical (default = \code{TRUE}) Whether to scale across
features \cr
\code{scale_cells} \tab logical (default = \code{TRUE}) Whether to scale across
cells/samples \cr
\code{scale_order} \tab character. One of either \code{"first_feats"} or
\code{"first_cells"}. When both \code{scale_feats} and \code{scale_cells} are \code{TRUE},
determines the order in which the 2 scaling operations are performed. \cr
\code{verbose} \tab logical (default = \code{TRUE}) Whether to be verbose
}
}

\seealso{
\link{process_param}

Other scaling parameters: 
\code{\link{scale_zscore}}
}
\concept{scaling parameters}
