% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto.R
\name{scale_spatial_locations}
\alias{scale_spatial_locations}
\title{Scale spatial locations}
\usage{
scale_spatial_locations(
  spatlocs,
  scale_factor = c(x = 1, y = 1, z = 1),
  scenter = c(x = 0, y = 0, z = 0)
)
}
\arguments{
\item{spatlocs}{spatial locations information to scale}

\item{scale_factor}{scaling factor to apply to coordinates.}

\item{scenter}{center from which to scale spatial coordinates. Given as vector
of xy(z) coordinates.}
}
\description{
Simple scaling of spatial locations by given \code{scale_factor}.
Values will be scaled from the coordinate origin or coordinates provided through
\code{scenter} param.
}
\details{
\code{scale_factor} either given as a single value where it will be applied to
x, y, and z (if available) dimensions or as a vector of named values for 'x',
'y', (and 'z').
}
\keyword{internal}
