% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{scale_zscore}
\alias{scale_zscore}
\title{Z Score Scaling}
\value{
scaled object
}
\description{
Wrapper around \code{base::scale()} to make it compatible with the
\code{\link[=processData]{processData()}} framework. Additionally provides a \code{MARGIN} param.

\deqn{\LARGE
z_{i,j} = \frac{x_{i,j} - \mu_i}{\sigma_i}
}

Where:
\itemize{
\item \eqn{x_{i,j}} is the original value for feature \eqn{i} in sample \eqn{j}
\item \eqn{\mu_i} is the mean of feature \eqn{i} across all samples
\item \eqn{\sigma_i} is the standard deviation of feature \eqn{i} across all
samples
\item \eqn{z_{i,j}} is the resulting scaled value
}
}
\section{params}{


\tabular{ll}{
\code{scale} \tab logical (default = \code{TRUE}) Whether to scale values \cr
\code{center} \tab logical (default = \code{TRUE}) Whether to center values\cr
\code{MARGIN} \tab numeric. Either 1 (rows) or 2 (cols). Direction along which
to perform the operation.
}
}

\seealso{
\link{process_param}

Other scaling parameters: 
\code{\link{scale_default}}
}
\concept{scaling parameters}
