% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.R
\name{screePlot}
\alias{screePlot}
\title{screePlot}
\usage{
screePlot(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  dim_reduction_name = NULL,
  name = deprecated(),
  expression_values = c("normalized", "scaled", "custom"),
  reduction = c("cells", "feats"),
  method = c("irlba", "exact", "random", "factominer"),
  rev = FALSE,
  feats_to_use = NULL,
  center = FALSE,
  scale_unit = FALSE,
  ncp = 100,
  ylim = c(0, 20),
  verbose = TRUE,
  show_plot = NULL,
  return_plot = NULL,
  save_plot = NULL,
  save_param = list(),
  default_save_name = "screePlot",
  ...
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit (e.g. "cell")}

\item{feat_type}{feature type (e.g. "rna", "dna", "protein")}

\item{dim_reduction_name}{name of PCA}

\item{name}{deprecated}

\item{expression_values}{expression values to use}

\item{reduction}{cells or features}

\item{method}{which implementation to use}

\item{rev}{do a reverse PCA}

\item{feats_to_use}{subset of features to use for PCA}

\item{center}{center data before PCA}

\item{scale_unit}{scale features before PCA}

\item{ncp}{numeric. max number of principal components to plot}

\item{ylim}{numeric. y-axis limits on scree plot}

\item{verbose}{be verbose}

\item{show_plot}{logical. show plot}

\item{return_plot}{logical. return ggplot object}

\item{save_plot}{logical. save the plot}

\item{save_param}{list of saving parameters, see 
\code{\link{showSaveParameters}}}

\item{default_save_name}{default save name for saving, don't change, 
change save_name in save_param}

\item{...}{additional arguments to pca function, see \code{\link{runPCA}}}
}
\value{
ggplot object for scree method
}
\description{
identify significant principal components (PCs) using an
screeplot (a.k.a. elbowplot)
}
\details{
Screeplot works by plotting the explained variance of each
 individual PC in a barplot allowing you to identify which PC provides a
 significant contribution (a.k.a 'elbow method'). \cr
 Screeplot will use an available pca object, based on the parameter 'name',
 or it will create it if it's not available (see \code{\link{runPCA}})
}
\examples{
g <- GiottoData::loadGiottoMini("visium")

screePlot(g)
}
