% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{select_expression_values}
\alias{select_expression_values}
\title{select_expression_values}
\usage{
select_expression_values(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=get_expression_values]{get_expression_values}}
  \describe{
    \item{\code{values}}{expression values to extract (e.g. "raw", "normalized", "scaled")}
    \item{\code{output}}{what object type to retrieve the expression as. Currently either
'matrix' for the matrix object contained in the exprObj or 'exprObj' (default) for
the exprObj itself are allowed.}
    \item{\code{gobject}}{giotto object}
    \item{\code{spat_unit}}{spatial unit (e.g. "cell")}
    \item{\code{feat_type}}{feature type (e.g. "rna", "dna", "protein")}
    \item{\code{set_defaults}}{set default spat_unit and feat_type. Change to FALSE only when}
  }}
}
\description{
select_expression_values
}
\seealso{
\code{\link{get_expression_values}}
}
\keyword{internal}
