% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{select_feature_info}
\alias{select_feature_info}
\title{Select feature info}
\usage{
select_feature_info(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=get_feature_info]{get_feature_info}}
  \describe{
    \item{\code{gobject}}{giotto object}
    \item{\code{feat_type}}{feature type (e.g. "rna", "dna", "protein")}
    \item{\code{set_defaults}}{set default spat_unit and feat_type. Change to FALSE only when
expression and spat_info are not expected to exist.}
  }}
}
\description{
Select feature info
}
\seealso{
\code{\link{get_feature_info}}
}
\keyword{internal}
