% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_in_situ_visuals.R
\name{select_gimage}
\alias{select_gimage}
\title{select_gimage}
\usage{
select_gimage(
  gobject,
  gimage = NULL,
  image_name = NULL,
  largeImage_name = NULL,
  spat_unit = NULL,
  spat_loc_name = NULL,
  feat_type = NULL
)
}
\description{
selects and creates giotto images for plotting
}
\keyword{internal}
