% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NN_network.R
\name{select_NearestNetwork}
\alias{select_NearestNetwork}
\title{select_NearestNetwork}
\usage{
select_NearestNetwork(
  gobject,
  nn_network_to_use = "sNN",
  network_name = "sNN.pca",
  output = c("igraph", "data.table")
)
}
\arguments{
\item{gobject}{giotto object}

\item{nn_network_to_use}{kNN or sNN}

\item{network_name}{name of NN network to be used}

\item{output}{return a igraph or data.table object}
}
\value{
igraph or data.table object
}
\description{
Extracts a NN-network from a Giotto object
}
\keyword{internal}
