% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{select_polygon_info}
\alias{select_polygon_info}
\title{Select polygon info}
\usage{
select_polygon_info(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=get_polygon_info]{get_polygon_info}}
  \describe{
    \item{\code{gobject}}{giotto object}
    \item{\code{polygon_name}}{name of polygons}
    \item{\code{polygon_overlap}}{include polygon overlap information}
  }}
}
\description{
Select polygon info
}
\seealso{
\code{\link{get_polygon_info}}
}
\keyword{internal}
