% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{select_spatial_locations}
\alias{select_spatial_locations}
\title{select_expression_values}
\usage{
select_spatial_locations(gobject, spat_loc_name = "raw")
}
\arguments{
\item{gobject}{giotto object}

\item{feat_type}{feature type}

\item{values}{expression values to extract}
}
\value{
expression matrix
}
\description{
helper function to select expression values
}
\keyword{internal}
