% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{select_spatialGrid}
\alias{select_spatialGrid}
\title{Select spatial grid}
\usage{
select_spatialGrid(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=get_spatialGrid]{get_spatialGrid}}
  \describe{
    \item{\code{gobject}}{giotto object}
    \item{\code{spat_unit}}{spatial unit}
    \item{\code{feat_type}}{feature type}
    \item{\code{name}}{name of spatial grid}
  }}
}
\description{
Select spatial grid
}
\seealso{
\code{\link{get_spatialGrid}}
}
\keyword{internal}
