% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{select_spatialNetwork}
\alias{select_spatialNetwork}
\title{Select spatial network}
\usage{
select_spatialNetwork(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=get_spatialNetwork]{get_spatialNetwork}}
  \describe{
    \item{\code{name}}{name of spatial network}
    \item{\code{output}}{object type to return as. Options: 'spatialNetworkObj' (default),
'networkDT' and 'networkDT_before_filter' for data.table outputs.}
    \item{\code{copy_obj}}{whether to copy/duplicate when getting the object (default = TRUE)}
    \item{\code{verbose}}{be verbose}
    \item{\code{gobject}}{giotto object}
    \item{\code{spat_unit}}{spatial unit (e.g. "cell")}
    \item{\code{set_defaults}}{set default spat_unit and feat_type. Change to FALSE only when}
  }}
}
\description{
Select spatial network
}
\seealso{
\code{\link{get_spatialNetwork}}
}
\keyword{internal}
