% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_genes.R
\name{selectPatternGenes}
\alias{selectPatternGenes}
\title{selectPatternGenes}
\usage{
selectPatternGenes(
  spatPatObj,
  dimensions = 1:5,
  top_pos_genes = 10,
  top_neg_genes = 10,
  min_pos_cor = 0.5,
  min_neg_cor = -0.5,
  return_top_selection = FALSE
)
}
\arguments{
\item{spatPatObj}{Output from detectSpatialPatterns}

\item{dimensions}{dimensions to identify correlated genes for.}

\item{top_pos_genes}{Top positively correlated genes.}

\item{top_neg_genes}{Top negatively correlated genes.}

\item{min_pos_cor}{Minimum positive correlation score to include a gene.}

\item{min_neg_cor}{Minimum negative correlation score to include a gene.}

\item{return_top_selection}{only return selection based on correlation criteria (boolean)}
}
\value{
Data.table with genes associated with selected dimension (PC).
}
\description{
Select genes correlated with spatial patterns
}
\details{
Description.
}
