% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto.R
\name{set_cell_and_feat_IDs}
\alias{set_cell_and_feat_IDs}
\title{Set cell and feature IDs}
\usage{
set_cell_and_feat_IDs(gobject)
}
\description{
sets cell and feature IDs based on provided expression data
}
\keyword{internal}
