% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{set_cell_id}
\alias{set_cell_id}
\title{Set cell IDs for a given spatial unit}
\usage{
set_cell_id(
  gobject,
  spat_unit = NULL,
  cell_IDs,
  set_defaults = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit (e.g. "cell")}

\item{cell_IDs}{character vector of cell IDs to set. (See details)}

\item{set_defaults}{set default spat_unit and feat_type. Change to FALSE only when
expression and spat_info are not expected to exist.}

\item{verbose}{be verbose}
}
\value{
giotto object with set cell_ID slot
}
\description{
Setter function for the cell_ID slot. Directly replaces (sets)
this slot
}
\details{
Data for each spatial unit is expected to agree on a single set of cell_IDs
that are shared across any feature types. These cell_IDs are stored within the
giotto object's \code{cell_ID} slot. \cr

Pass \code{NULL} to \code{cell_IDs} param in order to delete the entry. \cr
Pass \code{'initialize'} to \code{cell_IDs} param in order to initialize the
specified entry. \cr

\strong{NOTE:} The main purpose of the setter is to initialize, as cell_ID values
are AUTOMATICALLY updated every time \code{initialize()} is called on the
giotto object.
}
\seealso{
get_cell_id

Other functions to set data in giotto object: 
\code{\link{get_cell_id}()},
\code{\link{get_feat_id}()},
\code{\link{setCellMetadata}()},
\code{\link{setDimReduction}()},
\code{\link{setExpression}()},
\code{\link{setFeatureInfo}()},
\code{\link{setFeatureMetadata}()},
\code{\link{setGiottoImage}()},
\code{\link{setMultiomics}()},
\code{\link{setNearestNetwork}()},
\code{\link{setPolygonInfo}()},
\code{\link{setSpatialEnrichment}()},
\code{\link{setSpatialGrid}()},
\code{\link{setSpatialLocations}()},
\code{\link{setSpatialNetwork}()},
\code{\link{set_NearestNetwork}()},
\code{\link{set_cell_metadata}()},
\code{\link{set_dimReduction}()},
\code{\link{set_expression_values}()},
\code{\link{set_feat_id}()},
\code{\link{set_feature_info}()},
\code{\link{set_giottoImage}()},
\code{\link{set_multiomics}()},
\code{\link{set_polygon_info}()},
\code{\link{set_spatialGrid}()},
\code{\link{set_spatialNetwork}()},
\code{\link{set_spatial_enrichment}()},
\code{\link{set_spatial_locations}()}
}
\concept{functions to set data in giotto object}
\keyword{internal}
