% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{set_cell_metadata}
\alias{set_cell_metadata}
\title{Set cell metadata}
\usage{
set_cell_metadata(gobject, cell_metadata)
}
\arguments{
\item{gobject}{giotto object}

\item{cell_metadata}{cell_metadata supplied as a nested list with spat_unit$feat_type$name}
}
\description{
Set cell metadata
}
\keyword{internal}
