% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visuals.R
\name{set_default_cow_n_col}
\alias{set_default_cow_n_col}
\title{Set default ncols in plotting grid}
\usage{
set_default_cow_n_col(cow_n_col = NULL, nr_plots)
}
\description{
Set default ncols in plotting grid
}
\keyword{internal}
