% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{set_dimReduction}
\alias{set_dimReduction}
\title{Set dimension reduction}
\usage{
set_dimReduction(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  reduction = c("cells", "genes"),
  reduction_method = c("pca", "umap", "tsne"),
  name = "pca",
  dimObject
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit, e.g. "cell"}

\item{feat_type}{feature type, e.g. "rna", "dna", "protein"}

\item{reduction}{reduction on cells or features}

\item{reduction_method}{reduction method (e.g. "pca")}

\item{name}{name of reduction results}

\item{dimObject}{dimension object result to set}
}
\value{
giotto object
}
\description{
Function to set a dimension reduction slot
}
\examples{
expr_values = Giotto:::create_feats_to_use_matrix(gobject = my_giotto_object,
                                                  spat_unit = "cell",
                                                  feat_type = "rna",
                                                  sel_matrix = my_giotto_object@expression$cell$rna$raw,
                                                  feats_to_use = 'hvf')
pca_object <- Giotto:::runPCA_factominer(x = Giotto:::t_flex(expr_values),
                                         scale = TRUE,
                                         ncp = 100, rev = FALSE,
                                         set_seed = TRUE,
                                         seed_number = 1234)
my_giotto_object <- set_dimReduction(my_giotto_object,
                                     spat_unit = "cell",
                                     feat_type = "rna",
                                     reduction = "genes",
                                     reduction_method = "pca",
                                     name = "pca",
                                     dimObject = pca_object)
}
\seealso{
Other dimensional reduction data accessor functions: 
\code{\link{get_dimReduction}()}

Other functions to set data in giotto object: 
\code{\link{set_NearestNetwork}()},
\code{\link{set_expression_values}()},
\code{\link{set_feature_info}()},
\code{\link{set_giottoImage}()},
\code{\link{set_polygon_info}()},
\code{\link{set_spatialGrid}()},
\code{\link{set_spatialNetwork}()},
\code{\link{set_spatial_enrichment}()},
\code{\link{set_spatial_locations}()}
}
\concept{dimensional reduction data accessor functions}
\concept{functions to set data in giotto object}
