% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{set_feat_id}
\alias{set_feat_id}
\title{Set feat IDs for a given feature type}
\usage{
set_feat_id(gobject, feat_type = NULL, feat_IDs, set_defaults = TRUE)
}
\arguments{
\item{gobject}{giotto object}

\item{feat_type}{feature type (e.g. "rna", "dna", "protein")}

\item{set_defaults}{set default spat_unit and feat_type. Change to FALSE only when}
}
\value{
giotto object with set cell_ID slot
}
\description{
Across a single modality/feature type, all feature information is
expected to share a single set of feat_IDs. These feat_IDs are stored within
the giotto object's \code{feat_ID} slot. Getters and setters for this slot
directly (get) or replace (set) this slot.
}
\seealso{
get_feat_id

Other functions to set data in giotto object: 
\code{\link{get_cell_id}()},
\code{\link{get_feat_id}()},
\code{\link{setDimReduction}()},
\code{\link{setExpression}()},
\code{\link{setFeatureInfo}()},
\code{\link{setGiottoImage}()},
\code{\link{setMultiomics}()},
\code{\link{setNearestNetwork}()},
\code{\link{setPolygonInfo}()},
\code{\link{setSpatialEnrichment}()},
\code{\link{setSpatialGrid}()},
\code{\link{setSpatialLocations}()},
\code{\link{setSpatialNetwork}()},
\code{\link{set_NearestNetwork}()},
\code{\link{set_cell_id}()},
\code{\link{set_cell_metadata}()},
\code{\link{set_dimReduction}()},
\code{\link{set_expression_values}()},
\code{\link{set_feature_info}()},
\code{\link{set_feature_metadata}()},
\code{\link{set_giottoImage}()},
\code{\link{set_multiomics}()},
\code{\link{set_polygon_info}()},
\code{\link{set_spatialGrid}()},
\code{\link{set_spatialNetwork}()},
\code{\link{set_spatial_enrichment}()},
\code{\link{set_spatial_locations}()}
}
\concept{functions to set data in giotto object}
\keyword{internal}
