% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto.R
\name{set_giotto_python_path}
\alias{set_giotto_python_path}
\title{set_giotto_python_path}
\usage{
set_giotto_python_path(
  python_path = NULL,
  packages_to_install = c("pandas", "networkx", "python-igraph", "leidenalg",
    "python-louvain", "smfishHmrf")
)
}
\description{
sets the python path and/or install miniconda and the python modules
}
