% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{set_giottoImage}
\alias{set_giottoImage}
\title{set_giottoImage}
\usage{
set_giottoImage(
  gobject = NULL,
  image = NULL,
  image_type = NULL,
  name = NULL,
  verbose = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{image}{giotto image object to be attached without modification to the
giotto object}

\item{image_type}{type of giotto image object}

\item{name}{name of giotto image object}

\item{verbose}{be verbose}
}
\value{
a giotto object with image object directly attached
}
\description{
directly attach a giotto image to giotto object
}
\details{
Unlike the other setter functions, this is an internal function.
  Using this function, giotto image objects are directly attached to the
  gobject without further modifications of spatial positioning values
  within the image object that are generally needed in order for them to
  plot in the correct location relative to the other modalities of spatial data.
  Thus, this function should be used with care. For the more general-purpose
  method of attaching image objects, see \code{\link{addGiottoImage}}
}
\keyword{internal}
