% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{set_giottoImage_MG}
\alias{set_giottoImage_MG}
\title{Set \emph{magick}-based giotto \code{image}}
\usage{
set_giottoImage_MG(gobject, image_object, name = NULL, verbose = TRUE)
}
\arguments{
\item{gobject}{giotto object}

\item{image_object}{a giottoImage object}

\item{name}{name to assign giottoImage}

\item{verbose}{be verbose}
}
\value{
giotto object
}
\description{
Set a giottoImage for a giotto object with no additional modifications
}
\keyword{internal}
