% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{set_giottoLargeImage}
\alias{set_giottoLargeImage}
\title{set_giottoLargeImage}
\usage{
set_giottoLargeImage(gobject, largeImage_object, name = NULL, verbose = TRUE)
}
\arguments{
\item{gobject}{giotto object}

\item{largeImage_object}{a giottoLargeImage object}

\item{name}{name to assign giottoLargeImage}

\item{verbose}{be verbose}
}
\value{
a giotto object
}
\description{
set a giottoLargeImage for a giotto object with no additional modifications
}
\keyword{internal}
