% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnn.R
\name{set_multiomics}
\alias{set_multiomics}
\title{Set multiomics integration results}
\usage{
set_multiomics(
  gobject,
  result,
  spat_unit = NULL,
  feat_type = NULL,
  integration_method = "WNN",
  result_name = "theta_weighted_matrix",
  verbose = TRUE
)
}
\arguments{
\item{gobject}{A Giotto object}

\item{result}{A matrix or result from multiomics integration (e.g. theta weighted values from runWNN)}

\item{spat_unit}{spatial unit (e.g. 'cell')}

\item{feat_type}{(e.g. 'rna_protein')}

\item{integration_method}{multiomics integration method used. Default = 'WNN'}

\item{result_name}{Default = 'theta_weighted_matrix'}

\item{verbose}{be verbose}
}
\value{
A giotto object
}
\description{
Set a multiomics integration result in a Giotto object
}
\seealso{
Other multiomics accessor functions: 
\code{\link{get_multiomics}()}

Other functions to set data in giotto object: 
\code{\link{get_cell_id}()},
\code{\link{get_feat_id}()},
\code{\link{set_NearestNetwork}()},
\code{\link{set_cell_id}()},
\code{\link{set_cell_metadata}()},
\code{\link{set_dimReduction}()},
\code{\link{set_expression_values}()},
\code{\link{set_feat_id}()},
\code{\link{set_feature_info}()},
\code{\link{set_feature_metadata}()},
\code{\link{set_giottoImage}()},
\code{\link{set_polygon_info}()},
\code{\link{set_spatialGrid}()},
\code{\link{set_spatialNetwork}()},
\code{\link{set_spatial_enrichment}()},
\code{\link{set_spatial_locations}()}
}
\concept{functions to set data in giotto object}
\concept{multiomics accessor functions}
