% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{set_NearestNetwork}
\alias{set_NearestNetwork}
\title{set_NearestNetwork}
\usage{
set_NearestNetwork(
  gobject,
  feat_type = NULL,
  spat_unit = NULL,
  nn_network_to_use = "sNN",
  network_name = "sNN.pca",
  nn_network
)
}
\arguments{
\item{gobject}{giotto object}

\item{feat_type}{feature type}

\item{spat_unit}{spatial unit}

\item{nn_network_to_use}{kNN or sNN}

\item{network_name}{name of NN network to be used}

\item{nn_network}{nearest network}
}
\value{
giotto object
}
\description{
set a NN-network for a Giotto object
}
