% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{set_polygon_info}
\alias{set_polygon_info}
\title{Set polygon info}
\usage{
set_polygon_info(
  gobject,
  gpolygon,
  polygon_name = "cell",
  verbose = TRUE,
  initialize = FALSE
)
}
\arguments{
\item{gobject}{giotto object}

\item{gpolygon}{giottoPolygon object}

\item{polygon_name}{name of polygons. Default "cell" (only used when gpolygon
is length of 1)}

\item{initialize}{(default = FALSE) whether to initialize the gobject before
returning}
}
\value{
giotto object
}
\description{
Set giotto polygon spatVector
}
\seealso{
Other polygon info data accessor functions: 
\code{\link{getPolygonInfo}()},
\code{\link{get_polygon_info}()},
\code{\link{setPolygonInfo}()}

Other functions to set data in giotto object: 
\code{\link{get_cell_id}()},
\code{\link{get_feat_id}()},
\code{\link{setDimReduction}()},
\code{\link{setExpression}()},
\code{\link{setFeatureInfo}()},
\code{\link{setGiottoImage}()},
\code{\link{setMultiomics}()},
\code{\link{setNearestNetwork}()},
\code{\link{setPolygonInfo}()},
\code{\link{setSpatialEnrichment}()},
\code{\link{setSpatialGrid}()},
\code{\link{setSpatialLocations}()},
\code{\link{setSpatialNetwork}()},
\code{\link{set_NearestNetwork}()},
\code{\link{set_cell_id}()},
\code{\link{set_cell_metadata}()},
\code{\link{set_dimReduction}()},
\code{\link{set_expression_values}()},
\code{\link{set_feat_id}()},
\code{\link{set_feature_info}()},
\code{\link{set_feature_metadata}()},
\code{\link{set_giottoImage}()},
\code{\link{set_multiomics}()},
\code{\link{set_spatialGrid}()},
\code{\link{set_spatialNetwork}()},
\code{\link{set_spatial_enrichment}()},
\code{\link{set_spatial_locations}()}
}
\concept{functions to set data in giotto object}
\concept{polygon info data accessor functions}
