% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{set_row_order_dt}
\alias{set_row_order_dt}
\title{Set specific data.table row order}
\usage{
set_row_order_dt(x, neworder)
}
\arguments{
\item{x}{data.table}

\item{neworder}{numerical vector to reorder rows}
}
\description{
Set specific data.table row order
}
\keyword{internal}
