% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{set_spatialGrid}
\alias{set_spatialGrid}
\title{set_spatialGrid}
\usage{
set_spatialGrid(gobject, name = NULL, spatial_grid)
}
\arguments{
\item{gobject}{giotto object}

\item{name}{name of spatial grid}

\item{spatial_grid}{spatial grid}
}
\description{
function to set a spatial grid
}
