% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{setExpression}
\alias{setExpression}
\title{Set expression data}
\usage{
setExpression(
  gobject,
  x,
  spat_unit = NULL,
  feat_type = NULL,
  name = "raw",
  provenance = NULL,
  verbose = TRUE,
  initialize = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{x}{exprObj or list of exprObj to set. Passing NULL will remove a
specified set of expression data from the giotto object}

\item{spat_unit}{spatial unit (e.g. "cell")}

\item{feat_type}{feature type (e.g. "rna", "dna", "protein")}

\item{name}{name for the expression information}

\item{provenance}{provenance information (optional)
information for the giotto object. Pass NULL to remove an expression object}

\item{verbose}{be verbose}

\item{initialize}{(default = FALSE) whether to initialize the gobject before
returning}
}
\value{
giotto object
}
\description{
Function to set expression values for giotto object.
}
\seealso{
Other expression accessor functions: 
\code{\link{getExpression}()},
\code{\link{set_expression_values}()}

Other functions to set data in giotto object: 
\code{\link{get_cell_id}()},
\code{\link{get_feat_id}()},
\code{\link{setCellMetadata}()},
\code{\link{setDimReduction}()},
\code{\link{setFeatureInfo}()},
\code{\link{setFeatureMetadata}()},
\code{\link{setGiottoImage}()},
\code{\link{setMultiomics}()},
\code{\link{setNearestNetwork}()},
\code{\link{setPolygonInfo}()},
\code{\link{setSpatialEnrichment}()},
\code{\link{setSpatialGrid}()},
\code{\link{setSpatialLocations}()},
\code{\link{setSpatialNetwork}()},
\code{\link{set_NearestNetwork}()},
\code{\link{set_cell_id}()},
\code{\link{set_cell_metadata}()},
\code{\link{set_dimReduction}()},
\code{\link{set_expression_values}()},
\code{\link{set_feat_id}()},
\code{\link{set_feature_info}()},
\code{\link{set_giottoImage}()},
\code{\link{set_multiomics}()},
\code{\link{set_polygon_info}()},
\code{\link{set_spatialGrid}()},
\code{\link{set_spatialNetwork}()},
\code{\link{set_spatial_enrichment}()},
\code{\link{set_spatial_locations}()}
}
\concept{expression accessor functions}
\concept{functions to set data in giotto object}
