% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{setFeatureInfo}
\alias{setFeatureInfo}
\title{Set feature info}
\usage{
setFeatureInfo(gobject = NULL, gpoints, feat_type = NULL, verbose = TRUE)
}
\arguments{
\item{gobject}{giotto object}

\item{gpoints}{giotto points object}

\item{feat_type}{feature type (e.g. "rna", "dna", "protein")}

\item{verbose}{be verbose}
}
\value{
giotto object
}
\description{
Set giotto polygon spatVector for features
}
\seealso{
Other feature info data accessor functions: 
\code{\link{getFeatureInfo}()},
\code{\link{get_feature_info}()},
\code{\link{set_feature_info}()}

Other functions to set data in giotto object: 
\code{\link{get_cell_id}()},
\code{\link{get_feat_id}()},
\code{\link{setDimReduction}()},
\code{\link{setExpression}()},
\code{\link{setGiottoImage}()},
\code{\link{setMultiomics}()},
\code{\link{setNearestNetwork}()},
\code{\link{setPolygonInfo}()},
\code{\link{setSpatialEnrichment}()},
\code{\link{setSpatialGrid}()},
\code{\link{setSpatialLocations}()},
\code{\link{setSpatialNetwork}()},
\code{\link{set_NearestNetwork}()},
\code{\link{set_cell_id}()},
\code{\link{set_cell_metadata}()},
\code{\link{set_dimReduction}()},
\code{\link{set_expression_values}()},
\code{\link{set_feat_id}()},
\code{\link{set_feature_info}()},
\code{\link{set_feature_metadata}()},
\code{\link{set_giottoImage}()},
\code{\link{set_multiomics}()},
\code{\link{set_polygon_info}()},
\code{\link{set_spatialGrid}()},
\code{\link{set_spatialNetwork}()},
\code{\link{set_spatial_enrichment}()},
\code{\link{set_spatial_locations}()}
}
\concept{feature info data accessor functions}
\concept{functions to set data in giotto object}
