% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{setNearestNetwork}
\alias{setNearestNetwork}
\title{Set nearest neighbor network}
\usage{
setNearestNetwork(
  gobject,
  x,
  spat_unit = NULL,
  feat_type = NULL,
  nn_type = "sNN",
  name = "sNN.pca",
  provenance = NULL,
  verbose = TRUE,
  set_defaults = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{x}{nnNetObj or list of nnNetObj. Passing NULL will remove a specified
set of nearest neighbor network information from the gobject}

\item{spat_unit}{spatial unit (e.g. "cell")}

\item{feat_type}{feature type (e.g. "rna", "dna", "protein")}

\item{nn_type}{"kNN" or "sNN"}

\item{name}{name of NN network to be used
yet supported.}

\item{provenance}{provenance information (optional)}

\item{verbose}{be verbose}

\item{set_defaults}{set default spat_unit and feat_type. Change to FALSE only when
expression and spat_info are not expected to exist.}
}
\value{
giotto object
}
\description{
Set a NN-network for a Giotto object
}
\seealso{
Other expression space nearest network accessor functions: 
\code{\link{getNearestNetwork}()},
\code{\link{get_NearestNetwork}()},
\code{\link{set_NearestNetwork}()}

Other functions to set data in giotto object: 
\code{\link{get_cell_id}()},
\code{\link{get_feat_id}()},
\code{\link{setDimReduction}()},
\code{\link{setExpression}()},
\code{\link{setFeatureInfo}()},
\code{\link{setGiottoImage}()},
\code{\link{setMultiomics}()},
\code{\link{setPolygonInfo}()},
\code{\link{setSpatialEnrichment}()},
\code{\link{setSpatialGrid}()},
\code{\link{setSpatialLocations}()},
\code{\link{setSpatialNetwork}()},
\code{\link{set_NearestNetwork}()},
\code{\link{set_cell_id}()},
\code{\link{set_cell_metadata}()},
\code{\link{set_dimReduction}()},
\code{\link{set_expression_values}()},
\code{\link{set_feat_id}()},
\code{\link{set_feature_info}()},
\code{\link{set_feature_metadata}()},
\code{\link{set_giottoImage}()},
\code{\link{set_multiomics}()},
\code{\link{set_polygon_info}()},
\code{\link{set_spatialGrid}()},
\code{\link{set_spatialNetwork}()},
\code{\link{set_spatial_enrichment}()},
\code{\link{set_spatial_locations}()}
}
\concept{expression space nearest network accessor functions}
\concept{functions to set data in giotto object}
