% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{setPolygonInfo}
\alias{setPolygonInfo}
\title{Set polygon info}
\usage{
setPolygonInfo(
  gobject,
  gpolygon,
  polygon_name = "cell",
  input = c("table", "mask"),
  polygon_dfr_list_params = NULL,
  polygon_mask_list_params = NULL,
  calc_centroids = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{gpolygon}{single object or named list of objects to set as polygon
information (see details)}

\item{polygon_name}{(optional, character) When setting a single giottoPolygon
object, this param is ignored in favor of a default spatial unit if available
unless polygon_name is explicitly used in the call.
In all other cases, this param is the name to assign the generated polygon
information. If \code{gpolygon} is a single object then it is directly applied.
If \code{gpolygon} is an unnamed list then it will be used as the base of a
template for generating indexed default names for the polygon objects.}

\item{input}{(default = 'table') whether \code{gpolygon} is tabular data or
image ('mask') type information}

\item{polygon_dfr_list_params}{list parameters for
\code{\link{createGiottoPolygonsFromDfr}}. Used when \code{input} is 'table'}

\item{polygon_mask_list_params}{list parameters for
\code{\link{createGiottoPolygonsFromMask}}. Used when \code{input} is 'mask'}

\item{calc_centroids}{whether to generate centroids for the polygon(s)}
}
\value{
giotto object
}
\description{
Set polygon information into Giotto object
}
\details{
Inputs can be provided as either single objects or named lists of
objects. If the list is not named, then a generic name of the template
'cell_i' will be applied. \cr
If an input is a character string, then it is assumed that it is a
filepath. \cr
For required formatting when reading tabular data or objects, see
\code{\link{createGiottoPolygonsFromDfr}} details.
}
\seealso{
Other polygon info data accessor functions: 
\code{\link{getPolygonInfo}()},
\code{\link{get_polygon_info}()},
\code{\link{set_polygon_info}()}

Other functions to set data in giotto object: 
\code{\link{get_cell_id}()},
\code{\link{get_feat_id}()},
\code{\link{setDimReduction}()},
\code{\link{setExpression}()},
\code{\link{setFeatureInfo}()},
\code{\link{setGiottoImage}()},
\code{\link{setMultiomics}()},
\code{\link{setNearestNetwork}()},
\code{\link{setSpatialEnrichment}()},
\code{\link{setSpatialGrid}()},
\code{\link{setSpatialLocations}()},
\code{\link{setSpatialNetwork}()},
\code{\link{set_NearestNetwork}()},
\code{\link{set_cell_id}()},
\code{\link{set_cell_metadata}()},
\code{\link{set_dimReduction}()},
\code{\link{set_expression_values}()},
\code{\link{set_feat_id}()},
\code{\link{set_feature_info}()},
\code{\link{set_feature_metadata}()},
\code{\link{set_giottoImage}()},
\code{\link{set_multiomics}()},
\code{\link{set_polygon_info}()},
\code{\link{set_spatialGrid}()},
\code{\link{set_spatialNetwork}()},
\code{\link{set_spatial_enrichment}()},
\code{\link{set_spatial_locations}()}
}
\concept{functions to set data in giotto object}
\concept{polygon info data accessor functions}
