% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interoperability.R
\name{seuratToGiotto_OLD}
\alias{seuratToGiotto_OLD}
\title{seuratToGiotto_OLD}
\usage{
seuratToGiotto_OLD(obj_use = NULL, ...)
}
\arguments{
\item{obj_use}{Seurat object}

\item{...}{additional params to pass}
}
\value{
Giotto object
}
\description{
Converts Seurat object into a Giotto object. Deprecated, see \code{\link{giottoToSeurat}}
}
