% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_manipulation.R
\name{shift_spatial_network}
\alias{shift_spatial_network}
\title{Shift spatial network}
\usage{
shift_spatial_network(spatnet, dx = 0, dy = 0, dz = 0, copy_obj = TRUE)
}
\arguments{
\item{spatnet}{spatial network data.table}

\item{dx}{distance to shift on x axis}

\item{dy}{distance to shift on y axis}

\item{dz}{distance to shift on z axis}

\item{copy_obj}{copy/duplicate object (default = TRUE)}
}
\description{
Shift spatial network coordinates
}
\keyword{internal}
