% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_influence.R
\name{showCellProportionSwitchedPie}
\alias{showCellProportionSwitchedPie}
\title{showCellProportionSwitchedPie}
\usage{
showCellProportionSwitchedPie(
  gobject = NULL,
  spat_unit = NULL,
  feat_type = NULL
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit}

\item{feat_type}{feature type}
}
\description{
showCellProportionSwitchedPie
}
\details{
Creates a pie chart showing how many cells switched clusters after 
annotation resizing.
The function showPolygonSizeInfluence() must have been run on the Giotto 
Object for this function to run.
}
