% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_influence.r
\name{showCellProportionSwitchedSanKey}
\alias{showCellProportionSwitchedSanKey}
\title{showCellProportionSwitchedSanKey}
\usage{
showCellProportionSwitchedSanKey(
  gobject = NULL,
  spat_unit = NULL,
  alt_spat_unit = NULL,
  feat_type = NULL
)
}
\arguments{
\item{gobject}{giotto object which contains metadata for both spat_unit and alt_spat_unit}

\item{spat_unit}{spatial unit}

\item{alt_spat_unit}{alternative spatial unit which stores data after resizing annotations}

\item{feat_type}{feature type}
}
\description{
showCellProportionSwitchedSanKey
}
\details{
Creates a Sankey Diagram to illustrate cluster switching behavior.
Currently only supports displaying cluster switching for kmeans clusters.
}
