% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_visuals.R
\name{showClusterDendrogram}
\alias{showClusterDendrogram}
\title{showClusterDendrogram}
\usage{
showClusterDendrogram(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  expression_values = c("normalized", "scaled", "custom"),
  cluster_column,
  cor = c("pearson", "spearman"),
  distance = "ward.D",
  h = NULL,
  h_color = "red",
  rotate = FALSE,
  show_plot = NA,
  return_plot = NA,
  save_plot = NA,
  save_param = list(),
  default_save_name = "showClusterDendrogram",
  ...
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit (e.g. "cell")}

\item{feat_type}{feature type (e.g. "rna", "dna", "protein")}

\item{expression_values}{expression values to use (e.g. "normalized", "scaled", "custom")}

\item{cluster_column}{name of column to use for clusters (e.g. "leiden_clus")}

\item{cor}{correlation score to calculate distance (e.g. "pearson", "spearman")}

\item{distance}{distance method to use for hierarchical clustering, default to "ward.D"}

\item{h}{height of horizontal lines to plot}

\item{h_color}{color of horizontal lines}

\item{rotate}{rotate dendrogram 90 degrees}

\item{show_plot}{show plot. TRUE or FALSE}

\item{return_plot}{return ggplot object. TRUE or FALSE}

\item{save_plot}{directly save the plot. TRUE or FALSE}

\item{save_param}{list of saving parameters, see \code{\link{showSaveParameters}}}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}

\item{...}{additional parameters passed to \code{\link[ggdendro]{ggdendrogram}}}
}
\value{
ggplot
}
\description{
Creates dendrogram for selected clusters.
}
\details{
Expression correlation dendrogram for selected clusters.
}
