% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction_visuals.R
\name{showCPGscores}
\alias{showCPGscores}
\title{showCPGscores}
\usage{
showCPGscores(gobject, CPGscore, method = c("cell_barplot", "cell-cell",
  "cell_sankey"), min_cells = 5, min_pval = 0.05,
  min_spat_diff = 0.2, min_log2_fc = 0.5, direction = c("both", "up",
  "down"), cell_color_code = NULL, return_DT = F, show_plot = NA,
  return_plot = NA, save_plot = NA, save_param = list(),
  default_save_name = "showCPGscores")
}
\arguments{
\item{CPGscore}{CPGscore, output from getCellProximityGeneScores()}

\item{method}{visualization method}

\item{min_cells}{min number of cells threshold}

\item{min_pval}{p-value threshold}

\item{min_spat_diff}{spatial difference threshold}

\item{min_log2_fc}{min log2 fold-change}

\item{direction}{up or downregulation or both}

\item{cell_color_code}{color code for cell types}

\item{return_DT}{return filtered data.table (boolean)}

\item{show_plot}{show plot}

\item{return_plot}{return ggplot object}

\item{save_plot}{directly save the plot [boolean]}

\item{save_param}{list of saving parameters from all_plots_save_function()}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}
}
\value{
Gene to gene scores in data.table format
}
\description{
visualize Cell Proximity Gene enrichment scores
}
\details{
Give more details ...
}
\examples{
    showCPGscores(CPGscore)
}
