% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction_visuals.R
\name{showGeneExpressionProximityScore}
\alias{showGeneExpressionProximityScore}
\title{showGeneExpressionProximityScore}
\usage{
showGeneExpressionProximityScore(scores, selected_gene,
  sort_column = "diff_spat")
}
\arguments{
\item{scores}{CPscore, output from getAverageCellProximityGeneScores()}

\item{selected_gene}{gene to show}

\item{sort_column}{column name to use for sorting}
}
\value{
ggplot barplot
}
\description{
Create heatmap from cell-cell proximity scores
}
\details{
Give more details ...
}
\examples{
    showGeneExpressionProximityScore(scores)
}
