% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{showGiottoDimRed}
\alias{showGiottoDimRed}
\title{showGiottoDimRed}
\usage{
showGiottoDimRed(gobject, nrows = 3, ncols = 2)
}
\arguments{
\item{gobject}{giotto object}

\item{nrows}{number of coordinates rows to print}

\item{ncols}{number of coordinates columns to print}
}
\value{
prints the name and small subset of available dimension reduction coordinates
}
\description{
shows the available dimension reductions
}
