% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{showGiottoExpression}
\alias{showGiottoExpression}
\title{showGiottoExpression}
\usage{
showGiottoExpression(gobject, nrows = 4, ncols = 4)
}
\arguments{
\item{gobject}{giotto object}

\item{nrows}{number of rows to print for each matrix}

\item{ncols}{number of columns to print for each matrix}
}
\value{
prints the name and small subset of available matrices
}
\description{
shows the available matrices
}
\keyword{show}
