% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{showGiottoFeatInfo}
\alias{showGiottoFeatInfo}
\title{showGiottoFeatInfo}
\usage{
showGiottoFeatInfo(gobject)
}
\arguments{
\item{gobject}{giotto object}
}
\description{
show the available giotto spatial feature information
}
\keyword{show}
