% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{showGiottoFeatMetadata}
\alias{showGiottoFeatMetadata}
\title{showGiottoFeatMetadata}
\usage{
showGiottoFeatMetadata(gobject, nrows = 4)
}
\arguments{
\item{gobject}{giotto object}

\item{nrows}{number of rows to print for each metadata}
}
\value{
prints the name and small subset of available metadata
}
\description{
shows the available feature metadata
}
\seealso{
Other functions to show data in giotto object: 
\code{\link{showGiottoCellMetadata}()},
\code{\link{showGiottoDimRed}()},
\code{\link{showGiottoExpression}()},
\code{\link{showGiottoFeatInfo}()},
\code{\link{showGiottoImageNames}()},
\code{\link{showGiottoNearestNetworks}()},
\code{\link{showGiottoSpatEnrichments}()},
\code{\link{showGiottoSpatGrids}()},
\code{\link{showGiottoSpatLocs}()},
\code{\link{showGiottoSpatNetworks}()},
\code{\link{showGiottoSpatialInfo}()}
}
\concept{functions to show data in giotto object}
\keyword{show}
