% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto.R
\name{showGiottoInstructions}
\alias{showGiottoInstructions}
\title{showGiottoInstructions}
\usage{
showGiottoInstructions(gobject)
}
\arguments{
\item{gobject}{giotto object}
}
\value{
named vector with giotto instructions
}
\description{
Function to show instructions from giotto object
}
\examples{
    showGiottoInstructions()
}
