% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{showGiottoNearestNetworks}
\alias{showGiottoNearestNetworks}
\title{showGiottoNearestNetworks}
\usage{
showGiottoNearestNetworks(gobject, nrows = 3)
}
\arguments{
\item{gobject}{giotto object}

\item{nrows}{number of network rows to print}
}
\value{
prints the name and small subset of available nearest neighbor network info
}
\description{
shows the available nearest neighbor networks
}
\seealso{
Other functions to show data in giotto object: 
\code{\link{showGiottoCellMetadata}()},
\code{\link{showGiottoDimRed}()},
\code{\link{showGiottoExpression}()},
\code{\link{showGiottoFeatInfo}()},
\code{\link{showGiottoFeatMetadata}()},
\code{\link{showGiottoImageNames}()},
\code{\link{showGiottoSpatEnrichments}()},
\code{\link{showGiottoSpatGrids}()},
\code{\link{showGiottoSpatLocs}()},
\code{\link{showGiottoSpatNetworks}()},
\code{\link{showGiottoSpatialInfo}()}
}
\concept{functions to show data in giotto object}
\keyword{show}
