% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{showGiottoSpatEnrichments}
\alias{showGiottoSpatEnrichments}
\title{showGiottoSpatEnrichments}
\usage{
showGiottoSpatEnrichments(gobject, nrows = 4)
}
\arguments{
\item{gobject}{giotto object}

\item{nrows}{number of rows to print for each spatial enrichment data.table}
}
\value{
prints the name and small subset of available data.table
}
\description{
shows the available spatial enrichment results
}
