% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{showGiottoSpatGrids}
\alias{showGiottoSpatGrids}
\title{showGiottoSpatGrids}
\usage{
showGiottoSpatGrids(gobject, nrows = 4)
}
\arguments{
\item{gobject}{giotto object}

\item{nrows}{number of rows to print}
}
\value{
prints name of available spatial grids
}
\description{
Prints the available spatial grids that are attached to the Giotto object
}
