% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{showGrids}
\alias{showGrids}
\title{showGrids}
\usage{
showGrids(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=showGiottoSpatGrids]{showGiottoSpatGrids}}
  \describe{
    \item{\code{gobject}}{a giotto object}
    \item{\code{verbose}}{verbosity of function#'}
  }}
}
\description{
showGrids
}
\seealso{
\code{\link{showGiottoSpatGrids}}
}
