% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction_visuals.R
\name{showGTGscores}
\alias{showGTGscores}
\title{showGTGscores}
\usage{
showGTGscores(
  GTGscore,
  method = c("cell_barplot", "cell-cell", "cell_sankey"),
  min_cells = 5,
  min_pval = 0.05,
  min_spat_diff = 0.2,
  min_log2_fc = 0.5,
  direction = c("both", "up", "down"),
  cell_color_code = NULL,
  show_plot = T,
  specific_genes_1 = NULL,
  specific_genes_2 = NULL,
  first_cell_name = "ligand cell",
  second_cell_name = "receptor cell",
  return_DT = F
)
}
\arguments{
\item{method}{visualization method}

\item{min_cells}{min number of cells threshold}

\item{min_pval}{p-value threshold}

\item{min_spat_diff}{spatial difference threshold}

\item{min_log2_fc}{log2 fold-change threshold}

\item{direction}{up or downregulation or both}

\item{cell_color_code}{color code for cell types}

\item{show_plot}{print plot}

\item{specific_genes_1}{subset of genes, matched with specific_genes_2}

\item{specific_genes_2}{subset of genes, matched with specific_genes_1}

\item{first_cell_name}{name for first cells}

\item{second_cell_name}{name for second cells}

\item{CPGscore}{CPGscore, output from getCellProximityGeneScores()}
}
\value{
ggplot
}
\description{
visualize Cell Proximity Gene enrichment scores
}
\details{
Give more details ...
}
\examples{
    showGTGscores(CPGscore)
}
