% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{showImageNames}
\alias{showImageNames}
\title{showImageNames}
\usage{
showImageNames(gobject, verbose = TRUE)
}
\arguments{
\item{gobject}{a giotto object}

\item{verbose}{verbosity of function}
}
\value{
a vector of giotto image names attached to the giotto object
}
\description{
Show wich images are attached to the Giotto object (prints the names)
}
\examples{
    showImageNames(gobject)
}
