% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction_visuals.R
\name{showIntExpressionProximityScore}
\alias{showIntExpressionProximityScore}
\title{showIntExpressionProximityScore}
\usage{
showIntExpressionProximityScore(scores, selected_interaction,
  sort_column = "diff_spat", show_enriched_n = 5,
  show_depleted_n = 5)
}
\arguments{
\item{scores}{scores, output from getAverageCellProximityGeneScores()}

\item{selected_interaction}{interaction to show}

\item{sort_column}{column name to use for sorting}

\item{show_enriched_n}{show top enriched interactions}

\item{show_depleted_n}{show top depleted interactions}
}
\value{
ggplot barplot
}
\description{
Create heatmap from cell-cell proximity scores
}
\details{
Give more details ...
}
\examples{
    showIntExpressionProximityScore(scores)
}
