% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{showNetworks}
\alias{showNetworks}
\title{showNetworks}
\usage{
showNetworks(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=showGiottoSpatNetworks]{showGiottoSpatNetworks}}
  \describe{
    \item{\code{gobject}}{a giotto object}
    \item{\code{verbose}}{verbosity of function#'}
  }}
}
\description{
showNetworks
}
\seealso{
\code{\link{showGiottoSpatNetworks}}
}
