% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_genes.R
\name{showPattern3D}
\alias{showPattern3D}
\title{showPattern3D}
\usage{
showPattern3D(
  gobject,
  spatPatObj,
  dimension = 1,
  trim = c(0.02, 0.98),
  background_color = "white",
  grid_border_color = "grey",
  show_legend = T,
  point_size = 1,
  axis_scale = c("cube", "real", "custom"),
  custom_ratio = NULL,
  x_ticks = NULL,
  y_ticks = NULL,
  z_ticks = NULL,
  show_plot = NA,
  return_plot = NA,
  save_plot = NA,
  save_param = list(),
  default_save_name = "showPattern3D"
)
}
\arguments{
\item{gobject}{giotto object}

\item{spatPatObj}{Output from detectSpatialPatterns}

\item{dimension}{dimension to plot}

\item{trim}{Trim ends of the PC values.}

\item{background_color}{background color for plot}

\item{grid_border_color}{color for grid}

\item{show_legend}{show legend of plot}

\item{point_size}{adjust the point size}

\item{axis_scale}{scale the axis}

\item{custom_ratio}{cutomize the scale of the axis}

\item{x_ticks}{the tick number of x_axis}

\item{y_ticks}{the tick number of y_axis}

\item{z_ticks}{the tick number of z_axis}

\item{show_plot}{show plot}

\item{return_plot}{return plot object}

\item{save_plot}{directly save the plot [boolean]}

\item{save_param}{list of saving parameters, see \code{\link{showSaveParameters}}}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}
}
\value{
plotly
}
\description{
show patterns for 3D spatial data
}
