% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_genes.R
\name{showPatternGenes}
\alias{showPatternGenes}
\title{showPatternGenes}
\usage{
showPatternGenes(gobject, spatPatObj, dimension = 1, top_pos_genes = 5,
  top_neg_genes = 5, point_size = 1, show_plot = F,
  return_plot = TRUE, save_plot = F, save_param = list(...))
}
\arguments{
\item{gobject}{giotto object}

\item{spatPatObj}{Output from detectSpatialPatterns}

\item{dimension}{dimension to plot genes for.}

\item{top_pos_genes}{Top positively correlated genes.}

\item{top_neg_genes}{Top negatively correlated genes.}

\item{point_size}{size of points}

\item{show_plot}{show plot}

\item{return_plot}{return ggplot object}

\item{save_plot}{directly save the plot [boolean]}

\item{save_param}{list of saving parameters from all_plots_save_function()}
}
\value{
ggplot
}
\description{
show genes correlated with spatial patterns
}
\details{
Description.
}
\examples{
    showPatternGenes(gobject)
}
