% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{showProcessingSteps}
\alias{showProcessingSteps}
\title{showProcessingSteps}
\usage{
showProcessingSteps(gobject)
}
\arguments{
\item{gobject}{giotto object}
}
\value{
list of processing steps and names
}
\description{
shows the sequential processing steps that were performed
on a Giotto object in a summarized format
}
\examples{

data(mini_giotto_single_cell)

showProcessingSteps(mini_giotto_single_cell)

}
