% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_visuals.R
\name{showSaveParameters}
\alias{showSaveParameters}
\title{showSaveParameters}
\usage{
showSaveParameters()
}
\value{
Instruction on how to use the automatic plot saving options within Giotto
}
\description{
Description of Giotto saving options, links to \code{\link{all_plots_save_function}}
}
\examples{
    showSaveParameters()
}
