% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_genes.R
\name{showSpatialCorFeats}
\alias{showSpatialCorFeats}
\title{showSpatialCorFeats}
\usage{
showSpatialCorFeats(
  spatCorObject,
  use_clus_name = NULL,
  selected_clusters = NULL,
  feats = NULL,
  min_spat_cor = 0.5,
  min_expr_cor = NULL,
  min_cor_diff = NULL,
  min_rank_diff = NULL,
  show_top_feats = NULL
)
}
\arguments{
\item{spatCorObject}{spatial correlation object}

\item{use_clus_name}{cluster information to show}

\item{selected_clusters}{subset of clusters to show}

\item{feats}{subset of features to show}

\item{min_spat_cor}{filter on minimum spatial correlation}

\item{min_expr_cor}{filter on minimum single-cell expression correlation}

\item{min_cor_diff}{filter on minimum correlation difference
(spatial vs expression)}

\item{min_rank_diff}{filter on minimum correlation rank difference
(spatial vs expression)}

\item{show_top_feats}{show top features per gene}
}
\value{
data.table with filtered information
}
\description{
Shows and filters spatially correlated features
}
