% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction_visuals.R
\name{showTopGeneToGene}
\alias{showTopGeneToGene}
\title{showTopGeneToGene}
\usage{
showTopGeneToGene(
  GTGscore,
  top_interactions = 10,
  direction = c("increased", "decreased"),
  complement_data = T,
  subset_cell_ints = NULL,
  subset_genes = NULL
)
}
\arguments{
\item{GTGscore}{GTGscore, output from getGeneToGeneScores()}

\item{top_interactions}{number of top gene-gene enrichments to show}

\item{direction}{show top increased or decreased gene-gene enrichments}

\item{complement_data}{include non-enriched gene-gene scores from other cell-cell interactions}

\item{subset_cell_ints}{subset cell-cell interactions to show}

\item{subset_genes}{subset genes to show}
}
\value{
ggplot barplot
}
\description{
Show enriched/depleted gene-gene enrichments
}
\details{
Give more details ...
}
\examples{
    showTopGeneToGene(scores)
}
