% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_genes.R
\name{silhouetteRankTest}
\alias{silhouetteRankTest}
\title{silhouetteRankTest}
\usage{
silhouetteRankTest(
  gobject,
  expression_values = c("normalized", "scaled", "custom"),
  subset_genes = NULL,
  overwrite_input_bin = TRUE,
  rbp_ps = c(0.95, 0.99),
  examine_tops = c(0.005, 0.01, 0.05, 0.1, 0.3),
  matrix_type = "dissim",
  num_core = 4,
  parallel_path = "/usr/bin",
  output = NULL,
  query_sizes = 10L,
  verbose = FALSE
)
}
\arguments{
\item{gobject}{giotto object}

\item{expression_values}{expression values to use}

\item{subset_genes}{only run on this subset of genes}

\item{overwrite_input_bin}{overwrite input bin}

\item{rbp_ps}{fractional binarization thresholds}

\item{examine_tops}{top fractions to evaluate with silhouette}

\item{matrix_type}{type of matrix}

\item{num_core}{number of cores to use}

\item{parallel_path}{path to GNU parallel function}

\item{output}{output directory}

\item{query_sizes}{size of query}

\item{verbose}{be verbose}
}
\value{
data.table with spatial scores
}
\description{
Multi parameter aggregator version of \code{\link{silhouetteRank}}
}
