% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{simulatePatternGiottoObject}
\alias{simulatePatternGiottoObject}
\title{simulatePatternGiottoObject}
\usage{
simulatePatternGiottoObject(
  gobject,
  pattern_name = "pattern",
  pattern_cell_ids = NULL,
  spatial_prob = 0.95,
  gradient_direction = NULL,
  show_pattern = TRUE,
  pattern_colors = c(`in` = "green", out = "red")
)
}
\arguments{
\item{gobject}{Giotto object}

\item{pattern_name}{name of spatial pattern}

\item{pattern_cell_ids}{cell ids from original giotto object to create spatial pattern}

\item{spatial_prob}{probability for a high-expressing cell to be assigned to the spatial pattern}

\item{gradient_direction}{direction of gradient (default = NULL)}

\item{show_pattern}{show how pattern looks like}

\item{pattern_colors}{named vector with 2 colors}
}
\value{
Giotto object with simulated normalized and raw expression according to a spatial pattern
}
\description{
Forces all genes to follow a defined spatial pattern [beta]
}
\examples{
    simulatePatternGiottoObject(gobject)
}
