% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{slot}
\alias{slot}
\title{.hasSlot(object, name)}
\usage{
slot(object, name)
}
\description{
slotNames(x)
getSlots(x)
slot(object, name, check = TRUE) <- value
See \code{methods::\link[methods]{slot}} for details.
}
\keyword{internal}
