% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{slotNames}
\alias{slotNames}
\title{See \code{methods::\link[methods]{slotNames}} for details.}
\usage{
slotNames(object)
}
\description{
See \code{methods::\link[methods]{slotNames}} for details.
}
\keyword{internal}
