% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_structures.R
\name{smoothGiottoPolygons}
\alias{smoothGiottoPolygons}
\title{smoothGiottoPolygons}
\usage{
smoothGiottoPolygons(
  gpolygon,
  vertices = 20,
  k = 3,
  set_neg_to_zero = TRUE,
  ...
)
}
\arguments{
\item{gpolygon}{giotto polygon object}

\item{vertices}{number of vertices}

\item{k}{k}

\item{set_neg_to_zero}{set negative values to zero (default: TRUE)}

\item{...}{additional params to pass to \code{spline}}
}
\value{
Smoothed Giotto polygon object with reduced vertices
}
\description{
Smooths Giotto polygon object
}
\seealso{
\code{\link[stats]{spline}}
}
\concept{polygon}
