% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_enrichment.R
\name{solve_dampened_WLSj}
\alias{solve_dampened_WLSj}
\title{solve_dampened_WLSj}
\usage{
solve_dampened_WLSj(S, B, goldStandard, j)
}
\description{
solve WLS given a dampening constant
}
\keyword{internal}
