% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_genes.R
\name{spat_fish_func}
\alias{spat_fish_func}
\title{spat_fish_func}
\usage{
spat_fish_func(gene, bin_matrix, spat_mat, calc_hub = F, hub_min_int = 3)
}
\description{
performs fisher exact test
}
\keyword{internal}
