% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_genes.R
\name{spat_fisher_exact}
\alias{spat_fisher_exact}
\alias{.spat_fish_func}
\alias{.spat_fish_func_dt}
\title{Fisher exact test}
\usage{
.spat_fish_func(feat, bin_matrix, spat_mat, calc_hub = FALSE, hub_min_int = 3)

.spat_fish_func_dt(
  bin_matrix_DTm,
  spat_netw_min,
  calc_hub = FALSE,
  hub_min_int = 3,
  cores = NA
)
}
\description{
Perform fisher exact test
}
\section{Functions}{
\itemize{
\item \code{.spat_fish_func_dt()}: data.table implementation

}}
\keyword{internal}
