% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_genes.R
\name{spat_odds_ratio}
\alias{spat_odds_ratio}
\alias{.spat_or_func}
\alias{.spat_or_func_dt}
\title{Spatial odds-ratio}
\usage{
.spat_or_func(feat, bin_matrix, spat_mat, calc_hub = FALSE, hub_min_int = 3)

.spat_or_func_dt(
  bin_matrix_DTm,
  spat_netw_min,
  calc_hub = F,
  hub_min_int = 3,
  cores = NA
)
}
\description{
calculate odds-ratio
}
\section{Functions}{
\itemize{
\item \code{.spat_or_func_dt()}: data.table implementation

}}
\keyword{internal}
