% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_genes.R
\name{spat_OR_func_DT}
\alias{spat_OR_func_DT}
\title{spat_OR_func_DT}
\usage{
spat_OR_func_DT(
  bin_matrix_DTm,
  spat_netw_min,
  calc_hub = F,
  hub_min_int = 3,
  cores = NA
)
}
\description{
calculate odds-ratio with data.table implementation
}
\keyword{internal}
