% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction.R
\name{spatCellCellcom}
\alias{spatCellCellcom}
\title{spatCellCellcom}
\usage{
spatCellCellcom(
  gobject,
  feat_type = NULL,
  spat_unit = NULL,
  spatial_network_name = "Delaunay_network",
  cluster_column = "cell_types",
  random_iter = 1000,
  feat_set_1,
  feat_set_2,
  gene_set_1 = NULL,
  gene_set_2 = NULL,
  log2FC_addendum = 0.1,
  min_observations = 2,
  detailed = FALSE,
  adjust_method = c("fdr", "bonferroni", "BH", "holm", "hochberg", "hommel", "BY",
    "none"),
  adjust_target = c("feats", "cells"),
  do_parallel = TRUE,
  cores = NA,
  set_seed = TRUE,
  seed_number = 1234,
  verbose = c("a little", "a lot", "none")
)
}
\arguments{
\item{gobject}{giotto object to use}

\item{feat_type}{feature type}

\item{spat_unit}{spatial unit}

\item{spatial_network_name}{spatial network to use for identifying
interacting cells}

\item{cluster_column}{cluster column with cell type information}

\item{random_iter}{number of iterations}

\item{feat_set_1}{first specific feature set from feature pairs}

\item{feat_set_2}{second specific feature set from feature pairs}

\item{gene_set_1}{deprecated, use feat_set_1}

\item{gene_set_2}{deprecated, use feat_set_2}

\item{log2FC_addendum}{addendum to add when calculating log2FC}

\item{min_observations}{minimum number of interactions needed to be
considered}

\item{detailed}{provide more detailed information
(random variance and z-score)}

\item{adjust_method}{which method to adjust p-values}

\item{adjust_target}{adjust multiple hypotheses at the cell or feature level}

\item{do_parallel}{run calculations in parallel with mclapply}

\item{cores}{number of cores to use if do_parallel = TRUE}

\item{set_seed}{set a seed for reproducibility}

\item{seed_number}{seed number}

\item{verbose}{verbose}
}
\value{
Cell-Cell communication scores for feature pairs based on spatial
interaction
}
\description{
Spatial Cell-Cell communication scores based on spatial
expression of interacting cells
}
\details{
Statistical framework to identify if pairs of genes
(such as ligand-receptor combinations)
are expressed at higher levels than expected based on a reshuffled null
distribution of feature expression values in cells that are spatially in
proximity to each other.
\itemize{
\item \strong{LR_comb:} Pair of ligand and receptor
\item \strong{lig_cell_type:} cell type to assess expression level of ligand
\item \strong{lig_expr:} average expression of ligand in lig_cell_type
\item \strong{ligand:} ligand name
\item \strong{rec_cell_type:} cell type to assess expression level of receptor
\item \strong{rec_expr:} average expression of receptor in rec_cell_type
\item \strong{receptor:} receptor name
\item \strong{LR_expr:} combined average ligand and receptor expression
\item \strong{lig_nr:} total number of cells from lig_cell_type that spatially
interact with cells from rec_cell_type
\item \strong{rec_nr:} total number of cells from rec_cell_type that spatially
interact with cells from lig_cell_type
\item \strong{rand_expr:} average combined ligand and receptor expression from
random spatial permutations
\item \strong{av_diff:} average difference between LR_expr and rand_expr over all
random spatial permutations
\item \strong{sd_diff:} (optional) standard deviation of the difference between
LR_expr and rand_expr over all random spatial permutations
\item \strong{z_score:} (optional) z-score
\item \strong{log2fc:} log2 fold-change (LR_expr/rand_expr)
\item \strong{pvalue:} p-value
\item \strong{LR_cell_comb:} cell type pair combination
\item \strong{p.adj:} adjusted p-value
\item \strong{PI:} significance score: log2fc * -log10(p.adj)
}
}
