% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visuals.R
\name{spatDeconvPlot}
\alias{spatDeconvPlot}
\title{spatDeconvPlot}
\usage{
spatDeconvPlot(
  gobject,
  feat_type = NULL,
  deconv_name = "DWLS",
  show_image = F,
  gimage = NULL,
  image_name = "image",
  spat_loc_name = NULL,
  sdimx = "sdimx",
  sdimy = "sdimy",
  cell_color_code = NULL,
  line_color = NA,
  radius = 10,
  alpha = 1,
  legend_text = 8,
  background_color = "white",
  title = NULL,
  axis_text = 8,
  axis_title = 8,
  coord_fix_ratio = TRUE,
  show_plot = NA,
  return_plot = NA,
  save_plot = NA,
  save_param = list(),
  default_save_name = "spatDeconvPlot"
)
}
\arguments{
\item{gobject}{giotto object}

\item{deconv_name}{name of deconvolution results to use}

\item{show_image}{show a tissue background image}

\item{gimage}{a giotto image}

\item{image_name}{name of a giotto image}

\item{sdimx}{x-axis dimension name (default = 'sdimx')}

\item{sdimy}{y-axis dimension name (default = 'sdimy')}

\item{cell_color_code}{named vector with colors}

\item{line_color}{color of line within pie charts}

\item{radius}{radios of pie charts}

\item{alpha}{alpha of pie charts}

\item{legend_text}{size of legend text}

\item{background_color}{color of plot background}

\item{title}{title for plot (default = deconv_name)}

\item{axis_text}{size of axis text}

\item{axis_title}{size of axis title}

\item{coord_fix_ratio}{fix ratio between x and y-axis}

\item{show_plot}{show plot}

\item{return_plot}{return ggplot object}

\item{save_plot}{directly save the plot [boolean]}

\item{save_param}{list of saving parameters from \code{\link{all_plots_save_function}}}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}
}
\value{
ggplot
}
\description{
Visualize cell type enrichment / deconvolution results in a scatterpie
}
\details{
Description of parameters.
}
