% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visuals.R
\name{spatDimCellPlot2D}
\alias{spatDimCellPlot2D}
\title{spatDimCellPlot2D}
\usage{
spatDimCellPlot2D(
  gobject,
  show_image = F,
  gimage = NULL,
  image_name = "image",
  plot_alignment = c("vertical", "horizontal"),
  spat_enr_names = NULL,
  cell_annotation_values = NULL,
  dim_reduction_to_use = "umap",
  dim_reduction_name = "umap",
  dim1_to_use = 1,
  dim2_to_use = 2,
  sdimx = "sdimx",
  sdimy = "sdimy",
  cell_color_gradient = c("blue", "white", "red"),
  gradient_midpoint = NULL,
  gradient_limits = NULL,
  select_cell_groups = NULL,
  select_cells = NULL,
  dim_point_shape = c("border", "no_border"),
  dim_point_size = 1,
  dim_point_alpha = 1,
  dim_point_border_col = "black",
  dim_point_border_stroke = 0.1,
  spat_point_shape = c("border", "no_border", "voronoi"),
  spat_point_size = 1,
  spat_point_alpha = 1,
  spat_point_border_col = "black",
  spat_point_border_stroke = 0.1,
  dim_show_cluster_center = F,
  dim_show_center_label = T,
  dim_center_point_size = 4,
  dim_center_point_border_col = "black",
  dim_center_point_border_stroke = 0.1,
  dim_label_size = 4,
  dim_label_fontface = "bold",
  spat_show_cluster_center = F,
  spat_show_center_label = F,
  spat_center_point_size = 4,
  spat_center_point_border_col = "black",
  spat_center_point_border_stroke = 0.1,
  spat_label_size = 4,
  spat_label_fontface = "bold",
  show_NN_network = F,
  nn_network_to_use = "sNN",
  nn_network_name = "sNN.pca",
  dim_edge_alpha = 0.5,
  spat_show_network = F,
  spatial_network_name = "Delaunay_network",
  spat_network_color = "red",
  spat_network_alpha = 0.5,
  spat_show_grid = F,
  spatial_grid_name = "spatial_grid",
  spat_grid_color = "green",
  show_other_cells = TRUE,
  other_cell_color = "grey",
  dim_other_point_size = 0.5,
  spat_other_point_size = 0.5,
  spat_other_cells_alpha = 0.5,
  show_legend = T,
  legend_text = 8,
  legend_symbol_size = 1,
  dim_background_color = "white",
  spat_background_color = "white",
  vor_border_color = "white",
  vor_max_radius = 200,
  vor_alpha = 1,
  axis_text = 8,
  axis_title = 8,
  coord_fix_ratio = NULL,
  cow_n_col = 2,
  cow_rel_h = 1,
  cow_rel_w = 1,
  cow_align = "h",
  show_plot = NA,
  return_plot = NA,
  save_plot = NA,
  save_param = list(),
  default_save_name = "spatDimCellPlot2D"
)
}
\arguments{
\item{gobject}{giotto object}

\item{show_image}{show a tissue background image}

\item{gimage}{a giotto image}

\item{image_name}{name of a giotto image}

\item{plot_alignment}{direction to align plot}

\item{spat_enr_names}{names of spatial enrichment results to include}

\item{cell_annotation_values}{numeric cell annotation columns}

\item{dim_reduction_to_use}{dimension reduction to use}

\item{dim_reduction_name}{dimension reduction name}

\item{dim1_to_use}{dimension to use on x-axis}

\item{dim2_to_use}{dimension to use on y-axis}

\item{sdimx}{= spatial dimension to use on x-axis}

\item{sdimy}{= spatial dimension to use on y-axis}

\item{cell_color_gradient}{vector with 3 colors for numeric data}

\item{gradient_midpoint}{midpoint for color gradient}

\item{gradient_limits}{vector with lower and upper limits}

\item{select_cell_groups}{select subset of cells/clusters based on cell_color parameter}

\item{select_cells}{select subset of cells based on cell IDs}

\item{dim_point_shape}{dim reduction points with border or not (border or no_border)}

\item{dim_point_size}{size of points in dim. reduction space}

\item{dim_point_alpha}{transparancy of dim. reduction points}

\item{dim_point_border_col}{border color of points in dim. reduction space}

\item{dim_point_border_stroke}{border stroke of points in dim. reduction space}

\item{spat_point_shape}{shape of points (border, no_border or voronoi)}

\item{spat_point_size}{size of spatial points}

\item{spat_point_alpha}{transparancy of spatial points}

\item{spat_point_border_col}{border color of spatial points}

\item{spat_point_border_stroke}{border stroke of spatial points}

\item{dim_show_cluster_center}{show the center of each cluster}

\item{dim_show_center_label}{provide a label for each cluster}

\item{dim_center_point_size}{size of the center point}

\item{dim_center_point_border_col}{border color of center point}

\item{dim_center_point_border_stroke}{stroke size of center point}

\item{dim_label_size}{size of the center label}

\item{dim_label_fontface}{font of the center label}

\item{spat_show_cluster_center}{show the center of each cluster}

\item{spat_show_center_label}{provide a label for each cluster}

\item{spat_center_point_size}{size of the spatial center points}

\item{spat_center_point_border_col}{border color of the spatial center points}

\item{spat_center_point_border_stroke}{stroke size of the spatial center points}

\item{spat_label_size}{size of the center label}

\item{spat_label_fontface}{font of the center label}

\item{show_NN_network}{show underlying NN network}

\item{nn_network_to_use}{type of NN network to use (kNN vs sNN)}

\item{nn_network_name}{name of NN network to use, if show_NN_network = TRUE}

\item{dim_edge_alpha}{column to use for alpha of the edges}

\item{spat_show_network}{show spatial network}

\item{spatial_network_name}{name of spatial network to use}

\item{spat_network_color}{color of spatial network}

\item{spat_network_alpha}{alpha of spatial network}

\item{spat_show_grid}{show spatial grid}

\item{spatial_grid_name}{name of spatial grid to use}

\item{spat_grid_color}{color of spatial grid}

\item{show_other_cells}{display not selected cells}

\item{other_cell_color}{color of not selected cells}

\item{dim_other_point_size}{size of not selected dim cells}

\item{spat_other_point_size}{size of not selected spat cells}

\item{spat_other_cells_alpha}{alpha of not selected spat cells}

\item{show_legend}{show legend}

\item{legend_text}{size of legend text}

\item{legend_symbol_size}{size of legend symbols}

\item{dim_background_color}{background color of points in dim. reduction space}

\item{spat_background_color}{background color of spatial points}

\item{vor_border_color}{border colorr for voronoi plot}

\item{vor_max_radius}{maximum radius for voronoi 'cells'}

\item{vor_alpha}{transparancy of voronoi 'cells'}

\item{axis_text}{size of axis text}

\item{axis_title}{size of axis title}

\item{coord_fix_ratio}{ratio for coordinates}

\item{cow_n_col}{cowplot param: how many columns}

\item{cow_rel_h}{cowplot param: relative height}

\item{cow_rel_w}{cowplot param: relative width}

\item{cow_align}{cowplot param: how to align}

\item{show_plot}{show plot}

\item{return_plot}{return ggplot object}

\item{save_plot}{directly save the plot [boolean]}

\item{save_param}{list of saving parameters, see \code{\link{showSaveParameters}}}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}
}
\value{
ggplot
}
\description{
Visualize numerical features of cells according to spatial AND dimension reduction coordinates in 2D
}
\details{
Description of parameters.
}
\examples{

data(mini_giotto_single_cell)

# combine all metadata
combineMetadata(mini_giotto_single_cell, spat_enr_names = 'cluster_metagene')

# visualize total expression information
spatDimCellPlot2D(mini_giotto_single_cell, cell_annotation_values = 'total_expr')

# visualize enrichment results
spatDimCellPlot2D(mini_giotto_single_cell,
                 spat_enr_names = 'cluster_metagene',
                 cell_annotation_values = c('1','2'))

}
\seealso{
Other spatial and dimension reduction cell annotation visualizations: 
\code{\link{spatDimCellPlot}()}
}
\concept{spatial and dimension reduction cell annotation visualizations}
