% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visuals.R
\name{spatDimGenePlot}
\alias{spatDimGenePlot}
\title{spatDimGenePlot}
\usage{
spatDimGenePlot(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=spatDimGenePlot2D]{spatDimGenePlot2D}}
  \describe{
    \item{\code{gobject}}{giotto object}
    \item{\code{genes}}{genes to show}
    \item{\code{default_save_name}}{default save name for saving, don't change, change save_name in save_param}
  }}
}
\value{
ggplot
}
\description{
Visualize cells according to spatial AND dimension reduction coordinates in ggplot mode
}
\details{
Description of parameters.
}
\seealso{
\code{\link{spatDimGenePlot3D}}

Other spatial and dimension reduction gene expression visualizations: 
\code{\link{spatDimGenePlot2D}()},
\code{\link{spatDimGenePlot3D}()}
}
\concept{spatial and dimension reduction gene expression visualizations}
