% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visuals.R
\name{spatDimPlot2D}
\alias{spatDimPlot2D}
\title{spatDimPlot2D}
\usage{
spatDimPlot2D(gobject, plot_alignment = c("vertical", "horizontal"),
  dim_reduction_to_use = "umap", dim_reduction_name = "umap",
  dim1_to_use = 1, dim2_to_use = 2, sdimx = "sdimx",
  sdimy = "sdimy", show_NN_network = F, nn_network_to_use = "sNN",
  network_name = "sNN.pca", show_cluster_center = F,
  show_center_label = T, center_point_size = 4, label_size = 4,
  label_fontface = "bold", cell_color = NULL, color_as_factor = T,
  cell_color_code = NULL, select_cell_groups = NULL,
  select_cells = NULL, show_other_cells = T,
  other_cell_color = "lightgrey", other_point_size = 1,
  dim_point_size = 1, dim_point_border_col = "black",
  dim_point_border_stroke = 0.1, nn_network_alpha = 0.05,
  show_spatial_network = F, spatial_network_name = "spatial_network",
  spatial_network_color = "blue", show_spatial_grid = F,
  spatial_grid_name = "spatial_grid", spatial_grid_color = "blue",
  spatial_point_size = 1, spatial_point_border_col = "black",
  spatial_point_border_stroke = 0.1, spatial_other_point_size = 1,
  spatial_network_alpha = 0.5, spatial_other_cells_alpha = 0.5,
  dim_other_point_size = 1, show_legend = T, show_plot = NA,
  return_plot = NA, save_plot = NA, save_param = list(),
  default_save_name = "spatDimPlot2D")
}
\arguments{
\item{gobject}{giotto object}

\item{plot_alignment}{direction to align plot}

\item{dim_reduction_to_use}{dimension reduction to use}

\item{dim_reduction_name}{dimension reduction name}

\item{dim1_to_use}{dimension to use on x-axis}

\item{dim2_to_use}{dimension to use on y-axis}

\item{sdimx}{= spatial dimension to use on x-axis}

\item{sdimy}{= spatial dimension to use on y-axis}

\item{show_NN_network}{show underlying NN network}

\item{nn_network_to_use}{type of NN network to use (kNN vs sNN)}

\item{network_name}{name of NN network to use, if show_NN_network = TRUE}

\item{show_cluster_center}{show the center of each cluster}

\item{show_center_label}{provide a label for each cluster}

\item{center_point_size}{size of the center point}

\item{label_size}{size of the center label}

\item{label_fontface}{font of the center label}

\item{cell_color}{color for cells (see details)}

\item{color_as_factor}{convert color column to factor}

\item{cell_color_code}{named vector with colors}

\item{select_cell_groups}{select subset of cells/clusters based on cell_color parameter}

\item{select_cells}{select subset of cells based on cell IDs}

\item{show_other_cells}{display not selected cells}

\item{other_cell_color}{color of not selected cells}

\item{other_point_size}{size of not selected cells}

\item{dim_point_size}{size of points in dim. reduction space}

\item{dim_point_border_col}{border color of points in dim. reduction space}

\item{dim_point_border_stroke}{border stroke of points in dim. reduction space}

\item{nn_network_alpha}{column to use for alpha of the edges}

\item{show_spatial_network}{show spatial network}

\item{spatial_network_name}{name of spatial network to use}

\item{spatial_network_color}{color of spatial network}

\item{show_spatial_grid}{show spatial grid}

\item{spatial_grid_name}{name of spatial grid to use}

\item{spatial_grid_color}{color of spatial grid}

\item{spatial_point_size}{size of spatial points}

\item{spatial_point_border_col}{border color of spatial points}

\item{spatial_point_border_stroke}{border stroke of spatial points}

\item{spatial_other_point_size}{size of not selected spatial points}

\item{spatial_network_alpha}{alpha of spatial network}

\item{spatial_other_cells_alpha}{alpha of not selected spatial points}

\item{dim_other_point_size}{size of not selected dim. reduction points}

\item{show_legend}{show legend}

\item{show_plot}{show plot}

\item{return_plot}{return ggplot object}

\item{save_plot}{directly save the plot [boolean]}

\item{save_param}{list of saving parameters from all_plots_save_function()}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}
}
\value{
ggplot
}
\description{
Visualize cells according to spatial AND dimension reduction coordinates in ggplot2 mode
}
\details{
Description of parameters.
}
\examples{
    spatDimPlot2D(gobject)
}
