% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visuals.R
\name{spatDimPlot3D}
\alias{spatDimPlot3D}
\title{spatDimPlot3D}
\usage{
spatDimPlot3D(
  gobject,
  plot_alignment = c("horizontal", "vertical"),
  dim_reduction_to_use = "umap",
  dim_reduction_name = "umap",
  dim1_to_use = 1,
  dim2_to_use = 2,
  dim3_to_use = 3,
  sdimx = "sdimx",
  sdimy = "sdimy",
  sdimz = "sdimz",
  spat_enr_names = NULL,
  show_NN_network = FALSE,
  nn_network_to_use = "sNN",
  network_name = "sNN.pca",
  nn_network_color = "lightgray",
  nn_network_alpha = 0.5,
  show_cluster_center = F,
  show_center_label = T,
  center_point_size = 4,
  label_size = 16,
  select_cell_groups = NULL,
  select_cells = NULL,
  show_other_cells = T,
  other_cell_color = "lightgrey",
  other_point_size = 1.5,
  cell_color = NULL,
  color_as_factor = T,
  cell_color_code = NULL,
  dim_point_size = 3,
  show_spatial_network = F,
  spatial_network_name = "Delaunay_network",
  spatial_network_color = "lightgray",
  spatial_network_alpha = 0.5,
  show_spatial_grid = F,
  spatial_grid_name = "spatial_grid",
  spatial_grid_color = NULL,
  spatial_grid_alpha = 0.5,
  spatial_point_size = 3,
  axis_scale = c("cube", "real", "custom"),
  custom_ratio = NULL,
  x_ticks = NULL,
  y_ticks = NULL,
  z_ticks = NULL,
  legend_text_size = 12,
  show_plot = NA,
  return_plot = NA,
  save_plot = NA,
  save_param = list(),
  default_save_name = "spatDimPlot3D"
)
}
\arguments{
\item{gobject}{giotto object}

\item{plot_alignment}{direction to align plot}

\item{dim_reduction_to_use}{dimension reduction to use}

\item{dim_reduction_name}{dimension reduction name}

\item{dim1_to_use}{dimension to use on x-axis}

\item{dim2_to_use}{dimension to use on y-axis}

\item{dim3_to_use}{dimension to use on z-axis}

\item{sdimx}{= spatial dimension to use on x-axis}

\item{sdimy}{= spatial dimension to use on y-axis}

\item{sdimz}{= spatial dimension to use on z-axis}

\item{spat_enr_names}{names of spatial enrichment results to include}

\item{show_NN_network}{show underlying NN network}

\item{nn_network_to_use}{type of NN network to use (kNN vs sNN)}

\item{network_name}{name of NN network to use, if show_NN_network = TRUE}

\item{nn_network_color}{color of nn network}

\item{nn_network_alpha}{column to use for alpha of the edges}

\item{show_cluster_center}{show the center of each cluster}

\item{show_center_label}{provide a label for each cluster}

\item{center_point_size}{size of the center point}

\item{label_size}{size of the center label}

\item{select_cell_groups}{select subset of cells/clusters based on cell_color parameter}

\item{select_cells}{select subset of cells based on cell IDs}

\item{show_other_cells}{display not selected cells}

\item{other_cell_color}{color of not selected cells}

\item{other_point_size}{size of not selected cells}

\item{cell_color}{color for cells (see details)}

\item{color_as_factor}{convert color column to factor}

\item{cell_color_code}{named vector with colors}

\item{dim_point_size}{size of points in dim. reduction space}

\item{show_spatial_network}{show spatial network}

\item{spatial_network_name}{name of spatial network to use}

\item{spatial_network_color}{color of spatial network}

\item{spatial_network_alpha}{alpha of spatial network}

\item{show_spatial_grid}{show spatial grid}

\item{spatial_grid_name}{name of spatial grid to use}

\item{spatial_grid_color}{color of spatial grid}

\item{spatial_grid_alpha}{alpha of spatial grid}

\item{spatial_point_size}{size of spatial points}

\item{axis_scale}{the way to scale the axis}

\item{custom_ratio}{customize the scale of the plot}

\item{x_ticks}{set the number of ticks on the x-axis}

\item{y_ticks}{set the number of ticks on the y-axis}

\item{z_ticks}{set the number of ticks on the z-axis}

\item{legend_text_size}{size of legend}

\item{show_plot}{show plot}

\item{return_plot}{return ggplot object}

\item{save_plot}{directly save the plot [boolean]}

\item{save_param}{list of saving parameters, see \code{\link{showSaveParameters}}}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}
}
\value{
plotly
}
\description{
Visualize cells according to spatial AND dimension reduction coordinates in plotly mode
}
\details{
Description of parameters.
}
\seealso{
Other spatial and dimension reduction visualizations: 
\code{\link{spatDimPlot2D}()},
\code{\link{spatDimPlot}()}
}
\concept{spatial and dimension reduction visualizations}
