% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{spatEnrObj-class}
\alias{spatEnrObj-class}
\title{S4 spatEnrObj Class}
\description{
Framework to store spatial enrichment results
}
\section{Slots}{

\describe{
\item{\code{name}}{name of enrichment object}

\item{\code{method}}{method used to perform spatial enrichment}

\item{\code{enrichDT}}{spatial enrichment data.table}

\item{\code{spat_unit}}{spatial unit}

\item{\code{feat_type}}{feature type}

\item{\code{provenance}}{provenance information}

\item{\code{misc}}{misc}
}}

