% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{spatFeatData-class}
\alias{spatFeatData-class}
\title{Superclass for classes that contain both spatial and feature data}
\description{
Superclass for classes that contain both spatial and feature data
}
