% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visuals.R
\name{spatGenePlot}
\alias{spatGenePlot}
\title{spatGenePlot}
\usage{
spatGenePlot(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=spatGenePlot2D]{spatGenePlot2D}}
  \describe{
    \item{\code{gobject}}{giotto object}
    \item{\code{genes}}{genes to show}
    \item{\code{default_save_name}}{default save name for saving, don't change, change save_name in save_param}
  }}
}
\value{
ggplot
}
\description{
Visualize cells and gene expression according to spatial coordinates
}
\details{
Description of parameters.
}
\seealso{
\code{\link{spatGenePlot3D}} and \code{\link{spatGenePlot2D}}

Other spatial gene expression visualizations: 
\code{\link{spatGenePlot2D}()},
\code{\link{spatGenePlot3D}()}
}
\concept{spatial gene expression visualizations}
