% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visuals.R
\name{spatGenePlot3D}
\alias{spatGenePlot3D}
\title{spatGenePlot3D}
\usage{
spatGenePlot3D(gobject, expression_values = c("normalized", "scaled",
  "custom"), genes, show_network = F, network_color = NULL,
  spatial_network_name = "spatial_network", edge_alpha = NULL,
  show_grid = F, cluster_column = NULL, select_cell_groups = NULL,
  select_cells = NULL, show_other_cells = T,
  other_cell_color = "lightgrey", other_point_size = 1,
  genes_high_color = NULL, genes_mid_color = "white",
  genes_low_color = "blue", spatial_grid_name = "spatial_grid",
  point_size = 2, show_legend = T, axis_scale = c("cube", "real",
  "custom"), custom_ratio = NULL, x_ticks = NULL, y_ticks = NULL,
  z_ticks = NULL, show_plot = NA, return_plot = NA, save_plot = NA,
  save_param = list(), default_save_name = "spatGenePlot3D")
}
\arguments{
\item{gobject}{giotto object}

\item{expression_values}{gene expression values to use}

\item{genes}{genes to show}

\item{show_network}{show underlying spatial network}

\item{network_color}{color of spatial network}

\item{spatial_network_name}{name of spatial network to use}

\item{show_grid}{show spatial grid}

\item{genes_high_color}{color represents high gene expression}

\item{genes_mid_color}{color represents middle gene expression}

\item{genes_low_color}{color represents low gene expression}

\item{spatial_grid_name}{name of spatial grid to use}

\item{point_size}{size of point (cell)}

\item{show_legend}{show legend}

\item{show_plot}{show plots}

\item{return_plot}{return ggplot object}

\item{save_plot}{directly save the plot [boolean]}

\item{save_param}{list of saving parameters from all_plots_save_function()}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}

\item{grid_color}{color of spatial grid}

\item{midpoint}{expression midpoint}

\item{scale_alpha_with_expression}{scale expression with ggplot alpha parameter}

\item{...}{parameters for cowplot::save_plot()}
}
\value{
ggplot
}
\description{
Visualize cells and gene expression according to spatial coordinates
}
\details{
Description of parameters.
}
\examples{
    spatGenePlot3D(gobject)
}
