% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_genes.R
\name{Spatial_AEH}
\alias{Spatial_AEH}
\title{Spatial_AEH}
\usage{
Spatial_AEH(gobject = NULL, results = NULL, pattern_num = 5,
  l = 1.05, show_AEH = T, sdimx = NULL, sdimy = NULL,
  point_size = 3, point_alpha = 1, low_color = "blue",
  mid_color = "white", high_color = "red", midpoint = 0)
}
\arguments{
\item{gobject}{Giotto object}

\item{results}{output from spatial_DE}

\item{pattern_num}{the number of gene expression patterns}

\item{show_AEH}{show AEH plot}
}
\value{
a list or a dataframe of SVs
}
\description{
calculate automatic expression histology with spatialDE method
}
\details{
Description.
}
\examples{
    Spatial_DE(gobject)
}
