% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_genes.R
\name{Spatial_DE}
\alias{Spatial_DE}
\title{Spatial_DE}
\usage{
Spatial_DE(gobject = NULL, show_plot = T, size = c(4, 2, 1),
  color = c("blue", "green", "red"), sig_alpha = 0.5,
  unsig_alpha = 0.5)
}
\arguments{
\item{gobject}{Giotto object}

\item{show_plot}{show FSV plot}
}
\value{
a list or a dataframe of SVs
}
\description{
calculate spatial varible genes with spatialDE method
}
\details{
Description.
}
\examples{
    Spatial_DE(gobject)
}
