% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_genes.R
\name{spatialAEH}
\alias{spatialAEH}
\title{spatialAEH}
\usage{
spatialAEH(
  gobject = NULL,
  SpatialDE_results = NULL,
  name_pattern = "AEH_patterns",
  expression_values = c("raw", "normalized", "scaled", "custom"),
  pattern_num = 6,
  l = 1.05,
  python_path = NULL,
  return_gobject = TRUE
)
}
\arguments{
\item{gobject}{Giotto object}

\item{SpatialDE_results}{results of \code{\link{spatialDE}} function}

\item{name_pattern}{name for the computed spatial patterns}

\item{expression_values}{gene expression values to use}

\item{pattern_num}{number of spatial patterns to look for}

\item{l}{lengthscale}

\item{python_path}{specify specific path to python if required}

\item{return_gobject}{show plot}
}
\value{
An updated giotto object
}
\description{
Compute spatial variable genes with spatialDE method
}
\details{
This function is a wrapper for the SpatialAEH method implemented in the ...
}
