% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_enrichment.R
\name{spatialAutoCor}
\alias{spatialAutoCor}
\alias{spatialAutoCorGlobal}
\alias{spatialAutoCorLocal}
\title{Spatial autocorrelation}
\usage{
spatialAutoCorGlobal(
  gobject = NULL,
  spat_unit = NULL,
  feat_type = NULL,
  feats = NULL,
  method = c("moran", "geary"),
  data_to_use = c("expression", "cell_meta"),
  expression_values = c("normalized", "scaled", "custom"),
  meta_cols = NULL,
  spatial_network_to_use = "kNN_network",
  wm_method = c("distance", "adjacency"),
  wm_name = "spat_weights",
  node_values = NULL,
  weight_matrix = NULL,
  test_method = c("none", "monte_carlo"),
  mc_nsim = 99,
  cor_name = NULL,
  return_gobject = FALSE,
  verbose = TRUE
)

spatialAutoCorLocal(
  gobject = NULL,
  spat_unit = NULL,
  feat_type = NULL,
  feats = NULL,
  method = c("moran", "gi", "gi*", "mean"),
  data_to_use = c("expression", "cell_meta"),
  expression_values = c("normalized", "scaled", "custom"),
  meta_cols = NULL,
  spatial_network_to_use = "kNN_network",
  wm_method = c("distance", "adjacency"),
  wm_name = "spat_weights",
  node_values = NULL,
  weight_matrix = NULL,
  test_method = c("none"),
  enrich_name = NULL,
  return_gobject = TRUE,
  output = c("spatEnrObj", "data.table"),
  verbose = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit}

\item{feat_type}{feature type}

\item{feats}{features (expression) on which to run autocorrelation.
(leaving as NULL means that all features will be tested)}

\item{method}{method of autocorrelation. See details (default = 'moran')}

\item{data_to_use}{if using data from gobject, whether to test using expression
('expression') or cell metadata ('cell_meta')}

\item{expression_values}{name of expression information to use}

\item{meta_cols}{columns in cell metadata to test}

\item{spatial_network_to_use}{spatial network to use}

\item{wm_method}{type of weight matrix to generate from spatial network if no
weight matrix is found attached to the spatial network}

\item{wm_name}{name of attached weight matrix to use}

\item{node_values}{alternative method of directly supplying a set of node values}

\item{weight_matrix}{alternative method of directly supplying a spatial weight
matrix}

\item{test_method}{method to test values for significance (default is no
testing)}

\item{mc_nsim}{when \code{test_method = 'monte_carlo'} this is number of simulations
to perform}

\item{cor_name}{name to assign the results in global autocorrelation output}

\item{return_gobject}{(default = FALSE) whether to return results appended to}

\item{verbose}{be verbose}

\item{enrich_name}{name to assign local autocorrelation spatial enrichment results}

\item{output}{'spatEnrObj' or 'data.table'
metadata in the giotto object or as a data.table}
}
\description{
Find spatial autocorrelation. Note that \code{spatialAutoCorGlobal}
will return values as a data.table instead of appending information to the gobject.
\code{spatialAutoCorLocal} will append the results as a spatial enrichment object
by default. \cr
If providing external data using either the \code{node_values} and/or \code{weight_matrix}
params, the order of values provided should be the same as the ordering of the
columns and rows of the weight matrix.
}
\details{
\strong{Global Methods:}
\itemize{
  \item{\emph{Moran's I} 'moran'}
  \item{\emph{Geary's C} 'geary'}
}

\strong{Local Methods:}
\itemize{
  \item{\emph{Local Moran's I} 'moran'}
  \item{\emph{Getis-Ord Gi} 'Gi'}
  \item{\emph{Getis-Ord Gi*} 'Gi*'}
  \item{\emph{Local mean} 'mean'}
}
}
\section{Functions}{
\itemize{
\item \code{spatialAutoCorGlobal()}: Global autocorrelation (single value returned)

\item \code{spatialAutoCorLocal()}: Local autocorrelation (values generated for each spatial ID)

}}
