% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_genes.R
\name{spatialDE}
\alias{spatialDE}
\title{spatialDE}
\usage{
spatialDE(
  gobject = NULL,
  expression_values = c("raw", "normalized", "scaled", "custom"),
  size = c(4, 2, 1),
  color = c("blue", "green", "red"),
  sig_alpha = 0.5,
  unsig_alpha = 0.5,
  python_path = NULL,
  show_plot = NA,
  return_plot = NA,
  save_plot = NA,
  save_param = list(),
  default_save_name = "SpatialDE"
)
}
\arguments{
\item{gobject}{Giotto object}

\item{expression_values}{gene expression values to use}

\item{size}{size of plot}

\item{color}{low/medium/high color scheme for plot}

\item{sig_alpha}{alpha value for significance}

\item{unsig_alpha}{alpha value for unsignificance}

\item{python_path}{specify specific path to python if required}

\item{show_plot}{show plot}

\item{return_plot}{return ggplot object}

\item{save_plot}{directly save the plot [boolean]}

\item{save_param}{list of saving parameters, see \code{\link{showSaveParameters}}}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}
}
\value{
a list of data.frames with results and plot (optional)
}
\description{
Compute spatial variable genes with spatialDE method
}
\details{
This function is a wrapper for the SpatialDE method implemented in the ...
}
