% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interoperability.R
\name{spatialExperimentToGiotto}
\alias{spatialExperimentToGiotto}
\title{Utility function to convert a SpatialExperiment object to a Giotto object}
\usage{
spatialExperimentToGiotto(spe)
}
\arguments{
\item{spe}{Input SpatialExperiment object to convert to a Giotto object}
}
\value{
Output Giotto object
}
\description{
Utility function to convert a SpatialExperiment object to a Giotto object
}
\examples{
library(SpatialExperiment)
example(read10xVisium, echo = FALSE)
spatialExperimentToGiotto(spe)
}
