% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{spatialGridObj-class}
\alias{spatialGridObj-class}
\title{S4 spatialGridObj Class}
\description{
Framework to store spatial grid
}
\details{
This is an S4 object that defines a spatial grid. The structure of the grid is stored as a
\code{data.table} within the \code{gridDT} slot and is defined by start and stop spatial
locations along the spatial axes. The \code{data.table} also includes names for each cell
of the grid and names for each of the spatial axis locations that make up the cell.
Grids can be annotated with both spatial and feature information
}
\section{Slots}{

\describe{
\item{\code{name}}{name of spatialGridObj}

\item{\code{method}}{method used to generate spatial grid}

\item{\code{parameters}}{additional method-specific parameters used during spatial grid generation}

\item{\code{gridDT}}{data.table holding the spatial grid information}

\item{\code{spat_unit}}{spatial unit}

\item{\code{feat_type}}{feature type}

\item{\code{provenance}}{origin of aggregated information (if applicable)}

\item{\code{misc}}{misc}
}}

