% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{spatialLocationsObj-class}
\alias{spatialLocationsObj-class}
\title{S4 spatialLocationsObj Class}
\description{
Framework to store spatial location information
}
\section{Slots}{

\describe{
\item{\code{name}}{name of spatialLocationsObj}

\item{\code{coordinates}}{data.table of spatial coordinates/locations}

\item{\code{spat_unit}}{spatial unit tag}

\item{\code{provenance}}{origin of aggregated information (if applicable)}
}}

