% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{spatialNetworkObj-class}
\alias{spatialNetworkObj-class}
\title{S4 spatialNetworkObj Class}
\description{
Framework to store spatial network information
}
\details{
The generic access operators work with the data within the \code{networkDT}
slot (filtered).
}
\section{Slots}{

\describe{
\item{\code{name}}{name of spatialNetworkObj}

\item{\code{method}}{method used to generate spatial network}

\item{\code{parameters}}{additional method-specific parameters used during spatial network generation}

\item{\code{outputObj}}{network geometry object}

\item{\code{networkDT}}{data.table of network connections, distances, and weightings}

\item{\code{networkDT_before_filter}}{unfiltered data.table  of network connections, distances, and weightings}

\item{\code{cellShapeObj}}{network cell shape information}

\item{\code{crossSectionObjects}}{crossSectionObjects (see \code{\link{create_crossSection_object}})}

\item{\code{spat_unit}}{spatial unit tag}

\item{\code{provenance}}{origin of aggregated information (if applicable)}

\item{\code{misc}}{misc}
}}

