% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_clusters.R
\name{spatialSplitCluster}
\alias{spatialSplitCluster}
\title{Split cluster annotations based on a spatial network}
\usage{
spatialSplitCluster(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  spatial_network_name = "Delaunay_network",
  cluster_col,
  split_clus_name = paste0(cluster_col, "_split"),
  missing_id_name = "not_connected",
  return_gobject = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit (e.g. "cell")}

\item{feat_type}{feature type (e.g. "rna", "dna", "protein")}

\item{spatial_network_name}{character. Name of spatial network to use}

\item{cluster_col}{character. Column in metadata containing original
clustering}

\item{split_clus_name}{character. Name to assign the split cluster results}

\item{return_gobject}{logical. Return giotto object}

\item{include_all_ids}{logical. Include all ids, including vertex ids not
found in the spatial network}
}
\value{
giotto object with cluster annotations
}
\description{
Split cluster annotations based on a spatial network
}
\examples{
g <- GiottoData::loadGiottoMini("vizgen")
activeSpatUnit(g) <- "aggregate"
spatPlot2D(g, cell_color = "leiden_clus")

g <- spatialSplitCluster(g,
    cluster_col = "leiden_clus",
    split_clus_name = "new"
)
# don't show legend since there are too many categories generated
spatPlot2D(g, cell_color = "new", show_legend = FALSE)
}
