% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_structures.R
\name{spatialWeightMatrix}
\alias{spatialWeightMatrix}
\title{Create a spatial weight matrix}
\usage{
spatialWeightMatrix(
  gobject,
  spat_unit = NULL,
  spatial_network_to_use = "kNN_network",
  method = c("distance", "adjacency"),
  wm_name = "spat_weights",
  return_gobject = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit}

\item{spatial_network_to_use}{spatial network information to use}

\item{method}{type of weighted matrix to generate. See details}

\item{wm_name}{name to assign the weight matrix values}

\item{return_gobject}{(default = TRUE) whether to return as the giotto object
with attached results or the bare weighted matrix}

\item{verbose}{be verbose}
}
\description{
Generate spatial weight matrix based on the strength of spatial
interactions between nodes. Requires spatial networks to be first generated.
}
\details{
\itemize{
  \item{\code{"distance"} method is calculated using 1/(1+distance) to create an inverse
weighting based on the distance between nodes.}
  \item{\code{"adjacency"} method is a binary matrix with 1 signifying that two nodes
are connected in the spatial network and 0 indicating that they are not.}
}
}
