% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_in_situ_visuals.R
\name{spatInSituPlotHex}
\alias{spatInSituPlotHex}
\title{spatInSituPlotHex}
\usage{
spatInSituPlotHex(
  gobject,
  feats = NULL,
  feat_type = "rna",
  sdimx = "x",
  sdimy = "y",
  bins = 10,
  alpha = 0.5,
  show_polygon = TRUE,
  polygon_feat_type = "cell",
  polygon_color = "black",
  polygon_fill = NULL,
  polygon_fill_as_factor = NULL,
  polygon_alpha = 0.5,
  polygon_size = 0.5,
  coord_fix_ratio = NULL,
  axis_text = 8,
  axis_title = 8,
  legend_text = 6,
  background_color = "white",
  cow_n_col = 2,
  cow_rel_h = 1,
  cow_rel_w = 1,
  cow_align = "h",
  show_plot = NA,
  return_plot = NA,
  save_plot = NA,
  save_param = list(),
  default_save_name = "spatInSituPlotHex"
)
}
\arguments{
\item{gobject}{giotto object}

\item{feats}{features to plot}

\item{feat_type}{feature types of the feats}

\item{sdimx}{spatial dimension x}

\item{sdimy}{spatial dimension y}

\item{bins}{number of hexbins in one direction}

\item{alpha}{alpha of hexbin plot}

\item{show_polygon}{overlay polygon information (cell shape)}

\item{polygon_feat_type}{feature type associated with polygon information}

\item{polygon_color}{color for polygon border}

\item{polygon_fill}{fill color or column for polygon}

\item{polygon_fill_as_factor}{is fill color a factor}

\item{polygon_alpha}{alpha of polygon}

\item{polygon_size}{size of polygon border}

\item{coord_fix_ratio}{fix ratio between x and y-axis}

\item{axis_text}{axis text size}

\item{axis_title}{title text size}

\item{legend_text}{legend text size}

\item{background_color}{background color}

\item{cow_n_col}{cowplot param: how many columns}

\item{cow_rel_h}{cowplot param: relative height}

\item{cow_rel_w}{cowplot param: relative width}

\item{cow_align}{cowplot param: how to align}

\item{show_plot}{show plots}

\item{return_plot}{return ggplot object}

\item{save_plot}{directly save the plot [boolean]}

\item{save_param}{list of saving parameters, see \code{\link{showSaveParameters}}}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}
}
\value{
ggplot
}
\description{
Function to plot hexbins for features for multiple modalities at the spatial in situ level
}
\details{
TODO
}
\seealso{
Other In Situ visualizations: 
\code{\link{spatInSituPlotDensity}()},
\code{\link{spatInSituPlotPoints}()}
}
\concept{In Situ visualizations}
