% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_structures.R
\name{spatNetwDistributionsDistance}
\alias{spatNetwDistributionsDistance}
\title{spatNetwDistributionsDistance}
\usage{
spatNetwDistributionsDistance(
  gobject,
  spatial_network_name = "spatial_network",
  hist_bins = 30,
  test_distance_limit = NULL,
  ncol = 1,
  show_plot = NA,
  return_plot = NA,
  save_plot = NA,
  save_param = list(),
  default_save_name = "spatNetwDistributionsDistance"
)
}
\arguments{
\item{gobject}{Giotto object}

\item{spatial_network_name}{name of spatial network}

\item{hist_bins}{number of binds to use for the histogram}

\item{test_distance_limit}{effect of different distance threshold on k-neighbors}

\item{ncol}{number of columns to visualize the histograms in}

\item{show_plot}{show plot}

\item{return_plot}{return ggplot object}

\item{save_plot}{directly save the plot [boolean]}

\item{save_param}{list of saving parameters from \code{\link{all_plots_save_function}}}

\item{default_save_name}{default save name for saving, alternatively change save_name in save_param}
}
\value{
ggplot plot
}
\description{
This function return histograms displaying the distance distribution for each spatial k-neighbor
}
