% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_structures.R
\name{spatNetwDistributionsKneighbors}
\alias{spatNetwDistributionsKneighbors}
\title{spatNetwDistributionsKneighbors}
\usage{
spatNetwDistributionsKneighbors(
  gobject,
  spatial_network_name = "spatial_network",
  hist_bins = 30,
  show_plot = NA,
  return_plot = NA,
  save_plot = NA,
  save_param = list(),
  default_save_name = "spatNetwDistributionsKneighbors"
)
}
\arguments{
\item{gobject}{Giotto object}

\item{spatial_network_name}{name of spatial network}

\item{hist_bins}{number of binds to use for the histogram}

\item{show_plot}{show plot}

\item{return_plot}{return ggplot object}

\item{save_plot}{directly save the plot [boolean]}

\item{save_param}{list of saving parameters from \code{\link{all_plots_save_function}}}

\item{default_save_name}{default save name for saving, alternatively change save_name in save_param}
}
\value{
ggplot plot
}
\description{
This function returns a histogram displaying the number of k-neighbors distribution for each cell
}
