% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visuals.R
\name{spatPlot}
\alias{spatPlot}
\title{spatPlot}
\usage{
spatPlot(gobject, sdimx = "sdimx", sdimy = "sdimy", point_size = 3,
  point_border_col = "black", point_border_stroke = 0.1,
  cell_color = NULL, cell_color_code = NULL, color_as_factor = T,
  select_cell_groups = NULL, select_cells = NULL,
  show_other_cells = T, other_cell_color = "lightgrey",
  other_point_size = 1, show_network = F, network_color = NULL,
  network_alpha = 1, other_cells_alpha = 0.1,
  spatial_network_name = "spatial_network", show_grid = F,
  grid_color = NULL, spatial_grid_name = "spatial_grid",
  coord_fix_ratio = 0.6, title = "", show_legend = T,
  show_plot = NA, return_plot = NA, save_plot = NA,
  save_param = list(), default_save_name = "spatPlot")
}
\arguments{
\item{gobject}{giotto object}

\item{sdimx}{x-axis dimension name (default = 'sdimx')}

\item{sdimy}{y-axis dimension name (default = 'sdimy')}

\item{point_size}{size of point (cell)}

\item{point_border_col}{color of border around points}

\item{point_border_stroke}{stroke size of border around points}

\item{cell_color}{color for cells (see details)}

\item{cell_color_code}{named vector with colors}

\item{color_as_factor}{convert color column to factor}

\item{select_cell_groups}{select subset of cells/clusters based on cell_color parameter}

\item{select_cells}{select subset of cells based on cell IDs}

\item{show_other_cells}{display not selected cells}

\item{other_cell_color}{color of not selected cells}

\item{show_network}{show underlying spatial network}

\item{network_color}{color of spatial network}

\item{spatial_network_name}{name of spatial network to use}

\item{show_grid}{show spatial grid}

\item{grid_color}{color of spatial grid}

\item{spatial_grid_name}{name of spatial grid to use}

\item{coord_fix_ratio}{fix ratio between x and y-axis}

\item{title}{title of plot}

\item{show_legend}{show legend}

\item{show_plot}{show plot}

\item{return_plot}{return ggplot object}

\item{save_plot}{directly save the plot [boolean]}

\item{save_param}{list of saving parameters from all_plots_save_function()}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}
}
\value{
ggplot
}
\description{
Visualize cells according to spatial coordinates
}
\details{
Description of parameters.
}
\examples{
    spatPlot(gobject)
}
\seealso{
\code{\link{spatPlot3D}}
}
