% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visuals.R
\name{spatPlot2D_single}
\alias{spatPlot2D_single}
\title{spatPlot2D_single}
\usage{
spatPlot2D_single(
  gobject,
  show_image = F,
  gimage = NULL,
  image_name = "image",
  sdimx = "sdimx",
  sdimy = "sdimy",
  spat_enr_names = NULL,
  cell_color = NULL,
  color_as_factor = T,
  cell_color_code = NULL,
  cell_color_gradient = c("blue", "white", "red"),
  gradient_midpoint = NULL,
  gradient_limits = NULL,
  select_cell_groups = NULL,
  select_cells = NULL,
  point_shape = c("border", "no_border", "voronoi"),
  point_size = 3,
  point_alpha = 1,
  point_border_col = "black",
  point_border_stroke = 0.1,
  show_cluster_center = F,
  show_center_label = F,
  center_point_size = 4,
  center_point_border_col = "black",
  center_point_border_stroke = 0.1,
  label_size = 4,
  label_fontface = "bold",
  show_network = F,
  spatial_network_name = "Delaunay_network",
  network_color = NULL,
  network_alpha = 1,
  show_grid = F,
  spatial_grid_name = "spatial_grid",
  grid_color = NULL,
  show_other_cells = T,
  other_cell_color = "lightgrey",
  other_point_size = 1,
  other_cells_alpha = 0.1,
  coord_fix_ratio = NULL,
  title = NULL,
  show_legend = T,
  legend_text = 8,
  legend_symbol_size = 1,
  background_color = "white",
  vor_border_color = "white",
  vor_max_radius = 200,
  vor_alpha = 1,
  axis_text = 8,
  axis_title = 8,
  show_plot = NA,
  return_plot = NA,
  save_plot = NA,
  save_param = list(),
  default_save_name = "spatPlot2D_single"
)
}
\arguments{
\item{gobject}{giotto object}

\item{show_image}{show a tissue background image}

\item{gimage}{a giotto image}

\item{image_name}{name of a giotto image}

\item{sdimx}{x-axis dimension name (default = 'sdimx')}

\item{sdimy}{y-axis dimension name (default = 'sdimy')}

\item{spat_enr_names}{names of spatial enrichment results to include}

\item{cell_color}{color for cells (see details)}

\item{color_as_factor}{convert color column to factor}

\item{cell_color_code}{named vector with colors}

\item{cell_color_gradient}{vector with 3 colors for numeric data}

\item{gradient_midpoint}{midpoint for color gradient}

\item{gradient_limits}{vector with lower and upper limits}

\item{select_cell_groups}{select subset of cells/clusters based on cell_color parameter}

\item{select_cells}{select subset of cells based on cell IDs}

\item{point_shape}{shape of points (border, no_border or voronoi)}

\item{point_size}{size of point (cell)}

\item{point_alpha}{transparancy of point}

\item{point_border_col}{color of border around points}

\item{point_border_stroke}{stroke size of border around points}

\item{show_cluster_center}{plot center of selected clusters}

\item{show_center_label}{plot label of selected clusters}

\item{center_point_size}{size of center points}

\item{label_size}{size of labels}

\item{label_fontface}{font of labels}

\item{show_network}{show underlying spatial network}

\item{spatial_network_name}{name of spatial network to use}

\item{network_color}{color of spatial network}

\item{network_alpha}{alpha of spatial network}

\item{show_grid}{show spatial grid}

\item{spatial_grid_name}{name of spatial grid to use}

\item{grid_color}{color of spatial grid}

\item{show_other_cells}{display not selected cells}

\item{other_cell_color}{color of not selected cells}

\item{other_point_size}{point size of not selected cells}

\item{other_cells_alpha}{alpha of not selected cells}

\item{coord_fix_ratio}{fix ratio between x and y-axis}

\item{title}{title of plot}

\item{show_legend}{show legend}

\item{legend_text}{size of legend text}

\item{legend_symbol_size}{size of legend symbols}

\item{background_color}{color of plot background}

\item{vor_border_color}{border colorr for voronoi plot}

\item{vor_max_radius}{maximum radius for voronoi 'cells'}

\item{vor_alpha}{transparancy of voronoi 'cells'}

\item{axis_text}{size of axis text}

\item{axis_title}{size of axis title}

\item{show_plot}{show plot}

\item{return_plot}{return ggplot object}

\item{save_plot}{directly save the plot [boolean]}

\item{save_param}{list of saving parameters from \code{\link{all_plots_save_function}}}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}
}
\value{
ggplot
}
\description{
Visualize cells according to spatial coordinates
}
\details{
Description of parameters.
}
\seealso{
\code{\link{spatPlot3D}}
}
\keyword{internal}
