% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visuals.R
\name{spatPlot3D}
\alias{spatPlot3D}
\title{spatPlot3D}
\usage{
spatPlot3D(
  gobject,
  sdimx = "sdimx",
  sdimy = "sdimy",
  sdimz = "sdimz",
  spat_enr_names = NULL,
  point_size = 3,
  cell_color = NULL,
  cell_color_code = NULL,
  select_cell_groups = NULL,
  select_cells = NULL,
  show_other_cells = T,
  other_cell_color = "lightgrey",
  other_point_size = 0.5,
  other_cell_alpha = 0.5,
  show_network = F,
  spatial_network_name = "Delaunay_network",
  network_color = NULL,
  network_alpha = 1,
  show_grid = F,
  spatial_grid_name = "spatial_grid",
  grid_color = NULL,
  grid_alpha = 1,
  title = "",
  show_legend = T,
  axis_scale = c("cube", "real", "custom"),
  custom_ratio = NULL,
  x_ticks = NULL,
  y_ticks = NULL,
  z_ticks = NULL,
  show_plot = NA,
  return_plot = NA,
  save_plot = NA,
  save_param = list(),
  default_save_name = "spat3D"
)
}
\arguments{
\item{gobject}{giotto object}

\item{sdimx}{x-axis dimension name (default = 'sdimx')}

\item{sdimy}{y-axis dimension name (default = 'sdimy')}

\item{sdimz}{z-axis dimension name (default = 'sdimy')}

\item{spat_enr_names}{names of spatial enrichment results to include}

\item{point_size}{size of point (cell)}

\item{cell_color}{color for cells (see details)}

\item{cell_color_code}{named vector with colors}

\item{select_cell_groups}{select subset of cells/clusters based on cell_color parameter}

\item{select_cells}{select subset of cells based on cell IDs}

\item{show_other_cells}{display not selected cells}

\item{other_cell_color}{color of not selected cells}

\item{other_point_size}{size of not selected cells}

\item{other_cell_alpha}{alpha of not selected cells}

\item{show_network}{show underlying spatial network}

\item{spatial_network_name}{name of spatial network to use}

\item{network_color}{color of spatial network}

\item{network_alpha}{opacity of spatial network}

\item{show_grid}{show spatial grid}

\item{spatial_grid_name}{name of spatial grid to use}

\item{grid_color}{color of spatial grid}

\item{grid_alpha}{opacity of spatial grid}

\item{title}{title of plot}

\item{show_legend}{show legend}

\item{axis_scale}{the way to scale the axis}

\item{custom_ratio}{customize the scale of the plot}

\item{x_ticks}{set the number of ticks on the x-axis}

\item{y_ticks}{set the number of ticks on the y-axis}

\item{z_ticks}{set the number of ticks on the z-axis}

\item{show_plot}{show plot}

\item{return_plot}{return ggplot object}

\item{save_plot}{directly save the plot [boolean]}

\item{save_param}{list of saving parameters, see \code{\link{showSaveParameters}}}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}
}
\value{
ggplot
}
\description{
Visualize cells according to spatial coordinates
}
\seealso{
Other spatial visualizations: 
\code{\link{spatPlot2D}()},
\code{\link{spatPlot}()}
}
\concept{spatial visualizations}
