% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visuals.R
\name{spatPlot_2D_plotly}
\alias{spatPlot_2D_plotly}
\title{spatPlot_2D_plotly}
\usage{
spatPlot_2D_plotly(
  gobject,
  sdimx = NULL,
  sdimy = NULL,
  spat_enr_names = NULL,
  point_size = 3,
  cell_color = NULL,
  cell_color_code = NULL,
  color_as_factor = T,
  select_cell_groups = NULL,
  select_cells = NULL,
  show_other_cells = T,
  other_cell_color = "lightgrey",
  other_point_size = 0.5,
  show_network = FALSE,
  spatial_network_name = "spatial_network",
  network_color = "lightgray",
  network_alpha = 1,
  other_cell_alpha = 0.5,
  show_grid = FALSE,
  spatial_grid_name = "spatial_grid",
  grid_color = NULL,
  grid_alpha = 1,
  show_legend = T,
  axis_scale = c("cube", "real", "custom"),
  custom_ratio = NULL,
  x_ticks = NULL,
  y_ticks = NULL,
  show_plot = F
)
}
\value{
plotly object
}
\description{
Visualize cells at their 2D spatial locations with plotly
}
\keyword{internal}
