% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visuals.R
\name{spatPlotInteractive}
\alias{spatPlotInteractive}
\title{Select interactive polygons}
\usage{
spatPlotInteractive(p)
}
\arguments{
\item{p}{An image or plot to draw polygons on}
}
\value{
A `data.table` stored in polygon_results containing x,y coordinates from the plotted polygons.
}
\description{
Select interactive polygons
}
\examples{
\dontrun{
# Using a ggplot2 plot
libray(ggplot2)
df <- data.frame(x = 1:5, y = 1:5)
my_plot <- ggplot(df, aes(x,y)) + geom_point()
spatPlotInteractive(my_plot)

# Using a terra rast image
library(terra)
r = rast(system.file("ex/elev.tif", package="terra"))
spatPlotInteractive(r)

# Using an image contained in Giotto object
my_spatPlot <- spatPlot2D(gobject = my_giotto_object,
                          show_image = TRUE,
                          point_alpha = 0.75,
                          save_plot = FALSE)
spatPlotInteractive(my_spatPlot)
}
}
