% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_structures.R
\name{spatVector_to_dt}
\alias{spatVector_to_dt}
\title{Convert spatVector to data.table}
\usage{
spatVector_to_dt(spatvector, include_values = TRUE)
}
\description{
convert spatVector to data.table
}
\keyword{internal}
