% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_structures.R
\name{spatVector_to_dt2}
\alias{spatVector_to_dt2}
\title{Convert spatVector to data.table}
\usage{
spatVector_to_dt2(spatvector, include_values = TRUE)
}
\description{
convert spatVector to data.table. Faster and more barebones
}
\keyword{internal}
