% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spdep.R
\name{spdepAutoCorr}
\alias{spdepAutoCorr}
\title{Compute spatial auto correlation using spdep}
\usage{
spdepAutoCorr(
  gobject,
  method = c("geary.test", "lee.test", "lm.morantest", "moran.test"),
  spat_unit = NULL,
  feat_type = NULL,
  expression_values = "normalized",
  spatial_network_to_use = "spatial_network",
  return_gobject = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{gobject}{Input a Giotto object.}

\item{method}{Specify a method name to compute auto correlation.
Available methods include \code{"geary.test", "lee.test", "lm.morantest","moran.test"}.}

\item{spat_unit}{spatial unit}

\item{feat_type}{feature type}

\item{expression_values}{expression values to use, default = normalized}

\item{spatial_network_to_use}{spatial network to use, default = spatial_network}

\item{verbose}{be verbose}
}
\value{
A data table with computed values for each feature.
}
\description{
Compute spatial auto correlation using spdep
}
\examples{
\dontrun{
library(Giotto)
library(GiottoData)
library(spdep)
g = loadGiottoMini(dataset = "visium")
spdepAutoCorr(gobject = g, method = "moran.test")
}
}
