% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_manipulation.R
\name{spin-generic}
\alias{spin-generic}
\alias{spin,giottoPolygon-method}
\alias{spin,giottoPoints-method}
\alias{spin,spatLocsObj-method}
\title{Spin an object}
\usage{
\S4method{spin}{giottoPolygon}(x, angle, x0 = NULL, y0 = NULL)

\S4method{spin}{giottoPoints}(x, angle, x0 = NULL, y0 = NULL)

\S4method{spin}{spatLocsObj}(
  x,
  angle = NULL,
  x0 = NULL,
  y0 = NULL,
  z0 = NULL,
  xy_angle = NULL,
  zy_angle = NULL,
  xz_angle = NULL
)
}
\arguments{
\item{x}{object}

\item{angle}{numeric. Angle of rotation in degrees}

\item{x0}{numeric. x-coordinate of the center of rotation. Defaults to center x val if not given.}

\item{y0}{numeric. y-coordinate of the center of rotation. Defaults to center y val if not given.}

\item{z0}{spatLocsObj specific. Numeric. z-coordinate of the center of rotation.
Depending on if z data is present, defaults to either 0 or center z val if not given.}

\item{xy_angle}{spatLocsObj specific. xy plane rotation in degrees.
Overrides angle param}

\item{zy_angle}{spatLocsObj specific. zy plane rotation}

\item{xz_angle}{spatLocsObj specific. xz plane rotation}
}
\description{
Spin (rotate) an object spatially (limited to xy rotations)
}
\section{Functions}{
\itemize{
\item \code{spin(giottoPolygon)}: Spin a giottoPolygon object

\item \code{spin(giottoPoints)}: Spin a giottoPoints object

\item \code{spin(spatLocsObj)}: Spin a spatLocsObj

}}
