% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_structures.R
\name{spline_poly}
\alias{spline_poly}
\title{Convert spline to polygon}
\usage{
spline_poly(xy, vertices = 20, k = 3, ...)
}
\arguments{
\item{xy}{xy}

\item{vertices}{vertices}

\item{k}{k}

\item{...}{additional params to pass}
}
\description{
spline polynomial to smooth polygon
}
\keyword{internal}
