% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{split_dendrogram_in_two}
\alias{split_dendrogram_in_two}
\title{split_dendrogram_in_two}
\usage{
split_dendrogram_in_two(dend)
}
\arguments{
\item{dend}{dendrogram object}
}
\value{
list of two dendrograms and height of node
}
\description{
Merge selected clusters based on pairwise correlation scores and size of cluster.
}
\keyword{internal}
