% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_enrichment.R
\name{spot_deconvolution}
\alias{spot_deconvolution}
\title{spot_deconvolution}
\usage{
spot_deconvolution(expr, cluster_info, ct_exp, binary_matrix)
}
\description{
Rui to fill in
}
\keyword{internal}
