% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registration.R
\name{stack_spatlocs}
\alias{stack_spatlocs}
\title{stack_spatlocs}
\usage{
stack_spatlocs(spatlocs_list, z_vals)
}
\arguments{
\item{spatlocs_list}{list of spatial locations}

\item{z_vals}{z values to use for z stacking spatial locations list}
}
\description{
Adds z values and combines list of spatlocs into single z-stacked spatloc table
}
\keyword{internal}
