% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registration.R
\name{stackSpatLocs}
\alias{stackSpatLocs}
\title{stackSpatLocs}
\usage{
stackSpatLocs(spatlocs_list, z_list)
}
\arguments{
\item{spatlocs_list}{list of spatial locations}

\item{z_list}{z values to use for z stacking spatial locations list}
}
\description{
Adds z values and combines list of spatlocs into single z-stacked spatloc table
}
