% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{standardise_db}
\alias{standardise_db}
\title{standardise_db}
\usage{
standardise_db(x, center = TRUE, scale = TRUE, by = c("row", "col"), ...)
}
\arguments{
\item{x}{dbMatrix}

\item{center}{center data}

\item{scale}{scale data}

\item{by}{normalize by 'row' or 'col'}

\item{...}{additional params to pass}
}
\value{
standardized database table
}
\description{
standardizes a dbMatrix
}
\keyword{internal}
