% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R, R/utilities.R
\name{standardise_flex}
\alias{standardise_flex}
\title{standardise_flex}
\usage{
standardise_flex(x, center = TRUE, scale = TRUE)

standardise_flex(x, center = TRUE, scale = TRUE)
}
\arguments{
\item{x}{matrix}

\item{center}{center data}

\item{scale}{scale data}
}
\value{
standardized matrix
}
\description{
standardizes a matrix
}
\keyword{internal}
