% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.R
\name{standardise_giotto}
\alias{standardise_giotto}
\title{standardise_giotto}
\usage{
standardise_giotto(x, center = TRUE, scale = TRUE)
}
\arguments{
\item{x}{matrix}

\item{center}{center data}

\item{scale}{scale data}
}
\value{
standardized matrix
}
\description{
standardises a matrix
}
\keyword{internal}
