% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{stitchGiottoLargeImage}
\alias{stitchGiottoLargeImage}
\title{stitchGiottoLargeImage}
\usage{
stitchGiottoLargeImage(
  largeImage_list = NULL,
  gobject_list = NULL,
  largeImage_nameList = NULL,
  FOV_positions = NULL,
  FOV_xcol = NULL,
  FOV_ycol = NULL,
  FOV_inverty = FALSE,
  method = c("mosaic", "merge"),
  round_positions = FALSE,
  filename = NULL,
  dataType = NULL,
  fileType = NULL,
  dryRun = TRUE,
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{largeImage_list}{list of giottoLargeImage objects}

\item{gobject_list}{list of gobjects containing giottoLargeImages}

\item{largeImage_nameList}{list of names of giottoLargeImages within gobjects}

\item{FOV_positions}{dataframe of FOV positions. Values (if any) are directly added to current image mapping}

\item{FOV_xcol}{column name for FOV position x values}

\item{FOV_ycol}{column name for FOV position y values}

\item{FOV_inverty}{make FOV y position values negative}

\item{method}{method of stitching images (mosaic: average overlapping area values, merge: values get priority by order given)}

\item{round_positions}{round image positions. May be necessary to run.}

\item{filename}{file name to write the stitched image to. Defaults to "save_dir/stitch.tif" if save_dir param is found in first gobject Giotto instructions}

\item{dataType}{(optional) values for dataType are "INT1U", "INT2U", "INT2S", "INT4U", "INT4S", "FLT4S", "FLT8S". The first three letters indicate whether the dataType is integer (whole numbers) of a real number (decimal numbers), the fourth character indicates the number of bytes used (allowing for large numbers and/or more precision), and the "S" or "U" indicate whether the values are signed (both negative and positive) or unsigned (positive values only).}

\item{fileType}{(optional) image format (e.g. .tif) If not given, defaults to format given in the filename}

\item{dryRun}{display placeholder bounding boxes where images will be stitched}

\item{overwrite}{overwrite if filename already used. Defaults to TRUE}

\item{verbose}{be verbose}
}
\value{
largeGiottoImage object of stitched image
}
\description{
stitch multiple giottoLargeImages into single giottoLargeImage. Time consuming & save location recommended.
}
