% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_help.R
\name{stitchTileCoordinates}
\alias{stitchTileCoordinates}
\title{stitchTileCoordinates}
\usage{
stitchTileCoordinates(location_file, Xtilespan, Ytilespan)
}
\arguments{
\item{location_file}{location dataframe with X and Y coordinates}

\item{Xtilespan}{numerical value specifying the width of each tile}

\item{Ytilespan}{numerical value specifying the height of each tile}
}
\description{
Helper function to stitch tile coordinates together to form one complete picture
}
