% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{subset_cell_metadata}
\alias{subset_cell_metadata}
\title{Subset cell metadata}
\usage{
subset_cell_metadata(
  gobject,
  feat_type = NULL,
  cell_ids,
  spat_unit,
  all_feat_types = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{feat_type}{feature type (e.g. 'rna', 'protein')}

\item{cell_ids}{cell ids to keep}

\item{spat_unit}{spatial unit (e.g. 'cell')}

\item{all_feat_types}{(boolean) applies subset operation across the whole gobject
(ALL feature types), ignoring the \code{feat_type} input param. Defaults to TRUE.}
}
\description{
Subset cell metadata from giotto object
}
\keyword{internal}
