% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{subset_expression_data}
\alias{subset_expression_data}
\title{Subset expression data}
\usage{
subset_expression_data(
  gobject,
  filter_bool_feats,
  filter_bool_cells,
  feat_type,
  spat_unit
)
}
\description{
Subset expression data from giotto object
}
\keyword{internal}
