% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{subset_feature_info_data}
\alias{subset_feature_info_data}
\title{Subset feature info data}
\usage{
subset_feature_info_data(
  feat_info,
  feat_ids,
  feat_type = "rna",
  x_min = NULL,
  x_max = NULL,
  y_min = NULL,
  y_max = NULL
)
}
\description{
Subset  all spatial feature (points) data
}
\keyword{internal}
