% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{subset_feature_metadata}
\alias{subset_feature_metadata}
\title{Subset feature metadata}
\usage{
subset_feature_metadata(gobject, feat_type, spat_unit, filter_bool_feats)
}
\description{
Subset feature metadata from giotto object
}
\keyword{internal}
