% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{subset_giotto_points_object}
\alias{subset_giotto_points_object}
\title{Subset giotto points object}
\usage{
subset_giotto_points_object(
  gpoints,
  feat_ids = NULL,
  x_min = NULL,
  x_max = NULL,
  y_min = NULL,
  y_max = NULL
)
}
\description{
Subset a single giotto points object
}
\details{
Subset on feature ids and on x,y coordinates
}
\keyword{internal}
