% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{subset_giotto_polygon_object}
\alias{subset_giotto_polygon_object}
\title{Subset giotto polygon object}
\usage{
subset_giotto_polygon_object(gpolygon, cell_ids, feat_ids, feat_type)
}
\description{
Subset a single giotto polygon object
}
\keyword{internal}
