% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{subset_spatial_enrichment}
\alias{subset_spatial_enrichment}
\title{Subset spatial enrichment}
\usage{
subset_spatial_enrichment(gobject, spat_unit, feat_type, filter_bool_cells)
}
\description{
Subset spatial enrichment results from giotto object
}
\keyword{internal}
