% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{subset_spatial_info_data}
\alias{subset_spatial_info_data}
\title{Subset spatial info data}
\usage{
subset_spatial_info_data(
  spatial_info,
  cell_ids,
  poly_info = "cell",
  feat_ids,
  feat_type = NULL
)
}
\description{
Subset  all spatial info (polygon) data
}
\keyword{internal}
