% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{subset_spatial_locations}
\alias{subset_spatial_locations}
\title{Subset spatial locations}
\usage{
subset_spatial_locations(gobject, cell_ids, spat_unit)
}
\description{
Subset location data from giotto object
}
\keyword{internal}
