% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{subset_spatial_network}
\alias{subset_spatial_network}
\title{Subset spatial network}
\usage{
subset_spatial_network(gobject, spat_unit, cell_ids)
}
\description{
subset ALL spatial networks from giotto object of the given
spat_unit
}
\keyword{internal}
