% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{subsetGiotto}
\alias{subsetGiotto}
\title{subsetGiotto}
\usage{
subsetGiotto(gobject, cell_ids = NULL, gene_ids = NULL)
}
\arguments{
\item{gobject}{giotto object}

\item{cell_ids}{cell IDs to keep}

\item{gene_ids}{gene IDs to keep}
}
\value{
giotto object
}
\description{
subsets Giotto object including previous calculations
}
\examples{
    subsetGiotto(gobject)
}
