% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{subsetGiottoLocs}
\alias{subsetGiottoLocs}
\title{subsetGiottoLocs}
\usage{
subsetGiottoLocs(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  x_max = NULL,
  x_min = NULL,
  y_max = NULL,
  y_min = NULL,
  z_max = NULL,
  z_min = NULL,
  poly_info = "cell",
  return_gobject = T,
  verbose = FALSE
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit}

\item{feat_type}{feature type to use}

\item{x_max}{maximum x-coordinate}

\item{x_min}{minimum x-coordinate}

\item{y_max}{maximum y-coordinate}

\item{y_min}{minimum y-coordinate}

\item{z_max}{maximum z-coordinate}

\item{z_min}{minimum z-coordinate}

\item{poly_info}{polygon information to use}

\item{return_gobject}{return Giotto object}

\item{verbose}{be verbose}
}
\value{
giotto object
}
\description{
subsets Giotto object based on spatial locations
}
\details{
if return_gobject = FALSE, then a filtered combined metadata data.table will be returned
}
