% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{subsetGiottoLocsSubcellular}
\alias{subsetGiottoLocsSubcellular}
\title{Subset raw subcellular information by location}
\usage{
subsetGiottoLocsSubcellular(
  gobject,
  poly_info,
  feat_type = NULL,
  x_min = NULL,
  x_max = NULL,
  y_min = NULL,
  y_max = NULL,
  z_max = NULL,
  z_min = NULL,
  verbose = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{poly_info}{polygon information to use}

\item{feat_type}{feature type (e.g. "rna", "dna", "protein")}

\item{x_max, x_min, y_max, y_min, z_max, z_min}{minimum and maximum x, y, and z coordinates
to subset to}

\item{verbose}{be verbose}
}
\value{
giotto object
}
\description{
Subsets Giotto object based on spatial coordinates
}
\details{
Subsets a Giotto Subcellular object - without aggregated information - based on spatial coordinates
}
