% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{t_giotto}
\alias{t_giotto}
\title{t_giotto}
\usage{
t_giotto(mymatrix)
}
\arguments{
\item{mymatrix}{matrix object}
}
\value{
transposed matrix
}
\description{
t function that works with multiple matrix representations
}
