% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_structures.R
\name{tessellate}
\alias{tessellate}
\title{tessellate}
\usage{
tessellate(extent, shape = c("hexagon", "square"), radius, name = "grid")
}
\arguments{
\item{extent}{SpatExtent or anything else a SpatExtent can be extracted or created from}

\item{shape}{Shape of the tessellation grid. Available options are "hexagon" and "square".}

\item{radius}{numeric. Radius size of the tessellation grid.}

\item{name}{name of giottoPolygons grid to make}
}
\value{
A giottoPolygon
}
\description{
Generates a tessellated grid of polygons within the provided spatial extent
}
\details{
This function generates a tessellated grid of spatial locations based on the input spatial locations. The shape of the tessellation grid can be either hexagonal or square. The radius parameter determines the size of the grid cells or the bin size.
}
\concept{spatial location}
