% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_structures.R
\name{tessellateSpatLocs}
\alias{tessellateSpatLocs}
\title{tessellateSpatLocs}
\usage{
tessellateSpatLocs(spat_locs, shape = c("hexagon", "square"), radius)
}
\arguments{
\item{spat_locs}{Spatial locations to be tessellated.}

\item{shape}{Shape of the tessellation grid. Available options are "hexagon" and "square".}

\item{radius}{Radius size of the tessellation grid.}
}
\value{
A giotto object containing the tessellated spatial locations.
}
\description{
Generates a tessellated grid of spatial locations based on the input spatial locations.
}
\details{
This function generates a tessellated grid of spatial locations based on the input spatial locations. The shape of the tessellation grid can be either hexagonal or square. The radius parameter determines the size of the grid cells or the bin size.
}
\concept{spatial location}
