% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{time_format}
\alias{time_format}
\title{Format time for printing}
\usage{
time_format(secs)
}
\description{
Format time for printing
}
\details{
Code from \code{\link[data.table]{timetaken}}
}
\keyword{internal}
