% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registration.R
\name{trakem2_rigid_transf_extract}
\alias{trakem2_rigid_transf_extract}
\title{trakem2_rigid_transf_extract}
\usage{
trakem2_rigid_transf_extract(inputstring)
}
\arguments{
\item{inputstring}{string read in from TrakeEM2 xml file}
}
\description{
Extract rigid registration transformation values from FIJI TrakEM2 xml file. Generated through register_virtual_stack_slices.
}
\keyword{internal}
