% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_section.R
\name{transform_2d_mesh_to_3d_mesh}
\alias{transform_2d_mesh_to_3d_mesh}
\title{transform_2d_mesh_to_3d_mesh}
\usage{
transform_2d_mesh_to_3d_mesh(
  mesh_line_obj_2d,
  pca_out,
  center_vec,
  mesh_grid_n
)
}
\arguments{
\item{mesh_line_obj_2d}{mesh_line_obj_2d}

\item{pca_out}{pca_out}

\item{center_vec}{center_vec}

\item{mesh_grid_n}{mesh_grid_n}
}
\description{
transform 2d mesh to 3d mesh by reversing PCA
}
\keyword{internal}
