% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{transpose-generic}
\alias{transpose-generic}
\alias{t}
\alias{t,spatLocsObj-method}
\alias{t,spatialNetworkObj-method}
\alias{t.spatLocsObj}
\alias{t.spatialNetworkObj}
\title{Transpose}
\usage{
\S4method{t}{spatLocsObj}(x)

\S4method{t}{spatialNetworkObj}(x)

t.spatLocsObj(x)

t.spatialNetworkObj(x)
}
\arguments{
\item{x}{object to be transposed}
}
\description{
Transpose
}
