% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_genes.R
\name{trendSceek}
\alias{trendSceek}
\title{trendSceek}
\usage{
trendSceek(
  gobject,
  expression_values = c("normalized", "raw"),
  subset_genes = NULL,
  nrand = 100,
  ncores = 8,
  ...
)
}
\arguments{
\item{gobject}{Giotto object}

\item{expression_values}{gene expression values to use}

\item{subset_genes}{subset of genes to run trendsceek on}

\item{nrand}{An integer specifying the number of random resamplings of the mark distribution as to create the null-distribution.}

\item{ncores}{An integer specifying the number of cores to be used by BiocParallel}

\item{\dots}{Additional parameters to the \code{\link[trendsceek]{trendsceek_test}} function}
}
\value{
data.frame with trendsceek spatial genes results
}
\description{
Compute spatial variable genes with trendsceek method
}
\details{
This function is a wrapper for the trendsceek_test method implemented in the trendsceek package
}
