% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto.R
\name{update_giotto_params}
\alias{update_giotto_params}
\title{Update giotto parameters}
\usage{
update_giotto_params(
  gobject,
  description = "_test",
  return_gobject = TRUE,
  toplevel = 2
)
}
\description{
Update giotto parameters
}
\keyword{internal}
