% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{updateGiottoImageMG}
\alias{updateGiottoImageMG}
\title{updateGiottoImageMG}
\usage{
updateGiottoImageMG(
  gobject = NULL,
  image_name = NULL,
  giottoImage = NULL,
  xmax_adj = 0,
  xmin_adj = 0,
  ymax_adj = 0,
  ymin_adj = 0,
  x_shift = 0,
  y_shift = 0,
  scale_factor = NULL,
  scale_x = 1,
  scale_y = 1,
  order = c("first_adj", "first_scale"),
  xmin_set = NULL,
  xmax_set = NULL,
  ymin_set = NULL,
  ymax_set = NULL,
  return_gobject = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{image_name}{name of giottoImage}

\item{giottoImage}{giottoImage object}

\item{xmax_adj}{adjustment of the maximum x-value to align the image}

\item{xmin_adj}{adjustment of the minimum x-value to align the image}

\item{ymax_adj}{adjustment of the maximum y-value to align the image}

\item{ymin_adj}{adjustment of the minimum y-value to align the image}

\item{x_shift}{shift entire image in positive x direction}

\item{y_shift}{shift entire image in positive y direction}

\item{scale_factor}{set scale_x and scale_y at the same time}

\item{scale_x}{independently scale x axis image mapping from origin}

\item{scale_y}{independently scale y axis image mapping from origin}

\item{order}{perform fine adjustments (adjustments and shifts) or scaling first}

\item{xmin_set}{set image xmin boundary. Applied before adjustments}

\item{xmax_set}{set image xmax boundary. Applied before adjustments}

\item{ymin_set}{set image ymin boundary. Applied before adjustments}

\item{ymax_set}{set image ymax boundary. Applied before adjustments}

\item{return_gobject}{return a giotto object}

\item{verbose}{be verbose}
}
\value{
a giotto object or an updated giotto image if return_gobject = F
}
\description{
Updates the boundaries of a giotto image attached to a giotto object
}
