% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{updateGiottoLargeImage}
\alias{updateGiottoLargeImage}
\title{updateGiottoLargeImage}
\usage{
updateGiottoLargeImage(
  gobject = NULL,
  largeImage_name = NULL,
  giottoLargeImage = NULL,
  xmax_adj = 0,
  xmin_adj = 0,
  ymax_adj = 0,
  ymin_adj = 0,
  x_shift = 0,
  y_shift = 0,
  scale_factor = NULL,
  scale_x = 1,
  scale_y = 1,
  order = c("first_adj", "first_scale"),
  xmin_set = NULL,
  xmax_set = NULL,
  ymin_set = NULL,
  ymax_set = NULL,
  return_gobject = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{gobject}{giotto object containing giottoLargeImage}

\item{largeImage_name}{name of giottoLargeImage}

\item{giottoLargeImage}{giottoLargeImage object}

\item{xmax_adj}{adjustment of the maximum x-value to align the image}

\item{xmin_adj}{adjustment of the minimum x-value to align the image}

\item{ymax_adj}{adjustment of the maximum y-value to align the image}

\item{ymin_adj}{adjustment of the minimum y-value to align the image}

\item{x_shift}{shift entire image in positive x direction}

\item{y_shift}{shift entire image in positive y direction}

\item{scale_factor}{set scale_x and scale_y at the same time}

\item{scale_x}{independently scale x axis image mapping from origin}

\item{scale_y}{independently scale y axis image mapping from origin}

\item{order}{perform fine adjustments (adjustments and shifts) or scaling first}

\item{return_gobject}{return a giotto object if TRUE, a giottoLargeImage if FALSE}

\item{verbose}{be verbose}

\item{xmin_bound}{set image xmin boundary. Overrides minmax values as spatial anchor.}

\item{xmax_bound}{set image xmax boundary. Overrides minmax values as spatial anchor.}

\item{ymin_bound}{set image ymin boundary. Overrides minmax values as spatial anchor.}

\item{ymax_bound}{set image ymax boundary. Overrides minmax values as spatial anchor.}
}
\value{
a giotto object or an updated giottoLargeImage if return_gobject = FALSE
}
\description{
Updates the boundaries of a giottoLargeImage attached to a giotto object
}
