% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{updateGiottoObject}
\alias{updateGiottoObject}
\title{Update giotto object}
\usage{
updateGiottoObject(gobject)
}
\arguments{
\item{gobject}{giotto object to update}
}
\description{
Updates the giotto object for changes in structure for backwards
compatibility with earlier versions
}
\details{
Supported updates:
\itemize{
  \item{3.2.0 update adding multiomics slot}
  \item{master branch to suite - TODO}
}
}
\examples{
\dontrun{
gobject = updateGiottoObject(gobject)
}
}
