% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{updateGiottoPointsObject}
\alias{updateGiottoPointsObject}
\title{Update giotto points object}
\usage{
updateGiottoPointsObject(gpoints)
}
\arguments{
\item{gpoints}{giotto points object}
}
\description{
Update giotto points object
}
