% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{updateGiottoPolygonObject}
\alias{updateGiottoPolygonObject}
\title{Update giotto polygon object}
\usage{
updateGiottoPolygonObject(gpoly)
}
\arguments{
\item{gpoly}{giotto polygon object}
}
\description{
Update giotto polygon object
}
