% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{validObject}
\alias{validObject}
\title{See \code{methods::\link[methods]{validObject}} for details.}
\usage{
validObject(object, test = FALSE, complete = FALSE)
}
\description{
See \code{methods::\link[methods]{validObject}} for details.
}
\keyword{internal}
