% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_hmrf.R
\name{viewHMRFresults}
\alias{viewHMRFresults}
\title{viewHMRFresults}
\usage{
viewHMRFresults(gobject, HMRFoutput, k = NULL, betas_to_view = NULL,
  ...)
}
\arguments{
\item{gobject}{giotto object}

\item{HMRFoutput}{HMRF output from doHMRF}

\item{k}{number of HMRF domains}

\item{betas_to_view}{results from different betas that you want to view}

\item{...}{paramters to visPlot()}
}
\value{
spatial plots with HMRF domains
}
\description{
View results from doHMRF.
}
\details{
Description ...
}
\examples{
    viewHMRFresults(gobject)
}
\seealso{
\code{\link{visPlot}}
}
