% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_hmrf.R
\name{viewHMRFresults2D}
\alias{viewHMRFresults2D}
\title{viewHMRFresults2D}
\usage{
viewHMRFresults2D(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  HMRFoutput,
  k = NULL,
  betas_to_view = NULL,
  ...
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit}

\item{feat_type}{feature type}

\item{HMRFoutput}{HMRF output from doHMRF}

\item{k}{number of HMRF domains}

\item{betas_to_view}{results from different betas that you want to view}

\item{\dots}{additional parameters to spatPlot2D()}
}
\value{
spatial plots with HMRF domains
}
\description{
View results from doHMRF.
}
\seealso{
\code{\link{spatPlot2D}}
}
