% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_hmrf.R
\name{viewHMRFresults_V2}
\alias{viewHMRFresults_V2}
\title{view HMRF result}
\usage{
viewHMRFresults_V2(gobject, HMRFoutput, k, betas, third_dim = FALSE, ...)
}
\description{
function to plot spatial location with HMRF cluster of k and betas
}
\keyword{external}
