% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visuals.R
\name{visDimGenePlot_2D_ggplot}
\alias{visDimGenePlot_2D_ggplot}
\title{visDimGenePlot_2D_ggplot}
\usage{
visDimGenePlot_2D_ggplot(gobject, expression_values = c("normalized",
  "scaled", "custom"), genes = NULL, dim_reduction_to_use = "umap",
  dim_reduction_name = "umap", dim1_to_use = 1, dim2_to_use = 2,
  show_NN_network = F, nn_network_to_use = "sNN",
  network_name = "sNN.pca", network_color = "lightgray",
  edge_alpha = NULL, scale_alpha_with_expression = TRUE,
  point_size = 1, genes_high_color = "red",
  genes_mid_color = "white", genes_low_color = "blue",
  point_border_col = "black", point_border_stroke = 0.1,
  midpoint = 0, cow_n_col = 2, cow_rel_h = 1, cow_rel_w = 1,
  cow_align = "h", show_legend = T, show_plots = F)
}
\arguments{
\item{gobject}{giotto object}

\item{expression_values}{gene expression values to use}

\item{genes}{genes to show}

\item{dim_reduction_to_use}{dimension reduction to use}

\item{dim_reduction_name}{dimension reduction name}

\item{dim1_to_use}{dimension to use on x-axis}

\item{dim2_to_use}{dimension to use on y-axis}

\item{show_NN_network}{show underlying NN network}

\item{nn_network_to_use}{type of NN network to use (kNN vs sNN)}

\item{network_name}{name of NN network to use, if show_NN_network = TRUE}

\item{edge_alpha}{column to use for alpha of the edges}

\item{scale_alpha_with_expression}{scale expression with ggplot alpha parameter}

\item{point_size}{size of point (cell)}

\item{point_border_col}{color of border around points}

\item{point_border_stroke}{stroke size of border around points}

\item{midpoint}{size of point (cell)}

\item{cow_n_col}{cowplot param: how many columns}

\item{cow_rel_h}{cowplot param: relative height}

\item{cow_rel_w}{cowplot param: relative width}

\item{cow_align}{cowplot param: how to align}

\item{show_legend}{show legend}

\item{show_plots}{show plots}
}
\value{
ggplot
}
\description{
Visualize cells and gene expression according to dimension reduction coordinates
}
\details{
Description of parameters.
}
\examples{
    visDimGenePlot_2D_ggplot(gobject)
}
