% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visuals.R
\name{visForceLayoutPlot}
\alias{visForceLayoutPlot}
\title{visForceLayoutPlot}
\usage{
visForceLayoutPlot(gobject, nn_network_to_use = "sNN",
  network_name = "sNN.pca", layout_name = "layout", dim1_to_use = 1,
  dim2_to_use = 2, show_NN_network = T, cell_color = NULL,
  color_as_factor = F, cell_color_code = NULL, edge_alpha = NULL,
  point_size = 1, point.border.col = "black",
  point.border.stroke = 0.1, show.legend = T)
}
\arguments{
\item{gobject}{giotto object}

\item{nn_network_to_use}{type of NN network to use (kNN vs sNN)}

\item{network_name}{NN network to use}

\item{layout_name}{name of layout to use}

\item{dim1_to_use}{dimension to use on x-axis}

\item{dim2_to_use}{dimension to use on y-axis}

\item{show_NN_network}{show underlying NN network}

\item{cell_color}{color for cells (see details)}

\item{color_as_factor}{convert color column to factor}

\item{cell_color_code}{named vector with colors}

\item{edge_alpha}{column to use for alpha of the edges}

\item{point_size}{size of point (cell)}

\item{point.border.col}{color of border around points}

\item{point.border.stroke}{stroke size of border around points}

\item{show.legend}{show legend}
}
\value{
ggplot
}
\description{
Visualize cells according to forced layout algorithm coordinates
}
\details{
Description of parameters.
}
\examples{
    visForceLayoutPlot(gobject)
}
