% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visuals.R
\name{visGenePlot}
\alias{visGenePlot}
\title{visGenePlot}
\usage{
visGenePlot(gobject, expression_values = c("normalized", "scaled",
  "custom"), genes, show_network = F, network_color = NULL,
  spatial_network_name = "spatial_network", show_grid = F,
  grid_color = NULL, spatial_grid_name = "spatial_grid",
  midpoint = 0, scale_alpha_with_expression = TRUE, point_size = 1,
  show.legend = T, cow_n_col = 2, cow_rel_h = 1, cow_rel_w = 1,
  cow_align = "h", show_plots = F)
}
\arguments{
\item{gobject}{giotto object}

\item{expression_values}{gene expression values to use}

\item{genes}{genes to show}

\item{show_network}{show underlying spatial network}

\item{network_color}{color of spatial network}

\item{spatial_network_name}{name of spatial network to use}

\item{show_grid}{show spatial grid}

\item{grid_color}{color of spatial grid}

\item{spatial_grid_name}{name of spatial grid to use}

\item{midpoint}{expression midpoint}

\item{scale_alpha_with_expression}{scale expression with ggplot alpha parameter}

\item{point_size}{size of point (cell)}

\item{show.legend}{show legend}

\item{cow_n_col}{cowplot param: how many columns}

\item{cow_rel_h}{cowplot param: relative height}

\item{cow_rel_w}{cowplot param: relative width}

\item{cow_align}{cowplot param: how to align}

\item{show_plots}{show plots}
}
\value{
ggplot
}
\description{
Visualize cells and gene expression according to spatial coordinates
}
\details{
Description of parameters.
}
\examples{
    visGenePlot(gobject)
}
