% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visuals.R
\name{visGenePlot_3D_plotly}
\alias{visGenePlot_3D_plotly}
\title{visGenePlot_3D_plotly}
\usage{
visGenePlot_3D_plotly(
  gobject,
  expression_values = c("normalized", "scaled", "custom"),
  genes,
  show_network = F,
  network_color = NULL,
  spatial_network_name = "spatial_network",
  edge_alpha = NULL,
  show_grid = F,
  genes_high_color = NULL,
  genes_mid_color = "white",
  genes_low_color = "blue",
  spatial_grid_name = "spatial_grid",
  point_size = 1,
  show_legend = T,
  axis_scale = c("cube", "real", "custom"),
  custom_ratio = NULL,
  x_ticks = NULL,
  y_ticks = NULL,
  z_ticks = NULL,
  show_plots = F
)
}
\arguments{
\item{gobject}{giotto object}

\item{expression_values}{gene expression values to use}

\item{genes}{genes to show}

\item{show_network}{show underlying spatial network}

\item{network_color}{color of spatial network}

\item{spatial_network_name}{name of spatial network to use}

\item{show_grid}{show spatial grid}

\item{genes_high_color}{color represents high gene expression}

\item{genes_mid_color}{color represents middle gene expression}

\item{genes_low_color}{color represents low gene expression}

\item{spatial_grid_name}{name of spatial grid to use}

\item{point_size}{size of point (cell)}

\item{show_legend}{show legend}

\item{axis_scale}{three mode to adjust axis scale}

\item{x_ticks}{number of ticks on x axis}

\item{y_ticks}{number of ticks on y axis}

\item{z_ticks}{number of ticks on z axis}

\item{show_plots}{show plots}

\item{grid_color}{color of spatial grid}

\item{cow_n_col}{cowplot param: how many columns}

\item{cow_rel_h}{cowplot param: relative height}

\item{cow_rel_w}{cowplot param: relative width}

\item{cow_align}{cowplot param: how to align}
}
\value{
plotly
}
\description{
Visualize cells and gene expression according to spatial coordinates
}
\details{
Description of parameters.
}
\keyword{internal}
