% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{visium_micron_scalefactor}
\alias{visium_micron_scalefactor}
\alias{.visium_micron_scale}
\title{Calculate Pixel to Micron Scalefactor}
\usage{
.visium_micron_scale(json_scalefactors)
}
\arguments{
\item{json_scalefactors}{list of scalefactors from
.visium_read_scalefactors()}
}
\value{
scale factor for converting pixel to micron
}
\description{
Calculate Pixel to Micron Scalefactor
}
\details{
Calculates pixel to micron scalefactor.
Visium xy coordinates are based on the fullres image
The values provided are directly usable for generating polygon information
or calculating the micron size relative to spatial coordinates for this set
of spatial information.
}
\keyword{internal}
