% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visuals.R
\name{visPlot}
\alias{visPlot}
\title{visPlot}
\usage{
visPlot(gobject, sdimx = NULL, sdimy = NULL, sdimz = NULL,
  point_size = 1, point.border.col = "black",
  point.border.stroke = 0.1, cell_color = NULL,
  cell_color_code = NULL, color_as_factor = T, show_network = F,
  network_color = NULL, spatial_network_name = "spatial_network",
  show_grid = F, grid_color = NULL,
  spatial_grid_name = "spatial_grid", coord_fix_ratio = 0.6,
  title = "", show.legend = T, show.plot = F)
}
\arguments{
\item{gobject}{giotto object}

\item{sdimx}{x-axis dimension name (default = 'sdimx')}

\item{sdimy}{y-axis dimension name (default = 'sdimy')}

\item{sdimz}{z-axis dimension name (default = 'sdimz')}

\item{point_size}{size of point (cell)}

\item{point.border.col}{color of border around points}

\item{point.border.stroke}{stroke size of border around points}

\item{cell_color}{color for cells (see details)}

\item{cell_color_code}{named vector with colors}

\item{color_as_factor}{convert color column to factor}

\item{show_network}{show underlying spatial network}

\item{network_color}{color of spatial network}

\item{spatial_network_name}{name of spatial network to use}

\item{show_grid}{show spatial grid}

\item{grid_color}{color of spatial grid}

\item{spatial_grid_name}{name of spatial grid to use}

\item{coord_fix_ratio}{fix ratio between x and y-axis}

\item{title}{title of plot}

\item{show.legend}{show legend}

\item{show.plot}{show plot}
}
\value{
ggplot
}
\description{
Visualize cells according to spatial coordinates
}
\details{
Description of parameters.
}
\examples{
    visPlot(gobject)
}
