% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visuals.R
\name{visPlot_2D_plotly}
\alias{visPlot_2D_plotly}
\title{visPlot_2D_plotly}
\usage{
visPlot_2D_plotly(gobject, sdimx = NULL, sdimy = NULL,
  point_size = 3, cell_color = NULL, cell_color_code = NULL,
  color_as_factor = T, select_cell_groups = NULL,
  select_cells = NULL, show_other_cells = T,
  other_cell_color = "lightgrey", other_point_size = 0.5,
  show_network = F, network_color = "lightgray", network_alpha = 1,
  other_cell_alpha = 0.5, spatial_network_name = "spatial_network",
  show_grid = F, grid_color = NULL, grid_alpha = 1,
  spatial_grid_name = "spatial_grid", show_legend = T,
  axis_scale = c("cube", "real", "custom"), custom_ratio = NULL,
  x_ticks = NULL, y_ticks = NULL, show_plot = F)
}
\arguments{
\item{gobject}{giotto object}

\item{sdimx}{x-axis dimension name (default = 'sdimx')}

\item{sdimy}{y-axis dimension name (default = 'sdimy')}

\item{point_size}{size of point (cell)}

\item{cell_color}{color for cells (see details)}

\item{cell_color_code}{named vector with colors}

\item{color_as_factor}{convert color column to factor}

\item{select_cell_groups}{select a subset of the groups from cell_color}

\item{show_network}{show underlying spatial network}

\item{network_color}{color of spatial network}

\item{spatial_network_name}{name of spatial network to use}

\item{show_grid}{show spatial grid}

\item{grid_color}{color of spatial grid}

\item{grid_alpha}{alpha of spatial grid}

\item{spatial_grid_name}{name of spatial grid to use}

\item{show_legend}{show legend}

\item{show_plot}{show plot}
}
\value{
plotly
}
\description{
Visualize cells according to spatial coordinates
}
\details{
Description of parameters.
}
\examples{
    visPlot_2D_plotly(gobject)
}
