% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visuals.R
\name{visPlot_3D_plotly}
\alias{visPlot_3D_plotly}
\title{visPlot_3D_plotly}
\usage{
visPlot_3D_plotly(gobject, sdimx = NULL, sdimy = NULL, sdimz = NULL,
  point_size = 3, cell_color = NULL, cell_color_code = NULL,
  select_cell_groups = NULL, select_cells = NULL,
  show_other_cells = T, other_cell_color = "lightgrey",
  other_point_size = 0.5, show_network = F, network_color = NULL,
  network_alpha = 1, other_cell_alpha = 0.5,
  spatial_network_name = "spatial_network",
  spatial_grid_name = "spatial_grid", title = "", show_legend = T,
  axis_scale = c("cube", "real", "custom"), custom_ratio = NULL,
  x_ticks = NULL, y_ticks = NULL, z_ticks = NULL, show_plot = F)
}
\arguments{
\item{gobject}{giotto object}

\item{sdimx}{x-axis dimension name (default = 'sdimx')}

\item{sdimy}{y-axis dimension name (default = 'sdimy')}

\item{sdimz}{z-axis dimension name (default = 'sdimz')}

\item{point_size}{size of point (cell)}

\item{cell_color}{color for cells (see details)}

\item{cell_color_code}{named vector with colors}

\item{select_cell_groups}{select a subset of the groups from cell_color}

\item{show_network}{show underlying spatial network}

\item{network_color}{color of spatial network}

\item{spatial_network_name}{name of spatial network to use}

\item{spatial_grid_name}{name of spatial grid to use}

\item{title}{title of plot}

\item{show_legend}{show legend}

\item{show_plot}{show plot}

\item{point_border_col}{color of border around points}

\item{point_border_stroke}{stroke size of border around points}

\item{color_as_factor}{convert color column to factor}

\item{show_grid}{show spatial grid}

\item{grid_color}{color of spatial grid}

\item{coord_fix_ratio}{fix ratio between x and y-axis}
}
\value{
ggplot
}
\description{
Visualize cells according to spatial coordinates
}
\details{
Description of parameters.
}
\examples{
    visPlot_3D_plotly(gobject)
}
