% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visuals.R
\name{visSpatDimGenePlot}
\alias{visSpatDimGenePlot}
\title{visSpatDimGenePlot}
\usage{
visSpatDimGenePlot(gobject, plot_method = c("ggplot", "plotly"),
  expression_values = c("normalized", "scaled", "custom"),
  plot_alignment = c("horizontal", "vertical"),
  dim_reduction_to_use = "umap", dim_reduction_name = "umap",
  dim1_to_use = 1, dim2_to_use = 2, dim3_to_use = NULL,
  sdimx = NULL, sdimy = NULL, sdimz = NULL, genes,
  dim_point_border_col = "black", dim_point_border_stroke = 0.1,
  show_NN_network = F, nn_network_to_use = "sNN",
  network_name = "sNN.pca", edge_alpha_dim = NULL,
  scale_alpha_with_expression = TRUE, label_size = 16,
  genes_low_color = "blue", genes_mid_color = "white",
  genes_high_color = "red", dim_point_size = 3,
  nn_network_alpha = 0.5, show_spatial_network = F,
  spatial_network_name = "spatial_network",
  network_color = "lightgray", spatial_network_alpha = 0.5,
  show_spatial_grid = F, spatial_grid_name = "spatial_grid",
  spatial_grid_color = NULL, spatial_grid_alpha = 0.5,
  spatial_point_size = 3, spatial_point_border_col = "black",
  spatial_point_border_stroke = 0.1, legend_text_size = 12,
  axis_scale = c("cube", "real", "custom"), custom_ratio = NULL,
  x_ticks = NULL, y_ticks = NULL, z_ticks = NULL, midpoint = 0,
  point_size = 1, cow_n_col = 2, cow_rel_h = 1, cow_rel_w = 1,
  cow_align = "h", show_legend = T, show_plots = F)
}
\arguments{
\item{gobject}{giotto object}

\item{expression_values}{gene expression values to use}

\item{plot_alignment}{direction to align plot}

\item{dim_reduction_to_use}{dimension reduction to use}

\item{dim_reduction_name}{dimension reduction name}

\item{dim1_to_use}{dimension to use on x-axis}

\item{dim2_to_use}{dimension to use on y-axis}

\item{dim3_to_use}{dimension to use on z-axis}

\item{sdimx}{x-axis dimension name (default = 'sdimx')}

\item{sdimy}{y-axis dimension name (default = 'sdimy')}

\item{sdimz}{z-axis dimension name (default = 'sdimz')}

\item{genes}{genes to show}

\item{dim_point_border_col}{color of border around points}

\item{dim_point_border_stroke}{stroke size of border around points}

\item{show_NN_network}{show underlying NN network}

\item{nn_network_to_use}{type of NN network to use (kNN vs sNN)}

\item{network_name}{name of NN network to use, if show_NN_network = TRUE}

\item{edge_alpha_dim}{dim reduction plot: column to use for alpha of the edges}

\item{scale_alpha_with_expression}{scale expression with ggplot alpha parameter}

\item{label_size}{size for the label}

\item{genes_low_color}{color to represent low expression of gene}

\item{genes_high_color}{color to represent high expression of gene}

\item{dim_point_size}{dim reduction plot: point size}

\item{spatial_network_name}{name of spatial network to use}

\item{spatial_grid_name}{name of spatial grid to use}

\item{spatial_point_size}{spatial plot: point size}

\item{spatial_point_border_col}{color of border around points}

\item{spatial_point_border_stroke}{stroke size of border around points}

\item{legend_text_size}{the size of the text in legend}

\item{axis_scale}{three modes to adjust axis scale ratio}

\item{custom_ratio}{set the axis scale ratio on custom}

\item{x_ticks}{number of ticks on x axis}

\item{y_ticks}{number of ticks on y axis}

\item{z_ticks}{number of ticks on z axis}

\item{midpoint}{size of point (cell)}

\item{point_size}{size of point (cell)}

\item{cow_n_col}{cowplot param: how many columns}

\item{cow_rel_h}{cowplot param: relative height}

\item{cow_rel_w}{cowplot param: relative width}

\item{cow_align}{cowplot param: how to align}

\item{show_legend}{show legend}

\item{show_plot}{show plot}
}
\value{
ggplot or plotly
}
\description{
integration of visSpatDimGenePlot_2D(ggplot) and visSpatDimGenePlot_3D(plotly)
}
\details{
Description of parameters.
}
\examples{
    visSpatDimGenePlot(gobject)
}
