% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visuals.R
\name{visSpatDimGenePlot_2D}
\alias{visSpatDimGenePlot_2D}
\title{visSpatDimGenePlot_2D}
\usage{
visSpatDimGenePlot_2D(gobject, expression_values = c("normalized",
  "scaled", "custom"), plot_alignment = c("horizontal", "vertical"),
  genes, dim_reduction_to_use = "umap", dim_reduction_name = "umap",
  dim1_to_use = 1, dim2_to_use = 2, dim_point_size = 1,
  dim_point_border_col = "black", dim_point_border_stroke = 0.1,
  show_NN_network = F, nn_network_to_use = "sNN",
  network_name = "sNN.pca", edge_alpha_dim = NULL,
  scale_alpha_with_expression = TRUE,
  spatial_network_name = "spatial_network",
  spatial_grid_name = "spatial_grid", spatial_point_size = 1,
  spatial_point_border_col = "black",
  spatial_point_border_stroke = 0.1, midpoint = 0, point_size = 1,
  cow_n_col = 2, cow_rel_h = 1, cow_rel_w = 1, cow_align = "h",
  show_legend = T, show_plots = F)
}
\arguments{
\item{gobject}{giotto object}

\item{expression_values}{gene expression values to use}

\item{plot_alignment}{direction to align plot}

\item{genes}{genes to show}

\item{dim_reduction_to_use}{dimension reduction to use}

\item{dim_reduction_name}{dimension reduction name}

\item{dim1_to_use}{dimension to use on x-axis}

\item{dim2_to_use}{dimension to use on y-axis}

\item{dim_point_size}{dim reduction plot: point size}

\item{dim_point_border_col}{color of border around points}

\item{dim_point_border_stroke}{stroke size of border around points}

\item{show_NN_network}{show underlying NN network}

\item{nn_network_to_use}{type of NN network to use (kNN vs sNN)}

\item{network_name}{name of NN network to use, if show_NN_network = TRUE}

\item{edge_alpha_dim}{dim reduction plot: column to use for alpha of the edges}

\item{scale_alpha_with_expression}{scale expression with ggplot alpha parameter}

\item{spatial_network_name}{name of spatial network to use}

\item{spatial_grid_name}{name of spatial grid to use}

\item{spatial_point_size}{spatial plot: point size}

\item{spatial_point_border_col}{color of border around points}

\item{spatial_point_border_stroke}{stroke size of border around points}

\item{midpoint}{size of point (cell)}

\item{point_size}{size of point (cell)}

\item{cow_n_col}{cowplot param: how many columns}

\item{cow_rel_h}{cowplot param: relative height}

\item{cow_rel_w}{cowplot param: relative width}

\item{cow_align}{cowplot param: how to align}

\item{show_legend}{show legend}

\item{show_plot}{show plot}
}
\value{
ggplot
}
\description{
Visualize cells according to spatial AND dimension reduction coordinates in ggplot mode
}
\details{
Description of parameters.
}
\examples{
    visSpatDimGenePlot_2D(gobject)
}
