% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visuals.R
\name{visSpatDimPlot}
\alias{visSpatDimPlot}
\title{visSpatDimPlot}
\usage{
visSpatDimPlot(gobject, plot_alignment = c("vertical", "horizontal"),
  dim_reduction_to_use = "umap", dim_reduction_name = "umap",
  dim1_to_use = 1, dim2_to_use = 2, show_NN_network = F,
  nn_network_to_use = "sNN", network_name = "sNN.pca",
  cell_color = NULL, color_as_factor = T, cell_color_code = NULL,
  dim_point_size = 1, edge_alpha = 0.05, show_spatial_network = F,
  spatial_network_name = "spatial_network",
  spatial_network_color = NULL, show_spatial_grid = F,
  spatial_grid_name = "spatial_grid", spatial_grid_color = NULL,
  spatial_point_size = 1, show.legend = T, show.plot = F)
}
\arguments{
\item{gobject}{giotto object}

\item{plot_alignment}{direction to align plot}

\item{dim_reduction_to_use}{dimension reduction to use}

\item{dim_reduction_name}{dimension reduction name}

\item{dim1_to_use}{dimension to use on x-axis}

\item{dim2_to_use}{dimension to use on y-axis}

\item{show_NN_network}{show underlying NN network}

\item{nn_network_to_use}{type of NN network to use (kNN vs sNN)}

\item{network_name}{name of NN network to use, if show_NN_network = TRUE}

\item{cell_color}{color for cells (see details)}

\item{color_as_factor}{convert color column to factor}

\item{cell_color_code}{named vector with colors}

\item{edge_alpha}{column to use for alpha of the edges}

\item{show_spatial_network}{show spatial network}

\item{spatial_network_name}{name of spatial network to use}

\item{spatial_network_color}{color of spatial network}

\item{show_spatial_grid}{show spatial grid}

\item{spatial_grid_name}{name of spatial grid to use}

\item{spatial_grid_color}{color of spatial grid}

\item{show.legend}{show legend}

\item{show.plot}{show plot}
}
\value{
ggplot
}
\description{
Visualize cells according to spatial AND dimension reduction coordinates
}
\details{
Description of parameters.
}
\examples{
    visSpatDimPlot(gobject)
}
