% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{wrap-generic}
\alias{wrap-generic}
\alias{wrap,giottoPolygon-method}
\alias{wrap,giotto-method}
\alias{wrap,giottoPoints-method}
\alias{vect,packedGiottoPolygon-method}
\alias{vect,packedGiottoPoints-method}
\alias{vect,packedGiotto-method}
\title{Wrap giotto terra pointer information}
\usage{
\S4method{wrap}{giottoPolygon}(x)

\S4method{wrap}{giotto}(x)

\S4method{wrap}{giottoPoints}(x)

\S4method{vect}{packedGiottoPolygon}(x)

\S4method{vect}{packedGiottoPoints}(x)

\S4method{vect}{packedGiotto}(x)
}
\arguments{
\item{x}{giottoPolygon or giottoPoints}
}
\description{
Extension of wrap methods from terra for Giotto's terra-based S4
objects. Allows pointer information to be packaged into memory so that it can
be passed over a connection (e.g. nodes on a computer cluster)
}
\section{Functions}{
\itemize{
\item \code{wrap(giottoPolygon)}: Wrap giottoPolygon

\item \code{wrap(giotto)}: Wrap giotto

\item \code{wrap(giottoPoints)}: Wrap giottoPoints

\item \code{vect(packedGiottoPolygon)}: Unwrap giottoPolygon

\item \code{vect(packedGiottoPoints)}: Unwrap giottoPolygon

\item \code{vect(packedGiotto)}: Unwrap giotto

}}
