% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{wrap_msg}
\alias{wrap_msg}
\title{Wrap message}
\usage{
wrap_msg(..., sep = " ")
}
\arguments{
\item{...}{additional strings and/or elements to pass to cat}

\item{sep}{how to join elements of string (default is one space)}

\item{strWidth}{externally set wrapping width. (default value of 100 is not effected)}

\item{keywords}{internal}
}
\description{
Wrap message
}
