% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{wrap_txt}
\alias{wrap_txt}
\title{Wrap text}
\usage{
wrap_txt(..., sep = " ", strWidth = 100)
}
\arguments{
\item{...}{additional params to pass}

\item{sep}{how to join elements of string (default is one space)}

\item{strWidth}{externally set wrapping width. (default value of 100 is not effected)}
}
\description{
Wrap text
}
\keyword{internal}
