% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_viewer.R
\name{write_giotto_viewer_dim_reduction}
\alias{write_giotto_viewer_dim_reduction}
\title{write_giotto_viewer_dim_reduction}
\usage{
write_giotto_viewer_dim_reduction(
  dim_reduction_cell,
  dim_red = NULL,
  dim_red_name = NULL,
  dim_red_rounding = NULL,
  dim_red_rescale = c(-20, 20),
  output_directory = getwd()
)
}
\arguments{
\item{dim_reduction_cell}{dimension reduction slot from giotto object}

\item{dim_red}{high level name of dimension reduction}

\item{dim_red_name}{specific name of dimension reduction to use}

\item{dim_red_rounding}{numerical indicating how to round the coordinates}

\item{dim_red_rescale}{numericals to rescale the coordinates}

\item{output_directory}{directory where to save the files}
}
\value{
write a .txt and .annot file for the selection annotation
}
\description{
write out dimensional reduction data from a giotto object for the Viewer
}
\keyword{internal}
