% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_viewer.R
\name{write_giotto_viewer_numeric_annotation}
\alias{write_giotto_viewer_numeric_annotation}
\title{write_giotto_viewer_numeric_annotation}
\usage{
write_giotto_viewer_numeric_annotation(
  annotation,
  annot_name = "test",
  output_directory = getwd()
)
}
\arguments{
\item{annotation}{annotation from the data.table from giotto object}

\item{annot_name}{name of the annotation}

\item{output_directory}{directory where to save the files}
}
\value{
write a .txt and .annot file for the selection annotation
}
\description{
write out numeric annotation data from a giotto object for the Viewer
}
\keyword{internal}
