% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_help.R
\name{writeChatGPTqueryDEG}
\alias{writeChatGPTqueryDEG}
\title{writeChatGPTqueryDEG}
\usage{
writeChatGPTqueryDEG(
  DEG_output,
  top_n_genes = 10,
  tissue_type = "human breast cancer",
  folder_name = getwd(),
  file_name = "chatgpt_query.txt"
)
}
\arguments{
\item{DEG_output}{the output format from the differenetial expression functions}

\item{top_n_genes}{number of genes for each cluster}

\item{tissue_type}{tissue type}

\item{folder_name}{path to the folder where you want to save the .txt file}

\item{file_name}{name of .txt file}
}
\value{
writes a .txt file to the desired location
}
\description{
This function writes a query as a .txt file that can be used with
ChatGPT or a similar LLM service to find the most likely cell types based on the
top differential expressed genes (DEGs) between identified clusters.
}
\details{
This function does not run any LLM service. It simply creates the .txt
file that can then be used any LLM service (e.g. OpenAI, Gemini, ...)
}
