% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_hmrf.R
\name{writeHMRFresults}
\alias{writeHMRFresults}
\title{writeHMRFresults}
\usage{
writeHMRFresults(
  gobject,
  HMRFoutput,
  k = NULL,
  betas_to_view = NULL,
  print_command = F
)
}
\arguments{
\item{gobject}{giotto object}

\item{HMRFoutput}{HMRF output from doHMRF}

\item{k}{k to write results for}

\item{betas_to_view}{results from different betas that you want to view}

\item{print_command}{see the python command}
}
\value{
data.table with HMRF results for each b and the selected k
}
\description{
write results from doHMRF to a data.table.
}
