% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_section.R
\name{adapt_aspect_ratio}
\alias{adapt_aspect_ratio}
\title{adapt_aspect_ratio}
\usage{
adapt_aspect_ratio(
  current_ratio,
  cell_locations,
  sdimx = NULL,
  sdimy = NULL,
  sdimz = NULL,
  mesh_obj = NULL
)
}
\arguments{
\item{current_ratio}{current_ratio}

\item{cell_locations}{cell_locations}

\item{sdimx}{sdimx}

\item{sdimy}{sdimy}

\item{sdimz}{sdimz}

\item{mesh_obj}{mesh_obj}
}
\description{
adapt the aspact ratio after inserting cross section mesh grid lines
}
\keyword{internal}
