% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{addCellMetadata}
\alias{addCellMetadata}
\title{addCellMetadata}
\usage{
addCellMetadata(
  gobject,
  new_metadata,
  vector_name = NULL,
  by_column = FALSE,
  column_cell_ID = NULL
)
}
\arguments{
\item{gobject}{giotto object}

\item{new_metadata}{new cell metadata to use (data.table, data.frame, ...)}

\item{vector_name}{(optional) custom name if you provide a single vector}

\item{by_column}{merge metadata based on cell_ID column in pDataDT (default = FALSE)}

\item{column_cell_ID}{column name of new metadata to use if by_column = TRUE}
}
\value{
giotto object
}
\description{
adds cell metadata to the giotto object
}
\details{
You can add additional cell metadata in two manners:
\itemize{
  \item{1. Provide a data.table or data.frame with cell annotations in the same order as the cell_ID column in pDataDT(gobject) }
  \item{2. Provide a data.table or data.frame with cell annotations and specificy which column contains the cell IDs, these cell IDs need to match with the cell_ID column in pDataDT(gobject)}
}
}
