% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{addCellStatistics}
\alias{addCellStatistics}
\title{addCellStatistics}
\usage{
addCellStatistics(
  gobject,
  expression_values = c("normalized", "scaled", "custom"),
  detection_threshold = 0,
  return_gobject = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{expression_values}{expression values to use}

\item{detection_threshold}{detection threshold to consider a gene detected}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}
}
\value{
giotto object if return_gobject = TRUE
}
\description{
adds cells statistics to the giotto object
}
\details{
This function will add the following statistics to cell metadata:
\itemize{
  \item{nr_genes: }{Denotes in how many genes are detected per cell}
  \item{perc_genes: }{Denotes what percentage of genes is detected per cell}
  \item{total_expr: }{Shows the total sum of gene expression per cell}
}
}
\examples{

data(mini_giotto_single_cell)

updated_giotto_object = addCellStatistics(mini_giotto_single_cell)

}
