% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{addFeatsPerc}
\alias{addFeatsPerc}
\title{addFeatsPerc}
\usage{
addFeatsPerc(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  expression_values = c("normalized", "scaled", "custom"),
  feats = NULL,
  vector_name = "feat_perc",
  return_gobject = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit}

\item{feat_type}{feature type}

\item{expression_values}{expression values to use}

\item{feats}{vector of selected features}

\item{vector_name}{column name as seen in \code{\link{pDataDT}}}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}
}
\value{
giotto object if \code{return_gobject = TRUE}, else a vector with % 
results
}
\description{
Calculates the total percentage of (normalized) counts for a 
subset of selected genes
}
\examples{
g <- GiottoData::loadGiottoMini("visium")

addFeatsPerc(g, feats = c("Gm19935", "9630013A20Rik", "2900040C04Rik"))
}
