% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{addGeneMetadata}
\alias{addGeneMetadata}
\title{Add gene metadata}
\usage{
addGeneMetadata(gobject, new_metadata, by_column = F, column_gene_ID = NULL)
}
\arguments{
\item{gobject}{giotto object}

\item{new_metadata}{new metadata to use}

\item{by_column}{merge metadata based on gene_ID column in \code{\link{fDataDT}}}

\item{column_gene_ID}{column name of new metadata to use if \code{by_column = TRUE}}
}
\value{
giotto object
}
\description{
adds gene metadata to the giotto object
}
\details{
You can add additional gene metadata in two manners:
1. Provide a data.table or data.frame with gene annotations in the same order as the gene_ID column in fDataDT(gobject)
2. Provide a data.table or data.frame with gene annotations and specify which column contains the gene IDs,
these gene IDs need to match with the gene_ID column in fDataDT(gobject)
}
