% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{addGenesPerc}
\alias{addGenesPerc}
\title{addGenesPerc}
\usage{
addGenesPerc(
  gobject,
  expression_values = c("normalized", "scaled", "custom"),
  genes = NULL,
  vector_name = "gene_perc",
  return_gobject = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{expression_values}{expression values to use}

\item{genes}{vector of selected genes}

\item{vector_name}{column name as seen in pDataDT()}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}
}
\value{
giotto object if return_gobject = TRUE, else a vector with % results
}
\description{
calculates the total percentage of (normalized) counts for a subset of selected genes
}
\examples{

data(mini_giotto_single_cell)

# select genes (e.g. Rpl or mitochondrial)
random_genes = sample(slot(mini_giotto_single_cell, 'gene_ID'), 5)

# calculate percentage of those selected genes per cells/spot
updated_giotto_object = addGenesPerc(mini_giotto_single_cell,
                                     genes = random_genes,
                                     vector_name = 'random_gene_perc')

# visualize result in data.table format
pDataDT(updated_giotto_object)


}
