% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{addGeneStatistics}
\alias{addGeneStatistics}
\title{addGeneStatistics}
\usage{
addGeneStatistics(
  gobject,
  expression_values = c("normalized", "scaled", "custom"),
  detection_threshold = 0,
  return_gobject = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{expression_values}{expression values to use}

\item{detection_threshold}{detection threshold to consider a gene detected}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}
}
\value{
giotto object if return_gobject = TRUE
}
\description{
adds gene statistics to the giotto object
}
\details{
This function will add the following statistics to gene metadata:
\itemize{
  \item{nr_cells: }{Denotes in how many cells the gene is detected}
  \item{per_cells: }{Denotes in what percentage of cells the gene is detected}
  \item{total_expr: }{Shows the total sum of gene expression in all cells}
  \item{mean_expr: }{Average gene expression in all cells}
  \item{mean_expr_det: }{Average gene expression in cells with detectable levels of the gene}
}
}
\examples{

data(mini_giotto_single_cell)

updated_giotto_object = addGeneStatistics(mini_giotto_single_cell)

}
