% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{addGiottoImage}
\alias{addGiottoImage}
\title{addGiottoImage}
\usage{
addGiottoImage(
  gobject = NULL,
  images = NULL,
  largeImages = NULL,
  spat_loc_name = NULL,
  scale_factor = NULL,
  negative_y = TRUE
)
}
\arguments{
\item{gobject}{gobject to add images objects to}

\item{images}{list of giottoImages to add}

\item{largeImages}{list of giottoLargeImages to add}

\item{spat_loc_name}{provide spatial location slot in Giotto to align giottoImages. Defaults to first one}

\item{scale_factor}{provide scale of image pixel dimensions relative to spatial coordinates.}

\item{negative_y}{Map image to negative y spatial values if TRUE during automatic alignment. Meaning that origin is in upper left instead of lower left.}
}
\value{
an updated Giotto object with access to the list of images
}
\description{
Adds lists of giottoImages and giottoLargeImages to gobjects
}
\seealso{
Other basic image functions: 
\code{\link{plotGiottoImage}()},
\code{\link{reconnectGiottoImage}()},
\code{\link{updateGiottoImage}()}
}
\concept{basic image functions}
