% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{addGiottoImage}
\alias{addGiottoImage}
\title{addGiottoImage}
\usage{
addGiottoImage(gobject, images)
}
\arguments{
\item{gobject}{giotto object}

\item{images}{list of giotto image objects, see \code{\link{createGiottoImage}}}
}
\value{
an updated Giotto object with access to the list of images
}
\description{
Adds giotto image objects to your giotto object
}
