% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{addGiottoImageToSpatPlot}
\alias{addGiottoImageToSpatPlot}
\title{addGiottoImageToSpatPlot}
\usage{
addGiottoImageToSpatPlot(spatpl = NULL, gimage = NULL)
}
\arguments{
\item{spatpl}{a spatial ggplot object}

\item{gimage}{a giotto image, see \code{\link{createGiottoImage}}}
}
\value{
an updated spatial ggplot object
}
\description{
Add a giotto image to a spatial ggplot object post creation
}
