% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{addGiottoPoints}
\alias{addGiottoPoints}
\title{Add giotto points object to giotto object}
\usage{
addGiottoPoints(gobject, gpoints)
}
\arguments{
\item{gobject}{giotto object}

\item{gpoints}{list of giotto point objects, see \code{\link{createGiottoPoints}}}
}
\value{
giotto object
}
\description{
Adds Giotto points to an existing Giotto object
}
\concept{polygon}
