% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{addGiottoPolygons}
\alias{addGiottoPolygons}
\title{Add giotto polygons to giotto object}
\usage{
addGiottoPolygons(gobject, gpolygons)
}
\arguments{
\item{gobject}{giotto object}

\item{gpolygons}{list of giotto polygon objects,
see \code{\link{createGiottoPolygonsFromMask}} and \code{\link{createGiottoPolygonsFromDfr}}}
}
\value{
giotto object
}
\description{
Adds Giotto polygon to an existing Giotto object
}
\concept{polygon}
