% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_hmrf.R
\name{addHMRF}
\alias{addHMRF}
\title{addHMRF}
\usage{
addHMRF(gobject, HMRFoutput, k = NULL, betas_to_add = NULL, hmrf_name = NULL)
}
\arguments{
\item{gobject}{giotto object}

\item{HMRFoutput}{HMRF output from doHMRF()}

\item{k}{number of domains}

\item{betas_to_add}{results from different betas that you want to add}

\item{hmrf_name}{specify a custom name}
}
\value{
giotto object
}
\description{
Add selected results from doHMRF to the giotto object
}
