% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_structures.R
\name{addSpatialCentroidLocations}
\alias{addSpatialCentroidLocations}
\title{addSpatialCentroidLocations}
\usage{
addSpatialCentroidLocations(
  gobject,
  poly_info = "cell",
  spat_loc_name = "raw",
  return_gobject = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{poly_info}{polygon information}

\item{return_gobject}{return giotto object (default: TRUE)}

\item{feat_type}{feature type}
}
\value{

}
\description{
Calculates the centroid locations for the giotto polygons
}
\keyword{centroid}
