% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_structures.R
\name{addSpatialCentroidLocations}
\alias{addSpatialCentroidLocations}
\title{addSpatialCentroidLocations}
\usage{
addSpatialCentroidLocations(
  gobject,
  poly_info = "cell",
  feat_type = NULL,
  spat_loc_name = "raw",
  init_metadata = TRUE,
  return_gobject = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{poly_info}{polygon information}

\item{feat_type}{feature type}

\item{spat_loc_name}{name to give to the created spatial locations}

\item{init_metadata}{initialize cell and feature metadata for this spatial unit
(default = TRUE, but should be turned off if generated earlier in the workflow)}

\item{return_gobject}{return giotto object (default: TRUE)}

\item{verbose}{be verbose}
}
\value{
If \code{return_gobject = TRUE} the giotto object containing the calculated
  polygon centroids will be returned. If \code{return_gobject = FALSE} only the
  generated polygon centroids will be returned.
}
\description{
Calculates the centroid locations for the polygons within one or more selected layers
}
\concept{centroid}
