% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{addStatistics}
\alias{addStatistics}
\title{addStatistics}
\usage{
addStatistics(
  gobject,
  expression_values = c("normalized", "scaled", "custom"),
  detection_threshold = 0,
  return_gobject = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{expression_values}{expression values to use}

\item{detection_threshold}{detection threshold to consider a gene detected}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}
}
\value{
giotto object if return_gobject = TRUE, else a list with results
}
\description{
adds genes and cells statistics to the giotto object
}
\details{
See \code{\link{addGeneStatistics}} and \code{\link{addCellStatistics}}
}
\examples{

data(mini_giotto_single_cell)

updated_giotto_object = addStatistics(mini_giotto_single_cell)

}
