% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{addStatistics}
\alias{addStatistics}
\title{addStatistics}
\usage{
addStatistics(
  gobject,
  feat_type = NULL,
  spat_unit = NULL,
  stats = c("feature", "cell"),
  expression_values = c("normalized", "scaled", "custom"),
  detection_threshold = 0,
  return_gobject = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{feat_type}{feature type}

\item{spat_unit}{spatial unit}

\item{stats}{character. What statistics to add.
default = c("cell", "feature") See details}

\item{expression_values}{expression values to use}

\item{detection_threshold}{detection threshold to consider a feature detected}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}

\item{verbose}{be verbose}
}
\value{
giotto object if return_gobject = TRUE, else a list with results
}
\description{
Adds feature and cell statistics to the giotto object
}
\section{\code{stats} options}{
"feature" - includes \code{\link{addFeatStatistics}} results
"cell" - includes \code{\link{addCellStatistics}} results
"area" - includes polygon areas
}

\examples{
g <- GiottoData::loadGiottoMini("visium")

addStatistics(g)
}
