% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{aggregateStacksPolygons}
\alias{aggregateStacksPolygons}
\title{aggregateStacksPolygons}
\usage{
aggregateStacksPolygons(gobject, spat_units, new_spat_unit = "aggregate")
}
\arguments{
\item{gobject}{giotto object}

\item{spat_units}{spatial units to aggregate}

\item{new_spat_unit}{new name for aggregated spatial unit}
}
\value{
giotto object
}
\description{
aggregate polygons from different z-stacks
}
\seealso{
Other aggregate stacks: 
\code{\link{aggregateStacksExpression}()},
\code{\link{aggregateStacksLocations}()},
\code{\link{aggregateStacksPolygonOverlaps}()},
\code{\link{aggregateStacks}()}
}
\concept{aggregate stacks}
