% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_visuals.R
\name{all_plots_save_function}
\alias{all_plots_save_function}
\title{all_plots_save_function}
\usage{
all_plots_save_function(
  gobject,
  plot_object,
  save_dir = NULL,
  save_folder = NULL,
  save_name = NULL,
  default_save_name = "giotto_plot",
  save_format = NULL,
  show_saved_plot = F,
  ncol = 1,
  nrow = 1,
  scale = 1,
  base_width = NULL,
  base_height = NULL,
  base_aspect_ratio = NULL,
  units = NULL,
  dpi = NULL,
  limitsize = TRUE,
  ...
)
}
\arguments{
\item{gobject}{giotto object}

\item{plot_object}{object to plot}

\item{save_dir}{directory to save to}

\item{save_folder}{folder in save_dir to save to}

\item{save_name}{name of plot}

\item{default_save_name}{default name to save a plot}

\item{save_format}{format (e.g. png, tiff, pdf, ...)}

\item{show_saved_plot}{load & display the saved plot}

\item{ncol}{number of columns}

\item{nrow}{number of rows}

\item{scale}{scale}

\item{base_width}{width}

\item{base_height}{height}

\item{base_aspect_ratio}{aspect ratio}

\item{units}{units}

\item{dpi}{Plot resolution}

\item{limitsize}{When TRUE (the default), ggsave will not save images larger than 50x50 inches, to prevent the common error of specifying dimensions in pixels.}

\item{\dots}{additional parameters to ggplot_save_function or general_save_function}
}
\description{
Function to automatically save plots to directory of interest
}
\seealso{
\code{\link{general_save_function}}
}
\keyword{internal}
