% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interoperability.R
\name{anndataToGiotto}
\alias{anndataToGiotto}
\title{Convert anndata to Giotto}
\usage{
anndataToGiotto(
  anndata_path = NULL,
  n_key_added = NULL,
  spatial_n_key_added = NULL,
  deluanay_spat_net = TRUE,
  spat_unit = NULL,
  feat_type = NULL,
  python_path = NULL
)
}
\arguments{
\item{anndata_path}{path to the .h5ad file}

\item{n_key_added}{equivalent of "key_added" argument from scanpy.pp.neighbors().
If multiple spatial networks are in the anndata object, a list of key_added
        terms may be provided.
        If converting an anndata object from giottoToAnnData, a .txt file may be
        provided, which was generated in that function, 
            i.e. {spat_unit}_{feat_type}_nn_network_keys_added.txt
Cannot be "spatial". This becomes the name of the nearest network in the gobject.}

\item{spatial_n_key_added}{equivalent of "key_added" argument from squidpy.gr.spatial_neighbors. 
If multiple spatial networks are in the anndata object, a list of key_added
terms may be provided.
If converting an anndata object from giottoToAnnData, a .txt file may be
provided, which was generated in that function, 
    i.e. {spat_unit}_{feat_type}_spatial_network_keys_added.txt
Cannot be the same as n_key_added.}

\item{deluanay_spat_net}{binary parameter for spatial network. If TRUE, the spatial network is a deluanay network.}

\item{spat_unit}{desired spatial unit for conversion, default NULL}

\item{feat_type}{desired feature type for conversion, default NULL}

\item{python_path}{path to python executable within a conda/miniconda environment}
}
\value{
Giotto object
}
\description{
Converts a spatial anndata (e.g. scanpy) .h5ad file into a Giotto object
}
\details{
Function in beta. Converts a .h5ad file into a Giotto object.
   The returned Giotto Object will take default insructions with the
   exception of the python path, which may be customized.
   See \code{\link{changeGiottoInstructions}} to modify instructions after creation.
}
