% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interoperability.R
\name{anndataToGiotto}
\alias{anndataToGiotto}
\title{Convert anndata to Giotto}
\usage{
anndataToGiotto(anndata_path = NULL, instructions = NULL, ...)
}
\arguments{
\item{anndata_path}{path to the .h5ad file}

\item{instructions}{giotto instructions (optional)}
}
\value{
Giotto object
}
\description{
Converts a spatial anndata (e.g. scanpy) .h5ad file into a Giotto object
}
\details{
Function in beta. Converts a .h5ad file into a Giotto object.
}
