% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interoperability.R
\name{anndataToGiotto}
\alias{anndataToGiotto}
\title{anndataToGiotto}
\usage{
anndataToGiotto(
  anndata_path,
  metadata_cols = c("total_counts", "pct_counts_mt"),
  instructions = NULL,
  ...
)
}
\arguments{
\item{anndata_path}{path to the .h5ad file}

\item{metadata_cols}{metadata columns to include}

\item{instructions}{giotto instructions}

\item{\dots}{additional parameters to \code{\link{createGiottoObject}}}
}
\value{
Giotto object
}
\description{
Converts a spatial anndata (e.g. scanpy) .h5ad file into a Giotto object
}
\details{
Function in beta. Converts a .h5ad file into a Giotto object.
}
