% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{annotateGiotto}
\alias{annotateGiotto}
\title{annotateGiotto}
\usage{
annotateGiotto(gobject, annotation_vector = NULL,
  cluster_column = NULL, name = "cell_types")
}
\arguments{
\item{gobject}{giotto object}

\item{annotation_vector}{named annotation vector (names = cluster ids)}

\item{cluster_column}{cluster column to convert to annotation names}

\item{name}{new name for annotation column}
}
\value{
giotto object
}
\description{
adds cell annotation to giotto object based on clustering
}
\details{
Description of how to add cell metadata ...
}
\examples{
    annotateGiotto(gobject)
}
