% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_structures.R
\name{annotateSpatialGrid}
\alias{annotateSpatialGrid}
\title{annotateSpatialGrid}
\usage{
annotateSpatialGrid(
  gobject,
  spatial_grid_name = "spatial_grid",
  cluster_columns = NULL
)
}
\arguments{
\item{gobject}{Giotto object}

\item{spatial_grid_name}{name of spatial grid, see \code{\link{showGrids}}}

\item{cluster_columns}{names of cell metadata, see \code{\link{pDataDT}}}
}
\value{
annotated spatial grid data.table
}
\description{
annotate spatial grid with cell ID and cell metadata (optional)
}
