% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction.R
\name{annotateSpatialNetwork}
\alias{annotateSpatialNetwork}
\title{annotateSpatialNetwork}
\usage{
annotateSpatialNetwork(
  gobject,
  spatial_network_name = "spatial_network",
  cluster_column
)
}
\arguments{
\item{gobject}{giotto object}

\item{spatial_network_name}{name of spatial network to use}

\item{cluster_column}{name of column to use for clusters}
}
\value{
annotated network in data.table format
}
\description{
Annotate spatial network with cell metadata information.
}
\examples{
    annotateSpatialNetwork(gobject)
}
