% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_genes.R
\name{binSpect}
\alias{binSpect}
\title{binSpect}
\usage{
binSpect(
  gobject,
  bin_method = c("kmeans", "rank"),
  expression_values = c("normalized", "scaled", "custom"),
  subset_genes = NULL,
  spatial_network_name = "Delaunay_network",
  nstart = 3,
  iter_max = 10,
  percentage_rank = 30,
  do_fisher_test = TRUE,
  calc_hub = FALSE,
  hub_min_int = 3,
  get_av_expr = TRUE,
  get_high_expr = TRUE,
  do_parallel = TRUE,
  cores = NA,
  verbose = T
)
}
\arguments{
\item{gobject}{giotto object}

\item{bin_method}{method to binarize gene expression}

\item{expression_values}{expression values to use}

\item{subset_genes}{only select a subset of genes to test}

\item{spatial_network_name}{name of spatial network to use (default = 'spatial_network')}

\item{nstart}{kmeans: nstart parameter}

\item{iter_max}{kmeans: iter.max parameter}

\item{percentage_rank}{percentage of top cells for binarization}

\item{do_fisher_test}{perform fisher test}

\item{calc_hub}{calculate the number of hub cells}

\item{hub_min_int}{minimum number of cell-cell interactions for a hub cell}

\item{get_av_expr}{calculate the average expression per gene of the high expressing cells}

\item{get_high_expr}{calculate the number of high expressing cells  per gene}

\item{do_parallel}{run calculations in parallel with mclapply}

\item{cores}{number of cores to use if do_parallel = TRUE}

\item{verbose}{be verbose}
}
\value{
data.table with results (see details)
}
\description{
Previously: binGetSpatialGenes. BinSpect (Binary Spatial Extraction of genes) is a fast computational method
that identifies genes with a spatially coherent expression pattern.
}
\details{
We provide two ways to identify spatial genes based on gene expression binarization.
Both methods are identicial except for how binarization is performed.
\itemize{
  \item{1. binarize: }{Each gene is binarized (0 or 1) in each cell with \bold{kmeans} (k = 2) or based on \bold{rank} percentile}
  \item{2. network: }{Alll cells are connected through a spatial network based on the physical coordinates}
  \item{3. contingency table: }{A contingency table is calculated based on all edges of neighboring cells and the binarized expression (0-0, 0-1, 1-0 or 1-1)}
  \item{4. For each gene an odds-ratio (OR) and fisher.test (optional) is calculated}
}
Other statistics are provided (optional):
\itemize{
  \item{Number of cells with high expression (binary = 1)}
  \item{Average expression of each gene within high expressing cells }
  \item{Number of hub cells, these are high expressing cells that have a user defined number of high expressing neighbors}
}
By selecting a subset of likely spatial genes (e.g. soft thresholding highly variable genes) or using multiple cores can accelerate the speed.
}
\examples{
    binSpect(gobject)
}
