% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_structures.R
\name{calculate_distance_and_weight}
\alias{calculate_distance_and_weight}
\title{calculate_distance_and_weight}
\usage{
calculate_distance_and_weight(
  networkDT = NULL,
  sdimx = "sdimx",
  sdimy = "sdimy",
  sdimz = "sdimz",
  d2_or_d3 = c(2, 3)
)
}
\arguments{
\item{networkDT}{spatial network as data.table}

\item{sdimx}{spatial dimension x}

\item{sdimy}{spatial dimension y}

\item{sdimz}{spatial dimension z}

\item{d2_or_d3}{number of dimensions}
}
\description{
calculate_distance_and_weight
}
\keyword{internal}
