% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_genes.R
\name{calculateHVG}
\alias{calculateHVG}
\title{calculateHVG}
\usage{
calculateHVG(
  gobject,
  expression_values = c("normalized", "scaled", "custom"),
  method = c("cov_groups", "cov_loess"),
  reverse_log_scale = FALSE,
  logbase = 2,
  expression_threshold = 0,
  nr_expression_groups = 20,
  zscore_threshold = 1.5,
  HVGname = "hvg",
  difference_in_cov = 0.1,
  show_plot = NA,
  return_plot = NA,
  save_plot = NA,
  save_param = list(),
  default_save_name = "HVGplot",
  return_gobject = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{expression_values}{expression values to use}

\item{method}{method to calculate highly variable genes}

\item{reverse_log_scale}{reverse log-scale of expression values (default = FALSE)}

\item{logbase}{if reverse_log_scale is TRUE, which log base was used?}

\item{expression_threshold}{expression threshold to consider a gene detected}

\item{nr_expression_groups}{number of expression groups for cov_groups}

\item{zscore_threshold}{zscore to select hvg for cov_groups}

\item{HVGname}{name for highly variable genes in cell metadata}

\item{difference_in_cov}{minimum difference in coefficient of variance required}

\item{show_plot}{show plot}

\item{return_plot}{return ggplot object}

\item{save_plot}{directly save the plot [boolean]}

\item{save_param}{list of saving parameters from \code{\link{all_plots_save_function}}}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}
}
\value{
giotto object highly variable genes appended to gene metadata (fDataDT)
}
\description{
compute highly variable genes
}
\details{
Currently we provide 2 ways to calculate highly variable genes:

\bold{1. high coeff of variance (COV) within groups: } \cr
First genes are binned (\emph{nr_expression_groups}) into average expression groups and
the COV for each gene is converted into a z-score within each bin. Genes with a z-score
higher than the threshold (\emph{zscore_threshold}) are considered highly variable.  \cr

\bold{2. high COV based on loess regression prediction: } \cr
A predicted COV is calculated for each gene using loess regression (COV~log(mean expression))
Genes that show a higher than predicted COV (\emph{difference_in_cov}) are considered highly variable. \cr
}
\examples{

data(mini_giotto_single_cell) # loads existing Giotto object

# update a giotto object
mini_giotto_single_cell <- calculateHVG(gobject = mini_giotto_single_cell,
                                        zscore_threshold = 0.1,
                                        nr_expression_groups = 3)

# return a data.table with the high variable genes annotated
hvg_dt <- calculateHVG(gobject = mini_giotto_single_cell,
                       zscore_threshold = 0.1, nr_expression_groups = 3,
                       return_plot = FALSE, return_gobject = FALSE)

# return the ggplot object
hvg_plot <- calculateHVG(gobject = mini_giotto_single_cell,
                       zscore_threshold = 0.1, nr_expression_groups = 3,
                       return_plot = TRUE, return_gobject = FALSE)


}
