% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{calculateMetaTable}
\alias{calculateMetaTable}
\title{calculateMetaTable}
\usage{
calculateMetaTable(
  gobject,
  expression_values = c("normalized", "scaled", "custom"),
  metadata_cols = NULL,
  selected_genes = NULL
)
}
\arguments{
\item{gobject}{giotto object}

\item{expression_values}{expression values to use}

\item{metadata_cols}{annotation columns found in pDataDT(gobject)}

\item{selected_genes}{subset of genes to use}
}
\value{
data.table with average expression values for each gene per (combined) annotation
}
\description{
calculates the average gene expression for one or more (combined) annotation columns.
}
\examples{

data(mini_giotto_single_cell)

# show cell metadata
pDataDT(mini_giotto_single_cell)

# show average gene expression per annotated cell type
calculateMetaTable(mini_giotto_single_cell,
                   metadata_cols = 'cell_types')

}
