% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_structures.R
\name{calculateOverlapSerial}
\alias{calculateOverlapSerial}
\title{calculateOverlapSerial}
\usage{
calculateOverlapSerial(
  gobject,
  name_overlap = NULL,
  spatial_info = "cell",
  feat_info = "rna",
  poly_ID_names = "all",
  polygon_group_size = 500,
  return_gobject = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{gobject}{giotto object}

\item{name_overlap}{name for the overlap results (default to feat_info parameter)}

\item{spatial_info}{polygon information}

\item{feat_info}{feature information}

\item{poly_ID_names}{list of poly_IDs to use}

\item{polygon_group_size}{number of polygons to process per group}

\item{return_gobject}{return giotto object (default: TRUE)}

\item{verbose}{be verbose}
}
\value{
giotto object or spatVector with overlapping information
}
\description{
calculate overlap between cellular structures (polygons) and features (points)
}
\details{
Serial overlapping function that works on groups of polygons at a time.
  Number of polygons per group is defined by \code{polygon_group_size} param
}
\keyword{overlap}
