% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction_visuals.R
\name{cellProximityBarplot}
\alias{cellProximityBarplot}
\title{cellProximityBarplot}
\usage{
cellProximityBarplot(
  gobject,
  CPscore,
  min_orig_ints = 5,
  min_sim_ints = 5,
  p_val = 0.05,
  show_plot = NULL,
  return_plot = NULL,
  save_plot = NULL,
  save_param = list(),
  default_save_name = "cellProximityBarplot"
)
}
\arguments{
\item{gobject}{giotto object}

\item{CPscore}{CPscore, output from cellProximityEnrichment()}

\item{min_orig_ints}{filter on minimum original cell-cell interactions}

\item{min_sim_ints}{filter on minimum simulated cell-cell interactions}

\item{p_val}{p-value}

\item{show_plot}{logical. show plot}

\item{return_plot}{logical. return ggplot object}

\item{save_plot}{logical. save the plot}

\item{save_param}{list of saving parameters, see 
\code{\link{showSaveParameters}}}

\item{default_save_name}{default save name for saving, don't change, 
change save_name in save_param}
}
\value{
ggplot barplot
}
\description{
Create barplot from cell-cell proximity scores
}
\details{
This function creates a barplot that shows the  spatial proximity
 enrichment or depletion of cell type pairs.
}
\examples{
g <- GiottoData::loadGiottoMini("visium")

cellProximityBarplot(
    gobject = g,
    CPscore = cellProximityEnrichment(g, cluster_column = "leiden_clus")
)
}
