% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction.R
\name{cellProximityEnrichment}
\alias{cellProximityEnrichment}
\title{cellProximityEnrichment}
\usage{
cellProximityEnrichment(gobject,
  spatial_network_name = "spatial_network", cluster_column,
  number_of_simulations = 100)
}
\arguments{
\item{gobject}{giotto object}

\item{spatial_network_name}{name of spatial network to use}

\item{cluster_column}{name of column to use for clusters}

\item{number_of_simulations}{number of simulations to create expected observations}
}
\value{
Cell Proximity scores (CPscores) in data.table format
}
\description{
Compute cell-cell interaction enrichment (observed vs expected)
}
\examples{
    cellProximityEnrichment(gobject)
}
