% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction_spot.R
\name{cellProximityEnrichmentEachSpot}
\alias{cellProximityEnrichmentEachSpot}
\title{cellProximityEnrichmentEachSpot}
\usage{
cellProximityEnrichmentEachSpot(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  spatial_network_name = "spatial_network",
  cluster_column = "cell_ID"
)
}
\arguments{
\item{gobject}{giotto object}

\item{feat_type}{feature type}

\item{spatial_network_name}{name of spatial network to use}

\item{cluster_column}{name of column to use for clusters}
}
\value{
matrix that rownames are cell-cell interaction pairs and colnames are cell_IDs
}
\description{
Compute cell-cell interaction enrichment for each spot with its
interacted spots (observed)
}
