% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction_visuals.R
\name{cellProximityHeatmap}
\alias{cellProximityHeatmap}
\title{cellProximityHeatmap}
\usage{
cellProximityHeatmap(CPscore, scale = T, order_cell_types = T,
  color_breaks = NULL, color_names = NULL, show_plot = F,
  return_plot = TRUE, save_plot = FALSE, save_param = list(...))
}
\arguments{
\item{CPscore}{CPscore, output from cellProximityEnrichment()}

\item{scale}{scale cell-cell proximity interaction scores}

\item{order_cell_types}{order cell types based on enrichment correlation}

\item{color_breaks}{numerical vector of length 3 to represent min, mean and maximum}

\item{color_names}{character color vector of length 3}

\item{show_plot}{show plots}

\item{return_plot}{return ggplot object}

\item{save_plot}{directly save the plot [boolean]}

\item{save_param}{list of saving parameters from all_plots_save_function()}
}
\value{
ggplot heatmap
}
\description{
Create heatmap from cell-cell proximity scores
}
\details{
This function creates a heatmap that shows the  spatial proximity
 enrichment or depletion of cell type pairs.
}
\examples{
    cellProximityHeatmap(CPscore)
}
