% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction_visuals.R
\name{cellProximityHeatmap}
\alias{cellProximityHeatmap}
\title{cellProximityHeatmap}
\usage{
cellProximityHeatmap(
  gobject,
  CPscore,
  scale = T,
  order_cell_types = T,
  color_breaks = NULL,
  color_names = NULL,
  show_plot = NULL,
  return_plot = NULL,
  save_plot = NULL,
  save_param = list(),
  default_save_name = "cellProximityHeatmap"
)
}
\arguments{
\item{gobject}{giotto object}

\item{CPscore}{CPscore, output from cellProximityEnrichment()}

\item{scale}{scale cell-cell proximity interaction scores}

\item{order_cell_types}{order cell types based on enrichment correlation}

\item{color_breaks}{numerical vector of length 3 to represent min, mean and maximum}

\item{color_names}{character color vector of length 3}

\item{show_plot}{logical. show plot}

\item{return_plot}{logical. return ggplot object}

\item{save_plot}{logical. save the plot}

\item{save_param}{list of saving parameters, see \code{\link{showSaveParameters}}}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}
}
\value{
ggplot heatmap
}
\description{
Create heatmap from cell-cell proximity scores
}
\details{
This function creates a heatmap that shows the  spatial proximity
 enrichment or depletion of cell type pairs.
}
