% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction_visuals.R
\name{cellProximityNetwork}
\alias{cellProximityNetwork}
\title{cellProximityNetwork}
\usage{
cellProximityNetwork(
  gobject,
  CPscore,
  remove_self_edges = FALSE,
  self_loop_strength = 0.1,
  color_depletion = "lightgreen",
  color_enrichment = "red",
  rescale_edge_weights = TRUE,
  edge_weight_range_depletion = c(0.1, 1),
  edge_weight_range_enrichment = c(1, 5),
  layout = c("Fruchterman", "DrL", "Kamada-Kawai"),
  only_show_enrichment_edges = F,
  edge_width_range = c(0.1, 2),
  node_size = 4,
  node_text_size = 6,
  show_plot = NA,
  return_plot = NA,
  save_plot = NA,
  save_param = list(),
  default_save_name = "cellProximityNetwork"
)
}
\arguments{
\item{gobject}{giotto object}

\item{CPscore}{CPscore, output from cellProximityEnrichment()}

\item{remove_self_edges}{remove enrichment/depletion edges with itself}

\item{self_loop_strength}{size of self-loops}

\item{color_depletion}{color for depleted cell-cell interactions}

\item{color_enrichment}{color for enriched cell-cell interactions}

\item{rescale_edge_weights}{rescale edge weights (boolean)}

\item{edge_weight_range_depletion}{numerical vector of length 2 to rescale depleted edge weights}

\item{edge_weight_range_enrichment}{numerical vector of length 2 to rescale enriched edge weights}

\item{layout}{layout algorithm to use to draw nodes and edges}

\item{only_show_enrichment_edges}{show only the enriched pairwise scores}

\item{edge_width_range}{range of edge width}

\item{node_size}{size of nodes}

\item{node_text_size}{size of node labels}

\item{show_plot}{show plot}

\item{return_plot}{return ggplot object}

\item{save_plot}{directly save the plot [boolean]}

\item{save_param}{list of saving parameters from \code{\link{all_plots_save_function}}}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}
}
\value{
igraph plot
}
\description{
Create network from cell-cell proximity scores
}
\details{
This function creates a network that shows the  spatial proximity
 enrichment or depletion of cell type pairs.
}
