% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction_visuals.R
\name{cellProximitySpatPlot}
\alias{cellProximitySpatPlot}
\title{cellProximitySpatPlot}
\usage{
cellProximitySpatPlot(gobject, ...)
}
\arguments{
\item{gobject}{giotto object}

\item{interaction_name}{cell-cell interaction name}

\item{cluster_column}{cluster column with cell clusters}

\item{sdimx}{x-axis dimension name (default = 'sdimx')}

\item{sdimy}{y-axis dimension name (default = 'sdimy')}

\item{cell_color}{color for cells (see details)}

\item{cell_color_code}{named vector with colors}

\item{color_as_factor}{convert color column to factor}

\item{show_other_cells}{decide if show cells not in network}

\item{show_network}{show underlying spatial network}

\item{network_color}{color of spatial network}

\item{spatial_network_name}{name of spatial network to use}

\item{show_grid}{show spatial grid}

\item{grid_color}{color of spatial grid}

\item{spatial_grid_name}{name of spatial grid to use}

\item{coord_fix_ratio}{fix ratio between x and y-axis}

\item{show_legend}{show legend}

\item{point_size_select}{size of selected points}

\item{point_select_border_col}{border color of selected points}

\item{point_select_border_stroke}{stroke size of selected points}

\item{point_size_other}{size of other points}

\item{point_other_border_col}{border color of other points}

\item{point_other_border_stroke}{stroke size of other points}

\item{show_plot}{show plots}

\item{return_plot}{return ggplot object}

\item{save_plot}{directly save the plot [boolean]}

\item{save_param}{list of saving parameters from \code{\link{all_plots_save_function}}}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}
}
\value{
ggplot
}
\description{
Visualize 2D cell-cell interactions according to spatial coordinates in ggplot mode
}
\details{
Description of parameters.
}
\examples{
    cellProximitySpatPlot(gobject)
}
\seealso{
\code{\link{cellProximitySpatPlot2D}} and \code{\link{cellProximitySpatPlot3D}} for 3D
}
