% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction_visuals.R
\name{cellProximitySpatPlot3D}
\alias{cellProximitySpatPlot3D}
\title{cellProximitySpatPlot2D}
\usage{
cellProximitySpatPlot3D(
  gobject,
  interaction_name = NULL,
  cluster_column = NULL,
  sdimx = "sdimx",
  sdimy = "sdimy",
  sdimz = "sdimz",
  cell_color = NULL,
  cell_color_code = NULL,
  color_as_factor = T,
  show_other_cells = T,
  show_network = T,
  show_other_network = F,
  network_color = NULL,
  spatial_network_name = "Delaunay_network",
  show_grid = F,
  grid_color = NULL,
  spatial_grid_name = "spatial_grid",
  show_legend = T,
  point_size_select = 4,
  point_size_other = 2,
  point_alpha_other = 0.5,
  axis_scale = c("cube", "real", "custom"),
  custom_ratio = NULL,
  x_ticks = NULL,
  y_ticks = NULL,
  z_ticks = NULL,
  show_plot = NA,
  return_plot = NA,
  save_plot = NA,
  save_param = list(),
  default_save_name = "cellProximitySpatPlot3D",
  ...
)
}
\arguments{
\item{gobject}{giotto object}

\item{interaction_name}{cell-cell interaction name}

\item{cluster_column}{cluster column with cell clusters}

\item{sdimx}{x-axis dimension name (default = 'sdimx')}

\item{sdimy}{y-axis dimension name (default = 'sdimy')}

\item{sdimz}{z-axis dimension name (default = 'sdimz')}

\item{cell_color}{color for cells (see details)}

\item{cell_color_code}{named vector with colors}

\item{color_as_factor}{convert color column to factor}

\item{show_other_cells}{decide if show cells not in network}

\item{show_network}{show spatial network of selected cells}

\item{show_other_network}{show spatial network of not selected cells}

\item{network_color}{color of spatial network}

\item{spatial_network_name}{name of spatial network to use}

\item{show_grid}{show spatial grid}

\item{grid_color}{color of spatial grid}

\item{spatial_grid_name}{name of spatial grid to use}

\item{show_legend}{show legend}

\item{point_size_select}{size of selected points}

\item{point_size_other}{size of other points}

\item{point_alpha_other}{opacity of other points}

\item{axis_scale}{scale of axis}

\item{custom_ratio}{custom ratio of axes}

\item{x_ticks}{ticks on x-axis}

\item{y_ticks}{ticks on y-axis}

\item{z_ticks}{ticks on z-axis}

\item{show_plot}{show plots}

\item{return_plot}{return plotly object}

\item{save_plot}{directly save the plot [boolean]}

\item{save_param}{list of saving parameters from \code{\link{all_plots_save_function}}}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}

\item{\dots}{additional parameters}
}
\value{
plotly
}
\description{
Visualize 3D cell-cell interactions according to spatial coordinates in plotly mode
}
\details{
Description of parameters.
}
