% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction_visuals.R
\name{cellProximityVisPlot}
\alias{cellProximityVisPlot}
\title{cellProximityVisPlot}
\usage{
cellProximityVisPlot(
  gobject,
  interaction_name = NULL,
  cluster_column = NULL,
  sdimx = NULL,
  sdimy = NULL,
  sdimz = NULL,
  cell_color = NULL,
  cell_color_code = NULL,
  color_as_factor = T,
  show_other_cells = F,
  show_network = F,
  show_other_network = F,
  network_color = NULL,
  spatial_network_name = "Delaunay_network",
  show_grid = F,
  grid_color = NULL,
  spatial_grid_name = "spatial_grid",
  coord_fix_ratio = 1,
  show_legend = T,
  point_size_select = 2,
  point_select_border_col = "black",
  point_select_border_stroke = 0.05,
  point_size_other = 1,
  point_alpha_other = 0.3,
  point_other_border_col = "lightgrey",
  point_other_border_stroke = 0.01,
  axis_scale = c("cube", "real", "custom"),
  custom_ratio = NULL,
  x_ticks = NULL,
  y_ticks = NULL,
  z_ticks = NULL,
  plot_method = c("ggplot", "plotly"),
  ...
)
}
\arguments{
\item{gobject}{giotto object}

\item{interaction_name}{cell-cell interaction name}

\item{cluster_column}{cluster column with cell clusters}

\item{sdimx}{x-axis dimension name (default = 'sdimx')}

\item{sdimy}{y-axis dimension name (default = 'sdimy')}

\item{sdimz}{z-axis dimension name (default = 'sdimz')}

\item{cell_color}{color for cells (see details)}

\item{cell_color_code}{named vector with colors}

\item{color_as_factor}{convert color column to factor}

\item{show_other_cells}{show not selected cells}

\item{show_network}{show underlying spatial network}

\item{show_other_network}{show underlying spatial network of other cells}

\item{network_color}{color of spatial network}

\item{spatial_network_name}{name of spatial network to use}

\item{show_grid}{show spatial grid}

\item{grid_color}{color of spatial grid}

\item{spatial_grid_name}{name of spatial grid to use}

\item{coord_fix_ratio}{fix ratio between x and y-axis}

\item{show_legend}{show legend}

\item{point_size_select}{size of selected points}

\item{point_select_border_col}{border color of selected points}

\item{point_select_border_stroke}{stroke size of selected points}

\item{point_size_other}{size of other points}

\item{point_alpha_other}{alpha of other points}

\item{point_other_border_col}{border color of other points}

\item{point_other_border_stroke}{stroke size of other points}

\item{axis_scale}{scale of axis}

\item{custom_ratio}{custom ratio of scales}

\item{x_ticks}{x ticks}

\item{y_ticks}{y ticks}

\item{z_ticks}{z ticks}

\item{plot_method}{method to plot}

\item{\dots}{additional parameters}
}
\value{
ggplot or plotly
}
\description{
Visualize cell-cell interactions according to spatial coordinates
}
\details{
Description of parameters.
}
