% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto.R
\name{changeGiottoInstructions}
\alias{changeGiottoInstructions}
\title{changeGiottoInstructions}
\usage{
changeGiottoInstructions(
  gobject,
  params = NULL,
  new_values = NULL,
  return_gobject = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{params}{parameter(s) to change}

\item{new_values}{new value(s) for parameter(s)}

\item{return_gobject}{(boolean) return giotto object}
}
\value{
giotto object with one or more changed instructions
}
\description{
Function to change one or more instructions from giotto object
}
