% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto.R
\name{check_spatial_location_data}
\alias{check_spatial_location_data}
\title{Check spatial location data}
\usage{
check_spatial_location_data(gobject)
}
\description{
check cell ID (spatial unit) names between spatial location and expression data.
It will look for identical IDs after sorting.
}
\keyword{internal}
