% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_hmrf.R
\name{checkAndFixSpatialGenes}
\alias{checkAndFixSpatialGenes}
\title{checkAndFixSpatialGenes}
\usage{
checkAndFixSpatialGenes(gobject, use_spatial_genes, use_score = FALSE)
}
\arguments{
\item{gobject}{TBD}

\item{use_spatial_genes}{TBD}

\item{use_score}{TBD}
}
\description{
checkAndFixSpatialGenes
}
