% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_environment.R
\name{checkGiottoEnvironment}
\alias{checkGiottoEnvironment}
\title{checkGiottoEnvironment}
\usage{
checkGiottoEnvironment(verbose = TRUE)
}
\arguments{
\item{verbose}{be verbose}
}
\description{
checkGiottoEnvironment
}
\details{
Checks if a miniconda giotto environment can be found.
Can be installed with \code{\link{installGiottoEnvironment}}.
}
