% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{checkSpatLocsObj}
\alias{checkSpatLocsObj}
\title{Check spatialLocationsObj}
\usage{
checkSpatLocsObj(object)
}
\arguments{
\item{object}{S4 spatialLocationsObj to check}
}
\description{
Check function for S4 spatialLocationsObj
}
\keyword{internal}
