% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_hmrf.R
\name{chooseAvailableSpatialGenes}
\alias{chooseAvailableSpatialGenes}
\title{chooseAvailableSpatialGenes}
\usage{
chooseAvailableSpatialGenes(gobject, spat_unit = NULL, feat_type = NULL)
}
\arguments{
\item{gobject}{Giotto object}

\item{spat_unit}{spatial unit}

\item{feat_type}{feature type}
}
\description{
function to find the test name for existing spatial gene sets in Giotto
}
\details{
This function outputs the available test name for existing spatial gene sets in Giotto, 
which could be used in parameter ‘name’ in `filterSpatialGenes`. 
Priorities for showing the spatial gene test names are ‘binSpect’ > ‘silhouetteRankTest’ > ‘silhouetteRank’.
}
\keyword{internal}
