% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clusterCells}
\alias{clusterCells}
\title{clusterCells}
\usage{
clusterCells(gobject, nn_network_to_use = "sNN",
  network_name = "sNN.pca", cluster_method = c("python_leiden",
  "python_louvain", "louvain", "randomwalk", "sNNclust"),
  pyth_leid_resolution = 1, pyth_leid_weight_col = "weight",
  pyth_leid_part_type = c("RBConfigurationVertexPartition",
  "ModularityVertexPartition"), pyth_leid_init_memb = NULL,
  pyth_leid_iterations = 200, pyth_louv_resolution = 1,
  pyth_louv_weight_col = NULL, python_louv_random = F,
  python_path = NULL, louvain_gamma = 1, louvain_omega = 1,
  walk_steps = 4, walk_clusters = 10, walk_weights = NA,
  sNNclust_k = 20, sNNclust_eps = 4, sNNclust_minPts = 16,
  borderPoints = TRUE, name = "plouvain_clus", return_gobject = TRUE,
  set_seed = F, seed_number = 1234, ...)
}
\arguments{
\item{gobject}{giotto object}

\item{nn_network_to_use}{type of NN network to use (kNN vs sNN)}

\item{network_name}{name of NN network to use}

\item{cluster_method}{community cluster method to use}

\item{pyth_louv_resolution}{python louvain param: resolution}

\item{pyth_louv_weight_col}{python louvain param: weight column}

\item{python_louv_random}{python louvain param: random}

\item{python_path}{specify specific path to python if required}

\item{louvain_gamma}{louvain param: gamma or resolution}

\item{louvain_omega}{louvain param: omega}

\item{walk_steps}{randomwalk: number of steps}

\item{walk_clusters}{randomwalk: number of clusters}

\item{walk_weights}{randomwalk: weight column}

\item{sNNclust_k}{SNNclust: k neighbors to use}

\item{sNNclust_eps}{SNNclust: epsilon}

\item{sNNclust_minPts}{SNNclust: min points}

\item{borderPoints}{SNNclust: border points}

\item{name}{name for cluster}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}

\item{set_seed}{set seed}

\item{seed_number}{number for seed}

\item{...}{additional parameters}
}
\value{
giotto object appended with new cluster
}
\description{
cluster cells using a NN-network and community detection algorithms
}
\details{
Description of different clustering methods.
}
\examples{
    clusterCells(gobject)
}
