% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{color_tag}
\alias{color_tag}
\alias{use_color_text}
\alias{ansi_colors}
\alias{is_emacs_with_color}
\alias{emacs_version}
\title{Colorize print text}
\usage{
color_tag()

use_color_text()

ansi_colors()

is_emacs_with_color()

emacs_version()
}
\description{
Colorize print text
}
\details{
supported colors checking is modified from \pkg{cli}
\href{https://github.com/r-lib/cli/blob/HEAD/R/num-ansi-colors.R}{aab-num-ansi-colors.R}
}
\section{Functions}{
\itemize{
\item \code{use_color_text()}: Determine if system should print color

\item \code{ansi_colors()}: Determine if system can print at least 8 colors

\item \code{is_emacs_with_color()}: Determine if emacs can print color

\item \code{emacs_version()}: Determine emacs version

}}
\keyword{internal}
