% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction.R
\name{combineInteractionChangedGenes}
\alias{combineInteractionChangedGenes}
\title{combineInteractionChangedGenes}
\usage{
combineInteractionChangedGenes(
  cpgObject,
  selected_ints = NULL,
  selected_genes = NULL,
  specific_genes_1 = NULL,
  specific_genes_2 = NULL,
  min_cells = 5,
  min_int_cells = 3,
  min_fdr = 0.05,
  min_spat_diff = 0,
  min_log2_fc = 0.5,
  do_parallel = TRUE,
  cores = NA,
  verbose = T
)
}
\arguments{
\item{cpgObject}{ICG (interaction changed gene) score object}

\item{selected_ints}{subset of selected cell-cell interactions (optional)}

\item{selected_genes}{subset of selected genes (optional)}

\item{specific_genes_1}{specific geneset combo (need to position match specific_genes_2)}

\item{specific_genes_2}{specific geneset combo (need to position match specific_genes_1)}

\item{min_cells}{minimum number of target cell type}

\item{min_int_cells}{minimum number of interacting cell type}

\item{min_fdr}{minimum adjusted p-value}

\item{min_spat_diff}{minimum absolute spatial expression difference}

\item{min_log2_fc}{minimum absolute log2 fold-change}

\item{do_parallel}{run calculations in parallel with mclapply}

\item{cores}{number of cores to use if do_parallel = TRUE}

\item{verbose}{verbose}
}
\value{
cpgObject that contains the filtered differential gene scores
}
\description{
Combine ICG scores in a pairwise manner.
}
