% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{combineMetadata}
\alias{combineMetadata}
\title{combineMetadata}
\usage{
combineMetadata(gobject, spat_enr_names = NULL)
}
\arguments{
\item{gobject}{Giotto object}

\item{spat_enr_names}{names of spatial enrichment results to include}
}
\value{
Extended cell metadata in data.table format.
}
\description{
This function combines the cell metadata with spatial locations and
enrichment results from \code{\link{runSpatialEnrich}}
}
