% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_structures.R
\name{convert_to_full_spatial_network}
\alias{convert_to_full_spatial_network}
\title{convert_to_full_spatial_network}
\usage{
convert_to_full_spatial_network(reduced_spatial_network_DT)
}
\arguments{
\item{reduced_spatial_network_DT}{reduced spatial network in data.table format}
}
\description{
convert to a full spatial network
}
\keyword{internal}
