% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{copy-generic}
\alias{copy-generic}
\alias{copy}
\alias{copy,coordDataDT-method}
\title{Copy an entire object}
\usage{
copy(x)

\S4method{copy}{coordDataDT}(x)
}
\arguments{
\item{x}{a Giotto S4 class subobject}
}
\description{
S4 generic for Giotto's S4 subobjects to return with full copies of
certain subobjects that usually return referenced information.
}
\section{Functions}{
\itemize{
\item \code{copy,coordDataDT-method}: Copy \emph{data.table}-based spatial locations object.
}}

\seealso{
\code{\link[data.table]{copy}} \code{\link[terra]{deepcopy}}
}
