% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_section.R
\name{create_2d_mesh_grid_line_obj}
\alias{create_2d_mesh_grid_line_obj}
\title{create_2d_mesh_grid_line_obj}
\usage{
create_2d_mesh_grid_line_obj(x_min, x_max, y_min, y_max, mesh_grid_n)
}
\arguments{
\item{x_min}{x_min}

\item{x_max}{x_max}

\item{y_min}{y_min}

\item{y_max}{y_max}

\item{mesh_grid_n}{mesh_grid_n}
}
\description{
create 2d mesh grid line object
}
\keyword{internal}
