% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{create_average_detection_DT}
\alias{create_average_detection_DT}
\title{create_average_detection_DT}
\usage{
create_average_detection_DT(
  gobject,
  meta_data_name,
  expression_values = c("normalized", "scaled", "custom"),
  detection_threshold = 0
)
}
\arguments{
\item{gobject}{giotto object}

\item{meta_data_name}{name of metadata column to use}

\item{expression_values}{which expression values to use}

\item{detection_threshold}{detection threshold to consider a gene detected}
}
\value{
data.table with average gene epression values for each factor
}
\description{
calculates average gene detection for a cell metadata factor (e.g. cluster)
}
\keyword{internal}
