% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_structures.R
\name{create_delaunayNetwork3D}
\alias{create_delaunayNetwork3D}
\title{create_delaunayNetwork3D}
\usage{
create_delaunayNetwork3D(
  gobject,
  method = "delaunayn_geometry",
  sdimx = "sdimx",
  sdimy = "sdimy",
  sdimz = "sdimz",
  name = "delaunay_network_3D",
  maximum_distance = "auto",
  minimum_k = 0,
  options = "Pp",
  return_gobject = TRUE,
  ...
)
}
\description{
Create a spatial 3D Delaunay network.
}
\keyword{internal}
