% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_structures.R
\name{create_delaunayNetwork_deldir}
\alias{create_delaunayNetwork_deldir}
\title{create_delaunayNetwork_deldir}
\usage{
create_delaunayNetwork_deldir(
  spatial_locations,
  sdimx = "sdimx",
  sdimy = "sdimy",
  ...
)
}
\description{
Create a spatial Delaunay network with deldir
}
\keyword{internal}
