% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_structures.R
\name{create_delaunayNetwork_geometry_3D}
\alias{create_delaunayNetwork_geometry_3D}
\title{create_delaunayNetwork_geometry_3D}
\usage{
create_delaunayNetwork_geometry_3D(
  spatial_locations,
  sdimx = "sdimx",
  sdimy = "sdimy",
  sdimz = "sdimz",
  options = options,
  ...
)
}
\description{
Create a spatial 3D Delaunay network with geometry
}
\keyword{internal}
