% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{create_giotto_polygon_object}
\alias{create_giotto_polygon_object}
\title{Create a giotto polygon object}
\usage{
create_giotto_polygon_object(
  name = "cell",
  spatVector = NULL,
  spatVectorCentroids = NULL,
  overlaps = NULL,
  unique_IDs = NULL
)
}
\arguments{
\item{name}{name of polygon object}

\item{spatVector}{SpatVector of polygons}

\item{spatVectorCentroids}{(optional) SpatVector of polygon centroids}

\item{overlaps}{(optional) feature overlaps of polygons}

\item{unique_IDs}{unique polygon IDs for cacheing}
}
\description{
Create a giotto polygon object
}
\keyword{internal}
