% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.R
\name{create_screeplot}
\alias{create_screeplot}
\title{create_screeplot}
\usage{
create_screeplot(pca_obj, ncp = 20, ylim = c(0, 20))
}
\arguments{
\item{pca_obj}{pca dimension reduction object}

\item{ncp}{number of principal components to calculate}

\item{ylim}{y-axis limits on scree plot}
}
\value{
ggplot
}
\description{
create screeplot with ggplot
}
\keyword{internal}
