% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_structures.R
\name{create_spatialNetworkObject}
\alias{create_spatialNetworkObject}
\title{create_spatialNetworkObject}
\usage{
create_spatialNetworkObject(
  name = NULL,
  method = NULL,
  parameters = NULL,
  outputObj = NULL,
  networkDT = NULL,
  cellShapeObj = NULL,
  networkDT_before_filter = NULL,
  crossSectionObjects = NULL,
  misc = NULL
)
}
\arguments{
\item{name}{name}

\item{method}{method}

\item{parameters}{parameters}

\item{outputObj}{outputObj}

\item{networkDT}{networkDT}

\item{cellShapeObj}{cellShapeObj}

\item{networkDT_before_filter}{networkDT_before_filter}

\item{crossSectionObjects}{crossSectionObjects}
}
\description{
creates a spatial network object to store the created spatial network and additional information
}
\keyword{internal}
