% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_structures.R
\name{create_spatvector_object_from_dfr}
\alias{create_spatvector_object_from_dfr}
\title{Create terra spatvector object from a data.frame}
\usage{
create_spatvector_object_from_dfr(x, verbose = TRUE)
}
\arguments{
\item{x}{data.frame object}

\item{verbose}{be verbose}
}
\description{
create terra spatvector from a data.frame where cols 1 and 2 must
be x and y coordinates respectively. Additional columns are set as attributes
to the points where the first additional (col 3) should be the feat_ID.
}
\keyword{internal}
