% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_bento.R
\name{createBentoAdata}
\alias{createBentoAdata}
\title{Create bento adata object from gobject}
\usage{
createBentoAdata(gobject)
}
\arguments{
\item{gobject}{Giotto object}
}
\value{
bento_adata bento adata object
}
\description{
Create bento adata object from gobject
}
