% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_section.R
\name{createCrossSection}
\alias{createCrossSection}
\title{createCrossSection}
\usage{
createCrossSection(
  gobject,
  name = "cross_section",
  spatial_network_name = "Delaunay_network",
  thickness_unit = c("cell", "natural"),
  slice_thickness = 2,
  extend_ratio = 0.2,
  method = c("equation", "3 points", "point and norm vector",
    "point and two plane vectors"),
  point1 = NULL,
  point2 = NULL,
  point3 = NULL,
  normVector = NULL,
  planeVector1 = NULL,
  planeVector2 = NULL,
  equation = NULL,
  mesh_grid_n = 20,
  return_gobject = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}
}
\value{
giotto object with updated spatial network slot
}
\description{
Create a cross section.
}
\examples{

}
