% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createEpiObject.R
\name{createEpiObject}
\alias{createEpiObject}
\title{Create epigenetic giotto object}
\usage{
createEpiObject(
  fragmentsPath,
  genome = c("hg19", "hg38", "mm9", "mm10"),
  expression = NULL,
  expression_feat = "atac",
  spatial_locs = NULL,
  sampleNames = "sample1",
  minFrags = 0,
  maxFrags = 1e+07,
  minTSS = 0,
  addTileMat = TRUE,
  addGeneScoreMat = TRUE,
  offsetPlus = 0,
  offsetMinus = 0,
  TileMatParams = list(tileSize = 5000),
  force = TRUE,
  outputDirectory = getwd(),
  copyArrows = FALSE,
  useMatrix = "TileMatrix",
  name = "IterativeLSI",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{fragmentsPath}{A character vector containing the paths to the input files to use to generate the ArrowFiles. These files can be in one of the following formats: (i) scATAC tabix files, (ii) fragment files, or (iii) bam files.}

\item{genome}{A string indicating the default genome to be used for all ArchR functions. Currently supported values include "hg19","hg38","mm9", and "mm10". This value is stored as a global environment variable, not part of the ArchRProject. This can be overwritten on a per-function basis using the given function's geneAnnotationand genomeAnnotation parameter. For something other than one of the currently supported, see createGeneAnnnotation() and createGenomeAnnnotation()}

\item{expression}{expression information}

\item{expression_feat}{available features (e.g. atac, rna, ...)}

\item{spatial_locs}{data.table or data.frame with coordinates for cell centroids}

\item{sampleNames}{A character vector containing the names to assign to the samples}

\item{minFrags}{minFrags}

\item{maxFrags}{maxFrags}

\item{minTSS}{minTSS}

\item{addTileMat}{addTileMat}

\item{addGeneScoreMat}{addGeneScoreMat}

\item{offsetPlus}{offsetPlus}

\item{offsetMinus}{offsetMinus}

\item{TileMatParams}{TileMatParams}

\item{force}{TRUE}

\item{outputDirectory}{outputDirectory}

\item{copyArrows}{copyArrows}

\item{useMatrix}{useMatrix}

\item{name}{name}

\item{verbose}{Default = TRUE}

\item{...}{Additional params passed to `createGiottoObject`}
}
\value{
A Giotto object with at least an atac or epigenetic modality
}
\description{
Create epigenetic giotto object
}
