% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{createExprObj}
\alias{createExprObj}
\title{Create S4 exprObj}
\usage{
createExprObj(
  expression_data,
  name = "test",
  spat_unit = "cell",
  feat_type = "rna",
  provenance = NULL,
  misc = NULL,
  expression_matrix_class = c("dgCMatrix", "HDF5Matrix", "rhdf5"),
  h5_file = NULL
)
}
\arguments{
\item{expression_data}{expression data}

\item{name}{name of exprObj}

\item{spat_unit}{spatial unit of expression (e.g. 'cell')}

\item{feat_type}{feature type of expression (e.g. 'rna', 'protein')}

\item{provenance}{origin data of expression information (if applicable)}

\item{misc}{misc}
}
\description{
Create an S4 exprObj
}
