% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto.R
\name{createGiottoCosMxObject}
\alias{createGiottoCosMxObject}
\title{Create Nanostring CosMx Giotto Object}
\usage{
createGiottoCosMxObject(
  cosmx_dir = NULL,
  data_to_use = c("all", "subcellular", "aggregate"),
  FOVs = NULL,
  instructions = NULL,
  cores = NA,
  verbose = TRUE
)
}
\arguments{
\item{cosmx_dir}{full path to the exported cosmx directory}

\item{data_to_use}{which type(s) of expression data to build the gobject with
Default is \code{'all'} information available. \code{'subcellular'} loads the transcript
coordinates only. \code{'aggregate'} loads the provided aggregated expression matrix.}

\item{FOVs}{field of views to load (only affects subcellular data and images)}

\item{instructions}{list of instructions or output result from \code{\link{createGiottoInstructions}}}

\item{cores}{how many cores or threads to use to read data if paths are provided}

\item{verbose}{be verbose when building Giotto object}
}
\value{
a giotto object
}
\description{
Given the path to a CosMx experiment directory, creates a Giotto
object.
}
