% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_cosmx.R
\name{createGiottoCosMxObject}
\alias{createGiottoCosMxObject}
\title{Create Nanostring CosMx Giotto Object}
\usage{
createGiottoCosMxObject(
  cosmx_dir,
  version = "default",
  FOVs = NULL,
  feat_type = c("rna", "negprobes"),
  split_keyword = list("NegPrb"),
  load_images = list(composite = "composite", overlay = "overlay"),
  load_expression = FALSE,
  load_cellmeta = FALSE,
  fov_shifts_path = NULL,
  transcript_path = NULL,
  cell_labels_dir = NULL,
  expression_path = NULL,
  metadata_path = NULL,
  cores = determine_cores(),
  verbose = NULL,
  instructions = NULL,
  remove_unvalid_polygons = deprecated(),
  data_to_use = deprecated(),
  remove_background_polygon = deprecated(),
  background_algo = deprecated(),
  ...
)
}
\arguments{
\item{cosmx_dir}{full path to the exported cosmx directory}

\item{version}{character. Version of CosMx output. Current selections are
either "default", "v6", and "legacy" (for the NSCLC dataset).}

\item{FOVs}{field of views to load (only affects subcellular data and images)}

\item{feat_type}{character. feature type. Provide more than one value if
using the \code{split_keyword} param. For each set of keywords to split by, an
additional feat_type should be provided in the same order. Affects how
the transcripts information is loaded. Helpful for separating out the
QC probes. See details.}

\item{split_keyword}{list of character vectors of keywords to split the
transcripts based on their feat_ID. Keywords will be \code{grepl()}
matched against the feature IDs information. See details.}

\item{load_images}{named list of filepaths to directories. Loads the
composite and overlay images by default.}

\item{load_expression}{logical. (Default = FALSE) whether to load provided
expression matrix}

\item{load_cellmeta}{logical. (Default = TRUE) whether to load provided
cell metadata}

\item{fov_shifts_path}{Optional. Filepath to fov_positions_file}

\item{transcript_path}{Optional. Filepath to desired transcripts file to
load.}

\item{cell_labels_dir}{Optional. Path to directory containing CellLabels
images to load as polygons.}

\item{expression_path}{Optional. Filepath to cell feature matrix to load.}

\item{metadata_path}{Optional. Filepath to metadata file to load.}

\item{cores}{nthreads to use when loading in cell metadata and transcripts}

\item{verbose}{be verbose when building Giotto object}

\item{instructions}{list of instructions or output result
from \code{\link[GiottoClass]{createGiottoInstructions}}}

\item{remove_unvalid_polygons}{deprecated. Now always done}

\item{data_to_use}{deprecated. Not used}

\item{remove_background_polygon}{deprecated. Now always done}

\item{background_algo}{deprecated. Not used}

\item{...}{
  Arguments passed on to \code{\link[=importCosMx]{importCosMx}}
  \describe{
    \item{\code{slide}}{numeric. Slide number. Defaults to 1}
    \item{\code{micron}}{logical. Whether to scale spatial information as micron
instead of the default pixels}
    \item{\code{px2um}}{numeric. Scalefactor from pixels to micron. Defaults to 0.12028
based on `CosMx-ReadMe.html` info}
  }}
}
\value{
a giotto object
}
\description{
Given the path to a CosMx experiment directory, creates a Giotto
object.
}
\details{
[\strong{Expected Directory}] This function generates a giotto object when
given a link to a cosmx output directory. It expects the following items
within the directory where the \strong{bolded} portions are what this
function matches against:
\itemize{
\item{\strong{CellComposite} (folder of images)}
\item{\strong{CellLabels} (folder of images)}
\item{\strong{CellOverlay} (folder of images)}
\item{\strong{CompartmentLabels} (folder of images)}
\item{experimentname_\strong{exprMat_file}.csv (file)}
\item{experimentname_\strong{fov_positions_file}.csv (file)}
\item{experimentname_\strong{metadata_file}.csv (file)}
\item{experimentname_\strong{tx_file}.csv (file)}
}

[\strong{feat_type and split_keyword}]
Additional QC probe information is in the subcellular feature detections
information and must be separated from the gene expression information
during processing.
The QC probes have prefixes that allow them to be selected from the rest of
the feature IDs.
Giotto uses \code{feat_type} and \code{split_keyword} params to select these QC
probes out as separate feature types. See examples in
\verb{[GiottoClass::createGiottoPoints]} for how this works.

The Gene expression subset labeled as \code{rna} is accepted as the subset of
feat_IDs that do not get matched to any of the \code{split_keywords}.

[\strong{Images}] Images in the expected CellComposite and CellOverlay
folders will be loaded as giotto largeImage objects by default.
\code{\link{showGiottoImageNames}} can be used to see the available images.
}
