% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{createGiottoImage}
\alias{createGiottoImage}
\title{createGiottoImage}
\usage{
createGiottoImage(
  gobject = NULL,
  spatial_locs = NULL,
  mg_object,
  name = "image",
  xmax_adj = 0,
  xmin_adj = 0,
  ymax_adj = 0,
  ymin_adj = 0
)
}
\arguments{
\item{gobject}{giotto object}

\item{spatial_locs}{spatial locations (alternative if giobject = NULL)}

\item{mg_object}{magick image object}

\item{name}{name for the image}

\item{xmax_adj}{adjustment of the maximum x-value to align the image}

\item{xmin_adj}{adjustment of the minimum x-value to align the image}

\item{ymax_adj}{adjustment of the maximum y-value to align the image}

\item{ymin_adj}{adjustment of the minimum y-value to align the image}
}
\value{
a giotto image object
}
\description{
Creates a giotto image that can be added to a Giotto object and/or used to add an image to the spatial plotting functions
}
