% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto.R
\name{createGiottoInstructions}
\alias{createGiottoInstructions}
\title{createGiottoInstructions}
\usage{
createGiottoInstructions(python_path = NULL, save_dir = NULL,
  plot_format = NULL, dpi = NULL)
}
\arguments{
\item{python_path}{path to python bin to use}

\item{save_dir}{path of directory to use to save figures}

\item{dpi}{resolution for raster images}
}
\value{
named vector with giotto instructions
}
\description{
Function to create instructions for giotto functions
}
\examples{
    createGiottoInstructions()
}
