% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto.R
\name{createGiottoInstructions}
\alias{createGiottoInstructions}
\title{createGiottoInstructions}
\usage{
createGiottoInstructions(python_path = NULL, show_plot = NULL,
  return_plot = NULL, save_plot = NULL, save_dir = NULL,
  plot_format = NULL, dpi = NULL, units = NULL, height = NULL,
  width = NULL)
}
\arguments{
\item{python_path}{path to python bin to use}

\item{show_plot}{print plot to console, default = TRUE}

\item{return_plot}{return plot as object, default = TRUE}

\item{save_plot}{automatically save plot, dafault = FALSE}

\item{save_dir}{path of directory to use to save figures}

\item{dpi}{resolution for raster images to save}

\item{height}{height of the plots to save}

\item{width}{width of the plots to save}
}
\value{
named vector with giotto instructions
}
\description{
Function to create instructions for giotto functions
}
\examples{
    createGiottoInstructions()
}
