% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto.R
\name{createGiottoInstructions}
\alias{createGiottoInstructions}
\title{createGiottoInstructions}
\usage{
createGiottoInstructions(
  python_path = NULL,
  show_plot = NULL,
  return_plot = NULL,
  save_plot = NULL,
  save_dir = NULL,
  plot_format = NULL,
  dpi = NULL,
  units = NULL,
  height = NULL,
  width = NULL,
  is_docker = FALSE
)
}
\arguments{
\item{python_path}{path to python binary to use}

\item{show_plot}{print plot to console, default = TRUE}

\item{return_plot}{return plot as object, default = TRUE}

\item{save_plot}{automatically save plot, dafault = FALSE}

\item{save_dir}{path to directory where to save plots}

\item{plot_format}{format of plots (defaults to png)}

\item{dpi}{resolution for raster images}

\item{units}{units of format (defaults to in)}

\item{height}{height of plots}

\item{width}{width of  plots}

\item{is_docker}{using docker implementation of Giotto (defaults to FALSE)}
}
\value{
named vector with giotto instructions
}
\description{
Function to set global instructions for giotto functions
}
\seealso{
More online information can be found here \url{https://rubd.github.io/Giotto_site/articles/instructions_and_plotting.html}
}
