% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{createGiottoMerscopeObject}
\alias{createGiottoMerscopeObject}
\alias{createGiottoMerscopeObject_subcellular}
\alias{createGiottoMerscopeObject_aggregate}
\title{Create Vizgen MERSCOPE Giotto Object}
\usage{
createGiottoMerscopeObject(
  merscope_dir,
  data_to_use = c("subcellular", "aggregate"),
  FOVs = NULL,
  poly_z_indices = 1:7,
  calculate_overlap = TRUE,
  overlap_to_matrix = TRUE,
  aggregate_stack = TRUE,
  aggregate_stack_param = list(summarize_expression = "sum", summarize_locations =
    "mean", new_spat_unit = "cell"),
  instructions = NULL,
  cores = NA,
  verbose = TRUE
)

createGiottoMerscopeObject_subcellular(
  data_list,
  calculate_overlap = TRUE,
  overlap_to_matrix = TRUE,
  aggregate_stack = TRUE,
  aggregate_stack_param = list(summarize_expression = "sum", summarize_locations =
    "mean", new_spat_unit = "cell"),
  cores = NA,
  verbose = TRUE
)

createGiottoMerscopeObject_aggregate(data_list, cores = NA, verbose = TRUE)
}
\arguments{
\item{merscope_dir}{full path to the exported merscope directory}

\item{data_to_use}{which of either the 'subcellular' or 'aggregate' information
to use for object creation}

\item{FOVs}{which FOVs to use when building the subcellular object. (default is NULL)
NULL loads all FOVs (very slow)}

\item{calculate_overlap}{whether to run \code{\link{calculateOverlapRaster}}}

\item{overlap_to_matrix}{whether to run \code{\link{overlapToMatrix}}}

\item{aggregate_stack}{whether to run \code{\link{aggregateStacks}}}

\item{aggregate_stack_param}{params to pass to \code{\link{aggregateStacks}}}

\item{instructions}{list of instructions or output result from \code{\link[GiottoClass]{createGiottoInstructions}}}

\item{cores}{how many cores or threads to use to read data if paths are provided}

\item{verbose}{be verbose when building Giotto object}

\item{data_list}{list of loaded data from \code{\link{load_merscope_folder}}}
}
\value{
a giotto object
}
\description{
Given the path to a MERSCOPE experiment directory, creates a Giotto
object.
}
\details{
[\strong{Expected Directory}] This function generates a giotto object when given a
link to a MERSCOPE output directory. It expects the following items within the directory
where the \strong{bolded} portions are what this function matches against:
\itemize{
  \item{\strong{cell_boundaries} (folder .hdf5 files)}
  \item{\strong{images} (folder of .tif images and a scalefactor/transfrom table)}
  \item{\strong{cell_by_gene}.csv (file)}
  \item{cell_metadata\strong{fov_positions_file}.csv (file)}
  \item{detected_transcripts\strong{metadata_file}.csv (file)}
}
}
\section{Functions}{
\itemize{
\item \code{createGiottoMerscopeObject_subcellular()}: Create giotto object with 'subcellular' workflow

\item \code{createGiottoMerscopeObject_aggregate()}: Create giotto object with 'aggregate' workflow

}}
\keyword{internal}
