% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto.R
\name{createGiottoObjectSubcellular}
\alias{createGiottoObjectSubcellular}
\title{Create a giotto object from subcellular data}
\usage{
createGiottoObjectSubcellular(
  gpoints = NULL,
  gpolygons = NULL,
  polygon_mask_list_params = NULL,
  polygon_dfr_list_params = NULL,
  cell_metadata = NULL,
  feat_metadata = NULL,
  spatial_network = NULL,
  spatial_network_name = NULL,
  spatial_grid = NULL,
  spatial_grid_name = NULL,
  spatial_enrichment = NULL,
  spatial_enrichment_name = NULL,
  dimension_reduction = NULL,
  nn_network = NULL,
  images = NULL,
  instructions = NULL,
  cores = NA,
  verbose = TRUE
)
}
\arguments{
\item{gpoints}{giotto points}

\item{gpolygons}{giotto polygons}

\item{polygon_mask_list_params}{list parameters for \code{\link{createGiottoPolygonsFromMask}}}

\item{polygon_dfr_list_params}{list parameters for \code{\link{createGiottoPolygonsFromDfr}}}

\item{cell_metadata}{cell annotation metadata}

\item{feat_metadata}{feature annotation metadata for each unique feature}

\item{spatial_network}{list of spatial network(s)}

\item{spatial_network_name}{list of spatial network name(s)}

\item{spatial_grid}{list of spatial grid(s)}

\item{spatial_grid_name}{list of spatial grid name(s)}

\item{spatial_enrichment}{list of spatial enrichment score(s) for each spatial region}

\item{spatial_enrichment_name}{list of spatial enrichment name(s)}

\item{dimension_reduction}{list of dimension reduction(s)}

\item{nn_network}{list of nearest neighbor network(s)}

\item{images}{list of images}

\item{instructions}{list of instructions or output result from \code{\link{createGiottoInstructions}}}

\item{cores}{how many cores or threads to use to read data if paths are provided}

\item{verbose}{be verbose when building Giotto object}
}
\value{
giotto object
}
\description{
Function to create a giotto object starting from subcellular polygon (e.g. cell) and points (e.g. transcripts) information
}
\concept{giotto}
