% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{createGiottoPoints}
\alias{createGiottoPoints}
\title{Create giotto points object}
\usage{
createGiottoPoints(x, feat_type = "rna", verbose = TRUE, unique_IDs = NULL)
}
\arguments{
\item{x}{spatVector or data.frame-like object with points coordinate information (x, y, feat_ID)}

\item{feat_type}{feature type}

\item{verbose}{be verbose}

\item{unique_IDs}{(optional) character vector of unique IDs present within
the spatVector data. Provided for cacheing purposes}
}
\value{
giottoPoints
}
\description{
Creates Giotto point object from a structured dataframe-like object
}
\concept{polygon}
