% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_stomics.r
\name{createGiottoStereoSeqObject}
\alias{createGiottoStereoSeqObject}
\title{Create Stereo-seq Giotto Object}
\usage{
createGiottoStereoSeqObject(
  stereoseq_dir,
  type = c("squarebin", "cellbin"),
  bin_size = "bin100",
  gene_column = c("geneName", "geneID"),
  flip_spatial_locs = TRUE,
  verbose = TRUE,
  h5_file = NULL,
  instructions = NULL
)
}
\arguments{
\item{stereoseq_dir}{filepath to the exported Stereo-seq directory.}

\item{type}{character. Use "squarebin" to read expression from *.tissue.gef 
file (default) or "cellbin" to read expression from *.adjusted.cellbin.gef 
file.}

\item{bin_size}{bin size to select from *.tissue.gef file. Choose a value 
from "bin1", "bin5", "bin10", "bin20", "bin50", "bin100", "bin150", or 
"bin200". Only needed when using type = "squarebin".}

\item{gene_column}{(optional) character. Which column contains the gene names
within the geneExp information. Choose from "geneName" (default), or "geneID"}

\item{flip_spatial_locs}{Flip spatial locations in the y axis
from \code{\link{createGiottoInstructions}}}

\item{verbose}{logical. Be verbose.}

\item{h5_file}{(optional) name to create an on-disk HDF5 file.}

\item{instructions}{list of instructions or output result}
}
\value{
Giotto Stereo-seq object
}
\description{
Create Stereo-seq Giotto Object
}
