% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_stereoseq.r
\name{createGiottoStereoSeqObject}
\alias{createGiottoStereoSeqObject}
\title{Create Stereo-seq Giotto Object}
\usage{
createGiottoStereoSeqObject(
  stereoseq_dir,
  type = "squarebin",
  bin_size = "bin100",
  gene_column = "geneName",
  negative_y = TRUE,
  shift_polygon_y = 0,
  verbose = TRUE,
  h5_file = NULL,
  instructions = NULL
)
}
\arguments{
\item{stereoseq_dir}{filepath to the exported Stereo-seq directory.}

\item{type}{character. Use "squarebin" to read expression at different bin 
levels (default), "cellbin" to read expression at cell resolution, or 
"subcellular" to read individual transcripts and cell boundaries.}

\item{bin_size}{bin size. Choose a value from "bin1", "bin5", "bin10", 
"bin20", "bin50", "bin100", "bin150", or "bin200". List the available bins 
in your sample using rhdf5::h5ls("path/to/*.tissue.gef"). 
Only needed when using type = "squarebin".}

\item{gene_column}{(optional) character. Which column contains the gene names
within the geneExp information. Choose from "geneName" (default), or "geneID"}

\item{negative_y}{logical. Map data to negative y spatial values during 
automatic alignment (Default = TRUE). Meaning that origin is in upper left 
instead of lower left.
from \code{\link{createGiottoInstructions}}}

\item{verbose}{logical. Be verbose.}

\item{h5_file}{(optional) name to create an on-disk HDF5 file.}

\item{instructions}{list of instructions or output result}
}
\value{
Giotto Stereo-seq object
}
\description{
Create Stereo-seq Giotto Object
}
