% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_visiumHD.R
\name{createGiottoVisiumHDObject}
\alias{createGiottoVisiumHDObject}
\title{Create 10x VisiumHD Giotto Object}
\usage{
createGiottoVisiumHDObject(
  visiumHD_dir = NULL,
  expr_data = c("raw", "filter"),
  gene_column_index = 1,
  instructions = NULL,
  expression_matrix_class = c("dgCMatrix", "DelayedArray"),
  cores = NA,
  verbose = FALSE
)
}
\arguments{
\item{expr_data}{raw or filtered data (see details)}

\item{gene_column_index}{which column index to select (see details)}

\item{expression_matrix_class}{class of expression matrix to use
(e.g. "dgCMatrix", "DelayedArray")}

\item{cores}{how many cores or threads to use to read data if paths are
provided}

\item{verbose}{be verbose}

\item{VisiumHD_dir}{full path to the exported visiumHD directory}
}
\value{
giotto object
}
\description{
Given the path to a VisiumHD output folder, creates a
Giotto object
}
\details{
\itemize{
  \item{expr_data: raw will take expression data from raw_feature_bc_matrix 
  and filter from filtered_feature_bc_matrix}
  \item{gene_column_index: which gene identifiers (names) to use if there 
  are multiple columns (e.g. ensemble and gene symbol)}
}
}
