% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto.R
\name{createGiottoVisiumObject}
\alias{createGiottoVisiumObject}
\title{createGiottoVisiumObject}
\usage{
createGiottoVisiumObject(
  visium_dir = NULL,
  expr_data = c("raw", "filter"),
  gene_column_index = 1,
  h5_visium_path = NULL,
  h5_gene_ids = c("symbols", "ensembl"),
  h5_tissue_positions_path = NULL,
  h5_image_png_path = NULL,
  png_name = NULL,
  xmax_adj = 0,
  xmin_adj = 0,
  ymax_adj = 0,
  ymin_adj = 0,
  instructions = NULL,
  cores = NA,
  verbose = TRUE
)
}
\arguments{
\item{visium_dir}{path to the 10X visium directory [required]}

\item{expr_data}{raw or filtered data (see details)}

\item{gene_column_index}{which column index to select (see details)}

\item{h5_visium_path}{path to visium 10X .h5 file}

\item{h5_gene_ids}{gene names as symbols (default) or ensemble gene ids}

\item{h5_tissue_positions_path}{path to tissue locations (.csv file)}

\item{h5_image_png_path}{path to tissue .png file (optional)}

\item{png_name}{select name of png to use (see details)}

\item{xmax_adj}{adjustment of the maximum x-value to align the image}

\item{xmin_adj}{adjustment of the minimum x-value to align the image}

\item{ymax_adj}{adjustment of the maximum y-value to align the image}

\item{ymin_adj}{adjustment of the minimum y-value to align the image}

\item{instructions}{list of instructions or output result from \code{\link{createGiottoInstructions}}}

\item{cores}{how many cores or threads to use to read data if paths are provided}

\item{verbose}{be verbose}
}
\value{
giotto object
}
\description{
creates Giotto object directly from a 10X visium folder
}
\details{
If starting from a Visium 10X directory:
\itemize{
  \item{expr_data: raw will take expression data from raw_feature_bc_matrix and filter from filtered_feature_bc_matrix}
  \item{gene_column_index: which gene identifiers (names) to use if there are multiple columns (e.g. ensemble and gene symbol)}
  \item{png_name: by default the first png will be selected, provide the png name to override this (e.g. myimage.png)}
}

If starting from a Visium 10X .h5 file
\itemize{
  \item{h5_visium_path: full path to .h5 file: /your/path/to/visium_file.h5}
  \item{h5_tissue_positions_path: full path to spatial locations file: /you/path/to/tissue_positions_list.csv}
  \item{h5_image_png_path: full path to png: /your/path/to/images/tissue_lowres_image.png}
}
}
