% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto.R
\name{createGiottoXeniumObject}
\alias{createGiottoXeniumObject}
\title{Create 10x Xenium Giotto Object}
\usage{
createGiottoXeniumObject(
  xenium_dir,
  data_to_use = c("subcellular", "aggregate"),
  load_format = "csv",
  h5_expression = TRUE,
  h5_gene_ids = c("symbols", "ensembl"),
  gene_column_index = 1,
  bounds_to_load = c("cell"),
  instructions = NULL,
  cores = NA,
  verbose = TRUE
)
}
\arguments{
\item{xenium_dir}{full path to the exported xenium directory}

\item{data_to_use}{which type(s) of expression data to build the gobject with
(e.g. default: \strong{'subcellular'}, 'aggregate', or 'all')}

\item{load_format}{files formats from which to load the data. Either `csv` or
`parquet` currently supported.}

\item{h5_expression}{(boolean) whether to load cell_feature_matrix from .h5 file.
Default is \code{TRUE}}

\item{h5_gene_ids}{use gene symbols (default) or ensembl ids for the .h5 gene
expression matrix}

\item{gene_column_index}{which column from the features or genes .tsv file to use for row ids}

\item{bounds_to_load}{vector of boundary information to load (e.g. \code{'cell'}
or \code{'nucleus'} by themselves or \code{c('cell', 'nucleus')} to load both
at the same time.)}

\item{instructions}{list of instructions or output result from \code{\link{createGiottoInstructions}}}

\item{cores}{how many cores or threads to use to read data if paths are provided}

\item{verbose}{be verbose when building Giotto object}
}
\description{
Given the path to a Xenium experiment output folder, creates a Giotto
object
}
