% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_visuals.R
\name{createHeatmap_DT}
\alias{createHeatmap_DT}
\title{createHeatmap_DT}
\usage{
createHeatmap_DT(
  gobject,
  expression_values = c("normalized", "scaled", "custom"),
  genes,
  cluster_column = NULL,
  cluster_order = c("size", "correlation", "custom"),
  cluster_custom_order = NULL,
  cluster_cor_method = "pearson",
  cluster_hclust_method = "ward.D",
  gene_order = c("correlation", "custom"),
  gene_custom_order = NULL,
  gene_cor_method = "pearson",
  gene_hclust_method = "complete"
)
}
\arguments{
\item{gobject}{giotto object}

\item{expression_values}{expression values to use}

\item{genes}{genes to use}

\item{cluster_column}{name of column to use for clusters}

\item{cluster_order}{method to determine cluster order}

\item{cluster_custom_order}{custom order for clusters}

\item{cluster_cor_method}{method for cluster correlation}

\item{cluster_hclust_method}{method for hierarchical clustering of clusters}

\item{gene_order}{method to determine gene order}

\item{gene_custom_order}{custom order for genes}

\item{gene_cor_method}{method for gene correlation}

\item{gene_hclust_method}{method for hierarchical clustering of genes}
}
\value{
list
}
\description{
creates order for clusters
}
\details{
Creates input data.tables for plotHeatmap function.
}
\keyword{internal}
