% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_enrichment.R
\name{createSpatialEnrich}
\alias{createSpatialEnrich}
\title{createSpatialEnrich}
\usage{
createSpatialEnrich(
  gobject,
  enrich_method = c("PAGE", "rank", "hypergeometric"),
  sign_matrix,
  expression_values = c("normalized", "scaled", "custom"),
  reverse_log_scale = TRUE,
  logbase = 2,
  output_enrichment = c("original", "zscore"),
  name = "PAGE",
  return_gobject = TRUE
)
}
\arguments{
\item{gobject}{Giotto object}

\item{enrich_method}{method for gene signature enrichment calculation}

\item{sign_matrix}{Matrix of signature genes for each cell type / process}

\item{expression_values}{expression values to use}

\item{reverse_log_scale}{reverse expression values from log scale}

\item{logbase}{log base to use if reverse_log_scale = TRUE}

\item{output_enrichment}{how to return enrichment output}

\item{name}{to give to spatial enrichment results, default = PAGE}

\item{return_gobject}{return giotto object}
}
\value{
Giotto object or enrichment results if return_gobject = FALSE
}
\description{
Function to calculate gene signature enrichment scores per spatial position using a hypergeometric test.
}
\examples{
    createSpatialEnrich(gobject)
}
