% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_structures.R
\name{createSpatialKNNnetwork}
\alias{createSpatialKNNnetwork}
\title{createSpatialKNNnetwork}
\usage{
createSpatialKNNnetwork(
  gobject,
  method = "knn",
  name = "knn_network",
  k = 4,
  dimensions = "all",
  maximum_distance = NULL,
  minimum_k = 0,
  verbose = F,
  return_gobject = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}
}
\value{
giotto object with updated spatial network slot
}
\description{
Create a spatial kNN network.
}
\examples{
    createSpatialKNNnetwork(gobject)
}
