% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_structures.R
\name{createSpatialNetwork}
\alias{createSpatialNetwork}
\title{createSpatialNetwork}
\usage{
createSpatialNetwork(
  gobject,
  dimensions = "all",
  method = c("Delaunay", "kNN"),
  delaunay_method = c("delaunayn_geometry", "RTriangle", "deldir"),
  maximum_distance_delaunay = "auto",
  name = NULL,
  knn_method = "knn",
  k = 4,
  minimum_k = 0,
  maximum_distance_knn = NULL,
  verbose = F,
  return_gobject = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{dimensions}{which spatial dimensions to use (default = all)}

\item{method}{which method to use to create a spatial network}

\item{delaunay_method}{Delaunay method to use}

\item{maximum_distance_delaunay}{distance cuttof for nearest neighbors to consider for Delaunay network}

\item{name}{name for spatial network (default = 'spatial_network')}

\item{knn_method}{method to create kNN network}

\item{k}{number of nearest neighbors based on physical distance}

\item{minimum_k}{minimum nearest neigbhours if maximum_distance != NULL}

\item{maximum_distance_knn}{distance cuttof for nearest neighbors to consider for kNN network}

\item{verbose}{verbose}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}
}
\value{
giotto object with updated spatial network slot
}
\description{
Create a spatial network based on cell centroid physical distances.
}
\details{
Creates a spatial network connecting single-cells based on their physical distance to each other.
Number of neighbors can be determined by k, maximum distance from each cell with or without
setting a minimum k for each cell.

\strong{dimensions: } default = 'all' which takes all possible dimensions.
Alternatively you can provide a character vector that specififies the spatial dimensions to use, e.g. c("sdimx', "sdimy")
or a numerical vector, e.g. 2:3

\strong{maximum_distance: } to create a network based on maximum distance only, you also need to set k to a very high value, e.g. k = 100
}
\examples{
    createSpatialNetwork(gobject)
}
