% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_structures.R
\name{createSpatialNetwork}
\alias{createSpatialNetwork}
\title{createSpatialNetwork}
\usage{
createSpatialNetwork(gobject, k = 4, dimensions = "all",
  maximum_distance = NULL, minimum_k = 0, name = "spatial_network",
  verbose = F, return_gobject = TRUE)
}
\arguments{
\item{gobject}{giotto object}

\item{k}{number of nearest neighbors (spatial distance)}

\item{dimensions}{which spatial dimensions to use}

\item{maximum_distance}{cuttof for nearest neighbors to consider}

\item{minimum_k}{minimum nearest neigbhours if maximum_distance != NULL}

\item{name}{name for spatial network (default = 'spatial_network')}

\item{verbose}{verbose}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}
}
\value{
giotto object with updated spatial network slot
}
\description{
create a spatial network based on cell centroid distances
}
\details{
Description of spatial network creation ...
}
\examples{
    createSpatialNetwork(gobject)
}
