% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{cropGiottoLargeImage}
\alias{cropGiottoLargeImage}
\title{Crop a giotto largeImage object}
\usage{
cropGiottoLargeImage(
  gobject = NULL,
  largeImage_name = NULL,
  giottoLargeImage = NULL,
  crop_name = "image",
  crop_extent = NULL,
  xmax_crop = NULL,
  xmin_crop = NULL,
  ymax_crop = NULL,
  ymin_crop = NULL
)
}
\arguments{
\item{gobject}{gobject holding the giottoLargeImage}

\item{largeImage_name}{name of giottoLargeImage within gobject}

\item{giottoLargeImage}{alternative input param using giottoLargeImage object
instead of through \code{gobject} and \code{largeImage_name} params}

\item{crop_name}{arbitrary name for cropped giottoLargeImage}

\item{crop_extent}{terra extent object used to crop the giottoLargeImage}

\item{xmax_crop, xmin_crop, ymax_crop, ymin_crop}{crop min/max x and y bounds}
}
\value{
a giottoLargeImage object
}
\description{
Crop a giottoLargeImage based on crop_extent argument or given values
}
