% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_section.R
\name{crossSectionGenePlot}
\alias{crossSectionGenePlot}
\title{crossSectionGenePlot}
\usage{
crossSectionGenePlot(
  gobject = NULL,
  crossSection_obj = NULL,
  name = NULL,
  spatial_network_name = "Delaunay_network",
  default_save_name = "crossSectionGenePlot",
  ...
)
}
\arguments{
\item{gobject}{giotto object}

\item{crossSection_obj}{crossSection object}

\item{name}{name of virtual cross section to use}

\item{spatial_network_name}{name of spatial network to use}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}

\item{...}{parameters for spatGenePlot2D}
}
\value{
ggplot
}
\description{
Visualize cells and gene expression in a virtual cross section according to spatial coordinates
}
\details{
Description of parameters.
}
\seealso{
\code{\link{spatGenePlot3D}} and \code{\link{spatGenePlot2D}}
}
