% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_section.R
\name{crossSectionPlot3D}
\alias{crossSectionPlot3D}
\title{crossSectionPlot3D}
\usage{
crossSectionPlot3D(
  gobject,
  crossSection_obj = NULL,
  name = NULL,
  spatial_network_name = "Delaunay_network",
  show_other_cells = T,
  other_cell_color = alpha("lightgrey", 0),
  default_save_name = "crossSection3D",
  ...
)
}
\arguments{
\item{gobject}{giotto object}

\item{crossSection_obj}{cross section object as alternative input. default = NULL.}

\item{name}{name of virtual cross section to use}

\item{spatial_network_name}{name of spatial network to use}

\item{show_other_cells}{display not selected cells}

\item{other_cell_color}{color of cells outside the cross section. default = transparent.}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}

\item{...}{parameters for spatPlot3D}
}
\value{
ggplot
}
\description{
Visualize cells in a virtual cross section according to spatial coordinates
}
\details{
Description of parameters.
}
