% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_visuals.R
\name{decide_cluster_order}
\alias{decide_cluster_order}
\title{decide_cluster_order}
\usage{
decide_cluster_order(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  expression_values = c("normalized", "scaled", "custom"),
  feats,
  genes = NULL,
  cluster_column = NULL,
  cluster_order = c("size", "correlation", "custom"),
  cluster_custom_order = NULL,
  cor_method = "pearson",
  hclust_method = "ward.D"
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit (e.g. "cell")}

\item{feat_type}{feature type (e.g. "rna", "dna", "protein")}

\item{expression_values}{expression values to use (e.g. "normalized", "scaled", "custom")}

\item{feats}{features to use (e.g. genes)}

\item{genes}{deprecated, use feats}

\item{cluster_column}{name of column to use for clusters (e.g. "leiden_clus")}

\item{cluster_order}{method to determine cluster order (e.g. "size", "correlation", "custom")}

\item{cluster_custom_order}{custom order for clusters}

\item{cor_method}{method for correlation, default to 'pearson'}

\item{hclust_method}{method for hierarchical clustering, default to 'ward.D'}
}
\value{
custom
}
\description{
creates order for clusters
}
\details{
Calculates order for clusters.
}
\keyword{internal}
