% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_visuals.R
\name{decide_cluster_order}
\alias{decide_cluster_order}
\title{decide_cluster_order}
\usage{
decide_cluster_order(
  gobject,
  expression_values = c("normalized", "scaled", "custom"),
  genes,
  cluster_column = NULL,
  cluster_order = c("size", "correlation", "custom"),
  cluster_custom_order = NULL,
  cor_method = "pearson",
  hclust_method = "ward.D"
)
}
\arguments{
\item{gobject}{giotto object}

\item{expression_values}{expression values to use}

\item{genes}{genes to use}

\item{cluster_column}{name of column to use for clusters}

\item{cluster_order}{method to determine cluster order}

\item{cluster_custom_order}{custom order for clusters}

\item{cor_method}{method for correlation}

\item{hclust_method}{method for hierarchical clustering}
}
\value{
custom
}
\description{
creates order for clusters
}
\details{
Calculates order for clusters.
}
\keyword{internal}
