% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visuals.R
\name{dimCellPlot}
\alias{dimCellPlot}
\title{dimCellPlot}
\usage{
dimCellPlot(gobject, ...)
}
\arguments{
\item{gobject}{giotto object}

\item{...}{
  Arguments passed on to \code{\link[=dimCellPlot2D]{dimCellPlot2D}}
  \describe{
    \item{\code{dim_reduction_to_use}}{dimension reduction to use}
    \item{\code{dim_reduction_name}}{dimension reduction name}
    \item{\code{dim1_to_use}}{dimension to use on x-axis}
    \item{\code{dim2_to_use}}{dimension to use on y-axis}
    \item{\code{spat_enr_names}}{names of spatial enrichment results to include}
    \item{\code{cell_annotation_values}}{numeric cell annotation columns}
    \item{\code{show_NN_network}}{show underlying NN network}
    \item{\code{nn_network_to_use}}{type of NN network to use (kNN vs sNN)}
    \item{\code{network_name}}{name of NN network to use, if show_NN_network = TRUE}
    \item{\code{cell_color_code}}{named vector with colors for cell annotation values}
    \item{\code{cell_color_gradient}}{vector with 3 colors for numeric data}
    \item{\code{gradient_midpoint}}{midpoint for color gradient}
    \item{\code{gradient_limits}}{vector with lower and upper limits}
    \item{\code{select_cell_groups}}{select subset of cells/clusters based on cell_color parameter}
    \item{\code{select_cells}}{select subset of cells based on cell IDs}
    \item{\code{show_other_cells}}{display not selected cells}
    \item{\code{other_cell_color}}{color of not selected cells}
    \item{\code{other_point_size}}{size of not selected cells}
    \item{\code{show_cluster_center}}{plot center of selected clusters}
    \item{\code{show_center_label}}{plot label of selected clusters}
    \item{\code{center_point_size}}{size of center points}
    \item{\code{center_point_border_col}}{border color of center points}
    \item{\code{center_point_border_stroke}}{border stroke size of center points}
    \item{\code{label_size}}{size of labels}
    \item{\code{label_fontface}}{font of labels}
    \item{\code{edge_alpha}}{column to use for alpha of the edges}
    \item{\code{point_shape}}{point with border or not (border or no_border)}
    \item{\code{point_size}}{size of point (cell)}
    \item{\code{point_alpha}}{transparancy of dim. reduction points}
    \item{\code{point_border_col}}{color of border around points}
    \item{\code{point_border_stroke}}{stroke size of border around points}
    \item{\code{show_legend}}{show legend}
    \item{\code{legend_text}}{size of legend text}
    \item{\code{legend_symbol_size}}{size of legend symbols}
    \item{\code{background_color}}{color of plot background}
    \item{\code{axis_text}}{size of axis text}
    \item{\code{axis_title}}{size of axis title}
    \item{\code{cow_n_col}}{cowplot param: how many columns}
    \item{\code{cow_rel_h}}{cowplot param: relative height}
    \item{\code{cow_rel_w}}{cowplot param: relative width}
    \item{\code{cow_align}}{cowplot param: how to align}
    \item{\code{show_plot}}{show plot}
    \item{\code{return_plot}}{return ggplot object}
    \item{\code{save_plot}}{directly save the plot [boolean]}
    \item{\code{save_param}}{list of saving parameters, see \code{\link{showSaveParameters}}}
    \item{\code{default_save_name}}{default save name for saving, don't change, change save_name in save_param}
  }}
}
\value{
ggplot
}
\description{
Visualize cells according to dimension reduction coordinates
}
\details{
Description of parameters. For 3D plots see \code{\link{dimCellPlot2D}}
}
\examples{

data(mini_giotto_single_cell)

# combine all metadata
combineMetadata(mini_giotto_single_cell, spat_enr_names = 'cluster_metagene')

# visualize total expression information
dimCellPlot(mini_giotto_single_cell, cell_annotation_values = 'total_expr')

# visualize enrichment results
dimCellPlot(mini_giotto_single_cell,
               spat_enr_names = 'cluster_metagene',
               cell_annotation_values = c('1','2'))

}
\seealso{
Other dimension reduction cell annotation visualizations: 
\code{\link{dimCellPlot2D}()}
}
\concept{dimension reduction cell annotation visualizations}
