% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visuals.R
\name{dimGenePlot}
\alias{dimGenePlot}
\title{dimGenePlot}
\usage{
dimGenePlot(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=dimGenePlot2D]{dimGenePlot2D}}
  \describe{
    \item{\code{gobject}}{giotto object}
    \item{\code{genes}}{genes to show}
    \item{\code{default_save_name}}{default save name for saving, don't change, change save_name in save_param}
  }}
}
\value{
ggplot
}
\description{
Visualize gene expression according to dimension reduction coordinates
}
\details{
Description of parameters.
}
\seealso{
\code{\link{dimGenePlot3D}}

Other dimension reduction gene expression visualizations: 
\code{\link{dimGenePlot2D}()},
\code{\link{dimGenePlot3D}()}
}
\concept{dimension reduction gene expression visualizations}
