% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{dims-generic}
\alias{dims-generic}
\alias{nrow,giottoPoints-method}
\alias{nrow,giottoPolygon-method}
\alias{nrow,coordDataDT-method}
\alias{nrow,exprData-method}
\alias{ncol,exprData-method}
\alias{dim,exprData-method}
\title{Dimensions of giotto objects}
\usage{
\S4method{nrow}{giottoPoints}(x)

\S4method{nrow}{giottoPolygon}(x)

\S4method{nrow}{coordDataDT}(x)

\S4method{nrow}{exprData}(x)

\S4method{ncol}{exprData}(x)

\S4method{dim}{exprData}(x)
}
\arguments{
\item{x}{object to check dimensions of}
}
\description{
Find the dimensions of an object
}
\section{Functions}{
\itemize{
\item \code{nrow(giottoPoints)}: Find rows of giottoPoints object

\item \code{nrow(giottoPolygon)}: Find rows of giottoPolygon object

\item \code{nrow(coordDataDT)}: Find rows of giotto S4s with data.table based \code{coordinates} slots

\item \code{nrow(exprData)}: Find rows of giotto S4s with data.table based \code{coordinates} slots

\item \code{ncol(exprData)}: Find rows of giotto S4s with data.table based \code{coordinates} slots

\item \code{dim(exprData)}: Find rows of giotto S4s with data.table based \code{coordinates} slots

}}
