% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction.R
\name{do_multi_permuttest_random}
\alias{do_multi_permuttest_random}
\title{do_multi_permuttest_random}
\usage{
do_multi_permuttest_random(
  expr_values,
  select_ind,
  other_ind,
  mean_method,
  offset = 0.1,
  n = 100,
  cores = 2,
  set_seed = TRUE,
  seed_number = 1234
)
}
\description{
calculate multiple random values
}
\keyword{internal}
