% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_genes.R
\name{do_spatial_grid_averaging}
\alias{do_spatial_grid_averaging}
\title{do_spatial_grid_averaging}
\usage{
do_spatial_grid_averaging(
  gobject,
  expression_values = c("normalized", "scaled", "custom"),
  subset_genes = NULL,
  spatial_grid_name = "spatial_grid",
  min_cells_per_grid = 4
)
}
\arguments{
\item{gobject}{giotto object}

\item{expression_values}{gene expression values to use}

\item{subset_genes}{subset of genes to use}

\item{spatial_grid_name}{name of spatial grid to use}

\item{min_cells_per_grid}{minimum number of cells to consider a grid}
}
\value{
matrix with smoothened gene expression values based on spatial grid
}
\description{
smooth gene expression over a defined spatial grid
}
\keyword{internal}
