% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction.R
\name{do_ttest}
\alias{do_ttest}
\title{do_ttest}
\usage{
do_ttest(
  expr_values,
  select_ind,
  other_ind,
  adjust_method,
  mean_method,
  offset = 0.1
)
}
\description{
Performs t.test on subsets of a matrix
}
\keyword{internal}
