% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_segmentation.R
\name{doCellposeSegmentation}
\alias{doCellposeSegmentation}
\title{perform cellpose segmentation}
\usage{
doCellposeSegmentation(
  python_env = "giotto_cellpose",
  image_dir,
  mask_output,
  channel_1 = 0,
  channel_2 = 0,
  model_name = "cyto3",
  batch_size = 8,
  resample = TRUE,
  channel_axis = NULL,
  z_axis = NULL,
  normalize = TRUE,
  invert = FALSE,
  rescale = NULL,
  diameter = NULL,
  flow_threshold = 0.4,
  cellprob_threshold = 0,
  do_3D = FALSE,
  anisotropy = NULL,
  stitch_threshold = 0,
  min_size = 15,
  niter = NULL,
  augment = FALSE,
  tile = TRUE,
  tile_overlap = 0.1,
  bsize = 224,
  interp = TRUE,
  compute_masks = TRUE,
  progress = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{image_dir}{character, required. Provide a path to a gray scale or a three channel image.}

\item{mask_output}{required. Provide a path to the output mask file.}

\item{channel_1}{channel number for cytoplasm, default to 0(gray scale)}

\item{channel_2}{channel number for Nuclei, default to 0(gray scale)}

\item{model_name}{Name of the model to run inference. Default to 'cyto3', if you want to run cutomized trained model, place your model file in ~/.cellpose/models and specify your model name.}

\item{batch_size}{Cellpose Parameter, Number of 224x224 patches to run simultaneously on the GPU. Can make smaller or bigger depending on GPU memory usage. Defaults to 8.}

\item{resample}{Cellpose Parameter}

\item{channel_axis}{Cellpose Parameter}

\item{z_axis}{Cellpose Parameter}

\item{normalize}{Cellpose Parameter}

\item{invert}{Cellpose Parameter}

\item{rescale}{Cellpose Parameter}

\item{diameter}{Cellpose Parameter}

\item{flow_threshold}{Cellpose Parameter}

\item{cellprob_threshold}{Cellpose Parameter}

\item{do_3D}{Cellpose Parameter}

\item{anisotropy}{Cellpose Parameter}

\item{stitch_threshold}{Cellpose Parameter}

\item{min_size}{Cellpose Parameter}

\item{niter}{Cellpose Parameter}

\item{augment}{Cellpose Parameter}

\item{tile}{Cellpose Parameter}

\item{tile_overlap}{Cellpose Parameter}

\item{bsize}{Cellpose Parameter}

\item{interp}{Cellpose Parameter}

\item{compute_masks}{Cellpose Parameter}

\item{progress}{Cellpose Parameter}

\item{python_path}{python environment with cellpose installed. default = "giotto_cellpose".}
}
\value{
No return variable, as this will write directly to output path provided.
}
\description{
perform the Giotto Wrapper of cellpose segmentation. This is for a model inference to generate segmentation mask file from input image.
main parameters needed
}
\examples{
# example code
doCellposeSegmentation(image_dir = input_image, mask_output = output, channel_1 = 2, channel_2 = 1, model_name = "cyto3", batch_size = 4)
}
