% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_set_enrichment.R
\name{doFeatureSetEnrichment}
\alias{doFeatureSetEnrichment}
\title{doFeatureSetEnrichment}
\usage{
doFeatureSetEnrichment(
  dryrun = TRUE,
  path_to_GSEA = NULL,
  GSEA_dataset = NULL,
  GSEA_ranked_file = NULL,
  output_folder = NULL,
  name_analysis_folder = "my_GSEA_analysis",
  collapse = "false",
  mode = c("Abs_max_of_probes", "Max_probe", "Median_of_probes", "Mean_of_probes",
    "Sum_of_probes"),
  norm = "meandiv",
  nperm = 1000,
  scoring_scheme = "weighted",
  plot_top_x = 20,
  set_max = 500,
  set_min = 15
)
}
\arguments{
\item{dryrun}{do a dry run, default TRUE.}

\item{path_to_GSEA}{path to GSEA command line executable, e.g. gsea-XXX.jar.
See details (1.) for more information.}

\item{GSEA_dataset}{path to a Human/Mouse collection from GSEA, e.g. 
Hallmarks C1. See details (2.) for more information.}

\item{GSEA_ranked_file}{path to .rnk file for GSEA. See details (3.) for 
more information}

\item{output_folder}{path to which the GSEA results will be saved. Default 
is current working directory.}

\item{name_analysis_folder}{default output subdirectory prefix to which 
results are saved.
                            Will live within output_folder; equivalent of 
                            "Analysis Name" in GSEA Application.}

\item{collapse}{only 'false' is supported. This will use your dataset as-is, 
in the original format.}

\item{mode}{option selected in Advanced Field "Collapsing Mode for 
Probe Sets => 1 gene"}

\item{norm}{normalization mode; only meandiv is supported.}

\item{nperm}{number of permutations, default 1000}

\item{scoring_scheme}{Default "weighted", equivalent of 
"enrichment statistic" in GSEA Application}

\item{plot_top_x}{Default 20, number of enrichment plots to produce.}

\item{set_max}{default 500, equivalent to "max size; exclude larger sets" 
in Basic Fields in GSEA Application}

\item{set_min}{default 15, equivalent to "min size; exclude smaller sets" 
in Basic Fields in GSEA Application}
}
\value{
data.table
}
\description{
Preform Gene Set Enrichment Analysis using marker genes
}
\details{
NECESSARY PREREQUISITES
1. download and install the COMMAND line (all platforms) gsea-XXX.jar
https://www.gsea-msigdb.org/gsea/downloads.jsp
1.1. download zip file
1.2. unzip and move to known location 
(e.g. in path/to/your/applications/gsea/GSEA_4.3.2)

2. download the Human and Mouse collections
https://www.gsea-msigdb.org/gsea/msigdb/index.jsp or zipped folder 
https://www.gsea-msigdb.org/gsea/downloads.jsp (all downloaded)

3. create ranked gene lists
format: data.table or data.frame with 2 columns
column 1 = gene symbols, column 2 = weight or rank metric
for more details, see:
https://software.broadinstitute.org/cancer/software/gsea/wiki/index.php/Data_formats#RNK:_Ranked_list_file_format_.28.2A.rnk.29

For more information on parameter convetions,
please reference GSEA's documentation here:
https://www.gsea-msigdb.org/gsea/doc/GSEAUserGuideTEXT.htm#_Syntax
}
