% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_hmrf.R
\name{doHMRF}
\alias{doHMRF}
\title{doHMRF}
\usage{
doHMRF(gobject, expression_values = c("normalized", "scaled", "custom"),
  spatial_network_name = "spatial_network", spatial_genes = NULL,
  name = "test", k = 10, betas = c(0, 2, 50), tolerance = 1e-10,
  zscore = c("rowcol", "colrow", "None"), numinit = 100,
  python_path = NULL, output_folder = NULL)
}
\arguments{
\item{gobject}{giotto object}

\item{expression_values}{expression values to use}

\item{spatial_network_name}{name of spatial network to use for HMRF}

\item{spatial_genes}{spatial genes to use for HMRF}

\item{name}{name of HMRF run}

\item{k}{number of HMRF domains}

\item{betas}{betas to test for}

\item{tolerance}{tolerance}

\item{zscore}{zscore}

\item{numinit}{number of initializations}

\item{python_path}{python path to use}

\item{output_folder}{output folder to save results}
}
\value{
Creates a directory with results that can be viewed with viewHMRFresults
}
\description{
Run HMRF
}
\details{
Description of HMRF parameters ...
}
\examples{
    doHMRF(gobject)
}
