% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_hmrf.R
\name{doHMRF}
\alias{doHMRF}
\title{doHMRF}
\usage{
doHMRF(
  gobject,
  expression_values = c("normalized", "scaled", "custom"),
  spatial_network_name = "Delaunay_network",
  spatial_genes = NULL,
  spatial_dimensions = c("sdimx", "sdimy", "sdimz"),
  dim_reduction_to_use = NULL,
  dim_reduction_name = "pca",
  dimensions_to_use = 1:10,
  seed = 100,
  name = "test",
  k = 10,
  betas = c(0, 2, 50),
  tolerance = 1e-10,
  zscore = c("none", "rowcol", "colrow"),
  numinit = 100,
  python_path = NULL,
  output_folder = NULL,
  overwrite_output = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{expression_values}{expression values to use}

\item{spatial_network_name}{name of spatial network to use for HMRF}

\item{spatial_genes}{spatial genes to use for HMRF}

\item{spatial_dimensions}{select spatial dimensions to use, default is all possible dimensions}

\item{dim_reduction_to_use}{use another dimension reduction set as input}

\item{dim_reduction_name}{name of dimension reduction set to use}

\item{dimensions_to_use}{number of dimensions to use as input}

\item{seed}{seed to fix random number generator (for creating initialization of HMRF) (-1 if no fixing)}

\item{name}{name of HMRF run}

\item{k}{number of HMRF domains}

\item{betas}{betas to test for. three numbers: start_beta, beta_increment, num_betas e.g. c(0, 2.0, 50)}

\item{tolerance}{tolerance}

\item{zscore}{zscore}

\item{numinit}{number of initializations}

\item{python_path}{python path to use}

\item{output_folder}{output folder to save results}

\item{overwrite_output}{overwrite output folder}
}
\value{
Creates a directory with results that can be viewed with viewHMRFresults
}
\description{
Run HMRF
}
\details{
Description of HMRF parameters ...
}
