% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{doKmeans}
\alias{doKmeans}
\title{doKmeans}
\usage{
doKmeans(
  gobject,
  expression_values = c("normalized", "scaled", "custom"),
  genes_to_use = NULL,
  dim_reduction_to_use = c("cells", "pca", "umap", "tsne"),
  dim_reduction_name = "pca",
  dimensions_to_use = 1:10,
  distance_method = c("original", "pearson", "spearman", "euclidean", "maximum",
    "manhattan", "canberra", "binary", "minkowski"),
  centers = 10,
  iter_max = 100,
  nstart = 1000,
  algorithm = "Hartigan-Wong",
  name = "kmeans",
  return_gobject = TRUE,
  set_seed = T,
  seed_number = 1234
)
}
\arguments{
\item{gobject}{giotto object}

\item{expression_values}{expression values to use}

\item{genes_to_use}{subset of genes to use}

\item{dim_reduction_to_use}{dimension reduction to use}

\item{dim_reduction_name}{dimensions reduction name}

\item{dimensions_to_use}{dimensions to use}

\item{distance_method}{distance method}

\item{centers}{number of final clusters}

\item{iter_max}{kmeans maximum iterations}

\item{nstart}{kmeans nstart}

\item{algorithm}{kmeans algorithm}

\item{name}{name for kmeans clustering}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}

\item{set_seed}{set seed}

\item{seed_number}{number for seed}
}
\value{
giotto object with new clusters appended to cell metadata
}
\description{
cluster cells using kmeans algorithm
}
\details{
Description on how to use Kmeans clustering method.
}
\examples{

data(mini_giotto_single_cell)

mini_giotto_single_cell = doKmeans(mini_giotto_single_cell, centers = 4, name = 'kmeans_clus')
plotUMAP_2D(mini_giotto_single_cell, cell_color = 'kmeans_clus', point_size = 3)

}
\seealso{
\code{\link[stats]{kmeans}}
}
