% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{doLeidenCluster}
\alias{doLeidenCluster}
\title{doLeidenCluster}
\usage{
doLeidenCluster(gobject, nn_network_to_use = "sNN",
  network_name = "sNN.pca", python_path = NULL, resolution = 1,
  weight_col = "weight",
  partition_type = c("RBConfigurationVertexPartition",
  "ModularityVertexPartition"), init_membership = NULL,
  n_iterations = 200, name = "pleiden_clus", return_gobject = TRUE,
  set_seed = T, seed_number = 1234, ...)
}
\arguments{
\item{gobject}{giotto object}

\item{nn_network_to_use}{type of NN network to use (kNN vs sNN)}

\item{network_name}{name of NN network to use}

\item{python_path}{specify specific path to python if required}

\item{resolution}{resolution}

\item{weight_col}{weight column}

\item{partition_type}{partition type to use}

\item{init_membership}{initial membership of cells}

\item{n_iterations}{number of interations}

\item{name}{name for cluster}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}

\item{set_seed}{set seed}

\item{seed_number}{number for seed}

\item{...}{additional parameters}
}
\value{
giotto object appended with new cluster
}
\description{
cluster cells using a NN-network and the Leiden community detection algorithm
}
\details{
Description of Leiden clustering method.
}
\examples{
    doLeidenCluster(gobject)
}
