% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{doLeidenClusterIgraph}
\alias{doLeidenClusterIgraph}
\title{doLeidenClusterIgraph}
\usage{
doLeidenClusterIgraph(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  name = "leiden_clus",
  nn_network_to_use = "sNN",
  network_name = "sNN.pca",
  objective_function = c("modularity", "CPM"),
  weights = NULL,
  resolution_parameter = 1,
  beta = 0.01,
  initial_membership = NULL,
  n_iterations = 1000,
  return_gobject = TRUE,
  set_seed = TRUE,
  seed_number = 1234,
  ...
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit (e.g. "cell")}

\item{feat_type}{feature type (e.g. "rna", "dna", "protein")}

\item{name}{name for cluster, default to "leiden_clus"}

\item{nn_network_to_use}{type of NN network to use (kNN vs sNN), default to "sNN"}

\item{network_name}{name of NN network to use, default to "sNN.pca"}

\item{objective_function}{objective function for the leiden algo}

\item{weights}{weights of edges}

\item{resolution_parameter}{resolution, default = 1}

\item{beta}{leiden randomness}

\item{initial_membership}{initial membership of cells for the partition}

\item{n_iterations}{number of interations to run the Leiden algorithm.}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}

\item{set_seed}{set seed}

\item{seed_number}{number for seed}
}
\value{
giotto object with new clusters appended to cell metadata
}
\description{
cluster cells using a NN-network and the Leiden community
detection algorithm as implemented in igraph
}
\details{
This function is a wrapper for the Leiden algorithm implemented in igraph,
which can detect communities in graphs of millions of nodes (cells),
as long as they can fit in memory. See \code{\link[igraph]{cluster_leiden}} for more information.

Set \emph{weights = NULL} to use the vertices weights associated with the igraph network.
Set \emph{weights = NA} if you don't want to use vertices weights
}
