% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{doLeidenSubCluster}
\alias{doLeidenSubCluster}
\title{doLeidenSubCluster}
\usage{
doLeidenSubCluster(
  gobject,
  feat_type = NULL,
  name = "sub_pleiden_clus",
  cluster_column = NULL,
  selected_clusters = NULL,
  hvf_param = list(reverse_log_scale = TRUE, difference_in_cov = 1, expression_values =
    "normalized"),
  hvg_param = NULL,
  hvf_min_perc_cells = 5,
  hvg_min_perc_cells = NULL,
  hvf_mean_expr_det = 1,
  hvg_mean_expr_det = NULL,
  use_all_feats_as_hvf = FALSE,
  use_all_genes_as_hvg = NULL,
  min_nr_of_hvf = 5,
  min_nr_of_hvg = NULL,
  pca_param = list(expression_values = "normalized", scale_unit = TRUE),
  nn_param = list(dimensions_to_use = 1:20),
  k_neighbors = 10,
  resolution = 0.5,
  n_iterations = 500,
  python_path = NULL,
  nn_network_to_use = "sNN",
  network_name = "sNN.pca",
  return_gobject = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{feat_type}{feature type}

\item{name}{name for new clustering result}

\item{cluster_column}{cluster column to subcluster}

\item{selected_clusters}{only do subclustering on these clusters}

\item{hvf_param}{parameters for calculateHVf}

\item{hvg_param}{deprecatd, use hvf_param}

\item{hvf_min_perc_cells}{threshold for detection in min percentage of cells}

\item{hvg_min_perc_cells}{deprecated, use hvf_min_perc_cells}

\item{hvf_mean_expr_det}{threshold for mean expression level in cells with
detection}

\item{hvg_mean_expr_det}{deprecated, use hvf_mean_expr_det}

\item{use_all_feats_as_hvf}{forces all features to be HVF and to be used as
input for PCA}

\item{use_all_genes_as_hvg}{deprecated, use use_all_feats_as_hvf}

\item{min_nr_of_hvf}{minimum number of HVF, or all features will be used as
input for PCA}

\item{min_nr_of_hvg}{deprecated, use min_nr_of_hvf}

\item{pca_param}{parameters for runPCA}

\item{nn_param}{parameters for parameters for createNearestNetwork}

\item{k_neighbors}{number of k for createNearestNetwork}

\item{resolution}{resolution of Leiden clustering}

\item{n_iterations}{number of interations to run the Leiden algorithm.}

\item{python_path}{specify specific path to python if required}

\item{nn_network_to_use}{type of NN network to use (kNN vs sNN)}

\item{network_name}{name of NN network to use}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}

\item{verbose}{verbose}
}
\value{
giotto object with new subclusters appended to cell metadata
}
\description{
Further subcluster cells using a NN-network and the Leiden
algorithm
}
\details{
This function performs subclustering using the Leiden algorithm on
selected clusters.
The systematic steps are:
\itemize{
  \item{1. subset Giotto object}
  \item{2. identify highly variable fetures}
  \item{3. run PCA}
  \item{4. create nearest neighbouring network}
  \item{5. do Leiden clustering}
}
}
\seealso{
\code{\link{doLeidenCluster}}
}
