% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{doLouvainCluster}
\alias{doLouvainCluster}
\title{doLouvainCluster}
\usage{
doLouvainCluster(
  gobject,
  version = c("community", "multinet"),
  name = "louvain_clus",
  nn_network_to_use = "sNN",
  network_name = "sNN.pca",
  python_path = NULL,
  resolution = 1,
  weight_col = NULL,
  gamma = 1,
  omega = 1,
  louv_random = F,
  return_gobject = TRUE,
  set_seed = F,
  seed_number = 1234,
  ...
)
}
\arguments{
\item{gobject}{giotto object}

\item{version}{implemented version of Louvain clustering to use}

\item{name}{name for cluster}

\item{nn_network_to_use}{type of NN network to use (kNN vs sNN)}

\item{network_name}{name of NN network to use}

\item{python_path}{[community] specify specific path to python if required}

\item{resolution}{[community] resolution}

\item{weight_col}{weight column name}

\item{gamma}{[multinet] Resolution parameter for modularity in the generalized louvain method.}

\item{omega}{[multinet] Inter-layer weight parameter in the generalized louvain method}

\item{louv_random}{[community] Will randomize the node evaluation order and the community evaluation
order to get different partitions at each call}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}

\item{set_seed}{set seed}

\item{seed_number}{number for seed}

\item{\dots}{additional parameters}
}
\value{
giotto object with new clusters appended to cell metadata
}
\description{
cluster cells using a NN-network and the Louvain algorithm.
}
\details{
Louvain clustering using the community or multinet implementation of the louvain clustering algorithm.
}
\seealso{
\code{\link{doLouvainCluster_community}} and \code{\link{doLouvainCluster_multinet}}
}
