% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{doLouvainCluster_community}
\alias{doLouvainCluster_community}
\title{doLouvainCluster_community}
\usage{
doLouvainCluster_community(
  gobject,
  name = "louvain_clus",
  nn_network_to_use = "sNN",
  network_name = "sNN.pca",
  python_path = NULL,
  resolution = 1,
  weight_col = NULL,
  louv_random = F,
  return_gobject = TRUE,
  set_seed = F,
  seed_number = 1234
)
}
\arguments{
\item{gobject}{giotto object}

\item{name}{name for cluster}

\item{nn_network_to_use}{type of NN network to use (kNN vs sNN)}

\item{network_name}{name of NN network to use}

\item{python_path}{specify specific path to python if required}

\item{resolution}{resolution}

\item{weight_col}{weight column to use for edges}

\item{louv_random}{Will randomize the node evaluation order and the community evaluation
order to get different partitions at each call}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}

\item{set_seed}{set seed}

\item{seed_number}{number for seed}
}
\value{
giotto object with new clusters appended to cell metadata
}
\description{
cluster cells using a NN-network and the Louvain algorithm from the community module in Python
}
\details{
This function is a wrapper for the Louvain algorithm implemented in Python,
which can detect communities in graphs of nodes (cells).
See the \url{https://python-louvain.readthedocs.io/en/latest/index.html}{readthedocs}
page for more information.

Set \emph{weight_col = NULL} to give equal weight (=1) to each edge.
}
\keyword{internal}
