% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{doRandomWalkCluster}
\alias{doRandomWalkCluster}
\title{doRandomWalkCluster}
\usage{
doRandomWalkCluster(
  gobject,
  name = "random_walk_clus",
  nn_network_to_use = "sNN",
  network_name = "sNN.pca",
  walk_steps = 4,
  walk_clusters = 10,
  walk_weights = NA,
  return_gobject = TRUE,
  set_seed = FALSE,
  seed_number = 1234
)
}
\arguments{
\item{gobject}{giotto object}

\item{name}{name for cluster, default to "random_walk_clus"}

\item{nn_network_to_use}{type of NN network to use (kNN vs sNN), default to
"sNN"}

\item{network_name}{name of NN network to use, default to "sNN.pca"}

\item{walk_steps}{number of walking steps, default = 4}

\item{walk_clusters}{number of final clusters, default =  10}

\item{walk_weights}{cluster column defining the walk weights}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}

\item{set_seed}{set seed (default = FALSE)}

\item{seed_number}{number for seed}
}
\value{
giotto object with new clusters appended to cell metadata
}
\description{
Cluster cells using a random walk approach.
}
\details{
See \code{\link[igraph]{cluster_walktrap}} function from the igraph
package in R for more information.
}
\examples{
g <- GiottoData::loadGiottoMini("visium")

doRandomWalkCluster(g)
}
