% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_scrublet.R
\name{doScrubletDetect}
\alias{doScrubletDetect}
\title{doScrubletDetect}
\usage{
doScrubletDetect(
  gobject,
  feat_type = NULL,
  spat_unit = "cell",
  expression_values = "raw",
  expected_doublet_rate = 0.06,
  min_counts = 1,
  min_cells = 1,
  min_gene_variability_pctl = 85,
  n_prin_comps = 30,
  return_gobject = TRUE
)
}
\arguments{
\item{gobject}{giotto object containing expression data}

\item{feat_type}{feature type}

\item{spat_unit}{spatial unit}

\item{expression_values}{expression values to use}

\item{expected_doublet_rate}{expected transcriptomes that are doublets. 0.06 is from 10x Chromium guide.}

\item{min_counts}{scrublet internal data filtering, min counts found to be considered a cell}

\item{min_cells}{scrublet internal data filtering. min cells expressed to be considered a feat}

\item{min_gene_variability_pctl}{scrublet internal PCA generation. highly variable gene percentile cutoff}

\item{n_prin_comps}{number of PCs to use in PCA for detection}

\item{return_gobject}{return as gobject if TRUE, data.frame with cell_ID if fALSE}
}
\value{
list including doublet scores and classifications
}
\description{
run scrublet doublet detection for raw expression.
}
