% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{doSNNCluster}
\alias{doSNNCluster}
\title{doSNNCluster}
\usage{
doSNNCluster(
  gobject,
  name = "sNN_clus",
  nn_network_to_use = "kNN",
  network_name = "kNN.pca",
  k = 20,
  eps = 4,
  minPts = 16,
  borderPoints = TRUE,
  return_gobject = TRUE,
  set_seed = FALSE,
  seed_number = 1234
)
}
\arguments{
\item{gobject}{giotto object}

\item{name}{name for cluster, default to "sNN_clus"}

\item{nn_network_to_use}{type of NN network to use (only works on kNN), 
default to "kNN"}

\item{network_name}{name of kNN network to use, default to "kNN.pca"}

\item{k}{Neighborhood size for nearest neighbor sparsification to create 
the shared NN graph, default = 20}

\item{eps}{Two objects are only reachable from each other if they share at 
least eps nearest neighbors, default = 4}

\item{minPts}{minimum number of points that share at least eps nearest 
neighbors for a point to be considered a core points, default = 16}

\item{borderPoints}{should borderPoints be assigned to clusters like in 
DBSCAN? (default = TRUE)}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}

\item{set_seed}{set seed (default = FALSE)}

\item{seed_number}{number for seed}
}
\value{
giotto object with new clusters appended to cell metadata
}
\description{
Cluster cells using a SNN cluster approach.
}
\details{
See \code{\link[dbscan]{sNNclust}} from dbscan package
}
