% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{.cosmx_infer_fov_shifts}
\alias{.cosmx_infer_fov_shifts}
\title{Infer CosMx local to global shifts}
\usage{
.cosmx_infer_fov_shifts(tx_dt, meta_dt, flip_loc_y = NULL, navg = 100L)
}
\arguments{
\item{tx_dt}{transcript data.table input to use
(Only one of tx_dt or meta_dt should be used)}

\item{meta_dt}{cell metadata data.table input to use
(Only one of tx_dt or meta_dt should be used)}

\item{flip_loc_y}{whether a y flip needs to be performed on the local y
values before camparing with global y values}

\item{navg}{max n values to check per FOV to find average shift}
}
\value{
data.table with three columns. 1. FOV (integer), xshift (numeric),
yshift (numeric). Values should always be in pixels
}
\description{
From NanoString CosMx spatial info, infer the FOV shifts needed. These
values are needed for anything that requires the use of images, since those
do not come with spatial extent information embedded.
}
\keyword{internal}
