% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich-param.R
\name{enrich_hyper}
\alias{enrich_hyper}
\title{Hypergeometric enrichment}
\usage{
\S4method{analyzeData}{ANY,hyperEnrichParam}(x, param, ..., sign_matrix)
}
\arguments{
\item{x}{expression values. A `matrix`, a `Matrix`, or anything an
engine registered against [pageEnrichParam-class] accepts.}

\item{param}{a [hyperEnrichParam-class].}

\item{sign_matrix}{binary sign matrix, genes x cell types.}
}
\value{
a `data.table` of `cell_ID` and one column per cell type
}
\description{
Hypergeometric enrichment
}
