% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto.R
\name{evaluate_expr_matrix}
\alias{evaluate_expr_matrix}
\title{Evaluate expression matrix}
\usage{
evaluate_expr_matrix(inputmatrix, sparse = TRUE, cores = NA)
}
\arguments{
\item{inputmatrix}{inputmatrix to evaluate}

\item{sparse}{create sparse matrix (default = TRUE)}

\item{cores}{how many cores to use}
}
\value{
sparse matrix
}
\description{
Evaluate expression matrices that are provided as input and converts
them to preferred format for Giotto object. A filepath can also be provided through
\code{inputmatrix} param. If this is done, the function will attempt to read the
matrix file in using \code{\link{readExprMatrix}}.
}
\details{
The inputmatrix can be a matrix, sparse matrix, data.frame, data.table or path to any of these.
}
\keyword{internal}
