% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto.R
\name{evaluate_spatial_info}
\alias{evaluate_spatial_info}
\title{Evaluate spatial info}
\usage{
evaluate_spatial_info(spatial_info, cores = determine_cores(), spatial_locs)
}
\arguments{
\item{spatial_info}{spatial information to evaluate}

\item{cores}{how many cores to use}

\item{spatial_locs}{spatial location data.table to compare the cell IDs with}
}
\value{
data.table
}
\description{
Evaluate spatial information input
}
\keyword{internal}
