% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto.R
\name{evaluate_spatial_locations}
\alias{evaluate_spatial_locations}
\title{evaluate_spatial_locations}
\usage{
evaluate_spatial_locations(
  spatial_locs,
  cores = 1,
  dummy_n = 2,
  expr_matrix = NULL
)
}
\arguments{
\item{spatial_locs}{spatial locations to evaluate}

\item{cores}{how many cores to use}

\item{dummy_n}{number of rows to create dummy spaial locations}

\item{expr_matrix}{expression matrix to compare the cell IDs with}
}
\value{
data.table
}
\description{
Evaluate spatial location input
}
\keyword{internal}
