% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_viewer.R
\name{exportGiottoViewer}
\alias{exportGiottoViewer}
\title{exportGiottoViewer}
\usage{
exportGiottoViewer(
  gobject,
  output_directory = NULL,
  annotations,
  dim_reductions,
  dim_reduction_names,
  expression_values = c("normalized", "scaled", "custom"),
  dim_red_rounding = NULL,
  dim_red_rescale = c(-20, 20),
  expression_rounding = NULL,
  overwrite_dir = F,
  verbose = T
)
}
\arguments{
\item{gobject}{giotto object}

\item{output_directory}{directory where to save the files}

\item{annotations}{giotto cell annotations to view}

\item{dim_reductions}{high level dimension reductions to view}

\item{dim_reduction_names}{specific dimension reduction names}

\item{expression_values}{expression values to use in Viewer}

\item{dim_red_rounding}{numerical indicating how to round the coordinates}

\item{dim_red_rescale}{numericals to rescale the coordinates}

\item{expression_rounding}{numerical indicating how to round the expression data}

\item{overwrite_dir}{overwrite files in the directory if it already existed}

\item{verbose}{be verbose}
}
\value{
writes the necessary output to use in Giotto Viewer
}
\description{
compute highly variable genes
}
\details{
Giotto Viewer expects the results from Giotto Analyzer in a specific format, which is provided by this function.
}
\examples{
    exportGiottoViewer(gobject)
}
