% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_viewer.R
\name{exportGiottoViewer}
\alias{exportGiottoViewer}
\title{exportGiottoViewer}
\usage{
exportGiottoViewer(
  gobject,
  output_directory = NULL,
  spat_enr_names = NULL,
  factor_annotations = NULL,
  numeric_annotations = NULL,
  dim_reductions,
  dim_reduction_names,
  expression_values = c("scaled", "normalized", "custom"),
  dim_red_rounding = NULL,
  dim_red_rescale = c(-20, 20),
  expression_rounding = 2,
  overwrite_dir = T,
  verbose = T
)
}
\arguments{
\item{gobject}{giotto object}

\item{output_directory}{directory where to save the files}

\item{spat_enr_names}{spatial enrichment results to include for annotations}

\item{factor_annotations}{giotto cell annotations to view as factor}

\item{numeric_annotations}{giotto cell annotations to view as numeric}

\item{dim_reductions}{high level dimension reductions to view}

\item{dim_reduction_names}{specific dimension reduction names}

\item{expression_values}{expression values to use in Viewer}

\item{dim_red_rounding}{numerical indicating how to round the coordinates}

\item{dim_red_rescale}{numericals to rescale the coordinates}

\item{expression_rounding}{numerical indicating how to round the expression data}

\item{overwrite_dir}{overwrite files in the directory if it already existed}

\item{verbose}{be verbose}
}
\value{
writes the necessary output to use in Giotto Viewer
}
\description{
compute highly variable genes
}
\details{
Giotto Viewer expects the results from Giotto Analyzer in a specific format,
which is provided by this function. To include enrichment results from {\code{\link{createSpatialEnrich}}}
include the provided spatial enrichment name (default PAGE or rank)
and add the gene signature names (.e.g cell types) to the numeric annotations parameter.
}
\examples{

\dontrun{

data(mini_giotto_single_cell)
exportGiottoViewer(mini_giotto_single_cell)

}

}
