% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction.R
\name{exprCellCellcom}
\alias{exprCellCellcom}
\title{exprCellCellcom}
\usage{
exprCellCellcom(
  gobject,
  cluster_column = "cell_types",
  random_iter = 1000,
  gene_set_1,
  gene_set_2,
  log2FC_addendum = 0.1,
  detailed = FALSE,
  adjust_method = c("fdr", "bonferroni", "BH", "holm", "hochberg", "hommel", "BY",
    "none"),
  adjust_target = c("genes", "cells"),
  set_seed = TRUE,
  seed_number = 1234,
  verbose = T
)
}
\arguments{
\item{gobject}{giotto object to use}

\item{cluster_column}{cluster column with cell type information}

\item{random_iter}{number of iterations}

\item{gene_set_1}{first specific gene set from gene pairs}

\item{gene_set_2}{second specific gene set from gene pairs}

\item{log2FC_addendum}{addendum to add when calculating log2FC}

\item{detailed}{provide more detailed information (random variance and z-score)}

\item{adjust_method}{which method to adjust p-values}

\item{adjust_target}{adjust multiple hypotheses at the cell or gene level}

\item{set_seed}{set seed for random simulations (default = TRUE)}

\item{seed_number}{seed number}

\item{verbose}{verbose}
}
\value{
Cell-Cell communication scores for gene pairs based on expression only
}
\description{
Cell-Cell communication scores based on expression only
}
\details{
Statistical framework to identify if pairs of genes (such as ligand-receptor combinations)
are expressed at higher levels than expected based on a reshuffled null distribution of gene expression values,
without considering the spatial position of cells.
More details will follow soon.
}
