% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{extract-generic}
\alias{extract-generic}
\alias{$,coordDataDT-method}
\alias{$<-,coordDataDT-method}
\alias{$,metaData-method}
\alias{$<-,metaData-method}
\alias{[,coordDataDT,missing,ANY,missing-method}
\alias{[,coordDataDT,ANY,missing,missing-method}
\alias{[,coordDataDT,ANY,ANY,missing-method}
\alias{[,coordDataDT,missing,missing,missing-method}
\alias{[<-,coordDataDT,missing,missing,ANY-method}
\alias{[,metaData,missing,ANY,missing-method}
\alias{[,metaData,ANY,missing,missing-method}
\alias{[,metaData,ANY,ANY,missing-method}
\alias{[,metaData,missing,missing,missing-method}
\alias{[<-,metaData,missing,missing,ANY-method}
\alias{[,dimObj,ANY,ANY,missing-method}
\alias{[,dimObj,missing,missing,missing-method}
\alias{[<-,dimObj,missing,missing,ANY-method}
\alias{[,exprData,missing,ANY,missing-method}
\alias{[,exprData,ANY,missing,missing-method}
\alias{[,exprData,ANY,ANY,missing-method}
\alias{[,exprData,missing,missing,missing-method}
\alias{[<-,exprData,missing,missing,ANY-method}
\alias{[,spatNetData,missing,missing,missing-method}
\alias{[<-,spatNetData,missing,missing,ANY-method}
\alias{[,nnData,missing,missing,missing-method}
\alias{[<-,nnData,missing,missing,ANY-method}
\alias{[,enrData,missing,missing,missing-method}
\alias{[<-,enrData,missing,missing,ANY-method}
\alias{[,spatGridData,missing,missing,missing-method}
\alias{[<-,spatGridData,missing,missing,ANY-method}
\title{Extract or replace parts of an object}
\usage{
\S4method{$}{coordDataDT}(x, name)

\S4method{$}{coordDataDT}(x, name) <- value

\S4method{$}{metaData}(x, name)

\S4method{$}{metaData}(x, name) <- value

\S4method{[}{coordDataDT,missing,ANY,missing}(x, i, j)

\S4method{[}{coordDataDT,ANY,missing,missing}(x, i, j)

\S4method{[}{coordDataDT,ANY,ANY,missing}(x, i, j)

\S4method{[}{coordDataDT,missing,missing,missing}(x, i, j)

\S4method{[}{coordDataDT,missing,missing,ANY}(x, i, j) <- value

\S4method{[}{metaData,missing,ANY,missing}(x, i, j)

\S4method{[}{metaData,ANY,missing,missing}(x, i, j)

\S4method{[}{metaData,ANY,ANY,missing}(x, i, j)

\S4method{[}{metaData,missing,missing,missing}(x, i, j)

\S4method{[}{metaData,missing,missing,ANY}(x, i, j) <- value

\S4method{[}{dimObj,ANY,ANY,missing}(x, i, j)

\S4method{[}{dimObj,missing,missing,missing}(x, i, j)

\S4method{[}{dimObj,missing,missing,ANY}(x, i, j) <- value

\S4method{[}{exprData,missing,ANY,missing}(x, i, j)

\S4method{[}{exprData,ANY,missing,missing}(x, i, j)

\S4method{[}{exprData,ANY,ANY,missing}(x, i, j)

\S4method{[}{exprData,missing,missing,missing}(x, i, j)

\S4method{[}{exprData,missing,missing,ANY}(x, i, j) <- value

\S4method{[}{spatNetData,missing,missing,missing}(x, i, j)

\S4method{[}{spatNetData,missing,missing,ANY}(x, i, j) <- value

\S4method{[}{nnData,missing,missing,missing}(x, i, j)

\S4method{[}{nnData,missing,missing,ANY}(x, i, j) <- value

\S4method{[}{enrData,missing,missing,missing}(x, i, j)

\S4method{[}{enrData,missing,missing,ANY}(x, i, j) <- value

\S4method{[}{spatGridData,missing,missing,missing}(x, i, j)

\S4method{[}{spatGridData,missing,missing,ANY}(x, i, j) <- value
}
\arguments{
\item{x}{Giotto S4 object to extract columns from}

\item{name}{A literal character string (possibly backtick quoted).}

\item{value}{value(s) to set
This is normally matched to the colnames of the data.table object within the S4.}

\item{i, j}{indices specifying elements to extract or replace. Indices are
numeric or character vectors or empty}
}
\description{
Operators Giotto S4 internal data.tables to extract
or replace parts.
}
\section{\code{`$`} methods}{



  Select by colname from giotto S4 data.table coordinates slot.


  Select by colname from giotto S4 data.table metaDT slot.
}

\section{\code{`$<-`} methods}{



  Set values by colname into giotto S4 data.table coordinates slot.
  Works via data.table methods


  Set values by colname into giotto S4 data.table metaDT slot.
  Works via data.table methods
}

\section{\code{`[`} methods}{



  Select rows (i) and cols (j) from giotto S4 coordinates slot


  Return \code{coordinates} slot data.table from giotto S4


  Select rows (i) and cols (j) from giotto S4 metaDT slot


  Return \code{coordinates} slot data.table from giotto S4


  Select rows (i) and cols (j) from giotto S4 exprMat slot


  Return \code{exprMat} slot Matrix object from giotto S4


  Return \code{spatNetData} slot network data.table object from giotto S4


  Return \code{nnData} slot igraph object from giotto S4


  Return \code{enrData} slot enrichment data.table object from giotto S4


  Return \code{spatGridData} slot enrichment data.table object from giotto S4
}

\section{\code{`[<-`} methods}{

  TODO


  Assign to \code{coordinates} slot in giotto S4


  Assign to \code{metaDT} slot in giotto S4


  Assign to \code{coordinates} slot in giotto S4


  Assign to \code{exprMat} slot in giotto S4


  Return \code{spatNetData} slot network data.table object from giotto S4


  Return \code{nnData} slot igraph object from giotto S4


  Return \code{enrData} slot enrichment data.table object from giotto S4


  Return \code{spatGridData} slot enrichment data.table object from giotto S4
}

