% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{extractNearestNetwork}
\alias{extractNearestNetwork}
\title{Extract nearest network}
\usage{
extractNearestNetwork(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=get_NearestNetwork]{get_NearestNetwork}}
  \describe{
    \item{\code{nn_network_to_use}}{"kNN" or "sNN"}
    \item{\code{network_name}}{name of NN network to be used}
    \item{\code{output}}{return a igraph or data.table object. Default 'igraph'}
    \item{\code{gobject}}{giotto object}
    \item{\code{spat_unit}}{spatial unit (e.g. "cell")}
    \item{\code{feat_type}}{feature type (e.g. "rna", "dna", "protein")}
    \item{\code{set_defaults}}{set default spat_unit and feat_type. Change to FALSE only when}
  }}
}
\description{
Extract nearest network
}
\seealso{
\code{\link{get_NearestNetwork}}
}
\keyword{internal}
