% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NN_network.R
\name{extractNearestNetwork}
\alias{extractNearestNetwork}
\title{extractNearestNetwork}
\usage{
extractNearestNetwork(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=select_NearestNetwork]{select_NearestNetwork}}
  \describe{
    \item{\code{gobject}}{giotto object}
    \item{\code{nn_network_to_use}}{kNN or sNN}
    \item{\code{network_name}}{name of NN network to be used}
    \item{\code{output}}{return a igraph or data.table object}
  }}
}
\description{
Function to calculate gene signature enrichment scores per spatial position using a rank based approach.
}
\seealso{
\code{\link{select_NearestNetwork}}
}
\keyword{internal}
