% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{fDataDT}
\alias{fDataDT}
\title{fDataDT}
\usage{
fDataDT(gobject, spat_unit = NULL, feat_type = NULL, ...)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit}

\item{feat_type}{feature type}

\item{...}{additional params to pass}
}
\value{
data.table with feature metadata
}
\description{
show feature metadata
}
