% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{fDataDT}
\alias{fDataDT}
\title{fDataDT}
\usage{
fDataDT(gobject, feat_type = NULL, spat_unit = NULL)
}
\arguments{
\item{gobject}{giotto object}

\item{feat_type}{feature type}
}
\value{
data.table with gene metadata
}
\description{
show gene metadata
}
\examples{

data(mini_giotto_single_cell) # loads existing Giotto object
fDataDT(mini_giotto_single_cell)

}
