% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{filterCombinations}
\alias{filterCombinations}
\title{filterCombinations}
\usage{
filterCombinations(
  gobject,
  feat_type = NULL,
  spat_unit = NULL,
  expression_values = c("raw", "normalized", "scaled", "custom"),
  expression_thresholds = c(1, 2),
  feat_det_in_min_cells = c(5, 50),
  min_det_feats_per_cell = c(200, 400),
  scale_x_axis = "identity",
  x_axis_offset = 0,
  scale_y_axis = "identity",
  y_axis_offset = 0,
  show_plot = TRUE,
  return_plot = FALSE,
  save_plot = NA,
  save_param = list(),
  default_save_name = "filterCombinations"
)
}
\arguments{
\item{gobject}{giotto object}

\item{feat_type}{feature type (e.g. "rna", "dna", "protein")}

\item{spat_unit}{spatial unit (e.g. "cell")}

\item{expression_values}{expression values to use}

\item{expression_thresholds}{all thresholds to consider a gene expressed}

\item{feat_det_in_min_cells}{minimum # of cells that need to express a feature}

\item{min_det_feats_per_cell}{minimum # of features that need to be detected in a cell}

\item{scale_x_axis}{ggplot transformation for x-axis (e.g. log2)}

\item{x_axis_offset}{x-axis offset to be used together with the scaling transformation}

\item{scale_y_axis}{ggplot transformation for y-axis (e.g. log2)}

\item{y_axis_offset}{y-axis offset to be used together with the scaling transformation}

\item{show_plot}{logical. show plot}

\item{return_plot}{logical. return ggplot object}

\item{save_plot}{logical. save the plot}

\item{save_param}{list of saving parameters, see \code{\link{showSaveParameters}}}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}
}
\value{
list of data.table and ggplot object
}
\description{
Shows how many genes and cells are lost with combinations of thresholds.
}
\details{
Creates a scatterplot that visualizes the number of genes and cells that are
lost with a specific combination of a gene and cell threshold given an arbitrary cutoff
to call a gene expressed. This function can be used to make an informed decision at the
filtering step with filterGiotto.
}
