% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction.R
\name{filterCPG}
\alias{filterCPG}
\title{filterCPG}
\usage{
filterCPG(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=filterICG]{filterICG}}
  \describe{
    \item{\code{cpgObject}}{ICG (interaction changed gene) score object}
    \item{\code{min_cells}}{minimum number of source cell type}
    \item{\code{min_cells_expr}}{minimum expression level for source cell type}
    \item{\code{min_int_cells}}{minimum number of interacting neighbor cell type}
    \item{\code{min_int_cells_expr}}{minimum expression level for interacting neighbor cell type}
    \item{\code{min_fdr}}{minimum adjusted p-value}
    \item{\code{min_spat_diff}}{minimum absolute spatial expression difference}
    \item{\code{min_log2_fc}}{minimum log2 fold-change}
    \item{\code{min_zscore}}{minimum z-score change}
    \item{\code{zscores_column}}{calculate z-scores over cell types or genes}
    \item{\code{direction}}{differential expression directions to keep}
  }}
}
\description{
Filter Interaction Changed Gene scores.
}
\seealso{
\code{\link{filterICG}}
}
