% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{filterDistributions}
\alias{filterDistributions}
\title{filterDistributions}
\usage{
filterDistributions(gobject, expression_values = c("raw", "normalized",
  "scaled", "custom"), expression_threshold = 1, detection = c("genes",
  "cells"), plot_type = c("histogram", "violin"), nr_bins = 30,
  fill_color = "lightblue", scale_axis = "identity", axis_offset = 0,
  show_plot = TRUE)
}
\arguments{
\item{gobject}{giotto object}

\item{expression_values}{expression values to use}

\item{expression_threshold}{threshold to consider a gene expressed}

\item{detection}{look at genes or cells}

\item{plot_type}{type of plot}

\item{nr_bins}{number of bins for histogram plot}

\item{fill_color}{fill color for plots}

\item{scale_axis}{ggplot transformation for axis (e.g. log2)}

\item{axis_offset}{offset to be used together with the scaling transformation}

\item{show_plot}{show plot}
}
\value{
ggplot object
}
\description{
show gene or cell filter distributions
}
\examples{
    filterDistributions(gobject)
}
