% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filterDistributions}
\alias{filterDistributions}
\title{filterDistributions}
\usage{
filterDistributions(
  gobject,
  feat_type = NULL,
  spat_unit = NULL,
  expression_values = c("raw", "normalized", "scaled", "custom"),
  method = c("threshold", "sum", "mean"),
  expression_threshold = 1,
  detection = c("feats", "cells"),
  plot_type = c("histogram", "violin"),
  scale_y = NULL,
  nr_bins = 30,
  fill_color = "lightblue",
  scale_axis = "identity",
  axis_offset = 0,
  show_plot = NULL,
  return_plot = NULL,
  save_plot = NULL,
  save_param = list(),
  default_save_name = "filterDistributions"
)
}
\arguments{
\item{gobject}{giotto object}

\item{feat_type}{feature type}

\item{spat_unit}{spatial unit}

\item{expression_values}{expression values to use}

\item{method}{method to create distribution (see details)}

\item{expression_threshold}{threshold to consider a gene expressed}

\item{detection}{consider features (e.g. genes) or cells}

\item{plot_type}{type of plot}

\item{scale_y}{scale y-axis (e.g. "log"), NULL = no scaling}

\item{nr_bins}{number of bins for histogram plot}

\item{fill_color}{fill color for plots}

\item{scale_axis}{ggplot transformation for axis (e.g. log2)}

\item{axis_offset}{offset to be used together with the scaling
transformation}

\item{show_plot}{logical. show plot}

\item{return_plot}{logical. return ggplot object}

\item{save_plot}{logical. directly save the plot}

\item{save_param}{list of saving parameters from
\link[GiottoVisuals:plot_save]{GiottoVisuals::all_plots_save_function}}

\item{default_save_name}{default save name for saving, don't change,
change save_name in save_param}
}
\value{
ggplot object
}
\description{
show gene or cell distribution after filtering on expression
threshold
}
\details{
There are 3 ways to create a distribution profile and summarize it for
either the features or the cells (spatial units) \cr
\itemize{
\item{1. threshold: calculate features that cross a thresold (default)}
\item{2. sum: summarize the features, i.e. total of a feature}
\item{3. mean: calculate mean of the features, i.e. average expression}
}
}
\examples{
g <- GiottoData::loadGiottoMini("visium")

filterDistributions(g)
}
