% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{filterGiotto}
\alias{filterGiotto}
\title{filterGiotto}
\usage{
filterGiotto(gobject, expression_values = c("raw", "normalized",
  "scaled", "custom"), expression_threshold = 1,
  minimum_detected_genes = 100, minimum_expression_in_cell = 100,
  verbose = F)
}
\arguments{
\item{gobject}{giotto object}

\item{expression_values}{expression values to use}

\item{expression_threshold}{threshold to consider a gene expressed}

\item{minimum_detected_genes}{minimum # of genes that need to be detected in a cell}

\item{minimum_expression_in_cell}{minimum # of cells that need to express a gene}

\item{verbose}{verbose}
}
\value{
giotto object
}
\description{
filter Giotto object
}
\examples{
    filterGiotto(gobject)
}
