% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{findGiniMarkers}
\alias{findGiniMarkers}
\title{findGiniMarkers}
\usage{
findGiniMarkers(
  gobject,
  expression_values = c("normalized", "scaled", "custom"),
  cluster_column,
  subset_clusters = NULL,
  group_1 = NULL,
  group_2 = NULL,
  min_expr_gini_score = 0.5,
  min_det_gini_score = 0.5,
  detection_threshold = 0,
  rank_score = 1
)
}
\arguments{
\item{gobject}{giotto object}

\item{expression_values}{gene expression values to use}

\item{cluster_column}{clusters to use}

\item{subset_clusters}{selection of clusters to compare}

\item{group_1}{group 1 cluster IDs from cluster_column for pairwise comparison}

\item{group_2}{group 2 cluster IDs from cluster_column for pairwise comparison}

\item{min_expr_gini_score}{filter on minimum gini coefficient for expression}

\item{min_det_gini_score}{filter minimum gini coefficient for detection}

\item{detection_threshold}{detection threshold for gene expression}

\item{rank_score}{rank scores to include}
}
\value{
data.table with marker genes
}
\description{
Identify marker genes for selected clusters based on gini detection and expression scores.
}
\details{
Description of parameters.
}
\examples{
    findGiniMarkers(gobject)
}
