% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{findGiniMarkers}
\alias{findGiniMarkers}
\title{findGiniMarkers}
\usage{
findGiniMarkers(
  gobject,
  expression_values = c("normalized", "scaled", "custom"),
  cluster_column,
  subset_clusters = NULL,
  group_1 = NULL,
  group_2 = NULL,
  min_expr_gini_score = 0.2,
  min_det_gini_score = 0.2,
  detection_threshold = 0,
  rank_score = 1,
  min_genes = 5
)
}
\arguments{
\item{gobject}{giotto object}

\item{expression_values}{gene expression values to use}

\item{cluster_column}{clusters to use}

\item{subset_clusters}{selection of clusters to compare}

\item{group_1}{group 1 cluster IDs from cluster_column for pairwise comparison}

\item{group_2}{group 2 cluster IDs from cluster_column for pairwise comparison}

\item{min_expr_gini_score}{filter on minimum gini coefficient for expression}

\item{min_det_gini_score}{filter on minimum gini coefficient for detection}

\item{detection_threshold}{detection threshold for gene expression}

\item{rank_score}{rank scores for both detection and expression to include}

\item{min_genes}{minimum number of top genes to return}
}
\value{
data.table with marker genes
}
\description{
Identify marker genes for selected clusters based on gini detection and expression scores.
}
\details{
Detection of marker genes using the \url{https://en.wikipedia.org/wiki/Gini_coefficient}{gini}
coefficient is based on the following steps/principles per gene:
\itemize{
  \item{1. calculate average expression per cluster}
  \item{2. calculate detection fraction per cluster}
  \item{3. calculate gini-coefficient for av. expression values over all clusters}
  \item{4. calculate gini-coefficient for detection fractions over all clusters}
  \item{5. convert gini-scores to rank scores}
  \item{6. for each gene create combined score = detection rank x expression rank x expr gini-coefficient x detection gini-coefficient}
  \item{7. for each gene sort on expression and detection rank and combined score}
}

As a results "top gini" genes are genes that are very selectivily expressed in a specific cluster,
however not always expressed in all cells of that cluster. In other words highly specific, but
not necessarily sensitive at the single-cell level.

To perform differential expression between cluster groups you need to specificy cluster IDs
to the parameters \emph{group_1} and \emph{group_2}.
}
\examples{

data(mini_giotto_single_cell)

gini_markers = findGiniMarkers(gobject = mini_giotto_single_cell,
                               cluster_column = 'leiden_clus',
                               group_1 = 1,
                               group_2 = 2)

}
