% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{findGiniMarkers_one_vs_all}
\alias{findGiniMarkers_one_vs_all}
\title{findGiniMarkers_one_vs_all}
\usage{
findGiniMarkers_one_vs_all(
  gobject,
  expression_values = c("normalized", "scaled", "custom"),
  cluster_column,
  subset_clusters = NULL,
  min_expr_gini_score = 0.5,
  min_det_gini_score = 0.5,
  detection_threshold = 0,
  rank_score = 1,
  min_genes = 4,
  verbose = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{expression_values}{gene expression values to use}

\item{cluster_column}{clusters to use}

\item{subset_clusters}{selection of clusters to compare}

\item{min_expr_gini_score}{filter on minimum gini coefficient on expression}

\item{min_det_gini_score}{filter on minimum gini coefficient on detection}

\item{detection_threshold}{detection threshold for gene expression}

\item{rank_score}{rank scores for both detection and expression to include}

\item{min_genes}{minimum number of top genes to return}

\item{verbose}{be verbose}
}
\value{
data.table with marker genes
}
\description{
Identify marker genes for all clusters in a one vs all manner based on gini detection and expression scores.
}
\examples{

data(mini_giotto_single_cell)

gini_markers = findGiniMarkers_one_vs_all(gobject = mini_giotto_single_cell,
                                          cluster_column = 'leiden_clus')

}
\seealso{
\code{\link{findGiniMarkers}}
}
