% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{findMarkers}
\alias{findMarkers}
\title{findMarkers}
\usage{
findMarkers(gobject, expression_values = c("normalized", "scaled",
  "custom"), cluster_column, subset_clusters = NULL, group_1 = NULL,
  group_2 = NULL, method = c("scran"), ...)
}
\arguments{
\item{gobject}{giotto object}

\item{expression_values}{gene expression values to use}

\item{cluster_column}{clusters to use}

\item{subset_clusters}{selection of clusters to compare}

\item{group_1}{group 1 cluster IDs from cluster_column for pairwise comparison}

\item{group_2}{group 2 cluster IDs from cluster_column for pairwise comparison}

\item{method}{method to use}

\item{...}{additional parameters}
}
\value{
differentially expressed genes (DEG) output
}
\description{
Identify marker genes for selected clusters
}
\details{
Description of different DEG methods.
}
\examples{
    findMarkers(gobject)
}
