% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{findMastMarkers}
\alias{findMastMarkers}
\title{findMastMarkers}
\usage{
findMastMarkers(
  gobject,
  expression_values = c("normalized", "scaled", "custom"),
  cluster_column,
  group_1 = NULL,
  group_1_name = NULL,
  group_2 = NULL,
  group_2_name = NULL,
  adjust_columns = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{gobject}{giotto object}

\item{expression_values}{gene expression values to use}

\item{cluster_column}{clusters to use}

\item{group_1}{group 1 cluster IDs from cluster_column for pairwise comparison}

\item{group_1_name}{custom name for group_1 clusters}

\item{group_2}{group 2 cluster IDs from cluster_column for pairwise comparison}

\item{group_2_name}{custom name for group_2 clusters}

\item{adjust_columns}{column in pDataDT to adjust for (e.g. detection rate)}

\item{verbose}{be verbose}

\item{...}{additional parameters for the zlm function in MAST}
}
\value{
data.table with marker genes
}
\description{
Identify marker genes for selected clusters based on the MAST package.
}
\details{
This is a minimal convenience wrapper around the \code{\link[MAST]{zlm}}
from the MAST package to detect differentially expressed genes. Caution: with large datasets
MAST might take a long time to run and finish
}
\examples{

\dontrun{
data(mini_giotto_single_cell)

mast_markers = findMastMarkers(gobject = mini_giotto_single_cell,
                               cluster_column = 'leiden_clus',
                               group_1 = 1,
                               group_2 = 2)
}

}
