% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{findMastMarkers_one_vs_all}
\alias{findMastMarkers_one_vs_all}
\title{findMastMarkers_one_vs_all}
\usage{
findMastMarkers_one_vs_all(
  gobject,
  feat_type = NULL,
  spat_unit = NULL,
  expression_values = c("normalized", "scaled", "custom"),
  cluster_column,
  subset_clusters = NULL,
  adjust_columns = NULL,
  pval = 0.001,
  logFC = 1,
  min_feats = 10,
  min_genes = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{gobject}{giotto object}

\item{feat_type}{feature type}

\item{spat_unit}{spatial unit}

\item{expression_values}{feat expression values to use}

\item{cluster_column}{clusters to use}

\item{subset_clusters}{selection of clusters to compare}

\item{adjust_columns}{column in pDataDT to adjust for (e.g. detection rate)}

\item{pval}{filter on minimal p-value}

\item{logFC}{filter on logFC}

\item{min_feats}{minimum feats to keep per cluster, overrides pval and logFC}

\item{min_genes}{deprecated, use min_feats}

\item{verbose}{be verbose}

\item{...}{additional parameters for the zlm function in MAST}
}
\value{
data.table with marker feats
}
\description{
Identify marker feats for all clusters in a one vs all manner 
based on the MAST package.
}
\examples{
g <- GiottoData::loadGiottoMini("visium")

findMastMarkers_one_vs_all(gobject = g, cluster_column = "leiden_clus")
}
\seealso{
\code{\link{findMastMarkers}}
}
