% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{findScranMarkers}
\alias{findScranMarkers}
\title{findScranMarkers}
\usage{
findScranMarkers(
  gobject,
  expression_values = c("normalized", "scaled", "custom"),
  cluster_column,
  subset_clusters = NULL,
  group_1 = NULL,
  group_2 = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{gobject}{giotto object}

\item{expression_values}{gene expression values to use}

\item{cluster_column}{clusters to use}

\item{subset_clusters}{selection of clusters to compare}

\item{group_1}{group 1 cluster IDs from cluster_column for pairwise comparison}

\item{group_2}{group 2 cluster IDs from cluster_column for pairwise comparison}

\item{verbose}{be verbose (default = FALSE)}

\item{...}{additional parameters for the findMarkers function in scran}
}
\value{
data.table with marker genes
}
\description{
Identify marker genes for all or selected clusters based on scran's implementation of findMarkers.
}
\details{
This is a minimal convenience wrapper around
the \code{\link[scran]{findMarkers}} function from the scran package.

To perform differential expression between cluster groups you need to specificy cluster IDs
to the parameters \emph{group_1} and \emph{group_2}.
}
\examples{

data(mini_giotto_single_cell)

scran_markers = findScranMarkers(gobject = mini_giotto_single_cell,
                                 cluster_column = 'leiden_clus',
                                 group_1 = 1,
                                 group_2 = 2)

}
