% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{findScranMarkers_one_vs_all}
\alias{findScranMarkers_one_vs_all}
\title{findScranMarkers_one_vs_all}
\usage{
findScranMarkers_one_vs_all(
  gobject,
  expression_values = c("normalized", "scaled", "custom"),
  cluster_column,
  subset_clusters = NULL,
  pval = 0.01,
  logFC = 0.5,
  min_genes = 10,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{gobject}{giotto object}

\item{expression_values}{gene expression values to use}

\item{cluster_column}{clusters to use}

\item{subset_clusters}{subset of clusters to use}

\item{pval}{filter on minimal p-value}

\item{logFC}{filter on logFC}

\item{min_genes}{minimum genes to keep per cluster, overrides pval and logFC}

\item{verbose}{be verbose}

\item{...}{additional parameters for the findMarkers function in scran}
}
\value{
data.table with marker genes
}
\description{
Identify marker genes for all clusters in a one vs all manner based on scran's implementation of findMarkers.
}
\examples{

data(mini_giotto_single_cell)

scran_markers = findScranMarkers_one_vs_all(gobject = mini_giotto_single_cell,
                                            cluster_column = 'leiden_clus')

}
\seealso{
\code{\link{findScranMarkers}}
}
