% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interoperability.R
\name{gefToGiotto}
\alias{gefToGiotto}
\title{Convert gef to Giotto}
\usage{
gefToGiotto(gef_file, bin_size = "bin100", verbose = FALSE)
}
\arguments{
\item{gef_file}{path to .gef file}

\item{bin_size}{bin size to select from .gef file}

\item{verbose}{be verbose}
}
\description{
Converts .gef file (output stereo-seq pipeline) into
giotto subcellular object
}
\details{
Function in beta. Converts .gef object to Giotto object.

There are six possible choices for bin_size: bin1, bin10, bin20, bin50, bin100, bin200.

See SAW pipeline for additional information about the gef file.
}
