% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_help.R
\name{get10Xmatrix_h5}
\alias{get10Xmatrix_h5}
\title{get10Xmatrix_h5}
\usage{
get10Xmatrix_h5(path_to_data, gene_ids = c("symbols", "ensembl"))
}
\arguments{
\item{path_to_data}{path to the 10X .h5 file}

\item{gene_ids}{use gene symbols (default) or ensembl ids for the gene expression matrix}
}
\value{
(list of) sparse expression matrix from 10X
}
\description{
This function creates an expression matrix from a 10X h5 file path
}
\details{
If the .h5 10x file has multiple modalities (e.g. RNA and protein),
 multiple matrices will be returned
}
