% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_cell_metadata}
\alias{get_cell_metadata}
\title{Get cell metadata}
\usage{
get_cell_metadata(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  output = c("cellMetaObj", "data.table"),
  copy_obj = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit (e.g. "cell")}

\item{feat_type}{feature type (e.g. "rna", "dna", "protein")}

\item{output}{return as either 'data.table' or 'cellMetaObj'}
}
\description{
Get cell metadata from giotto object
}
\seealso{
pDataDT
}
\keyword{internal}
