% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_dimReduction}
\alias{get_dimReduction}
\title{get_dimReduction}
\usage{
get_dimReduction(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  reduction = c("cells", "feats"),
  reduction_method = c("pca", "umap", "tsne"),
  name = "pca",
  return_dimObj = FALSE
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit}

\item{feat_type}{feature type}

\item{reduction}{reduction on cells or features}

\item{reduction_method}{reduction method (e.g. pca)}

\item{name}{name of reduction results}

\item{return_dimObj}{return full dimension object result}
}
\value{
dim reduction coordinates (default) or dim reduction object
}
\description{
function to get a dimension reduction object
}
\seealso{
Other dimensional reduction data accessor functions: 
\code{\link{set_dimReduction}()}

Other functions to get data from giotto object: 
\code{\link{get_NearestNetwork}()},
\code{\link{get_expression_values}()},
\code{\link{get_feature_info}()},
\code{\link{get_giottoImage}()},
\code{\link{get_polygon_info}()},
\code{\link{get_spatialGrid}()},
\code{\link{get_spatialNetwork}()},
\code{\link{get_spatial_enrichment}()},
\code{\link{get_spatial_locations}()}
}
\concept{dimensional reduction data accessor functions}
\concept{functions to get data from giotto object}
