% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_dimReduction}
\alias{get_dimReduction}
\title{get_dimReduction}
\usage{
get_dimReduction(
  gobject,
  reduction = c("cells", "genes"),
  reduction_method = c("pca", "umap", "tsne"),
  name = "pca",
  return_dimObj = FALSE
)
}
\value{
dim reduction coordinates (default) or dim reduction object
}
\description{
function to get a dimension reduction object
}
