% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_section.R
\name{get_distance}
\alias{get_distance}
\title{get_distance}
\usage{
get_distance(networkDT, method = c("mean", "median"))
}
\arguments{
\item{networkDT}{networkDT}

\item{method}{method}
}
\description{
estimate average distance between neighboring cells with network table as input
}
\keyword{internal}
