% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_expression_values}
\alias{get_expression_values}
\title{get_expression_values}
\usage{
get_expression_values(gobject, feat_type = NULL, spat_unit = NULL, values)
}
\arguments{
\item{gobject}{giotto object}

\item{feat_type}{feature type}

\item{spat_unit}{spatial unit}

\item{values}{expression values to extract}
}
\value{
expression matrix
}
\description{
function to get expression values from giotto object
}
