% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_expression_values}
\alias{get_expression_values}
\title{Get expression values}
\usage{
get_expression_values(
  gobject,
  values = NULL,
  spat_unit = NULL,
  feat_type = NULL,
  output = c("exprObj", "matrix"),
  set_defaults = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{values}{expression values to extract (e.g. "raw", "normalized", "scaled")}

\item{spat_unit}{spatial unit (e.g. "cell")}

\item{feat_type}{feature type (e.g. "rna", "dna", "protein")}

\item{output}{what object type to retrieve the expression as. Currently either
'matrix' for the matrix object contained in the exprObj or 'exprObj' (default) for
the exprObj itself are allowed.}

\item{set_defaults}{set default spat_unit and feat_type. Change to FALSE only when}
}
\value{
expression matrix
}
\description{
Function to get expression values from giotto object
}
\seealso{
Other expression accessor functions: 
\code{\link{getExpression}()},
\code{\link{setExpression}()},
\code{\link{set_expression_values}()}

Other functions to get data from giotto object: 
\code{\link{getDimReduction}()},
\code{\link{getExpression}()},
\code{\link{getFeatureInfo}()},
\code{\link{getGiottoImage}()},
\code{\link{getMultiomics}()},
\code{\link{getNearestNetwork}()},
\code{\link{getPolygonInfo}()},
\code{\link{getSpatialEnrichment}()},
\code{\link{getSpatialGrid}()},
\code{\link{getSpatialLocations}()},
\code{\link{getSpatialNetwork}()},
\code{\link{get_NearestNetwork}()},
\code{\link{get_dimReduction}()},
\code{\link{get_feature_info}()},
\code{\link{get_giottoImage}()},
\code{\link{get_multiomics}()},
\code{\link{get_polygon_info}()},
\code{\link{get_spatialGrid}()},
\code{\link{get_spatialNetwork}()},
\code{\link{get_spatial_enrichment}()},
\code{\link{get_spatial_locations}()}
}
\concept{expression accessor functions}
\concept{functions to get data from giotto object}
