% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_feature_info}
\alias{get_feature_info}
\title{get_feature_info}
\usage{
get_feature_info(gobject, feat_name = "rna")
}
\arguments{
\item{gobject}{giotto object}

\item{feat_name}{name of feature}
}
\description{
get giotto points spatVector
}
