% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_feature_info}
\alias{get_feature_info}
\title{Get feature info}
\usage{
get_feature_info(gobject, feat_type = NULL)
}
\arguments{
\item{gobject}{giotto object}

\item{feat_type}{name of feature}
}
\description{
Get giotto points spatVector
}
\seealso{
Other feature info data accessor functions: 
\code{\link{set_feature_info}()}

Other functions to get data from giotto object: 
\code{\link{get_NearestNetwork}()},
\code{\link{get_dimReduction}()},
\code{\link{get_expression_values}()},
\code{\link{get_giottoImage}()},
\code{\link{get_polygon_info}()},
\code{\link{get_spatialGrid}()},
\code{\link{get_spatialNetwork}()},
\code{\link{get_spatial_enrichment}()},
\code{\link{get_spatial_locations}()}
}
\concept{feature info data accessor functions}
\concept{functions to get data from giotto object}
