% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_feature_metadata}
\alias{get_feature_metadata}
\title{Get feature metadata}
\usage{
get_feature_metadata(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  output = c("featMetaObj", "data.table"),
  copy_obj = TRUE,
  set_defaults = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit (e.g. "cell")}

\item{feat_type}{feature type (e.g. "rna", "dna", "protein")}

\item{output}{return as either 'data.table' or 'featMetaObj'}

\item{copy_obj}{whether to perform a deepcopy of the data.table information}

\item{set_defaults}{set default spat_unit and feat_type. Change to FALSE only when
expression and spat_info are not expected to exist.}
}
\description{
Get feature metadata from giotto object
}
\seealso{
fDataDT
}
\keyword{internal}
