% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{get_GiottoLargeImage}
\alias{get_GiottoLargeImage}
\title{get_GiottoLargeImage}
\usage{
get_GiottoLargeImage(gobject = NULL, largeImage_name = NULL)
}
\arguments{
\item{gobject}{giotto object}

\item{largeImage_name}{name of giottoLargeImage}
}
\value{
a giottoLargeImage
}
\description{
get a giottoLargeImage from a giottoObject
}
\keyword{internal}
