% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_giottoLargeImage}
\alias{get_giottoLargeImage}
\title{Get \emph{terra}-based giotto \code{largeImage}}
\usage{
get_giottoLargeImage(gobject = NULL, name = NULL)
}
\arguments{
\item{gobject}{giotto object}

\item{name}{name of giottoLargeImage \code{\link{showGiottoImageNames}}}
}
\value{
a giottoLargeImage
}
\description{
Set a giottoLargeImage from a giottoObject
}
\keyword{internal}
