% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_polygon_info}
\alias{get_polygon_info}
\title{get_polygon_info}
\usage{
get_polygon_info(gobject, polygon_name = "cell", polygon_overlap = NULL)
}
\arguments{
\item{gobject}{giotto object}

\item{polygon_name}{name of polygons}

\item{polygon_overlap}{include polygon overlap information}
}
\description{
get giotto polygon spatVector
}
\seealso{
Other polygon info data accessor functions: 
\code{\link{set_polygon_info}()}

Other functions to get data from giotto object: 
\code{\link{get_NearestNetwork}()},
\code{\link{get_dimReduction}()},
\code{\link{get_expression_values}()},
\code{\link{get_feature_info}()},
\code{\link{get_giottoImage}()},
\code{\link{get_spatialGrid}()},
\code{\link{get_spatialNetwork}()},
\code{\link{get_spatial_enrichment}()},
\code{\link{get_spatial_locations}()}
}
\concept{functions to get data from giotto object}
\concept{polygon info data accessor functions}
