% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_polygon_info}
\alias{get_polygon_info}
\title{Get polygon info}
\usage{
get_polygon_info(
  gobject,
  polygon_name = NULL,
  polygon_overlap = NULL,
  return_giottoPolygon = FALSE
)
}
\arguments{
\item{gobject}{giotto object}

\item{polygon_name}{name of polygons. Default "cell"}

\item{polygon_overlap}{include polygon overlap information}

\item{return_giottoPolygon}{(Defaults to FALSE) Return as giottoPolygon S4 object}
}
\description{
Get giotto polygon spatVector
}
\seealso{
Other polygon info data accessor functions: 
\code{\link{getPolygonInfo}()},
\code{\link{setPolygonInfo}()},
\code{\link{set_polygon_info}()}

Other functions to get data from giotto object: 
\code{\link{getDimReduction}()},
\code{\link{getExpression}()},
\code{\link{getFeatureInfo}()},
\code{\link{getGiottoImage}()},
\code{\link{getMultiomics}()},
\code{\link{getNearestNetwork}()},
\code{\link{getPolygonInfo}()},
\code{\link{getSpatialEnrichment}()},
\code{\link{getSpatialGrid}()},
\code{\link{getSpatialLocations}()},
\code{\link{getSpatialNetwork}()},
\code{\link{get_NearestNetwork}()},
\code{\link{get_dimReduction}()},
\code{\link{get_expression_values}()},
\code{\link{get_feature_info}()},
\code{\link{get_giottoImage}()},
\code{\link{get_multiomics}()},
\code{\link{get_spatialGrid}()},
\code{\link{get_spatialNetwork}()},
\code{\link{get_spatial_enrichment}()},
\code{\link{get_spatial_locations}()}
}
\concept{functions to get data from giotto object}
\concept{polygon info data accessor functions}
