% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_section.R
\name{get_sectionThickness}
\alias{get_sectionThickness}
\title{get_sectionThickness}
\usage{
get_sectionThickness(
  gobject,
  thickness_unit = c("cell", "natural"),
  slice_thickness = 2,
  spatial_network_name = "Delaunay_network",
  cell_distance_estimate_method = c("mean", "median"),
  plane_equation = NULL
)
}
\arguments{
\item{gobject}{gobject}

\item{thickness_unit}{thickness_unit}

\item{spatial_network_name}{spatial_network_name}

\item{cell_distance_estimate_method}{cell_distance_estimate_method}

\item{plane_equation}{plane_equation}
}
\description{
get section thickness
}
\keyword{internal}
