% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_spatial_enrichment}
\alias{get_spatial_enrichment}
\title{get_spatial_enrichment}
\usage{
get_spatial_enrichment(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  enrichm_name = "DWLS"
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit}

\item{feat_type}{feature type}

\item{enrichm_name}{name of spatial enrichment results}
}
\value{
data.table with fractions
}
\description{
function to get a spatial enrichment data.table
}
