% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_spatial_enrichment}
\alias{get_spatial_enrichment}
\title{Get spatial enrichment}
\usage{
get_spatial_enrichment(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  enrichm_name = "DWLS"
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit}

\item{feat_type}{feature type}

\item{enrichm_name}{name of spatial enrichment results}
}
\value{
data.table with fractions
}
\description{
Function to get a spatial enrichment data.table
}
\seealso{
Other spatial enrichment data accessor functions: 
\code{\link{set_spatial_enrichment}()}

Other functions to get data from giotto object: 
\code{\link{get_NearestNetwork}()},
\code{\link{get_dimReduction}()},
\code{\link{get_expression_values}()},
\code{\link{get_feature_info}()},
\code{\link{get_giottoImage}()},
\code{\link{get_polygon_info}()},
\code{\link{get_spatialGrid}()},
\code{\link{get_spatialNetwork}()},
\code{\link{get_spatial_locations}()}
}
\concept{functions to get data from giotto object}
\concept{spatial enrichment data accessor functions}
