% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_spatial_locations}
\alias{get_spatial_locations}
\title{Get spatial locations}
\usage{
get_spatial_locations(
  gobject,
  spat_unit = NULL,
  spat_loc_name = NULL,
  output = c("spatLocsObj", "data.table"),
  copy_obj = TRUE,
  verbose = TRUE,
  set_defaults = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit (e.g. "cell")}

\item{spat_loc_name}{name of spatial locations (defaults to first name in spatial_locs slot, e.g. "raw")}

\item{output}{what object type to get the spatial locations as. Default is as
a 'spatLocsObj'. Returning as 'data.table' is also possible.}

\item{copy_obj}{whether to copy/duplicate when getting the object (default = TRUE)}

\item{verbose}{be verbose}

\item{set_defaults}{set default spat_unit and feat_type. Change to FALSE only when
expression and spat_info are not expected to exist.}
}
\value{
data.table with coordinates or spatLocsObj depending on \code{output}
}
\description{
Function to get a spatial location data.table
}
\seealso{
Other spatial location data accessor functions: 
\code{\link{set_spatial_locations}()}

Other functions to get data from giotto object: 
\code{\link{get_NearestNetwork}()},
\code{\link{get_dimReduction}()},
\code{\link{get_expression_values}()},
\code{\link{get_feature_info}()},
\code{\link{get_giottoImage}()},
\code{\link{get_polygon_info}()},
\code{\link{get_spatialGrid}()},
\code{\link{get_spatialNetwork}()},
\code{\link{get_spatial_enrichment}()}
}
\concept{functions to get data from giotto object}
\concept{spatial location data accessor functions}
