% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_spatialNetwork}
\alias{get_spatialNetwork}
\title{Get spatial network}
\usage{
get_spatialNetwork(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  name = NULL,
  return_network_Obj = FALSE
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit}

\item{feat_type}{feature type}

\item{name}{name of spatial network}

\item{return_network_Obj}{return network object (default = FALSE)}
}
\description{
Function to get a spatial network
}
\seealso{
Other spatial network data accessor functions: 
\code{\link{set_spatialNetwork}()}

Other functions to get data from giotto object: 
\code{\link{get_NearestNetwork}()},
\code{\link{get_dimReduction}()},
\code{\link{get_expression_values}()},
\code{\link{get_feature_info}()},
\code{\link{get_giottoImage}()},
\code{\link{get_polygon_info}()},
\code{\link{get_spatialGrid}()},
\code{\link{get_spatial_enrichment}()},
\code{\link{get_spatial_locations}()}
}
\concept{functions to get data from giotto object}
\concept{spatial network data accessor functions}
