% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_spatialNetwork}
\alias{get_spatialNetwork}
\title{Get spatial network}
\usage{
get_spatialNetwork(
  gobject,
  spat_unit = NULL,
  name = NULL,
  output = c("spatialNetworkObj", "networkDT", "networkDT_before_filter", "outputObj"),
  set_defaults = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit (e.g. "cell")}

\item{name}{name of spatial network}

\item{output}{object type to return as. Options: 'spatialNetworkObj' (default),
'networkDT' and 'networkDT_before_filter' for data.table outputs.}

\item{set_defaults}{set default spat_unit and feat_type. Change to FALSE only when}
}
\description{
Function to get a spatial network
}
\seealso{
Other spatial network data accessor functions: 
\code{\link{set_spatialNetwork}()}

Other functions to get data from giotto object: 
\code{\link{get_NearestNetwork}()},
\code{\link{get_dimReduction}()},
\code{\link{get_expression_values}()},
\code{\link{get_feature_info}()},
\code{\link{get_giottoImage}()},
\code{\link{get_polygon_info}()},
\code{\link{get_spatialGrid}()},
\code{\link{get_spatial_enrichment}()},
\code{\link{get_spatial_locations}()}
}
\concept{functions to get data from giotto object}
\concept{spatial network data accessor functions}
