% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactivity.R
\name{getCellsFromPolygon}
\alias{getCellsFromPolygon}
\title{Get cells located within the polygons area}
\usage{
getCellsFromPolygon(
  gobject,
  polygon_name = "selections",
  spat_unit = "cell",
  spat_loc_name = "raw",
  polygons = NULL
)
}
\arguments{
\item{gobject}{A Giotto object}

\item{polygon_name}{name of polygon selections}

\item{spat_unit}{spatial unit, default = 'cell'}

\item{spat_loc_name}{name of spatial locations to use, default = 'raw'}

\item{polygons}{character. A vector with polygon names to extract cells 
from. If NULL, cells from all polygons are retrieved}
}
\value{
A terra 'SpatVector' with cell ID, x y coordinates, and polygon ID 
where each cell is located in.
}
\description{
Get cells located within the polygons area
}
\examples{
## Plot interactive polygons
g <- GiottoData::loadGiottoMini("visium")
my_polygon_coords <- data.frame(poly_ID = rep("polygon1", 3), 
sdimx = c(5477, 5959, 4720), sdimy = c(-4125, -2808, -5202))

## Add polygon coordinates to Giotto object
my_giotto_polygons <- createGiottoPolygonsFromDfr(my_polygon_coords, 
name = "selections")
g <- addGiottoPolygons(
    gobject = g,
    gpolygons = list(my_giotto_polygons)
)

## Add polygon IDs to cell metadata
addPolygonCells(g)

## Get only cells from polygon 1
getCellsFromPolygon(g)

}
