% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{getDendrogramSplits}
\alias{getDendrogramSplits}
\title{getDendrogramSplits}
\usage{
getDendrogramSplits(
  gobject,
  expression_values = c("normalized", "scaled", "custom"),
  cluster_column,
  cor = c("pearson", "spearman"),
  distance = "ward.D",
  h = NULL,
  h_color = "red",
  show_dend = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{expression_values}{expression values to use}

\item{cluster_column}{name of column to use for clusters}

\item{cor}{correlation score to calculate distance}

\item{distance}{distance method to use for hierarchical clustering}

\item{h}{height of horizontal lines to plot}

\item{h_color}{color of horizontal lines}

\item{show_dend}{show dendrogram}

\item{verbose}{be verbose}
}
\value{
data.table object
}
\description{
Split dendrogram at each node and keep the leave (label) information..
}
\details{
Creates a data.table with three columns and each row represents a node in the
dendrogram. For each node the height of the node is given together with the two
subdendrograms. This information can be used to determine in a hierarchical manner
differentially expressed marker genes at each node.
}
\examples{
\dontrun{
data("mini_giotto_single_cell")

splits = getDendrogramSplits(mini_giotto_single_cell, cluster_column = 'leiden_clus')
}
}
