% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction.R
\name{getGeneToGeneScores}
\alias{getGeneToGeneScores}
\title{getGeneToGeneScores}
\usage{
getGeneToGeneScores(
  CPGscore,
  selected_genes = NULL,
  specific_genes_1 = NULL,
  specific_genes_2 = NULL,
  min_cells = 5,
  min_fdr = 0.05,
  min_spat_diff = 0.2,
  min_log2_fc = 0.5,
  direction = c("both", "up", "down"),
  fold_change_addendum = 0.1,
  verbose = TRUE
)
}
\arguments{
\item{CPGscore}{CPGscore, output from getCellProximityGeneScores()}

\item{selected_genes}{select subset of genes}

\item{specific_genes_1}{specific source genes (see details)}

\item{specific_genes_2}{specific target genes (see details)}

\item{min_cells}{min number of cells threshold}

\item{min_spat_diff}{spatial difference threshold}

\item{min_log2_fc}{log2 fold-change threshold}

\item{direction}{up or downregulation or both}

\item{fold_change_addendum}{constant to add when calculating log2 fold-change}

\item{verbose}{verbose}

\item{min_pval}{p-value threshold}
}
\value{
Gene to gene scores in data.table format
}
\description{
Compute gene-gene enrichment scores.
}
\details{
Give more details ...
}
\examples{
    getGeneToGeneScores(CPGscore)
}
