% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{getGiottoImage}
\alias{getGiottoImage}
\title{Get giotto image object}
\usage{
getGiottoImage(
  gobject = NULL,
  image_type = c("image", "largeImage"),
  name = NULL
)
}
\arguments{
\item{gobject}{giotto object}

\item{image_type}{type of giotto image object. Either "image" or "largeImage"}

\item{name}{name of a giotto image object \code{\link{showGiottoImageNames}}}
}
\value{
a giotto image object
}
\description{
Get giotto image object from gobject
}
\seealso{
Other image data accessor functions: 
\code{\link{get_giottoImage}()},
\code{\link{setGiottoImage}()},
\code{\link{set_giottoImage}()}

Other functions to get data from giotto object: 
\code{\link{getDimReduction}()},
\code{\link{getExpression}()},
\code{\link{getFeatureInfo}()},
\code{\link{getMultiomics}()},
\code{\link{getNearestNetwork}()},
\code{\link{getPolygonInfo}()},
\code{\link{getSpatialEnrichment}()},
\code{\link{getSpatialGrid}()},
\code{\link{getSpatialLocations}()},
\code{\link{getSpatialNetwork}()},
\code{\link{get_NearestNetwork}()},
\code{\link{get_dimReduction}()},
\code{\link{get_expression_values}()},
\code{\link{get_feature_info}()},
\code{\link{get_giottoImage}()},
\code{\link{get_multiomics}()},
\code{\link{get_polygon_info}()},
\code{\link{get_spatialGrid}()},
\code{\link{get_spatialNetwork}()},
\code{\link{get_spatial_enrichment}()},
\code{\link{get_spatial_locations}()}
}
\concept{functions to get data from giotto object}
\concept{image data accessor functions}
