% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{getGiottoImage}
\alias{getGiottoImage}
\title{getGiottoImage}
\usage{
getGiottoImage(gobject, image_name)
}
\arguments{
\item{gobject}{giotto object}

\item{image_name}{name of giotto image \code{\link{showGiottoImageNames}}}
}
\value{
a giotto image
}
\description{
get get a giotto image from a giotto object
}
