% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_help.R
\name{getSpatialDataset}
\alias{getSpatialDataset}
\title{getSpatialDataset}
\usage{
getSpatialDataset(
  dataset = c("ST_OB_1", "ST_OB_2", "codex_spleen", "cycif_PDAC", "starmap_3D_cortex",
    "osmfish_SS_cortex", "merfish_preoptic", "seqfish_SS_cortex", "seqfish_OB"),
  directory = getwd()
)
}
\arguments{
\item{dataset}{dataset to download}

\item{directory}{directory to save the data to}
}
\description{
This package will automatically download the spatial locations and
expression matrix for the chosen dataset. These files are already in the right format
to create a Giotto object.
}
