% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{giottoPoints-class}
\alias{giottoPoints-class}
\alias{giottoPoints}
\title{S4 giotto points Class}
\description{
Giotto class to store and operate on points data
}
\details{
Contains vector-type feature data
}
\section{Slots}{

\describe{
\item{\code{feat_type}}{name of feature type}

\item{\code{spatVector}}{terra spatVector to store point shapes}

\item{\code{networks}}{feature networks}

\item{\code{unique_ID_cache}}{cached unique feature IDs that should match the spatVector slot}
}}

\concept{giotto points class}
