% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{giottoPolygon-class}
\alias{giottoPolygon-class}
\alias{giottoPolygon}
\title{S4 giotto polygon Class}
\description{
Giotto class to store and operate on polygon-like data
}
\details{
holds polygon data
}
\section{Slots}{

\describe{
\item{\code{name}}{name of polygon shapes}

\item{\code{spatVector}}{terra spatVector to store polygon shapes}

\item{\code{spatVectorCentroids}}{centroids of polygon shapes}

\item{\code{overlaps}}{information about overlapping points and polygons}

\item{\code{unique_ID_cache}}{cached unique spatial IDs that should match the spatVector slot}
}}

\concept{giotto polygon class}
