% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_toSeurat.R
\name{giottoToSeurat}
\alias{giottoToSeurat}
\title{giottoToSeurat}
\usage{
giottoToSeurat(obj_use = NULL, ...)
}
\arguments{
\item{obj_use}{Giotto object}
}
\value{
Seurat object
}
\description{
Converts Giotto object into a Seurat object
}
