% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interoperability.R
\name{giottoToSeurat}
\alias{giottoToSeurat}
\title{Convert Giotto to Seurat}
\usage{
giottoToSeurat(obj_use = NULL, ...)
}
\arguments{
\item{obj_use}{Giotto object}

\item{...}{additional params to pass to \code{\link{get_spatial_locations}}}
}
\value{
Seurat object
}
\description{
Converts Giotto object into a Seurat object
}
