% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interoperability.R
\name{giottoToSpatialExperiment}
\alias{giottoToSpatialExperiment}
\title{Utility function to convert a Giotto object to a SpatialExperiment object.}
\usage{
giottoToSpatialExperiment(giottoObj, verbose = TRUE)
}
\arguments{
\item{giottoObj}{Input Giotto object to convert to a SpatialExperiment object.}

\item{verbose}{A boolean value specifying if progress messages should be displayed
or not. Default \code{TRUE}.}
}
\value{
A SpatialExperiment object that contains data from the input Giotto object.
}
\description{
Utility function to convert a Giotto object to a SpatialExperiment object.
}
\examples{
\dontrun{
mini_gobject <- GiottoData::loadGiottoMini('vizgen')
giottoToSpatialExperiment(mini_gobject)
}
}
