% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gobject_mini_vizgen}
\alias{gobject_mini_vizgen}
\title{gobject_mini_vizgen}
\format{
An object of class \code{"giotto"};
see \code{\link[Giotto]{createGiottoObject}} or
 \code{\link[Giotto]{createGiottoObjectSubcellular}}.
}
\usage{
loadGiottoMini(dataset = 'vizgen')
}
\description{
gobject_mini_vizgen
}
\references{
www.vizgen.com
}
\keyword{datasets}
