% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_genes.R
\name{heatmSpatialCorGenes}
\alias{heatmSpatialCorGenes}
\title{heatmSpatialCorGenes}
\usage{
heatmSpatialCorGenes(
  spatCorObject,
  use_clus_name = NULL,
  show_cluster_annot = TRUE,
  ...
)
}
\arguments{
\item{use_clus_name}{name of clusters to visualize (from clusterSpatialCorGenes())}

\item{show_cluster_annot}{show cluster annotation on top of heatmap}

\item{...}{additional parameters to the \code{\link[ComplexHeatmap]{Heatmap}} function from ComplexHeatmap}
}
\value{
Heatmap generated by ComplexHeatmap
}
\description{
Create heatmap of spatially correlated genes
}
\examples{
    heatmSpatialCorGenes(gobject)
}
