% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_genes.R
\name{heatmSpatialCorGenes}
\alias{heatmSpatialCorGenes}
\title{heatmSpatialCorGenes}
\usage{
heatmSpatialCorGenes(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=heatmSpatialCorFeats]{heatmSpatialCorFeats}}
  \describe{
    \item{\code{gobject}}{giotto object}
    \item{\code{spatCorObject}}{spatial correlation object}
    \item{\code{use_clus_name}}{name of clusters to visualize
(from clusterSpatialCorFeats())}
    \item{\code{show_cluster_annot}}{show cluster annotation on top of heatmap}
    \item{\code{show_row_dend}}{show row dendrogram}
    \item{\code{show_column_dend}}{show column dendrogram}
    \item{\code{show_row_names}}{show row names}
    \item{\code{show_column_names}}{show column names}
    \item{\code{show_plot}}{show plot}
    \item{\code{return_plot}}{return ggplot object}
    \item{\code{save_plot}}{directly save the plot [boolean]}
    \item{\code{save_param}}{list of saving parameters, see
\code{\link{showSaveParameters}}}
    \item{\code{default_save_name}}{default save name for saving, don't change,
change save_name in save_param}
  }}
}
\description{
Create heatmap of spatially correlated genes
}
\seealso{
\code{\link{heatmSpatialCorFeats}}
}
