% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_genes.R
\name{heatmSpatialCorGenes}
\alias{heatmSpatialCorGenes}
\title{heatmSpatialCorGenes}
\usage{
heatmSpatialCorGenes(
  gobject,
  spatCorObject,
  use_clus_name = NULL,
  show_cluster_annot = TRUE,
  show_row_dend = T,
  show_column_dend = F,
  show_row_names = F,
  show_column_names = F,
  show_plot = NA,
  return_plot = NA,
  save_plot = NA,
  save_param = list(),
  default_save_name = "heatmSpatialCorGenes",
  ...
)
}
\arguments{
\item{gobject}{giotto object}

\item{spatCorObject}{spatial correlation object}

\item{use_clus_name}{name of clusters to visualize (from clusterSpatialCorGenes())}

\item{show_cluster_annot}{show cluster annotation on top of heatmap}

\item{show_row_dend}{show row dendrogram}

\item{show_column_dend}{show column dendrogram}

\item{show_row_names}{show row names}

\item{show_column_names}{show column names}

\item{show_plot}{show plot}

\item{return_plot}{return ggplot object}

\item{save_plot}{directly save the plot [boolean]}

\item{save_param}{list of saving parameters, see \code{\link{showSaveParameters}}}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}

\item{\dots}{additional parameters to the \code{\link[ComplexHeatmap]{Heatmap}} function from ComplexHeatmap}
}
\value{
Heatmap generated by ComplexHeatmap
}
\description{
Create heatmap of spatially correlated genes
}
