% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_enrichment.R
\name{hyperGeometricEnrich}
\alias{hyperGeometricEnrich}
\title{hyperGeometricEnrich}
\usage{
hyperGeometricEnrich(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=runHyperGeometricEnrich]{runHyperGeometricEnrich}}
  \describe{
    \item{\code{gobject}}{Giotto object}
    \item{\code{sign_matrix}}{Matrix of signature genes for each cell type / process}
    \item{\code{expression_values}}{expression values to use}
    \item{\code{reverse_log_scale}}{reverse expression values from log scale}
    \item{\code{logbase}}{log base to use if reverse_log_scale = TRUE}
    \item{\code{top_percentage}}{percentage of cells that will be considered to have gene expression with matrix binarization}
    \item{\code{output_enrichment}}{how to return enrichment output}
    \item{\code{p_value}}{calculate p-values (boolean, default = FALSE)}
    \item{\code{name}}{to give to spatial enrichment results, default = rank}
    \item{\code{return_gobject}}{return giotto object}
  }}
}
\description{
Function to calculate gene signature enrichment scores per spatial position using a hypergeometric test.
}
\seealso{
\code{\link{runHyperGeometricEnrich}}
}
