% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{importCosMx}
\alias{importCosMx}
\title{Import a CosMx Assay}
\usage{
importCosMx(cosmx_dir = NULL, fovs = NULL)
}
\arguments{
\item{cosmx_dir}{CosMx output directory}

\item{fovs}{numeric. (optional) If provided, will load specific fovs.
Otherwise, all FOVs will be loaded}
}
\value{
CosmxReader object
}
\description{
Giotto import functionalities for CosMx datasets. This function generates
a `CosmxReader` instance that has convenient reader functions for converting
individual pieces of CosMx data into Giotto-compatible representations when
the params `cosmx_dir` and `fovs` (if only a subset is desired) are provided.
A function that creates the full `giotto` object is also available.
These functions should have all param values provided as defaults, but
can be flexibly modified to do things such as look in alternative
directories or paths.
}
\examples{
# Create a `CosmxReader` object
reader <- importCosMx()

\dontrun{
# Set the cosmx_dir and fov parameters
reader$cosmx_dir <- "path to cosmx dir"
reader$fov <- c(1, 4)

# Load polygons, transcripts, and images
polys <- reader$load_polys()
tx <- reader$load_transcripts()
imgs <- reader$load_images()

# Create a `giotto` object and add the loaded data
g <- giotto()
g <- setGiotto(g, tx[["rna"]])
g <- setGiotto(g, polys)
g <- addGiottoLargeImage(g, largeImages = imgs)
force(g)
}
}
