% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_xenium.R
\name{importXenium}
\alias{importXenium}
\title{Import a 10X Xenium Assay}
\usage{
importXenium(xenium_dir = NULL, qv_threshold = 20)
}
\arguments{
\item{xenium_dir}{Xenium output directory}
}
\value{
`XeniumReader` object
}
\description{
Giotto import functionalities for Xenium datasets. This function creates a
`XeniumReader` instance that has convenient reader functions for converting
individual pieces of Xenium data into Giotto-compatible representations.

These functions should have all param values provided as defaults, but
can be flexibly modified to do things such as look in alternative
directories or paths
}
