% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_environment.R
\name{installGiottoEnvironment}
\alias{installGiottoEnvironment}
\title{installGiottoEnvironment}
\usage{
installGiottoEnvironment(
  packages_to_install = c("pandas", "networkx", "python-igraph", "leidenalg",
    "python-louvain", "python.app", "scikit-learn"),
  force_miniconda = FALSE,
  force_environment = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{packages_to_install}{all python modules (packages) that should be installed for Giotto to work}

\item{force_miniconda}{force reinstallation of miniconda}

\item{force_environment}{force reinstallation of the giotto environment}

\item{verbose}{be verbose}
}
\value{
installs a giotto environment using the reticulate miniconda system
}
\description{
Installs a giotto environment
}
\details{
This function will install a local giotto environment using
the miniconda system as implemented by reticulate. Once this giotto environment is
installed it will be automatically detected when you run the Giotto toolbox. If you want to use
your own python path then you can set the python_path in the \code{\link{createGiottoInstructions}}
and provide the instructions to the \code{\link{createGiottoObject}} function.
}
\examples{
\dontrun{

  # this command will install r-miniconda
  # and a giotto environment with all necessary python modules
  installGiottoEnvironment()
}

}
