% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registration.R
\name{InteractiveLandmarkSelection}
\alias{InteractiveLandmarkSelection}
\title{Record landmarks by interactive selection}
\usage{
InteractiveLandmarkSelection(source, target)
}
\arguments{
\item{source_image}{the image to be plotted on the left, and landmarks will output in the first of the list. Input can be a ggplot object, a GiottoImage, or a character represent a path to a image}

\item{target_image}{the image to be plotted on the right, and landmarks will output in the second of the list. Input can be a ggplot object, a GiottoImage, or a character represent a path to a image}
}
\value{
a list of landmarks
}
\description{
Record landmarks by interactive selection
}
