% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.R
\name{jackstrawPlot}
\alias{jackstrawPlot}
\title{jackstrawPlot}
\usage{
jackstrawPlot(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  expression_values = c("normalized", "scaled", "custom"),
  reduction = c("cells", "feats"),
  feats_to_use = NULL,
  center = FALSE,
  scale_unit = FALSE,
  ncp = 20,
  ylim = c(0, 1),
  iter = 10,
  threshold = 0.01,
  verbose = TRUE,
  show_plot = NULL,
  return_plot = NULL,
  save_plot = NULL,
  save_param = list(),
  default_save_name = "jackstrawPlot"
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit (e.g. "cell")}

\item{feat_type}{feature type (e.g. "rna", "dna", "protein")}

\item{expression_values}{expression values to use}

\item{reduction}{cells or genes}

\item{feats_to_use}{subset of features to use for PCA}

\item{center}{center data before PCA}

\item{scale_unit}{scale features before PCA}

\item{ncp}{number of principal components to calculate}

\item{ylim}{y-axis limits on jackstraw plot}

\item{iter}{number of interations for jackstraw}

\item{threshold}{p-value threshold to call a PC significant}

\item{verbose}{show progress of jackstraw method}

\item{show_plot}{logical. show plot}

\item{return_plot}{logical. return ggplot object}

\item{save_plot}{logical. save the plot}

\item{save_param}{list of saving parameters, see \code{\link{showSaveParameters}}}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}
}
\value{
ggplot object for jackstraw method
}
\description{
identify significant prinicipal components (PCs)
}
\details{
The Jackstraw method uses the \code{\link[jackstraw]{permutationPA}} function. By
 systematically permuting genes it identifies robust, and thus significant, PCs.
 \cr
}
