% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.R
\name{jackstrawPlot}
\alias{jackstrawPlot}
\title{jackstrawPlot}
\usage{
jackstrawPlot(
  gobject,
  expression_values = c("normalized", "scaled", "custom"),
  reduction = c("cells", "genes"),
  genes_to_use = NULL,
  center = FALSE,
  scale_unit = FALSE,
  ncp = 20,
  ylim = c(0, 1),
  iter = 10,
  threshold = 0.01,
  verbose = TRUE,
  show_plot = NA,
  return_plot = NA,
  save_plot = NA,
  save_param = list(),
  default_save_name = "jackstrawPlot"
)
}
\arguments{
\item{gobject}{giotto object}

\item{expression_values}{expression values to use}

\item{reduction}{cells or genes}

\item{genes_to_use}{subset of genes to use for PCA}

\item{center}{center data before PCA}

\item{scale_unit}{scale features before PCA}

\item{ncp}{number of principal components to calculate}

\item{ylim}{y-axis limits on jackstraw plot}

\item{iter}{number of interations for jackstraw}

\item{threshold}{p-value threshold to call a PC significant}

\item{verbose}{show progress of jackstraw method}

\item{show_plot}{show plot}

\item{return_plot}{return ggplot object}

\item{save_plot}{directly save the plot [boolean]}

\item{save_param}{list of saving parameters from all_plots_save_function()}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}
}
\value{
ggplot object for jackstraw method
}
\description{
identify significant prinicipal components (PCs)
}
\details{
The Jackstraw method uses the \code{\link[jackstraw]{permutationPA}} function. By
 systematically permuting genes it identifies robust, and thus significant, PCs.
 \cr
}
\examples{

\donttest{

data(mini_giotto_single_cell)

# jackstraw package is required to run
jackstrawPlot(mini_giotto_single_cell, ncp = 10)

}

}
