% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto.R
\name{joinGiottoObjects}
\alias{joinGiottoObjects}
\title{Join giotto objects}
\usage{
joinGiottoObjects(
  gobject_list,
  gobject_names = NULL,
  join_method = c("shift", "z_stack", "no_change"),
  z_vals = 1000,
  x_shift = NULL,
  y_shift = NULL,
  x_padding = NULL,
  y_padding = NULL,
  verbose = TRUE
)
}
\arguments{
\item{gobject_list}{list of giotto objects}

\item{gobject_names}{unique giotto names for each giotto object}

\item{join_method}{method to join giotto objects, see details}

\item{z_vals}{distance(s) along z-axis if method is z-stack (default is step of 1000)}

\item{x_shift}{list of values to shift along x-axis if method is shift}

\item{y_shift}{list of values to shift along y-axis if method is shift}

\item{x_padding}{padding between datasets/images if method is shift}

\item{y_padding}{padding between datasets/images if method is shift}

\item{verbose}{be verbose
Preview where each gobject will be in space with bounding polygons}
}
\value{
giotto object
}
\description{
Function to join multiple giotto objects together
}
\details{
This function joins both the expression and spatial information of
multiple giotto objects into a single one. Giotto supports multiple ways of
joining spatial information as selected through param \code{join_method}:

\itemize{
  \item{\strong{\code{"shift"}}} {
    (default) Spatial locations of different datasets are shifted
    by numeric vectors of values supplied through \code{x_shift}, \code{y_shift},
    \code{x_padding}, and \code{y_padding}. This is particularly useful for data
    that is provided as tiles or ROIs or when analyzing multiple spatial datasets
    together and keeping their spatial data separate.

    \strong{If shift values are given then a value is needed for each giotto object
    to be joined in \code{gobject_list}. Order matters.}

    If a regular step value is desired instead of a specific list of values, use
    \code{x_padding} and \code{y_padding}. Both shift and padding values can be used
    at the same time.

    Leaving \code{x_shift} and \code{y_shift} values as \code{NULL} will have Giotto
    estimate an appropriate \code{x_shift} value based on the x dimension of
    available image objects. If no image objects are available, a default behavior of
    \code{x_padding = 1000} will be applied.
  }
  \item{\strong{\code{"z_stack"}}} {
    Datasets are spatially combined with no change to x and y
    spatial locations, but a z value is incorporated for each dataset based on input
    supplied through param \code{z_vals}. To specify a z value for each dataset to
    join, a numeric vector must be given with a value for each element in \code{gobject_list}.
    Order matters.

    Alternatively, a single numeric value can be supplied to \code{z_vals} in which
    case this input will be treated as a z step value.
  }
  \item{\strong{\code{"no_change"}}} {
    No changes are applied to the spatial locations of the datasets when joining.
  }
}
}
\concept{giotto}
