% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_help.R
\name{kmeans_binarize_wrapper}
\alias{kmeans_binarize_wrapper}
\title{kmeans_binarize_wrapper}
\usage{
kmeans_binarize_wrapper(
  gobject,
  expression_values = c("normalized", "scaled", "custom"),
  subset_genes = NULL,
  kmeans_algo = c("kmeans", "kmeans_arma", "kmeans_arma_subset"),
  nstart = 3,
  iter_max = 10,
  extreme_nr = 50,
  sample_nr = 50,
  set.seed = NULL
)
}
\description{
wrapper for different binarization functions
}
\keyword{internal}
