% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{list_cell_metadata}
\alias{list_cell_metadata}
\title{list_cell_metadata}
\usage{
list_cell_metadata(gobject, spat_unit = NULL, feat_type = NULL)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit (e.g. "cell")}

\item{feat_type}{feature type (e.g. "rna", "dna", "protein")}
}
\value{
names and locations of available cell metadata as data.table
}
\description{
lists the available cell metadata
}
