% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{list_dim_reductions}
\alias{list_dim_reductions}
\title{list_dim_reductions}
\usage{
list_dim_reductions(
  gobject,
  data_type = NULL,
  spat_unit = NULL,
  feat_type = NULL,
  dim_type = NULL
)
}
\arguments{
\item{gobject}{giotto object}

\item{data_type}{cells or feats data used in dim reduction}

\item{spat_unit}{spatial unit}

\item{feat_type}{feature type}

\item{dim_type}{dimensional reduction method}
}
\value{
names and locations of dimension reduction as a dataframe
}
\description{
return the available dimension reductions
}
