% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{list_expression}
\alias{list_expression}
\title{list_expression}
\usage{
list_expression(gobject, spat_unit = NULL, feat_type = NULL)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit (e.g. "cell")}

\item{feat_type}{feature type (e.g. "rna", "dna", "protein")}
}
\value{
names and locations of available matrices as data.table. col order matters.
}
\description{
lists the available matrices
}
