% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{list_giotto_data}
\alias{list_giotto_data}
\title{list_giotto_data}
\usage{
list_giotto_data(gobject = NULL, slot = NULL, ...)
}
\arguments{
\item{gobject}{giotto object}

\item{slot}{giotto object slot of interest (e.g. "expression", "spatial_locs", etc.)}

\item{...}{additional params to pass}
}
\value{
names and locations of data within giotto object slot
}
\description{
list the available data within specified giotto object slot
}
\keyword{internal}
