% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{list_image_names}
\alias{list_image_names}
\title{list_image_names}
\usage{
list_image_names(gobject, img_type = NULL)
}
\arguments{
\item{gobject}{a giotto object}

\item{img_type}{image or largeImage}
}
\value{
a vector of giotto image names attached to the giotto object
}
\description{
Prints the available giotto images that are attached to the Giotto object
}
