% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{list_spatial_enrichments}
\alias{list_spatial_enrichments}
\title{list_spatial_enrichments}
\usage{
list_spatial_enrichments(gobject, spat_unit = NULL, feat_type = NULL)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit}
}
\value{
names and locations of available data.table as dataframe
}
\description{
return the available spatial enrichment results
}
