% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{list_spatial_grids}
\alias{list_spatial_grids}
\title{list_spatial_grids}
\usage{
list_spatial_grids(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  return_uniques = FALSE
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit (e.g. "cell")}

\item{feat_type}{feature type (e.g. "rna", "dna", "protein")}

\item{return_uniques}{return unique nesting names (ignores if final object exists/is correct class)}
}
\value{
data.table of names and locations of available spatial grids. col order matters
}
\description{
return the available spatial grids that are attached to the Giotto object
}
