% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{list_spatial_locations}
\alias{list_spatial_locations}
\title{list_spatial_locations}
\usage{
list_spatial_locations(gobject, spat_unit = NULL)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit (e.g. "cell")}
}
\value{
names and locations of available data.table as data.table
}
\description{
shows the available spatial locations
}
