% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{list_spatial_networks}
\alias{list_spatial_networks}
\title{list_spatial_networks}
\usage{
list_spatial_networks(gobject, spat_unit = NULL, feat_type = NULL)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit}

\item{feat_type}{feature type}
}
\value{
data.table of names and locations of available spatial networks. col order matters
}
\description{
return the available spatial networks that are attached to the Giotto object
}
