% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONTraC_wrapper.R
\name{load_cell_bin_niche_cluster}
\alias{load_cell_bin_niche_cluster}
\title{load_cell_bin_niche_cluster}
\usage{
load_cell_bin_niche_cluster(gobject, ontrac_results_dir = getwd())
}
\arguments{
\item{gobject}{giotto object}

\item{ontrac_results_dir}{the directory where the ONTraC results are saved}
}
\value{
gobject with cell-level binarized niche cluster
}
\description{
load cell-level binarized niche cluster
}
\details{
This function loads the ONTraC outputed cell-level binarized niche
cluster into the giotto object.
}
