% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_create.R
\name{load_merscope_folder}
\alias{load_merscope_folder}
\alias{load_merscope_folder_subcellular}
\alias{load_merscope_folder_aggregate}
\title{Load MERSCOPE data from folder}
\usage{
load_merscope_folder(
  dir_items,
  data_to_use,
  fovs = NULL,
  cores = NA,
  verbose = TRUE
)

load_merscope_folder_subcellular(
  dir_items,
  data_to_use,
  cores = NA,
  verbose = TRUE,
  fovs = NULL
)

load_merscope_folder_aggregate(
  dir_items,
  data_to_use,
  cores = NA,
  verbose = TRUE
)
}
\arguments{
\item{dir_items}{list of full filepaths from \code{\link{read_merscope_folder}}}

\item{data_to_use}{which of either the 'subcellular' or 'aggregate' information
to use for object creation}

\item{cores}{how many cores or threads to use to read data if paths are provided}

\item{verbose}{be verbose when building Giotto object}
}
\value{
list of loaded-in MERSCOPE data
}
\description{
Load MERSCOPE data from folder
}
\section{Functions}{
\itemize{
\item \code{load_merscope_folder_subcellular()}: Load items for 'subcellular' workflow

\item \code{load_merscope_folder_aggregate()}: Load items for 'aggregate' workflow

}}
\keyword{internal}
