% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_help.R
\name{loadGiotto}
\alias{loadGiotto}
\title{loadGiotto}
\usage{
loadGiotto(path_to_folder, verbose = TRUE)
}
\arguments{
\item{path_to_folder}{path to folder where Giotto object was stored with \code{\link{saveGiotto}}}

\item{verbose}{be verbose}
}
\value{
Giotto object
}
\description{
Saves a Giotto object to a specific folder structure
}
\details{
Works together with \code{\link{saveGiotto}} to save and re-load
Giotto objects.
}
