% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_hmrf.R
\name{loadHMRF}
\alias{loadHMRF}
\title{loadHMRF}
\usage{
loadHMRF(
  name_used = "test",
  output_folder_used,
  k_used = 10,
  betas_used,
  python_path_used
)
}
\arguments{
\item{name_used}{name of HMRF that was run}

\item{output_folder_used}{output folder that was used}

\item{k_used}{number of HMRF domains that was tested}

\item{betas_used}{betas that were tested}

\item{python_path_used}{python path that was used}
}
\value{
reloads a previous ran HMRF from doHRMF
}
\description{
load previous HMRF
}
\details{
Description of HMRF parameters ...
}
