% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction.R
\name{make_simulated_network}
\alias{make_simulated_network}
\title{make_simulated_network}
\usage{
make_simulated_network(
  gobject,
  spatial_network_name = "Delaunay_network",
  cluster_column,
  number_of_simulations = 100,
  set_seed = TRUE,
  seed_number = 1234
)
}
\description{
Simulate random network.
}
\keyword{internal}
