% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto_structures.R
\name{makePseudoVisium}
\alias{makePseudoVisium}
\title{makePseudoVisium}
\usage{
makePseudoVisium(spat_locs = NULL, center_center_distance_um = 0.5)
}
\arguments{
\item{spat_locs}{Spatial locations of giotto object with columns sdimx and sdimy}

\item{center_center_distance_um}{Center-to-center distance of input spat_locs in micrometers}
}
\value{
A data.table of vertices for the pseudo-visium spots.
}
\description{
Generates a pseudo-visium grid of spots based on the input spatial locations
}
\details{
This function generates a pseudo-Visium grid of spots based on the input spatial locations. The center-to-center distance parameter is used to determine the size of the spots
}
\concept{spatial location}
