% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_enrichment.R
\name{makeSignMatrixDWLS}
\alias{makeSignMatrixDWLS}
\title{makeSignMatrixDWLS}
\usage{
makeSignMatrixDWLS(gobject, sign_gene, cell_type)
}
\arguments{
\item{gobject}{Giotto object of single cell}

\item{sign_gene}{all of DE genes (signature)}

\item{cell_type}{cell type for single cells}
}
\value{
matrix
}
\description{
Function to convert signature genes into an average expression matrix format that can be used for
spatialDWLS. gobject is the single cell Giotto object. sign_gene is in character format contains all DEGs.
cell_type are all the cell types used for single cell analysis.
}
\seealso{
\code{\link{spatialDWLS}}
}
