% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_enrichment.R
\name{makeSignMatrixDWLSfromMatrix}
\alias{makeSignMatrixDWLSfromMatrix}
\title{makeSignMatrixDWLSfromMatrix}
\usage{
makeSignMatrixDWLSfromMatrix(matrix, sign_gene, cell_type_vector)
}
\arguments{
\item{matrix}{scRNA-seq matrix}

\item{sign_gene}{genes to use (e.g. marker genes)}

\item{cell_type_vector}{vector with cell types (length = ncol(matrix))}
}
\value{
matrix
}
\description{
Function to convert a single-cell RNAseq matrix into a format
 that can be used with \code{\link{runDWLSDeconv}}.
}
\seealso{
\code{\link{runDWLSDeconv}}
}
