% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_enrichment.R
\name{makeSignMatrixDWLSfromMatrix}
\alias{makeSignMatrixDWLSfromMatrix}
\title{makeSignMatrixDWLSfromMatrix}
\usage{
makeSignMatrixDWLSfromMatrix(matrix, sign_gene, cell_type_vector)
}
\arguments{
\item{matrix}{scRNA-seq matrix}

\item{sign_gene}{genes to use (e.g. marker genes)}

\item{cell_type_vector}{vector with cell types (length = ncol(matrix))}
}
\value{
matrix
}
\description{
Function to convert a single-cell RNAseq matrix into a format
 that can be used with \code{\link{runDWLSDeconv}}.
}
\examples{
sign_gene <- c("Bcl11b", "Lmo1", "F3", "Cnih3", "Ppp1r3c", "Rims2", "Gfap",
"Gjc3", "Chrna4", "Prkcd", "Prr18", "Grb14", "Tprn", "Clic1", "Olig2",
"Hrh3", "Tmbim1", "Carhsp1", "Tmem88b", "Ugt8a", "Arpp19", "Lamp5",
"Galnt6", "Hlf", "Hs3st2", "Tbr1", "Myl4", "Cygb", "Ttc9b","Ipcef1")

sign_matrix <- matrix(rnorm(length(sign_gene)*3), nrow = length(sign_gene))
rownames(sign_matrix) <- sign_gene
colnames(sign_matrix) <- c("cell_type1", "cell_type2", "cell_type3")

makeSignMatrixDWLSfromMatrix(matrix = sign_matrix, sign_gene = sign_gene,
cell_type_vector = c("cell_type1", "cell_type2", "cell_type3"))
}
\seealso{
\code{\link{runDWLSDeconv}}
}
