% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_enrichment.R
\name{makeSignMatrixRank}
\alias{makeSignMatrixRank}
\title{makeSignMatrixRank}
\usage{
makeSignMatrixRank(sc_matrix, sc_cluster_ids, gobject = NULL)
}
\arguments{
\item{sign_names}{vector with names for each provided gene signature}

\item{sign_list}{list of genes (signature)}
}
\value{
matrix
}
\description{
Function to convert a single-cell count matrix
and a corresponding single-cell cluster vector into
a rank matrix that can be used with the Rank enrichment option.
}
\examples{
    makeSignMatrixRank()
}
\seealso{
\code{\link{rankEnrich}}
}
