% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_help.R
\name{my_rowMeans}
\alias{my_rowMeans}
\title{my_rowMeans}
\usage{
my_rowMeans(x, method = c("arithmic", "geometric"), offset = 0.1)
}
\description{
arithmic or geometric rowMeans that works for a single column
}
\keyword{internal}
