% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\docType{class}
\name{NNClusParam-class}
\alias{NNClusParam-class}
\alias{NNClusParam}
\title{NNClusParam}
\description{
Virtual subclass of \code{BlusterParam} that designates network clustering
methods which are supported via Giotto-specific methods instead of directly
via \pkg{bluster}.
}
\keyword{internal}
