% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{norm_default}
\alias{norm_default}
\title{Default Giotto Normalization}
\description{
Expression matrix normalization method.

Steps:
\enumerate{
\item \link[=norm_library]{Total library size} normalization and scaling by
a custom scale-factor.
\item \link[=norm_log]{Log} transformation of data.
}
}
\section{params}{


\tabular{ll}{
\code{library_size_norm} \tab logical (default = \code{TRUE}). whether to perform
library size normalization \cr
\code{scalefactor} \tab numeric (default = 6000). Scalefactor to use after
library size normalization. (skipped if \code{library_size_norm = FALSE}) \cr
\code{log_norm} \tab logical (default = \code{TRUE}). Whether to transform values to
log-scale. \cr
\code{log_offset} \tab numeric (default = 1). If \code{log_norm = TRUE}, offset
value to add to expression values to avoid \code{log(0)} \cr
\code{logbase} \tab numeric (default = 2). If \code{log_norm = TRUE}, log base to
use to log normalize expression values
}
}

\seealso{
\link{process_param}

Other normalization parameters: 
\code{\link{norm_l2}},
\code{\link{norm_library}},
\code{\link{norm_log}},
\code{\link{norm_osmfish}},
\code{\link{norm_pearson}},
\code{\link{norm_quantile}},
\code{\link{norm_tfidf}}
}
\concept{normalization parameters}
