% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{norm_log}
\alias{norm_log}
\title{Log Normalization}
\description{
Apply a log normalization

\deqn{\LARGE
x'_{i,j} = \frac{\log(x_{i,j} + b)}{\log(a)}
}
Where:
\itemize{
\item (\eqn{x_{i,j}}) is the raw count for feature \eqn{i} in sample \eqn{j}
\item (\eqn{x'_{i,j}}) is the log normalized expression value for feature
\eqn{i} in sample \eqn{j}
\item (\eqn{a}) is the log base
\item (\eqn{b}) is an offset value
}
}
\section{params}{


\tabular{ll}{
\code{base} \tab numeric (default = 2) log base to use. Expressed as \eqn{a} in
the above equation. \cr
\code{offset} \tab numeric (default = 1). Offset to add to expression values to
avoid \eqn{\log(0)}. Expressed as \eqn{b} in the above equation.
}
}

\seealso{
\link{process_param}

Other normalization parameters: 
\code{\link{norm_default}},
\code{\link{norm_l2}},
\code{\link{norm_library}},
\code{\link{norm_osmfish}},
\code{\link{norm_pearson}},
\code{\link{norm_quantile}},
\code{\link{norm_tfidf}}
}
\concept{normalization parameters}
