% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{norm_osmfish}
\alias{norm_osmfish}
\title{osmFISH Normalization}
\value{
normalized object
}
\description{
Normalization method as provided by the osmFISH paper

Steps:
\enumerate{
\item First normalize genes, for each gene divide the counts by the total gene
count and multiply by the total number of genes.
\item Next normalize cells, for each cell divide the normalized gene counts by
the total counts per cell and multiply by the total number of cells.
}

\deqn{\LARGE
x'_{i,j} = \frac{x_{i,j}}{\sum_j x_{i,j}} \times n_{\text{features}}
}

\deqn{\LARGE
x''_{i,j} = \frac{x'_{i,j}}{\sum_i x'_{i,j}} \times n_{\text{samples}}
}

Where:
\itemize{
\item (\eqn{x_{i,j}}) is the raw count for feature \eqn{i} in sample \eqn{j}
\item (\eqn{x'_{i,j}}) is the feature normalized expression value
\item (\eqn{x''_{i,j}}) is the final normalized expression value after both
feature and cell normalization
\item (\eqn{n_{\text{samples}}}) is the total number of cells
(columns in matrix)
\item (\eqn{n_{\text{features}}}) is the total number of cells
(rows in matrix)
}
}
\section{params}{

None
}

\seealso{
\link{process_param}

Other normalization parameters: 
\code{\link{norm_arcsinh}},
\code{\link{norm_default}},
\code{\link{norm_l2}},
\code{\link{norm_library}},
\code{\link{norm_log}},
\code{\link{norm_pearson}},
\code{\link{norm_quantile}},
\code{\link{norm_tfidf}}
}
\concept{normalization parameters}
