% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{norm_tfidf}
\alias{norm_tfidf}
\title{TF-IDF Normalization}
\description{
TF-IDF (Term Frequency-Inverse Document Frequency) normalization is borrowed
from natural language processing to identify features that are highly expressed
in specific samples but not widely expressed across the entire dataset.

\deqn{\LARGE
TF_{i,j} = \frac{x_{i,j}}{\sum_{i} x_{i,j}}
}

\deqn{\LARGE
IDF_{i} = \log(1 + \frac{n_{samples}}{1 + n_{samples \: where \: feature \: i > 0}})
}

\deqn{\LARGE
TFIDF_{i,j} = TF_{i,j} \times IDF_{i}
}

Where:
\itemize{
\item (\eqn{x_{i,j}}) is the raw count for feature \eqn{i} in sample \eqn{j}
\item (\eqn{TF_{i,j}}) is the term frequency of feature \eqn{i} in sample \eqn{j}
\item (\eqn{IDF_{i}}) is the inverse document frequency of feature \eqn{i}
\item (\eqn{TFIDF_{i,j}}) is the final TF-IDF normalized value
}
}
\section{Note}{
\link[=norm_l2]{L2} normalization is commonly performed after TF-IDF normalization
}

\section{params}{

None
}

\seealso{
\link{process_param}

Other normalization parameters: 
\code{\link{norm_default}},
\code{\link{norm_l2}},
\code{\link{norm_library}},
\code{\link{norm_log}},
\code{\link{norm_osmfish}},
\code{\link{norm_pearson}},
\code{\link{norm_quantile}}
}
\concept{normalization parameters}
