% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalizeGiotto}
\alias{normalizeGiotto}
\title{normalizeGiotto}
\usage{
normalizeGiotto(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  expression_values = "raw",
  norm_methods = c("standard", "pearson_resid", "osmFISH", "quantile"),
  library_size_norm = TRUE,
  scalefactor = 6000,
  log_norm = TRUE,
  log_offset = 1,
  logbase = 2,
  scale_feats = TRUE,
  scale_genes = NULL,
  scale_cells = TRUE,
  scale_order = c("first_feats", "first_cells"),
  theta = 100,
  update_slot = "scaled",
  verbose = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit}

\item{feat_type}{feature type}

\item{expression_values}{expression values to use}

\item{norm_methods}{normalization method to use}

\item{library_size_norm}{normalize cells by library size}

\item{scalefactor}{scale factor to use after library size normalization}

\item{log_norm}{transform values to log-scale}

\item{log_offset}{offset value to add to expression matrix, default = 1}

\item{logbase}{log base to use to log normalize expression values}

\item{scale_feats}{z-score genes over all cells}

\item{scale_genes}{deprecated, use scale_feats}

\item{scale_cells}{z-score cells over all genes}

\item{scale_order}{order to scale feats and cells}

\item{theta}{theta parameter for the pearson residual normalization step}

\item{update_slot}{slot or name to use for the results from osmFISH and
pearson residual normalization}

\item{verbose}{be verbose}
}
\value{
giotto object
}
\description{
fast normalize and/or scale expresion values of Giotto object
}
\details{
Currently there are two 'methods' to normalize your raw counts data.

A. The standard method follows the standard protocol which can be adjusted
using the provided parameters and follows the following order: \cr
\itemize{
  \item{1. Data normalization for total library size and scaling by a custom 
  scale-factor.}
  \item{2. Log transformation of data.}
  \item{3. Z-scoring of data by genes and/or cells.}
}
B. The normalization method as provided by the osmFISH paper is also 
implemented: \cr
\itemize{
  \item{1. First normalize genes, for each gene divide the counts by the 
  total gene count and multiply by the total number of genes.}
  \item{2. Next normalize cells, for each cell divide the normalized gene 
  counts by the total counts per cell and multiply by the total number of 
  cells.}
}
C. The normalization method as provided by Lause/Kobak et al is also 
implemented: \cr
\itemize{
  \item{1. First calculate expected values based on Pearson correlations.}
  \item{2. Next calculate z-scores based on observed and expected values.}
}
D. Quantile normalization across features
\itemize{
  \item{1. Rank feature expression}
  \item{2. Define a common distribution by sorting expression values per 
  feature then finding the mean across all features per index}
  \item{3. Apply common distribution to expression information by using 
  the ranks from step 1 as indices}
}
By default the latter two results will be saved in the Giotto slot for
scaled expression, this can be changed by changing the update_slot parameters
}
\examples{
g <- GiottoData::loadGiottoMini("visium")

normalizeGiotto(g) # default is method A
}
