% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{normalizeGiotto}
\alias{normalizeGiotto}
\title{normalizeGiotto}
\usage{
normalizeGiotto(gobject, norm_methods = c("standard", "osmFISH"),
  library_size_norm = TRUE, scalefactor = 6000, log_norm = TRUE,
  logbase = 2, scale_genes = T, scale_cells = T,
  scale_order = c("first_genes", "first_cells"), verbose = F)
}
\arguments{
\item{gobject}{giotto object}

\item{norm_methods}{normalization method to use}

\item{library_size_norm}{normalize cells by library size}

\item{scalefactor}{scale factor to use after library size normalization}

\item{log_norm}{transform values to log-scale}

\item{logbase}{log base to use to log normalize expression values}

\item{scale_genes}{z-score genes over all cells}

\item{scale_cells}{z-score cells over all genes}

\item{scale_order}{order to scale genes and cells}

\item{verbose}{be verbose}
}
\value{
giotto object
}
\description{
normalize and/or scale expresion values of Giotto object
}
\details{
Currently there are two 'methods' to normalize your raw counts data.

A. The standard method follows the standard protocol which can be adjusted using
the provided parameters and follows the following order: \n
1. First the data can be normalized according to total library size and subsequently
scaled by a provided scale-factor. \n
2. The data can then be transformed to a log-scale if required. \n
3. Scaling (z-score) of genes and cells can be performed together or separately.

B. The normalization method as provided by the osmFISH paper is also implemented:
1. First normalize genes, for each gene divide the counts by the total gene count and
multiply by the total number of genes. \n
2. Next normalize cells, for each cell divide the normalized gene counts by the total
counts per cell and multiply by the total number of cells. \n
This data will be saved in the Giotto slot for custom expression.
}
\examples{
    normalizeGiotto(gobject)
}
