% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_hmrf.R
\name{numPts_below_line}
\alias{numPts_below_line}
\title{numPts_below_line}
\usage{
numPts_below_line(myVector, slope, x)
}
\arguments{
\item{myVector}{TBD}

\item{slope}{TBD}

\item{x}{TBD}
}
\description{
numPts_below_line
}
\keyword{internal}
