% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{objName-generic}
\alias{objName-generic}
\alias{objName}
\alias{objName<-}
\alias{objName,nameData-method}
\alias{objName<-,nameData-method}
\title{Giotto object name information}
\usage{
objName(x)

\S4method{objName}{nameData}(x)

\S4method{objName}{nameData}(x) <- value
}
\arguments{
\item{x}{a Giotto S4 class subobject with name data}

\item{value}{value to set as object name}
}
\description{
access and set name slot fo S4 subobject
}
\section{Functions}{
\itemize{
\item \code{objName(nameData)}: Get name information

\item \code{objName(nameData) <- value}: Set name information

}}
