% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_enrichment.R
\name{optimize_solveDampenedWLS}
\alias{optimize_solveDampenedWLS}
\title{optimize_solveDampenedWLS}
\usage{
optimize_solveDampenedWLS(S, B, constant_J)
}
\description{
Rui to fill in
}
\keyword{internal}
