% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_enrichment.R
\name{PAGEEnrich}
\alias{PAGEEnrich}
\title{PAGEEnrich}
\usage{
PAGEEnrich(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=runPAGEEnrich]{runPAGEEnrich}}
  \describe{
    \item{\code{gobject}}{Giotto object}
    \item{\code{sign_matrix}}{Matrix of signature genes for each cell type / process}
    \item{\code{expression_values}}{expression values to use}
    \item{\code{min_overlap_genes}}{minimum number of overlapping genes in sign_matrix required to calculate enrichment}
    \item{\code{reverse_log_scale}}{reverse expression values from log scale}
    \item{\code{logbase}}{log base to use if reverse_log_scale = TRUE}
    \item{\code{output_enrichment}}{how to return enrichment output}
    \item{\code{p_value}}{calculate p-values (boolean, default = FALSE)}
    \item{\code{include_depletion}}{calculate both enrichment and depletion}
    \item{\code{n_times}}{number of permutations to calculate for p_value}
    \item{\code{max_block}}{number of lines to process together (default = 20e6)}
    \item{\code{name}}{to give to spatial enrichment results, default = PAGE}
    \item{\code{verbose}}{be verbose}
    \item{\code{return_gobject}}{return giotto object}
  }}
}
\description{
Function to calculate gene signature enrichment scores per spatial position using PAGE.
}
\seealso{
\code{\link{runPAGEEnrich}}
}
