% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{pDataDT}
\alias{pDataDT}
\title{pDataDT}
\usage{
pDataDT(gobject, feat_type = NULL)
}
\arguments{
\item{gobject}{giotto object}

\item{feat_type}{feature type}
}
\value{
data.table with cell metadata
}
\description{
show cell metadata
}
\examples{

data(mini_giotto_single_cell) # loads existing Giotto object
pDataDT(mini_giotto_single_cell)

}
