% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot-generic}
\alias{plot-generic}
\alias{plot}
\alias{plot,giottoImage,missing-method}
\alias{plot,giottoLargeImage,missing-method}
\alias{plot,giottoPolygon,missing-method}
\alias{plot,giottoPoints,missing-method}
\alias{plot,spatLocsObj,missing-method}
\title{Preview a Giotto spatial object}
\usage{
\S4method{plot}{giottoImage,missing}(x, y, ...)

\S4method{plot}{giottoLargeImage,missing}(x, y, ...)

\S4method{plot}{giottoPolygon,missing}(x, y, point_size = 0.1, type = c("poly", "centroid"), ...)

\S4method{plot}{giottoPoints,missing}(x, y, point_size = 0.1, feats = NULL, ...)

\S4method{plot}{spatLocsObj,missing}(x, y, ...)
}
\arguments{
\item{x}{giotto image, giottoPolygon, or giottoPoints object}

\item{y}{Not used.}

\item{\dots}{additional parameters to pass}

\item{point_size}{size of points when plotting giottoPoints}

\item{type}{what to plot: either 'poly' (default) or polygon 'centroid'}

\item{feats}{specific features to plot within giottoPoints object (defaults to NULL, meaning all available features)}
}
\description{
S4 generic for previewing Giotto's image and subcellular objects.
}
\section{Functions}{
\itemize{
\item \code{plot(x = giottoImage, y = missing)}: Plot \emph{magick}-based giottoImage object. ... param passes to \code{\link{plot_giottoImage_MG}}

\item \code{plot(x = giottoLargeImage, y = missing)}: Plot \emph{terra}-based giottoLargeImage object. ... param passes to \code{\link{plot_giottoLargeImage}}

\item \code{plot(x = giottoPolygon, y = missing)}: Plot \emph{terra}-based giottoPolygon object. ... param passes to \code{\link[terra]{plot}}

\item \code{plot(x = giottoPoints, y = missing)}: \emph{terra}-based giottoPoint object. ... param passes to \code{\link[terra]{plot}}

\item \code{plot(x = spatLocsObj, y = missing)}: Plot a spatLocsObj

}}
\concept{plot}
