% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visuals.R
\name{plot_auto_largeImage_resample}
\alias{plot_auto_largeImage_resample}
\title{Optimized largeImage resampling}
\usage{
plot_auto_largeImage_resample(
  gobject,
  giottoLargeImage = NULL,
  largeImage_name = NULL,
  spat_unit = NULL,
  spat_loc_name = NULL,
  include_image_in_border = TRUE,
  flex_resample = TRUE,
  max_crop = 1e+08,
  max_resample_scale = 100
)
}
\arguments{
\item{gobject}{\code{gobject} containing \code{largeImage} object}

\item{giottoLargeImage}{\code{largeImage} object to resample if not provided through
\code{gobject} and \code{largeImage_name}}

\item{largeImage_name}{name of \code{largeImage} in \code{gobject}}

\item{spat_unit}{spatial unit}

\item{spat_loc_name}{name of spatial locations to plot}

\item{include_image_in_border}{[boolean] expand the extent sampled to also show image in 
border regions not included in spatlocs. This prevents images in plots from
being sharply cut off around the furthest spatial locations. (default is \code{TRUE})}

\item{flex_resample}{[boolean] Whether to allow automatic selection of sampling
workflow as defined in details sections. (default is \code{TRUE})}

\item{max_crop}{maximum crop size allowed for \strong{method A} before switching to
\strong{method B} (see description)}

\item{max_resample_scale}{maximum cells allowed to resample to compensate for
decreased resolution when cropping after sampling}
}
\value{
a \code{giottoLargeImage} cropped and resampled properly for plotting
}
\description{
Downsample \code{largeImage} for plotting. Uses \code{\link[terra]{spatSample}}
to load only a portion of the original image, speeding up plotting and lowering memory
footprint.

Default behavior of \code{spatSample} is to crop if only a smaller ROI is needed for plotting
followed by the sampling process in order to reduce wasted sampling by focusing the
sample space. For very large ROIs, this crop can be time intensive and require
writing to disk.

This function examines the ROI dimensions as defined through the limits of the spatial 
locations to be plotted, and decides between the following two methods in order to
avoid this issue:
\itemize{
  \item{\strong{Method A.} First crop original image and then sample n values where
  n = 500,000 to generate final image}
  \item{\strong{Method B.} First oversample n values and then crop, where n = 500,000
  scaled by a value >1. Scaling factor increases the smaller the ROI is and
  is defined by: original dimensions/crop dimensions where the larger ratio between
  x and y dims is chosen. Scale factor is capped by \code{max_resample_scale}}
}
Control points for this function are set by \code{max_crop} which decides the max
ROI area after which switchover to method B happens in order to avoid laborious crops 
and \code{max_resample_scale} which determines the maximum scale factor for number
of values to sample. Both values can be adjusted depending on system resources.
Additionally, \code{flex_resample} determines if this switching behavior happens.
When set to \code{FALSE}, only method A is used.
}
\seealso{
\code{\link[terra]{spatSample}}
}
\keyword{internal}
