% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_in_situ_visuals.R
\name{plot_feature_hexbin_layer}
\alias{plot_feature_hexbin_layer}
\title{plot_feature_hexbin_layer}
\usage{
plot_feature_hexbin_layer(
  ggobject = NULL,
  spatial_feat_info,
  sel_feat,
  sdimx = "x",
  sdimy = "y",
  bins = 10,
  alpha = 0.5
)
}
\value{
ggplot
}
\description{
low level function to plot hexbins at the spatial in situ level
}
\details{
This function can plot one feature for one modality.
}
\keyword{internal}
