% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{plot_giottoLargeImage}
\alias{plot_giottoLargeImage}
\title{plot_giottoLargeImage}
\usage{
plot_giottoLargeImage(
  gobject = NULL,
  largeImage_name = NULL,
  giottoLargeImage = NULL,
  crop_extent = NULL,
  xmax_crop = NULL,
  xmin_crop = NULL,
  ymax_crop = NULL,
  ymin_crop = NULL,
  max_intensity = NULL
)
}
\arguments{
\item{gobject}{giotto object}

\item{largeImage_name}{name of giottoLargeImage}

\item{giottoLargeImage}{giottoLargeImage object}

\item{crop_extent}{(optional) extent object to focus on specific region of image}

\item{xmax_crop, xmin_crop, ymax_crop, ymin_crop}{(optional) crop min/max x and y bounds}

\item{max_intensity}{(optional) value to treat as maximum intensity in color scale}
}
\value{
plot
}
\description{
Plot a \emph{downsampled} version of giottoLargeImage. Cropping can increase plot resolution of region of interest.
}
\keyword{internal}
