% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd.R
\name{plot_params}
\alias{plot_params}
\title{Params documentation template: plot_params}
\arguments{
\item{group_by}{character. Create multiple plots based on cell annotation 
column}

\item{group_by_subset}{character. subset the group_by factor column}

\item{gradient_midpoint}{numeric. midpoint for color gradient}

\item{gradient_style}{either 'divergent' (midpoint is used in color scaling)
or 'sequential' (scaled based on data range)}

\item{gradient_limits}{numeric vector with lower and upper limits}

\item{gradient_color}{character. continuous colors to use. palette to
use or vector of colors to use (minimum of 2).}

\item{select_cell_groups}{select subset of cells/clusters based on 
cell_color parameter}

\item{select_cells}{select subset of cells based on cell IDs}

\item{show_other_cells}{display not selected cells}

\item{other_cell_color}{color for not selected cells}

\item{other_cell_alpha}{(0 to 1) alpha for not selected cells}

\item{other_cells_alpha}{(0 to 1) alpha for not selected cells}

\item{other_point_size}{point size for not selected cells}

\item{show_cluster_center}{plot center of selected clusters}

\item{show_center_label}{plot label of selected clusters}

\item{center_point_size}{size of center points}

\item{center_point_border_col}{border color of center points}

\item{center_point_border_stroke}{border stroke size of center points}

\item{label_size}{size of labels}

\item{label_fontface}{font of labels}

\item{edge_alpha}{column to use for alpha of the edges}

\item{point_shape}{point with border or not (border or no_border)}

\item{point_size}{size of point (cell)}

\item{point_alpha}{transparency of points}

\item{point_border_col}{color of border around points}

\item{point_border_stroke}{stroke size of border around points}

\item{title}{character. title for plot, defaults to cell_color parameter}

\item{show_legend}{logical. show legend}

\item{legend_text}{size of legend text}

\item{legend_symbol_size}{size of legend symbols}

\item{background_color}{color of plot background}

\item{axis_text}{size of axis text}

\item{axis_title}{size of axis title}
}
\value{
plot
}
\description{
Params documentation template: plot_params
}
\keyword{internal}
