% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visuals.R
\name{plot_spat_image_layer_ggplot}
\alias{plot_spat_image_layer_ggplot}
\title{plot_spat_image_layer_ggplot}
\usage{
plot_spat_image_layer_ggplot(
  ggplot,
  gobject,
  gimage,
  sdimx = NULL,
  sdimy = NULL
)
}
\arguments{
\item{gobject}{giotto object}

\item{gimage}{a giotto image}

\item{sdimx}{x-axis dimension name (default = 'sdimx')}

\item{sdimy}{y-axis dimension name (default = 'sdimy')}
}
\value{
ggplot
}
\description{
create background image in ggplot
}
\keyword{internal}
