% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction_visuals.R
\name{plotCCcomHeatmap}
\alias{plotCCcomHeatmap}
\title{plotCCcomHeatmap}
\usage{
plotCCcomHeatmap(
  gobject,
  comScores,
  selected_LR = NULL,
  selected_cell_LR = NULL,
  show_LR_names = TRUE,
  show_cell_LR_names = TRUE,
  show = c("PI", "LR_expr", "log2fc"),
  cor_method = c("pearson", "kendall", "spearman"),
  aggl_method = c("ward.D", "ward.D2", "single", "complete", "average", "mcquitty",
    "median", "centroid"),
  show_plot = NA,
  return_plot = NA,
  save_plot = NA,
  save_param = list(),
  default_save_name = "plotCCcomHeatmap"
)
}
\arguments{
\item{gobject}{giotto object}

\item{comScores}{communinication scores from \code{\link{exprCellCellcom}} or \code{\link{spatCellCellcom}}}

\item{selected_LR}{selected ligand-receptor combinations}

\item{selected_cell_LR}{selected cell-cell combinations for ligand-receptor combinations}

\item{show_LR_names}{show ligand-receptor names}

\item{show_cell_LR_names}{show cell-cell names}

\item{show}{values to show on heatmap}

\item{cor_method}{correlation method used for clustering}

\item{aggl_method}{agglomeration method used by hclust}

\item{show_plot}{show plots}

\item{return_plot}{return plotting object}

\item{save_plot}{directly save the plot [boolean]}

\item{save_param}{list of saving parameters from \code{\link{all_plots_save_function}}}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}
}
\value{
ggplot
}
\description{
Plots heatmap for ligand-receptor communication scores in cell-cell interactions
}
