% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{plotGiottoImage}
\alias{plotGiottoImage}
\title{Plot a giotto image object}
\usage{
plotGiottoImage(
  gobject = NULL,
  image_name = NULL,
  image_type = NULL,
  giottoImage = NULL,
  giottoLargeImage = NULL,
  largeImage_crop_params_list = NULL,
  largeImage_max_intensity = NULL
)
}
\arguments{
\item{gobject}{gobject containing giotto image object}

\item{image_name}{name of giotto image object}

\item{image_type}{type of giotto image object to plot}

\item{giottoImage}{giottoImage object to plot directly}

\item{giottoLargeImage}{giottoLargeImage object to plot directly}

\item{largeImage_crop_params_list}{(optional) named list of params for focusing
on a specified region of a giottoLargeImage.}

\item{largeImage_max_intensity}{(optional) assign override value to treat as
maximum intensity in color scale when plotting giottoLargeImage}
}
\description{
Display a giotto image in the viewer panel. Image object to plot
  can be specified by providing the giotto object containing the image (\code{gobject}),
  the image object name (\code{image_name}), and the image object type (\code{image_type}).
  Alternatively, image objects can be directly plotted through their respective
  associated params.
}
\section{largeImage-specific additional params}{

  \code{largeImage_crop_params_list} accepts a named list of the following
    possible params to define a region of interest (ROI) to plot through either
    a terra extent object OR x and y min and max bounds given as numerics:
  \itemize{
    \item{\code{crop_extent} -- terra extent object to define crop ROI}
    \item{\code{xmax_crop} -- x max of ROI}
    \item{\code{xmin_crop} -- x min of ROI}
    \item{\code{ymax_crop} -- y max of ROI}
    \item{\code{ymin_crop} -- y min of ROI}
  }
  \code{largeImage_max_intensity} accepts a numeric value to set the max
    value in the plotting color scale. Can be used in case there are high
    outlier intensity values in the image and a preview with alternative 
    color scaling is desired.
}

\seealso{
Other basic image functions: 
\code{\link{addGiottoImage}()},
\code{\link{reconnectGiottoImage}()},
\code{\link{updateGiottoImage}()}
}
\concept{basic image functions}
