% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{plotGiottoImage}
\alias{plotGiottoImage}
\title{plotGiottoImage}
\usage{
plotGiottoImage(gobject, image_name)
}
\arguments{
\item{gobject}{giotto object}

\item{image_name}{name of giotto image \code{\link{showGiottoImageNames}}}
}
\value{
plot
}
\description{
get plot a giotto image from a giotto object
}
