% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{plotGiottoLargeImage}
\alias{plotGiottoLargeImage}
\title{plotGiottoLargeImage}
\usage{
plotGiottoLargeImage(
  gobject = NULL,
  largeImage_name = NULL,
  giottoLargeImage = NULL,
  crop_extent = NULL,
  xmax_crop = NULL,
  xmin_crop = NULL,
  ymax_crop = NULL,
  ymin_crop = NULL,
  max_intensity = NULL
)
}
\arguments{
\item{gobject}{giotto object}

\item{largeImage_name}{name of giottoLargeImage}

\item{max_intensity}{value to treat as maximum intensity in color scale}

\item{zoom_extent}{extent object to focus on specific region of image}

\item{xmax_zoom}{assign zoom boundary}

\item{xmin_zoom}{assign zoom boundary}

\item{ymax_zoom}{assign zoom boundary}

\item{ymin_zoom}{assign zoom boundary}
}
\value{
plot
}
\description{
plot a downsampled version of giottoLargeImage
}
