% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_visuals.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{plotHeatmap}
\usage{
plotHeatmap(gobject, expression_values = c("normalized", "scaled",
  "custom"), genes, cluster_column = NULL, cluster_order = c("size",
  "correlation", "custom"), cluster_custom_order = NULL,
  cluster_color_code = NULL, cluster_cor_method = "pearson",
  cluster_hclust_method = "ward.D", gene_order = c("custom",
  "correlation"), gene_custom_order = NULL,
  gene_cor_method = "pearson", gene_hclust_method = "complete",
  show_values = c("rescaled", "z-scaled", "original"),
  size_vertical_lines = 1.1, gradient_colors = c("blue", "yellow",
  "red"), gene_label_selection = NULL, axis_text_y_size = NULL,
  legend_nrows = 1, show_plot = NA, return_plot = NA,
  save_plot = NA, save_param = list(),
  default_save_name = "plotHeatmap")
}
\arguments{
\item{gobject}{giotto object}

\item{expression_values}{expression values to use}

\item{genes}{genes to use}

\item{cluster_column}{name of column to use for clusters}

\item{cluster_order}{method to determine cluster order}

\item{cluster_custom_order}{custom order for clusters}

\item{cluster_color_code}{color code for clusters}

\item{cluster_cor_method}{method for cluster correlation}

\item{cluster_hclust_method}{method for hierarchical clustering of clusters}

\item{gene_order}{method to determine gene order}

\item{gene_custom_order}{custom order for genes}

\item{gene_cor_method}{method for gene correlation}

\item{gene_hclust_method}{method for hierarchical clustering of genes}

\item{show_values}{which values to show on heatmap}

\item{size_vertical_lines}{sizes for vertical lines}

\item{gradient_colors}{colors for heatmap gradient}

\item{gene_label_selection}{subset of genes to show on y-axis}

\item{axis_text_y_size}{size for y-axis text}

\item{legend_nrows}{number of rows for the cluster legend}

\item{show_plot}{show plot}

\item{return_plot}{return ggplot object}

\item{save_plot}{directly save the plot [boolean]}

\item{save_param}{list of saving parameters from all_plots_save_function()}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}
}
\value{
ggplot
}
\description{
creates order for clusters
}
\details{
Creates heatmap for genes and clusters.
}
\examples{
    plotHeatmap(gobject)
}
