% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction_visuals.R
\name{plotInteractionChangedFeats}
\alias{plotInteractionChangedFeats}
\title{Plot interaction changed features}
\usage{
plotInteractionChangedFeats(
  gobject,
  icfObject,
  source_type,
  source_markers,
  ICF_feats,
  cell_color_code = NULL,
  show_plot = NULL,
  return_plot = NULL,
  save_plot = NULL,
  save_param = list(),
  default_save_name = "plotInteractionChangedFeats"
)
}
\arguments{
\item{gobject}{giotto object}

\item{icfObject}{ICF (interaction changed feature) score object}

\item{source_type}{cell type of the source cell}

\item{source_markers}{markers for the source cell type}

\item{ICF_feats}{named character vector of ICF features}

\item{cell_color_code}{character. discrete colors to use. palette to use or
named vector of colors}

\item{show_plot}{logical. show plot}

\item{return_plot}{logical. return ggplot object}

\item{save_plot}{logical. save the plot}

\item{save_param}{list of saving parameters, see \code{\link{showSaveParameters}}}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}
}
\value{
plot
}
\description{
Create barplot to visualize interaction changed features
}
