% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_visuals.R
\name{plotly_grid}
\alias{plotly_grid}
\title{plotly_grid}
\usage{
plotly_grid(
  spatial_grid,
  x_start = "x_start",
  y_start = "y_start",
  x_end = "x_end",
  y_end = "y_end"
)
}
\arguments{
\item{spatial_grid}{spatial_grid in giotto object}
}
\value{
edges in spatial grid as data.table()
}
\description{
provide grid segment to draw in plot_ly()
}
\keyword{internal}
