% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_visuals.R
\name{plotly_network}
\alias{plotly_network}
\title{plotly_network}
\usage{
plotly_network(
  network,
  x = "sdimx_begin",
  y = "sdimy_begin",
  z = "sdimz_begin",
  x_end = "sdimx_end",
  y_end = "sdimy_end",
  z_end = "sdimz_end"
)
}
\arguments{
\item{gobject}{network in giotto object}
}
\value{
edges in network as data.table()
}
\description{
provide network segment to draw in 3D plot_ly()
}
\keyword{internal}
