% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_visuals.R
\name{plotMetaDataCellsHeatmap}
\alias{plotMetaDataCellsHeatmap}
\title{plotMetaDataCellsHeatmap}
\usage{
plotMetaDataCellsHeatmap(
  gobject,
  metadata_cols = NULL,
  spat_enr_names = NULL,
  value_cols = NULL,
  first_meta_col = NULL,
  second_meta_col = NULL,
  show_values = c("zscores", "original", "zscores_rescaled"),
  custom_cluster_order = NULL,
  clus_cor_method = "pearson",
  clus_cluster_method = "complete",
  custom_values_order = NULL,
  values_cor_method = "pearson",
  values_cluster_method = "complete",
  midpoint = 0,
  x_text_size = 8,
  x_text_angle = 45,
  y_text_size = 8,
  strip_text_size = 8,
  show_plot = NA,
  return_plot = NA,
  save_plot = NA,
  save_param = list(),
  default_save_name = "plotMetaDataCellsHeatmap"
)
}
\arguments{
\item{gobject}{giotto object}

\item{metadata_cols}{annotation columns found in pDataDT(gobject)}

\item{spat_enr_names}{spatial enrichment results to include}

\item{value_cols}{value columns to use}

\item{first_meta_col}{if more than 1 metadata column, select the x-axis factor}

\item{second_meta_col}{if more than 1 metadata column, select the facetting factor}

\item{show_values}{which values to show on heatmap}

\item{custom_cluster_order}{custom cluster order (default = NULL)}

\item{clus_cor_method}{correlation method for clusters}

\item{clus_cluster_method}{hierarchical cluster method for the clusters}

\item{custom_values_order}{custom values order (default = NULL)}

\item{values_cor_method}{correlation method for values}

\item{values_cluster_method}{hierarchical cluster method for the values}

\item{midpoint}{midpoint of show_values}

\item{x_text_size}{size of x-axis text}

\item{x_text_angle}{angle of x-axis text}

\item{y_text_size}{size of y-axis text}

\item{strip_text_size}{size of strip text}

\item{show_plot}{show plot}

\item{return_plot}{return ggplot object}

\item{save_plot}{directly save the plot [boolean]}

\item{save_param}{list of saving parameters, see \code{\link{showSaveParameters}}}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}
}
\value{
ggplot or data.table
}
\description{
Creates heatmap for numeric cell metadata within aggregated clusters.
}
\details{
Creates heatmap for the average values of selected value columns in the different annotation groups.
}
\seealso{
\code{\link{plotMetaDataHeatmap}} for gene expression instead of numeric cell annotation data.
}
