% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_visuals.R
\name{plotMetaDataHeatmap}
\alias{plotMetaDataHeatmap}
\title{plotMetaDataHeatmap}
\usage{
plotMetaDataHeatmap(
  gobject,
  expression_values = c("normalized", "scaled", "custom"),
  metadata_cols = NULL,
  selected_genes = NULL,
  first_meta_col = NULL,
  second_meta_col = NULL,
  show_values = c("zscores", "original", "zscores_rescaled"),
  custom_cluster_order = NULL,
  clus_cor_method = "pearson",
  clus_cluster_method = "complete",
  custom_gene_order = NULL,
  gene_cor_method = "pearson",
  gene_cluster_method = "complete",
  midpoint = 0,
  x_text_size = 8,
  x_text_angle = 45,
  y_text_size = 8,
  strip_text_size = 8,
  show_plot = NA,
  return_plot = NA,
  save_plot = NA,
  save_param = list(),
  default_save_name = "plotMetaDataHeatmap"
)
}
\arguments{
\item{gobject}{giotto object}

\item{expression_values}{expression values to use}

\item{metadata_cols}{annotation columns found in pDataDT(gobject)}

\item{selected_genes}{subset of genes to use}

\item{first_meta_col}{if more than 1 metadata column, select the x-axis factor}

\item{second_meta_col}{if more than 1 metadata column, select the facetting factor}

\item{show_values}{which values to show on heatmap}

\item{custom_cluster_order}{custom cluster order (default = NULL)}

\item{clus_cor_method}{correlation method for clusters}

\item{clus_cluster_method}{hierarchical cluster method for the clusters}

\item{custom_gene_order}{custom gene order (default = NULL)}

\item{gene_cor_method}{correlation method for genes}

\item{gene_cluster_method}{hierarchical cluster method for the genes}

\item{midpoint}{midpoint of show_values}

\item{x_text_size}{size of x-axis text}

\item{x_text_angle}{angle of x-axis text}

\item{y_text_size}{size of y-axis text}

\item{strip_text_size}{size of strip text}

\item{show_plot}{show plot}

\item{return_plot}{return ggplot object}

\item{save_plot}{directly save the plot [boolean]}

\item{save_param}{list of saving parameters from all_plots_save_function()}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}
}
\value{
ggplot or data.table
}
\description{
creates order for clusters
}
\details{
Creates heatmap for average the average expression of selected genes in the different annotation groups
}
\examples{
    plotMetaDataHeatmap(gobject)
}
