% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_visuals.R
\name{plotMetaDataHeatmap}
\alias{plotMetaDataHeatmap}
\title{plotMetaDataHeatmap}
\usage{
plotMetaDataHeatmap(
  gobject,
  expression_values = c("normalized", "scaled", "custom"),
  metadata_cols = NULL,
  selected_genes = NULL,
  first_meta_col = NULL,
  second_meta_col = NULL,
  show_values = c("zscores", "original", "zscores_rescaled"),
  custom_cluster_order = NULL,
  clus_cor_method = "pearson",
  clus_cluster_method = "complete",
  custom_gene_order = NULL,
  gene_cor_method = "pearson",
  gene_cluster_method = "complete",
  gradient_color = c("blue", "white", "red"),
  gradient_midpoint = 0,
  gradient_limits = NULL,
  x_text_size = 10,
  x_text_angle = 45,
  y_text_size = 10,
  strip_text_size = 8,
  show_plot = NA,
  return_plot = NA,
  save_plot = NA,
  save_param = list(),
  default_save_name = "plotMetaDataHeatmap"
)
}
\arguments{
\item{gobject}{giotto object}

\item{expression_values}{expression values to use}

\item{metadata_cols}{annotation columns found in pDataDT(gobject)}

\item{selected_genes}{subset of genes to use}

\item{first_meta_col}{if more than 1 metadata column, select the x-axis factor}

\item{second_meta_col}{if more than 1 metadata column, select the facetting factor}

\item{show_values}{which values to show on heatmap}

\item{custom_cluster_order}{custom cluster order (default = NULL)}

\item{clus_cor_method}{correlation method for clusters}

\item{clus_cluster_method}{hierarchical cluster method for the clusters}

\item{custom_gene_order}{custom gene order (default = NULL)}

\item{gene_cor_method}{correlation method for genes}

\item{gene_cluster_method}{hierarchical cluster method for the genes}

\item{gradient_color}{vector with 3 colors for numeric data}

\item{gradient_midpoint}{midpoint for color gradient}

\item{gradient_limits}{vector with lower and upper limits}

\item{x_text_size}{size of x-axis text}

\item{x_text_angle}{angle of x-axis text}

\item{y_text_size}{size of y-axis text}

\item{strip_text_size}{size of strip text}

\item{show_plot}{show plot}

\item{return_plot}{return ggplot object}

\item{save_plot}{directly save the plot [boolean]}

\item{save_param}{list of saving parameters, see \code{\link{showSaveParameters}}}

\item{default_save_name}{default save name}
}
\value{
ggplot or data.table
}
\description{
Creates heatmap for genes within aggregated clusters.
}
\details{
Creates heatmap for the average expression of selected genes in the different annotation/cluster groups.
Calculation of cluster or gene order is done on the provided expression values, but visualization
is by default on the z-scores. Other options are the original values or z-scores rescaled per gene (-1 to 1).
}
\examples{
\dontrun{

data(mini_giotto_single_cell)

# get all genes
all_genes = slot(mini_giotto_single_cell, 'gene_ID')

# look at cell metadata
cell_metadata = pDataDT(mini_giotto_single_cell)

# plot heatmap per cell type, a column name from cell_metadata
plotMetaDataHeatmap(mini_giotto_single_cell,
                    selected_genes = all_genes[1:10],
                    metadata_cols = 'cell_types')

}
}
\seealso{
\code{\link{plotMetaDataCellsHeatmap}} for numeric cell annotation instead of gene expression.
}
