% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction_visuals.R
\name{plotRankSpatvsExpr}
\alias{plotRankSpatvsExpr}
\title{plotRankSpatvsExpr}
\usage{
plotRankSpatvsExpr(
  gobject,
  combCC,
  expr_rnk_column = "LR_expr_rnk",
  spat_rnk_column = "LR_spat_rnk",
  midpoint = 10,
  size_range = c(0.01, 1.5),
  xlims = NULL,
  ylims = NULL,
  selected_ranks = c(1, 10, 20),
  show_plot = NA,
  return_plot = NA,
  save_plot = NA,
  save_param = list(),
  default_save_name = "plotRankSpatvsExpr"
)
}
\arguments{
\item{gobject}{giotto object}

\item{combCC}{combined communinication scores from \code{\link{combCCcom}}}

\item{expr_rnk_column}{column with expression rank information to use}

\item{spat_rnk_column}{column with spatial rank information to use}

\item{midpoint}{midpoint of colors}

\item{size_range}{size ranges of dotplot}

\item{xlims}{x-limits, numerical vector of 2}

\item{ylims}{y-limits, numerical vector of 2}

\item{selected_ranks}{numerical vector, will be used to print out the percentage of top spatial ranks are recovered}

\item{show_plot}{show plots}

\item{return_plot}{return plotting object}

\item{save_plot}{directly save the plot [boolean]}

\item{save_param}{list of saving parameters from \code{\link{all_plots_save_function}}}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}
}
\value{
ggplot
}
\description{
Plots dotplot to compare ligand-receptor rankings from spatial and expression information
}
