% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction_visuals.R
\name{plotRecovery}
\alias{plotRecovery}
\title{plotRecovery}
\usage{
plotRecovery(
  gobject,
  combCC,
  expr_rnk_column = "exprPI_rnk",
  spat_rnk_column = "spatPI_rnk",
  ground_truth = c("spatial", "expression"),
  show_plot = NULL,
  return_plot = NULL,
  save_plot = NULL,
  save_param = list(),
  default_save_name = "plotRecovery"
)
}
\arguments{
\item{gobject}{giotto object}

\item{combCC}{combined communication scores from \code{\link{combCCcom}}}

\item{expr_rnk_column}{column with expression rank information to use}

\item{spat_rnk_column}{column with spatial rank information to use}

\item{ground_truth}{what to consider as ground truth (default: spatial)}

\item{show_plot}{logical. show plot}

\item{return_plot}{logical. return ggplot object}

\item{save_plot}{logical. save the plot}

\item{save_param}{list of saving parameters, see \code{\link{showSaveParameters}}}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}
}
\value{
ggplot
}
\description{
Plots recovery plot to compare ligand-receptor rankings from 
spatial and expression information
}
\examples{
g <- GiottoData::loadGiottoMini("visium")

exprCC <- exprCellCellcom(g, cluster_column = "leiden_clus", 
feat_set_1 = "Gm19935", feat_set_2 = "9630013A20Rik")
spatialCC <- spatCellCellcom(gobject = g, cluster_column = "leiden_clus",
feat_set_1 = "Gm19935", feat_set_2 = "9630013A20Rik", verbose = "a lot",
random_iter = 10)

combCC <- combCCcom(spatialCC = spatialCC, exprCC = exprCC)

plotRecovery(gobject = g, combCC = combCC)
}
