% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_structures.R
\name{plotStatDelaunayNetwork}
\alias{plotStatDelaunayNetwork}
\title{plotStatDelaunayNetwork}
\usage{
plotStatDelaunayNetwork(
  gobject,
  method = c("deldir", "delaunayn_geometry", "RTriangle"),
  dimensions = "all",
  maximum_distance = "auto",
  minimum_k = 0,
  options = "Pp",
  Y = TRUE,
  j = TRUE,
  S = 0,
  show_plot = NA,
  return_plot = NA,
  save_plot = NA,
  save_param = list(),
  default_save_name = "plotStatDelaunayNetwork",
  ...
)
}
\arguments{
\item{gobject}{giotto object}

\item{method}{package to use to create a Delaunay network}

\item{dimensions}{which spatial dimensions to use (maximum 2 dimensions)}

\item{maximum_distance}{distance cuttof for Delaunay neighbors to consider}

\item{minimum_k}{minimum neigbhours if maximum_distance != NULL}

\item{options}{(geometry) String containing extra control options for the underlying Qhull command; see the Qhull documentation (../doc/qhull/html/qdelaun.html) for the available options. (default = 'Pp', do not report precision problems)}

\item{Y}{(RTriangle) If TRUE prohibits the insertion of Steiner points on the mesh boundary.}

\item{j}{(RTriangle) If TRUE jettisons vertices that are not part of the final triangulation from the output.}

\item{S}{(RTriangle) Specifies the maximum number of added Steiner points.}

\item{show_plot}{show plots}

\item{return_plot}{return ggplot object}

\item{save_plot}{directly save the plot [boolean]}

\item{save_param}{list of saving parameters, see \code{\link{showSaveParameters}}}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}

\item{\dots}{Other parameters}
}
\value{
giotto object with updated spatial network slot
}
\description{
Plots network statistics for a Delaunay network..
}
