% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto.R
\name{print.giotto}
\alias{print.giotto}
\title{Prints giotto object.}
\usage{
\method{print}{giotto}(object, nr_genes = 5, nr_cells = 5)
}
\arguments{
\item{object}{giotto object}

\item{nr_genes}{number of genes (rows) to print}

\item{nr_cells}{number of cells (columns) to print}
}
\description{
Prints giotto object
}
