% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{process_param}
\alias{process_param}
\alias{normParam}
\alias{scaleParam}
\alias{adjustParam}
\title{Data Processing Parameter Class Factories}
\usage{
normParam(method = "default", ...)

scaleParam(method = "default", ...)

adjustParam(method = "limma", ...)
}
\arguments{
\item{method}{character. Name of method to use. See details.}

\item{\dots}{(optional) Additional named parameters relevant to the param
class.}
}
\description{
Data processing operations in Giotto Suite can be divided into
normalization, scaling, and adjustments
}
\section{normParam methods}{

\itemize{
\item \code{\link[=norm_default]{"default"}} - default Giotto normalizations steps
(library + log norms)
\item \code{\link[=norm_library]{"library"}} - library normalization
\item \code{\link[=norm_log]{"log"}} - log normalization
\item \code{\link[=norm_osmfish]{"osmfish"}} - osmfish normalization method
\item \code{\link[=norm_pearson]{"pearson"}} - Lause/Kobak 2020 pearson residuals
normalization
\item \code{\link[=norm_quantile]{"quantile"}} - quantile normalization
\item \code{\link[=norm_tfidf]{"tf-idf"}} - Term Frequency-Inverse Document Frequency
\item \code{\link[=norm_l2]{"l2"}} - L2 normalization (also known as Euclidean
normalization)
}
}

\section{scaleParam methods}{

\itemize{
\item \code{\link[=scale_default]{"default"}} - default Giotto scaling steps (scale along
features then cells)
\item \code{\link[=scale_zscore]{"zscore"}} - essentially the same as \code{base::scale()}, but
with a \code{MARGIN} param allowing scaling long either cols or rows
}
}

\section{adjustParam methods}{

\itemize{
\item \code{"limma"} - limma batch correction
}
}

