% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\docType{class}
\name{process_param}
\alias{process_param}
\alias{defaultNormParam-class}
\alias{libraryNormParam-class}
\alias{logNormParam-class}
\alias{osmFISHNormParam-class}
\alias{pearsonResidNormParam-class}
\alias{quantileNormParam-class}
\alias{tfidfNormParam-class}
\alias{l2NormParam-class}
\alias{arcsinhNormParam-class}
\alias{defaultScaleParam-class}
\alias{zscoreScaleParam-class}
\alias{limmaAdjustParam-class}
\alias{binarizeThreshParam-class}
\alias{minmaxThreshParam-class}
\alias{normParam}
\alias{scaleParam}
\alias{adjustParam}
\alias{thresholdParam}
\title{Data Processing Parameter Classes}
\usage{
normParam(method = "default", ...)

scaleParam(method = "default", ...)

adjustParam(method = "limma", ...)

thresholdParam(method = "binarize", ...)
}
\arguments{
\item{method}{character. Name of method to use. See details.}

\item{\dots}{(optional) Additional named parameters relevant to the param
class.}
}
\description{
Data processing operations in Giotto Suite can be divided into
normalization, scaling, and adjustments. These operations can be selected
via the factory functions \code{normParam()}, \code{scaleParam()}, and \code{adjustParam()},
respectively.

Requested operations are generated as method-specific param classes that
contain all the parameters needed to perform them, editable through \verb{$<-}.
}
\details{
Generated params are S4 objects inheriting from \code{processParam} and one of
\code{normParam}, \code{scaleParam}, and \code{adjustParam}.
}
\section{normParam methods}{

\itemize{
\item \code{\link[=norm_default]{"default"}} - default Giotto normalizations steps
(library + log norms)
\item \code{\link[=norm_library]{"library"}} - library normalization
\item \code{\link[=norm_log]{"log"}} - log normalization
\item \code{\link[=norm_osmfish]{"osmfish"}} - osmfish normalization method
\item \code{\link[=norm_pearson]{"pearson"}} - Lause/Kobak 2020 pearson residuals
normalization
\item \code{\link[=norm_quantile]{"quantile"}} - quantile normalization
\item \code{\link[=norm_tfidf]{"tf-idf"}} - Term Frequency-Inverse Document Frequency
\item \code{\link[=norm_l2]{"l2"}} - L2 normalization (also known as Euclidean
normalization)
\item \code{\link[=norm_arcsinh]{"arcsinh"}} - arcsinh transformation
}
}

\section{scaleParam methods}{

\itemize{
\item \code{\link[=scale_default]{"default"}} - default Giotto scaling steps (scale along
features then cells)
\item \code{\link[=scale_zscore]{"zscore"}} - essentially the same as \code{base::scale()}, but
with a \code{MARGIN} param allowing scaling long either cols or rows
}
}

\section{adjustParam methods}{

\itemize{
\item \code{\link[=adjust_limma]{"limma"}} - limma batch correction
}
}

\section{thresholdParam methods}{

\itemize{
\item \code{\link[=threshold_binarize]{"binarize"}} - data binarization (matrices and
rasters)
\item \code{\link[=threshold_minmax]{"minmax"}} - value restriction/clamping (matrices
and rasters)
}
}

\seealso{
\code{\link[=processData]{processData()}} for the generic used to apply these params

\code{\link[=processExpression]{processExpression()}} for the way to use this framework with the
\code{giotto} object
}
