% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{processData}
\alias{processData}
\alias{processData,exprObj,list-method}
\alias{processData,exprObj,normParam-method}
\alias{processData,exprObj,adjustParam-method}
\alias{processData,exprObj,scaleParam-method}
\alias{processData,allMatrix,list-method}
\title{Composable Data Processing}
\usage{
\S4method{processData}{exprObj,list}(x, param, name = "scaled", ...)

\S4method{processData}{exprObj,normParam}(x, param, name = "normalized", ...)

\S4method{processData}{exprObj,adjustParam}(x, param, name = "custom", ...)

\S4method{processData}{exprObj,scaleParam}(x, param, name = "scaled", ...)

\S4method{processData}{allMatrix,list}(x, param, ...)
}
\arguments{
\item{x}{data to transform}

\item{param}{S4 parameter class defining the transform operation and
params affecting it. Can also be a list of several of these objects, acting
as a pipeline.}

\item{name}{character. \link[GiottoClass:giotto_schema]{Object name} to assign
to the output.}

\item{\dots}{additional params to pass}
}
\value{
The same class as \code{x}
}
\description{
Perform data transformations, or set up chains of transformations and
operations to be applied to matrix type data. \code{processData()} is a generic
for which methods can be defined off both \code{x} (the data to transform),
and \code{param} (the transform operation).
}
\examples{
m <- matrix(c(0, 0, 3, 2, 0, 5, 4, 0, 0, 1, 12, 0), nrow = 3)

# single operation
lib_norm <- normParam("library")
lib_norm$scalefactor <- 5000 # alter a default param of library norm
processData(m, lib_norm)

# chained operations
log_norm <- normParam("log")
zscore_cols <- scaleParam("zscore")
zscore_rows <- scaleParam("zscore", MARGIN = 1)
# this is essentially the same as the default giotto normalization
# only difference is the library norm scalefactor change.
processData(m, list(lib_norm, log_norm, zscore_cols, zscore_rows))
}
\seealso{
\link{process_param} for processing operations that can be performed
through \code{processData()}

\code{\link[=processExpression]{processExpression()}} for the way to use this framework with the
\code{giotto} object
}
