% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{processExpression}
\alias{processExpression}
\title{Expression Data Processing}
\usage{
processExpression(
  gobject,
  param,
  name = NULL,
  expression_values = "raw",
  spat_unit = NULL,
  feat_type = NULL,
  return_gobject = TRUE,
  ...
)
}
\arguments{
\item{gobject}{\code{giotto} object}

\item{param}{S4 parameter class defining the transform operation and
params affecting it. Can also be a list of several of these objects, acting
as a pipeline.}

\item{name}{character (optional). \link[GiottoClass:giotto_schema]{Object name}
to assign to the output. Default \code{name} changes based on \code{param} input:
\itemize{
\item when \code{param} is \code{list} or \code{scaleParam}: \code{name = "scaled"}
\item when \code{param} is \code{normParam}: \code{name = "normalized"}
\item when \code{param} is \code{adjustParam}: \code{name = "custom"}
\item when \code{param} is \code{osmFISHNormParam}: \code{name = "custom"}
\item when \code{param} is \code{pearsonResidNormParam}: \code{name = "scaled"}
}}

\item{expression_values}{character. Name of matrix to use}

\item{spat_unit}{character (optional). spatial unit to use}

\item{feat_type}{character (optional). feature type to use}

\item{return_gobject}{logical (optional). Whether to return the \code{gobject}.
When FALSE, the \code{exprObj} is returned instead.}

\item{...}{additional params to pass}
}
\value{
A \code{giotto} object when \code{return_gobject = TRUE}. Otherwise, an
\code{exprObj}
}
\description{
Perform data transformations, or set up chains of transformations and
operations to be applied to expression type data in the \code{giotto} object.
}
\examples{
g <- GiottoData::loadGiottoMini("visium")
# single operation
processExpression(g, normParam("library"), name = "library")

# single operation with changed parameter
lib <- normParam("library")
lib$scalefactor = 1000
processExpression(g, lib, name = "library2")

# return the exprObj instead
processExpression(g, lib, name = "library2", return_gobject = FALSE)

# chained operation (this is the Giotto standard normalization)
processExpression(g,
    list(
        normParam("library"),
        normParam("log"),
        scaleParam("zscore", MARGIN = 1),
        scaleParam("zscore", MARGIN = 2)
    ),
    name = "scaled2"
)
}
\seealso{
\link{process_param} for processing operations that can be performed

\code{\link[=processData]{processData()}} for the lower level generic handling these operations
}
