% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{processGiotto}
\alias{processGiotto}
\title{processGiotto}
\usage{
processGiotto(
  gobject,
  filter_params = list(),
  norm_params = list(),
  stat_params = list(),
  adjust_params = list(),
  verbose = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{filter_params}{additional parameters to filterGiotto}

\item{norm_params}{additional parameters to normalizeGiotto}

\item{stat_params}{additional parameters to addStatistics}

\item{adjust_params}{additional parameters to adjustGiottoMatrix}

\item{verbose}{be verbose (default is TRUE)}
}
\value{
giotto object
}
\description{
Wrapper for the different Giotto object processing functions
}
\details{
See \code{\link{filterGiotto}}, \code{\link{normalizeGiotto}},
\code{\link{addStatistics}} and \code{\link{adjustGiottoMatrix}} for more
information about the different parameters in each step. If you do not provide
them it will use the default values.
}
\examples{

data(mini_giotto_single_cell)

processed_object = processGiotto(mini_giotto_single_cell,
                                 filter_params = list(gene_det_in_min_cells = 10,
                                 min_det_genes_per_cell = 10))

}
