% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_genes.R
\name{rankSpatialCorGroups}
\alias{rankSpatialCorGroups}
\title{rankSpatialCorGroups}
\usage{
rankSpatialCorGroups(
  gobject,
  spatCorObject,
  use_clus_name = NULL,
  show_plot = NA,
  return_plot = FALSE,
  save_plot = NA,
  save_param = list(),
  default_save_name = "rankSpatialCorGroups"
)
}
\arguments{
\item{gobject}{giotto object}

\item{spatCorObject}{spatial correlation object}

\item{use_clus_name}{name of clusters to visualize (from clusterSpatialCorGenes())}

\item{show_plot}{show plot}

\item{return_plot}{return ggplot object}

\item{save_plot}{directly save the plot [boolean]}

\item{save_param}{list of saving parameters, see \code{\link{showSaveParameters}}}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}
}
\value{
data.table with positive (within group) and negative (outside group) scores
}
\description{
Rank spatial correlated clusters according to correlation structure
}
