% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{read_cell_metadata}
\alias{read_cell_metadata}
\alias{readCellMetadata}
\title{Read cell metadata}
\usage{
readCellMetadata(
  data_list,
  default_spat_unit = NULL,
  default_feat_type = NULL,
  provenance = NULL,
  verbose = TRUE
)

read_cell_metadata(
  metadata,
  default_spat_unit = NULL,
  default_feat_type = NULL,
  provenance = NULL,
  verbose = TRUE
)
}
\arguments{
\item{data_list}{nested list of cell metadata information}

\item{provenance}{provenance information (optional)}

\item{verbose}{be verbose}

\item{metadata}{nested list of cell metadata information}
}
\description{
read cell metadata from list

read cell metadata from list
}
\keyword{internal}
