% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_section.R
\name{read_crossSection}
\alias{read_crossSection}
\title{read_crossSection}
\usage{
read_crossSection(gobject, name = NULL, spatial_network_name = NULL)
}
\description{
read a cross section object from a giotto object
}
