% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_section.R
\name{read_crossSection}
\alias{read_crossSection}
\title{read_crossSection}
\usage{
read_crossSection(gobject, name = NULL, spatial_network_name = NULL)
}
\arguments{
\item{gobject}{gobject}

\item{name}{name}

\item{spatial_network_name}{spatial_network_name}
}
\description{
read a cross section object from a giotto object
}
\keyword{internal}
