% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{read_feature_metadata}
\alias{read_feature_metadata}
\alias{readFeatMetadata}
\title{Read feature metadata}
\usage{
readFeatMetadata(
  data_list,
  default_spat_unit = NULL,
  default_feat_type = NULL,
  provenance = NULL,
  verbose = TRUE
)
}
\arguments{
\item{data_list}{nested list of feature metadata information}

\item{provenance}{provenance information (optional)}

\item{verbose}{be verbose}
}
\description{
read feature metadata from listt
}
