% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{readDimReducData}
\alias{readDimReducData}
\title{Read dimensional reduction data}
\usage{
readDimReducData(
  data_list,
  default_spat_unit = NULL,
  default_feat_type = NULL,
  reduction = c("cells", "feats"),
  provenance = NULL,
  verbose = TRUE
)
}
\arguments{
\item{data_list}{(nested) list of dimension reduction input data}

\item{default_spat_unit}{(optional) default spat_unit to use}

\item{default_feat_type}{(optional) default feat_type to use}

\item{reduction}{whether dim reduction was performed on 'cels' or 'feats'}

\item{provenance}{(optional) provenance information}

\item{verbose}{be verbose}
}
\description{
read dimension reduction results from list
}
