% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto.R
\name{readExprMatrix}
\alias{readExprMatrix}
\title{readExprMatrix}
\usage{
readExprMatrix(path, cores = NA)
}
\arguments{
\item{path}{path to the expression matrix}
}
\value{
sparse matrix
}
\description{
Function to read an expression matrix into a sparse matrix.
}
\details{
The expression matrix needs to have both unique column names and row names
}
\examples{
    readExprMatrix()
}
