% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto.R
\name{readGiottoInstructions}
\alias{readGiottoInstructions}
\title{readGiottoInstrunctions}
\usage{
readGiottoInstructions(giotto_instructions, param = NULL)
}
\arguments{
\item{giotto_instructions}{giotto object or result from createGiottoInstructions()}

\item{param}{parameter to retrieve}
}
\value{
specific parameter
}
\description{
Retrieves the instruction associated with the provided parameter
}
