% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_help.R
\name{readPolygonFilesVizgen}
\alias{readPolygonFilesVizgen}
\title{readPolygonFilesVizgen}
\usage{
readPolygonFilesVizgen(
  gobject,
  boundaries_path,
  fovs = NULL,
  polygon_feat_types = 0:6,
  flip_x_axis = F,
  flip_y_axis = F,
  smooth_polygons = TRUE,
  smooth_vertices = 60,
  set_neg_to_zero = FALSE,
  return_gobject = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{boundaries_path}{path to the cell_boundaries folder}

\item{fovs}{selected fovs, if NULL select all fovs within Giotto object}

\item{polygon_feat_types}{a vector containing the polygon feature types}

\item{flip_x_axis}{flip x axis of polygon coordinates (multiply by -1)}

\item{flip_y_axis}{flip y axis of polygon coordinates (multiply by -1)}

\item{smooth_polygons}{smooth polygons (default = TRUE)}

\item{smooth_vertices}{number of vertices for smoothing}

\item{set_neg_to_zero}{set negative values to zero when smoothing}

\item{return_gobject}{return giotto object}

\item{verbose}{be verbose}
}
\description{
Read selected polygon files for the FOVs present in the Giotto
object and add the smoothed polygons to the object
}
\seealso{
\code{\link{smoothGiottoPolygons}}
}
