% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_help.R
\name{readPolygonFilesVizgen}
\alias{readPolygonFilesVizgen}
\title{readPolygonFilesVizgen}
\usage{
readPolygonFilesVizgen(gobject, boundaries_path, polygon_feat_types = 0:6)
}
\arguments{
\item{gobject}{giotto object}

\item{boundaries_path}{path to the cell_boundaries folder}

\item{polygon_feat_types}{a vector containing the polygon feature types}

\item{verbose}{be verbose
considered in the analysis}
}
\description{
Read selected polygon files for the FOVs present in the Giotto
object and add the smoothed polygons to the object
}
