% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_help.R
\name{readPolygonFilesVizgenHDF5}
\alias{readPolygonFilesVizgenHDF5}
\title{readPolygonFilesVizgenHDF5}
\usage{
readPolygonFilesVizgenHDF5(
  boundaries_path,
  fovs = NULL,
  polygon_feat_types = 0:6,
  custom_polygon_names = NULL,
  flip_x_axis = F,
  flip_y_axis = F,
  smooth_polygons = TRUE,
  smooth_vertices = 60,
  set_neg_to_zero = FALSE,
  H5Fopen_flags = h5default("H5F_ACC_RD"),
  verbose = TRUE
)
}
\arguments{
\item{boundaries_path}{path to the cell_boundaries folder}

\item{fovs}{subset of fovs to use}

\item{polygon_feat_types}{a vector containing the polygon feature types}

\item{custom_polygon_names}{a character vector to provide custom polygon names
(optional)}

\item{flip_x_axis}{flip x axis of polygon coordinates (multiply by -1)}

\item{flip_y_axis}{flip y axis of polygon coordinates (multiply by -1)}

\item{smooth_polygons}{smooth polygons (default = TRUE)}

\item{smooth_vertices}{number of vertices for smoothing}

\item{set_neg_to_zero}{set negative values to zero when smoothing}

\item{H5Fopen_flags}{see \code{\link[rhdf5]{H5Fopen}} for more details}

\item{verbose}{be verbose}
}
\description{
Read and create polygons for all cells, or for only selected FOVs.
}
\details{
Set H5Fopen_flags to "H5F_ACC_RDONLY" if you encounter permission issues.
}
\seealso{
\code{\link{smoothGiottoPolygons}}
}
