% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_help.R
\name{readPolygonFilesVizgenHDF5}
\alias{readPolygonFilesVizgenHDF5}
\title{readPolygonFilesVizgenHDF5}
\usage{
readPolygonFilesVizgenHDF5(
  boundaries_path,
  fovs = NULL,
  z_indices = 1L:7L,
  segm_to_use = 1L,
  custom_polygon_names = NULL,
  flip_x_axis = FALSE,
  flip_y_axis = TRUE,
  calc_centroids = FALSE,
  smooth_polygons = TRUE,
  smooth_vertices = 60L,
  set_neg_to_zero = FALSE,
  H5Fopen_flags = "H5F_ACC_RDWR",
  cores = determine_cores(),
  create_gpoly_parallel = TRUE,
  create_gpoly_bin = FALSE,
  verbose = TRUE,
  output = c("giottoPolygon", "data.table"),
  polygon_feat_types = NULL
)
}
\arguments{
\item{boundaries_path}{path to the cell_boundaries folder}

\item{fovs}{subset of fovs to use}

\item{z_indices}{z indices of polygons to use}

\item{segm_to_use}{segmentation results to use (usually = 1. Depends on if
alternative segmentations were generated)}

\item{custom_polygon_names}{a character vector to provide custom polygon
names (optional)}

\item{flip_x_axis}{flip x axis of polygon coordinates (multiply by -1)}

\item{flip_y_axis}{flip y axis of polygon coordinates (multiply by -1)}

\item{calc_centroids}{calculate centroids (default = FALSE)}

\item{smooth_polygons}{smooth polygons (default = TRUE)}

\item{smooth_vertices}{number of vertices for smoothing}

\item{set_neg_to_zero}{set negative values to zero when smoothing}

\item{H5Fopen_flags}{see \code{\link[rhdf5]{H5Fopen}} for more details}

\item{cores}{cores to use}

\item{create_gpoly_parallel}{(default = TRUE) Whether to run gpoly creation
in parallel}

\item{create_gpoly_bin}{(Optional, default = FALSE) Parallelization option.
Accepts integer values as an binning size when generating giottoPolygon
objects}

\item{verbose}{be verbose}

\item{output}{whether to return as list of giottoPolygon or data.table}

\item{polygon_feat_types}{deprecated. Use \code{z_indices}}
}
\value{
list of giottoPolygon or data.table
}
\description{
Read polygon info for all cells or for only selected FOVs from
Vizgen HDF5 files. Data is returned as a list of giottoPolygons or
data.tables of the requested z indices.
}
\details{
Set H5Fopen_flags to "H5F_ACC_RDONLY" if you encounter permission
issues.
}
\seealso{
\code{\link{smoothGiottoPolygons}}
}
