% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto.R
\name{readSpatLocData}
\alias{readSpatLocData}
\title{Read spatial location data}
\usage{
readSpatLocData(
  data_list,
  default_spat_unit = NULL,
  provenance = NULL,
  cores = determine_cores(),
  verbose = TRUE
)
}
\arguments{
\item{data_list}{(nested) list of spatial locations input data}

\item{provenance}{provenance information (optional)}

\item{cores}{how many cores to use}

\item{verbose}{be verbose}
}
\value{
list of spatLocsObj
}
\description{
read spatial locations/coordinates from nested list and generate
list of Giotto spatLocsObj
}
