% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{readSpatNetData}
\alias{readSpatNetData}
\title{Read spatial networks}
\usage{
readSpatNetData(
  data_list,
  default_spat_unit = NULL,
  provenance = NULL,
  verbose = TRUE
)
}
\arguments{
\item{data_list}{(nested) list of spatial network input data}

\item{default_spat_unit}{(optional) default spat_unit to use}

\item{provenance}{(optional) provenance information}

\item{verbose}{be verbose}
}
\description{
read spatial networks data from list
}
