% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{reconnectGiottoImage}
\alias{reconnectGiottoImage}
\title{Reconnect images with dead pointers}
\usage{
reconnectGiottoImage(
  gobject,
  auto_reconnect = TRUE,
  image_name = NULL,
  largeImage_name = NULL,
  image_path = NULL,
  largeImage_path = NULL,
  verbose = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{auto_reconnect}{automatically reconnect images if TRUE. manual if FALSE}

\item{image_name}{names of images to reconnect}

\item{largeImage_name}{name of large images to reconnect}

\item{image_path}{named list of paths to images to reconnect to giottoImages}

\item{largeImage_path}{named list of paths to images to reconnect to giottoLargeImages}

\item{verbose}{be verbose}
}
\value{
a giotto object with updated image pointer
}
\description{
reconnect a gobject's dead image pointers using filepaths to 
  the original source image files
}
\details{
Inputs can either be given as both image name (\code{image_name}/\code{largeImage_name})
  and filepath (\code{image_path}/\code{largeImage_path}) args or as only a
  a named list through a filepath argument alone.
  If \code{auto_reconnect = TRUE} then no additional params need to be supplied.
  As long as giotto image objects were directly created using filepaths, those
  filepaths are stored within the image objects and will be referenced during
  reconnection. Issues will only arise if giotto image objects were created directly
  from the underlying image handling package objects (\emph{magick} or \emph{raster objects}) or
  if image files have been moved since the the giotto image object was generated.
  In such cases, use manual reconnection by setting \code{auto_reconnect = FALSE}.
}
