% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suite_reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{addCellMetadata}
\alias{addGiottoImage}
\alias{addGiottoImageMG}
\alias{addGiottoLargeImage}
\alias{addGiottoPoints}
\alias{addGiottoPoints3D}
\alias{addGiottoPolygons}
\alias{addSpatialCentroidLocations}
\alias{addSpatialCentroidLocationsLayer}
\alias{aggregateStacks}
\alias{aggregateStacksExpression}
\alias{aggregateStacksLocations}
\alias{aggregateStacksPolygonOverlaps}
\alias{aggregateStacksPolygons}
\alias{aggregateStacksPolygonsOLD}
\alias{anndataToGiotto}
\alias{annotateGiotto}
\alias{annotateSpatialGrid}
\alias{annotateSpatialNetwork}
\alias{calculateMetaTable}
\alias{calculateMetaTableCells}
\alias{calculateOverlapParallel}
\alias{calculateOverlapPolygonImages}
\alias{calculateOverlapRaster}
\alias{calculateOverlapSerial}
\alias{changeGiottoInstructions}
\alias{changeImageBg}
\alias{circleVertices}
\alias{combineCellData}
\alias{combineFeatureData}
\alias{combineFeatureOverlapData}
\alias{combineMetadata}
\alias{combineSpatialCellFeatureInfo}
\alias{combineSpatialCellMetadataInfo}
\alias{convertGiottoLargeImageToMG}
\alias{createCellMetaObj}
\alias{createDimObj}
\alias{createExprObj}
\alias{createFeatMetaObj}
\alias{createGiottoImage}
\alias{createGiottoInstructions}
\alias{createGiottoLargeImage}
\alias{createGiottoLargeImageList}
\alias{createGiottoObject}
\alias{createGiottoObjectSubcellular}
\alias{createGiottoPoints}
\alias{createGiottoPolygonsFromDfr}
\alias{createGiottoPolygonsFromGeoJSON}
\alias{createGiottoPolygonsFromMask}
\alias{createMetafeats}
\alias{createNearestNetObj}
\alias{createSpatEnrObj}
\alias{createSpatLocsObj}
\alias{createSpatNetObj}
\alias{createSpatialDefaultGrid}
\alias{createSpatialDelaunayNetwork}
\alias{createSpatialFeaturesKNNnetwork}
\alias{createSpatialGrid}
\alias{createSpatialKNNnetwork}
\alias{createSpatialNetwork}
\alias{createSpatialWeightMatrix}
\alias{createNearestNetwork}
\alias{cropGiottoLargeImage}
\alias{distGiottoImage}
\alias{estimateImageBg}
\alias{fDataDT}
\alias{featType}
\alias{featureNetwork}
\alias{gefToGiotto}
\alias{getCellMetadata}
\alias{getDimReduction}
\alias{getExpression}
\alias{getFeatureInfo}
\alias{getFeatureMetadata}
\alias{getGiottoImage}
\alias{getMultiomics}
\alias{getNearestNetwork}
\alias{getPolygonInfo}
\alias{getRainbowColors}
\alias{getSpatialEnrichment}
\alias{getSpatialGrid}
\alias{getSpatialLocations}
\alias{getSpatialNetwork}
\alias{giotto}
\alias{giottoImage}
\alias{giottoLargeImage}
\alias{giottoMasterToSuite}
\alias{giottoPoints}
\alias{giottoPolygon}
\alias{giottoToAnnData}
\alias{giottoToSeurat}
\alias{giottoToSeuratV4}
\alias{giottoToSeuratV5}
\alias{giottoToSpatialExperiment}
\alias{hexVertices}
\alias{joinGiottoObjects}
\alias{loadGiotto}
\alias{makePseudoVisium}
\alias{objHistory}
\alias{objName}
\alias{overlapImagesToMatrix}
\alias{overlapToMatrix}
\alias{overlapToMatrixMultiPoly}
\alias{pDataDT}
\alias{plotGiottoImage}
\alias{polyStamp}
\alias{prov}
\alias{readCellMetadata}
\alias{readDimReducData}
\alias{readExprData}
\alias{readExprMatrix}
\alias{readFeatData}
\alias{readFeatMetadata}
\alias{readGiottoInstructions}
\alias{readNearestNetData}
\alias{readPolygonData}
\alias{readSpatEnrichData}
\alias{readSpatLocsData}
\alias{readSpatNetData}
\alias{reconnectGiottoImage}
\alias{rectVertices}
\alias{removeCellAnnotation}
\alias{removeFeatAnnotation}
\alias{replaceGiottoInstructions}
\alias{rescalePolygons}
\alias{saveGiotto}
\alias{setDimReduction}
\alias{setExpression}
\alias{setFeatureInfo}
\alias{setGiottoImage}
\alias{setMultiomics}
\alias{setNearestNetwork}
\alias{setPolygonInfo}
\alias{setSpatialEnrichment}
\alias{setSpatialGrid}
\alias{setSpatialLocations}
\alias{setSpatialNetwork}
\alias{seuratToGiotto}
\alias{seuratToGiottoV4}
\alias{seuratToGiottoV5}
\alias{showGiottoCellMetadata}
\alias{showGiottoDimRed}
\alias{showGiottoExpression}
\alias{showGiottoFeatInfo}
\alias{showGiottoFeatMetadata}
\alias{showGiottoImageNames}
\alias{showGiottoInstructions}
\alias{showGiottoNearestNetworks}
\alias{showGiottoSpatEnrichments}
\alias{showGiottoSpatGrids}
\alias{showGiottoSpatLocs}
\alias{showGiottoSpatNetworks}
\alias{showGiottoSpatialInfo}
\alias{showGrids}
\alias{showNetworks}
\alias{showProcessingSteps}
\alias{tessellate}
\alias{dimCellPlot2D}
\alias{dimFeatPlot2D}
\alias{dimGenePlot3D}
\alias{dimPlot}
\alias{dimPlot2D}
\alias{dimPlot3D}
\alias{getDistinctColors}
\alias{plotHeatmap}
\alias{plotMetaDataCellsHeatmap}
\alias{plotMetaDataHeatmap}
\alias{plotPCA}
\alias{plotPCA_2D}
\alias{plotPCA_3D}
\alias{plotStatDelaunayNetwork}
\alias{plotTSNE}
\alias{plotTSNE_2D}
\alias{plotTSNE_3D}
\alias{plotUMAP}
\alias{plotUMAP_2D}
\alias{plotUMAP_3D}
\alias{showClusterDendrogram}
\alias{showClusterHeatmap}
\alias{showSaveParameters}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{GiottoClass}{\code{\link[GiottoClass]{addCellMetadata}}, \code{\link[GiottoClass]{addGiottoImage}}, \code{\link[GiottoClass]{addGiottoImageMG}}, \code{\link[GiottoClass]{addGiottoLargeImage}}, \code{\link[GiottoClass]{addGiottoPoints}}, \code{\link[GiottoClass:addGiottoPoints]{addGiottoPoints3D}}, \code{\link[GiottoClass]{addGiottoPolygons}}, \code{\link[GiottoClass]{addSpatialCentroidLocations}}, \code{\link[GiottoClass]{addSpatialCentroidLocationsLayer}}, \code{\link[GiottoClass]{aggregateStacks}}, \code{\link[GiottoClass]{aggregateStacksExpression}}, \code{\link[GiottoClass]{aggregateStacksLocations}}, \code{\link[GiottoClass]{aggregateStacksPolygonOverlaps}}, \code{\link[GiottoClass]{aggregateStacksPolygons}}, \code{\link[GiottoClass]{aggregateStacksPolygonsOLD}}, \code{\link[GiottoClass]{anndataToGiotto}}, \code{\link[GiottoClass]{annotateGiotto}}, \code{\link[GiottoClass]{annotateSpatialGrid}}, \code{\link[GiottoClass]{annotateSpatialNetwork}}, \code{\link[GiottoClass]{calculateMetaTable}}, \code{\link[GiottoClass]{calculateMetaTableCells}}, \code{\link[GiottoClass]{calculateOverlapParallel}}, \code{\link[GiottoClass]{calculateOverlapPolygonImages}}, \code{\link[GiottoClass]{calculateOverlapRaster}}, \code{\link[GiottoClass]{calculateOverlapSerial}}, \code{\link[GiottoClass]{changeGiottoInstructions}}, \code{\link[GiottoClass]{changeImageBg}}, \code{\link[GiottoClass]{circleVertices}}, \code{\link[GiottoClass]{combineCellData}}, \code{\link[GiottoClass]{combineFeatureData}}, \code{\link[GiottoClass]{combineFeatureOverlapData}}, \code{\link[GiottoClass]{combineMetadata}}, \code{\link[GiottoClass]{combineSpatialCellFeatureInfo}}, \code{\link[GiottoClass]{combineSpatialCellMetadataInfo}}, \code{\link[GiottoClass]{convertGiottoLargeImageToMG}}, \code{\link[GiottoClass]{createCellMetaObj}}, \code{\link[GiottoClass]{createDimObj}}, \code{\link[GiottoClass]{createExprObj}}, \code{\link[GiottoClass]{createFeatMetaObj}}, \code{\link[GiottoClass]{createGiottoImage}}, \code{\link[GiottoClass]{createGiottoInstructions}}, \code{\link[GiottoClass]{createGiottoLargeImage}}, \code{\link[GiottoClass]{createGiottoLargeImageList}}, \code{\link[GiottoClass]{createGiottoObject}}, \code{\link[GiottoClass]{createGiottoObjectSubcellular}}, \code{\link[GiottoClass]{createGiottoPoints}}, \code{\link[GiottoClass:createGiottoPolygon]{createGiottoPolygonsFromDfr}}, \code{\link[GiottoClass:createGiottoPolygon]{createGiottoPolygonsFromGeoJSON}}, \code{\link[GiottoClass:createGiottoPolygon]{createGiottoPolygonsFromMask}}, \code{\link[GiottoClass]{createMetafeats}}, \code{\link[GiottoClass]{createNearestNetObj}}, \code{\link[GiottoClass]{createNearestNetwork}}, \code{\link[GiottoClass]{createSpatEnrObj}}, \code{\link[GiottoClass]{createSpatialDefaultGrid}}, \code{\link[GiottoClass]{createSpatialDelaunayNetwork}}, \code{\link[GiottoClass]{createSpatialFeaturesKNNnetwork}}, \code{\link[GiottoClass]{createSpatialGrid}}, \code{\link[GiottoClass]{createSpatialKNNnetwork}}, \code{\link[GiottoClass]{createSpatialNetwork}}, \code{\link[GiottoClass]{createSpatialWeightMatrix}}, \code{\link[GiottoClass]{createSpatLocsObj}}, \code{\link[GiottoClass]{createSpatNetObj}}, \code{\link[GiottoClass]{cropGiottoLargeImage}}, \code{\link[GiottoClass]{distGiottoImage}}, \code{\link[GiottoClass]{estimateImageBg}}, \code{\link[GiottoClass]{fDataDT}}, \code{\link[GiottoClass:featType-generic]{featType}}, \code{\link[GiottoClass:featureNetwork-class]{featureNetwork}}, \code{\link[GiottoClass]{gefToGiotto}}, \code{\link[GiottoClass]{getCellMetadata}}, \code{\link[GiottoClass]{getDimReduction}}, \code{\link[GiottoClass]{getExpression}}, \code{\link[GiottoClass]{getFeatureInfo}}, \code{\link[GiottoClass]{getFeatureMetadata}}, \code{\link[GiottoClass]{getGiottoImage}}, \code{\link[GiottoClass]{getMultiomics}}, \code{\link[GiottoClass]{getNearestNetwork}}, \code{\link[GiottoClass]{getPolygonInfo}}, \code{\link[GiottoClass]{getRainbowColors}}, \code{\link[GiottoClass]{getSpatialEnrichment}}, \code{\link[GiottoClass]{getSpatialEnrichment}}, \code{\link[GiottoClass]{getSpatialGrid}}, \code{\link[GiottoClass]{getSpatialGrid}}, \code{\link[GiottoClass]{getSpatialLocations}}, \code{\link[GiottoClass]{getSpatialLocations}}, \code{\link[GiottoClass]{getSpatialNetwork}}, \code{\link[GiottoClass]{getSpatialNetwork}}, \code{\link[GiottoClass:giotto-class]{giotto}}, \code{\link[GiottoClass:giottoImage-class]{giottoImage}}, \code{\link[GiottoClass:giottoLargeImage-class]{giottoLargeImage}}, \code{\link[GiottoClass]{giottoMasterToSuite}}, \code{\link[GiottoClass:giottoPoints-class]{giottoPoints}}, \code{\link[GiottoClass:giottoPolygon-class]{giottoPolygon}}, \code{\link[GiottoClass]{giottoToAnnData}}, \code{\link[GiottoClass]{giottoToSeurat}}, \code{\link[GiottoClass]{giottoToSeuratV4}}, \code{\link[GiottoClass]{giottoToSeuratV5}}, \code{\link[GiottoClass]{giottoToSpatialExperiment}}, \code{\link[GiottoClass]{hexVertices}}, \code{\link[GiottoClass]{joinGiottoObjects}}, \code{\link[GiottoClass]{loadGiotto}}, \code{\link[GiottoClass]{makePseudoVisium}}, \code{\link[GiottoClass]{objHistory}}, \code{\link[GiottoClass:objName-generic]{objName}}, \code{\link[GiottoClass]{overlapImagesToMatrix}}, \code{\link[GiottoClass]{overlapToMatrix}}, \code{\link[GiottoClass]{overlapToMatrixMultiPoly}}, \code{\link[GiottoClass]{pDataDT}}, \code{\link[GiottoClass]{plotGiottoImage}}, \code{\link[GiottoClass]{polyStamp}}, \code{\link[GiottoClass:prov-generic]{prov}}, \code{\link[GiottoClass:read_cell_metadata]{readCellMetadata}}, \code{\link[GiottoClass]{readDimReducData}}, \code{\link[GiottoClass]{readExprData}}, \code{\link[GiottoClass]{readExprMatrix}}, \code{\link[GiottoClass]{readFeatData}}, \code{\link[GiottoClass:read_feature_metadata]{readFeatMetadata}}, \code{\link[GiottoClass]{readGiottoInstructions}}, \code{\link[GiottoClass]{readNearestNetData}}, \code{\link[GiottoClass]{readPolygonData}}, \code{\link[GiottoClass]{readSpatEnrichData}}, \code{\link[GiottoClass]{readSpatLocsData}}, \code{\link[GiottoClass]{readSpatNetData}}, \code{\link[GiottoClass]{reconnectGiottoImage}}, \code{\link[GiottoClass]{rectVertices}}, \code{\link[GiottoClass]{removeCellAnnotation}}, \code{\link[GiottoClass]{removeFeatAnnotation}}, \code{\link[GiottoClass]{replaceGiottoInstructions}}, \code{\link[GiottoClass]{rescalePolygons}}, \code{\link[GiottoClass]{saveGiotto}}, \code{\link[GiottoClass]{setDimReduction}}, \code{\link[GiottoClass]{setExpression}}, \code{\link[GiottoClass]{setFeatureInfo}}, \code{\link[GiottoClass]{setGiottoImage}}, \code{\link[GiottoClass]{setMultiomics}}, \code{\link[GiottoClass]{setNearestNetwork}}, \code{\link[GiottoClass]{setPolygonInfo}}, \code{\link[GiottoClass]{setSpatialEnrichment}}, \code{\link[GiottoClass]{setSpatialGrid}}, \code{\link[GiottoClass]{setSpatialLocations}}, \code{\link[GiottoClass]{setSpatialNetwork}}, \code{\link[GiottoClass]{seuratToGiotto}}, \code{\link[GiottoClass]{seuratToGiottoV4}}, \code{\link[GiottoClass]{seuratToGiottoV5}}, \code{\link[GiottoClass]{showGiottoCellMetadata}}, \code{\link[GiottoClass]{showGiottoDimRed}}, \code{\link[GiottoClass]{showGiottoExpression}}, \code{\link[GiottoClass]{showGiottoFeatInfo}}, \code{\link[GiottoClass]{showGiottoFeatMetadata}}, \code{\link[GiottoClass]{showGiottoImageNames}}, \code{\link[GiottoClass]{showGiottoInstructions}}, \code{\link[GiottoClass]{showGiottoNearestNetworks}}, \code{\link[GiottoClass]{showGiottoSpatEnrichments}}, \code{\link[GiottoClass]{showGiottoSpatGrids}}, \code{\link[GiottoClass]{showGiottoSpatialInfo}}, \code{\link[GiottoClass]{showGiottoSpatLocs}}, \code{\link[GiottoClass]{showGiottoSpatNetworks}}, \code{\link[GiottoClass]{showGrids}}, \code{\link[GiottoClass]{showNetworks}}, \code{\link[GiottoClass]{showProcessingSteps}}, \code{\link[GiottoClass]{tessellate}}}

  \item{GiottoUtils}{\code{\link[GiottoUtils:pipe]{\%>\%}}}

  \item{GiottoVisuals}{\code{\link[GiottoVisuals:dimCellPlot]{dimCellPlot2D}}, \code{\link[GiottoVisuals]{dimFeatPlot2D}}, \code{\link[GiottoVisuals]{dimGenePlot3D}}, \code{\link[GiottoVisuals]{dimPlot}}, \code{\link[GiottoVisuals:dimPlot]{dimPlot2D}}, \code{\link[GiottoVisuals:dimPlot]{dimPlot3D}}, \code{\link[GiottoVisuals]{getDistinctColors}}, \code{\link[GiottoVisuals]{plotHeatmap}}, \code{\link[GiottoVisuals]{plotMetaDataCellsHeatmap}}, \code{\link[GiottoVisuals]{plotMetaDataHeatmap}}, \code{\link[GiottoVisuals]{plotPCA}}, \code{\link[GiottoVisuals]{plotPCA_2D}}, \code{\link[GiottoVisuals]{plotPCA_3D}}, \code{\link[GiottoVisuals]{plotStatDelaunayNetwork}}, \code{\link[GiottoVisuals]{plotTSNE}}, \code{\link[GiottoVisuals]{plotTSNE_2D}}, \code{\link[GiottoVisuals]{plotTSNE_3D}}, \code{\link[GiottoVisuals]{plotUMAP}}, \code{\link[GiottoVisuals]{plotUMAP_2D}}, \code{\link[GiottoVisuals]{plotUMAP_3D}}, \code{\link[GiottoVisuals]{showClusterDendrogram}}, \code{\link[GiottoVisuals]{showClusterHeatmap}}, \code{\link[GiottoVisuals]{showSaveParameters}}}
}}

