% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suite_reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{getRainbowColors}
\alias{getDistinctColors}
\alias{activeSpatUnit}
\alias{activeSpatUnit<-}
\alias{activeFeatType}
\alias{activeFeatType<-}
\alias{ext}
\alias{ext<-}
\alias{featType}
\alias{featType<-}
\alias{spatUnit}
\alias{spatUnit<-}
\alias{instructions}
\alias{instructions<-}
\alias{prov}
\alias{prov<-}
\alias{objName}
\alias{objName<-}
\alias{as.points}
\alias{as.polygons}
\alias{as.terra}
\alias{as.stars}
\alias{as.sf}
\alias{as.sp}
\alias{overlaps}
\alias{calculateOverlap}
\alias{overlapToMatrix}
\alias{centroids}
\alias{copy}
\alias{crop}
\alias{density}
\alias{flip}
\alias{spin}
\alias{spatShift}
\alias{featIDs}
\alias{spatIDs}
\alias{setGiotto}
\alias{rescale}
\alias{vect}
\alias{wrap}
\alias{createGiottoPoints}
\alias{createGiottoPolygon}
\alias{addCellMetadata}
\alias{addFeatMetadata}
\alias{addGiottoImage}
\alias{addGiottoImageMG}
\alias{addGiottoLargeImage}
\alias{addGiottoPoints}
\alias{addGiottoPoints3D}
\alias{addGiottoPolygons}
\alias{addNetworkLayout}
\alias{addSpatialCentroidLocations}
\alias{addSpatialCentroidLocationsLayer}
\alias{aggregateStacks}
\alias{aggregateStacksExpression}
\alias{aggregateStacksLocations}
\alias{aggregateStacksPolygonOverlaps}
\alias{aggregateStacksPolygons}
\alias{anndataToGiotto}
\alias{annotateGiotto}
\alias{annotateSpatialGrid}
\alias{annotateSpatialNetwork}
\alias{calculateMetaTable}
\alias{calculateMetaTableCells}
\alias{calculateOverlapParallel}
\alias{calculateOverlapPolygonImages}
\alias{calculateOverlapRaster}
\alias{calculateOverlapSerial}
\alias{calculateSpatCellMetadataProportions}
\alias{changeGiottoInstructions}
\alias{changeImageBg}
\alias{checkGiottoEnvironment}
\alias{circleVertices}
\alias{combineCellData}
\alias{combineFeatureData}
\alias{combineFeatureOverlapData}
\alias{combineMetadata}
\alias{combineSpatialCellMetadataInfo}
\alias{combineToMultiPolygon}
\alias{convertGiottoLargeImageToMG}
\alias{createBentoAdata}
\alias{createCellMetaObj}
\alias{createDimObj}
\alias{createExprObj}
\alias{createFeatMetaObj}
\alias{createGiottoImage}
\alias{createGiottoInstructions}
\alias{createGiottoLargeImage}
\alias{createGiottoLargeImageList}
\alias{createGiottoObject}
\alias{createGiottoObjectSubcellular}
\alias{createGiottoPolygonsFromDfr}
\alias{createGiottoPolygonsFromGeoJSON}
\alias{createGiottoPolygonsFromMask}
\alias{createMetafeats}
\alias{createNearestNetObj}
\alias{createSpatEnrObj}
\alias{createSpatLocsObj}
\alias{createSpatNetObj}
\alias{createSpatialDefaultGrid}
\alias{createSpatialDelaunayNetwork}
\alias{createSpatialFeaturesKNNnetwork}
\alias{createSpatialGrid}
\alias{createSpatialKNNnetwork}
\alias{createSpatialNetwork}
\alias{createSpatialWeightMatrix}
\alias{createNearestNetwork}
\alias{cropGiottoLargeImage}
\alias{distGiottoImage}
\alias{estimateImageBg}
\alias{fDataDT}
\alias{featureNetwork}
\alias{gefToGiotto}
\alias{getCellMetadata}
\alias{getDimReduction}
\alias{getExpression}
\alias{getFeatureInfo}
\alias{getFeatureMetadata}
\alias{getGiottoImage}
\alias{getMultiomics}
\alias{getNearestNetwork}
\alias{getPolygonInfo}
\alias{getSpatialEnrichment}
\alias{getSpatialGrid}
\alias{getSpatialLocations}
\alias{getSpatialNetwork}
\alias{giotto}
\alias{giottoImage}
\alias{giottoLargeImage}
\alias{giottoMasterToSuite}
\alias{giottoPoints}
\alias{giottoPolygon}
\alias{giottoToAnnData}
\alias{giottoToSeurat}
\alias{giottoToSeuratV4}
\alias{giottoToSeuratV5}
\alias{giottoToSpatialExperiment}
\alias{hexVertices}
\alias{hist}
\alias{installGiottoEnvironment}
\alias{joinGiottoObjects}
\alias{loadGiotto}
\alias{makePseudoVisium}
\alias{objHistory}
\alias{orthoGrid}
\alias{overlapImagesToMatrix}
\alias{overlapToMatrixMultiPoly}
\alias{pDataDT}
\alias{plotGiottoImage}
\alias{polyStamp}
\alias{readCellMetadata}
\alias{readDimReducData}
\alias{readExprData}
\alias{readExprMatrix}
\alias{readFeatData}
\alias{readFeatMetadata}
\alias{readGiottoInstructions}
\alias{readNearestNetData}
\alias{readPolygonData}
\alias{readSpatEnrichData}
\alias{readSpatLocsData}
\alias{readSpatNetData}
\alias{reconnectGiottoImage}
\alias{rectVertices}
\alias{removeCellAnnotation}
\alias{removeFeatAnnotation}
\alias{removeGiottoEnvironment}
\alias{replaceGiottoInstructions}
\alias{rescalePolygons}
\alias{saveGiotto}
\alias{setCellMetadata}
\alias{setDimReduction}
\alias{setExpression}
\alias{setFeatureInfo}
\alias{setFeatureMetadata}
\alias{setGiottoImage}
\alias{setMultiomics}
\alias{setNearestNetwork}
\alias{setPolygonInfo}
\alias{setSpatialEnrichment}
\alias{setSpatialGrid}
\alias{setSpatialLocations}
\alias{setSpatialNetwork}
\alias{seuratToGiotto}
\alias{seuratToGiottoV4}
\alias{seuratToGiottoV5}
\alias{showGiottoCellMetadata}
\alias{showGiottoDimRed}
\alias{showGiottoExpression}
\alias{showGiottoFeatInfo}
\alias{showGiottoFeatMetadata}
\alias{showGiottoImageNames}
\alias{showGiottoInstructions}
\alias{showGiottoNearestNetworks}
\alias{showGiottoSpatEnrichments}
\alias{showGiottoSpatGrids}
\alias{showGiottoSpatLocs}
\alias{showGiottoSpatNetworks}
\alias{showGiottoSpatialInfo}
\alias{showProcessingSteps}
\alias{smoothGiottoPolygons}
\alias{spatQueryGiottoPolygons}
\alias{spatialExperimentToGiotto}
\alias{stitchFieldCoordinates}
\alias{stitchGiottoLargeImage}
\alias{subsetGiotto}
\alias{subsetGiottoLocs}
\alias{subsetGiottoLocsMulti}
\alias{subsetGiottoLocsSubcellular}
\alias{tessellate}
\alias{triGrid}
\alias{updateGiottoImage}
\alias{updateGiottoImageMG}
\alias{updateGiottoLargeImage}
\alias{updateGiottoObject}
\alias{updateGiottoPointsObject}
\alias{updateGiottoPolygonObject}
\alias{writeGiottoLargeImage}
\alias{addGiottoImageToSpatPlot}
\alias{dotPlot}
\alias{dimCellPlot}
\alias{dimCellPlot2D}
\alias{dimFeatPlot2D}
\alias{dimGenePlot3D}
\alias{dimFeatPlot3D}
\alias{dimPlot}
\alias{dimPlot2D}
\alias{dimPlot3D}
\alias{getColors}
\alias{giottoSankeyPlan}
\alias{plotHeatmap}
\alias{plotMetaDataCellsHeatmap}
\alias{plotMetaDataHeatmap}
\alias{plotPCA}
\alias{plotPCA_2D}
\alias{plotPCA_3D}
\alias{plotStatDelaunayNetwork}
\alias{plotTSNE}
\alias{plotTSNE_2D}
\alias{plotTSNE_3D}
\alias{plotUMAP}
\alias{plotUMAP_2D}
\alias{plotUMAP_3D}
\alias{sankeyLabel}
\alias{sankeyLabel<-}
\alias{sankeyPlot}
\alias{sankeyRelate}
\alias{sankeyRelate<-}
\alias{sankeySet}
\alias{sankeySetAddresses}
\alias{showColorInstructions}
\alias{showClusterDendrogram}
\alias{showClusterHeatmap}
\alias{showSaveParameters}
\alias{spatCellPlot}
\alias{spatCellPlot2D}
\alias{spatDeconvPlot}
\alias{spatDimCellPlot}
\alias{spatDimCellPlot2D}
\alias{spatDimFeatPlot2D}
\alias{spatDimGenePlot3D}
\alias{spatDimFeatPlot3D}
\alias{spatDimPlot}
\alias{spatDimPlot2D}
\alias{spatDimPlot3D}
\alias{spatFeatPlot2D}
\alias{spatFeatPlot2D_single}
\alias{spatGenePlot3D}
\alias{spatFeatPlot3D}
\alias{spatPlot}
\alias{spatPlot2D}
\alias{spatPlot3D}
\alias{spatInSituPlotDensity}
\alias{spatInSituPlotHex}
\alias{spatInSituPlotPoints}
\alias{spatNetwDistributions}
\alias{spatNetwDistributionsDistance}
\alias{spatNetwDistributionsKneighbors}
\alias{subsetSankeySet}
\alias{violinPlot}
\title{Objects exported from other packages}
\value{
pipe operator

vector of colors

vector of colors
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{GiottoClass}{\code{\link[GiottoClass:activeFeatType-generic]{activeFeatType}}, \code{\link[GiottoClass:activeFeatType-generic]{activeFeatType<-}}, \code{\link[GiottoClass:activeSpatUnit-generic]{activeSpatUnit}}, \code{\link[GiottoClass:activeSpatUnit-generic]{activeSpatUnit<-}}, \code{\link[GiottoClass]{addCellMetadata}}, \code{\link[GiottoClass]{addFeatMetadata}}, \code{\link[GiottoClass]{addGiottoImage}}, \code{\link[GiottoClass]{addGiottoImageMG}}, \code{\link[GiottoClass]{addGiottoLargeImage}}, \code{\link[GiottoClass]{addGiottoPoints}}, \code{\link[GiottoClass:addGiottoPoints]{addGiottoPoints3D}}, \code{\link[GiottoClass]{addGiottoPolygons}}, \code{\link[GiottoClass]{addNetworkLayout}}, \code{\link[GiottoClass]{addSpatialCentroidLocations}}, \code{\link[GiottoClass]{addSpatialCentroidLocationsLayer}}, \code{\link[GiottoClass]{aggregateStacks}}, \code{\link[GiottoClass]{aggregateStacksExpression}}, \code{\link[GiottoClass]{aggregateStacksLocations}}, \code{\link[GiottoClass]{aggregateStacksPolygonOverlaps}}, \code{\link[GiottoClass]{aggregateStacksPolygons}}, \code{\link[GiottoClass]{anndataToGiotto}}, \code{\link[GiottoClass]{annotateGiotto}}, \code{\link[GiottoClass]{annotateSpatialGrid}}, \code{\link[GiottoClass]{annotateSpatialNetwork}}, \code{\link[GiottoClass]{as.points}}, \code{\link[GiottoClass]{as.polygons}}, \code{\link[GiottoClass:r_spatial_conversions]{as.sf}}, \code{\link[GiottoClass:r_spatial_conversions]{as.sp}}, \code{\link[GiottoClass:r_spatial_conversions]{as.stars}}, \code{\link[GiottoClass:r_spatial_conversions]{as.terra}}, \code{\link[GiottoClass]{calculateMetaTable}}, \code{\link[GiottoClass]{calculateMetaTableCells}}, \code{\link[GiottoClass]{calculateOverlap}}, \code{\link[GiottoClass]{calculateOverlapParallel}}, \code{\link[GiottoClass]{calculateOverlapPolygonImages}}, \code{\link[GiottoClass]{calculateOverlapRaster}}, \code{\link[GiottoClass]{calculateOverlapSerial}}, \code{\link[GiottoClass]{calculateSpatCellMetadataProportions}}, \code{\link[GiottoClass:centroids-generic]{centroids}}, \code{\link[GiottoClass]{changeGiottoInstructions}}, \code{\link[GiottoClass]{changeImageBg}}, \code{\link[GiottoClass:giotto_python]{checkGiottoEnvironment}}, \code{\link[GiottoClass]{circleVertices}}, \code{\link[GiottoClass]{combineCellData}}, \code{\link[GiottoClass]{combineFeatureData}}, \code{\link[GiottoClass]{combineFeatureOverlapData}}, \code{\link[GiottoClass]{combineMetadata}}, \code{\link[GiottoClass]{combineSpatialCellMetadataInfo}}, \code{\link[GiottoClass]{combineToMultiPolygon}}, \code{\link[GiottoClass]{convertGiottoLargeImageToMG}}, \code{\link[GiottoClass]{copy}}, \code{\link[GiottoClass]{createBentoAdata}}, \code{\link[GiottoClass]{createCellMetaObj}}, \code{\link[GiottoClass]{createDimObj}}, \code{\link[GiottoClass]{createExprObj}}, \code{\link[GiottoClass]{createFeatMetaObj}}, \code{\link[GiottoClass]{createGiottoImage}}, \code{\link[GiottoClass:giotto_instructions]{createGiottoInstructions}}, \code{\link[GiottoClass]{createGiottoLargeImage}}, \code{\link[GiottoClass]{createGiottoLargeImageList}}, \code{\link[GiottoClass:create_giotto]{createGiottoObject}}, \code{\link[GiottoClass:create_giotto]{createGiottoObjectSubcellular}}, \code{\link[GiottoClass]{createGiottoPoints}}, \code{\link[GiottoClass]{createGiottoPolygon}}, \code{\link[GiottoClass:createGiottoPolygon]{createGiottoPolygonsFromDfr}}, \code{\link[GiottoClass:createGiottoPolygon]{createGiottoPolygonsFromGeoJSON}}, \code{\link[GiottoClass:createGiottoPolygon]{createGiottoPolygonsFromMask}}, \code{\link[GiottoClass]{createMetafeats}}, \code{\link[GiottoClass]{createNearestNetObj}}, \code{\link[GiottoClass]{createNearestNetwork}}, \code{\link[GiottoClass]{createSpatEnrObj}}, \code{\link[GiottoClass]{createSpatialDefaultGrid}}, \code{\link[GiottoClass]{createSpatialDelaunayNetwork}}, \code{\link[GiottoClass]{createSpatialFeaturesKNNnetwork}}, \code{\link[GiottoClass]{createSpatialGrid}}, \code{\link[GiottoClass]{createSpatialKNNnetwork}}, \code{\link[GiottoClass]{createSpatialNetwork}}, \code{\link[GiottoClass]{createSpatialWeightMatrix}}, \code{\link[GiottoClass]{createSpatLocsObj}}, \code{\link[GiottoClass]{createSpatNetObj}}, \code{\link[GiottoClass]{crop}}, \code{\link[GiottoClass]{cropGiottoLargeImage}}, \code{\link[GiottoClass]{density}}, \code{\link[GiottoClass]{distGiottoImage}}, \code{\link[GiottoClass]{estimateImageBg}}, \code{\link[GiottoClass]{ext}}, \code{\link[GiottoClass:ext]{ext<-}}, \code{\link[GiottoClass]{fDataDT}}, \code{\link[GiottoClass:spatIDs-generic]{featIDs}}, \code{\link[GiottoClass:giotto_schema]{featType}}, \code{\link[GiottoClass:giotto_schema]{featType<-}}, \code{\link[GiottoClass:featureNetwork-class]{featureNetwork}}, \code{\link[GiottoClass]{flip}}, \code{\link[GiottoClass]{gefToGiotto}}, \code{\link[GiottoClass]{getCellMetadata}}, \code{\link[GiottoClass]{getDimReduction}}, \code{\link[GiottoClass]{getExpression}}, \code{\link[GiottoClass]{getFeatureInfo}}, \code{\link[GiottoClass]{getFeatureMetadata}}, \code{\link[GiottoClass]{getGiottoImage}}, \code{\link[GiottoClass]{getMultiomics}}, \code{\link[GiottoClass]{getNearestNetwork}}, \code{\link[GiottoClass]{getPolygonInfo}}, \code{\link[GiottoClass]{getSpatialEnrichment}}, \code{\link[GiottoClass]{getSpatialEnrichment}}, \code{\link[GiottoClass]{getSpatialGrid}}, \code{\link[GiottoClass]{getSpatialGrid}}, \code{\link[GiottoClass]{getSpatialLocations}}, \code{\link[GiottoClass]{getSpatialLocations}}, \code{\link[GiottoClass]{getSpatialNetwork}}, \code{\link[GiottoClass]{getSpatialNetwork}}, \code{\link[GiottoClass:giotto-class]{giotto}}, \code{\link[GiottoClass:giottoImage-class]{giottoImage}}, \code{\link[GiottoClass:giottoLargeImage-class]{giottoLargeImage}}, \code{\link[GiottoClass]{giottoMasterToSuite}}, \code{\link[GiottoClass:giottoPoints-class]{giottoPoints}}, \code{\link[GiottoClass:giottoPolygon-class]{giottoPolygon}}, \code{\link[GiottoClass]{giottoToAnnData}}, \code{\link[GiottoClass]{giottoToSeurat}}, \code{\link[GiottoClass]{giottoToSeuratV4}}, \code{\link[GiottoClass]{giottoToSeuratV5}}, \code{\link[GiottoClass]{giottoToSpatialExperiment}}, \code{\link[GiottoClass]{hexVertices}}, \code{\link[GiottoClass]{hist}}, \code{\link[GiottoClass:giotto_python]{installGiottoEnvironment}}, \code{\link[GiottoClass:giotto_instructions]{instructions}}, \code{\link[GiottoClass:giotto_instructions]{instructions<-}}, \code{\link[GiottoClass]{joinGiottoObjects}}, \code{\link[GiottoClass]{loadGiotto}}, \code{\link[GiottoClass]{makePseudoVisium}}, \code{\link[GiottoClass]{objHistory}}, \code{\link[GiottoClass:giotto_schema]{objName}}, \code{\link[GiottoClass:giotto_schema]{objName<-}}, \code{\link[GiottoClass:generate_grid]{orthoGrid}}, \code{\link[GiottoClass]{overlapImagesToMatrix}}, \code{\link[GiottoClass:overlaps-generic]{overlaps}}, \code{\link[GiottoClass]{overlapToMatrix}}, \code{\link[GiottoClass]{overlapToMatrixMultiPoly}}, \code{\link[GiottoClass]{pDataDT}}, \code{\link[GiottoClass]{plotGiottoImage}}, \code{\link[GiottoClass]{polyStamp}}, \code{\link[GiottoClass:giotto_schema]{prov}}, \code{\link[GiottoClass:giotto_schema]{prov<-}}, \code{\link[GiottoClass]{readCellMetadata}}, \code{\link[GiottoClass]{readDimReducData}}, \code{\link[GiottoClass]{readExprData}}, \code{\link[GiottoClass]{readExprMatrix}}, \code{\link[GiottoClass]{readFeatData}}, \code{\link[GiottoClass]{readFeatMetadata}}, \code{\link[GiottoClass]{readGiottoInstructions}}, \code{\link[GiottoClass]{readNearestNetData}}, \code{\link[GiottoClass]{readPolygonData}}, \code{\link[GiottoClass]{readSpatEnrichData}}, \code{\link[GiottoClass]{readSpatLocsData}}, \code{\link[GiottoClass]{readSpatNetData}}, \code{\link[GiottoClass]{reconnectGiottoImage}}, \code{\link[GiottoClass]{rectVertices}}, \code{\link[GiottoClass]{removeCellAnnotation}}, \code{\link[GiottoClass]{removeFeatAnnotation}}, \code{\link[GiottoClass:giotto_python]{removeGiottoEnvironment}}, \code{\link[GiottoClass]{replaceGiottoInstructions}}, \code{\link[GiottoClass]{rescale}}, \code{\link[GiottoClass]{rescalePolygons}}, \code{\link[GiottoClass]{saveGiotto}}, \code{\link[GiottoClass]{setCellMetadata}}, \code{\link[GiottoClass]{setDimReduction}}, \code{\link[GiottoClass]{setExpression}}, \code{\link[GiottoClass]{setFeatureInfo}}, \code{\link[GiottoClass]{setFeatureMetadata}}, \code{\link[GiottoClass]{setGiotto}}, \code{\link[GiottoClass]{setGiottoImage}}, \code{\link[GiottoClass]{setMultiomics}}, \code{\link[GiottoClass]{setNearestNetwork}}, \code{\link[GiottoClass]{setPolygonInfo}}, \code{\link[GiottoClass]{setSpatialEnrichment}}, \code{\link[GiottoClass]{setSpatialGrid}}, \code{\link[GiottoClass]{setSpatialLocations}}, \code{\link[GiottoClass]{setSpatialNetwork}}, \code{\link[GiottoClass]{seuratToGiotto}}, \code{\link[GiottoClass]{seuratToGiottoV4}}, \code{\link[GiottoClass]{seuratToGiottoV5}}, \code{\link[GiottoClass]{showGiottoCellMetadata}}, \code{\link[GiottoClass]{showGiottoDimRed}}, \code{\link[GiottoClass]{showGiottoExpression}}, \code{\link[GiottoClass]{showGiottoFeatInfo}}, \code{\link[GiottoClass]{showGiottoFeatMetadata}}, \code{\link[GiottoClass]{showGiottoImageNames}}, \code{\link[GiottoClass]{showGiottoInstructions}}, \code{\link[GiottoClass]{showGiottoNearestNetworks}}, \code{\link[GiottoClass]{showGiottoSpatEnrichments}}, \code{\link[GiottoClass]{showGiottoSpatGrids}}, \code{\link[GiottoClass]{showGiottoSpatialInfo}}, \code{\link[GiottoClass]{showGiottoSpatLocs}}, \code{\link[GiottoClass]{showGiottoSpatNetworks}}, \code{\link[GiottoClass]{showProcessingSteps}}, \code{\link[GiottoClass]{smoothGiottoPolygons}}, \code{\link[GiottoClass]{spatialExperimentToGiotto}}, \code{\link[GiottoClass:spatIDs-generic]{spatIDs}}, \code{\link[GiottoClass]{spatQueryGiottoPolygons}}, \code{\link[GiottoClass]{spatShift}}, \code{\link[GiottoClass:giotto_schema]{spatUnit}}, \code{\link[GiottoClass:giotto_schema]{spatUnit<-}}, \code{\link[GiottoClass]{spin}}, \code{\link[GiottoClass]{stitchFieldCoordinates}}, \code{\link[GiottoClass]{stitchFieldCoordinates}}, \code{\link[GiottoClass]{stitchGiottoLargeImage}}, \code{\link[GiottoClass]{subsetGiotto}}, \code{\link[GiottoClass]{subsetGiottoLocs}}, \code{\link[GiottoClass]{subsetGiottoLocsMulti}}, \code{\link[GiottoClass]{subsetGiottoLocsSubcellular}}, \code{\link[GiottoClass]{tessellate}}, \code{\link[GiottoClass:generate_grid]{triGrid}}, \code{\link[GiottoClass]{updateGiottoImage}}, \code{\link[GiottoClass]{updateGiottoImageMG}}, \code{\link[GiottoClass]{updateGiottoLargeImage}}, \code{\link[GiottoClass]{updateGiottoObject}}, \code{\link[GiottoClass]{updateGiottoPointsObject}}, \code{\link[GiottoClass]{updateGiottoPolygonObject}}, \code{\link[GiottoClass:wrap]{vect}}, \code{\link[GiottoClass]{wrap}}, \code{\link[GiottoClass]{writeGiottoLargeImage}}}

  \item{GiottoUtils}{\code{\link[GiottoUtils:pipe]{\%>\%}}, \code{\link[GiottoUtils]{getDistinctColors}}, \code{\link[GiottoUtils]{getRainbowColors}}}

  \item{GiottoVisuals}{\code{\link[GiottoVisuals]{addGiottoImageToSpatPlot}}, \code{\link[GiottoVisuals]{dimCellPlot}}, \code{\link[GiottoVisuals:dimCellPlot]{dimCellPlot2D}}, \code{\link[GiottoVisuals]{dimFeatPlot2D}}, \code{\link[GiottoVisuals]{dimFeatPlot3D}}, \code{\link[GiottoVisuals:dimFeatPlot3D]{dimGenePlot3D}}, \code{\link[GiottoVisuals]{dimPlot}}, \code{\link[GiottoVisuals:dimPlot]{dimPlot2D}}, \code{\link[GiottoVisuals:dimPlot]{dimPlot3D}}, \code{\link[GiottoVisuals]{dotPlot}}, \code{\link[GiottoVisuals]{getColors}}, \code{\link[GiottoVisuals]{giottoSankeyPlan}}, \code{\link[GiottoVisuals]{plotHeatmap}}, \code{\link[GiottoVisuals]{plotMetaDataCellsHeatmap}}, \code{\link[GiottoVisuals]{plotMetaDataHeatmap}}, \code{\link[GiottoVisuals]{plotPCA}}, \code{\link[GiottoVisuals]{plotPCA_2D}}, \code{\link[GiottoVisuals]{plotPCA_3D}}, \code{\link[GiottoVisuals]{plotStatDelaunayNetwork}}, \code{\link[GiottoVisuals]{plotTSNE}}, \code{\link[GiottoVisuals]{plotTSNE_2D}}, \code{\link[GiottoVisuals]{plotTSNE_3D}}, \code{\link[GiottoVisuals]{plotUMAP}}, \code{\link[GiottoVisuals]{plotUMAP_2D}}, \code{\link[GiottoVisuals]{plotUMAP_3D}}, \code{\link[GiottoVisuals]{sankeyLabel}}, \code{\link[GiottoVisuals:sankeyLabel]{sankeyLabel<-}}, \code{\link[GiottoVisuals]{sankeyPlot}}, \code{\link[GiottoVisuals]{sankeyRelate}}, \code{\link[GiottoVisuals:sankeyRelate]{sankeyRelate<-}}, \code{\link[GiottoVisuals]{sankeySet}}, \code{\link[GiottoVisuals]{sankeySetAddresses}}, \code{\link[GiottoVisuals]{showClusterDendrogram}}, \code{\link[GiottoVisuals]{showClusterHeatmap}}, \code{\link[GiottoVisuals]{showColorInstructions}}, \code{\link[GiottoVisuals]{showSaveParameters}}, \code{\link[GiottoVisuals]{spatCellPlot}}, \code{\link[GiottoVisuals:spatCellPlot]{spatCellPlot2D}}, \code{\link[GiottoVisuals]{spatDeconvPlot}}, \code{\link[GiottoVisuals]{spatDimCellPlot}}, \code{\link[GiottoVisuals]{spatDimCellPlot2D}}, \code{\link[GiottoVisuals]{spatDimFeatPlot2D}}, \code{\link[GiottoVisuals]{spatDimFeatPlot3D}}, \code{\link[GiottoVisuals:spatDimFeatPlot3D]{spatDimGenePlot3D}}, \code{\link[GiottoVisuals]{spatDimPlot}}, \code{\link[GiottoVisuals:spatDimPlot]{spatDimPlot2D}}, \code{\link[GiottoVisuals]{spatDimPlot3D}}, \code{\link[GiottoVisuals]{spatFeatPlot2D}}, \code{\link[GiottoVisuals]{spatFeatPlot2D_single}}, \code{\link[GiottoVisuals]{spatFeatPlot3D}}, \code{\link[GiottoVisuals:spatFeatPlot3D]{spatGenePlot3D}}, \code{\link[GiottoVisuals]{spatInSituPlotDensity}}, \code{\link[GiottoVisuals]{spatInSituPlotHex}}, \code{\link[GiottoVisuals]{spatInSituPlotPoints}}, \code{\link[GiottoVisuals]{spatNetwDistributions}}, \code{\link[GiottoVisuals]{spatNetwDistributionsDistance}}, \code{\link[GiottoVisuals]{spatNetwDistributionsKneighbors}}, \code{\link[GiottoVisuals]{spatPlot}}, \code{\link[GiottoVisuals:spatPlot]{spatPlot2D}}, \code{\link[GiottoVisuals:spatPlot]{spatPlot3D}}, \code{\link[GiottoVisuals]{subsetSankeySet}}, \code{\link[GiottoVisuals]{violinPlot}}}
}}

