% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registration.R
\name{registerGiottoObjectList}
\alias{registerGiottoObjectList}
\title{registerGiottoObjectList}
\usage{
registerGiottoObjectList(
  gobject_list,
  image = "image",
  registered_image_name = "image",
  image_list = NULL,
  save_dir = NULL,
  spat_loc_values = NULL,
  spat_loc_name = "raw",
  method = c("fiji", "rvision"),
  fiji_xml_files,
  fiji_registered_images,
  scaling = 1,
  allow_rvision_autoscale = TRUE,
  auto_comp_reg_border = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{gobject_list}{list of gobjects to register}

\item{registered_image_name}{arbitrary slot name for registered images. Defaults to the slot of unregistered images.}

\item{spat_loc_values}{spatial locations to use}

\item{spat_loc_name}{arbitrary name for registered spatial locations. Defaults to replacement of spat_loc_values (optional)}

\item{method}{method used to align gobjects. Current options are either using FIJI register_virtual_stack_slices output or rvision}

\item{fiji_xml_files}{record of transformations performed in xml format exported during image registration using FIJI register_virtual_stack_slices}

\item{fiji_registered_images}{registered images output by FIJI register_virtual_stack_slices}

\item{scaling}{scaling to be applied to spatial coordinates}

\item{allow_rvision_autoscale}{Whether or not to allow rvision to automatically scale the images when performing image registration}

\item{verbose}{be verbose}
}
\value{
list of registered giotto objects where the registered images and spatial locations
}
\description{
Wrapper function for registerGiottoObjectListFiji and registerGiottoObjectListRvision
}
