% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{removeCellAnnotation}
\alias{removeCellAnnotation}
\title{removeCellAnnotation}
\usage{
removeCellAnnotation(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  columns = NULL,
  return_gobject = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit}

\item{feat_type}{feature type}

\item{columns}{names of columns to remove}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}
}
\value{
giotto object
}
\description{
removes cell annotation from a Giotto object for a specific feature modality (default = 'rna')
}
\details{
if return_gobject = FALSE, it will return the cell metadata
}
