% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{removeCellAnnotation}
\alias{removeCellAnnotation}
\title{removeCellAnnotation}
\usage{
removeCellAnnotation(gobject, columns = NULL, return_gobject = TRUE)
}
\arguments{
\item{gobject}{giotto object}

\item{columns}{names of columns to remove}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}
}
\value{
giotto object
}
\description{
removes cell annotation of giotto object
}
\details{
if return_gobject = FALSE, it will return the cell metadata
}
\examples{

data(mini_giotto_single_cell) # load full mini giotto object

# show cell metadata
pDataDT(mini_giotto_single_cell)

# remove cell_types column
mini_giotto_single_cell = removeCellAnnotation(mini_giotto_single_cell,
                                               columns = 'cell_types')

}
