% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{removeGeneAnnotation}
\alias{removeGeneAnnotation}
\title{removeGeneAnnotation}
\usage{
removeGeneAnnotation(gobject, columns = NULL, return_gobject = TRUE)
}
\arguments{
\item{gobject}{giotto object}

\item{columns}{names of columns to remove}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}
}
\value{
giotto object
}
\description{
removes gene annotation of giotto object
}
\details{
if return_gobject = FALSE, it will return the gene metadata
}
\examples{

data(mini_giotto_single_cell) # load full mini giotto object

# show gene metadata
fDataDT(mini_giotto_single_cell)

# remove nr_cells column
mini_giotto_single_cell = removeGeneAnnotation(mini_giotto_single_cell,
                                               columns = 'nr_cells')

}
