% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_environment.R
\name{removeGiottoEnvironment}
\alias{removeGiottoEnvironment}
\title{removeGiottoEnvironment}
\usage{
removeGiottoEnvironment(mini_path = NULL, verbose = TRUE)
}
\arguments{
\item{mini_path}{path to anaconda/miniconda.
Default: reticulate::miniconda_path()
i.e. "C:/my/conda/lives/here" OR "C:\\my\\conda\\lives\\here"}

\item{verbose}{be verbose}
}
\description{
removeGiottoEnvironment
}
\details{
Removes a previously installed giotto environment.
See \code{\link{installGiottoEnvironment}}.
}
