% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto.R
\name{replaceGiottoInstructions}
\alias{replaceGiottoInstructions}
\title{replaceGiottoInstructions}
\usage{
replaceGiottoInstructions(gobject, instructions = NULL)
}
\arguments{
\item{gobject}{giotto object}

\item{instructions}{new instructions (e.g. result from createGiottoInstructions)}
}
\value{
giotto object with replaces instructions
}
\description{
Function to replace all instructions from giotto object
}
