% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registration.R
\name{rotate_spatial_locations}
\alias{rotate_spatial_locations}
\title{rotate_spatial_locations}
\usage{
rotate_spatial_locations(spatlocs, rotateradians)
}
\arguments{
\item{spatlocs}{spatial locations to use}

\item{rotateradians}{radians by which the x and y values will be rotated in a counter clockwise manner}
}
\description{
Rotate given X Y coordinates by given radians in counter clockwise manner about the coordinate origin
}
\keyword{internal}
