% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_genes.R
\name{run_spatial_sim_tests_one_rep}
\alias{run_spatial_sim_tests_one_rep}
\title{run_spatial_sim_tests_one_rep}
\usage{
run_spatial_sim_tests_one_rep(
  gobject,
  pattern_name = "pattern",
  pattern_cell_ids = NULL,
  gene_name = NULL,
  spatial_prob = 0.95,
  show_pattern = FALSE,
  spatial_network_name = "kNN_network",
  spat_methods = c("binSpect_single", "binSpect_multi", "spatialDE", "spark",
    "silhouetteRank"),
  spat_methods_params = list(NA, NA, NA, NA, NA),
  spat_methods_names = c("binSpect_single", "binSpect_multi", "spatialDE", "spark",
    "silhouetteRank"),
  save_plot = FALSE,
  save_raw = FALSE,
  save_norm = FALSE,
  save_dir = "~",
  save_name = "plot",
  run_simulations = TRUE,
  ...
)
}
\description{
runs all spatial tests for 1 probability and 1 rep
}
\keyword{internal}
