% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_enrichment.R
\name{runDWLSDeconv}
\alias{runDWLSDeconv}
\title{runDWLSDeconv}
\usage{
runDWLSDeconv(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  expression_values = c("normalized"),
  logbase = 2,
  cluster_column = "leiden_clus",
  sign_matrix,
  n_cell = 50,
  cutoff = 2,
  name = NULL,
  return_gobject = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit}

\item{feat_type}{feature type}

\item{expression_values}{expression values to use}

\item{logbase}{base used for log normalization}

\item{cluster_column}{name of cluster column}

\item{sign_matrix}{sig matrix for deconvolution}

\item{n_cell}{number of cells per spot}

\item{cutoff}{cut off (default = 2)}

\item{name}{name to give to spatial deconvolution results, default = DWLS}

\item{return_gobject}{return giotto object}
}
\value{
giotto object or deconvolution results
}
\description{
Function to perform DWLS deconvolution based on single cell 
expression data
}
\examples{
g <- GiottoData::loadGiottoMini("visium")
x <- findMarkers_one_vs_all(g, 
cluster_column = "leiden_clus", min_feats = 20)
sign_gene <- x$feats

sign_matrix <- matrix(rnorm(length(sign_gene)*8, mean  = 10), 
nrow = length(sign_gene))
rownames(sign_matrix) <- sign_gene
colnames(sign_matrix) <- paste0("celltype_",unique(x$cluster))

runDWLSDeconv(gobject = g, sign_matrix = sign_matrix)
}
\seealso{
\url{https://github.com/dtsoucas/DWLS} for the \emph{DWLS} bulk 
deconvolution method, and \doi{10.1186/s13059-021-02362-7} for 
\emph{spatialDWLS}, the spatial implementation used here.
}
