% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_enrichment.R
\name{runDWLSDeconv}
\alias{runDWLSDeconv}
\title{runDWLSDeconv}
\usage{
runDWLSDeconv(
  gobject,
  expression_values = c("normalized"),
  logbase = 2,
  cluster_column = "leiden_clus",
  sign_matrix,
  n_cell = 50,
  cutoff = 2,
  name = NULL,
  return_gobject = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{expression_values}{expression values to use}

\item{logbase}{base used for log normalization}

\item{cluster_column}{name of cluster column}

\item{sign_matrix}{sig matrix for deconvolution}

\item{n_cell}{number of cells per spot}

\item{cutoff}{cut off (default = 2)}

\item{name}{name to give to spatial deconvolution results, default = DWLS}

\item{return_gobject}{return giotto object}
}
\value{
giotto object or deconvolution results
}
\description{
Function to perform DWLS deconvolution based on single cell expression data
}
