% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registration.R
\name{fiji}
\alias{fiji}
\title{Find Fiji location}
\usage{
fiji(fijiPath = NULL)
}
\arguments{
\item{fijiPath}{manually set filepath to Fiji executable}
}
\value{
character
}
\description{
\code{fiji} returns path to preferred Fiji executable. \cr
  This function is modified from jimpipeline by jefferislab
}
\examples{
\dontrun{
# Path to current Fiji executable
fiji()

# Set path to preferred Fiji executable (this will be remembered)
# you can also set options(giotto.fiji="/some/path")
fiji("/Applications/Fiji.app/Contents/MacOS/ImageJ-macosx")
}
}
