% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnn.R
\name{runIntegratedUMAP}
\alias{runIntegratedUMAP}
\title{Run integrated UMAP}
\usage{
runIntegratedUMAP(
  gobject,
  spat_unit = "cell",
  modality1 = "rna",
  modality2 = "protein",
  k = 20,
  spread = 5,
  min_dist = 0.01,
  ...
)
}
\arguments{
\item{gobject}{A giotto object}

\item{spat_unit}{spatial unit}

\item{modality1}{modality 1 name. Default = "rna"}

\item{modality2}{modality 2 name. Default = "protein"}

\item{k}{k number}

\item{spread}{UMAP param: spread}

\item{min_dist}{UMAP param: min_dist}

\item{...}{additional UMAP parameters}
}
\value{
A Giotto object with integrated UMAP
}
\description{
Run integrated UMAP
}
