% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONTraC_wrapper.R
\name{runONTraCV1}
\alias{runONTraCV1}
\title{runONTraCV1}
\usage{
runONTraCV1(
  ONTraC_input,
  dataset,
  preprocessing_dir,
  GNN_dir,
  NTScore_dir,
  n_cpu = 4L,
  n_neighbors = 50L,
  n_local = 20L,
  device = c("cpu", "cuda"),
  epochs = 1000L,
  patience = 100L,
  min_delta = 0.001,
  min_epochs = 50L,
  batch_size = 0L,
  seed = 42L,
  lr = 0.03,
  hidden_feats = 4L,
  k = 6L,
  modularity_loss_weight = 0.3,
  purity_loss_weight = 300,
  regularization_loss_weight = 0.1,
  beta = 0.03,
  envname = "giotto_ontrac_env",
  shell = FALSE
)
}
\arguments{
\item{dataset}{the path to the input data file}

\item{preprocessing_dir}{the directory to save the preprocessing results}

\item{GNN_dir}{the directory to save the GNN results}

\item{NTScore_dir}{the directory to save the NTScore results}

\item{envname, }{the name of the conda environment. Default is
"giotto_ontrac_env"}
}
\value{
none
}
\description{
run ONTraC
}
\details{
This function runs ONTraC
}
\examples{
runONTraCV1(
    dataset = "ONTraC_dataset_input.csv",
    preprocessing_dir = "preprocessing_dir",
    GNN_dir = "GNN_dir",
    NTScore_dir = "NTScore_dir",
    envname = "giotto_ontrac_env"
)
}
