% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.R
\name{runPCA}
\alias{runPCA}
\title{runPCA}
\usage{
runPCA(
  gobject,
  expression_values = c("normalized", "scaled", "custom"),
  reduction = c("cells", "genes"),
  name = "pca",
  genes_to_use = NULL,
  return_gobject = TRUE,
  center = F,
  scale_unit = F,
  ncp = 100,
  ...
)
}
\arguments{
\item{gobject}{giotto object}

\item{expression_values}{expression values to use}

\item{reduction}{cells or genes}

\item{name}{arbitrary name for PCA run}

\item{genes_to_use}{subset of genes to use for PCA}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}

\item{center}{center data first (default = FALSE)}

\item{scale_unit}{scale features before PCA}

\item{ncp}{number of principal components to calculate}

\item{...}{additional parameters for PCA (see details)}
}
\value{
giotto object with updated PCA dimension recuction
}
\description{
runs a Principal Component Analysis
}
\details{
See \code{\link[FactoMineR]{PCA}} for more information about other parameters.
}
\examples{
    runPCA(gobject)
}
