% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.R
\name{runPCA}
\alias{runPCA}
\title{runPCA}
\usage{
runPCA(
  gobject,
  expression_values = c("normalized", "scaled", "custom"),
  reduction = c("cells", "genes"),
  name = "pca",
  genes_to_use = "hvg",
  return_gobject = TRUE,
  center = F,
  scale_unit = F,
  ncp = 100,
  method = c("irlba", "factominer"),
  rev = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{gobject}{giotto object}

\item{expression_values}{expression values to use}

\item{reduction}{cells or genes}

\item{name}{arbitrary name for PCA run}

\item{genes_to_use}{subset of genes to use for PCA}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}

\item{center}{center data first (default = FALSE)}

\item{scale_unit}{scale features before PCA}

\item{ncp}{number of principal components to calculate}

\item{method}{which implementation to use}

\item{rev}{do a reverse PCA}

\item{verbose}{verbosity of the function}

\item{...}{additional parameters for PCA (see details)}
}
\value{
giotto object with updated PCA dimension recuction
}
\description{
runs a Principal Component Analysis
}
\details{
See \code{\link[irlba]{prcomp_irlba}} and \code{\link[FactoMineR]{PCA}} for more information about other parameters.
\itemize{
  \item genes_to_use = NULL: will use all genes from the selected matrix
  \item genes_to_use = <hvg name>: can be used to select a column name of
  highly variable genes (see \code{\link{calculateHVG}})
  \item genes_to_use = c('geneA', 'geneB', ...): will use all manually provided genes
}
}
\examples{
    runPCA(gobject)
}
