% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.R
\name{runPCA}
\alias{runPCA}
\title{runPCA}
\usage{
runPCA(gobject, expression_values = c("normalized", "scaled", "custom"),
  reduction = c("cells", "genes"), name = "pca", genes_to_use = NULL,
  return_gobject = TRUE, scale.unit = F, ncp = 200, ...)
}
\arguments{
\item{gobject}{giotto object}

\item{expression_values}{expression values to use}

\item{reduction}{cells or genes}

\item{name}{arbitrary name for PCA run}

\item{genes_to_use}{subset of genes to use for PCA}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}

\item{scale.unit}{scale features before PCA}

\item{ncp}{number of principal components to calculate}

\item{...}{additional parameters for PCA}
}
\value{
giotto object with updated PCA dimension recuction
}
\description{
run PCA
}
\details{
Description of PCA steps...
}
\examples{
    runPCA(gobject)
}
