% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.R
\name{runPCA}
\alias{runPCA}
\title{runPCA}
\usage{
runPCA(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  expression_values = c("normalized", "scaled", "custom"),
  reduction = c("cells", "feats"),
  name = NULL,
  feats_to_use = "hvf",
  return_gobject = TRUE,
  center = TRUE,
  scale_unit = TRUE,
  ncp = 100,
  method = c("irlba", "exact", "random", "factominer"),
  method_params = BiocParallel::SerialParam(),
  rev = FALSE,
  set_seed = TRUE,
  seed_number = 1234,
  verbose = TRUE,
  toplevel = 1L,
  ...
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit}

\item{feat_type}{feature type}

\item{expression_values}{expression values to use}

\item{reduction}{cells or genes}

\item{name}{arbitrary name for PCA run}

\item{feats_to_use}{subset of features to use for PCA}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}

\item{center}{center data first (default = TRUE)}

\item{scale_unit}{scale features before PCA (default = TRUE)}

\item{ncp}{number of principal components to calculate (see details)}

\item{method}{which implementation to use}

\item{method_params}{BiocParallelParam object}

\item{rev}{do a reverse PCA}

\item{set_seed}{use of seed}

\item{seed_number}{seed number to use}

\item{verbose}{verbosity of the function}

\item{toplevel}{relative stackframe where call was made}

\item{...}{additional parameters for PCA (see details)}
}
\value{
giotto object with updated PCA dimension recuction
}
\description{
runs a Principal Component Analysis
}
\details{
See \code{\link[BiocSingular]{runPCA}} and
\code{\link[FactoMineR]{PCA}} for more information about other parameters.
With the feats_to_use param, you can control which features are used to
calculate your PCA, which can be useful for making sure that your downstream
dimension reduction and clusterings are based on your features of interest.
\itemize{
  \item feats_to_use = NULL: will use all features from the selected matrix
  \item feats_to_use = <hvg name>: can be used to select a column name of
  highly variable features, created by (see \code{\link{calculateHVF}})
  \item feats_to_use = c('geneA', 'geneB', ...): will use all manually
  provided features
}
By default the number of principle components that we calculate is 100, which
may not encompass all the variation within the dataset. Setting ncp to NULL
will calculate all the principle components.
}
\examples{
g <- GiottoData::loadGiottoMini("visium")

runPCA(g)
}
