% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.R
\name{runPCA}
\alias{runPCA}
\title{runPCA}
\usage{
runPCA(
  gobject,
  feat_type = NULL,
  expression_values = c("normalized", "scaled", "custom"),
  reduction = c("cells", "feats"),
  name = NULL,
  feats_to_use = "hvf",
  genes_to_use = NULL,
  return_gobject = TRUE,
  center = TRUE,
  scale_unit = TRUE,
  ncp = 100,
  method = c("irlba", "factominer"),
  rev = FALSE,
  set_seed = TRUE,
  seed_number = 1234,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{gobject}{giotto object}

\item{feat_type}{feature type}

\item{expression_values}{expression values to use}

\item{reduction}{cells or genes}

\item{name}{arbitrary name for PCA run}

\item{feats_to_use}{subset of features to use for PCA}

\item{genes_to_use}{deprecated use feats_to_use}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}

\item{center}{center data first (default = TRUE)}

\item{scale_unit}{scale features before PCA (default = TRUE)}

\item{ncp}{number of principal components to calculate}

\item{method}{which implementation to use}

\item{rev}{do a reverse PCA}

\item{set_seed}{use of seed}

\item{seed_number}{seed number to use}

\item{verbose}{verbosity of the function}

\item{...}{additional parameters for PCA (see details)}
}
\value{
giotto object with updated PCA dimension recuction
}
\description{
runs a Principal Component Analysis
}
\details{
See \code{\link[irlba]{prcomp_irlba}} and \code{\link[FactoMineR]{PCA}} for more information about other parameters.
\itemize{
  \item feats_to_use = NULL: will use all features from the selected matrix
  \item feats_to_use = <hvg name>: can be used to select a column name of
  highly variable features, created by (see \code{\link{calculateHVF}})
  \item feats_to_use = c('geneA', 'geneB', ...): will use all manually provided features
}
}
\examples{

data(mini_giotto_single_cell)

# run PCA
mini_giotto_single_cell <- runPCA(gobject = mini_giotto_single_cell,
                                  center = TRUE, scale_unit = TRUE)

# plot PCA results
plotPCA(mini_giotto_single_cell)

}
