% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.R
\name{runPCA_factominer}
\alias{runPCA_factominer}
\title{runPCA_factominer}
\usage{
runPCA_factominer(
  x,
  ncp = 100,
  scale = TRUE,
  rev = FALSE,
  set_seed = TRUE,
  seed_number = 1234,
  ...
)
}
\arguments{
\item{x}{matrix or object that can be converted to matrix}

\item{ncp}{number of principal components to calculate}

\item{scale}{scale features}

\item{rev}{reverse PCA}

\item{set_seed}{use of seed}

\item{seed_number}{seed number to use}
}
\value{
list of eigenvalues, loadings and pca coordinates
}
\description{
performs PCA based on the factominer package
}
\keyword{internal}
