% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_enrichment.R
\name{runSpatialDeconv}
\alias{runSpatialDeconv}
\title{runSpatialDeconv}
\usage{
runSpatialDeconv(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  deconv_method = c("DWLS"),
  expression_values = c("normalized"),
  logbase = 2,
  cluster_column = "leiden_clus",
  sign_matrix,
  n_cell = 50,
  cutoff = 2,
  name = NULL,
  return_gobject = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit}

\item{feat_type}{feature type}

\item{deconv_method}{method to use for deconvolution}

\item{expression_values}{expression values to use}

\item{logbase}{base used for log normalization}

\item{cluster_column}{name of cluster column}

\item{sign_matrix}{signature matrix for deconvolution}

\item{n_cell}{number of cells per spot}

\item{cutoff}{cut off (default = 2)}

\item{name}{name to give to spatial deconvolution results}

\item{return_gobject}{return giotto object}
}
\value{
giotto object or deconvolution results
}
\description{
Function to perform deconvolution based on single cell expression data
}
\seealso{
\url{https://github.com/dtsoucas/DWLS} for the \emph{DWLS} bulk deconvolution method,
and \url{https://doi.org/10.1186/s13059-021-02362-7} for \emph{spatialDWLS}, the spatial implementation used here.
}
