% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_enrichment.R
\name{runSpatialEnrich}
\alias{runSpatialEnrich}
\title{runSpatialEnrich}
\usage{
runSpatialEnrich(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  enrich_method = c("PAGE", "rank", "hypergeometric"),
  sign_matrix,
  expression_values = c("normalized", "scaled", "custom"),
  min_overlap_genes = 5,
  reverse_log_scale = TRUE,
  logbase = 2,
  p_value = FALSE,
  include_depletion = FALSE,
  ties_method = c("average", "max"),
  n_times = 1000,
  rbp_p = 0.99,
  num_agg = 100,
  max_block = 2e+07,
  top_percentage = 5,
  output_enrichment = c("original", "zscore"),
  name = NULL,
  verbose = TRUE,
  return_gobject = TRUE
)
}
\arguments{
\item{gobject}{Giotto object}

\item{spat_unit}{spatial unit}

\item{feat_type}{feature type}

\item{enrich_method}{method for gene signature enrichment calculation}

\item{sign_matrix}{Matrix of signature genes for each cell type / process}

\item{expression_values}{expression values to use}

\item{min_overlap_genes}{minimum number of overlapping genes in sign_matrix
required to calculate enrichment (PAGE)}

\item{reverse_log_scale}{reverse expression values from log scale}

\item{logbase}{log base to use if reverse_log_scale = TRUE}

\item{p_value}{calculate p-value (default = FALSE)}

\item{include_depletion}{(PAGE) also test for depletion, not enrichment
only (default = FALSE)}

\item{ties_method}{(rank) how to rank tied expression values, \code{"average"}
(default) or \code{"max"}}

\item{n_times}{(page/rank) number of permutation iterations to calculate
p-value}

\item{rbp_p}{(rank) fractional binarization threshold (default = 0.99)}

\item{num_agg}{(rank) number of top genes to aggregate (default = 100)}

\item{max_block}{number of lines to process together (default = 20e6)}

\item{top_percentage}{(hyper) percentage of cells that will be considered
to have gene expression with matrix binarization}

\item{output_enrichment}{how to return enrichment output}

\item{name}{to give to spatial enrichment results, default = PAGE}

\item{verbose}{be verbose}

\item{return_gobject}{return giotto object}
}
\value{
Giotto object or enrichment results if return_gobject = FALSE
}
\description{
Function to calculate gene signature enrichment scores per
spatial position using an enrichment test.
}
\details{
For details see the individual functions:
\itemize{
\item \strong{PAGE:} \code{\link{runPAGEEnrich}}
\item \strong{Rank:} \code{\link{runRankEnrich}}
\item \strong{Hypergeometric:} \code{\link{runHyperGeometricEnrich}}
}
}
\examples{
g <- GiottoData::loadGiottoMini("visium")
x <- findMarkers_one_vs_all(g,
    cluster_column = "leiden_clus", min_feats = 20
)
sign_gene <- x$feats

sign_matrix <- matrix(rnorm(length(sign_gene) * 8, mean = 10),
    nrow = length(sign_gene)
)
rownames(sign_matrix) <- sign_gene
colnames(sign_matrix) <- paste0("celltype_", unique(x$cluster))

runSpatialEnrich(gobject = g, sign_matrix = sign_matrix)
}
