% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_enrichment.R
\name{runSpatialEnrich}
\alias{runSpatialEnrich}
\title{runSpatialEnrich}
\usage{
runSpatialEnrich(
  gobject,
  enrich_method = c("PAGE", "rank", "hypergeometric"),
  sign_matrix,
  expression_values = c("normalized", "scaled", "custom"),
  reverse_log_scale = TRUE,
  logbase = 2,
  p_value = FALSE,
  n_times = 1000,
  top_percentage = 5,
  output_enrichment = c("original", "zscore"),
  name = NULL,
  return_gobject = TRUE
)
}
\arguments{
\item{gobject}{Giotto object}

\item{enrich_method}{method for gene signature enrichment calculation}

\item{sign_matrix}{Matrix of signature genes for each cell type / process}

\item{expression_values}{expression values to use}

\item{reverse_log_scale}{reverse expression values from log scale}

\item{logbase}{log base to use if reverse_log_scale = TRUE}

\item{p_value}{calculate p-value (default = FALSE)}

\item{n_times}{(page/rank) number of permutation iterations to calculate p-value}

\item{top_percentage}{(hyper) percentage of cells that will be considered to have gene expression with matrix binarization}

\item{output_enrichment}{how to return enrichment output}

\item{name}{to give to spatial enrichment results, default = PAGE}

\item{return_gobject}{return giotto object}
}
\value{
Giotto object or enrichment results if return_gobject = FALSE
}
\description{
Function to calculate gene signature enrichment scores per spatial position using an enrichment test.
}
\details{
For details see the individual functions:
\itemize{
  \item{PAGE: }{\code{\link{runPAGEEnrich}}}
  \item{PAGE: }{\code{\link{runRankEnrich}}}
  \item{PAGE: }{\code{\link{runHyperGeometricEnrich}}}
}
}
