% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.R
\name{runtSNE}
\alias{runtSNE}
\title{runtSNE}
\usage{
runtSNE(
  gobject,
  expression_values = c("normalized", "scaled", "custom"),
  reduction = c("cells", "genes"),
  dim_reduction_to_use = "pca",
  dim_reduction_name = "pca",
  dimensions_to_use = 1:10,
  name = "tsne",
  genes_to_use = NULL,
  return_gobject = TRUE,
  dims = 2,
  perplexity = 30,
  theta = 0.5,
  do_PCA_first = F,
  set_seed = T,
  seed_number = 1234,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{gobject}{giotto object}

\item{expression_values}{expression values to use}

\item{reduction}{cells or genes}

\item{dim_reduction_to_use}{use another dimension reduction set as input}

\item{dim_reduction_name}{name of dimension reduction set to use}

\item{dimensions_to_use}{number of dimensions to use as input}

\item{name}{arbitrary name for tSNE run}

\item{genes_to_use}{if dim_reduction_to_use = NULL, which genes to use}

\item{return_gobject}{boolean: return giotto object (default = TRUE)}

\item{dims}{tSNE param: number of dimensions to return}

\item{perplexity}{tSNE param: perplexity}

\item{theta}{tSNE param: theta}

\item{do_PCA_first}{tSNE param: do PCA before tSNE (default = FALSE)}

\item{set_seed}{use of seed}

\item{seed_number}{seed number to use}

\item{verbose}{verbosity of the function}

\item{...}{additional tSNE parameters}
}
\value{
giotto object with updated tSNE dimension recuction
}
\description{
run tSNE
}
\details{
See \code{\link[Rtsne]{Rtsne}} for more information about these and other parameters. \cr
\itemize{
  \item Input for tSNE dimension reduction can be another dimension reduction (default = 'pca')
  \item To use gene expression as input set dim_reduction_to_use = NULL
  \item If dim_reduction_to_use = NULL, genes_to_use can be used to select a column name of
  highly variable genes (see \code{\link{calculateHVG}}) or simply provide a vector of genes
  \item multiple tSNE results can be stored by changing the \emph{name} of the analysis
}
}
\examples{

data(mini_giotto_single_cell)

mini_giotto_single_cell <- runtSNE(mini_giotto_single_cell,
                                   dimensions_to_use = 1:3,
                                   n_threads = 1,
                                   n_neighbors = 3,
                                   perplexity = 1)

plotTSNE(gobject = mini_giotto_single_cell)

}
