% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnn.R
\name{runWNN}
\alias{runWNN}
\title{Multi omics integration with WNN}
\usage{
runWNN(
  gobject,
  spat_unit = "cell",
  modality_1 = "rna",
  modality_2 = "protein",
  pca_name_modality_1 = "rna.pca",
  pca_name_modality_2 = "protein.pca",
  k = 20,
  integrated_feat_type = NULL,
  matrix_result_name = NULL,
  w_name_modality_1 = NULL,
  w_name_modality_2 = NULL,
  verbose = FALSE
)
}
\arguments{
\item{gobject}{A Giotto object with individual PCA modalities pre-calculated}

\item{spat_unit}{spatial unit}

\item{modality_1}{modality 1 name. Default = "rna"}

\item{modality_2}{modality 2 name. Default = "protein"}

\item{pca_name_modality_1}{Default = 'rna.pca'}

\item{pca_name_modality_2}{Default = 'protein.pca'}

\item{k}{k number, default = 20}

\item{integrated_feat_type}{integrated feature type (e.g. 'rna_protein')}

\item{matrix_result_name}{Default = 'theta_weighted_matrix'}

\item{w_name_modality_1}{name for modality 1 weights}

\item{w_name_modality_2}{name for modality 2 weights}

\item{verbose}{be verbose}
}
\value{
A Giotto object with integrated UMAP (integrated.umap) within the dimension_reduction slot and Leiden clusters (integrated_leiden_clus) in the cellular metadata.
}
\description{
Multi omics integration with WNN
}
