% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnn.R
\name{runWNN}
\alias{runWNN}
\title{Multi omics integration with WNN}
\usage{
runWNN(
  gobject,
  spat_unit = "cell",
  modality_1 = "modality1",
  modality_2 = "modality2",
  k = 20
)
}
\arguments{
\item{gobject}{A Giotto object with individual PCA modalities pre-calculated}

\item{spat_unit}{spatial unit}

\item{modality_1}{modality 1 name. Default = "rna"}

\item{modality_2}{modality 2 name. Default = "protein"}

\item{k}{k number, default = 20}
}
\value{
A Giotto object with integrated UMAP (integrated.umap) within the dimension_reduction slot and Leiden clusters (integrated_leiden_clus) in the cellular metadata.
}
\description{
Multi omics integration with WNN
}
