% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_hmrf.R
\name{sampling_sp_genes}
\alias{sampling_sp_genes}
\title{sampling_sp_genes}
\usage{
sampling_sp_genes(clust, sample_rate = 2, target = 500, seed = 10)
}
\arguments{
\item{seed}{}
}
\description{
function to select a set of spatial genes
}
\keyword{internal}
