% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_hmrf.R
\name{sampling_sp_genes}
\alias{sampling_sp_genes}
\title{sampling_sp_genes}
\usage{
sampling_sp_genes(clust, sample_rate = 2, target = 500, seed = 10)
}
\arguments{
\item{clust}{spatial gene clusters}

\item{sample_rate}{sampling rate, takes values equal or greater than 1}

\item{target}{target length of gene list}

\item{seed}{random seed}
}
\value{
list
}
\description{
function to select a set of spatial genes
}
\details{
This function samples a subset of spatial genes among different clusters,
with size n = target.
Number of samples from each cluster denpends on the relative proportion of
each cluster.
Changing from equal size by setting sample_rate = 1 to with exact proportion
of each cluster by setting sample_rate = +Inf
}
\keyword{internal}
