% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registration.R
\name{scale_spatial_locations}
\alias{scale_spatial_locations}
\title{scale_spatial_locations}
\usage{
scale_spatial_locations(spatlocs, scalefactor)
}
\arguments{
\item{spatlocs}{spatial locations to scale}

\item{scalefactor}{scaling factor to apply to coordinates}
}
\description{
Scale the X and Y coordinates given by the scale factor input
}
\keyword{internal}
