% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.R
\name{screePlot}
\alias{screePlot}
\title{screePlot}
\usage{
screePlot(
  gobject,
  name = "pca",
  expression_values = c("normalized", "scaled", "custom"),
  reduction = c("cells", "genes"),
  genes_to_use = NULL,
  center = F,
  scale_unit = F,
  ncp = 100,
  ylim = c(0, 20),
  verbose = T,
  show_plot = NA,
  return_plot = NA,
  save_plot = NA,
  save_param = list(),
  default_save_name = "screePlot"
)
}
\arguments{
\item{gobject}{giotto object}

\item{name}{name of PCA object if available}

\item{expression_values}{expression values to use}

\item{reduction}{cells or genes}

\item{genes_to_use}{subset of genes to use for PCA}

\item{center}{center data before PCA}

\item{scale_unit}{scale features before PCA}

\item{ncp}{number of principal components to calculate}

\item{ylim}{y-axis limits on scree plot}

\item{verbose}{verobsity}

\item{show_plot}{show plot}

\item{return_plot}{return ggplot object}

\item{save_plot}{directly save the plot [boolean]}

\item{save_param}{list of saving parameters from all_plots_save_function()}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}
}
\value{
ggplot object for scree method
}
\description{
identify significant prinicipal components (PCs) using an screeplot (a.k.a. elbowplot)
}
\details{
Screeplot works by plotting the explained variance of each
 individual PC in a barplot allowing you to identify which PC does not show a significant
 contribution anymore ( = 'elbow method'). \cr
}
\examples{
    screePlot(gobject)
}
