% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{select_dimReduction}
\alias{select_dimReduction}
\title{select_dimReduction}
\usage{
select_dimReduction(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=get_dimReduction]{get_dimReduction}}
  \describe{
    \item{\code{gobject}}{giotto object}
    \item{\code{spat_unit}}{spatial unit (e.g. "cell")}
    \item{\code{feat_type}}{feature type (e.g. "rna", "dna", "protein")}
    \item{\code{reduction}}{reduction on cells or features (e.g. "cells", "feats")}
    \item{\code{reduction_method}}{reduction method (e.g. "pca", "umap", "tsne")}
    \item{\code{name}}{name of reduction results}
    \item{\code{return_dimObj}}{return full dimension object result. Default = FALSE}
  }}
}
\description{
select_dimReduction
}
\seealso{
\code{\link{get_dimReduction}}
}
\keyword{internal}
