% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.R
\name{select_dimReduction}
\alias{select_dimReduction}
\title{select_dimReduction}
\usage{
select_dimReduction(
  gobject,
  reduction = c("cells", "genes"),
  reduction_method = c("pca", "umap", "tsne"),
  name = "pca",
  return_dimObj = FALSE
)
}
\value{
dim reduction coordinates (default) or dim reduction object
}
\description{
Creates an object that stores a dimension reduction output
}
\keyword{internal}
