% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{select_expression_values}
\alias{select_expression_values}
\title{select_expression_values}
\usage{
select_expression_values(gobject, values)
}
\arguments{
\item{gobject}{giotto object}

\item{values}{expression values to extract}
}
\value{
expression matrix
}
\description{
helper function to select expression values
}
\keyword{internal}
