% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{select_feature_info}
\alias{select_feature_info}
\title{Select feature info}
\usage{
select_feature_info(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=get_feature_info]{get_feature_info}}
  \describe{
    \item{\code{gobject}}{giotto object}
    \item{\code{feat_type}}{name of feature (e.g. "rna", "dna", "protein")}
  }}
}
\description{
Select feature info
}
\seealso{
\code{\link{get_feature_info}}
}
\keyword{internal}
