% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{select_spatial_locations}
\alias{select_spatial_locations}
\title{select_spatial_locations}
\usage{
select_spatial_locations(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=get_spatial_locations]{get_spatial_locations}}
  \describe{
    \item{\code{spat_loc_name}}{name of spatial locations (defaults to first name in spatial_locs slot, e.g. "raw")}
    \item{\code{output}}{what object type to get the spatial locations as. Default is as
a 'spatLocsObj'. Returning as 'data.table' is also possible.}
    \item{\code{copy_obj}}{whether to copy/duplicate when getting the object (default = TRUE)}
    \item{\code{verbose}}{be verbose}
    \item{\code{gobject}}{giotto object}
    \item{\code{spat_unit}}{spatial unit (e.g. "cell")}
    \item{\code{set_defaults}}{set default spat_unit and feat_type. Change to FALSE only when}
  }}
}
\description{
select_spatial_locations
}
\seealso{
\code{\link{get_spatial_locations}}
}
\keyword{internal}
