% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_structures.R
\name{select_spatialGrid}
\alias{select_spatialGrid}
\title{select_spatialGrid}
\usage{
select_spatialGrid(gobject, name = NULL, return_grid_Obj = FALSE)
}
\description{
accessor function to select spatial grid
}
\keyword{internal}
