% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{select_spatialNetwork}
\alias{select_spatialNetwork}
\title{Select spatial network}
\usage{
select_spatialNetwork(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=get_spatialNetwork]{get_spatialNetwork}}
  \describe{
    \item{\code{gobject}}{giotto object}
    \item{\code{spat_unit}}{spatial unit (e.g. "cell")}
    \item{\code{name}}{name of spatial network}
    \item{\code{return_network_Obj}}{return network object (default = FALSE)}
  }}
}
\description{
Select spatial network
}
\seealso{
\code{\link{get_spatialNetwork}}
}
\keyword{internal}
