% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_structures.R
\name{select_spatialNetwork}
\alias{select_spatialNetwork}
\title{select_spatialNetwork}
\usage{
select_spatialNetwork(gobject, name = NULL, return_network_Obj = FALSE)
}
\description{
function to select a spatial network
}
\keyword{internal}
