% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{set_cell_metadata}
\alias{set_cell_metadata}
\title{Set cell metadata}
\usage{
set_cell_metadata(
  gobject,
  metadata,
  spat_unit = NULL,
  feat_type = NULL,
  provenance = NULL,
  verbose = TRUE,
  set_defaults = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{metadata}{cellMetaObj or data.table containing cell metadata. Setting NULL will remove
the object. Passing 'initialize' will reset the object.}

\item{spat_unit}{spatial unit (e.g. "cell")}

\item{feat_type}{feature type (e.g. "rna", "dna", "protein")}

\item{provenance}{provenance information to set}

\item{verbose}{be verbose}

\item{set_defaults}{set default spat_unit and feat_type. Change to FALSE only when}
}
\value{
giotto object
}
\description{
Function to set cell metadata information into giotto object
}
\seealso{
Other functions to set data in giotto object: 
\code{\link{get_cell_id}()},
\code{\link{get_feat_id}()},
\code{\link{set_NearestNetwork}()},
\code{\link{set_cell_id}()},
\code{\link{set_dimReduction}()},
\code{\link{set_expression_values}()},
\code{\link{set_feat_id}()},
\code{\link{set_feature_info}()},
\code{\link{set_feature_metadata}()},
\code{\link{set_giottoImage}()},
\code{\link{set_multiomics}()},
\code{\link{set_polygon_info}()},
\code{\link{set_spatialGrid}()},
\code{\link{set_spatialNetwork}()},
\code{\link{set_spatial_enrichment}()},
\code{\link{set_spatial_locations}()}
}
\concept{functions to set data in giotto object}
