% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{set_CellMetadata}
\alias{set_CellMetadata}
\title{Set cell metadata}
\usage{
set_CellMetadata(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  metadata,
  verbose = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit (e.g. "cell")}

\item{feat_type}{feature type (e.g. "rna", "dna", "protein")}

\item{metadata}{cellMetaObj or data.table containing cell metadata. Setting NULL will remove
the object}

\item{verbose}{be verbose}
}
\value{
giotto object
}
\description{
Function to set cell metadata information into giotto object
}
\seealso{
Other functions to set data in giotto object: 
\code{\link{get_cell_id}()},
\code{\link{get_feat_id}()},
\code{\link{set_FeatMetadata}()},
\code{\link{set_NearestNetwork}()},
\code{\link{set_cell_id}()},
\code{\link{set_dimReduction}()},
\code{\link{set_expression_values}()},
\code{\link{set_feat_id}()},
\code{\link{set_feature_info}()},
\code{\link{set_giottoImage}()},
\code{\link{set_polygon_info}()},
\code{\link{set_spatialGrid}()},
\code{\link{set_spatialNetwork}()},
\code{\link{set_spatial_enrichment}()},
\code{\link{set_spatial_locations}()}
}
\concept{functions to set data in giotto object}
