% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{set_default_feat_type}
\alias{set_default_feat_type}
\title{set_default_feat_type}
\usage{
set_default_feat_type(gobject, feat_type = NULL, spat_unit)
}
\arguments{
\item{gobject}{giotto object}

\item{feat_type}{feature type (e.g. "rna", "dna", "protein")}

\item{spat_unit}{spatial unit (e.g. "cell")}
}
\description{
set_default_feat_type
}
\keyword{internal}
