% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{set_dimReduction}
\alias{set_dimReduction}
\title{set_dimReduction}
\usage{
set_dimReduction(
  gobject,
  feat_type = NULL,
  spat_unit = NULL,
  reduction = c("cells", "genes"),
  reduction_method = c("pca", "umap", "tsne"),
  name = "pca",
  dimObject
)
}
\arguments{
\item{gobject}{giotto object}

\item{reduction}{reduction with cells or genes}

\item{reduction_method}{reduction name}

\item{name}{reduction name}

\item{dimObject}{dimension object}
}
\value{
giotto object
}
\description{
function to set a dimension reduction slot
}
