% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{set_dimReduction}
\alias{set_dimReduction}
\title{set_dimReduction}
\usage{
set_dimReduction(
  gobject,
  spat_unit = NULL,
  reduction = c("cells", "genes"),
  reduction_method = c("pca", "umap", "tsne"),
  name = "pca",
  dimObject
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit}

\item{reduction}{reduction on cells or features}

\item{reduction_method}{reduction method (e.g. pca)}

\item{name}{name of reduction results}

\item{dimObject}{dimension object result to set}
}
\value{
giotto object
}
\description{
function to set a dimension reduction slot
}
