% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{set_dimReduction}
\alias{set_dimReduction}
\title{Set dimension reduction}
\usage{
set_dimReduction(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  reduction = c("cells", "genes"),
  reduction_method = c("pca", "umap", "tsne"),
  name = "pca",
  dimObject
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit (e.g. "cell")}

\item{feat_type}{feature type (e.g. "rna", "dna", "protein")}

\item{reduction}{reduction on cells or features}

\item{reduction_method}{reduction method (e.g. "pca")}

\item{name}{name of reduction results}

\item{dimObject}{dimension object result to set}
}
\value{
giotto object
}
\description{
Function to set a dimension reduction slot
}
\seealso{
Other dimensional reduction data accessor functions: 
\code{\link{get_dimReduction}()}
}
\concept{dimensional reduction data accessor functions}
\concept{functions to set data in giotto object                                    dimObject = pca_object)}
