% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{set_dimReduction}
\alias{set_dimReduction}
\title{Set dimension reduction}
\usage{
set_dimReduction(
  gobject,
  dimObject,
  spat_unit = NULL,
  feat_type = NULL,
  reduction = c("cells", "genes"),
  reduction_method = c("pca", "umap", "tsne"),
  name = "pca",
  provenance = NULL,
  verbose = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{dimObject}{dimension object result to set}

\item{spat_unit}{spatial unit (e.g. "cell")}

\item{feat_type}{feature type (e.g. "rna", "dna", "protein")}

\item{reduction}{reduction on cells or features}

\item{reduction_method}{reduction method (e.g. "pca")}

\item{name}{name of reduction results}

\item{provenance}{provenance information (optional)}

\item{verbose}{be verbose}
}
\value{
giotto object
}
\description{
Function to set a dimension reduction slot
}
\seealso{
Other dimensional reduction data accessor functions: 
\code{\link{get_dimReduction}()}

Other functions to set data in giotto object: 
\code{\link{get_cell_id}()},
\code{\link{get_feat_id}()},
\code{\link{set_NearestNetwork}()},
\code{\link{set_cell_id}()},
\code{\link{set_cell_metadata}()},
\code{\link{set_expression_values}()},
\code{\link{set_feat_id}()},
\code{\link{set_feature_info}()},
\code{\link{set_feature_metadata}()},
\code{\link{set_giottoImage}()},
\code{\link{set_polygon_info}()},
\code{\link{set_spatialGrid}()},
\code{\link{set_spatialNetwork}()},
\code{\link{set_spatial_enrichment}()},
\code{\link{set_spatial_locations}()}
}
\concept{dimensional reduction data accessor functions}
\concept{functions to set data in giotto object}
