% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{set_expression_values}
\alias{set_expression_values}
\title{Set expression values}
\usage{
set_expression_values(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  name = "test",
  values
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit,  e.g. "cell"}

\item{feat_type}{feature type, e.g. "rna", "dna", "protein"}

\item{name}{name for the expression slot}

\item{values}{expression values}
}
\value{
giotto object
}
\description{
Function to set expression values for giotto object
}
\examples{
my_matrix <- matrix(rnorm(7486491), nrow = 11827, ncol = 633)
my_giotto_object <- set_expression_values(my_giotto_object,
                                          spat_unit = "cell",
                                          feat_type = "rna",
                                          name = "test",
                                          values = my_matrix)
}
\seealso{
Other expression accessor functions: 
\code{\link{get_expression_values}()}

Other functions to set data in giotto object: 
\code{\link{set_NearestNetwork}()},
\code{\link{set_dimReduction}()},
\code{\link{set_feature_info}()},
\code{\link{set_giottoImage}()},
\code{\link{set_polygon_info}()},
\code{\link{set_spatialGrid}()},
\code{\link{set_spatialNetwork}()},
\code{\link{set_spatial_enrichment}()},
\code{\link{set_spatial_locations}()}
}
\concept{expression accessor functions}
\concept{functions to set data in giotto object}
