% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{set_expression_values}
\alias{set_expression_values}
\title{Set expression values}
\usage{
set_expression_values(
  gobject,
  values,
  spat_unit = NULL,
  feat_type = NULL,
  name = "test",
  provenance = NULL,
  verbose = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{values}{exprObj or matrix of expression values. If NULL, then the object
will be removed.}

\item{spat_unit}{spatial unit (e.g. "cell")}

\item{feat_type}{feature type (e.g. "rna", "dna", "protein")}

\item{name}{name for the expression slot}

\item{provenance}{provenance information (optional)}

\item{verbose}{be verbose}
}
\value{
giotto object
}
\description{
Function to set expression values for giotto object
}
\seealso{
Other expression accessor functions: 
\code{\link{get_expression_values}()}

Other functions to set data in giotto object: 
\code{\link{get_cell_id}()},
\code{\link{get_feat_id}()},
\code{\link{set_NearestNetwork}()},
\code{\link{set_cell_id}()},
\code{\link{set_cell_metadata}()},
\code{\link{set_dimReduction}()},
\code{\link{set_feat_id}()},
\code{\link{set_feature_info}()},
\code{\link{set_feature_metadata}()},
\code{\link{set_giottoImage}()},
\code{\link{set_polygon_info}()},
\code{\link{set_spatialGrid}()},
\code{\link{set_spatialNetwork}()},
\code{\link{set_spatial_enrichment}()},
\code{\link{set_spatial_locations}()}
}
\concept{expression accessor functions}
\concept{functions to set data in giotto object}
