% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{set_expression_values}
\alias{set_expression_values}
\title{set_expression_values}
\usage{
set_expression_values(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  name = "test",
  values
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit}

\item{feat_type}{feature type}

\item{name}{name for the expression slot}

\item{values}{expression values}
}
\value{
giotto object
}
\description{
function to set expression values for giotto object
}
