% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{set_FeatMetadata}
\alias{set_FeatMetadata}
\title{set_FeatMetadata}
\usage{
set_FeatMetadata(gobject, spat_unit = NULL, feat_type = NULL, meta_dt = NULL)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit (e.g. "cell")}

\item{feat_type}{feature type (e.g. "rna", "dna", "protein")}

\item{meta_dt}{data.table object containing feature metadata}
}
\value{
giotto object
}
\description{
set specific slots of feature metadata, overwriting all other slots
}
\seealso{
Other functions to set data in giotto object: 
\code{\link{get_cell_id}()},
\code{\link{get_feat_id}()},
\code{\link{set_CellMetadata}()},
\code{\link{set_NearestNetwork}()},
\code{\link{set_cell_id}()},
\code{\link{set_dimReduction}()},
\code{\link{set_expression_values}()},
\code{\link{set_feat_id}()},
\code{\link{set_feature_info}()},
\code{\link{set_giottoImage}()},
\code{\link{set_polygon_info}()},
\code{\link{set_spatialGrid}()},
\code{\link{set_spatialNetwork}()},
\code{\link{set_spatial_enrichment}()},
\code{\link{set_spatial_locations}()}
}
\concept{functions to set data in giotto object}
