% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{set_feature_info}
\alias{set_feature_info}
\title{set_feature_info}
\usage{
set_feature_info(gobject, feat_name = NULL, gpolygon)
}
\arguments{
\item{gobject}{giotto object}

\item{feat_name}{name of feat}

\item{gpolygon}{giotto polygon}
}
\description{
set giotto polygon spatVector for features
}
