% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto.R, R/python_environment.R
\name{set_giotto_python_path}
\alias{set_giotto_python_path}
\title{set_giotto_python_path}
\usage{
set_giotto_python_path(python_path = NULL)

set_giotto_python_path(python_path = NULL)
}
\description{
sets the python path and/or installs miniconda and the python modules

sets the python path
}
\keyword{internal}
