% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{set_giottoImage}
\alias{set_giottoImage}
\title{Set giotto image object}
\usage{
set_giottoImage(
  gobject = NULL,
  image = NULL,
  image_type = NULL,
  name = NULL,
  verbose = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{image}{giotto image object to be attached without modification to the
giotto object}

\item{image_type}{type of giotto image object. Either "image" or "largeImage"}

\item{name}{name of giotto image object}

\item{verbose}{be verbose}
}
\value{
giotto object
}
\description{
Directly attach a giotto image to giotto object
}
\details{
\emph{\strong{Use with care!}} This function directly attaches giotto image
  objects to the gobject without further modifications of spatial positioning values
  within the image object that are generally needed in order for them to
  plot in the correct location relative to the other modalities of spatial data. \cr
  For the more general-purpose method of attaching image objects, see \code{\link{addGiottoImage}}
}
\seealso{
\code{\link{addGiottoImage}}

Other image data accessor functions: 
\code{\link{get_giottoImage}()}

Other functions to set data in giotto object: 
\code{\link{set_NearestNetwork}()},
\code{\link{set_expression_values}()},
\code{\link{set_feature_info}()},
\code{\link{set_polygon_info}()},
\code{\link{set_spatialGrid}()},
\code{\link{set_spatialNetwork}()},
\code{\link{set_spatial_enrichment}()},
\code{\link{set_spatial_locations}()}
}
\concept{functions to set data in giotto object}
\concept{image data accessor functions}
