% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{set_NearestNetwork}
\alias{set_NearestNetwork}
\title{Set nearest network}
\usage{
set_NearestNetwork(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  nn_network_to_use = "sNN",
  network_name = "sNN.pca",
  nn_network
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit (e.g. "cell")}

\item{feat_type}{feature type (e.g. "rna", "dna", "protein")}

\item{nn_network_to_use}{"kNN" or "sNN"}

\item{network_name}{name of NN network to be used}

\item{nn_network}{nearest network}
}
\value{
giotto object
}
\description{
Set a NN-network for a Giotto object
}
\seealso{
Other expression space nearest network accessor functions: 
\code{\link{get_NearestNetwork}()}

Other functions to set data in giotto object: 
\code{\link{get_cell_id}()},
\code{\link{get_feat_id}()},
\code{\link{set_CellMetadata}()},
\code{\link{set_FeatMetadata}()},
\code{\link{set_cell_id}()},
\code{\link{set_dimReduction}()},
\code{\link{set_expression_values}()},
\code{\link{set_feat_id}()},
\code{\link{set_feature_info}()},
\code{\link{set_giottoImage}()},
\code{\link{set_polygon_info}()},
\code{\link{set_spatialGrid}()},
\code{\link{set_spatialNetwork}()},
\code{\link{set_spatial_enrichment}()},
\code{\link{set_spatial_locations}()}
}
\concept{expression space nearest network accessor functions}
\concept{functions to set data in giotto object}
