% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{set_polygon_info}
\alias{set_polygon_info}
\title{set_polygon_info}
\usage{
set_polygon_info(gobject, polygon_name = "cell", gpolygon)
}
\arguments{
\item{gobject}{giotto object}

\item{polygon_name}{name of polygons}

\item{gpolygon}{giotto polygon}
}
\description{
set giotto polygon spatVector
}
