% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{set_polygon_info}
\alias{set_polygon_info}
\title{Set polygon info}
\usage{
set_polygon_info(gobject, polygon_name = "cell", gpolygon, verbose = TRUE)
}
\arguments{
\item{gobject}{giotto object}

\item{polygon_name}{name of polygons. Default "cell"}

\item{gpolygon}{giotto polygon}

\item{verbose}{verbosity}
}
\value{
giotto object
}
\description{
Set giotto polygon spatVector
}
\seealso{
Other polygon info data accessor functions: 
\code{\link{get_polygon_info}()}

Other functions to set data in giotto object: 
\code{\link{set_CellMetadata}()},
\code{\link{set_FeatMetadata}()},
\code{\link{set_NearestNetwork}()},
\code{\link{set_dimReduction}()},
\code{\link{set_expression_values}()},
\code{\link{set_feature_info}()},
\code{\link{set_giottoImage}()},
\code{\link{set_spatialGrid}()},
\code{\link{set_spatialNetwork}()},
\code{\link{set_spatial_enrichment}()},
\code{\link{set_spatial_locations}()}
}
\concept{functions to set data in giotto object}
\concept{polygon info data accessor functions}
