% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{set_spatial_enrichment}
\alias{set_spatial_enrichment}
\title{set_spatial_enrichment}
\usage{
set_spatial_enrichment(gobject, enrichm_name = "enrichment", spatenrichment)
}
\arguments{
\item{gobject}{giotto object}

\item{enrichm_name}{name of spatial enrichment results}

\item{spatenrichment}{spatial enrichment results}
}
\value{
giotto object
}
\description{
function to set a spatial enrichment slot
}
