% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{set_spatial_enrichment}
\alias{set_spatial_enrichment}
\title{Set spatial enrichment}
\usage{
set_spatial_enrichment(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  enrichm_name = "enrichment",
  spatenrichment
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit (e.g. "cell")}

\item{feat_type}{feature type (e.g. "rna", "dna", "protein")}

\item{enrichm_name}{name of spatial enrichment results. Default "DWLS"}

\item{spatenrichment}{spatial enrichment results}
}
\value{
giotto object
}
\description{
Function to set a spatial enrichment slot
}
\seealso{
Other spatial enrichment data accessor functions: 
\code{\link{get_spatial_enrichment}()}

Other functions to set data in giotto object: 
\code{\link{set_CellMetadata}()},
\code{\link{set_FeatMetadata}()},
\code{\link{set_NearestNetwork}()},
\code{\link{set_dimReduction}()},
\code{\link{set_expression_values}()},
\code{\link{set_feature_info}()},
\code{\link{set_giottoImage}()},
\code{\link{set_polygon_info}()},
\code{\link{set_spatialGrid}()},
\code{\link{set_spatialNetwork}()},
\code{\link{set_spatial_locations}()}
}
\concept{functions to set data in giotto object}
\concept{spatial enrichment data accessor functions}
