% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{set_spatial_locations}
\alias{set_spatial_locations}
\title{Set spatial locations}
\usage{
set_spatial_locations(
  gobject,
  spat_unit = NULL,
  spat_loc_name = "raw",
  spatlocs,
  verbose = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit (e.g. "cell")}

\item{spat_loc_name}{name of spatial locations, default "raw"}

\item{spatlocs}{spatial locations}

\item{verbose}{be verbose}
}
\value{
giotto object
}
\description{
Function to set a spatial location slot
}
\seealso{
Other spatial location data accessor functions: 
\code{\link{get_spatial_locations}()}

Other functions to set data in giotto object: 
\code{\link{set_NearestNetwork}()},
\code{\link{set_expression_values}()},
\code{\link{set_feature_info}()},
\code{\link{set_giottoImage}()},
\code{\link{set_polygon_info}()},
\code{\link{set_spatialGrid}()},
\code{\link{set_spatialNetwork}()},
\code{\link{set_spatial_enrichment}()}
}
\concept{functions to set data in giotto object}
\concept{spatial location data accessor functions}
