% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{set_spatial_locations}
\alias{set_spatial_locations}
\title{Set spatial locations}
\usage{
set_spatial_locations(
  gobject,
  spatlocs,
  spat_unit = NULL,
  spat_loc_name = "raw",
  provenance = NULL,
  verbose = TRUE,
  set_defaults = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{spatlocs}{spatial locations (accepts either \code{data.table} or
\code{spatLocsObj})}

\item{spat_unit}{spatial unit (e.g. "cell")}

\item{spat_loc_name}{name of spatial locations, default "raw"}

\item{provenance}{provenance information (optional)}

\item{verbose}{be verbose}

\item{set_defaults}{set default spat_unit and feat_type. Change to FALSE only when
expression and spat_info are not expected to exist.}
}
\value{
giotto object
}
\description{
Function to set a spatial location slot
}
\details{
If a \code{spatLocsObj} is provided to \code{spatlocs} param then any
attached name and spat_unit info will be used for input to this function's
\code{spat_loc_name} and \code{spat_unit}params, BUT will be overridden by any
alternative specific inputs to those params. \cr
ie: a \code{spatLocsObj} with spat_unit slot == 'cell' will be automatically
nested by spat_unit 'cell' when using \code{set_spatial_locations} as long as
param \code{spat_unit = NULL}. BUT if param \code{spat_unit = 'nucleus'} then
the \code{spatLocsObj} will be nested by spat_unit 'nucleus' instead and
its spat_unit slot will be changed to 'nucleus'
}
\seealso{
Other spatial location data accessor functions: 
\code{\link{get_spatial_locations}()}

Other functions to set data in giotto object: 
\code{\link{get_cell_id}()},
\code{\link{get_feat_id}()},
\code{\link{set_NearestNetwork}()},
\code{\link{set_cell_id}()},
\code{\link{set_cell_metadata}()},
\code{\link{set_dimReduction}()},
\code{\link{set_expression_values}()},
\code{\link{set_feat_id}()},
\code{\link{set_feature_info}()},
\code{\link{set_feature_metadata}()},
\code{\link{set_giottoImage}()},
\code{\link{set_multiomics}()},
\code{\link{set_polygon_info}()},
\code{\link{set_spatialGrid}()},
\code{\link{set_spatialNetwork}()},
\code{\link{set_spatial_enrichment}()}
}
\concept{functions to set data in giotto object}
\concept{spatial location data accessor functions}
