% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{set_spatial_locations}
\alias{set_spatial_locations}
\title{set_spatial_locations}
\usage{
set_spatial_locations(gobject, spat_loc_name = "raw", spatlocs)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_loc_name}{name of spatial locations}

\item{spatlocs}{spatial locations}
}
\value{
giotto object
}
\description{
function to set a spatial location slot
}
