% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{set_spatialGrid}
\alias{set_spatialGrid}
\title{Set spatial grid}
\usage{
set_spatialGrid(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  name = NULL,
  spatial_grid
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit (e.g. "cell")}

\item{feat_type}{feature type (e.g. "rna", "dna", "protein")}

\item{name}{name of spatial grid}

\item{spatial_grid}{spatial grid object}
}
\value{
giotto object
}
\description{
Function to set a spatial grid
}
\seealso{
Other spatial grid data accessor functions: 
\code{\link{get_spatialGrid}()}

Other functions to set data in giotto object: 
\code{\link{get_cell_id}()},
\code{\link{get_feat_id}()},
\code{\link{set_CellMetadata}()},
\code{\link{set_FeatMetadata}()},
\code{\link{set_NearestNetwork}()},
\code{\link{set_cell_id}()},
\code{\link{set_dimReduction}()},
\code{\link{set_expression_values}()},
\code{\link{set_feat_id}()},
\code{\link{set_feature_info}()},
\code{\link{set_giottoImage}()},
\code{\link{set_polygon_info}()},
\code{\link{set_spatialNetwork}()},
\code{\link{set_spatial_enrichment}()},
\code{\link{set_spatial_locations}()}
}
\concept{functions to set data in giotto object}
\concept{spatial grid data accessor functions}
