% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{set_spatialNetwork}
\alias{set_spatialNetwork}
\title{set_spatialNetwork}
\usage{
set_spatialNetwork(gobject, name = NULL, spatial_network)
}
\arguments{
\item{gobject}{giotto object}

\item{name}{name of spatial network}

\item{spatial_network}{spatial network}
}
\description{
function to set a spatial network
}
