% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interoperability.R
\name{seuratToGiotto}
\alias{seuratToGiotto}
\title{seuratToGiotto}
\usage{
seuratToGiotto(
  sobject,
  spatial_assay = "Spatial",
  dim_reduction = c("pca", "umap"),
  subcellular_assay = "Vizgen"
)
}
\arguments{
\item{sobject}{Seurat object}
}
\value{
Giotto object
}
\description{
Converts Seurat object into a Giotto object
}
