% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interoperability.R
\name{seuratToGiotto}
\alias{seuratToGiotto}
\title{Convert a Seurat object to a Giotto object}
\usage{
seuratToGiotto(
  sobject,
  spatial_assay = "Spatial",
  dim_reduction = c("pca", "umap"),
  subcellular_assay = "Vizgen"
)
}
\arguments{
\item{sobject}{Input Seurat object to convert to Giotto object}

\item{spatial_assay}{Specify name of the spatial assay slot in Seurat. Default is \code{"Spatial"}.}

\item{dim_reduction}{Specify which dimensional reduction computations to fetch from
input Seurat object. Default is \code{"c('pca', 'umap')"}.}

\item{subcellular_assay}{Specify name of the subcellular assay in input object.
Default is \code{"Vizgen"}.}
}
\value{
A Giotto object converted from Seurat object with all computations stored in it.
}
\description{
Convert a Seurat object to a Giotto object
}
