% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto.R, R/images.R
\docType{methods}
\name{show,giotto-method}
\alias{show,giotto-method}
\alias{show,giottoImage-method}
\title{show method for giotto class}
\usage{
\S4method{show}{giotto}(object)

\S4method{show}{giottoImage}(object)
}
\arguments{
\item{object}{giottoImage object}
}
\description{
show method for giotto class

show method for giottoImage class
}
