% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{show,giotto-method}
\alias{show,giotto-method}
\alias{show,aggregateExprObj-method}
\alias{show,dimObj-method}
\alias{show,spatialLocationsObj-method}
\alias{show,spatialNetworkObj-method}
\alias{show,spatialGridObj-method}
\alias{show,giottoImage-method}
\alias{show,giottoLargeImage-method}
\title{show method for giotto class}
\usage{
\S4method{show}{giotto}(object)

\S4method{show}{aggregateExprObj}(object)

\S4method{show}{dimObj}(object)

\S4method{show}{spatialLocationsObj}(object)

\S4method{show}{spatialNetworkObj}(object)

\S4method{show}{spatialGridObj}(object)

\S4method{show}{giottoImage}(object)

\S4method{show}{giottoLargeImage}(object)
}
\arguments{
\item{object}{giottoLargeImage object}
}
\description{
show method for giotto class

show method for aggregateExprObj class

show method for dimObj class

show method for spatialLocationsObj class

show method for spatialNetworkObj class

show method for spatialGridObj class

show method for giottoImage class

show method for giottoLargeImage class
}
