% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.R, R/images.R, R/giotto.R,
%   R/spatial_structures.R
\docType{methods}
\name{show,dimObj-method}
\alias{show,dimObj-method}
\alias{show,giottoImage-method}
\alias{show,giottoLargeImage-method}
\alias{show,giotto-method}
\alias{show,aggregateExprObj-method}
\alias{show,spatialNetworkObj-method}
\alias{show,spatialGridObj-method}
\title{show method for dimObj class}
\usage{
\S4method{show}{dimObj}(object)

\S4method{show}{giottoImage}(object)

\S4method{show}{giottoLargeImage}(object)

\S4method{show}{giotto}(object)

\S4method{show}{aggregateExprObj}(object)

\S4method{show}{spatialNetworkObj}(object)

\S4method{show}{spatialGridObj}(object)
}
\arguments{
\item{object}{spatial grid object}
}
\description{
show method for dimObj class

show method for giottoImage class

show method for giottoLargeImage class

show method for aggregateExprObj class

show method for spatialNetworkObj class

show method for spatialGridObj class
}
