% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giotto.R
\docType{methods}
\name{show,giotto-method}
\alias{show,giotto-method}
\title{show method for giotto class}
\usage{
\S4method{show}{giotto}(object)
}
\arguments{
\item{object}{giotto object}
}
\description{
show method for giotto class
}
