% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_visuals.R
\name{showClusterDendrogram}
\alias{showClusterDendrogram}
\title{showClusterDendrogram}
\usage{
showClusterDendrogram(gobject, expression_values = c("normalized",
  "scaled", "custom"), cluster_column, cor = c("pearson", "spearman"),
  distance = "ward.D", h = NULL, h_color = "red")
}
\arguments{
\item{gobject}{giotto object}

\item{expression_values}{expression values to use}

\item{cluster_column}{name of column to use for clusters}

\item{cor}{correlation score to calculate distance}

\item{distance}{distance method to use for hierarchical clustering}

\item{h}{height of horizontal lines to plot}

\item{h_color}{color of horizontal lines}
}
\value{
ggplot
}
\description{
Creates dendrogram based on identified clusters
}
\details{
Correlation dendrogram of selected clustering.
}
\examples{
    showClusterDendrogram(gobject)
}
