% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_visuals.R
\name{showClusterDendrogram}
\alias{showClusterDendrogram}
\title{showClusterDendrogram}
\usage{
showClusterDendrogram(
  gobject,
  expression_values = c("normalized", "scaled", "custom"),
  cluster_column,
  cor = c("pearson", "spearman"),
  distance = "ward.D",
  h = NULL,
  h_color = "red",
  rotate = FALSE,
  show_plot = NA,
  return_plot = NA,
  save_plot = NA,
  save_param = list(),
  default_save_name = "showClusterDendrogram",
  ...
)
}
\arguments{
\item{gobject}{giotto object}

\item{expression_values}{expression values to use}

\item{cluster_column}{name of column to use for clusters}

\item{cor}{correlation score to calculate distance}

\item{distance}{distance method to use for hierarchical clustering}

\item{h}{height of horizontal lines to plot}

\item{h_color}{color of horizontal lines}

\item{rotate}{rotate dendrogram 90 degrees}

\item{show_plot}{show plot}

\item{return_plot}{return ggplot object}

\item{save_plot}{directly save the plot [boolean]}

\item{save_param}{list of saving parameters, see \code{\link{showSaveParameters}}}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}

\item{...}{additional parameters for ggdendrogram()}
}
\value{
ggplot
}
\description{
Creates dendrogram for selected clusters.
}
\details{
Expression correlation dendrogram for selected clusters.
}
\examples{

data(mini_giotto_single_cell)

# cell metadata
cell_metadata = pDataDT(mini_giotto_single_cell)

# create heatmap
showClusterDendrogram(mini_giotto_single_cell,
                      cluster_column = 'cell_types')

}
