% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_visuals.R
\name{showClusterHeatmap}
\alias{showClusterHeatmap}
\title{showClusterHeatmap}
\usage{
showClusterHeatmap(gobject, expression_values = c("normalized", "scaled",
  "custom"), cluster_column, cor = c("pearson", "spearman"),
  distance = "ward.D")
}
\arguments{
\item{gobject}{giotto object}

\item{expression_values}{expression values to use}

\item{cluster_column}{name of column to use for clusters}

\item{cor}{correlation score to calculate distance}

\item{distance}{distance method to use for hierarchical clustering}
}
\value{
ggplot
}
\description{
Creates heatmap based on identified clusters
}
\details{
Correlation heatmap of selected clustering.
}
\examples{
    showClusterHeatmap(gobject)
}
