% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_visuals.R
\name{showClusterHeatmap}
\alias{showClusterHeatmap}
\title{showClusterHeatmap}
\usage{
showClusterHeatmap(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  expression_values = c("normalized", "scaled", "custom"),
  feats = "all",
  genes = NULL,
  cluster_column,
  cor = c("pearson", "spearman"),
  distance = "ward.D",
  show_plot = NA,
  return_plot = NA,
  save_plot = NA,
  save_param = list(),
  default_save_name = "showClusterHeatmap",
  ...
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit (e.g. "cell")}

\item{feat_type}{feature type (e.g. "rna", "dna", "protein")}

\item{expression_values}{expression values to use (e.g. "normalized", "scaled", "custom")}

\item{feats}{vector of features to use, default to 'all'}

\item{genes}{deprecated. Replaced by \code{feats} param}

\item{cluster_column}{name of column to use for clusters (e.g. "leiden_clus")}

\item{cor}{correlation score to calculate distance (e.g. "pearson", "spearman")}

\item{distance}{distance method to use for hierarchical clustering, default to "ward.D"}

\item{show_plot}{show plot. TRUE or FALSE}

\item{return_plot}{return ggplot object. TRUE or FALSE}

\item{save_plot}{directly save the plot. TRUE or FALSE}

\item{save_param}{list of saving parameters, see \code{\link{showSaveParameters}}}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}

\item{...}{additional parameters passed to \code{\link{ComplexHeatmap::Heatmap}} function}
}
\value{
ggplot
}
\description{
Creates heatmap based on identified clusters
}
\details{
Correlation heatmap of selected clusters.
}
