% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_visuals.R
\name{showClusterHeatmap}
\alias{showClusterHeatmap}
\title{showClusterHeatmap}
\usage{
showClusterHeatmap(gobject, expression_values = c("normalized", "scaled",
  "custom"), genes = "all", cluster_column, cor = c("pearson",
  "spearman"), distance = "ward.D", ...)
}
\arguments{
\item{gobject}{giotto object}

\item{expression_values}{expression values to use}

\item{genes}{vector of genes to use, default to 'all'}

\item{cluster_column}{name of column to use for clusters}

\item{cor}{correlation score to calculate distance}

\item{distance}{distance method to use for hierarchical clustering}

\item{...}{additional parameters for the Heatmap function from ComplexHeatmap}
}
\value{
ggplot
}
\description{
Creates heatmap based on identified clusters
}
\details{
Correlation heatmap of selected clusters.
}
\examples{
    showClusterHeatmap(gobject)
}
