% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_visuals.R
\name{showClusterHeatmap}
\alias{showClusterHeatmap}
\title{showClusterHeatmap}
\usage{
showClusterHeatmap(
  gobject,
  expression_values = c("normalized", "scaled", "custom"),
  genes = "all",
  cluster_column,
  cor = c("pearson", "spearman"),
  distance = "ward.D",
  show_plot = NA,
  return_plot = NA,
  save_plot = NA,
  save_param = list(),
  default_save_name = "showClusterHeatmap",
  ...
)
}
\arguments{
\item{gobject}{giotto object}

\item{expression_values}{expression values to use}

\item{genes}{vector of genes to use, default to 'all'}

\item{cluster_column}{name of column to use for clusters}

\item{cor}{correlation score to calculate distance}

\item{distance}{distance method to use for hierarchical clustering}

\item{show_plot}{show plot}

\item{return_plot}{return ggplot object}

\item{save_plot}{directly save the plot [boolean]}

\item{save_param}{list of saving parameters, see \code{\link{showSaveParameters}}}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}

\item{...}{additional parameters for the Heatmap function from ComplexHeatmap}
}
\value{
ggplot
}
\description{
Creates heatmap based on identified clusters
}
\details{
Correlation heatmap of selected clusters.
}
\examples{

data(mini_giotto_single_cell)

# cell metadata
cell_metadata = pDataDT(mini_giotto_single_cell)

# create heatmap
showClusterHeatmap(mini_giotto_single_cell,
                   cluster_column = 'cell_types')

}
