% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction_visuals.R
\name{showCPGscores}
\alias{showCPGscores}
\title{showCPGscores}
\usage{
showCPGscores(CPGscore, method = c("cell_barplot", "cell-cell",
  "cell_sankey"), min_cells = 5, min_pval = 0.05,
  min_spat_diff = 0.2, cell_color_code = NULL, show.plot = T)
}
\arguments{
\item{CPGscore}{CPGscore, output from getCellProximityGeneScores()}

\item{method}{visualization method}

\item{min_cells}{min number of cells threshold}

\item{min_pval}{p-value threshold}

\item{min_spat_diff}{spatial difference threshold}

\item{cell_color_code}{color code for cell types}

\item{show.plot}{print plot}
}
\value{
Gene to gene scores in data.table format
}
\description{
visualize Cell Proximity Gene enrichment scores
}
\details{
Give more details ...
}
\examples{
    showCPGscores(CPGscore)
}
