% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{showGiottoExpression}
\alias{showGiottoExpression}
\title{showGiottoExpression}
\usage{
showGiottoExpression(gobject)
}
\arguments{
\item{gobject}{giotto object}
}
\value{
prints the name and small subset of available matrices
}
\description{
shows the available matrices
}
\seealso{
Other functions to show data in giotto object: 
\code{\link{showGiottoCellMetadata}()},
\code{\link{showGiottoDimRed}()},
\code{\link{showGiottoFeatInfo}()},
\code{\link{showGiottoFeatMetadata}()},
\code{\link{showGiottoImageNames}()},
\code{\link{showGiottoNearestNetworks}()},
\code{\link{showGiottoSpatEnrichments}()},
\code{\link{showGiottoSpatGrids}()},
\code{\link{showGiottoSpatLocs}()},
\code{\link{showGiottoSpatNetworks}()},
\code{\link{showGiottoSpatialInfo}()}
}
\concept{functions to show data in giotto object}
\keyword{show}
