% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{showGiottoFeatInfo}
\alias{showGiottoFeatInfo}
\title{showGiottoFeatInfo}
\usage{
showGiottoFeatInfo(gobject)
}
\arguments{
\item{gobject}{giotto object}
}
\description{
show the available giotto spatial feature information
}
\seealso{
Other functions to show data in giotto object: 
\code{\link{showGiottoDimRed}()},
\code{\link{showGiottoExpression}()},
\code{\link{showGiottoImageNames}()},
\code{\link{showGiottoSpatEnrichments}()},
\code{\link{showGiottoSpatGrids}()},
\code{\link{showGiottoSpatLocs}()},
\code{\link{showGiottoSpatNetworks}()},
\code{\link{showGiottoSpatialInfo}()}
}
\concept{functions to show data in giotto object}
\keyword{show}
