% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{showGiottoImageNames}
\alias{showGiottoImageNames}
\title{showGiottoImageNames}
\usage{
showGiottoImageNames(gobject)
}
\arguments{
\item{gobject}{a giotto object}
}
\value{
prints names of available giotto image objects
}
\description{
Prints the available giotto images that are attached to the Giotto object
}
\keyword{show}
