% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{showGiottoImageNames}
\alias{showGiottoImageNames}
\title{showGiottoImageNames}
\usage{
showGiottoImageNames(gobject, verbose = TRUE)
}
\arguments{
\item{gobject}{a giotto object}

\item{verbose}{verbosity of function}
}
\value{
a vector of giotto image names attached to the giotto object
}
\description{
Prints the available giotto images that are attached to the Giotto object
}
