% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{showGiottoSpatGrids}
\alias{showGiottoSpatGrids}
\title{showGiottoSpatGrids}
\usage{
showGiottoSpatGrids(
  gobject,
  feat_type = NULL,
  spat_unit = NULL,
  verbose = TRUE
)
}
\arguments{
\item{gobject}{a giotto object}

\item{verbose}{verbosity of function#'}
}
\value{
vector
}
\description{
Prints the available spatial grids that are attached to the Giotto object
}
