% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{showGiottoSpatialInfo}
\alias{showGiottoSpatialInfo}
\title{showGiottoSpatialInfo}
\usage{
showGiottoSpatialInfo(gobject)
}
\arguments{
\item{gobject}{giotto object}
}
\description{
show the available giotto spatial polygon information
}
\keyword{show}
