% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{showGiottoSpatLocs}
\alias{showGiottoSpatLocs}
\title{showGiottoSpatLocs}
\usage{
showGiottoSpatLocs(gobject, nrows = 4)
}
\arguments{
\item{gobject}{giotto object}

\item{nrows}{number of rows to print for each spatial location data.table}
}
\value{
prints the name and small subset of available data.table
}
\description{
shows the available spatial locations
}
\seealso{
Other functions to show data in giotto object: 
\code{\link{showGiottoCellMetadata}()},
\code{\link{showGiottoDimRed}()},
\code{\link{showGiottoExpression}()},
\code{\link{showGiottoFeatInfo}()},
\code{\link{showGiottoFeatMetadata}()},
\code{\link{showGiottoImageNames}()},
\code{\link{showGiottoNearestNetworks}()},
\code{\link{showGiottoSpatEnrichments}()},
\code{\link{showGiottoSpatGrids}()},
\code{\link{showGiottoSpatNetworks}()},
\code{\link{showGiottoSpatialInfo}()}
}
\concept{functions to show data in giotto object}
\keyword{show}
