% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{showGiottoSpatNetworks}
\alias{showGiottoSpatNetworks}
\title{showGiottoSpatNetworks}
\usage{
showGiottoSpatNetworks(
  gobject,
  feat_type = NULL,
  spat_unit = NULL,
  verbose = TRUE
)
}
\arguments{
\item{gobject}{a giotto object}

\item{feat_type}{feature type}

\item{spat_unit}{spatial unit}

\item{verbose}{verbosity of function#'}
}
\value{
vector
}
\description{
Prints the available spatial networks that are attached to the Giotto object
}
