% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_structures.R
\name{showNetworks}
\alias{showNetworks}
\title{showNetworks}
\usage{
showNetworks(gobject, verbose = TRUE)
}
\arguments{
\item{gobject}{a giotto object}

\item{verbose}{verbosity of function#'}
}
\value{
vector
}
\description{
Prints the available spatial networks that are attached to the Giotto object
}
