% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_genes.R
\name{showPattern}
\alias{showPattern}
\title{showPattern}
\usage{
showPattern(spatPatObj, dimension = 1, trim = c(0.02, 0.98),
  background.color = "white", grid.border.color = "grey",
  show.legend = T, show.plot = F)
}
\arguments{
\item{spatPatObj}{Output from detectSpatialPatterns}

\item{dimension}{dimension to plot}

\item{trim}{Trim ends of the PC values.}

\item{background.color}{background color for plot}

\item{grid.border.color}{color for grid}

\item{show.plot}{Show the plot.}
}
\value{
ggplot
}
\description{
create a spatial grid
}
\details{
Description.
}
\examples{
    showPattern(gobject)
}
