% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_genes.R
\name{showPattern}
\alias{showPattern}
\title{showPattern}
\usage{
showPattern(gobject, spatPatObj, ...)
}
\arguments{
\item{gobject}{giotto object}

\item{spatPatObj}{Output from detectSpatialPatterns}

\item{...}{
  Arguments passed on to \code{\link[=showPattern2D]{showPattern2D}}
  \describe{
    \item{\code{dimension}}{dimension to plot}
    \item{\code{trim}}{Trim ends of the PC values.}
    \item{\code{background_color}}{background color for plot}
    \item{\code{grid_border_color}}{color for grid}
    \item{\code{show_legend}}{show legend of ggplot}
    \item{\code{point_size}}{size of points}
    \item{\code{show_plot}}{show plot}
    \item{\code{return_plot}}{return ggplot object}
    \item{\code{save_plot}}{directly save the plot [boolean]}
    \item{\code{save_param}}{list of saving parameters, see \code{\link{showSaveParameters}}}
    \item{\code{default_save_name}}{default save name for saving, don't change, change save_name in save_param}
  }}
}
\value{
ggplot
}
\description{
show patterns for 2D spatial data
}
\seealso{
\code{\link{showPattern2D}}
}
