% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_genes.R
\name{showPatternGenes}
\alias{showPatternGenes}
\title{showPatternGenes}
\usage{
showPatternGenes(spatPatObj, dimension = 1, top_pos_genes = 5,
  top_neg_genes = 5, point.size = 1, show.plot = F)
}
\arguments{
\item{spatPatObj}{Output from detectSpatialPatterns}

\item{dimension}{dimension to plot genes for.}

\item{top_pos_genes}{Top positively correlated genes.}

\item{top_neg_genes}{Top negatively correlated genes.}

\item{point.size}{size of points}

\item{show.plot}{Show the plot.}
}
\value{
ggplot
}
\description{
create a spatial grid
}
\details{
Description.
}
\examples{
    showPatternGenes(gobject)
}
