% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.R
\name{signPCA}
\alias{signPCA}
\title{signPCA}
\usage{
signPCA(
  gobject,
  method = c("screeplot", "jackstraw"),
  expression_values = c("normalized", "scaled", "custom"),
  reduction = c("cells", "genes"),
  genes_to_use = NULL,
  scale_unit = T,
  ncp = 50,
  scree_labels = T,
  scree_ylim = c(0, 10),
  jack_iter = 10,
  jack_threshold = 0.01,
  jack_verbose = T,
  show_plot = NA,
  return_plot = NA,
  save_plot = NA,
  save_param = list(),
  default_save_name = "signPCA",
  ...
)
}
\arguments{
\item{gobject}{giotto object}

\item{method}{method to use to identify significant PCs}

\item{expression_values}{expression values to use}

\item{reduction}{cells or genes}

\item{genes_to_use}{subset of genes to use for PCA}

\item{scale_unit}{scale features before PCA}

\item{ncp}{number of principal components to calculate}

\item{scree_labels}{show labels on scree plot}

\item{scree_ylim}{y-axis limits on scree plot}

\item{jack_iter}{number of interations for jackstraw}

\item{jack_threshold}{p-value threshold to call a PC significant}

\item{jack_verbose}{show progress of jackstraw method}

\item{show_plot}{show plot}

\item{return_plot}{return ggplot object}

\item{save_plot}{directly save the plot [boolean]}

\item{save_param}{list of saving parameters from all_plots_save_function()}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}

\item{...}{additional parameters for PCA}
}
\value{
ggplot object for scree method and maxtrix of p-values for jackstraw
}
\description{
identify significant prinicipal components (PCs)
}
\details{
Two different methods can be used to assess the number of relevant or significant
 prinicipal components (PC's). \cr
 1. Screeplot works by plotting the explained variance of each
 individual PC in a barplot allowing you to identify which PC does not show a significant
 contribution anymore ( = 'elbow method'). \cr
 2. The Jackstraw method uses the \code{\link[jackstraw]{permutationPA}} function. By
 systematically permuting genes it identifies robust, and thus significant, PCs.
 \cr multiple PCA results can be stored by changing the \emph{name} parameter
}
\examples{
    signPCA(gobject)
}
