% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.R
\name{signPCA}
\alias{signPCA}
\title{signPCA}
\usage{
signPCA(gobject, method = c("screeplot", "jackstraw"),
  expression_values = c("normalized", "scaled", "custom"),
  reduction = c("cells", "genes"), genes_to_use = NULL,
  scale_unit = T, ncp = 50, scree_labels = T, scree_ylim = c(0,
  10), jack_iter = 10, jack_threshold = 0.01, jack_verbose = T,
  show_plot = T, ...)
}
\arguments{
\item{gobject}{giotto object}

\item{method}{method to use to identify significant PCs}

\item{expression_values}{expression values to use}

\item{reduction}{cells or genes}

\item{genes_to_use}{subset of genes to use for PCA}

\item{scale_unit}{scale features before PCA}

\item{ncp}{number of principal components to calculate}

\item{scree_labels}{show labels on scree plot}

\item{scree_ylim}{y-axis limits on scree plot}

\item{jack_iter}{number of interations for jackstraw}

\item{jack_threshold}{p-value threshold to call a PC significant}

\item{jack_verbose}{show progress of jackstraw method}

\item{show_plot}{show plots}

\item{...}{additional parameters for PCA}
}
\value{
ggplot object for scree method and maxtrix of p-values for jackstraw
}
\description{
identify significant prinicipal components (PCs)
}
\details{
Description of PCA steps...
}
\examples{
    signPCA(gobject)
}
