% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.R
\name{signPCA}
\alias{signPCA}
\title{signPCA}
\usage{
signPCA(
  gobject,
  name = "pca",
  method = c("screeplot", "jackstraw"),
  expression_values = c("normalized", "scaled", "custom"),
  reduction = c("cells", "genes"),
  pca_method = c("irlba", "factominer"),
  rev = FALSE,
  genes_to_use = NULL,
  center = T,
  scale_unit = T,
  ncp = 50,
  scree_ylim = c(0, 10),
  jack_iter = 10,
  jack_threshold = 0.01,
  jack_ylim = c(0, 1),
  verbose = TRUE,
  show_plot = NA,
  return_plot = NA,
  save_plot = NA,
  save_param = list(),
  default_save_name = "signPCA"
)
}
\arguments{
\item{gobject}{giotto object}

\item{name}{name of PCA object if available}

\item{method}{method to use to identify significant PCs}

\item{expression_values}{expression values to use}

\item{reduction}{cells or genes}

\item{pca_method}{which implementation to use}

\item{rev}{do a reverse PCA}

\item{genes_to_use}{subset of genes to use for PCA}

\item{center}{center data before PCA}

\item{scale_unit}{scale features before PCA}

\item{ncp}{number of principal components to calculate}

\item{scree_ylim}{y-axis limits on scree plot}

\item{jack_iter}{number of interations for jackstraw}

\item{jack_threshold}{p-value threshold to call a PC significant}

\item{jack_ylim}{y-axis limits on jackstraw plot}

\item{verbose}{verbosity}

\item{show_plot}{show plot}

\item{return_plot}{return ggplot object}

\item{save_plot}{directly save the plot [boolean]}

\item{save_param}{list of saving parameters from all_plots_save_function()}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}
}
\value{
ggplot object for scree method and maxtrix of p-values for jackstraw
}
\description{
identify significant prinicipal components (PCs)
}
\details{
Two different methods can be used to assess the number of relevant or significant
 prinicipal components (PC's). \cr
 1. Screeplot works by plotting the explained variance of each
 individual PC in a barplot allowing you to identify which PC provides a significant
 contribution  (a.k.a. 'elbow method'). \cr
 2. The Jackstraw method uses the \code{\link[jackstraw]{permutationPA}} function. By
 systematically permuting genes it identifies robust, and thus significant, PCs.
 \cr
}
