% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_genes.R
\name{simulateOneGenePatternGiottoObject}
\alias{simulateOneGenePatternGiottoObject}
\title{simulateOneGenePatternGiottoObject}
\usage{
simulateOneGenePatternGiottoObject(
  gobject,
  pattern_name = "pattern",
  pattern_cell_ids = NULL,
  gene_name = NULL,
  spatial_prob = 0.95,
  gradient_direction = NULL,
  show_pattern = TRUE,
  pattern_colors = c(`in` = "green", out = "red"),
  ...
)
}
\arguments{
\item{gobject}{giotto object}

\item{pattern_name}{name of spatial pattern}

\item{pattern_cell_ids}{cell ids that make up the spatial pattern}

\item{gene_name}{selected gene}

\item{spatial_prob}{probability for a high expressing gene value to be part of the spatial pattern}

\item{gradient_direction}{direction of gradient}

\item{show_pattern}{show the discrete spatial pattern}

\item{pattern_colors}{2 color vector for the spatial pattern}

\item{\dots}{additional parameters for (re-)normalizing}
}
\value{
Reprocessed Giotto object for which one gene has a forced spatial pattern
}
\description{
Create a simulated spatial pattern for one selected gnee
}
