% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_help.R
\name{sort_combine_two_DT_columns}
\alias{sort_combine_two_DT_columns}
\title{sort_combine_two_DT_columns}
\usage{
sort_combine_two_DT_columns(DT, column1, column2, myname = "unif_gene_gene")
}
\description{
fast sorting and pasting of 2 character columns in a data.table
}
\keyword{internal}
