% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_genes.R
\name{spark}
\alias{spark}
\title{spark}
\usage{
spark(
  gobject,
  percentage = 0.1,
  min_count = 10,
  expression_values = "raw",
  num_core = 5,
  covariates = NULL,
  return_object = c("data.table", "spark"),
  ...
)
}
\arguments{
\item{gobject}{giotto object}

\item{percentage}{The percentage of cells that are expressed for analysis}

\item{min_count}{minimum number of counts for a gene to be included}

\item{expression_values}{type of values to use (raw by default)}

\item{num_core}{number of cores to use}

\item{covariates}{The covariates in experiments, i.e. confounding factors/batch effect. Column name of giotto cell metadata.}

\item{return_object}{type of result to return (data.table or spark object)}

\item{\dots}{Additional parameters to the \code{\link[SPARK]{spark.vc}} function}
}
\value{
data.table with SPARK spatial genes results or the SPARK object
}
\description{
Compute spatially expressed genes with SPARK method
}
\details{
This function is a wrapper for the method implemented in the SPARK package:
\itemize{
 \item{1. CreateSPARKObject }{create a SPARK object from a Giotto object}
 \item{2. spark.vc }{ Fits the count-based spatial model to estimate the parameters,
 see \code{\link[SPARK]{spark.vc}} for additional parameters}
 \item{3. spark.test }{ Testing multiple kernel matrices}
}
}
