% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction_spot.R
\name{spatCellCellcomSpots}
\alias{spatCellCellcomSpots}
\title{spatCellCellcomSpots}
\usage{
spatCellCellcomSpots(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  ave_celltype_exp,
  spatial_network_name = "Delaunay_network",
  cluster_column = "cell_ID",
  random_iter = 1000,
  feature_set_1,
  feature_set_2,
  min_observations = 2,
  expression_values = c("normalized", "scaled", "custom"),
  detailed = FALSE,
  adjust_method = c("fdr", "bonferroni", "BH", "holm", "hochberg", "hommel", "BY",
    "none"),
  adjust_target = c("features", "cells"),
  do_parallel = TRUE,
  cores = NA,
  set_seed = TRUE,
  seed_number = 1234,
  verbose = c("a little", "a lot", "none")
)
}
\arguments{
\item{gobject}{giotto object to use}

\item{spat_unit}{spatial unit (e.g. 'cell')}

\item{feat_type}{feature type (e.g. 'rna')}

\item{ave_celltype_exp}{Matrix with average expression per cell type}

\item{spatial_network_name}{spatial network to use for identifying
interacting cells}

\item{cluster_column}{cluster column with cell type information}

\item{random_iter}{number of iterations}

\item{feature_set_1}{first specific feature set from feature pairs}

\item{feature_set_2}{second specific feature set from feature pairs}

\item{min_observations}{minimum number of interactions needed to be
considered}

\item{expression_values}{(e.g. 'normalized', 'scaled', 'custom')}

\item{detailed}{provide more detailed information
(random variance and z-score)}

\item{adjust_method}{which method to adjust p-values}

\item{adjust_target}{adjust multiple hypotheses at the cell or feature
level}

\item{do_parallel}{run calculations in parallel with mclapply}

\item{cores}{number of cores to use if do_parallel = TRUE}

\item{set_seed}{set a seed for reproducibility}

\item{seed_number}{seed number}

\item{verbose}{verbose (e.g. 'a little', 'a lot', 'none')}
}
\value{
Cell-Cell communication scores for feature pairs based on spatial
interaction
}
\description{
Spatial Cell-Cell communication scores based on spatial
expression of interacting cells at spots resolution
}
\details{
Statistical framework to identify if pairs of features
(such as ligand-receptor combinations) are expressed at higher levels than
expected based on a reshuffled null distribution of feature expression
values in cells that are spatially in proximity to each other.
\itemize{
 * LR_comb:Pair of ligand and receptor
 * lig_cell_type: cell type to assess expression level of ligand
 * lig_expr: average expression residual(observed - DWLS_predicted) of
 ligand in lig_cell_type
 * ligand: ligand name
 * rec_cell_type: cell type to assess expression level of receptor
 * rec_expr: average expression residual(observed - DWLS_predicted) of
 receptor in rec_cell_type
 * receptor: receptor name
 * LR_expr: combined average ligand and receptor expression residual
 * lig_nr: total number of cells from lig_cell_type that spatially interact
 with cells from rec_cell_type
 * rec_nr: total number of cells from rec_cell_type that spatially interact
 with cells from lig_cell_type
 * rand_expr: average combined ligand and receptor expression residual from
 random spatial permutations
 * av_diff: average difference between LR_expr and rand_expr over all random
 spatial permutations
 * sd_diff: (optional) standard deviation of the difference between LR_expr
 and rand_expr over all random spatial permutations
 * z_score: (optional) z-score
 * log2fc: LR_expr - rand_expr
 * pvalue: p-value
 * LR_cell_comb: cell type pair combination
 * p.adj: adjusted p-value
 * PI: significanc score: log2fc \* -log10(p.adj)
}
}
