% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visuals.R
\name{spatDimPlot}
\alias{spatDimPlot}
\title{spatDimPlot}
\usage{
spatDimPlot(gobject, ...)
}
\arguments{
\item{gobject}{giotto object}

\item{plot_alignment}{direction to align plot}

\item{dim_reduction_to_use}{dimension reduction to use}

\item{dim_reduction_name}{dimension reduction name}

\item{dim1_to_use}{dimension to use on x-axis}

\item{dim2_to_use}{dimension to use on y-axis}

\item{sdimx}{= spatial dimension to use on x-axis}

\item{sdimy}{= spatial dimension to use on y-axis}

\item{spat_enr_names}{names of spatial enrichment results to include}

\item{cell_color}{color for cells (see details)}

\item{color_as_factor}{convert color column to factor}

\item{cell_color_code}{named vector with colors}

\item{cell_color_gradient}{vector with 3 colors for numeric data}

\item{gradient_midpoint}{midpoint for color gradient}

\item{gradient_limits}{vector with lower and upper limits}

\item{select_cell_groups}{select subset of cells/clusters based on cell_color parameter}

\item{select_cells}{select subset of cells based on cell IDs}

\item{dim_point_size}{size of points in dim. reduction space}

\item{dim_point_border_col}{border color of points in dim. reduction space}

\item{dim_point_border_stroke}{border stroke of points in dim. reduction space}

\item{spat_point_size}{size of spatial points}

\item{spat_point_border_col}{border color of spatial points}

\item{spat_point_border_stroke}{border stroke of spatial points}

\item{dim_show_cluster_center}{show the center of each cluster}

\item{dim_show_center_label}{provide a label for each cluster}

\item{dim_center_point_size}{size of the center point}

\item{dim_center_point_border_col}{border color of center point}

\item{dim_center_point_border_stroke}{stroke size of center point}

\item{dim_label_size}{size of the center label}

\item{dim_label_fontface}{font of the center label}

\item{spat_show_cluster_center}{show the center of each cluster}

\item{spat_show_center_label}{provide a label for each cluster}

\item{spat_center_point_size}{size of the center point}

\item{spat_label_size}{size of the center label}

\item{spat_label_fontface}{font of the center label}

\item{show_NN_network}{show underlying NN network}

\item{nn_network_to_use}{type of NN network to use (kNN vs sNN)}

\item{network_name}{name of NN network to use, if show_NN_network = TRUE}

\item{nn_network_alpha}{column to use for alpha of the edges}

\item{show_spatial_network}{show spatial network}

\item{spat_network_name}{name of spatial network to use}

\item{spat_network_color}{color of spatial network}

\item{show_spatial_grid}{show spatial grid}

\item{spat_grid_name}{name of spatial grid to use}

\item{spat_grid_color}{color of spatial grid}

\item{show_other_cells}{display not selected cells}

\item{other_cell_color}{color of not selected cells}

\item{dim_other_point_size}{size of not selected dim cells}

\item{spat_other_point_size}{size of not selected spat cells}

\item{spat_other_cells_alpha}{alpha of not selected spat cells}

\item{dim_show_legend}{show legend of dimension reduction plot}

\item{spat_show_legend}{show legend of spatial plot}

\item{show_plot}{show plot}

\item{return_plot}{return ggplot object}

\item{save_plot}{directly save the plot [boolean]}

\item{save_param}{list of saving parameters from all_plots_save_function()}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}
}
\value{
ggplot
}
\description{
Visualize cells according to spatial AND dimension reduction coordinates 2D
}
\details{
Description of parameters.
}
\examples{
    spatDimPlot(gobject)
}
\seealso{
\code{\link{spatDimPlot2D}} and \code{\link{spatDimPlot3D}} for 3D visualization.
}
