% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visuals.R
\name{spatDimPlot}
\alias{spatDimPlot}
\title{spatDimPlot}
\usage{
spatDimPlot(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=spatDimPlot2D]{spatDimPlot2D}}
  \describe{
    \item{\code{gobject}}{giotto object}
    \item{\code{show_image}}{show a tissue background image}
    \item{\code{gimage}}{a giotto image}
    \item{\code{image_name}}{name of a giotto image}
    \item{\code{plot_alignment}}{direction to align plot}
    \item{\code{dim_reduction_to_use}}{dimension reduction to use}
    \item{\code{dim_reduction_name}}{dimension reduction name}
    \item{\code{dim1_to_use}}{dimension to use on x-axis}
    \item{\code{dim2_to_use}}{dimension to use on y-axis}
    \item{\code{sdimx}}{= spatial dimension to use on x-axis}
    \item{\code{sdimy}}{= spatial dimension to use on y-axis}
    \item{\code{spat_enr_names}}{names of spatial enrichment results to include}
    \item{\code{cell_color}}{color for cells (see details)}
    \item{\code{color_as_factor}}{convert color column to factor}
    \item{\code{cell_color_code}}{named vector with colors}
    \item{\code{cell_color_gradient}}{vector with 3 colors for numeric data}
    \item{\code{gradient_midpoint}}{midpoint for color gradient}
    \item{\code{gradient_limits}}{vector with lower and upper limits}
    \item{\code{select_cell_groups}}{select subset of cells/clusters based on cell_color parameter}
    \item{\code{select_cells}}{select subset of cells based on cell IDs}
    \item{\code{dim_point_shape}}{point with border or not (border or no_border)}
    \item{\code{dim_point_size}}{size of points in dim. reduction space}
    \item{\code{dim_point_alpha}}{transparancy of point in dim. reduction space}
    \item{\code{dim_point_border_col}}{border color of points in dim. reduction space}
    \item{\code{dim_point_border_stroke}}{border stroke of points in dim. reduction space}
    \item{\code{spat_point_shape}}{shape of points (border, no_border or voronoi)}
    \item{\code{spat_point_size}}{size of spatial points}
    \item{\code{spat_point_alpha}}{transparancy of spatial points}
    \item{\code{spat_point_border_col}}{border color of spatial points}
    \item{\code{spat_point_border_stroke}}{border stroke of spatial points}
    \item{\code{dim_show_cluster_center}}{show the center of each cluster}
    \item{\code{dim_show_center_label}}{provide a label for each cluster}
    \item{\code{dim_center_point_size}}{size of the center point}
    \item{\code{dim_center_point_border_col}}{border color of center point}
    \item{\code{dim_center_point_border_stroke}}{stroke size of center point}
    \item{\code{dim_label_size}}{size of the center label}
    \item{\code{dim_label_fontface}}{font of the center label}
    \item{\code{spat_show_cluster_center}}{show the center of each cluster}
    \item{\code{spat_show_center_label}}{provide a label for each cluster}
    \item{\code{spat_center_point_size}}{size of the center point}
    \item{\code{spat_center_point_border_col}}{border color of spatial center points}
    \item{\code{spat_center_point_border_stroke}}{border strike size of spatial center points}
    \item{\code{spat_label_size}}{size of the center label}
    \item{\code{spat_label_fontface}}{font of the center label}
    \item{\code{show_NN_network}}{show underlying NN network}
    \item{\code{nn_network_to_use}}{type of NN network to use (kNN vs sNN)}
    \item{\code{network_name}}{name of NN network to use, if show_NN_network = TRUE}
    \item{\code{nn_network_alpha}}{column to use for alpha of the edges}
    \item{\code{show_spatial_network}}{show spatial network}
    \item{\code{spat_network_name}}{name of spatial network to use}
    \item{\code{spat_network_color}}{color of spatial network}
    \item{\code{spat_network_alpha}}{alpha of spatial network}
    \item{\code{show_spatial_grid}}{show spatial grid}
    \item{\code{spat_grid_name}}{name of spatial grid to use}
    \item{\code{spat_grid_color}}{color of spatial grid}
    \item{\code{show_other_cells}}{display not selected cells}
    \item{\code{other_cell_color}}{color of not selected cells}
    \item{\code{dim_other_point_size}}{size of not selected dim cells}
    \item{\code{spat_other_point_size}}{size of not selected spat cells}
    \item{\code{spat_other_cells_alpha}}{alpha of not selected spat cells}
    \item{\code{dim_show_legend}}{show legend of dimension reduction plot}
    \item{\code{spat_show_legend}}{show legend of spatial plot}
    \item{\code{legend_text}}{size of legend text}
    \item{\code{legend_symbol_size}}{size of legend symbols}
    \item{\code{dim_background_color}}{background color of points in dim. reduction space}
    \item{\code{spat_background_color}}{background color of spatial points}
    \item{\code{vor_border_color}}{border colorr for voronoi plot}
    \item{\code{vor_max_radius}}{maximum radius for voronoi 'cells'}
    \item{\code{vor_alpha}}{transparancy of voronoi 'cells'}
    \item{\code{axis_text}}{size of axis text}
    \item{\code{axis_title}}{size of axis title}
    \item{\code{show_plot}}{show plot}
    \item{\code{return_plot}}{return ggplot object}
    \item{\code{save_plot}}{directly save the plot [boolean]}
    \item{\code{save_param}}{list of saving parameters, see \code{\link{showSaveParameters}}}
    \item{\code{default_save_name}}{default save name for saving, don't change, change save_name in save_param}
  }}
}
\value{
ggplot
}
\description{
Visualize cells according to spatial AND dimension reduction coordinates 2D
}
\details{
Description of parameters.
}
\examples{

data(mini_giotto_single_cell)

spatDimPlot(mini_giotto_single_cell)
spatDimPlot(mini_giotto_single_cell, cell_color = 'cell_types',
             spat_point_size = 3, dim_point_size = 3)

}
\seealso{
\code{\link{spatDimPlot2D}} and \code{\link{spatDimPlot3D}} for 3D visualization.

Other spatial and dimension reduction visualizations: 
\code{\link{spatDimPlot2D}()},
\code{\link{spatDimPlot3D}()}
}
\concept{spatial and dimension reduction visualizations}
