% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visuals.R
\name{spatGenePlot}
\alias{spatGenePlot}
\title{spatGenePlot}
\usage{
spatGenePlot(gobject, ...)
}
\arguments{
\item{gobject}{giotto object}

\item{...}{parameters for cowplot::save_plot()}

\item{expression_values}{gene expression values to use}

\item{genes}{genes to show}

\item{genes_high_color}{color represents high gene expression}

\item{genes_mid_color}{color represents middle gene expression}

\item{genes_low_color}{color represents low gene expression}

\item{show_network}{show underlying spatial network}

\item{network_color}{color of spatial network}

\item{spatial_network_name}{name of spatial network to use}

\item{show_grid}{show spatial grid}

\item{grid_color}{color of spatial grid}

\item{spatial_grid_name}{name of spatial grid to use}

\item{midpoint}{expression midpoint}

\item{scale_alpha_with_expression}{scale expression with ggplot alpha parameter}

\item{point_size}{size of point (cell)}

\item{point_border_col}{color of border around points}

\item{point_border_stroke}{stroke size of border around points}

\item{cow_n_col}{cowplot param: how many columns}

\item{cow_rel_h}{cowplot param: relative height}

\item{cow_rel_w}{cowplot param: relative width}

\item{cow_align}{cowplot param: how to align}

\item{show_legend}{show legend}

\item{show_plot}{show plots}

\item{return_plot}{return ggplot object}

\item{save_plot}{directly save the plot [boolean]}

\item{save_param}{list of saving parameters from all_plots_save_function()}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}
}
\value{
ggplot
}
\description{
Visualize cells and gene expression according to spatial coordinates
}
\details{
Description of parameters.
}
\examples{
    spatGenePlot(gobject)
}
\seealso{
\code{\link{spatGenePlot3D}} and \code{\link{spatGenePlot2D}}
}
