% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visuals.R
\name{spatGenePlot}
\alias{spatGenePlot}
\title{spatGenePlot}
\usage{
spatGenePlot(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=spatGenePlot2D]{spatGenePlot2D}}
  \describe{
    \item{\code{gobject}}{giotto object}
    \item{\code{show_image}}{show a tissue background image}
    \item{\code{gimage}}{a giotto image}
    \item{\code{image_name}}{name of a giotto image}
    \item{\code{sdimx}}{x-axis dimension name (default = 'sdimx')}
    \item{\code{sdimy}}{y-axis dimension name (default = 'sdimy')}
    \item{\code{expression_values}}{gene expression values to use}
    \item{\code{genes}}{genes to show}
    \item{\code{cell_color_gradient}}{vector with 3 colors for numeric data}
    \item{\code{gradient_midpoint}}{midpoint for color gradient}
    \item{\code{gradient_limits}}{vector with lower and upper limits}
    \item{\code{show_network}}{show underlying spatial network}
    \item{\code{network_color}}{color of spatial network}
    \item{\code{spatial_network_name}}{name of spatial network to use}
    \item{\code{edge_alpha}}{alpha of edge}
    \item{\code{show_grid}}{show spatial grid}
    \item{\code{grid_color}}{color of spatial grid}
    \item{\code{spatial_grid_name}}{name of spatial grid to use}
    \item{\code{midpoint}}{expression midpoint}
    \item{\code{scale_alpha_with_expression}}{scale expression with ggplot alpha parameter}
    \item{\code{point_shape}}{shape of points (border, no_border or voronoi)}
    \item{\code{point_size}}{size of point (cell)}
    \item{\code{point_alpha}}{transparancy of points}
    \item{\code{point_border_col}}{color of border around points}
    \item{\code{point_border_stroke}}{stroke size of border around points}
    \item{\code{cow_n_col}}{cowplot param: how many columns}
    \item{\code{cow_rel_h}}{cowplot param: relative height}
    \item{\code{cow_rel_w}}{cowplot param: relative width}
    \item{\code{cow_align}}{cowplot param: how to align}
    \item{\code{show_legend}}{show legend}
    \item{\code{legend_text}}{size of legend text}
    \item{\code{background_color}}{color of plot background}
    \item{\code{vor_border_color}}{border colorr for voronoi plot}
    \item{\code{vor_max_radius}}{maximum radius for voronoi 'cells'}
    \item{\code{vor_alpha}}{transparancy of voronoi 'cells'}
    \item{\code{axis_text}}{size of axis text}
    \item{\code{axis_title}}{size of axis title}
    \item{\code{show_plot}}{show plots}
    \item{\code{return_plot}}{return ggplot object}
    \item{\code{save_plot}}{directly save the plot [boolean]}
    \item{\code{save_param}}{list of saving parameters, see \code{\link{showSaveParameters}}}
    \item{\code{default_save_name}}{default save name for saving, don't change, change save_name in save_param}
  }}
}
\value{
ggplot
}
\description{
Visualize cells and gene expression according to spatial coordinates
}
\details{
Description of parameters.
}
\examples{

data(mini_giotto_single_cell)

all_genes = slot(mini_giotto_single_cell, 'gene_ID')
selected_genes = all_genes[1:2]
spatGenePlot(mini_giotto_single_cell, genes = selected_genes, point_size = 3)

}
\seealso{
\code{\link{spatGenePlot3D}} and \code{\link{spatGenePlot2D}}

Other spatial gene expression visualizations: 
\code{\link{spatGenePlot2D}()},
\code{\link{spatGenePlot3D}()}
}
\concept{spatial gene expression visualizations}
