% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visuals.R
\name{spatGenePlot2D}
\alias{spatGenePlot2D}
\title{spatGenePlot2D}
\usage{
spatGenePlot2D(gobject, expression_values = c("normalized", "scaled",
  "custom"), genes, genes_high_color = "darkred",
  genes_mid_color = "white", genes_low_color = "darkblue",
  show_network = F, network_color = NULL,
  spatial_network_name = "spatial_network", edge_alpha = NULL,
  show_grid = F, grid_color = NULL,
  spatial_grid_name = "spatial_grid", midpoint = 0,
  scale_alpha_with_expression = TRUE, point_size = 1,
  point_border_col = "black", point_border_stroke = 0.1,
  show_legend = T, cow_n_col = 2, cow_rel_h = 1, cow_rel_w = 1,
  cow_align = "h", show_plots = F, return_plot = TRUE,
  save_plot = FALSE, save_dir = NULL, save_folder = NULL,
  save_name = NULL, save_format = NULL, show_saved_plot = FALSE, ...)
}
\arguments{
\item{gobject}{giotto object}

\item{expression_values}{gene expression values to use}

\item{genes}{genes to show}

\item{genes_high_color}{color represents high gene expression}

\item{genes_mid_color}{color represents middle gene expression}

\item{genes_low_color}{color represents low gene expression}

\item{show_network}{show underlying spatial network}

\item{network_color}{color of spatial network}

\item{spatial_network_name}{name of spatial network to use}

\item{show_grid}{show spatial grid}

\item{grid_color}{color of spatial grid}

\item{spatial_grid_name}{name of spatial grid to use}

\item{midpoint}{expression midpoint}

\item{scale_alpha_with_expression}{scale expression with ggplot alpha parameter}

\item{point_size}{size of point (cell)}

\item{point_border_col}{color of border around points}

\item{point_border_stroke}{stroke size of border around points}

\item{show_legend}{show legend}

\item{cow_n_col}{cowplot param: how many columns}

\item{cow_rel_h}{cowplot param: relative height}

\item{cow_rel_w}{cowplot param: relative width}

\item{cow_align}{cowplot param: how to align}

\item{show_plots}{show plots}

\item{return_plot}{return ggplot object}

\item{save_plot}{directly save the plot [boolean]}

\item{save_dir}{directory to save the plot}

\item{save_folder}{(optional) folder in directory to save the plot}

\item{save_name}{name of plot}

\item{save_format}{format of plot (e.g. tiff, png, pdf, ...)}

\item{show_saved_plot}{load & display the saved plot}

\item{...}{parameters for cowplot::save_plot()}
}
\value{
ggplot
}
\description{
Visualize cells and gene expression according to spatial coordinates
}
\details{
Description of parameters.
}
\examples{
    spatGenePlot2D(gobject)
}
