% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interoperability.R
\name{spatialExperimentToGiotto}
\alias{spatialExperimentToGiotto}
\title{Utility function to convert a SpatialExperiment object to a Giotto object}
\usage{
spatialExperimentToGiotto(
  spe,
  python_path,
  nn_network = NULL,
  sp_network = NULL,
  verbose = TRUE
)
}
\arguments{
\item{spe}{Input SpatialExperiment object to convert to a Giotto object.}

\item{python_path}{Specify the path to python.}

\item{nn_network}{Specify the name of the nearest neighbour network(s)
in the input SpatialExperiment object. Default \code{NULL} will use
all existing networks.}

\item{sp_network}{Specify the name of the spatial network(s) in the input
SpatialExperiment object. Default \code{NULL} will use all existing
networks. This can be a vector of multiple network names.}

\item{verbose}{A boolean value specifying if progress messages should
be displayed or not. Default \code{TRUE}.}
}
\value{
Giotto object
}
\description{
Utility function to convert a SpatialExperiment object to a Giotto object
}
\examples{
\dontrun{
library(SpatialExperiment)
example(read10xVisium, echo = FALSE)
spatialExperimentToGiotto(spe)
}
}
