% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_in_situ_visuals.R
\name{spatInSituPlotPoints}
\alias{spatInSituPlotPoints}
\title{spatInSituPlotPoints}
\usage{
spatInSituPlotPoints(
  gobject,
  show_image = F,
  gimage = NULL,
  image_name = NULL,
  largeImage_name = NULL,
  spat_unit = NULL,
  spat_loc_name = NULL,
  feats = NULL,
  feat_type = "rna",
  feats_color_code = NULL,
  feat_shape_code = NULL,
  sdimx = "x",
  sdimy = "y",
  point_size = 1.5,
  stroke = 0.5,
  expand_counts = FALSE,
  count_info_column = "count",
  jitter = c(0, 0),
  show_polygon = TRUE,
  use_overlap = TRUE,
  polygon_feat_type = "cell",
  polygon_color = "black",
  polygon_bg_color = "black",
  polygon_fill = NULL,
  polygon_fill_gradient = c("blue", "white", "red"),
  polygon_fill_gradient_midpoint = NULL,
  polygon_fill_as_factor = NULL,
  polygon_fill_code = NULL,
  polygon_alpha = 0.5,
  polygon_line_size = 2,
  axis_text = 8,
  axis_title = 8,
  legend_text = 6,
  coord_fix_ratio = 1,
  background_color = "black",
  show_legend = TRUE,
  plot_method = c("ggplot", "scattermore", "scattermost"),
  plot_last = c("polygons", "points"),
  show_plot = NA,
  return_plot = NA,
  save_plot = NA,
  save_param = list(),
  default_save_name = "spatInSituPlotPoints",
  verbose = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{show_image}{show a tissue background image}

\item{gimage}{a giotto image}

\item{image_name}{name of a giotto image}

\item{largeImage_name}{name of a giottoLargeImage}

\item{spat_unit}{spatial unit}

\item{spat_loc_name}{name of spatial locations}

\item{feats}{features to plot}

\item{feat_type}{feature types of the feats}

\item{feats_color_code}{code to color the provided features}

\item{feat_shape_code}{code to shape the provided feature types}

\item{sdimx}{spatial dimension x}

\item{sdimy}{spatial dimension y}

\item{point_size}{size of the points}

\item{expand_counts}{expand feature coordinate counts (see details)}

\item{count_info_column}{column name with count information (if expand_counts = TRUE)}

\item{jitter}{maximum x,y jitter provided as c(x, y)}

\item{show_polygon}{overlay polygon information (e.g. cell shape)}

\item{use_overlap}{use polygon and feature coordinates overlap results}

\item{polygon_feat_type}{feature type associated with polygon information}

\item{polygon_color}{color for polygon border}

\item{polygon_bg_color}{color for polygon background (overruled by polygon_fill)}

\item{polygon_fill}{fill color or column for polygon}

\item{polygon_fill_gradient}{polygon fill gradient colors given in order from low to high}

\item{polygon_fill_gradient_midpoint}{value to set as gradient midpoint (optional). If
left as \code{NULL}, the median value detected will be chosen}

\item{polygon_fill_as_factor}{is fill color a factor}

\item{polygon_fill_code}{code to color the fill column}

\item{polygon_alpha}{alpha of polygon}

\item{polygon_line_size}{line width of the polygon's outline}

\item{axis_text}{axis text size}

\item{axis_title}{title text size}

\item{legend_text}{legend text size}

\item{coord_fix_ratio}{fix ratio of coordinates}

\item{background_color}{background color}

\item{show_legend}{show legend}

\item{plot_method}{method to plot points}

\item{plot_last}{which layer to show on top of plot, polygons (default) or points.}

\item{show_plot}{show plots}

\item{return_plot}{return ggplot object}

\item{save_plot}{directly save the plot [boolean]}

\item{save_param}{list of saving parameters, see \code{\link{showSaveParameters}}}

\item{default_save_name}{default save name for saving, don't change, change save_name in save_param}

\item{verbose}{verbosity}

\item{line}{line size of point shapes}
}
\value{
ggplot
}
\description{
Function to plot multiple features for multiple modalities at the spatial in situ level
}
\details{
TODO
}
\seealso{
Other In Situ visualizations: 
\code{\link{spatInSituPlotDensity}()},
\code{\link{spatInSituPlotHex}()}
}
\concept{In Situ visualizations}
