% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{spatUnit-generic}
\alias{spatUnit-generic}
\alias{spatUnit}
\alias{spatUnit<-}
\alias{spatUnit,spatData-method}
\alias{spatUnit,giottoPolygon-method}
\alias{spatUnit<-,spatData-method}
\alias{spatUnit<-,giottoPolygon-method}
\title{Spatial unit information}
\usage{
spatUnit(x)

\S4method{spatUnit}{spatData}(x)

\S4method{spatUnit}{giottoPolygon}(x)

\S4method{spatUnit}{spatData}(x) <- value

\S4method{spatUnit}{giottoPolygon}(x) <- value
}
\arguments{
\item{x}{a Giotto S4 class subobject with spatial unit}

\item{value}{value to set as spatial unit}
}
\description{
access and set spat_unit slot of S4 subobject
}
\section{Functions}{
\itemize{
\item \code{spatUnit(spatData)}: Get spatial unit information

\item \code{spatUnit(giottoPolygon)}: Get spatial unit information

\item \code{spatUnit(spatData) <- value}: Set spatial unit information

\item \code{spatUnit(giottoPolygon) <- value}: Set giottoPolygon spat_unit

}}
