% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction.R
\name{specificCellCellcommunicationScores}
\alias{specificCellCellcommunicationScores}
\title{specificCellCellcommunicationScores}
\usage{
specificCellCellcommunicationScores(gobject,
  spatial_network_name = "spatial_network",
  cluster_column = "cell_types", random_iter = 100,
  cell_type_1 = "astrocyte", cell_type_2 = "endothelial", gene_set_1,
  gene_set_2, log2FC_addendum = 0.1, min_observations = 2,
  verbose = T)
}
\arguments{
\item{gobject}{giotto object to use}

\item{spatial_network_name}{spatial network to use for identifying interacting cells}

\item{cluster_column}{cluster column with cell type information}

\item{random_iter}{number of iterations}

\item{cell_type_1}{first cell type}

\item{cell_type_2}{second cell type}

\item{gene_set_1}{first specific gene set from gene pairs}

\item{gene_set_2}{second specific gene set from gene pairs}

\item{log2FC_addendum}{addendum to add when calculating log2FC}

\item{min_observations}{minimum number of interactions needed to be considered}

\item{verbose}{verbose}
}
\value{
Cell-Cell communication scores for gene pairs based on spatial interaction
}
\description{
Specific Cell-Cell communication scores based on spatial expression of interacting cells
}
\details{
Details will follow.
}
\examples{
    specificCellCellcommunicationScores(gobject)
}
