% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interaction.R
\name{specificCellCellcommunicationScores}
\alias{specificCellCellcommunicationScores}
\title{specificCellCellcommunicationScores}
\usage{
specificCellCellcommunicationScores(
  gobject,
  spatial_network_name = "Delaunay_network",
  cluster_column = "cell_types",
  random_iter = 100,
  cell_type_1 = "astrocyte",
  cell_type_2 = "endothelial",
  gene_set_1,
  gene_set_2,
  log2FC_addendum = 0.1,
  min_observations = 2,
  detailed = FALSE,
  adjust_method = c("fdr", "bonferroni", "BH", "holm", "hochberg", "hommel", "BY",
    "none"),
  adjust_target = c("genes", "cells"),
  set_seed = FALSE,
  seed_number = 1234,
  verbose = T
)
}
\arguments{
\item{gobject}{giotto object to use}

\item{spatial_network_name}{spatial network to use for identifying interacting cells}

\item{cluster_column}{cluster column with cell type information}

\item{random_iter}{number of iterations}

\item{cell_type_1}{first cell type}

\item{cell_type_2}{second cell type}

\item{gene_set_1}{first specific gene set from gene pairs}

\item{gene_set_2}{second specific gene set from gene pairs}

\item{log2FC_addendum}{addendum to add when calculating log2FC}

\item{min_observations}{minimum number of interactions needed to be considered}

\item{detailed}{provide more detailed information (random variance and z-score)}

\item{adjust_method}{which method to adjust p-values}

\item{adjust_target}{adjust multiple hypotheses at the cell or gene level}

\item{set_seed}{set a seed for reproducibility}

\item{seed_number}{seed number}

\item{verbose}{verbose}
}
\value{
Cell-Cell communication scores for gene pairs based on spatial interaction
}
\description{
Specific Cell-Cell communication scores based on spatial expression of interacting cells
}
\details{
Statistical framework to identify if pairs of genes (such as ligand-receptor combinations)
are expressed at higher levels than expected based on a reshuffled null distribution
of gene expression values in cells that are spatially in proximity to eachother.
\itemize{
 \item{LR_comb:}{Pair of ligand and receptor}
 \item{lig_cell_type:}{ cell type to assess expression level of ligand }
 \item{lig_expr:}{ average expression of ligand in lig_cell_type }
 \item{ligand:}{ ligand name }
 \item{rec_cell_type:}{ cell type to assess expression level of receptor }
 \item{rec_expr:}{ average expression of receptor in rec_cell_type}
 \item{receptor:}{ receptor name }
 \item{LR_expr:}{ combined average ligand and receptor expression }
 \item{lig_nr:}{ total number of cells from lig_cell_type that spatially interact with cells from rec_cell_type }
 \item{rec_nr:}{ total number of cells from rec_cell_type that spatially interact with cells from lig_cell_type }
 \item{rand_expr:}{ average combined ligand and receptor expression from random spatial permutations }
 \item{av_diff:}{ average difference between LR_expr and rand_expr over all random spatial permutations }
 \item{sd_diff:}{ (optional) standard deviation of the difference between LR_expr and rand_expr over all random spatial permutations }
 \item{z_score:}{ (optinal) z-score }
 \item{log2fc:}{ log2 fold-change (LR_expr/rand_expr) }
 \item{pvalue:}{ p-value }
 \item{LR_cell_comb:}{ cell type pair combination }
 \item{p.adj:}{ adjusted p-value }
 \item{PI:}{ significanc score: log2fc * -log10(p.adj) }
}
}
