% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{subset_cell_metadata}
\alias{subset_cell_metadata}
\title{Subset cell metadata}
\usage{
subset_cell_metadata(gobject, feat_type, filter_bool_cells, spat_unit)
}
\description{
Subset cell metadata from giotto object
}
\keyword{internal}
