% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{subset_dimension_reduction}
\alias{subset_dimension_reduction}
\title{Subset dimension reduction}
\usage{
subset_dimension_reduction(gobject, spat_unit, feat_type, cells_to_keep)
}
\description{
Subset dimension reduction results from giotto object
}
\keyword{internal}
