% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{subset_expression_data}
\alias{subset_expression_data}
\title{Subset expression data}
\usage{
subset_expression_data(
  gobject,
  cell_ids,
  feat_ids,
  feat_type,
  spat_unit,
  all_spat_units,
  all_feat_types
)
}
\description{
Subset expression data from giotto object
}
\keyword{internal}
