% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{subset_feature_metadata}
\alias{subset_feature_metadata}
\title{Subset feature metadata}
\usage{
subset_feature_metadata(
  gobject,
  feat_type,
  spat_unit = NULL,
  filter_bool_feats,
  all_spat_units = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{feat_type}{feature type (e.g. 'rna', 'protein')}

\item{spat_unit}{spatial unit (e.g. 'cell')}

\item{filter_bool_feats}{boolean filter to subset feat metadata}

\item{all_spat_units}{(boolean) applies subset operation across the whole gobject
(ALL spat_units), ignoring the \code{spat_unit} input param. Defaults to TRUE.}
}
\description{
Subset feature metadata from giotto object
}
\keyword{internal}
