% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{subset_nearest_network}
\alias{subset_nearest_network}
\title{Subset nearest network}
\usage{
subset_nearest_network(gobject, spat_unit, feat_type, cell_ids)
}
\description{
Subset nearest network results from giotto object
}
\keyword{internal}
