% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{subsetGiotto}
\alias{subsetGiotto}
\title{subsetGiotto}
\usage{
subsetGiotto(
  gobject,
  cell_ids = NULL,
  feat_type = NULL,
  feat_ids = NULL,
  gene_ids = NULL,
  verbose = FALSE
)
}
\arguments{
\item{gobject}{giotto object}

\item{cell_ids}{cell IDs to keep}

\item{feat_type}{feature type to use}

\item{feat_ids}{feature IDs to keep}

\item{verbose}{be verbose}
}
\value{
giotto object
}
\description{
subsets Giotto object including previous analyses.
}
\examples{
\donttest{

data(mini_giotto_single_cell)

random_cells = sample(slot(mini_giotto_single_cell, 'cell_ID'), 10)
random_genes = sample(slot(mini_giotto_single_cell, 'gene_ID'), 10)

subset_obj = subsetGiotto(mini_giotto_single_cell,
                         cell_ids = random_cells,
                         feat_ids = random_genes)

}

}
