% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{subsetGiotto}
\alias{subsetGiotto}
\title{subsetGiotto}
\usage{
subsetGiotto(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  cell_ids = NULL,
  feat_ids = NULL,
  gene_ids = NULL,
  poly_info = NULL,
  x_max = NULL,
  x_min = NULL,
  y_max = NULL,
  y_min = NULL,
  verbose = TRUE,
  toplevel_params = 2
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit}

\item{feat_type}{feature type to use}

\item{cell_ids}{cell IDs to keep}

\item{feat_ids}{feature IDs to keep}

\item{gene_ids}{deprecated, use feat_ids}

\item{poly_info}{polygon information to use}

\item{x_max}{maximum x-coordinate for feature coordinates}

\item{x_min}{minimum x-coordinate for feature coordinates}

\item{y_max}{maximum y-coordinate for feature coordinates}

\item{y_min}{minimum y-coordinate for feature coordinates}

\item{verbose}{be verbose}

\item{toplevel_params}{parameters to extract}
}
\value{
giotto object
}
\description{
subsets Giotto object including previous analyses.
}
\details{
Subsets a Giotto object for a specific spatial unit and feature type
}
\examples{
\donttest{

data(mini_giotto_single_cell)

random_cells = sample(slot(mini_giotto_single_cell, 'cell_ID'), 10)
random_genes = sample(slot(mini_giotto_single_cell, 'gene_ID'), 10)

subset_obj = subsetGiotto(mini_giotto_single_cell,
                         cell_ids = random_cells,
                         feat_ids = random_genes)

}

}
