% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{subsetGiottoLocs}
\alias{subsetGiottoLocs}
\title{subsetGiottoLocs}
\usage{
subsetGiottoLocs(
  gobject,
  x_max = NULL,
  x_min = NULL,
  y_max = NULL,
  y_min = NULL,
  z_max = NULL,
  z_min = NULL,
  return_gobject = T
)
}
\arguments{
\item{gobject}{giotto object}

\item{x_max}{maximum x-coordinate}

\item{x_min}{minimum x-coordinate}

\item{y_max}{maximum y-coordinate}

\item{y_min}{minimum y-coordinate}

\item{z_max}{maximum z-coordinate}

\item{z_min}{minimum z-coordinate}

\item{return_gobject}{return Giotto object}
}
\value{
giotto object
}
\description{
subsets Giotto object based on spatial locations
}
\details{
if return_gobject = FALSE, then a filtered combined metadata data.table will be returned
}
\examples{
    subsetGiottoLocs(gobject)
}
