% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_giotto.R
\name{subsetGiottoLocsMulti}
\alias{subsetGiottoLocsMulti}
\title{Subset by spatial locations -- multi}
\usage{
subsetGiottoLocsMulti(
  gobject,
  spat_unit = NULL,
  feat_type = NULL,
  spat_loc_name = NULL,
  x_max = NULL,
  x_min = NULL,
  y_max = NULL,
  y_min = NULL,
  z_max = NULL,
  z_min = NULL,
  poly_info = NULL,
  return_gobject = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{gobject}{giotto object}

\item{spat_unit}{spatial unit}

\item{feat_type}{feature type to use}

\item{spat_loc_name}{name of spatial locations to use}

\item{x_max, x_min, y_max, y_min, z_max, z_min}{minimum and maximum x, y, and z coordinates
to subset to}

\item{poly_info}{polygon information to use}

\item{return_gobject}{return Giotto object}

\item{verbose}{be verbose}
}
\value{
giotto object
}
\description{
Subsets Giotto object based on spatial locations
}
\details{
Subsets a Giotto based on spatial locations for multiple spatial units
if return_gobject = FALSE, then a filtered combined metadata data.table will be returned
}
