% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{transpose-generic}
\alias{transpose-generic}
\alias{t}
\alias{t,spatLocsObj-method}
\alias{t,spatialNetworkObj-method}
\title{Transpose}
\usage{
\S4method{t}{spatLocsObj}(x)

\S4method{t}{spatialNetworkObj}(x)
}
\arguments{
\item{x}{object to be transposed}
}
\description{
Transpose
}
\section{Functions}{
\itemize{
\item \code{t(spatLocsObj)}: Transpose a spatLocObj

\item \code{t(spatialNetworkObj)}: Transpose a spatialNeworkObj

}}
