% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{updateGiottoImage}
\alias{updateGiottoImage}
\title{updateGiottoImage}
\usage{
updateGiottoImage(
  gobject = NULL,
  image_name = NULL,
  largeImage_name = NULL,
  xmax_adj = 0,
  xmin_adj = 0,
  ymax_adj = 0,
  ymin_adj = 0,
  x_shift = 0,
  y_shift = 0,
  scale_x = 1,
  scale_y = 1,
  order = c("first_adj", "first_scale"),
  xmax_set = NULL,
  xmin_set = NULL,
  ymax_set = NULL,
  ymin_set = NULL,
  return_gobject = TRUE
)
}
\arguments{
\item{gobject}{gobject containing giottoImage or giottoLargeImage}

\item{image_name}{name of giottoImage}

\item{largeImage_name}{name of giottoLargeImage}

\item{xmax_adj}{increase of the maximum x image boundary}

\item{xmin_adj}{decrease of the minimum x image boundary}

\item{ymax_adj}{increase of the maximum y image boundary}

\item{ymin_adj}{decrease of the minimum y image boundary}

\item{x_shift}{shift entire image in positive x direction}

\item{y_shift}{shift entire image in positive y direction}

\item{scale_x}{scale x axis image mapping from coordinate origin}

\item{scale_y}{scale y axis image mapping from coordinate origin}

\item{order}{order of operations between fine adjustments (adjustment and shift
parameters) and scaling}

\item{xmax_set}{directly set image xmax boundary. Overrides minmax values as spatial anchor.}

\item{xmin_set}{directly set image xmin boundary. Overrides minmax values as spatial anchor.}

\item{ymax_set}{directly set image ymax boundary. Overrides minmax values as spatial anchor.}

\item{ymin_set}{directly set image ymin boundary. Overrides minmax values as spatial anchor.}

\item{return_gobject}{return a giotto object}
}
\value{
a giotto object or an updated giotto image object if return_gobject = F
}
\description{
Updates the spatial positioning and sizing of a giottoImage or
  giottoLargeImage attached to a giotto object.
}
\details{
This function works for all image objects associated with Giotto.
}
