% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{updateGiottoImage}
\alias{updateGiottoImage}
\title{updateGiottoImage}
\usage{
updateGiottoImage(
  gobject = NULL,
  image_name = NULL,
  largeImage_name = NULL,
  xmax_adj = 0,
  xmin_adj = 0,
  ymax_adj = 0,
  ymin_adj = 0,
  x_shift = 0,
  y_shift = 0,
  scale_x = 1,
  scale_y = 1,
  order = c("first_adj", "first_scale"),
  xmax_set = NULL,
  xmin_set = NULL,
  ymax_set = NULL,
  ymin_set = NULL,
  return_gobject = TRUE
)
}
\arguments{
\item{gobject}{gobject containing giottoImage or giottoLargeImage}

\item{image_name}{name of giottoImage}

\item{largeImage_name}{name of giottoLargeImage}

\item{xmax_adj, xmin_adj, ymax_adj, ymin_adj}{adjust image boundaries by increasing 
maximum and decreasing minimum bounds respectively of xy bounds}

\item{x_shift, y_shift}{shift entire image along xy axes}

\item{scale_x, scale_y}{scale xy axis image mapping from coordinate origin}

\item{order}{order of operations between fine adjustments (adjustment and shift
parameters) and scaling}

\item{xmin_set, xmax_set, ymin_set, ymax_set}{directly set xy image boundaries.
Overrides minmax values as spatial anchor.}

\item{return_gobject}{return a giotto object}
}
\value{
a giotto object or an updated giotto image object if return_gobject = F
}
\description{
Updates the spatial positioning and sizing of a giottoImage or
  giottoLargeImage attached to a giotto object.
}
\details{
This function works for all image objects associated with Giotto.
}
\seealso{
Other basic image functions: 
\code{\link{addGiottoImage}()},
\code{\link{plotGiottoImage}()},
\code{\link{reconnectGiottoImage}()}
}
\concept{basic image functions}
