% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{updateGiottoLargeImage}
\alias{updateGiottoLargeImage}
\title{updateGiottoLargeImage}
\usage{
updateGiottoLargeImage(
  gobject = NULL,
  largeImage_name = NULL,
  giottoLargeImage = NULL,
  xmax_adj = 0,
  xmin_adj = 0,
  ymax_adj = 0,
  ymin_adj = 0,
  x_shift = 0,
  y_shift = 0,
  scale_factor = NULL,
  scale_x = 1,
  scale_y = 1,
  order = c("first_adj", "first_scale"),
  xmin_set = NULL,
  xmax_set = NULL,
  ymin_set = NULL,
  ymax_set = NULL,
  return_gobject = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{gobject}{\code{giotto} object containing giotto \code{largeImage} object}

\item{largeImage_name}{name of giotto \code{largeImage} object}

\item{giottoLargeImage}{\code{largeImage} object to directly update}

\item{xmax_adj, xmin_adj, ymax_adj, ymin_adj}{adjust image boundaries by increasing 
maximum and decreasing minimum bounds respectively of xy bounds}

\item{x_shift, y_shift}{shift entire image along xy axes}

\item{scale_factor}{set \code{scale_x} and \code{scale_y} params at the same time}

\item{scale_x, scale_y}{independently scale x or y axis image mapping from coordinate origin}

\item{order}{order of operations between fine adjustments (adjustment and shift
parameters) and scaling}

\item{xmin_set, xmax_set, ymin_set, ymax_set}{directly set xy image boundaries.
Overrides minmax values as spatial anchor.}

\item{return_gobject}{return a \code{giotto} object if \code{TRUE}, a giotto 
\code{largeImage} object if \code{FALSE}}

\item{verbose}{be verbose}
}
\value{
a \code{giotto} object or an updated giotto \code{largeImage} object if
  \code{return_gobject = FALSE}
}
\description{
Updates the boundaries of a giotto \code{largeImage} object attached to 
  a \code{giotto} object if both \code{gobject} and \code{largeImage_name} params
  are given. Alternatively can directly accept and return as \code{largeImage}
}
