% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_hmrf.R
\name{viewHMRFresults}
\alias{viewHMRFresults}
\title{viewHMRFresults}
\usage{
viewHMRFresults(
  gobject,
  HMRFoutput,
  k = NULL,
  betas_to_view = NULL,
  third_dim = FALSE,
  ...
)
}
\arguments{
\item{gobject}{giotto object}

\item{HMRFoutput}{HMRF output from doHMRF}

\item{k}{number of HMRF domains}

\item{betas_to_view}{results from different betas that you want to view}

\item{third_dim}{3D data (boolean)}

\item{\dots}{additional paramters (see details)}
}
\value{
spatial plots with HMRF domains
}
\description{
View results from doHMRF.
}
\seealso{
\code{\link{spatPlot2D}} and \code{\link{spatPlot3D}}
}
