% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_hmrf.R
\name{viewHMRFresults3D}
\alias{viewHMRFresults3D}
\title{viewHMRFresults3D}
\usage{
viewHMRFresults3D(gobject, HMRFoutput, k = NULL, betas_to_view = NULL, ...)
}
\arguments{
\item{gobject}{giotto object}

\item{HMRFoutput}{HMRF output from doHMRF}

\item{k}{number of HMRF domains}

\item{betas_to_view}{results from different betas that you want to view}

\item{\dots}{additional parameters to spatPlot3D()}
}
\value{
spatial plots with HMRF domains
}
\description{
View results from doHMRF.
}
\seealso{
\code{\link{spatPlot3D}}
}
