% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_hmrf.R
\name{viewHMRFresults_V2}
\alias{viewHMRFresults_V2}
\title{viewHMRFresults_V2}
\usage{
viewHMRFresults_V2(
  gobject,
  k,
  betas,
  hmrf_name,
  spat_unit = NULL,
  feat_type = NULL,
  third_dim = FALSE,
  cow_n_col = 2,
  cow_rel_h = 1,
  cow_rel_w = 1,
  cow_align = "h",
  show_plot = TRUE,
  save_plot = TRUE,
  return_plot = TRUE,
  default_save_name = "HMRF_result",
  save_param = list(),
  ...
)
}
\arguments{
\item{gobject}{giotto object}

\item{k}{number of clusters in hmrf results}

\item{betas}{values of beta to plot}

\item{hmrf_name}{name of hmrf models}

\item{spat_unit}{spatial unit}

\item{feat_type}{feature type}

\item{third_dim}{index of 3D plot}

\item{cow_n_col}{number of columns to show in figure}

\item{cow_rel_h}{relative height}

\item{cow_rel_w}{relative width}

\item{cow_align}{cowplot alignment parameter}

\item{show_plot}{option to show plot}

\item{save_plot}{option to save plot}

\item{return_plot}{if function return a plot}

\item{default_save_name}{names of figure file to save}

\item{save_param}{other saving parameters}

\item{\dots}{additional params to pass to plotting}
}
\description{
function to view HMRF results with multiple betas
}
\details{
This function plots spatial map of HMRF domain clusters for multiple beta with the name (hmrf_name),
matching the first part of the cell meta column names with HMRF clusters (for example name of ‘hmrf1 k=8 b=0.00’ is ‘hmrf1’)
}
