% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_visuals.R
\name{violinPlot}
\alias{violinPlot}
\title{violinPlot}
\usage{
violinPlot(gobject, expression_values = c("normalized", "scaled",
  "custom"), genes, cluster_column, cluster_custom_order = NULL,
  color_violin = c("genes", "cluster"), cluster_color_code = NULL,
  strip.text = 7, axis.text.x.size = 10, axis.text.y.size = 6)
}
\arguments{
\item{gobject}{giotto object}

\item{expression_values}{expression values to use}

\item{genes}{genes to plot}

\item{cluster_column}{name of column to use for clusters}

\item{color_violin}{color violinplots according to genes or clusters}
}
\value{
ggplot
}
\description{
Creates heatmap based on identified clusters
}
\details{
Correlation heatmap of clusters vs genes.
}
\examples{
    violinPlot(gobject)
}
