% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visuals.R
\name{visForceLayoutPlot}
\alias{visForceLayoutPlot}
\title{visForceLayoutPlot}
\usage{
visForceLayoutPlot(
  gobject,
  nn_network_to_use = "sNN",
  network_name = "sNN.pca",
  layout_name = "layout",
  dim1_to_use = 1,
  dim2_to_use = 2,
  show_NN_network = T,
  cell_color = NULL,
  color_as_factor = F,
  cell_color_code = NULL,
  edge_alpha = NULL,
  point_size = 1,
  point_border_col = "black",
  point_border_stroke = 0.1,
  show_legend = T,
  show_plot = F,
  return_plot = TRUE,
  save_plot = F,
  save_dir = NULL,
  save_folder = NULL,
  save_name = NULL,
  save_format = NULL,
  show_saved_plot = F,
  ...
)
}
\arguments{
\item{gobject}{giotto object}

\item{nn_network_to_use}{type of NN network to use (kNN vs sNN)}

\item{network_name}{NN network to use}

\item{layout_name}{name of layout to use}

\item{dim1_to_use}{dimension to use on x-axis}

\item{dim2_to_use}{dimension to use on y-axis}

\item{show_NN_network}{show underlying NN network}

\item{cell_color}{color for cells (see details)}

\item{color_as_factor}{convert color column to factor}

\item{cell_color_code}{named vector with colors}

\item{edge_alpha}{column to use for alpha of the edges}

\item{point_size}{size of point (cell)}

\item{point_border_col}{color of border around points}

\item{point_border_stroke}{stroke size of border around points}

\item{show_legend}{show legend}

\item{show_plot}{show plot}

\item{return_plot}{return ggplot object}

\item{save_plot}{directly save the plot [boolean]}

\item{save_dir}{directory to save the plot}

\item{save_folder}{(optional) folder in directory to save the plot}

\item{save_name}{name of plot}

\item{save_format}{format of plot (e.g. tiff, png, pdf, ...)}

\item{show_saved_plot}{load & display the saved plot}
}
\value{
ggplot
}
\description{
Visualize cells according to forced layout algorithm coordinates
}
\details{
Description of parameters.
}
\examples{
    visForceLayoutPlot(gobject)
}
