% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visuals.R
\name{visPlot}
\alias{visPlot}
\title{visPlot}
\usage{
visPlot(gobject, sdimx = NULL, sdimy = NULL, sdimz = NULL,
  point_size = 3, point_border_col = "black",
  point_border_stroke = 0.1, cell_color = NULL,
  cell_color_code = NULL, color_as_factor = T,
  select_cell_groups = NULL, select_cells = NULL,
  show_other_cells = T, other_cell_color = "lightgrey",
  show_network = F, network_color = NULL, network_alpha = 1,
  other_cells_alpha = 0.1, spatial_network_name = "spatial_network",
  show_grid = F, grid_color = NULL, grid_alpha = 1,
  spatial_grid_name = "spatial_grid", coord_fix_ratio = 0.6,
  title = "", show_legend = T, axis_scale = c("cube", "real",
  "custom"), custom_ratio = NULL, x_ticks = NULL, y_ticks = NULL,
  z_ticks = NULL, plot_method = c("ggplot", "plotly"), show_plot = F)
}
\arguments{
\item{gobject}{giotto object}

\item{sdimx}{x-axis dimension name (default = 'sdimx')}

\item{sdimy}{y-axis dimension name (default = 'sdimy')}

\item{sdimz}{z-axis dimension name (default = 'sdimz')}

\item{point_size}{size of point (cell)}

\item{point_border_col}{color of border around points}

\item{point_border_stroke}{stroke size of border around points}

\item{cell_color}{color for cells (see details)}

\item{cell_color_code}{named vector with colors}

\item{color_as_factor}{convert color column to factor}

\item{select_cell_groups}{select subset of cells/clusters based on cell_color parameter}

\item{select_cells}{select subset of cells based on cell IDs}

\item{show_other_cells}{display not selected cells}

\item{other_cell_color}{color of not selected cells}

\item{show_network}{show underlying spatial network}

\item{network_color}{color of spatial network}

\item{spatial_network_name}{name of spatial network to use}

\item{show_grid}{show spatial grid}

\item{grid_color}{color of spatial grid}

\item{spatial_grid_name}{name of spatial grid to use}

\item{coord_fix_ratio}{fix ratio between x and y-axis}

\item{title}{title of plot}

\item{show_legend}{show legend}

\item{show_plot}{show plot}
}
\value{
ggplot
}
\description{
Visualize cells according to spatial coordinates
}
\details{
Description of parameters.
}
\examples{
    visPlot(gobject)
}
