% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{wrap_msg}
\alias{wrap_msg}
\title{Wrap message}
\usage{
wrap_msg(..., collapse = "")
}
\arguments{
\item{...}{additional strings and/or elements to pass to cat}

\item{collapse}{how to join elements of string (default is no space)}
}
\description{
Wrap message
}
\keyword{internal}
