% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_registration.R
\name{xy_translate_spatial_locations}
\alias{xy_translate_spatial_locations}
\title{xy_translate_spatial_locations}
\usage{
xy_translate_spatial_locations(spatlocs, xtranslate, ytranslate)
}
\arguments{
\item{spatlocs}{spatial locations to use}

\item{xtranslate}{value to translate coordinates in the positive x direction}

\item{ytranslate}{value to translate coordinates in the positive y direction}
}
\description{
Translate given X Y coordinates by given x and y translation values
}
\keyword{internal}
